% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentsforamazonbedrock_operations.R
\name{agentsforamazonbedrock_delete_agent_alias}
\alias{agentsforamazonbedrock_delete_agent_alias}
\title{Deletes an alias of an agent}
\usage{
agentsforamazonbedrock_delete_agent_alias(agentAliasId, agentId)
}
\arguments{
\item{agentAliasId}{[required] The unique identifier of the alias to delete.}

\item{agentId}{[required] The unique identifier of the agent that the alias belongs to.}
}
\description{
Deletes an alias of an agent.

See \url{https://www.paws-r-sdk.com/docs/agentsforamazonbedrock_delete_agent_alias/} for full documentation.
}
\keyword{internal}
