% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentsforamazonbedrock_operations.R
\name{agentsforamazonbedrock_delete_data_source}
\alias{agentsforamazonbedrock_delete_data_source}
\title{Deletes a data source from a knowledge base}
\usage{
agentsforamazonbedrock_delete_data_source(dataSourceId, knowledgeBaseId)
}
\arguments{
\item{dataSourceId}{[required] The unique identifier of the data source to delete.}

\item{knowledgeBaseId}{[required] The unique identifier of the knowledge base from which to delete the
data source.}
}
\description{
Deletes a data source from a knowledge base.

See \url{https://www.paws-r-sdk.com/docs/agentsforamazonbedrock_delete_data_source/} for full documentation.
}
\keyword{internal}
