% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentsforamazonbedrock_operations.R
\name{agentsforamazonbedrock_delete_flow_alias}
\alias{agentsforamazonbedrock_delete_flow_alias}
\title{Deletes an alias of a flow}
\usage{
agentsforamazonbedrock_delete_flow_alias(aliasIdentifier, flowIdentifier)
}
\arguments{
\item{aliasIdentifier}{[required] The unique identifier of the alias to be deleted.}

\item{flowIdentifier}{[required] The unique identifier of the flow that the alias belongs to.}
}
\description{
Deletes an alias of a flow.

See \url{https://www.paws-r-sdk.com/docs/agentsforamazonbedrock_delete_flow_alias/} for full documentation.
}
\keyword{internal}
