% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentsforamazonbedrock_operations.R
\name{agentsforamazonbedrock_disassociate_agent_collaborator}
\alias{agentsforamazonbedrock_disassociate_agent_collaborator}
\title{Disassociates an agent collaborator}
\usage{
agentsforamazonbedrock_disassociate_agent_collaborator(
  agentId,
  agentVersion,
  collaboratorId
)
}
\arguments{
\item{agentId}{[required] An agent ID.}

\item{agentVersion}{[required] The agent's version.}

\item{collaboratorId}{[required] The collaborator's ID.}
}
\description{
Disassociates an agent collaborator.

See \url{https://www.paws-r-sdk.com/docs/agentsforamazonbedrock_disassociate_agent_collaborator/} for full documentation.
}
\keyword{internal}
