% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentsforamazonbedrock_operations.R
\name{agentsforamazonbedrock_get_agent_action_group}
\alias{agentsforamazonbedrock_get_agent_action_group}
\title{Gets information about an action group for an agent}
\usage{
agentsforamazonbedrock_get_agent_action_group(
  actionGroupId,
  agentId,
  agentVersion
)
}
\arguments{
\item{actionGroupId}{[required] The unique identifier of the action group for which to get information.}

\item{agentId}{[required] The unique identifier of the agent that the action group belongs to.}

\item{agentVersion}{[required] The version of the agent that the action group belongs to.}
}
\description{
Gets information about an action group for an agent.

See \url{https://www.paws-r-sdk.com/docs/agentsforamazonbedrock_get_agent_action_group/} for full documentation.
}
\keyword{internal}
