% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentsforamazonbedrock_operations.R
\name{agentsforamazonbedrock_get_agent_knowledge_base}
\alias{agentsforamazonbedrock_get_agent_knowledge_base}
\title{Gets information about a knowledge base associated with an agent}
\usage{
agentsforamazonbedrock_get_agent_knowledge_base(
  agentId,
  agentVersion,
  knowledgeBaseId
)
}
\arguments{
\item{agentId}{[required] The unique identifier of the agent with which the knowledge base is
associated.}

\item{agentVersion}{[required] The version of the agent with which the knowledge base is associated.}

\item{knowledgeBaseId}{[required] The unique identifier of the knowledge base associated with the agent.}
}
\description{
Gets information about a knowledge base associated with an agent.

See \url{https://www.paws-r-sdk.com/docs/agentsforamazonbedrock_get_agent_knowledge_base/} for full documentation.
}
\keyword{internal}
