% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentsforamazonbedrock_operations.R
\name{agentsforamazonbedrock_get_agent_version}
\alias{agentsforamazonbedrock_get_agent_version}
\title{Gets details about a version of an agent}
\usage{
agentsforamazonbedrock_get_agent_version(agentId, agentVersion)
}
\arguments{
\item{agentId}{[required] The unique identifier of the agent.}

\item{agentVersion}{[required] The version of the agent.}
}
\description{
Gets details about a version of an agent.

See \url{https://www.paws-r-sdk.com/docs/agentsforamazonbedrock_get_agent_version/} for full documentation.
}
\keyword{internal}
