% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentsforamazonbedrock_operations.R
\name{agentsforamazonbedrock_get_data_source}
\alias{agentsforamazonbedrock_get_data_source}
\title{Gets information about a data source}
\usage{
agentsforamazonbedrock_get_data_source(dataSourceId, knowledgeBaseId)
}
\arguments{
\item{dataSourceId}{[required] The unique identifier of the data source.}

\item{knowledgeBaseId}{[required] The unique identifier of the knowledge base for the data source.}
}
\description{
Gets information about a data source.

See \url{https://www.paws-r-sdk.com/docs/agentsforamazonbedrock_get_data_source/} for full documentation.
}
\keyword{internal}
