% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentsforamazonbedrock_operations.R
\name{agentsforamazonbedrock_get_flow_alias}
\alias{agentsforamazonbedrock_get_flow_alias}
\title{Retrieves information about a flow}
\usage{
agentsforamazonbedrock_get_flow_alias(aliasIdentifier, flowIdentifier)
}
\arguments{
\item{aliasIdentifier}{[required] The unique identifier of the alias for which to retrieve information.}

\item{flowIdentifier}{[required] The unique identifier of the flow that the alias belongs to.}
}
\description{
Retrieves information about a flow. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/flows-deploy.html}{Deploy a flow in Amazon Bedrock} in the Amazon Bedrock User Guide.

See \url{https://www.paws-r-sdk.com/docs/agentsforamazonbedrock_get_flow_alias/} for full documentation.
}
\keyword{internal}
