% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentsforamazonbedrock_operations.R
\name{agentsforamazonbedrock_get_ingestion_job}
\alias{agentsforamazonbedrock_get_ingestion_job}
\title{Gets information about a data ingestion job}
\usage{
agentsforamazonbedrock_get_ingestion_job(
  dataSourceId,
  ingestionJobId,
  knowledgeBaseId
)
}
\arguments{
\item{dataSourceId}{[required] The unique identifier of the data source for the data ingestion job you
want to get information on.}

\item{ingestionJobId}{[required] The unique identifier of the data ingestion job you want to get
information on.}

\item{knowledgeBaseId}{[required] The unique identifier of the knowledge base for the data ingestion job
you want to get information on.}
}
\description{
Gets information about a data ingestion job. Data sources are ingested into your knowledge base so that Large Language Models (LLMs) can use your data.

See \url{https://www.paws-r-sdk.com/docs/agentsforamazonbedrock_get_ingestion_job/} for full documentation.
}
\keyword{internal}
