% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentsforamazonbedrock_operations.R
\name{agentsforamazonbedrock_get_knowledge_base}
\alias{agentsforamazonbedrock_get_knowledge_base}
\title{Gets information about a knoweldge base}
\usage{
agentsforamazonbedrock_get_knowledge_base(knowledgeBaseId)
}
\arguments{
\item{knowledgeBaseId}{[required] The unique identifier of the knowledge base you want to get information
on.}
}
\description{
Gets information about a knoweldge base.

See \url{https://www.paws-r-sdk.com/docs/agentsforamazonbedrock_get_knowledge_base/} for full documentation.
}
\keyword{internal}
