% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentsforamazonbedrock_operations.R
\name{agentsforamazonbedrock_stop_ingestion_job}
\alias{agentsforamazonbedrock_stop_ingestion_job}
\title{Stops a currently running data ingestion job}
\usage{
agentsforamazonbedrock_stop_ingestion_job(
  dataSourceId,
  ingestionJobId,
  knowledgeBaseId
)
}
\arguments{
\item{dataSourceId}{[required] The unique identifier of the data source for the data ingestion job you
want to stop.}

\item{ingestionJobId}{[required] The unique identifier of the data ingestion job you want to stop.}

\item{knowledgeBaseId}{[required] The unique identifier of the knowledge base for the data ingestion job
you want to stop.}
}
\description{
Stops a currently running data ingestion job. You can send a \code{\link[=agentsforamazonbedrock_start_ingestion_job]{start_ingestion_job}} request again to ingest the rest of your data when you are ready.

See \url{https://www.paws-r-sdk.com/docs/agentsforamazonbedrock_stop_ingestion_job/} for full documentation.
}
\keyword{internal}
