% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentsforamazonbedrock_operations.R
\name{agentsforamazonbedrock_update_agent_alias}
\alias{agentsforamazonbedrock_update_agent_alias}
\title{Updates configurations for an alias of an agent}
\usage{
agentsforamazonbedrock_update_agent_alias(
  agentAliasId,
  agentAliasName,
  agentId,
  description = NULL,
  routingConfiguration = NULL
)
}
\arguments{
\item{agentAliasId}{[required] The unique identifier of the alias.}

\item{agentAliasName}{[required] Specifies a new name for the alias.}

\item{agentId}{[required] The unique identifier of the agent.}

\item{description}{Specifies a new description for the alias.}

\item{routingConfiguration}{Contains details about the routing configuration of the alias.}
}
\description{
Updates configurations for an alias of an agent.

See \url{https://www.paws-r-sdk.com/docs/agentsforamazonbedrock_update_agent_alias/} for full documentation.
}
\keyword{internal}
