% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentsforamazonbedrockruntime_operations.R
\name{agentsforamazonbedrockruntime_generate_query}
\alias{agentsforamazonbedrockruntime_generate_query}
\title{Generates an SQL query from a natural language query}
\usage{
agentsforamazonbedrockruntime_generate_query(
  queryGenerationInput,
  transformationConfiguration
)
}
\arguments{
\item{queryGenerationInput}{[required] Specifies information about a natural language query to transform into
SQL.}

\item{transformationConfiguration}{[required] Specifies configurations for transforming the natural language query
into SQL.}
}
\description{
Generates an SQL query from a natural language query. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/knowledge-base-generate-query.html}{Generate a query for structured data} in the Amazon Bedrock User Guide.

See \url{https://www.paws-r-sdk.com/docs/agentsforamazonbedrockruntime_generate_query/} for full documentation.
}
\keyword{internal}
