% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentsforamazonbedrockruntime_operations.R
\name{agentsforamazonbedrockruntime_retrieve}
\alias{agentsforamazonbedrockruntime_retrieve}
\title{Queries a knowledge base and retrieves information from it}
\usage{
agentsforamazonbedrockruntime_retrieve(
  guardrailConfiguration = NULL,
  knowledgeBaseId,
  nextToken = NULL,
  retrievalConfiguration = NULL,
  retrievalQuery
)
}
\arguments{
\item{guardrailConfiguration}{Guardrail settings.}

\item{knowledgeBaseId}{[required] The unique identifier of the knowledge base to query.}

\item{nextToken}{If there are more results than can fit in the response, the response
returns a \code{nextToken}. Use this token in the \code{nextToken} field of
another request to retrieve the next batch of results.}

\item{retrievalConfiguration}{Contains configurations for the knowledge base query and retrieval
process. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/kb-test-config.html}{Query configurations}.}

\item{retrievalQuery}{[required] Contains the query to send the knowledge base.}
}
\description{
Queries a knowledge base and retrieves information from it.

See \url{https://www.paws-r-sdk.com/docs/agentsforamazonbedrockruntime_retrieve/} for full documentation.
}
\keyword{internal}
