% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentedairuntime_operations.R
\name{augmentedairuntime_delete_human_loop}
\alias{augmentedairuntime_delete_human_loop}
\title{Deletes the specified human loop for a flow definition}
\usage{
augmentedairuntime_delete_human_loop(HumanLoopName)
}
\arguments{
\item{HumanLoopName}{[required] The name of the human loop that you want to delete.}
}
\description{
Deletes the specified human loop for a flow definition.

See \url{https://www.paws-r-sdk.com/docs/augmentedairuntime_delete_human_loop/} for full documentation.
}
\keyword{internal}
