% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentedairuntime_operations.R
\name{augmentedairuntime_describe_human_loop}
\alias{augmentedairuntime_describe_human_loop}
\title{Returns information about the specified human loop}
\usage{
augmentedairuntime_describe_human_loop(HumanLoopName)
}
\arguments{
\item{HumanLoopName}{[required] The name of the human loop that you want information about.}
}
\description{
Returns information about the specified human loop. If the human loop was deleted, this operation will return a \code{ResourceNotFoundException} error.

See \url{https://paws-r.github.io/docs/augmentedairuntime/describe_human_loop.html} for full documentation.
}
\keyword{internal}
