% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentedairuntime_operations.R
\name{augmentedairuntime_start_human_loop}
\alias{augmentedairuntime_start_human_loop}
\title{Starts a human loop, provided that at least one activation condition is
met}
\usage{
augmentedairuntime_start_human_loop(
  HumanLoopName,
  FlowDefinitionArn,
  HumanLoopInput,
  DataAttributes = NULL
)
}
\arguments{
\item{HumanLoopName}{[required] The name of the human loop.}

\item{FlowDefinitionArn}{[required] The Amazon Resource Name (ARN) of the flow definition associated with
this human loop.}

\item{HumanLoopInput}{[required] An object that contains information about the human loop.}

\item{DataAttributes}{Attributes of the specified data. Use \code{DataAttributes} to specify if
your data is free of personally identifiable information and/or free of
adult content.}
}
\description{
Starts a human loop, provided that at least one activation condition is met.

See \url{https://paws-r.github.io/docs/augmentedairuntime/start_human_loop.html} for full documentation.
}
\keyword{internal}
