% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_batch_delete_evaluation_job}
\alias{bedrock_batch_delete_evaluation_job}
\title{Creates a batch deletion job}
\usage{
bedrock_batch_delete_evaluation_job(jobIdentifiers)
}
\arguments{
\item{jobIdentifiers}{[required] An array of model evaluation job ARNs to be deleted.}
}
\description{
Creates a batch deletion job. A model evaluation job can only be deleted if it has following status \code{FAILED}, \code{COMPLETED}, and \code{STOPPED}. You can request up to 25 model evaluation jobs be deleted in a single request.

See \url{https://www.paws-r-sdk.com/docs/bedrock_batch_delete_evaluation_job/} for full documentation.
}
\keyword{internal}
