% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_create_foundation_model_agreement}
\alias{bedrock_create_foundation_model_agreement}
\title{Request a model access agreement for the specified model}
\usage{
bedrock_create_foundation_model_agreement(offerToken, modelId)
}
\arguments{
\item{offerToken}{[required] An offer token encapsulates the information for an offer.}

\item{modelId}{[required] Model Id of the model for the access request.}
}
\description{
Request a model access agreement for the specified model.

See \url{https://www.paws-r-sdk.com/docs/bedrock_create_foundation_model_agreement/} for full documentation.
}
\keyword{internal}
