% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_delete_automated_reasoning_policy_test_case}
\alias{bedrock_delete_automated_reasoning_policy_test_case}
\title{Deletes an Automated Reasoning policy test}
\usage{
bedrock_delete_automated_reasoning_policy_test_case(
  policyArn,
  testCaseId,
  lastUpdatedAt
)
}
\arguments{
\item{policyArn}{[required] The Amazon Resource Name (ARN) of the Automated Reasoning policy that
contains the test.}

\item{testCaseId}{[required] The unique identifier of the test to delete.}

\item{lastUpdatedAt}{[required] The timestamp when the test was last updated. This is used as a
concurrency token to prevent conflicting modifications.}
}
\description{
Deletes an Automated Reasoning policy test. This operation is idempotent; if you delete a test more than once, each call succeeds.

See \url{https://www.paws-r-sdk.com/docs/bedrock_delete_automated_reasoning_policy_test_case/} for full documentation.
}
\keyword{internal}
