% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_delete_custom_model_deployment}
\alias{bedrock_delete_custom_model_deployment}
\title{Deletes a custom model deployment}
\usage{
bedrock_delete_custom_model_deployment(customModelDeploymentIdentifier)
}
\arguments{
\item{customModelDeploymentIdentifier}{[required] The Amazon Resource Name (ARN) or name of the custom model deployment to
delete.}
}
\description{
Deletes a custom model deployment. This operation stops the deployment and removes it from your account. After deletion, the deployment ARN can no longer be used for inference requests.

See \url{https://www.paws-r-sdk.com/docs/bedrock_delete_custom_model_deployment/} for full documentation.
}
\keyword{internal}
