% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_delete_foundation_model_agreement}
\alias{bedrock_delete_foundation_model_agreement}
\title{Delete the model access agreement for the specified model}
\usage{
bedrock_delete_foundation_model_agreement(modelId)
}
\arguments{
\item{modelId}{[required] Model Id of the model access to delete.}
}
\description{
Delete the model access agreement for the specified model.

See \url{https://www.paws-r-sdk.com/docs/bedrock_delete_foundation_model_agreement/} for full documentation.
}
\keyword{internal}
