% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_delete_guardrail}
\alias{bedrock_delete_guardrail}
\title{Deletes a guardrail}
\usage{
bedrock_delete_guardrail(guardrailIdentifier, guardrailVersion = NULL)
}
\arguments{
\item{guardrailIdentifier}{[required] The unique identifier of the guardrail. This can be an ID or the ARN.}

\item{guardrailVersion}{The version of the guardrail.}
}
\description{
Deletes a guardrail.

See \url{https://www.paws-r-sdk.com/docs/bedrock_delete_guardrail/} for full documentation.
}
\keyword{internal}
