% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_delete_marketplace_model_endpoint}
\alias{bedrock_delete_marketplace_model_endpoint}
\title{Deletes an endpoint for a model from Amazon Bedrock Marketplace}
\usage{
bedrock_delete_marketplace_model_endpoint(endpointArn)
}
\arguments{
\item{endpointArn}{[required] The Amazon Resource Name (ARN) of the endpoint you want to delete.}
}
\description{
Deletes an endpoint for a model from Amazon Bedrock Marketplace.

See \url{https://www.paws-r-sdk.com/docs/bedrock_delete_marketplace_model_endpoint/} for full documentation.
}
\keyword{internal}
