% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_delete_model_invocation_logging_configuration}
\alias{bedrock_delete_model_invocation_logging_configuration}
\title{Delete the invocation logging}
\usage{
bedrock_delete_model_invocation_logging_configuration()
}
\description{
Delete the invocation logging.

See \url{https://www.paws-r-sdk.com/docs/bedrock_delete_model_invocation_logging_configuration/} for full documentation.
}
\keyword{internal}
