% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_delete_prompt_router}
\alias{bedrock_delete_prompt_router}
\title{Deletes a specified prompt router}
\usage{
bedrock_delete_prompt_router(promptRouterArn)
}
\arguments{
\item{promptRouterArn}{[required] The Amazon Resource Name (ARN) of the prompt router to delete.}
}
\description{
Deletes a specified prompt router. This action cannot be undone.

See \url{https://www.paws-r-sdk.com/docs/bedrock_delete_prompt_router/} for full documentation.
}
\keyword{internal}
