% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_delete_resource_policy}
\alias{bedrock_delete_resource_policy}
\title{Deletes a previously created Bedrock resource policy}
\usage{
bedrock_delete_resource_policy(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the Bedrock resource to which this resource policy applies.}
}
\description{
Deletes a previously created Bedrock resource policy.

See \url{https://www.paws-r-sdk.com/docs/bedrock_delete_resource_policy/} for full documentation.
}
\keyword{internal}
