% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_deregister_marketplace_model_endpoint}
\alias{bedrock_deregister_marketplace_model_endpoint}
\title{Deregisters an endpoint for a model from Amazon Bedrock Marketplace}
\usage{
bedrock_deregister_marketplace_model_endpoint(endpointArn)
}
\arguments{
\item{endpointArn}{[required] The Amazon Resource Name (ARN) of the endpoint you want to deregister.}
}
\description{
Deregisters an endpoint for a model from Amazon Bedrock Marketplace. This operation removes the endpoint's association with Amazon Bedrock but does not delete the underlying Amazon SageMaker endpoint.

See \url{https://www.paws-r-sdk.com/docs/bedrock_deregister_marketplace_model_endpoint/} for full documentation.
}
\keyword{internal}
