% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_get_account_data_retention}
\alias{bedrock_get_account_data_retention}
\title{Returns the account-wide data retention mode for Amazon Bedrock}
\usage{
bedrock_get_account_data_retention()
}
\description{
Returns the account-wide data retention mode for Amazon Bedrock.

See \url{https://www.paws-r-sdk.com/docs/bedrock_get_account_data_retention/} for full documentation.
}
\keyword{internal}
