% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_get_advanced_prompt_optimization_job}
\alias{bedrock_get_advanced_prompt_optimization_job}
\title{Gets information about an advanced prompt optimization job}
\usage{
bedrock_get_advanced_prompt_optimization_job(jobIdentifier)
}
\arguments{
\item{jobIdentifier}{[required] The ARN or ID of the advanced prompt optimization job.}
}
\description{
Gets information about an advanced prompt optimization job.

See \url{https://www.paws-r-sdk.com/docs/bedrock_get_advanced_prompt_optimization_job/} for full documentation.
}
\keyword{internal}
