% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_get_automated_reasoning_policy_build_workflow}
\alias{bedrock_get_automated_reasoning_policy_build_workflow}
\title{Retrieves detailed information about an Automated Reasoning policy build
workflow, including its status, configuration, and metadata}
\usage{
bedrock_get_automated_reasoning_policy_build_workflow(
  policyArn,
  buildWorkflowId
)
}
\arguments{
\item{policyArn}{[required] The Amazon Resource Name (ARN) of the Automated Reasoning policy whose
build workflow you want to retrieve.}

\item{buildWorkflowId}{[required] The unique identifier of the build workflow to retrieve.}
}
\description{
Retrieves detailed information about an Automated Reasoning policy build workflow, including its status, configuration, and metadata.

See \url{https://www.paws-r-sdk.com/docs/bedrock_get_automated_reasoning_policy_build_workflow/} for full documentation.
}
\keyword{internal}
