% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_get_automated_reasoning_policy_next_scenario}
\alias{bedrock_get_automated_reasoning_policy_next_scenario}
\title{Retrieves the next test scenario for validating an Automated Reasoning
policy}
\usage{
bedrock_get_automated_reasoning_policy_next_scenario(
  policyArn,
  buildWorkflowId
)
}
\arguments{
\item{policyArn}{[required] The Amazon Resource Name (ARN) of the Automated Reasoning policy for
which you want to get the next test scenario.}

\item{buildWorkflowId}{[required] The unique identifier of the build workflow associated with the test
scenarios.}
}
\description{
Retrieves the next test scenario for validating an Automated Reasoning policy. This is used during the interactive policy refinement process to test policy behavior.

See \url{https://www.paws-r-sdk.com/docs/bedrock_get_automated_reasoning_policy_next_scenario/} for full documentation.
}
\keyword{internal}
