% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_get_automated_reasoning_policy_test_case}
\alias{bedrock_get_automated_reasoning_policy_test_case}
\title{Retrieves details about a specific Automated Reasoning policy test}
\usage{
bedrock_get_automated_reasoning_policy_test_case(policyArn, testCaseId)
}
\arguments{
\item{policyArn}{[required] The Amazon Resource Name (ARN) of the Automated Reasoning policy that
contains the test.}

\item{testCaseId}{[required] The unique identifier of the test to retrieve.}
}
\description{
Retrieves details about a specific Automated Reasoning policy test.

See \url{https://www.paws-r-sdk.com/docs/bedrock_get_automated_reasoning_policy_test_case/} for full documentation.
}
\keyword{internal}
