% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_get_automated_reasoning_policy_test_result}
\alias{bedrock_get_automated_reasoning_policy_test_result}
\title{Retrieves the test result for a specific Automated Reasoning policy test}
\usage{
bedrock_get_automated_reasoning_policy_test_result(
  policyArn,
  buildWorkflowId,
  testCaseId
)
}
\arguments{
\item{policyArn}{[required] The Amazon Resource Name (ARN) of the Automated Reasoning policy.}

\item{buildWorkflowId}{[required] The build workflow identifier. The build workflow must display a
\code{COMPLETED} status to get results.}

\item{testCaseId}{[required] The unique identifier of the test for which to retrieve results.}
}
\description{
Retrieves the test result for a specific Automated Reasoning policy test. Returns detailed validation findings and execution status.

See \url{https://www.paws-r-sdk.com/docs/bedrock_get_automated_reasoning_policy_test_result/} for full documentation.
}
\keyword{internal}
