% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_get_custom_model}
\alias{bedrock_get_custom_model}
\title{Get the properties associated with a Amazon Bedrock custom model that
you have created}
\usage{
bedrock_get_custom_model(modelIdentifier)
}
\arguments{
\item{modelIdentifier}{[required] Name or Amazon Resource Name (ARN) of the custom model.}
}
\description{
Get the properties associated with a Amazon Bedrock custom model that you have created. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/custom-models.html}{Custom models} in the \href{https://docs.aws.amazon.com/bedrock/latest/userguide/what-is-bedrock.html?trkcampaign=podcast_ep46_es}{Amazon Bedrock User Guide}.

See \url{https://www.paws-r-sdk.com/docs/bedrock_get_custom_model/} for full documentation.
}
\keyword{internal}
