% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_get_evaluation_job}
\alias{bedrock_get_evaluation_job}
\title{Gets information about an evaluation job, such as the status of the job}
\usage{
bedrock_get_evaluation_job(jobIdentifier)
}
\arguments{
\item{jobIdentifier}{[required] The Amazon Resource Name (ARN) of the evaluation job you want get
information on.}
}
\description{
Gets information about an evaluation job, such as the status of the job.

See \url{https://www.paws-r-sdk.com/docs/bedrock_get_evaluation_job/} for full documentation.
}
\keyword{internal}
