% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_get_foundation_model}
\alias{bedrock_get_foundation_model}
\title{Get details about a Bedrock foundation model}
\usage{
bedrock_get_foundation_model(modelIdentifier)
}
\arguments{
\item{modelIdentifier}{[required] The model identifier.}
}
\description{
Get details about a Bedrock foundation model.

See \url{https://www.paws-r-sdk.com/docs/bedrock_get_foundation_model/} for full documentation.
}
\keyword{internal}
