% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_get_foundation_model_availability}
\alias{bedrock_get_foundation_model_availability}
\title{Get information about the Foundation model availability}
\usage{
bedrock_get_foundation_model_availability(modelId)
}
\arguments{
\item{modelId}{[required] The model Id of the foundation model.}
}
\description{
Get information about the Foundation model availability.

See \url{https://www.paws-r-sdk.com/docs/bedrock_get_foundation_model_availability/} for full documentation.
}
\keyword{internal}
