% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_get_guardrail}
\alias{bedrock_get_guardrail}
\title{Gets details about a guardrail}
\usage{
bedrock_get_guardrail(guardrailIdentifier, guardrailVersion = NULL)
}
\arguments{
\item{guardrailIdentifier}{[required] The unique identifier of the guardrail for which to get details.}

\item{guardrailVersion}{The version of the guardrail for which to get details. If you don't
specify a version, the response returns details for the \code{DRAFT} version.}
}
\description{
Gets details about a guardrail. If you don't specify a version, the response returns details for the \code{DRAFT} version.

See \url{https://www.paws-r-sdk.com/docs/bedrock_get_guardrail/} for full documentation.
}
\keyword{internal}
