% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_get_imported_model}
\alias{bedrock_get_imported_model}
\title{Gets properties associated with a customized model you imported}
\usage{
bedrock_get_imported_model(modelIdentifier)
}
\arguments{
\item{modelIdentifier}{[required] Name or Amazon Resource Name (ARN) of the imported model.}
}
\description{
Gets properties associated with a customized model you imported.

See \url{https://www.paws-r-sdk.com/docs/bedrock_get_imported_model/} for full documentation.
}
\keyword{internal}
