% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_get_inference_profile}
\alias{bedrock_get_inference_profile}
\title{Gets information about an inference profile}
\usage{
bedrock_get_inference_profile(inferenceProfileIdentifier)
}
\arguments{
\item{inferenceProfileIdentifier}{[required] The unique identifier of the inference profile.}
}
\description{
Gets information about an inference profile. For more information, see the Amazon Bedrock User Guide.

See \url{https://www.paws-r-sdk.com/docs/bedrock_get_inference_profile/} for full documentation.
}
\keyword{internal}
