% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_get_marketplace_model_endpoint}
\alias{bedrock_get_marketplace_model_endpoint}
\title{Retrieves details about a specific endpoint for a model from Amazon
Bedrock Marketplace}
\usage{
bedrock_get_marketplace_model_endpoint(endpointArn)
}
\arguments{
\item{endpointArn}{[required] The Amazon Resource Name (ARN) of the endpoint you want to get
information about.}
}
\description{
Retrieves details about a specific endpoint for a model from Amazon Bedrock Marketplace.

See \url{https://www.paws-r-sdk.com/docs/bedrock_get_marketplace_model_endpoint/} for full documentation.
}
\keyword{internal}
