% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_get_model_invocation_job}
\alias{bedrock_get_model_invocation_job}
\title{Gets details about a batch inference job}
\usage{
bedrock_get_model_invocation_job(jobIdentifier)
}
\arguments{
\item{jobIdentifier}{[required] The Amazon Resource Name (ARN) of the batch inference job.}
}
\description{
Gets details about a batch inference job. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/batch-inference-manage.html#batch-inference-view}{View details about a batch inference job}

See \url{https://www.paws-r-sdk.com/docs/bedrock_get_model_invocation_job/} for full documentation.
}
\keyword{internal}
