% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_get_model_invocation_logging_configuration}
\alias{bedrock_get_model_invocation_logging_configuration}
\title{Get the current configuration values for model invocation logging}
\usage{
bedrock_get_model_invocation_logging_configuration()
}
\description{
Get the current configuration values for model invocation logging.

See \url{https://www.paws-r-sdk.com/docs/bedrock_get_model_invocation_logging_configuration/} for full documentation.
}
\keyword{internal}
