% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_get_prompt_router}
\alias{bedrock_get_prompt_router}
\title{Retrieves details about a prompt router}
\usage{
bedrock_get_prompt_router(promptRouterArn)
}
\arguments{
\item{promptRouterArn}{[required] The prompt router's ARN}
}
\description{
Retrieves details about a prompt router.

See \url{https://www.paws-r-sdk.com/docs/bedrock_get_prompt_router/} for full documentation.
}
\keyword{internal}
