% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_get_resource_policy}
\alias{bedrock_get_resource_policy}
\title{Gets the resource policy document for a Bedrock resource}
\usage{
bedrock_get_resource_policy(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the Bedrock resource to which this resource policy applies.}
}
\description{
Gets the resource policy document for a Bedrock resource

See \url{https://www.paws-r-sdk.com/docs/bedrock_get_resource_policy/} for full documentation.
}
\keyword{internal}
