% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_get_use_case_for_model_access}
\alias{bedrock_get_use_case_for_model_access}
\title{Get usecase for model access}
\usage{
bedrock_get_use_case_for_model_access()
}
\description{
Get usecase for model access.

See \url{https://www.paws-r-sdk.com/docs/bedrock_get_use_case_for_model_access/} for full documentation.
}
\keyword{internal}
