% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_list_advanced_prompt_optimization_jobs}
\alias{bedrock_list_advanced_prompt_optimization_jobs}
\title{Lists the advanced prompt optimization jobs in your account}
\usage{
bedrock_list_advanced_prompt_optimization_jobs(
  maxResults = NULL,
  nextToken = NULL,
  sortBy = NULL,
  sortOrder = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of results to return in the response.}

\item{nextToken}{If the total number of results is greater than the \code{maxResults} value provided in the request, use this token in a subsequent request to get the next set of results.}

\item{sortBy}{The field to sort the results by.}

\item{sortOrder}{The sort order for the results.}
}
\description{
Lists the advanced prompt optimization jobs in your account.

See \url{https://www.paws-r-sdk.com/docs/bedrock_list_advanced_prompt_optimization_jobs/} for full documentation.
}
\keyword{internal}
