% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_list_automated_reasoning_policy_test_cases}
\alias{bedrock_list_automated_reasoning_policy_test_cases}
\title{Lists tests for an Automated Reasoning policy}
\usage{
bedrock_list_automated_reasoning_policy_test_cases(
  policyArn,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{policyArn}{[required] The Amazon Resource Name (ARN) of the Automated Reasoning policy for
which to list tests.}

\item{nextToken}{The pagination token from a previous request to retrieve the next page
of results.}

\item{maxResults}{The maximum number of tests to return in a single call.}
}
\description{
Lists tests for an Automated Reasoning policy. We recommend using pagination to ensure that the operation returns quickly and successfully.

See \url{https://www.paws-r-sdk.com/docs/bedrock_list_automated_reasoning_policy_test_cases/} for full documentation.
}
\keyword{internal}
