% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_list_automated_reasoning_policy_test_results}
\alias{bedrock_list_automated_reasoning_policy_test_results}
\title{Lists test results for an Automated Reasoning policy, showing how the
policy performed against various test scenarios and validation checks}
\usage{
bedrock_list_automated_reasoning_policy_test_results(
  policyArn,
  buildWorkflowId,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{policyArn}{[required] The Amazon Resource Name (ARN) of the Automated Reasoning policy whose
test results you want to list.}

\item{buildWorkflowId}{[required] The unique identifier of the build workflow whose test results you want
to list.}

\item{nextToken}{A pagination token from a previous request to continue listing test
results from where the previous request left off.}

\item{maxResults}{The maximum number of test results to return in a single response. Valid
range is 1-100.}
}
\description{
Lists test results for an Automated Reasoning policy, showing how the policy performed against various test scenarios and validation checks.

See \url{https://www.paws-r-sdk.com/docs/bedrock_list_automated_reasoning_policy_test_results/} for full documentation.
}
\keyword{internal}
