% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_list_custom_model_deployments}
\alias{bedrock_list_custom_model_deployments}
\title{Lists custom model deployments in your account}
\usage{
bedrock_list_custom_model_deployments(
  createdBefore = NULL,
  createdAfter = NULL,
  nameContains = NULL,
  maxResults = NULL,
  nextToken = NULL,
  sortBy = NULL,
  sortOrder = NULL,
  statusEquals = NULL,
  modelArnEquals = NULL
)
}
\arguments{
\item{createdBefore}{Filters deployments created before the specified date and time.}

\item{createdAfter}{Filters deployments created after the specified date and time.}

\item{nameContains}{Filters deployments whose names contain the specified string.}

\item{maxResults}{The maximum number of results to return in a single call.}

\item{nextToken}{The token for the next set of results. Use this token to retrieve
additional results when the response is truncated.}

\item{sortBy}{The field to sort the results by. The only supported value is
\code{CreationTime}.}

\item{sortOrder}{The sort order for the results. Valid values are \code{Ascending} and
\code{Descending}. Default is \code{Descending}.}

\item{statusEquals}{Filters deployments by status. Valid values are \code{CREATING}, \code{ACTIVE},
and \code{FAILED}.}

\item{modelArnEquals}{Filters deployments by the Amazon Resource Name (ARN) of the associated
custom model.}
}
\description{
Lists custom model deployments in your account. You can filter the results by creation time, name, status, and associated model. Use this operation to manage and monitor your custom model deployments.

See \url{https://www.paws-r-sdk.com/docs/bedrock_list_custom_model_deployments/} for full documentation.
}
\keyword{internal}
