% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_list_evaluation_jobs}
\alias{bedrock_list_evaluation_jobs}
\title{Lists model evaluation jobs}
\usage{
bedrock_list_evaluation_jobs(
  creationTimeAfter = NULL,
  creationTimeBefore = NULL,
  statusEquals = NULL,
  nameContains = NULL,
  maxResults = NULL,
  nextToken = NULL,
  sortBy = NULL,
  sortOrder = NULL
)
}
\arguments{
\item{creationTimeAfter}{A filter that includes model evaluation jobs created after the time
specified.}

\item{creationTimeBefore}{A filter that includes model evaluation jobs created prior to the time
specified.}

\item{statusEquals}{Only return jobs where the status condition is met.}

\item{nameContains}{Query parameter string for model evaluation job names.}

\item{maxResults}{The maximum number of results to return.}

\item{nextToken}{Continuation token from the previous response, for Amazon Bedrock to
list the next set of results.}

\item{sortBy}{Allows you to sort model evaluation jobs by when they were created.}

\item{sortOrder}{How you want the order of jobs sorted.}
}
\description{
Lists model evaluation jobs.

See \url{https://www.paws-r-sdk.com/docs/bedrock_list_evaluation_jobs/} for full documentation.
}
\keyword{internal}
