% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_list_foundation_model_agreement_offers}
\alias{bedrock_list_foundation_model_agreement_offers}
\title{Get the offers associated with the specified model}
\usage{
bedrock_list_foundation_model_agreement_offers(modelId, offerType = NULL)
}
\arguments{
\item{modelId}{[required] Model Id of the foundation model.}

\item{offerType}{Type of offer associated with the model.}
}
\description{
Get the offers associated with the specified model.

See \url{https://www.paws-r-sdk.com/docs/bedrock_list_foundation_model_agreement_offers/} for full documentation.
}
\keyword{internal}
