% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_list_marketplace_model_endpoints}
\alias{bedrock_list_marketplace_model_endpoints}
\title{Lists the endpoints for models from Amazon Bedrock Marketplace in your
Amazon Web Services account}
\usage{
bedrock_list_marketplace_model_endpoints(
  maxResults = NULL,
  nextToken = NULL,
  modelSourceEquals = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of results to return in a single call. If more
results are available, the operation returns a \code{NextToken} value.}

\item{nextToken}{The token for the next set of results. You receive this token from a
previous
\code{\link[=bedrock_list_marketplace_model_endpoints]{list_marketplace_model_endpoints}}
call.}

\item{modelSourceEquals}{If specified, only endpoints for the given model source identifier are
returned.}
}
\description{
Lists the endpoints for models from Amazon Bedrock Marketplace in your Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/bedrock_list_marketplace_model_endpoints/} for full documentation.
}
\keyword{internal}
