% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_list_model_customization_jobs}
\alias{bedrock_list_model_customization_jobs}
\title{Returns a list of model customization jobs that you have submitted}
\usage{
bedrock_list_model_customization_jobs(
  creationTimeAfter = NULL,
  creationTimeBefore = NULL,
  statusEquals = NULL,
  nameContains = NULL,
  maxResults = NULL,
  nextToken = NULL,
  sortBy = NULL,
  sortOrder = NULL
)
}
\arguments{
\item{creationTimeAfter}{Return customization jobs created after the specified time.}

\item{creationTimeBefore}{Return customization jobs created before the specified time.}

\item{statusEquals}{Return customization jobs with the specified status.}

\item{nameContains}{Return customization jobs only if the job name contains these
characters.}

\item{maxResults}{Maximum number of results to return in the response.}

\item{nextToken}{Continuation token from the previous response, for Amazon Bedrock to
list the next set of results.}

\item{sortBy}{The field to sort by in the returned list of jobs.}

\item{sortOrder}{The sort order of the results.}
}
\description{
Returns a list of model customization jobs that you have submitted. You can filter the jobs to return based on one or more criteria.

See \url{https://www.paws-r-sdk.com/docs/bedrock_list_model_customization_jobs/} for full documentation.
}
\keyword{internal}
