% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_list_tags_for_resource}
\alias{bedrock_list_tags_for_resource}
\title{List the tags associated with the specified resource}
\usage{
bedrock_list_tags_for_resource(resourceARN)
}
\arguments{
\item{resourceARN}{[required] The ARN of the resource.}
}
\description{
List the tags associated with the specified resource.

See \url{https://www.paws-r-sdk.com/docs/bedrock_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
