% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_put_account_data_retention}
\alias{bedrock_put_account_data_retention}
\title{Sets the account-wide data retention mode for Amazon Bedrock}
\usage{
bedrock_put_account_data_retention(mode)
}
\arguments{
\item{mode}{[required] The data retention mode to set for the account.}
}
\description{
Sets the account-wide data retention mode for Amazon Bedrock.

See \url{https://www.paws-r-sdk.com/docs/bedrock_put_account_data_retention/} for full documentation.
}
\keyword{internal}
