% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_put_enforced_guardrail_configuration}
\alias{bedrock_put_enforced_guardrail_configuration}
\title{Sets the account-level enforced guardrail configuration}
\usage{
bedrock_put_enforced_guardrail_configuration(
  configId = NULL,
  guardrailInferenceConfig
)
}
\arguments{
\item{configId}{Unique ID for the account enforced configuration.}

\item{guardrailInferenceConfig}{[required] Account-level enforced guardrail input configuration.}
}
\description{
Sets the account-level enforced guardrail configuration.

See \url{https://www.paws-r-sdk.com/docs/bedrock_put_enforced_guardrail_configuration/} for full documentation.
}
\keyword{internal}
