% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_put_model_invocation_logging_configuration}
\alias{bedrock_put_model_invocation_logging_configuration}
\title{Set the configuration values for model invocation logging}
\usage{
bedrock_put_model_invocation_logging_configuration(loggingConfig)
}
\arguments{
\item{loggingConfig}{[required] The logging configuration values to set.}
}
\description{
Set the configuration values for model invocation logging.

See \url{https://www.paws-r-sdk.com/docs/bedrock_put_model_invocation_logging_configuration/} for full documentation.
}
\keyword{internal}
