% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_put_resource_policy}
\alias{bedrock_put_resource_policy}
\title{Adds a resource policy for a Bedrock resource}
\usage{
bedrock_put_resource_policy(resourceArn, resourcePolicy)
}
\arguments{
\item{resourceArn}{[required] The ARN of the Bedrock resource to which this resource policy applies.}

\item{resourcePolicy}{[required] The JSON string representing the Bedrock resource policy.}
}
\description{
Adds a resource policy for a Bedrock resource.

See \url{https://www.paws-r-sdk.com/docs/bedrock_put_resource_policy/} for full documentation.
}
\keyword{internal}
