% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_put_use_case_for_model_access}
\alias{bedrock_put_use_case_for_model_access}
\title{Put usecase for model access}
\usage{
bedrock_put_use_case_for_model_access(formData)
}
\arguments{
\item{formData}{[required] Put customer profile Request.}
}
\description{
Put usecase for model access.

See \url{https://www.paws-r-sdk.com/docs/bedrock_put_use_case_for_model_access/} for full documentation.
}
\keyword{internal}
