% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_register_marketplace_model_endpoint}
\alias{bedrock_register_marketplace_model_endpoint}
\title{Registers an existing Amazon SageMaker endpoint with Amazon Bedrock
Marketplace, allowing it to be used with Amazon Bedrock APIs}
\usage{
bedrock_register_marketplace_model_endpoint(
  endpointIdentifier,
  modelSourceIdentifier
)
}
\arguments{
\item{endpointIdentifier}{[required] The ARN of the Amazon SageMaker endpoint you want to register with
Amazon Bedrock Marketplace.}

\item{modelSourceIdentifier}{[required] The ARN of the model from Amazon Bedrock Marketplace that is deployed on
the endpoint.}
}
\description{
Registers an existing Amazon SageMaker endpoint with Amazon Bedrock Marketplace, allowing it to be used with Amazon Bedrock APIs.

See \url{https://www.paws-r-sdk.com/docs/bedrock_register_marketplace_model_endpoint/} for full documentation.
}
\keyword{internal}
