% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_stop_evaluation_job}
\alias{bedrock_stop_evaluation_job}
\title{Stops an in progress model evaluation job}
\usage{
bedrock_stop_evaluation_job(jobIdentifier)
}
\arguments{
\item{jobIdentifier}{[required] The ARN of the model evaluation job you want to stop.}
}
\description{
Stops an in progress model evaluation job.

See \url{https://www.paws-r-sdk.com/docs/bedrock_stop_evaluation_job/} for full documentation.
}
\keyword{internal}
