% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_stop_model_invocation_job}
\alias{bedrock_stop_model_invocation_job}
\title{Stops a batch inference job}
\usage{
bedrock_stop_model_invocation_job(jobIdentifier)
}
\arguments{
\item{jobIdentifier}{[required] The Amazon Resource Name (ARN) of the batch inference job to stop.}
}
\description{
Stops a batch inference job. You're only charged for tokens that were already processed. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/batch-inference-manage.html#batch-inference-stop}{Stop a batch inference job}.

See \url{https://www.paws-r-sdk.com/docs/bedrock_stop_model_invocation_job/} for full documentation.
}
\keyword{internal}
