% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_update_custom_model_deployment}
\alias{bedrock_update_custom_model_deployment}
\title{Updates a custom model deployment with a new custom model}
\usage{
bedrock_update_custom_model_deployment(
  modelArn,
  customModelDeploymentIdentifier
)
}
\arguments{
\item{modelArn}{[required] ARN of the new custom model to deploy. This replaces the currently
deployed model.}

\item{customModelDeploymentIdentifier}{[required] Identifier of the custom model deployment to update with the new custom
model.}
}
\description{
Updates a custom model deployment with a new custom model. This allows you to deploy updated models without creating new deployment endpoints.

See \url{https://www.paws-r-sdk.com/docs/bedrock_update_custom_model_deployment/} for full documentation.
}
\keyword{internal}
