% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_update_marketplace_model_endpoint}
\alias{bedrock_update_marketplace_model_endpoint}
\title{Updates the configuration of an existing endpoint for a model from
Amazon Bedrock Marketplace}
\usage{
bedrock_update_marketplace_model_endpoint(
  endpointArn,
  endpointConfig,
  clientRequestToken = NULL
)
}
\arguments{
\item{endpointArn}{[required] The Amazon Resource Name (ARN) of the endpoint you want to update.}

\item{endpointConfig}{[required] The new configuration for the endpoint, including the number and type of
instances to use.}

\item{clientRequestToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. This token is listed as not required because
Amazon Web Services SDKs automatically generate it for you and set this
parameter. If you're not using the Amazon Web Services SDK or the CLI,
you must provide this token or the action will fail.}
}
\description{
Updates the configuration of an existing endpoint for a model from Amazon Bedrock Marketplace.

See \url{https://www.paws-r-sdk.com/docs/bedrock_update_marketplace_model_endpoint/} for full documentation.
}
\keyword{internal}
