% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_delete_agent_action_group}
\alias{bedrockagent_delete_agent_action_group}
\title{Deletes an action group in an agent}
\usage{
bedrockagent_delete_agent_action_group(
  agentId,
  agentVersion,
  actionGroupId,
  skipResourceInUseCheck = NULL
)
}
\arguments{
\item{agentId}{[required] The unique identifier of the agent that the action group belongs to.}

\item{agentVersion}{[required] The version of the agent that the action group belongs to.}

\item{actionGroupId}{[required] The unique identifier of the action group to delete.}

\item{skipResourceInUseCheck}{By default, this value is \code{false} and deletion is stopped if the
resource is in use. If you set it to \code{true}, the resource will be
deleted even if the resource is in use.}
}
\description{
Deletes an action group in an agent.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_delete_agent_action_group/} for full documentation.
}
\keyword{internal}
