% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_delete_agent_alias}
\alias{bedrockagent_delete_agent_alias}
\title{Deletes an alias of an agent}
\usage{
bedrockagent_delete_agent_alias(agentId, agentAliasId)
}
\arguments{
\item{agentId}{[required] The unique identifier of the agent that the alias belongs to.}

\item{agentAliasId}{[required] The unique identifier of the alias to delete.}
}
\description{
Deletes an alias of an agent.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_delete_agent_alias/} for full documentation.
}
\keyword{internal}
