% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_delete_agent_version}
\alias{bedrockagent_delete_agent_version}
\title{Deletes a version of an agent}
\usage{
bedrockagent_delete_agent_version(
  agentId,
  agentVersion,
  skipResourceInUseCheck = NULL
)
}
\arguments{
\item{agentId}{[required] The unique identifier of the agent that the version belongs to.}

\item{agentVersion}{[required] The version of the agent to delete.}

\item{skipResourceInUseCheck}{By default, this value is \code{false} and deletion is stopped if the
resource is in use. If you set it to \code{true}, the resource will be
deleted even if the resource is in use.}
}
\description{
Deletes a version of an agent.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_delete_agent_version/} for full documentation.
}
\keyword{internal}
