% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_delete_data_source}
\alias{bedrockagent_delete_data_source}
\title{Deletes a data source from a knowledge base}
\usage{
bedrockagent_delete_data_source(knowledgeBaseId, dataSourceId)
}
\arguments{
\item{knowledgeBaseId}{[required] The unique identifier of the knowledge base from which to delete the
data source.}

\item{dataSourceId}{[required] The unique identifier of the data source to delete.}
}
\description{
Deletes a data source from a knowledge base.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_delete_data_source/} for full documentation.
}
\keyword{internal}
