% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_delete_flow_alias}
\alias{bedrockagent_delete_flow_alias}
\title{Deletes an alias of a flow}
\usage{
bedrockagent_delete_flow_alias(flowIdentifier, aliasIdentifier)
}
\arguments{
\item{flowIdentifier}{[required] The unique identifier of the flow that the alias belongs to.}

\item{aliasIdentifier}{[required] The unique identifier of the alias to be deleted.}
}
\description{
Deletes an alias of a flow.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_delete_flow_alias/} for full documentation.
}
\keyword{internal}
