% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_delete_knowledge_base}
\alias{bedrockagent_delete_knowledge_base}
\title{Deletes a knowledge base}
\usage{
bedrockagent_delete_knowledge_base(knowledgeBaseId)
}
\arguments{
\item{knowledgeBaseId}{[required] The unique identifier of the knowledge base to delete.}
}
\description{
Deletes a knowledge base. Before deleting a knowledge base, you should disassociate the knowledge base from any agents that it is associated with by making a \code{\link[=bedrockagent_disassociate_agent_knowledge_base]{disassociate_agent_knowledge_base}} request.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_delete_knowledge_base/} for full documentation.
}
\keyword{internal}
