% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_delete_resource_policy}
\alias{bedrockagent_delete_resource_policy}
\title{Removes the resource policy associated with a knowledge base}
\usage{
bedrockagent_delete_resource_policy(resourceArn, expectedRevisionId = NULL)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the knowledge base to remove the resource policy from.}

\item{expectedRevisionId}{The expected revision identifier of the resource policy. Use this to prevent conflicts when multiple users update the same policy concurrently.}
}
\description{
Removes the resource policy associated with a knowledge base. After deletion, other AWS accounts can no longer access the knowledge base using cross-account permissions.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_delete_resource_policy/} for full documentation.
}
\keyword{internal}
