% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_disassociate_agent_knowledge_base}
\alias{bedrockagent_disassociate_agent_knowledge_base}
\title{Disassociates a knowledge base from an agent}
\usage{
bedrockagent_disassociate_agent_knowledge_base(
  agentId,
  agentVersion,
  knowledgeBaseId
)
}
\arguments{
\item{agentId}{[required] The unique identifier of the agent from which to disassociate the
knowledge base.}

\item{agentVersion}{[required] The version of the agent from which to disassociate the knowledge base.}

\item{knowledgeBaseId}{[required] The unique identifier of the knowledge base to disassociate.}
}
\description{
Disassociates a knowledge base from an agent.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_disassociate_agent_knowledge_base/} for full documentation.
}
\keyword{internal}
