% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_get_agent}
\alias{bedrockagent_get_agent}
\title{Gets information about an agent}
\usage{
bedrockagent_get_agent(agentId)
}
\arguments{
\item{agentId}{[required] The unique identifier of the agent.}
}
\description{
Gets information about an agent.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_get_agent/} for full documentation.
}
\keyword{internal}
