% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_get_agent_action_group}
\alias{bedrockagent_get_agent_action_group}
\title{Gets information about an action group for an agent}
\usage{
bedrockagent_get_agent_action_group(actionGroupId, agentId, agentVersion)
}
\arguments{
\item{actionGroupId}{[required] The unique identifier of the action group for which to get information.}

\item{agentId}{[required] The unique identifier of the agent that the action group belongs to.}

\item{agentVersion}{[required] The version of the agent that the action group belongs to.}
}
\description{
Gets information about an action group for an agent.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_get_agent_action_group/} for full documentation.
}
\keyword{internal}
