% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_get_agent_alias}
\alias{bedrockagent_get_agent_alias}
\title{Gets information about an alias of an agent}
\usage{
bedrockagent_get_agent_alias(agentId, agentAliasId)
}
\arguments{
\item{agentId}{[required] The unique identifier of the agent to which the alias to get information
belongs.}

\item{agentAliasId}{[required] The unique identifier of the alias for which to get information.}
}
\description{
Gets information about an alias of an agent.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_get_agent_alias/} for full documentation.
}
\keyword{internal}
