% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_get_agent_collaborator}
\alias{bedrockagent_get_agent_collaborator}
\title{Retrieves information about an agent's collaborator}
\usage{
bedrockagent_get_agent_collaborator(agentId, agentVersion, collaboratorId)
}
\arguments{
\item{agentId}{[required] The agent's ID.}

\item{agentVersion}{[required] The agent's version.}

\item{collaboratorId}{[required] The collaborator's ID.}
}
\description{
Retrieves information about an agent's collaborator.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_get_agent_collaborator/} for full documentation.
}
\keyword{internal}
