% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_get_data_source}
\alias{bedrockagent_get_data_source}
\title{Gets information about a data source}
\usage{
bedrockagent_get_data_source(knowledgeBaseId, dataSourceId)
}
\arguments{
\item{knowledgeBaseId}{[required] The unique identifier of the knowledge base for the data source.}

\item{dataSourceId}{[required] The unique identifier of the data source.}
}
\description{
Gets information about a data source.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_get_data_source/} for full documentation.
}
\keyword{internal}
