% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_get_flow_alias}
\alias{bedrockagent_get_flow_alias}
\title{Retrieves information about a flow}
\usage{
bedrockagent_get_flow_alias(flowIdentifier, aliasIdentifier)
}
\arguments{
\item{flowIdentifier}{[required] The unique identifier of the flow that the alias belongs to.}

\item{aliasIdentifier}{[required] The unique identifier of the alias for which to retrieve information.}
}
\description{
Retrieves information about a flow. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/flows-deploy.html}{Deploy a flow in Amazon Bedrock} in the Amazon Bedrock User Guide.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_get_flow_alias/} for full documentation.
}
\keyword{internal}
