% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_get_knowledge_base}
\alias{bedrockagent_get_knowledge_base}
\title{Gets information about a knowledge base}
\usage{
bedrockagent_get_knowledge_base(knowledgeBaseId)
}
\arguments{
\item{knowledgeBaseId}{[required] The unique identifier of the knowledge base you want to get information
on.}
}
\description{
Gets information about a knowledge base.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_get_knowledge_base/} for full documentation.
}
\keyword{internal}
