% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_get_resource_policy}
\alias{bedrockagent_get_resource_policy}
\title{Retrieves the resource policy associated with a knowledge base}
\usage{
bedrockagent_get_resource_policy(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the knowledge base to retrieve the resource policy for.}
}
\description{
Retrieves the resource policy associated with a knowledge base.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_get_resource_policy/} for full documentation.
}
\keyword{internal}
