% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_get_vpc_configuration}
\alias{bedrockagent_get_vpc_configuration}
\title{Returns the details and current status of a single VPC configuration}
\usage{
bedrockagent_get_vpc_configuration(knowledgeBaseId, vpcConfigurationId)
}
\arguments{
\item{knowledgeBaseId}{[required] The unique identifier of the knowledge base that owns the VPC configuration.}

\item{vpcConfigurationId}{[required] The unique identifier of the VPC configuration to retrieve.}
}
\description{
Returns the details and current status of a single VPC configuration. Use this operation to poll for the outcome of an asynchronous create or delete.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_get_vpc_configuration/} for full documentation.
}
\keyword{internal}
