% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_list_agent_aliases}
\alias{bedrockagent_list_agent_aliases}
\title{Lists the aliases of an agent and information about each one}
\usage{
bedrockagent_list_agent_aliases(agentId, maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{agentId}{[required] The unique identifier of the agent.}

\item{maxResults}{The maximum number of results to return in the response. If the total
number of results is greater than this value, use the token returned in
the response in the \code{nextToken} field when making another request to
return the next batch of results.}

\item{nextToken}{If the total number of results is greater than the \code{maxResults} value
provided in the request, enter the token returned in the \code{nextToken}
field in the response in this field to return the next batch of results.}
}
\description{
Lists the aliases of an agent and information about each one.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_list_agent_aliases/} for full documentation.
}
\keyword{internal}
