% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_list_agent_collaborators}
\alias{bedrockagent_list_agent_collaborators}
\title{Retrieve a list of an agent's collaborators}
\usage{
bedrockagent_list_agent_collaborators(
  agentId,
  agentVersion,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{agentId}{[required] The agent's ID.}

\item{agentVersion}{[required] The agent's version.}

\item{maxResults}{The maximum number of agent collaborators to return in one page of
results.}

\item{nextToken}{Specify the pagination token from a previous request to retrieve the
next page of results.}
}
\description{
Retrieve a list of an agent's collaborators.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_list_agent_collaborators/} for full documentation.
}
\keyword{internal}
