% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_list_flow_aliases}
\alias{bedrockagent_list_flow_aliases}
\title{Returns a list of aliases for a flow}
\usage{
bedrockagent_list_flow_aliases(
  flowIdentifier,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{flowIdentifier}{[required] The unique identifier of the flow for which aliases are being returned.}

\item{maxResults}{The maximum number of results to return in the response. If the total
number of results is greater than this value, use the token returned in
the response in the \code{nextToken} field when making another request to
return the next batch of results.}

\item{nextToken}{If the total number of results is greater than the \code{maxResults} value
provided in the request, enter the token returned in the \code{nextToken}
field in the response in this field to return the next batch of results.}
}
\description{
Returns a list of aliases for a flow.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_list_flow_aliases/} for full documentation.
}
\keyword{internal}
