% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_list_flows}
\alias{bedrockagent_list_flows}
\title{Returns a list of flows and information about each flow}
\usage{
bedrockagent_list_flows(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of results to return in the response. If the total
number of results is greater than this value, use the token returned in
the response in the \code{nextToken} field when making another request to
return the next batch of results.}

\item{nextToken}{If the total number of results is greater than the \code{maxResults} value
provided in the request, enter the token returned in the \code{nextToken}
field in the response in this field to return the next batch of results.}
}
\description{
Returns a list of flows and information about each flow. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/}{Manage a flow in Amazon Bedrock} in the Amazon Bedrock User Guide.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_list_flows/} for full documentation.
}
\keyword{internal}
