% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_list_vpc_configurations}
\alias{bedrockagent_list_vpc_configurations}
\title{Returns a paginated list of the VPC configurations for a knowledge base}
\usage{
bedrockagent_list_vpc_configurations(
  knowledgeBaseId,
  statusFilter = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{knowledgeBaseId}{[required] The unique identifier of the knowledge base whose VPC configurations you want to list.}

\item{statusFilter}{The status to filter the results by. Only VPC configurations with the specified status are returned.}

\item{maxResults}{The maximum number of results to return in the response. If more results are available, the response returns a \code{nextToken}.}

\item{nextToken}{A pagination token to retrieve the next page of results, returned in a previous response when more results are available.}
}
\description{
Returns a paginated list of the VPC configurations for a knowledge base. You can optionally filter by status. Use the \code{nextToken} parameter to retrieve additional results.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_list_vpc_configurations/} for full documentation.
}
\keyword{internal}
