% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_prepare_agent}
\alias{bedrockagent_prepare_agent}
\title{Creates a DRAFT version of the agent that can be used for internal
testing}
\usage{
bedrockagent_prepare_agent(agentId)
}
\arguments{
\item{agentId}{[required] The unique identifier of the agent for which to create a \code{DRAFT}
version.}
}
\description{
Creates a \code{DRAFT} version of the agent that can be used for internal testing.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_prepare_agent/} for full documentation.
}
\keyword{internal}
