% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_prepare_flow}
\alias{bedrockagent_prepare_flow}
\title{Prepares the DRAFT version of a flow so that it can be invoked}
\usage{
bedrockagent_prepare_flow(flowIdentifier)
}
\arguments{
\item{flowIdentifier}{[required] The unique identifier of the flow.}
}
\description{
Prepares the \code{DRAFT} version of a flow so that it can be invoked. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/flows-test.html}{Test a flow in Amazon Bedrock} in the Amazon Bedrock User Guide.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_prepare_flow/} for full documentation.
}
\keyword{internal}
