% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_put_resource_policy}
\alias{bedrockagent_put_resource_policy}
\title{Associates a resource policy with a knowledge base}
\usage{
bedrockagent_put_resource_policy(
  resourceArn,
  policy,
  expectedRevisionId = NULL
)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the knowledge base to attach the resource policy to.}

\item{policy}{[required] The JSON-formatted resource policy to associate with the knowledge base.}

\item{expectedRevisionId}{The expected revision identifier of the resource policy. Use this to prevent conflicts when multiple users update the same policy concurrently. Specify the \code{revisionId} from the most recent \code{\link[=bedrockagent_get_resource_policy]{get_resource_policy}} or \code{\link[=bedrockagent_put_resource_policy]{put_resource_policy}} response.}
}
\description{
Associates a resource policy with a knowledge base. A resource policy allows other AWS accounts to access the knowledge base. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/kb-managed-cross-account.html}{Cross-account access for knowledge bases}.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_put_resource_policy/} for full documentation.
}
\keyword{internal}
