% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_tag_resource}
\alias{bedrockagent_tag_resource}
\title{Associate tags with a resource}
\usage{
bedrockagent_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource to tag.}

\item{tags}{[required] An object containing key-value pairs that define the tags to attach to the resource.}
}
\description{
Associate tags with a resource. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/what-is-bedrock.html?trkcampaign=podcast_ep46_es}{Tagging resources} in the Amazon Bedrock User Guide.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_tag_resource/} for full documentation.
}
\keyword{internal}
