% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_untag_resource}
\alias{bedrockagent_untag_resource}
\title{Remove tags from a resource}
\usage{
bedrockagent_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource from which to remove
tags.}

\item{tagKeys}{[required] A list of keys of the tags to remove from the resource.}
}
\description{
Remove tags from a resource.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_untag_resource/} for full documentation.
}
\keyword{internal}
