% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_update_agent_action_group}
\alias{bedrockagent_update_agent_action_group}
\title{Updates the configuration for an action group for an agent}
\usage{
bedrockagent_update_agent_action_group(
  agentId,
  agentVersion,
  actionGroupId,
  actionGroupName,
  description = NULL,
  parentActionGroupSignature = NULL,
  parentActionGroupSignatureParams = NULL,
  actionGroupExecutor = NULL,
  actionGroupState = NULL,
  apiSchema = NULL,
  functionSchema = NULL
)
}
\arguments{
\item{agentId}{[required] The unique identifier of the agent for which to update the action group.}

\item{agentVersion}{[required] The unique identifier of the agent version for which to update the
action group.}

\item{actionGroupId}{[required] The unique identifier of the action group.}

\item{actionGroupName}{[required] Specifies a new name for the action group.}

\item{description}{Specifies a new name for the action group.}

\item{parentActionGroupSignature}{Update the built-in or computer use action for this action group. If you
specify a value, you must leave the \code{description}, \code{apiSchema}, and
\code{actionGroupExecutor} fields empty for this action group.
\itemize{
\item To allow your agent to request the user for additional information
when trying to complete a task, set this field to
\code{AMAZON.UserInput}.
\item To allow your agent to generate, run, and troubleshoot code when
trying to complete a task, set this field to
\code{AMAZON.CodeInterpreter}.
\item To allow your agent to use an Anthropic computer use tool, specify
one of the following values.

Computer use is a new Anthropic Claude model capability (in beta)
available with Anthropic Claude 3.7 Sonnet and Claude 3.5 Sonnet v2
only. When operating computer use functionality, we recommend taking
additional security precautions, such as executing computer actions
in virtual environments with restricted data access and limited
internet connectivity. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/agents-computer-use.html}{Configure an Amazon Bedrock Agent to complete tasks with computer use tools}.
\itemize{
\item \code{ANTHROPIC.Computer} - Gives the agent permission to use the
mouse and keyboard and take screenshots.
\item \code{ANTHROPIC.TextEditor} - Gives the agent permission to view,
create and edit files.
\item \code{ANTHROPIC.Bash} - Gives the agent permission to run commands in
a bash shell.
}
}

During orchestration, if your agent determines that it needs to invoke
an API in an action group, but doesn't have enough information to
complete the API request, it will invoke this action group instead and
return an
\href{https://docs.aws.amazon.com/bedrock/latest/APIReference/API_agent-runtime_Observation.html}{Observation}
reprompting the user for more information.}

\item{parentActionGroupSignatureParams}{The configuration settings for a computer use action.

Computer use is a new Anthropic Claude model capability (in beta)
available with Claude 3.7 Sonnet and Claude 3.5 Sonnet v2 only. For more
information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/agents-computer-use.html}{Configure an Amazon Bedrock Agent to complete tasks with computer use tools}.}

\item{actionGroupExecutor}{The Amazon Resource Name (ARN) of the Lambda function containing the
business logic that is carried out upon invoking the action.}

\item{actionGroupState}{Specifies whether the action group is available for the agent to invoke
or not when sending an
\href{https://docs.aws.amazon.com/bedrock/latest/APIReference/API_agent-runtime_InvokeAgent.html}{InvokeAgent}
request.}

\item{apiSchema}{Contains either details about the S3 object containing the OpenAPI
schema for the action group or the JSON or YAML-formatted payload
defining the schema. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/agents-api-schema.html}{Action group OpenAPI schemas}.}

\item{functionSchema}{Contains details about the function schema for the action group or the
JSON or YAML-formatted payload defining the schema.}
}
\description{
Updates the configuration for an action group for an agent.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_update_agent_action_group/} for full documentation.
}
\keyword{internal}
