% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_update_agent_collaborator}
\alias{bedrockagent_update_agent_collaborator}
\title{Updates an agent's collaborator}
\usage{
bedrockagent_update_agent_collaborator(
  agentDescriptor,
  agentId,
  agentVersion,
  collaborationInstruction,
  collaboratorId,
  collaboratorName,
  relayConversationHistory = NULL
)
}
\arguments{
\item{agentDescriptor}{[required] An agent descriptor for the agent collaborator.}

\item{agentId}{[required] The agent's ID.}

\item{agentVersion}{[required] The agent's version.}

\item{collaborationInstruction}{[required] Instruction for the collaborator.}

\item{collaboratorId}{[required] The collaborator's ID.}

\item{collaboratorName}{[required] The collaborator's name.}

\item{relayConversationHistory}{A relay conversation history for the collaborator.}
}
\description{
Updates an agent's collaborator.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_update_agent_collaborator/} for full documentation.
}
\keyword{internal}
