% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_update_data_source}
\alias{bedrockagent_update_data_source}
\title{Updates the configurations for a data source connector}
\usage{
bedrockagent_update_data_source(
  knowledgeBaseId,
  dataSourceId,
  name,
  description = NULL,
  dataSourceConfiguration,
  dataDeletionPolicy = NULL,
  serverSideEncryptionConfiguration = NULL,
  vectorIngestionConfiguration = NULL
)
}
\arguments{
\item{knowledgeBaseId}{[required] The unique identifier of the knowledge base for the data source.}

\item{dataSourceId}{[required] The unique identifier of the data source.}

\item{name}{[required] Specifies a new name for the data source.}

\item{description}{Specifies a new description for the data source.}

\item{dataSourceConfiguration}{[required] The connection configuration for the data source that you want to
update.}

\item{dataDeletionPolicy}{The data deletion policy for the data source that you want to update.}

\item{serverSideEncryptionConfiguration}{Contains details about server-side encryption of the data source.}

\item{vectorIngestionConfiguration}{Contains details about how to ingest the documents in the data source.}
}
\description{
Updates the configurations for a data source connector.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_update_data_source/} for full documentation.
}
\keyword{internal}
