% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_batch_create_memory_records}
\alias{bedrockagentcore_batch_create_memory_records}
\title{Creates multiple memory records in a single batch operation for the
specified memory with custom content}
\usage{
bedrockagentcore_batch_create_memory_records(
  memoryId,
  records,
  clientToken = NULL
)
}
\arguments{
\item{memoryId}{[required] The unique ID of the memory resource where records will be created.}

\item{records}{[required] A list of memory record creation inputs to be processed in the batch
operation.}

\item{clientToken}{A unique, case-sensitive identifier to ensure idempotent processing of
the batch request.}
}
\description{
Creates multiple memory records in a single batch operation for the specified memory with custom content.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_batch_create_memory_records/} for full documentation.
}
\keyword{internal}
