% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_batch_delete_memory_records}
\alias{bedrockagentcore_batch_delete_memory_records}
\title{Deletes multiple memory records in a single batch operation from the
specified memory}
\usage{
bedrockagentcore_batch_delete_memory_records(memoryId, records)
}
\arguments{
\item{memoryId}{[required] The unique ID of the memory resource where records will be deleted.}

\item{records}{[required] A list of memory record deletion inputs to be processed in the batch
operation.}
}
\description{
Deletes multiple memory records in a single batch operation from the specified memory.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_batch_delete_memory_records/} for full documentation.
}
\keyword{internal}
