% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_complete_resource_token_auth}
\alias{bedrockagentcore_complete_resource_token_auth}
\title{Confirms the user authentication session for obtaining OAuth2}
\usage{
bedrockagentcore_complete_resource_token_auth(userIdentifier, sessionUri)
}
\arguments{
\item{userIdentifier}{[required] The OAuth2.0 token or user ID that was used to generate the workload
access token used for initiating the user authorization flow to retrieve
OAuth2.0 tokens.}

\item{sessionUri}{[required] Unique identifier for the user's authentication session for retrieving
OAuth2 tokens. This ID tracks the authorization flow state across
multiple requests and responses during the OAuth2 authentication
process.}
}
\description{
Confirms the user authentication session for obtaining OAuth2.0 tokens for a resource.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_complete_resource_token_auth/} for full documentation.
}
\keyword{internal}
