% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_create_payment_instrument}
\alias{bedrockagentcore_create_payment_instrument}
\title{Create a new payment instrument for a connector}
\usage{
bedrockagentcore_create_payment_instrument(
  userId = NULL,
  agentName = NULL,
  paymentManagerArn,
  paymentConnectorId,
  paymentInstrumentType,
  paymentInstrumentDetails,
  clientToken = NULL
)
}
\arguments{
\item{userId}{The user ID associated with this payment instrument.}

\item{agentName}{The agent name associated with this request, used for observability.}

\item{paymentManagerArn}{[required] The ARN of the payment manager that owns this payment instrument.}

\item{paymentConnectorId}{[required] The ID of the payment connector to use for this instrument.}

\item{paymentInstrumentType}{[required] The type of payment instrument being created.}

\item{paymentInstrumentDetails}{[required] The details of the payment instrument.}

\item{clientToken}{Idempotency token to ensure request uniqueness.}
}
\description{
Create a new payment instrument for a connector

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_create_payment_instrument/} for full documentation.
}
\keyword{internal}
