% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_delete_ab_test}
\alias{bedrockagentcore_delete_ab_test}
\title{Deletes an A/B test and its associated gateway rules}
\usage{
bedrockagentcore_delete_ab_test(abTestId)
}
\arguments{
\item{abTestId}{[required] The unique identifier of the A/B test to delete.}
}
\description{
Deletes an A/B test and its associated gateway rules.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_delete_ab_test/} for full documentation.
}
\keyword{internal}
