% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_delete_batch_evaluation}
\alias{bedrockagentcore_delete_batch_evaluation}
\title{Deletes a batch evaluation and its associated results}
\usage{
bedrockagentcore_delete_batch_evaluation(batchEvaluationId)
}
\arguments{
\item{batchEvaluationId}{[required] The unique identifier of the batch evaluation to delete.}
}
\description{
Deletes a batch evaluation and its associated results.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_delete_batch_evaluation/} for full documentation.
}
\keyword{internal}
