% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_delete_capacity_provider_session}
\alias{bedrockagentcore_delete_capacity_provider_session}
\title{Deletes a session associated with a capacity provider in Amazon Bedrock
AgentCore and makes the session unavailable for further use}
\usage{
bedrockagentcore_delete_capacity_provider_session(
  capacityProviderId,
  sessionId
)
}
\arguments{
\item{capacityProviderId}{[required] The unique identifier of the capacity provider associated with the session.}

\item{sessionId}{[required] The unique identifier of the capacity provider session to delete.}
}
\description{
Deletes a session associated with a capacity provider in Amazon Bedrock AgentCore and makes the session unavailable for further use. To delete a capacity provider session, specify both the capacity provider identifier and the session ID. After you delete a session, you cannot restart it.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_delete_capacity_provider_session/} for full documentation.
}
\keyword{internal}
