% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_delete_event}
\alias{bedrockagentcore_delete_event}
\title{Deletes an event from an AgentCore Memory resource}
\usage{
bedrockagentcore_delete_event(memoryId, sessionId, eventId, actorId)
}
\arguments{
\item{memoryId}{[required] The identifier of the AgentCore Memory resource from which to delete the
event.}

\item{sessionId}{[required] The identifier of the session containing the event to delete.}

\item{eventId}{[required] The identifier of the event to delete.}

\item{actorId}{[required] The identifier of the actor associated with the event to delete.}
}
\description{
Deletes an event from an AgentCore Memory resource. When you delete an event, it is permanently removed.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_delete_event/} for full documentation.
}
\keyword{internal}
