% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_delete_recommendation}
\alias{bedrockagentcore_delete_recommendation}
\title{Deletes a recommendation and its associated results}
\usage{
bedrockagentcore_delete_recommendation(recommendationId)
}
\arguments{
\item{recommendationId}{[required] The unique identifier of the recommendation to delete.}
}
\description{
Deletes a recommendation and its associated results.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_delete_recommendation/} for full documentation.
}
\keyword{internal}
