% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_get_ab_test}
\alias{bedrockagentcore_get_ab_test}
\title{Retrieves detailed information about an A/B test, including its
configuration, status, and statistical results}
\usage{
bedrockagentcore_get_ab_test(abTestId)
}
\arguments{
\item{abTestId}{[required] The unique identifier of the A/B test to retrieve.}
}
\description{
Retrieves detailed information about an A/B test, including its configuration, status, and statistical results.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_get_ab_test/} for full documentation.
}
\keyword{internal}
