% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_get_agent_card}
\alias{bedrockagentcore_get_agent_card}
\title{Retrieves the A2A agent card associated with an AgentCore Runtime agent}
\usage{
bedrockagentcore_get_agent_card(
  runtimeSessionId = NULL,
  agentRuntimeArn,
  qualifier = NULL
)
}
\arguments{
\item{runtimeSessionId}{The session ID that the AgentCore Runtime agent is using.}

\item{agentRuntimeArn}{[required] The ARN of the AgentCore Runtime agent for which you want to get the A2A
agent card.}

\item{qualifier}{Optional qualifier to specify an agent alias, such as \code{prod}code\> or
\code{dev}. If you don't provide a value, the DEFAULT alias is used.}
}
\description{
Retrieves the A2A agent card associated with an AgentCore Runtime agent.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_get_agent_card/} for full documentation.
}
\keyword{internal}
