% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_get_batch_evaluation}
\alias{bedrockagentcore_get_batch_evaluation}
\title{Retrieves detailed information about a batch evaluation, including its
status, configuration, results, and any error details}
\usage{
bedrockagentcore_get_batch_evaluation(batchEvaluationId)
}
\arguments{
\item{batchEvaluationId}{[required] The unique identifier of the batch evaluation to retrieve.}
}
\description{
Retrieves detailed information about a batch evaluation, including its status, configuration, results, and any error details.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_get_batch_evaluation/} for full documentation.
}
\keyword{internal}
