% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_get_code_interpreter_session}
\alias{bedrockagentcore_get_code_interpreter_session}
\title{Retrieves detailed information about a specific code interpreter session
in Amazon Bedrock AgentCore}
\usage{
bedrockagentcore_get_code_interpreter_session(
  codeInterpreterIdentifier,
  sessionId
)
}
\arguments{
\item{codeInterpreterIdentifier}{[required] The unique identifier of the code interpreter associated with the
session.}

\item{sessionId}{[required] The unique identifier of the code interpreter session to retrieve.}
}
\description{
Retrieves detailed information about a specific code interpreter session in Amazon Bedrock AgentCore. This operation returns the session's configuration, current status, and metadata.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_get_code_interpreter_session/} for full documentation.
}
\keyword{internal}
