% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_get_event}
\alias{bedrockagentcore_get_event}
\title{Retrieves information about a specific event in an AgentCore Memory
resource}
\usage{
bedrockagentcore_get_event(memoryId, sessionId, actorId, eventId)
}
\arguments{
\item{memoryId}{[required] The identifier of the AgentCore Memory resource containing the event.}

\item{sessionId}{[required] The identifier of the session containing the event.}

\item{actorId}{[required] The identifier of the actor associated with the event.}

\item{eventId}{[required] The identifier of the event to retrieve.}
}
\description{
Retrieves information about a specific event in an AgentCore Memory resource.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_get_event/} for full documentation.
}
\keyword{internal}
