% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_get_payment_instrument}
\alias{bedrockagentcore_get_payment_instrument}
\title{Get a payment instrument by ID}
\usage{
bedrockagentcore_get_payment_instrument(
  userId = NULL,
  agentName = NULL,
  paymentManagerArn,
  paymentConnectorId = NULL,
  paymentInstrumentId
)
}
\arguments{
\item{userId}{The user ID associated with this payment instrument.}

\item{agentName}{The agent name associated with this request, used for observability.}

\item{paymentManagerArn}{[required] The ARN of the payment manager that owns this payment instrument.}

\item{paymentConnectorId}{The ID of the payment connector.}

\item{paymentInstrumentId}{[required] The ID of the payment instrument to retrieve.}
}
\description{
Get a payment instrument by ID

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_get_payment_instrument/} for full documentation.
}
\keyword{internal}
