% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_get_payment_session}
\alias{bedrockagentcore_get_payment_session}
\title{Get a payment session}
\usage{
bedrockagentcore_get_payment_session(
  userId = NULL,
  agentName = NULL,
  paymentManagerArn,
  paymentSessionId
)
}
\arguments{
\item{userId}{The user ID associated with this payment session.}

\item{agentName}{The agent name associated with this request, used for observability.}

\item{paymentManagerArn}{[required] The ARN of the payment manager that owns this session.}

\item{paymentSessionId}{[required] The ID of the payment session to retrieve.}
}
\description{
Get a payment session

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_get_payment_session/} for full documentation.
}
\keyword{internal}
