% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_get_recommendation}
\alias{bedrockagentcore_get_recommendation}
\title{Retrieves detailed information about a recommendation, including its
configuration, status, and results}
\usage{
bedrockagentcore_get_recommendation(recommendationId)
}
\arguments{
\item{recommendationId}{[required] The unique identifier of the recommendation to retrieve.}
}
\description{
Retrieves detailed information about a recommendation, including its configuration, status, and results.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_get_recommendation/} for full documentation.
}
\keyword{internal}
