% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_get_resource_api_key}
\alias{bedrockagentcore_get_resource_api_key}
\title{Retrieves the API key associated with an API key credential provider}
\usage{
bedrockagentcore_get_resource_api_key(
  workloadIdentityToken,
  resourceCredentialProviderName
)
}
\arguments{
\item{workloadIdentityToken}{[required] The identity token of the workload from which you want to retrieve the
API key.}

\item{resourceCredentialProviderName}{[required] The credential provider name for the resource from which you are
retrieving the API key.}
}
\description{
Retrieves the API key associated with an API key credential provider.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_get_resource_api_key/} for full documentation.
}
\keyword{internal}
