% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_get_resource_oauth_2_token}
\alias{bedrockagentcore_get_resource_oauth_2_token}
\title{Returns the OAuth 2}
\usage{
bedrockagentcore_get_resource_oauth_2_token(
  workloadIdentityToken,
  resourceCredentialProviderName,
  scopes,
  oauth2Flow,
  sessionUri = NULL,
  resourceOauth2ReturnUrl = NULL,
  forceAuthentication = NULL,
  customParameters = NULL,
  customState = NULL,
  resources = NULL,
  audiences = NULL
)
}
\arguments{
\item{workloadIdentityToken}{[required] The identity token of the workload from which you want to retrieve the
OAuth2 token.}

\item{resourceCredentialProviderName}{[required] The name of the resource's credential provider.}

\item{scopes}{[required] The OAuth scopes being requested.}

\item{oauth2Flow}{[required] The type of flow to be performed.}

\item{sessionUri}{Unique identifier for the user's authentication session for retrieving
OAuth2 tokens. This ID tracks the authorization flow state across
multiple requests and responses during the OAuth2 authentication
process.}

\item{resourceOauth2ReturnUrl}{The callback URL to redirect to after the OAuth 2.0 token retrieval is
complete. This URL must be one of the provided URLs configured for the
workload identity.}

\item{forceAuthentication}{Indicates whether to always initiate a new three-legged OAuth (3LO)
flow, regardless of any existing session.}

\item{customParameters}{A map of custom parameters to include in the authorization request to
the resource credential provider. These parameters are in addition to
the standard OAuth 2.0 flow parameters, and will not override them.}

\item{customState}{An opaque string that will be sent back to the callback URL provided in
resourceOauth2ReturnUrl. This state should be used to protect the
callback URL of your application against CSRF attacks by ensuring the
response corresponds to the original request.}

\item{resources}{The resources to include in the token request. These are used to specify
the target resources for which the OAuth2 token is being requested.}

\item{audiences}{The audiences to include in the token request. These are used to specify
the intended recipients of the OAuth2 token.}
}
\description{
Returns the OAuth 2.0 token of the provided resource.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_get_resource_oauth_2_token/} for full documentation.
}
\keyword{internal}
