% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_get_resource_payment_token}
\alias{bedrockagentcore_get_resource_payment_token}
\title{Generates authentication tokens for payment providers that use
vendor-specific authentication mechanisms}
\usage{
bedrockagentcore_get_resource_payment_token(
  workloadIdentityToken,
  resourceCredentialProviderName,
  paymentTokenRequest
)
}
\arguments{
\item{workloadIdentityToken}{[required] Workload access token for authorization. Named workloadIdentityToken for
consistency with APIKey and OAuth2CredentialProvider.}

\item{resourceCredentialProviderName}{[required] Name of the payment credential provider to use}

\item{paymentTokenRequest}{[required] Vendor-specific token request input Contains all request parameters in a
type-safe, vendor-specific structure}
}
\description{
Generates authentication tokens for payment providers that use vendor-specific authentication mechanisms.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_get_resource_payment_token/} for full documentation.
}
\keyword{internal}
