% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_get_workload_access_token}
\alias{bedrockagentcore_get_workload_access_token}
\title{Obtains a workload access token for agentic workloads not acting on
behalf of a user}
\usage{
bedrockagentcore_get_workload_access_token(workloadName)
}
\arguments{
\item{workloadName}{[required] The unique identifier for the registered workload.}
}
\description{
Obtains a workload access token for agentic workloads not acting on behalf of a user.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_get_workload_access_token/} for full documentation.
}
\keyword{internal}
