% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_get_workload_access_token_for_jwt}
\alias{bedrockagentcore_get_workload_access_token_for_jwt}
\title{Obtains a workload access token for agentic workloads acting on behalf
of a user, using a JWT token}
\usage{
bedrockagentcore_get_workload_access_token_for_jwt(workloadName, userToken)
}
\arguments{
\item{workloadName}{[required] The unique identifier for the registered workload.}

\item{userToken}{[required] The OAuth 2.0 token issued by the user's identity provider.}
}
\description{
Obtains a workload access token for agentic workloads acting on behalf of a user, using a JWT token.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_get_workload_access_token_for_jwt/} for full documentation.
}
\keyword{internal}
