% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_get_workload_access_token_for_user_id}
\alias{bedrockagentcore_get_workload_access_token_for_user_id}
\title{Obtains a workload access token for agentic workloads acting on behalf
of a user, using the user's ID}
\usage{
bedrockagentcore_get_workload_access_token_for_user_id(workloadName, userId)
}
\arguments{
\item{workloadName}{[required] The name of the workload from which you want to retrieve the access
token.}

\item{userId}{[required] The ID of the user for whom you are retrieving the access token.}
}
\description{
Obtains a workload access token for agentic workloads acting on behalf of a user, using the user's ID.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_get_workload_access_token_for_user_id/} for full documentation.
}
\keyword{internal}
