% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_ingest_data}
\alias{bedrockagentcore_ingest_data}
\title{Submits content directly for ingestion to generate long-term memory
records in a AgentCore Memory resource}
\usage{
bedrockagentcore_ingest_data(
  memoryId,
  source,
  contentTimestamp,
  actorId,
  sessionId = NULL,
  extractionConfig = NULL,
  metadata = NULL,
  clientToken = NULL
)
}
\arguments{
\item{memoryId}{[required] The identifier of the AgentCore Memory resource to ingest content into.}

\item{source}{[required] The content to ingest. Only inline content is supported.}

\item{contentTimestamp}{[required] The timestamp of when the content occurred.}

\item{actorId}{[required] The identifier of the actor associated with this content. An actor represents an entity that participates in sessions and generates content.}

\item{sessionId}{The identifier of the session that the content belongs to. If not provided, a session identifier is generated and returned in the response.}

\item{extractionConfig}{The extraction configuration for long-term memory records. Use this parameter to specify namespace variable keys and their values for namespace substitution during extraction.}

\item{metadata}{The key-value metadata to attach to the content.}

\item{clientToken}{A unique, case-sensitive identifier to ensure that the operation completes no more than one time. If this token matches a previous request, AgentCore ignores the request, but does not return an error.}
}
\description{
Submits content directly for ingestion to generate long-term memory records in a AgentCore Memory resource.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_ingest_data/} for full documentation.
}
\keyword{internal}
