% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_list_actors}
\alias{bedrockagentcore_list_actors}
\title{Lists all actors in an AgentCore Memory resource}
\usage{
bedrockagentcore_list_actors(memoryId, maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{memoryId}{[required] The identifier of the AgentCore Memory resource for which to list
actors.}

\item{maxResults}{The maximum number of results to return in a single call. The default
value is 20.}

\item{nextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}
}
\description{
Lists all actors in an AgentCore Memory resource. We recommend using pagination to ensure that the operation returns quickly and successfully.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_list_actors/} for full documentation.
}
\keyword{internal}
