% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_list_batch_evaluations}
\alias{bedrockagentcore_list_batch_evaluations}
\title{Lists all batch evaluations in the account, providing summary
information about each evaluation's status and configuration}
\usage{
bedrockagentcore_list_batch_evaluations(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of results to return in the response. If the total
number of results is greater than this value, use the token returned in
the response in the \code{nextToken} field when making another request to
return the next batch of results.}

\item{nextToken}{If the total number of results is greater than the \code{maxResults} value
provided in the request, enter the token returned in the \code{nextToken}
field in the response in this field to return the next batch of results.}
}
\description{
Lists all batch evaluations in the account, providing summary information about each evaluation's status and configuration.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_list_batch_evaluations/} for full documentation.
}
\keyword{internal}
