% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_list_memory_records}
\alias{bedrockagentcore_list_memory_records}
\title{Lists memory records in an AgentCore Memory resource based on specified
criteria}
\usage{
bedrockagentcore_list_memory_records(
  memoryId,
  namespace = NULL,
  namespacePath = NULL,
  memoryStrategyId = NULL,
  maxResults = NULL,
  nextToken = NULL,
  metadataFilters = NULL
)
}
\arguments{
\item{memoryId}{[required] The identifier of the AgentCore Memory resource for which to list memory
records.}

\item{namespace}{The namespace prefix to filter memory records by. Returns all memory
records in namespaces that start with the provided prefix.}

\item{namespacePath}{Use namespacePath for hierarchical retrievals. Return all memory records
where namespace falls under the same parent hierarchy.}

\item{memoryStrategyId}{The memory strategy identifier to filter memory records by. If
specified, only memory records with this strategy ID are returned.}

\item{maxResults}{The maximum number of results to return in a single call. The default
value is 20.}

\item{nextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{metadataFilters}{A list of metadata filter expressions to scope the returned memory
records.}
}
\description{
Lists memory records in an AgentCore Memory resource based on specified criteria. We recommend using pagination to ensure that the operation returns quickly and successfully.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_list_memory_records/} for full documentation.
}
\keyword{internal}
