% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_list_payment_instruments}
\alias{bedrockagentcore_list_payment_instruments}
\title{List payment instruments for a manager}
\usage{
bedrockagentcore_list_payment_instruments(
  userId = NULL,
  agentName = NULL,
  paymentManagerArn,
  paymentConnectorId = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{userId}{The user ID associated with the payment instruments.}

\item{agentName}{The agent name associated with this request, used for observability.}

\item{paymentManagerArn}{[required] The ARN of the payment manager that owns the payment instruments.}

\item{paymentConnectorId}{The ID of the payment connector to filter by.}

\item{nextToken}{Token for pagination to retrieve the next set of results.}

\item{maxResults}{Maximum number of results to return in a single response.}
}
\description{
List payment instruments for a manager

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_list_payment_instruments/} for full documentation.
}
\keyword{internal}
