% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_list_payment_sessions}
\alias{bedrockagentcore_list_payment_sessions}
\title{List payment manager sessions}
\usage{
bedrockagentcore_list_payment_sessions(
  userId = NULL,
  agentName = NULL,
  paymentManagerArn,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{userId}{The user ID associated with the payment sessions.}

\item{agentName}{The agent name associated with this request, used for observability.}

\item{paymentManagerArn}{[required] The ARN of the payment manager that owns the sessions.}

\item{nextToken}{Token for pagination to retrieve the next set of results.}

\item{maxResults}{Maximum number of results to return in a single response.}
}
\description{
List payment manager sessions

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_list_payment_sessions/} for full documentation.
}
\keyword{internal}
