% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_list_recommendations}
\alias{bedrockagentcore_list_recommendations}
\title{Lists all recommendations in the account, with optional filtering by
status}
\usage{
bedrockagentcore_list_recommendations(
  maxResults = NULL,
  nextToken = NULL,
  statusFilter = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of results to return in the response. If the total
number of results is greater than this value, use the token returned in
the response in the \code{nextToken} field when making another request to
return the next batch of results.}

\item{nextToken}{If the total number of results is greater than the \code{maxResults} value
provided in the request, enter the token returned in the \code{nextToken}
field in the response in this field to return the next batch of results.}

\item{statusFilter}{Optional filter to return only recommendations with the specified
status.}
}
\description{
Lists all recommendations in the account, with optional filtering by status.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_list_recommendations/} for full documentation.
}
\keyword{internal}
