% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_list_sessions}
\alias{bedrockagentcore_list_sessions}
\title{Lists sessions in an AgentCore Memory resource based on specified
criteria}
\usage{
bedrockagentcore_list_sessions(
  memoryId,
  actorId,
  maxResults = NULL,
  nextToken = NULL,
  filter = NULL
)
}
\arguments{
\item{memoryId}{[required] The identifier of the AgentCore Memory resource for which to list
sessions.}

\item{actorId}{[required] The identifier of the actor for which to list sessions.}

\item{maxResults}{The maximum number of results to return in a single call. The default
value is 20.}

\item{nextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{filter}{Filter criteria to apply when listing sessions.}
}
\description{
Lists sessions in an AgentCore Memory resource based on specified criteria. We recommend using pagination to ensure that the operation returns quickly and successfully.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_list_sessions/} for full documentation.
}
\keyword{internal}
