% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_process_payment}
\alias{bedrockagentcore_process_payment}
\title{Processes a payment using a payment instrument within a payment session}
\usage{
bedrockagentcore_process_payment(
  userId = NULL,
  agentName = NULL,
  paymentManagerArn,
  paymentSessionId,
  paymentInstrumentId,
  paymentType,
  paymentInput,
  clientToken = NULL
)
}
\arguments{
\item{userId}{The user ID associated with this payment.}

\item{agentName}{The agent name associated with this request, used for observability.}

\item{paymentManagerArn}{[required] The ARN of the payment manager.}

\item{paymentSessionId}{[required] The ID of the payment session.}

\item{paymentInstrumentId}{[required] The ID of the payment instrument to use.}

\item{paymentType}{[required] The type of payment to process.}

\item{paymentInput}{[required] The payment input details specific to the payment type.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}
}
\description{
Processes a payment using a payment instrument within a payment session.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_process_payment/} for full documentation.
}
\keyword{internal}
