% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_start_memory_extraction_job}
\alias{bedrockagentcore_start_memory_extraction_job}
\title{Starts a memory extraction job that processes events that failed
extraction previously in an AgentCore Memory resource and produces
structured memory records}
\usage{
bedrockagentcore_start_memory_extraction_job(
  memoryId,
  extractionJob,
  clientToken = NULL
)
}
\arguments{
\item{memoryId}{[required] The unique identifier of the memory for which to start extraction jobs.}

\item{extractionJob}{[required] Extraction job to start in this operation.}

\item{clientToken}{A unique, case-sensitive identifier to ensure idempotent processing of
the request.}
}
\description{
Starts a memory extraction job that processes events that failed extraction previously in an AgentCore Memory resource and produces structured memory records. When earlier extraction attempts have left events unprocessed, this job will pick up and extract those as well.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_start_memory_extraction_job/} for full documentation.
}
\keyword{internal}
