% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_stop_batch_evaluation}
\alias{bedrockagentcore_stop_batch_evaluation}
\title{Stops a running batch evaluation}
\usage{
bedrockagentcore_stop_batch_evaluation(batchEvaluationId)
}
\arguments{
\item{batchEvaluationId}{[required] The unique identifier of the batch evaluation to stop.}
}
\description{
Stops a running batch evaluation. Sessions that have already been evaluated retain their results.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_stop_batch_evaluation/} for full documentation.
}
\keyword{internal}
