% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_stop_browser_session}
\alias{bedrockagentcore_stop_browser_session}
\title{Terminates an active browser session in Amazon Bedrock AgentCore}
\usage{
bedrockagentcore_stop_browser_session(
  traceId = NULL,
  traceParent = NULL,
  browserIdentifier,
  sessionId,
  clientToken = NULL
)
}
\arguments{
\item{traceId}{The trace identifier for request tracking.}

\item{traceParent}{The parent trace information for distributed tracing.}

\item{browserIdentifier}{[required] The unique identifier of the browser associated with the session.}

\item{sessionId}{[required] The unique identifier of the browser session to stop.}

\item{clientToken}{A unique, case-sensitive identifier to ensure that the API request
completes no more than one time. If this token matches a previous
request, Amazon Bedrock AgentCore ignores the request, but does not
return an error.}
}
\description{
Terminates an active browser session in Amazon Bedrock AgentCore. This operation stops the session, releases associated resources, and makes the session unavailable for further use.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_stop_browser_session/} for full documentation.
}
\keyword{internal}
