% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_update_browser_stream}
\alias{bedrockagentcore_update_browser_stream}
\title{Updates a browser stream}
\usage{
bedrockagentcore_update_browser_stream(
  browserIdentifier,
  sessionId,
  streamUpdate,
  clientToken = NULL
)
}
\arguments{
\item{browserIdentifier}{[required] The identifier of the browser.}

\item{sessionId}{[required] The identifier of the browser session.}

\item{streamUpdate}{[required] The update to apply to the browser stream.}

\item{clientToken}{A unique, case-sensitive identifier to ensure that the operation
completes no more than one time. If this token matches a previous
request, Amazon Bedrock ignores the request, but does not return an
error.}
}
\description{
Updates a browser stream. To use this operation, you must have permissions to perform the bedrock:UpdateBrowserStream action.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_update_browser_stream/} for full documentation.
}
\keyword{internal}
