% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_create_agent_runtime_endpoint}
\alias{bedrockagentcorecontrol_create_agent_runtime_endpoint}
\title{Creates an AgentCore Runtime endpoint}
\usage{
bedrockagentcorecontrol_create_agent_runtime_endpoint(
  agentRuntimeId,
  name,
  agentRuntimeVersion = NULL,
  description = NULL,
  clientToken = NULL,
  tags = NULL
)
}
\arguments{
\item{agentRuntimeId}{[required] The unique identifier of the AgentCore Runtime to create an endpoint
for.}

\item{name}{[required] The name of the AgentCore Runtime endpoint.}

\item{agentRuntimeVersion}{The version of the AgentCore Runtime to use for the endpoint.}

\item{description}{The description of the AgentCore Runtime endpoint.}

\item{clientToken}{A unique, case-sensitive identifier to ensure idempotency of the
request.}

\item{tags}{A map of tag keys and values to assign to the agent runtime endpoint.
Tags enable you to categorize your resources in different ways, for
example, by purpose, owner, or environment.}
}
\description{
Creates an AgentCore Runtime endpoint.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_create_agent_runtime_endpoint/} for full documentation.
}
\keyword{internal}
