% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_create_capacity_provider}
\alias{bedrockagentcorecontrol_create_capacity_provider}
\title{Creates a capacity provider}
\usage{
bedrockagentcorecontrol_create_capacity_provider(
  name,
  description = NULL,
  permissionsConfiguration,
  clientToken = NULL,
  tags = NULL,
  computeConfiguration
)
}
\arguments{
\item{name}{[required] The name of the capacity provider. The name must be unique within your account.}

\item{description}{An optional description of the capacity provider. If you don't specify a description, the service creates the capacity provider without one.}

\item{permissionsConfiguration}{[required] The permissions configuration for the capacity provider. This specifies the IAM role that AgentCore uses to manage the Amazon EC2 instances on your behalf.}

\item{clientToken}{A unique, case-sensitive identifier to ensure that the API request completes no more than one time. If you don't specify this field, a value is randomly generated for you. If this token matches a previous request, the service ignores the request, but doesn't return an error. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring idempotency}.}

\item{tags}{A map of tag keys and values to associate with the capacity provider. If you don't specify tags, the capacity provider is created with no tags.}

\item{computeConfiguration}{[required] The compute configuration for the capacity provider. This defines the Amazon EC2 compute resources used to launch instances: the operating system, allowed instance types, networking, and storage.}
}
\description{
Creates a capacity provider. A capacity provider defines the Amazon EC2 infrastructure for AgentCore Runtime, including the operating system, allowed instance types, networking, and storage. It also specifies the IAM permissions that AgentCore uses to manage those instances.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_create_capacity_provider/} for full documentation.
}
\keyword{internal}
