% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_create_code_interpreter}
\alias{bedrockagentcorecontrol_create_code_interpreter}
\title{Creates a custom code interpreter}
\usage{
bedrockagentcorecontrol_create_code_interpreter(
  name,
  description = NULL,
  executionRoleArn = NULL,
  networkConfiguration,
  certificates = NULL,
  filesystemConfigurations = NULL,
  clientToken = NULL,
  tags = NULL
)
}
\arguments{
\item{name}{[required] The name of the code interpreter. The name must be unique within your account.}

\item{description}{The description of the code interpreter.}

\item{executionRoleArn}{The Amazon Resource Name (ARN) of the IAM role that provides permissions for the code interpreter to access Amazon Web Services services.}

\item{networkConfiguration}{[required] The network configuration for the code interpreter. This configuration specifies the network mode for the code interpreter.}

\item{certificates}{A list of certificates to install in the code interpreter.}

\item{filesystemConfigurations}{The file system configurations to mount into the code interpreter. Use these configurations to mount your own Amazon Simple Storage Service (Amazon S3) Files or Amazon Elastic File System (Amazon EFS) access points. Your sessions can then access your data. If you don't specify this field, no file systems are mounted.}

\item{clientToken}{A unique, case-sensitive identifier to ensure that the operation completes no more than one time. If this token matches a previous request, Amazon Bedrock AgentCore ignores the request but does not return an error.}

\item{tags}{A map of tag keys and values to assign to the code interpreter. Tags enable you to categorize your resources in different ways, for example, by purpose, owner, or environment.}
}
\description{
Creates a custom code interpreter.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_create_code_interpreter/} for full documentation.
}
\keyword{internal}
