% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_create_consent_portal}
\alias{bedrockagentcorecontrol_create_consent_portal}
\title{Creates a new consent portal}
\usage{
bedrockagentcorecontrol_create_consent_portal(
  executionRoleArn,
  idpConfig,
  name,
  sources,
  description = NULL,
  tags = NULL
)
}
\arguments{
\item{executionRoleArn}{[required] The Amazon Resource Name (ARN) of the IAM role that the consent portal assumes to access the resources defined in its sources.}

\item{idpConfig}{[required] The identity provider configuration that the consent portal uses to authenticate end users.}

\item{name}{[required] The name of the consent portal. The name must be unique within your account.}

\item{sources}{[required] The resources served by the consent portal. Currently, we only support type \code{agentcore-gateway}.}

\item{description}{The description of the consent portal.}

\item{tags}{A map of tag keys and values to assign to the consent portal. Tags enable you to categorize your resources in different ways, for example, by purpose, owner, or environment.}
}
\description{
Creates a new consent portal.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_create_consent_portal/} for full documentation.
}
\keyword{internal}
