% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_create_dataset}
\alias{bedrockagentcorecontrol_create_dataset}
\title{Creates a new dataset resource asynchronously}
\usage{
bedrockagentcorecontrol_create_dataset(
  clientToken = NULL,
  datasetName,
  description = NULL,
  source,
  schemaType,
  kmsKeyArn = NULL,
  tags = NULL
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier to ensure that the API request completes no more than one time. If you don't specify this field, a value is randomly generated for you. If this token matches a previous request, the service ignores the request, but doesn't return an error. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring idempotency}.}

\item{datasetName}{[required] Human-readable name for the dataset. Must be unique within the account. Immutable after creation.}

\item{description}{A description of the dataset.}

\item{source}{[required] Source of initial examples. Provide either inline examples or an S3 URI pointing to a JSONL file.}

\item{schemaType}{[required] Versioned schema type governing the structure of examples. Immutable after creation.}

\item{kmsKeyArn}{Optional KMS key ARN for server-side encryption on service Amazon S3 writes.}

\item{tags}{A map of tag keys and values to assign to the dataset.}
}
\description{
Creates a new dataset resource asynchronously. Returns immediately with status CREATING. Poll \code{\link[=bedrockagentcorecontrol_get_dataset]{get_dataset}} until status transitions to ACTIVE or CREATE_FAILED.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_create_dataset/} for full documentation.
}
\keyword{internal}
