% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_create_harness_endpoint}
\alias{bedrockagentcorecontrol_create_harness_endpoint}
\title{Operation to create a harness endpoint}
\usage{
bedrockagentcorecontrol_create_harness_endpoint(
  harnessId,
  endpointName,
  targetVersion = NULL,
  description = NULL,
  clientToken = NULL,
  tags = NULL
)
}
\arguments{
\item{harnessId}{[required] The ID of the harness to create an endpoint for.}

\item{endpointName}{[required] The name of the endpoint. Must start with a letter and contain only alphanumeric characters and underscores.}

\item{targetVersion}{The harness version that the endpoint points to and serves invocations from.}

\item{description}{A description of the endpoint.}

\item{clientToken}{A unique, case-sensitive identifier to ensure idempotency of the request.}

\item{tags}{Tags to apply to the endpoint resource.}
}
\description{
Operation to create a harness endpoint.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_create_harness_endpoint/} for full documentation.
}
\keyword{internal}
