% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_create_oauth_2_credential_provider}
\alias{bedrockagentcorecontrol_create_oauth_2_credential_provider}
\title{Creates a new OAuth2 credential provider}
\usage{
bedrockagentcorecontrol_create_oauth_2_credential_provider(
  name,
  credentialProviderVendor,
  oauth2ProviderConfigInput,
  tags = NULL
)
}
\arguments{
\item{name}{[required] The name of the OAuth2 credential provider. The name must be unique
within your account.}

\item{credentialProviderVendor}{[required] The vendor of the OAuth2 credential provider. This specifies which
OAuth2 implementation to use.}

\item{oauth2ProviderConfigInput}{[required] The configuration settings for the OAuth2 provider, including client ID,
client secret, and other vendor-specific settings.}

\item{tags}{A map of tag keys and values to assign to the OAuth2 credential
provider. Tags enable you to categorize your resources in different
ways, for example, by purpose, owner, or environment.}
}
\description{
Creates a new OAuth2 credential provider.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_create_oauth_2_credential_provider/} for full documentation.
}
\keyword{internal}
