% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_create_payment_manager}
\alias{bedrockagentcorecontrol_create_payment_manager}
\title{Creates a new payment manager in your Amazon Web Services account}
\usage{
bedrockagentcorecontrol_create_payment_manager(
  name,
  description = NULL,
  authorizerType,
  authorizerConfiguration = NULL,
  roleArn,
  clientToken = NULL,
  tags = NULL,
  kmsKeyArn = NULL
)
}
\arguments{
\item{name}{[required] The name of the payment manager.}

\item{description}{A description of the payment manager.}

\item{authorizerType}{[required] The type of authorizer to use for the payment manager.
\itemize{
\item \code{CUSTOM_JWT} - Authorize with a bearer token.
\item \code{AWS_IAM} - Authorize with your Amazon Web Services IAM credentials.
}}

\item{authorizerConfiguration}{The authorizer configuration for the payment manager.}

\item{roleArn}{[required] The Amazon Resource Name (ARN) of the IAM role that the payment manager assumes to access resources on your behalf.}

\item{clientToken}{A unique, case-sensitive identifier to ensure that the API request completes no more than one time. If you don't specify this field, a value is randomly generated for you. If this token matches a previous request, the service ignores the request, but doesn't return an error. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring idempotency}.}

\item{tags}{A map of tag keys and values to assign to the payment manager.}

\item{kmsKeyArn}{The Amazon Resource Name (ARN) of the customer managed KMS key to use for encrypting sensitive payment manager data at rest. If you don't specify a key, the data is encrypted with an Amazon Web Services owned key.}
}
\description{
Creates a new payment manager in your Amazon Web Services account. A payment manager serves as the top-level resource for managing payment processing capabilities, including payment connectors that integrate with supported payment providers.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_create_payment_manager/} for full documentation.
}
\keyword{internal}
