% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_create_workload_identity}
\alias{bedrockagentcorecontrol_create_workload_identity}
\title{Creates a new workload identity}
\usage{
bedrockagentcorecontrol_create_workload_identity(
  name,
  allowedResourceOauth2ReturnUrls = NULL,
  tags = NULL
)
}
\arguments{
\item{name}{[required] The name of the workload identity. The name must be unique within your
account.}

\item{allowedResourceOauth2ReturnUrls}{The list of allowed OAuth2 return URLs for resources associated with
this workload identity.}

\item{tags}{A map of tag keys and values to assign to the workload identity. Tags
enable you to categorize your resources in different ways, for example,
by purpose, owner, or environment.}
}
\description{
Creates a new workload identity.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_create_workload_identity/} for full documentation.
}
\keyword{internal}
