% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_delete_agent_runtime_endpoint}
\alias{bedrockagentcorecontrol_delete_agent_runtime_endpoint}
\title{Deletes an AgentCore Runtime endpoint}
\usage{
bedrockagentcorecontrol_delete_agent_runtime_endpoint(
  agentRuntimeId,
  endpointName,
  clientToken = NULL
)
}
\arguments{
\item{agentRuntimeId}{[required] The unique identifier of the AgentCore Runtime associated with the endpoint.}

\item{endpointName}{[required] The name of the AgentCore Runtime endpoint to delete.}

\item{clientToken}{A unique, case-sensitive identifier to ensure idempotency of the request.}
}
\description{
Deletes an AgentCore Runtime endpoint.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_delete_agent_runtime_endpoint/} for full documentation.
}
\keyword{internal}
