% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_delete_api_key_credential_provider}
\alias{bedrockagentcorecontrol_delete_api_key_credential_provider}
\title{Deletes an API key credential provider}
\usage{
bedrockagentcorecontrol_delete_api_key_credential_provider(name)
}
\arguments{
\item{name}{[required] The name of the API key credential provider to delete.}
}
\description{
Deletes an API key credential provider.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_delete_api_key_credential_provider/} for full documentation.
}
\keyword{internal}
