% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_delete_configuration_bundle}
\alias{bedrockagentcorecontrol_delete_configuration_bundle}
\title{Deletes a configuration bundle and all of its versions}
\usage{
bedrockagentcorecontrol_delete_configuration_bundle(bundleId)
}
\arguments{
\item{bundleId}{[required] The unique identifier of the configuration bundle to delete.}
}
\description{
Deletes a configuration bundle and all of its versions.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_delete_configuration_bundle/} for full documentation.
}
\keyword{internal}
