% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_delete_consent_portal}
\alias{bedrockagentcorecontrol_delete_consent_portal}
\title{Deletes a consent portal}
\usage{
bedrockagentcorecontrol_delete_consent_portal(consentPortalIdentifier)
}
\arguments{
\item{consentPortalIdentifier}{[required] The identifier of the consent portal. You can specify either the consent portal ID or its Amazon Resource Name (ARN).}
}
\description{
Deletes a consent portal.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_delete_consent_portal/} for full documentation.
}
\keyword{internal}
