% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_delete_evaluator}
\alias{bedrockagentcorecontrol_delete_evaluator}
\title{Deletes a custom evaluator}
\usage{
bedrockagentcorecontrol_delete_evaluator(evaluatorId)
}
\arguments{
\item{evaluatorId}{[required] The unique identifier of the evaluator to delete.}
}
\description{
Deletes a custom evaluator. Builtin evaluators cannot be deleted. The evaluator must not be referenced by any active online evaluation configurations.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_delete_evaluator/} for full documentation.
}
\keyword{internal}
