% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_delete_gateway}
\alias{bedrockagentcorecontrol_delete_gateway}
\title{Deletes a gateway}
\usage{
bedrockagentcorecontrol_delete_gateway(gatewayIdentifier)
}
\arguments{
\item{gatewayIdentifier}{[required] The identifier of the gateway to delete.}
}
\description{
Deletes a gateway.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_delete_gateway/} for full documentation.
}
\keyword{internal}
