% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_delete_gateway_rule}
\alias{bedrockagentcorecontrol_delete_gateway_rule}
\title{Deletes a gateway rule}
\usage{
bedrockagentcorecontrol_delete_gateway_rule(gatewayIdentifier, ruleId)
}
\arguments{
\item{gatewayIdentifier}{[required] The identifier of the gateway containing the rule.}

\item{ruleId}{[required] The unique identifier of the rule to delete.}
}
\description{
Deletes a gateway rule.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_delete_gateway_rule/} for full documentation.
}
\keyword{internal}
