% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_delete_gateway_target}
\alias{bedrockagentcorecontrol_delete_gateway_target}
\title{Deletes a gateway target}
\usage{
bedrockagentcorecontrol_delete_gateway_target(gatewayIdentifier, targetId)
}
\arguments{
\item{gatewayIdentifier}{[required] The unique identifier of the gateway associated with the target.}

\item{targetId}{[required] The unique identifier of the gateway target to delete.}
}
\description{
Deletes a gateway target.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_delete_gateway_target/} for full documentation.
}
\keyword{internal}
