% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_delete_harness}
\alias{bedrockagentcorecontrol_delete_harness}
\title{Operation to delete a Harness}
\usage{
bedrockagentcorecontrol_delete_harness(
  harnessId,
  clientToken = NULL,
  deleteManagedMemory = NULL
)
}
\arguments{
\item{harnessId}{[required] The ID of the harness to delete.}

\item{clientToken}{A unique, case-sensitive identifier to ensure idempotency of the request.}

\item{deleteManagedMemory}{Whether to delete the managed memory on harness deletion. Default: true. If false, the memory is disassociated and becomes a regular customer-owned resource.}
}
\description{
Operation to delete a Harness.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_delete_harness/} for full documentation.
}
\keyword{internal}
