% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_delete_memory}
\alias{bedrockagentcorecontrol_delete_memory}
\title{Deletes an Amazon Bedrock AgentCore Memory resource}
\usage{
bedrockagentcorecontrol_delete_memory(clientToken = NULL, memoryId)
}
\arguments{
\item{clientToken}{A client token is used for keeping track of idempotent requests. It can contain a session id which can be around 250 chars, combined with a unique AWS identifier.}

\item{memoryId}{[required] The unique identifier of the memory to delete.}
}
\description{
Deletes an Amazon Bedrock AgentCore Memory resource. When you delete a memory resource, it is permanently removed.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_delete_memory/} for full documentation.
}
\keyword{internal}
