% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_delete_oauth_2_credential_provider}
\alias{bedrockagentcorecontrol_delete_oauth_2_credential_provider}
\title{Deletes an OAuth2 credential provider}
\usage{
bedrockagentcorecontrol_delete_oauth_2_credential_provider(name)
}
\arguments{
\item{name}{[required] The name of the OAuth2 credential provider to delete.}
}
\description{
Deletes an OAuth2 credential provider.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_delete_oauth_2_credential_provider/} for full documentation.
}
\keyword{internal}
