% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_delete_online_evaluation_config}
\alias{bedrockagentcorecontrol_delete_online_evaluation_config}
\title{Deletes an online evaluation configuration and stops any ongoing
evaluation processes associated with it}
\usage{
bedrockagentcorecontrol_delete_online_evaluation_config(
  onlineEvaluationConfigId
)
}
\arguments{
\item{onlineEvaluationConfigId}{[required] The unique identifier of the online evaluation configuration to delete.}
}
\description{
Deletes an online evaluation configuration and stops any ongoing evaluation processes associated with it.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_delete_online_evaluation_config/} for full documentation.
}
\keyword{internal}
