% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_delete_payment_connector}
\alias{bedrockagentcorecontrol_delete_payment_connector}
\title{Deletes a payment connector}
\usage{
bedrockagentcorecontrol_delete_payment_connector(
  paymentManagerId,
  paymentConnectorId,
  clientToken = NULL
)
}
\arguments{
\item{paymentManagerId}{[required] The unique identifier of the parent payment manager.}

\item{paymentConnectorId}{[required] The unique identifier of the payment connector to delete.}

\item{clientToken}{A unique, case-sensitive identifier to ensure that the API request
completes no more than one time. If you don't specify this field, a
value is randomly generated for you. If this token matches a previous
request, the service ignores the request, but doesn't return an error.
For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring idempotency}.}
}
\description{
Deletes a payment connector.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_delete_payment_connector/} for full documentation.
}
\keyword{internal}
