% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_delete_payment_credential_provider}
\alias{bedrockagentcorecontrol_delete_payment_credential_provider}
\title{Deletes a payment credential provider and its associated stored
credentials}
\usage{
bedrockagentcorecontrol_delete_payment_credential_provider(name)
}
\arguments{
\item{name}{[required] The name of the payment credential provider to delete.}
}
\description{
Deletes a payment credential provider and its associated stored credentials.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_delete_payment_credential_provider/} for full documentation.
}
\keyword{internal}
