% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_delete_payment_manager}
\alias{bedrockagentcorecontrol_delete_payment_manager}
\title{Deletes a payment manager}
\usage{
bedrockagentcorecontrol_delete_payment_manager(
  paymentManagerId,
  clientToken = NULL
)
}
\arguments{
\item{paymentManagerId}{[required] The unique identifier of the payment manager to delete.}

\item{clientToken}{A unique, case-sensitive identifier to ensure that the API request
completes no more than one time. If you don't specify this field, a
value is randomly generated for you. If this token matches a previous
request, the service ignores the request, but doesn't return an error.
For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring idempotency}.}
}
\description{
Deletes a payment manager. All payment connectors associated with the payment manager must be deleted before the payment manager can be deleted. This operation initiates the deletion process asynchronously.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_delete_payment_manager/} for full documentation.
}
\keyword{internal}
