% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_delete_registry}
\alias{bedrockagentcorecontrol_delete_registry}
\title{Deletes a registry}
\usage{
bedrockagentcorecontrol_delete_registry(registryId)
}
\arguments{
\item{registryId}{[required] The identifier of the registry to delete. You can specify either the
Amazon Resource Name (ARN) or the ID of the registry.}
}
\description{
Deletes a registry. The registry must contain zero records before it can be deleted. This operation initiates the deletion process asynchronously.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_delete_registry/} for full documentation.
}
\keyword{internal}
