% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_delete_registry_record}
\alias{bedrockagentcorecontrol_delete_registry_record}
\title{Deletes a registry record}
\usage{
bedrockagentcorecontrol_delete_registry_record(registryId, recordId)
}
\arguments{
\item{registryId}{[required] The identifier of the registry containing the record. You can specify
either the Amazon Resource Name (ARN) or the ID of the registry.}

\item{recordId}{[required] The identifier of the registry record to delete. You can specify either
the Amazon Resource Name (ARN) or the ID of the record.}
}
\description{
Deletes a registry record. The record's status transitions to \code{DELETING} and the record is removed asynchronously.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_delete_registry_record/} for full documentation.
}
\keyword{internal}
