% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_get_agent_runtime}
\alias{bedrockagentcorecontrol_get_agent_runtime}
\title{Gets an Amazon Bedrock AgentCore Runtime}
\usage{
bedrockagentcorecontrol_get_agent_runtime(
  agentRuntimeId,
  agentRuntimeVersion = NULL
)
}
\arguments{
\item{agentRuntimeId}{[required] The unique identifier of the AgentCore Runtime to retrieve.}

\item{agentRuntimeVersion}{The version of the AgentCore Runtime to retrieve.}
}
\description{
Gets an Amazon Bedrock AgentCore Runtime.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_get_agent_runtime/} for full documentation.
}
\keyword{internal}
