% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_get_agent_runtime_endpoint}
\alias{bedrockagentcorecontrol_get_agent_runtime_endpoint}
\title{Gets information about an Amazon Secure AgentEndpoint}
\usage{
bedrockagentcorecontrol_get_agent_runtime_endpoint(
  agentRuntimeId,
  endpointName
)
}
\arguments{
\item{agentRuntimeId}{[required] The unique identifier of the AgentCore Runtime associated with the
endpoint.}

\item{endpointName}{[required] The name of the AgentCore Runtime endpoint to retrieve.}
}
\description{
Gets information about an Amazon Secure AgentEndpoint.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_get_agent_runtime_endpoint/} for full documentation.
}
\keyword{internal}
