% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_get_browser}
\alias{bedrockagentcorecontrol_get_browser}
\title{Gets information about a custom browser}
\usage{
bedrockagentcorecontrol_get_browser(browserId)
}
\arguments{
\item{browserId}{[required] The unique identifier of the browser to retrieve.}
}
\description{
Gets information about a custom browser.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_get_browser/} for full documentation.
}
\keyword{internal}
