% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_get_browser_profile}
\alias{bedrockagentcorecontrol_get_browser_profile}
\title{Gets information about a browser profile}
\usage{
bedrockagentcorecontrol_get_browser_profile(profileId)
}
\arguments{
\item{profileId}{[required] The unique identifier of the browser profile to retrieve.}
}
\description{
Gets information about a browser profile.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_get_browser_profile/} for full documentation.
}
\keyword{internal}
