% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_get_capacity_provider}
\alias{bedrockagentcorecontrol_get_capacity_provider}
\title{Retrieves information about a capacity provider, including its status,
permissions configuration, and compute configuration}
\usage{
bedrockagentcorecontrol_get_capacity_provider(capacityProviderId)
}
\arguments{
\item{capacityProviderId}{[required] The unique identifier of the capacity provider.}
}
\description{
Retrieves information about a capacity provider, including its status, permissions configuration, and compute configuration.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_get_capacity_provider/} for full documentation.
}
\keyword{internal}
