% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_get_code_interpreter}
\alias{bedrockagentcorecontrol_get_code_interpreter}
\title{Gets information about a custom code interpreter}
\usage{
bedrockagentcorecontrol_get_code_interpreter(codeInterpreterId)
}
\arguments{
\item{codeInterpreterId}{[required] The unique identifier of the code interpreter to retrieve.}
}
\description{
Gets information about a custom code interpreter.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_get_code_interpreter/} for full documentation.
}
\keyword{internal}
