% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_get_configuration_bundle}
\alias{bedrockagentcorecontrol_get_configuration_bundle}
\title{Gets the latest version of a configuration bundle}
\usage{
bedrockagentcorecontrol_get_configuration_bundle(bundleId, branchName = NULL)
}
\arguments{
\item{bundleId}{[required] The unique identifier of the configuration bundle to retrieve.}

\item{branchName}{The branch name to get the latest version from. If not specified,
returns the latest version on the mainline branch.}
}
\description{
Gets the latest version of a configuration bundle. By default, returns the latest version on the mainline branch. Use \code{\link[=bedrockagentcorecontrol_get_configuration_bundle_version]{get_configuration_bundle_version}} to retrieve a specific historical version.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_get_configuration_bundle/} for full documentation.
}
\keyword{internal}
