% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_get_configuration_bundle_version}
\alias{bedrockagentcorecontrol_get_configuration_bundle_version}
\title{Gets a specific version of a configuration bundle by its version
identifier}
\usage{
bedrockagentcorecontrol_get_configuration_bundle_version(bundleId, versionId)
}
\arguments{
\item{bundleId}{[required] The unique identifier of the configuration bundle.}

\item{versionId}{[required] The version identifier of the configuration bundle version to retrieve.}
}
\description{
Gets a specific version of a configuration bundle by its version identifier.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_get_configuration_bundle_version/} for full documentation.
}
\keyword{internal}
