% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_get_gateway}
\alias{bedrockagentcorecontrol_get_gateway}
\title{Retrieves information about a specific Gateway}
\usage{
bedrockagentcorecontrol_get_gateway(gatewayIdentifier)
}
\arguments{
\item{gatewayIdentifier}{[required] The identifier of the gateway to retrieve.}
}
\description{
Retrieves information about a specific Gateway.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_get_gateway/} for full documentation.
}
\keyword{internal}
