% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_get_gateway_rate_limit}
\alias{bedrockagentcorecontrol_get_gateway_rate_limit}
\title{Retrieves information about a gateway rate limit}
\usage{
bedrockagentcorecontrol_get_gateway_rate_limit(gatewayIdentifier, rateLimitId)
}
\arguments{
\item{gatewayIdentifier}{[required] The unique identifier of the gateway.}

\item{rateLimitId}{[required] The unique identifier of the rate limit to retrieve.}
}
\description{
Retrieves information about a gateway rate limit.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_get_gateway_rate_limit/} for full documentation.
}
\keyword{internal}
