% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_get_gateway_target}
\alias{bedrockagentcorecontrol_get_gateway_target}
\title{Retrieves information about a specific gateway target}
\usage{
bedrockagentcorecontrol_get_gateway_target(gatewayIdentifier, targetId)
}
\arguments{
\item{gatewayIdentifier}{[required] The identifier of the gateway that contains the target.}

\item{targetId}{[required] The unique identifier of the target to retrieve.}
}
\description{
Retrieves information about a specific gateway target.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_get_gateway_target/} for full documentation.
}
\keyword{internal}
