% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_get_harness_endpoint}
\alias{bedrockagentcorecontrol_get_harness_endpoint}
\title{Operation to get a single harness endpoint}
\usage{
bedrockagentcorecontrol_get_harness_endpoint(harnessId, endpointName)
}
\arguments{
\item{harnessId}{[required] The ID of the harness that the endpoint belongs to.}

\item{endpointName}{[required] The name of the endpoint to retrieve.}
}
\description{
Operation to get a single harness endpoint.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_get_harness_endpoint/} for full documentation.
}
\keyword{internal}
