% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_get_memory}
\alias{bedrockagentcorecontrol_get_memory}
\title{Retrieve an existing Amazon Bedrock AgentCore Memory resource}
\usage{
bedrockagentcorecontrol_get_memory(memoryId, view = NULL)
}
\arguments{
\item{memoryId}{[required] The unique identifier of the memory to retrieve.}

\item{view}{The level of detail to return for the memory.}
}
\description{
Retrieve an existing Amazon Bedrock AgentCore Memory resource.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_get_memory/} for full documentation.
}
\keyword{internal}
