% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_get_oauth_2_credential_provider}
\alias{bedrockagentcorecontrol_get_oauth_2_credential_provider}
\title{Retrieves information about an OAuth2 credential provider}
\usage{
bedrockagentcorecontrol_get_oauth_2_credential_provider(name)
}
\arguments{
\item{name}{[required] The name of the OAuth2 credential provider to retrieve.}
}
\description{
Retrieves information about an OAuth2 credential provider.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_get_oauth_2_credential_provider/} for full documentation.
}
\keyword{internal}
