% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_get_online_evaluation_config}
\alias{bedrockagentcorecontrol_get_online_evaluation_config}
\title{Retrieves detailed information about an online evaluation configuration,
including its rules, data sources, evaluators, and execution status}
\usage{
bedrockagentcorecontrol_get_online_evaluation_config(onlineEvaluationConfigId)
}
\arguments{
\item{onlineEvaluationConfigId}{[required] The unique identifier of the online evaluation configuration to
retrieve.}
}
\description{
Retrieves detailed information about an online evaluation configuration, including its rules, data sources, evaluators, and execution status.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_get_online_evaluation_config/} for full documentation.
}
\keyword{internal}
