% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_get_payment_connector}
\alias{bedrockagentcorecontrol_get_payment_connector}
\title{Retrieves information about a specific payment connector}
\usage{
bedrockagentcorecontrol_get_payment_connector(
  paymentManagerId,
  paymentConnectorId
)
}
\arguments{
\item{paymentManagerId}{[required] The unique identifier of the parent payment manager.}

\item{paymentConnectorId}{[required] The unique identifier of the payment connector to retrieve.}
}
\description{
Retrieves information about a specific payment connector.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_get_payment_connector/} for full documentation.
}
\keyword{internal}
