% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_get_payment_manager}
\alias{bedrockagentcorecontrol_get_payment_manager}
\title{Retrieves information about a specific payment manager}
\usage{
bedrockagentcorecontrol_get_payment_manager(paymentManagerId)
}
\arguments{
\item{paymentManagerId}{[required] The unique identifier of the payment manager to retrieve.}
}
\description{
Retrieves information about a specific payment manager.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_get_payment_manager/} for full documentation.
}
\keyword{internal}
