% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_get_registry}
\alias{bedrockagentcorecontrol_get_registry}
\title{Retrieves information about a specific registry}
\usage{
bedrockagentcorecontrol_get_registry(registryId)
}
\arguments{
\item{registryId}{[required] The identifier of the registry to retrieve. You can specify either the
Amazon Resource Name (ARN) or the ID of the registry.}
}
\description{
Retrieves information about a specific registry.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_get_registry/} for full documentation.
}
\keyword{internal}
