% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_get_registry_record}
\alias{bedrockagentcorecontrol_get_registry_record}
\title{Retrieves information about a specific registry record}
\usage{
bedrockagentcorecontrol_get_registry_record(registryId, recordId)
}
\arguments{
\item{registryId}{[required] The identifier of the registry containing the record. You can specify
either the Amazon Resource Name (ARN) or the ID of the registry.}

\item{recordId}{[required] The identifier of the registry record to retrieve. You can specify
either the Amazon Resource Name (ARN) or the ID of the record.}
}
\description{
Retrieves information about a specific registry record.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_get_registry_record/} for full documentation.
}
\keyword{internal}
