% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_get_resource_policy}
\alias{bedrockagentcorecontrol_get_resource_policy}
\title{Retrieves the resource-based policy for a specified resource}
\usage{
bedrockagentcorecontrol_get_resource_policy(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource for which to retrieve the
resource policy.}
}
\description{
Retrieves the resource-based policy for a specified resource.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_get_resource_policy/} for full documentation.
}
\keyword{internal}
