% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_get_token_vault}
\alias{bedrockagentcorecontrol_get_token_vault}
\title{Retrieves information about a token vault}
\usage{
bedrockagentcorecontrol_get_token_vault(tokenVaultId = NULL)
}
\arguments{
\item{tokenVaultId}{The unique identifier of the token vault to retrieve.}
}
\description{
Retrieves information about a token vault.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_get_token_vault/} for full documentation.
}
\keyword{internal}
