% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_get_workload_identity}
\alias{bedrockagentcorecontrol_get_workload_identity}
\title{Retrieves information about a workload identity}
\usage{
bedrockagentcorecontrol_get_workload_identity(name)
}
\arguments{
\item{name}{[required] The name of the workload identity to retrieve.}
}
\description{
Retrieves information about a workload identity.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_get_workload_identity/} for full documentation.
}
\keyword{internal}
