% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_list_agent_runtime_versions_by_capacity_provider}
\alias{bedrockagentcorecontrol_list_agent_runtime_versions_by_capacity_provider}
\title{Lists the agent runtime versions that are associated with a capacity
provider}
\usage{
bedrockagentcorecontrol_list_agent_runtime_versions_by_capacity_provider(
  capacityProviderId,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{capacityProviderId}{[required] The unique identifier of the capacity provider.}

\item{maxResults}{The maximum number of results to return in the response. If the total number of results is greater than this value, use the token returned in the response in the \code{nextToken} field when making another request to return the next batch of results.}

\item{nextToken}{If the total number of results is greater than the \code{maxResults} value provided in the request, enter the token returned in the \code{nextToken} field in the response in this field to return the next batch of results.}
}
\description{
Lists the agent runtime versions that are associated with a capacity provider. Use this operation to identify the runtimes you must disassociate before you can delete the capacity provider. Results are paginated; use the \code{nextToken} parameter to retrieve additional results.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_list_agent_runtime_versions_by_capacity_provider/} for full documentation.
}
\keyword{internal}
