% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_list_agent_runtime_endpoints}
\alias{bedrockagentcorecontrol_list_agent_runtime_endpoints}
\title{Lists all endpoints for a specific Amazon Secure Agent}
\usage{
bedrockagentcorecontrol_list_agent_runtime_endpoints(
  agentRuntimeId,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{agentRuntimeId}{[required] The unique identifier of the AgentCore Runtime to list endpoints for.}

\item{maxResults}{The maximum number of results to return in the response.}

\item{nextToken}{A token to retrieve the next page of results.}
}
\description{
Lists all endpoints for a specific Amazon Secure Agent.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_list_agent_runtime_endpoints/} for full documentation.
}
\keyword{internal}
