% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_list_browsers}
\alias{bedrockagentcorecontrol_list_browsers}
\title{Lists all custom browsers in your account}
\usage{
bedrockagentcorecontrol_list_browsers(
  maxResults = NULL,
  nextToken = NULL,
  type = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of results to return in a single call. The default
value is 10. The maximum value is 50.}

\item{nextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{type}{The type of browsers to list. If not specified, all browser types are
returned.}
}
\description{
Lists all custom browsers in your account.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_list_browsers/} for full documentation.
}
\keyword{internal}
