% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_list_capacity_providers}
\alias{bedrockagentcorecontrol_list_capacity_providers}
\title{Lists the capacity providers in your account and returns summary
information for each one}
\usage{
bedrockagentcorecontrol_list_capacity_providers(
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of results to return in the response. If the total number of results is greater than this value, use the token returned in the response in the \code{nextToken} field when making another request to return the next batch of results.}

\item{nextToken}{If the total number of results is greater than the \code{maxResults} value provided in the request, enter the token returned in the \code{nextToken} field in the response in this field to return the next batch of results.}
}
\description{
Lists the capacity providers in your account and returns summary information for each one. To retrieve the full configuration for a specific capacity provider, use \code{\link[=bedrockagentcorecontrol_get_capacity_provider]{get_capacity_provider}}. Results are paginated; use the \code{nextToken} parameter to retrieve additional results.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_list_capacity_providers/} for full documentation.
}
\keyword{internal}
