% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_list_consent_portals}
\alias{bedrockagentcorecontrol_list_consent_portals}
\title{Lists all of the consent portals in your account}
\usage{
bedrockagentcorecontrol_list_consent_portals(
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of consent portals to return in a single call.}

\item{nextToken}{A token to retrieve the next page of results. Use the value returned in a previous response to request the next page.}
}
\description{
Lists all of the consent portals in your account.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_list_consent_portals/} for full documentation.
}
\keyword{internal}
