% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_list_dataset_versions}
\alias{bedrockagentcorecontrol_list_dataset_versions}
\title{Lists all published versions of a dataset, sorted by version number
descending (newest first)}
\usage{
bedrockagentcorecontrol_list_dataset_versions(
  datasetId,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{datasetId}{[required] The unique identifier of the dataset.}

\item{nextToken}{The token for the next page of results.}

\item{maxResults}{The maximum number of versions to return per page.}
}
\description{
Lists all published versions of a dataset, sorted by version number descending (newest first). Does not include the DRAFT working copy.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_list_dataset_versions/} for full documentation.
}
\keyword{internal}
