% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_list_datasets}
\alias{bedrockagentcorecontrol_list_datasets}
\title{Lists all datasets in the caller's account, paginated}
\usage{
bedrockagentcorecontrol_list_datasets(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{The token for the next page of results.}

\item{maxResults}{The maximum number of datasets to return per page.}
}
\description{
Lists all datasets in the caller's account, paginated.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_list_datasets/} for full documentation.
}
\keyword{internal}
