% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_list_evaluators}
\alias{bedrockagentcorecontrol_list_evaluators}
\title{Lists all available evaluators, including both builtin evaluators
provided by the service and custom evaluators created by the user}
\usage{
bedrockagentcorecontrol_list_evaluators(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{The pagination token from a previous request to retrieve the next page
of results.}

\item{maxResults}{The maximum number of evaluators to return in a single response.}
}
\description{
Lists all available evaluators, including both builtin evaluators provided by the service and custom evaluators created by the user.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_list_evaluators/} for full documentation.
}
\keyword{internal}
