% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_list_gateway_rate_limits}
\alias{bedrockagentcorecontrol_list_gateway_rate_limits}
\title{Lists all rate limits for a gateway}
\usage{
bedrockagentcorecontrol_list_gateway_rate_limits(
  gatewayIdentifier,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{gatewayIdentifier}{[required] The unique identifier of the gateway.}

\item{maxResults}{The maximum number of results to return in the response. If the total number of results is greater than this value, use the token returned in the response in the \code{nextToken} field when making another request to return the next batch of results.}

\item{nextToken}{The token to use to retrieve the next page of results. Use the value returned in a previous \code{\link[=bedrockagentcorecontrol_list_gateway_rate_limits]{list_gateway_rate_limits}} response.}
}
\description{
Lists all rate limits for a gateway. Results are paginated. Use the \code{nextToken} parameter to retrieve additional results.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_list_gateway_rate_limits/} for full documentation.
}
\keyword{internal}
