% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_list_gateway_rules}
\alias{bedrockagentcorecontrol_list_gateway_rules}
\title{Lists all rules for a gateway}
\usage{
bedrockagentcorecontrol_list_gateway_rules(
  gatewayIdentifier,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{gatewayIdentifier}{[required] The identifier of the gateway to list rules for.}

\item{maxResults}{The maximum number of results to return in the response. If the total
number of results is greater than this value, use the token returned in
the response in the \code{nextToken} field when making another request to
return the next batch of results.}

\item{nextToken}{The pagination token from a previous request.}
}
\description{
Lists all rules for a gateway.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_list_gateway_rules/} for full documentation.
}
\keyword{internal}
