% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_list_gateway_targets}
\alias{bedrockagentcorecontrol_list_gateway_targets}
\title{Lists all targets for a specific gateway}
\usage{
bedrockagentcorecontrol_list_gateway_targets(
  gatewayIdentifier,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{gatewayIdentifier}{[required] The identifier of the gateway to list targets for.}

\item{maxResults}{The maximum number of results to return in the response. If the total
number of results is greater than this value, use the token returned in
the response in the \code{nextToken} field when making another request to
return the next batch of results.}

\item{nextToken}{If the total number of results is greater than the \code{maxResults} value
provided in the request, enter the token returned in the \code{nextToken}
field in the response in this field to return the next batch of results.}
}
\description{
Lists all targets for a specific gateway.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_list_gateway_targets/} for full documentation.
}
\keyword{internal}
