% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_list_harness_endpoints}
\alias{bedrockagentcorecontrol_list_harness_endpoints}
\title{Operation to list the endpoints of a harness}
\usage{
bedrockagentcorecontrol_list_harness_endpoints(
  harnessId,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{harnessId}{[required] The ID of the harness whose endpoints are listed.}

\item{maxResults}{The maximum number of results to return in a single call.}

\item{nextToken}{The token for the next set of results.}
}
\description{
Operation to list the endpoints of a harness.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_list_harness_endpoints/} for full documentation.
}
\keyword{internal}
