% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_list_memories}
\alias{bedrockagentcorecontrol_list_memories}
\title{Lists the available Amazon Bedrock AgentCore Memory resources in the
current Amazon Web Services Region}
\usage{
bedrockagentcorecontrol_list_memories(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of results to return in a single call. The default
value is 10. The maximum value is 50.}

\item{nextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}
}
\description{
Lists the available Amazon Bedrock AgentCore Memory resources in the current Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_list_memories/} for full documentation.
}
\keyword{internal}
