% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_list_online_evaluation_configs}
\alias{bedrockagentcorecontrol_list_online_evaluation_configs}
\title{Lists all online evaluation configurations in the account, providing
summary information about each configuration's status and settings}
\usage{
bedrockagentcorecontrol_list_online_evaluation_configs(
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{nextToken}{The pagination token from a previous request to retrieve the next page
of results.}

\item{maxResults}{The maximum number of online evaluation configurations to return in a
single response.}
}
\description{
Lists all online evaluation configurations in the account, providing summary information about each configuration's status and settings.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_list_online_evaluation_configs/} for full documentation.
}
\keyword{internal}
