% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_list_policy_engines}
\alias{bedrockagentcorecontrol_list_policy_engines}
\title{Retrieves a list of policy engines within the AgentCore Policy system}
\usage{
bedrockagentcorecontrol_list_policy_engines(
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{nextToken}{A pagination token returned from a previous
\code{\link[=bedrockagentcorecontrol_list_policy_engines]{list_policy_engines}}
call. Use this token to retrieve the next page of results when the
response is paginated.}

\item{maxResults}{The maximum number of policy engines to return in a single response. If
not specified, the default is 10 policy engines per page, with a maximum
of 100 per page.}
}
\description{
Retrieves a list of policy engines within the AgentCore Policy system. This operation supports pagination to help administrators discover and manage policy engines across their account. Each policy engine serves as a container for related policies.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_list_policy_engines/} for full documentation.
}
\keyword{internal}
