% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_list_tags_for_resource}
\alias{bedrockagentcorecontrol_list_tags_for_resource}
\title{Lists the tags associated with the specified resource}
\usage{
bedrockagentcorecontrol_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource for which you want to
list tags.}
}
\description{
Lists the tags associated with the specified resource.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
