% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_list_workload_identities}
\alias{bedrockagentcorecontrol_list_workload_identities}
\title{Lists all workload identities in your account}
\usage{
bedrockagentcorecontrol_list_workload_identities(
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{nextToken}{Pagination token.}

\item{maxResults}{Maximum number of results to return.}
}
\description{
Lists all workload identities in your account.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_list_workload_identities/} for full documentation.
}
\keyword{internal}
