% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_put_resource_policy}
\alias{bedrockagentcorecontrol_put_resource_policy}
\title{Creates or updates a resource-based policy for a resource with the
specified resourceArn}
\usage{
bedrockagentcorecontrol_put_resource_policy(resourceArn, policy)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource for which to create or
update the resource policy.}

\item{policy}{[required] The resource policy to create or update.}
}
\description{
Creates or updates a resource-based policy for a resource with the specified resourceArn.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_put_resource_policy/} for full documentation.
}
\keyword{internal}
