% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_rotate_payment_connector_credentials}
\alias{bedrockagentcorecontrol_rotate_payment_connector_credentials}
\title{Replaces the service-managed credentials of a payment connector with
newly issued credentials}
\usage{
bedrockagentcorecontrol_rotate_payment_connector_credentials(
  paymentManagerId,
  paymentConnectorId,
  credentialsToRotate,
  clientToken = NULL
)
}
\arguments{
\item{paymentManagerId}{[required] The unique identifier of the parent payment manager.}

\item{paymentConnectorId}{[required] The unique identifier of the payment connector whose credentials you want to rotate.}

\item{credentialsToRotate}{[required] The credentials to rotate. Specify the member that matches the payment connector's \code{type}. Each credential that you select is rotated independently.}

\item{clientToken}{A unique, case-sensitive identifier to ensure that the API request completes no more than one time. If you don't specify this field, a value is randomly generated for you. If this token matches a previous request, the service ignores the request, but doesn't return an error. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring idempotency}.}
}
\description{
Replaces the service-managed credentials of a payment connector with newly issued credentials.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_rotate_payment_connector_credentials/} for full documentation.
}
\keyword{internal}
