% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_set_token_vault_cmk}
\alias{bedrockagentcorecontrol_set_token_vault_cmk}
\title{Sets the customer master key (CMK) for a token vault}
\usage{
bedrockagentcorecontrol_set_token_vault_cmk(
  tokenVaultId = NULL,
  kmsConfiguration
)
}
\arguments{
\item{tokenVaultId}{The unique identifier of the token vault to update.}

\item{kmsConfiguration}{[required] The KMS configuration for the token vault, including the key type and
KMS key ARN.}
}
\description{
Sets the customer master key (CMK) for a token vault.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_set_token_vault_cmk/} for full documentation.
}
\keyword{internal}
