% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_submit_registry_record_for_approval}
\alias{bedrockagentcorecontrol_submit_registry_record_for_approval}
\title{Submits a registry record for approval}
\usage{
bedrockagentcorecontrol_submit_registry_record_for_approval(
  registryId,
  recordId
)
}
\arguments{
\item{registryId}{[required] The identifier of the registry containing the record. You can specify
either the Amazon Resource Name (ARN) or the ID of the registry.}

\item{recordId}{[required] The identifier of the registry record to submit for approval. You can
specify either the Amazon Resource Name (ARN) or the ID of the record.}
}
\description{
Submits a registry record for approval. This transitions the record from \code{DRAFT} status to \code{PENDING_APPROVAL} status. If the registry has auto-approval enabled, the record is automatically approved.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_submit_registry_record_for_approval/} for full documentation.
}
\keyword{internal}
