% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_synchronize_gateway_targets}
\alias{bedrockagentcorecontrol_synchronize_gateway_targets}
\title{Synchronizes the gateway targets by fetching the latest tool definitions
from the target endpoints}
\usage{
bedrockagentcorecontrol_synchronize_gateway_targets(
  gatewayIdentifier,
  targetIdList
)
}
\arguments{
\item{gatewayIdentifier}{[required] The gateway Identifier.}

\item{targetIdList}{[required] The target ID list.}
}
\description{
Synchronizes the gateway targets by fetching the latest tool definitions from the target endpoints.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_synchronize_gateway_targets/} for full documentation.
}
\keyword{internal}
