% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_tag_resource}
\alias{bedrockagentcorecontrol_tag_resource}
\title{Associates the specified tags to a resource with the specified
resourceArn}
\usage{
bedrockagentcorecontrol_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to tag.}

\item{tags}{[required] The tags to add to the resource. A tag is a key-value pair.}
}
\description{
Associates the specified tags to a resource with the specified resourceArn. If existing tags on a resource are not specified in the request parameters, they are not changed. When a resource is deleted, the tags associated with that resource are also deleted.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_tag_resource/} for full documentation.
}
\keyword{internal}
