% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_untag_resource}
\alias{bedrockagentcorecontrol_untag_resource}
\title{Removes the specified tags from the specified resource}
\usage{
bedrockagentcorecontrol_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to untag.}

\item{tagKeys}{[required] The tag keys of the tags to remove from the resource.}
}
\description{
Removes the specified tags from the specified resource.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_untag_resource/} for full documentation.
}
\keyword{internal}
