% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_update_agent_runtime_endpoint}
\alias{bedrockagentcorecontrol_update_agent_runtime_endpoint}
\title{Updates an existing Amazon Bedrock AgentCore Runtime endpoint}
\usage{
bedrockagentcorecontrol_update_agent_runtime_endpoint(
  agentRuntimeId,
  endpointName,
  agentRuntimeVersion = NULL,
  description = NULL,
  clientToken = NULL
)
}
\arguments{
\item{agentRuntimeId}{[required] The unique identifier of the AgentCore Runtime associated with the
endpoint.}

\item{endpointName}{[required] The name of the AgentCore Runtime endpoint to update.}

\item{agentRuntimeVersion}{The updated version of the AgentCore Runtime for the endpoint.}

\item{description}{The updated description of the AgentCore Runtime endpoint.}

\item{clientToken}{A unique, case-sensitive identifier to ensure idempotency of the
request.}
}
\description{
Updates an existing Amazon Bedrock AgentCore Runtime endpoint.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_update_agent_runtime_endpoint/} for full documentation.
}
\keyword{internal}
