% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_update_capacity_provider}
\alias{bedrockagentcorecontrol_update_capacity_provider}
\title{Updates a capacity provider}
\usage{
bedrockagentcorecontrol_update_capacity_provider(
  capacityProviderId,
  description = NULL,
  clientToken = NULL
)
}
\arguments{
\item{capacityProviderId}{[required] The unique identifier of the capacity provider to update.}

\item{description}{The updated description of the capacity provider.}

\item{clientToken}{A unique, case-sensitive identifier to ensure that the API request completes no more than one time. If you don't specify this field, a value is randomly generated for you. If this token matches a previous request, the service ignores the request, but doesn't return an error. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring idempotency}.}
}
\description{
Updates a capacity provider. Only the description can be changed. To change other configuration, such as instance types, networking, or storage, create a new capacity provider.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_update_capacity_provider/} for full documentation.
}
\keyword{internal}
