% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_update_consent_portal}
\alias{bedrockagentcorecontrol_update_consent_portal}
\title{Updates an existing consent portal}
\usage{
bedrockagentcorecontrol_update_consent_portal(
  consentPortalIdentifier,
  executionRoleArn = NULL,
  idpConfig = NULL,
  description = NULL
)
}
\arguments{
\item{consentPortalIdentifier}{[required] The identifier of the consent portal. You can specify either the consent portal ID or its Amazon Resource Name (ARN).}

\item{executionRoleArn}{The Amazon Resource Name (ARN) of the IAM role that the consent portal assumes to access the resources defined in its sources.}

\item{idpConfig}{The identity provider configuration that the consent portal uses to authenticate end users.}

\item{description}{The description of the consent portal.}
}
\description{
Updates an existing consent portal.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_update_consent_portal/} for full documentation.
}
\keyword{internal}
