% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_update_gateway}
\alias{bedrockagentcorecontrol_update_gateway}
\title{Updates an existing gateway}
\usage{
bedrockagentcorecontrol_update_gateway(
  gatewayIdentifier,
  name,
  description = NULL,
  roleArn,
  protocolType = NULL,
  protocolConfiguration = NULL,
  authorizerType,
  authorizerConfiguration = NULL,
  kmsKeyArn = NULL,
  customTransformConfiguration = NULL,
  interceptorConfigurations = NULL,
  policyEngineConfiguration = NULL,
  exceptionLevel = NULL,
  wafConfiguration = NULL
)
}
\arguments{
\item{gatewayIdentifier}{[required] The identifier of the gateway to update.}

\item{name}{[required] The name of the gateway. This name must be the same as the one when the gateway was created.}

\item{description}{The updated description for the gateway.}

\item{roleArn}{[required] The updated IAM role ARN that provides permissions for the gateway.}

\item{protocolType}{The updated protocol type for the gateway.}

\item{protocolConfiguration}{The configuration for a gateway protocol. This structure defines how the gateway communicates with external services.}

\item{authorizerType}{[required] The updated authorizer type for the gateway.}

\item{authorizerConfiguration}{The updated authorizer configuration for the gateway.}

\item{kmsKeyArn}{The updated ARN of the KMS key used to encrypt the gateway.}

\item{customTransformConfiguration}{The updated custom transformation configuration for the gateway. This configuration defines how the gateway transforms requests and responses.}

\item{interceptorConfigurations}{The updated interceptor configurations for the gateway.}

\item{policyEngineConfiguration}{The updated policy engine configuration for the gateway. A policy engine is a collection of policies that evaluates and authorizes agent tool calls. When associated with a gateway, the policy engine intercepts all agent requests and determines whether to allow or deny each action based on the defined policies.}

\item{exceptionLevel}{The level of detail in error messages returned when invoking the gateway.
\itemize{
\item If the value is \code{DEBUG}, granular exception messages are returned to help a user debug the gateway.
\item If the value is omitted, a generic error message is returned to the end user.
}}

\item{wafConfiguration}{The updated Amazon Web Services WAF configuration for the gateway.}
}
\description{
Updates an existing gateway.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_update_gateway/} for full documentation.
}
\keyword{internal}
