% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_update_gateway_target}
\alias{bedrockagentcorecontrol_update_gateway_target}
\title{Updates an existing gateway target}
\usage{
bedrockagentcorecontrol_update_gateway_target(
  gatewayIdentifier,
  targetId,
  name = NULL,
  description = NULL,
  targetConfiguration,
  credentialProviderConfigurations = NULL,
  metadataConfiguration = NULL,
  privateEndpoint = NULL,
  certificateConfigurations = NULL
)
}
\arguments{
\item{gatewayIdentifier}{[required] The unique identifier of the gateway associated with the target.}

\item{targetId}{[required] The unique identifier of the gateway target to update.}

\item{name}{The updated name for the gateway target.}

\item{description}{The updated description for the gateway target.}

\item{targetConfiguration}{[required] The configuration for a gateway target. This structure defines how the gateway connects to and interacts with the target endpoint.}

\item{credentialProviderConfigurations}{The updated credential provider configurations for the gateway target.}

\item{metadataConfiguration}{Configuration for HTTP header and query parameter propagation to the gateway target.}

\item{privateEndpoint}{The private endpoint configuration for the gateway target. Use this to connect the gateway to private resources in your VPC.}

\item{certificateConfigurations}{The private certificate authority (CA) configurations for the gateway target. Use this to have the gateway trust a private CA when it establishes TLS connections to the target endpoint. Provide each certificate by reference to an Amazon S3 object or an Amazon Web Services Secrets Manager secret. You can specify only one certificate authority configuration in this list. To remove a previously configured certificate authority, omit this field on update.}
}
\description{
Updates an existing gateway target.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_update_gateway_target/} for full documentation.
}
\keyword{internal}
