% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_update_harness_endpoint}
\alias{bedrockagentcorecontrol_update_harness_endpoint}
\title{Operation to update a harness endpoint}
\usage{
bedrockagentcorecontrol_update_harness_endpoint(
  harnessId,
  endpointName,
  targetVersion = NULL,
  description = NULL,
  clientToken = NULL
)
}
\arguments{
\item{harnessId}{[required] The ID of the harness that the endpoint belongs to.}

\item{endpointName}{[required] The name of the endpoint to update.}

\item{targetVersion}{The harness version that the endpoint points to. If not specified, the existing value is retained.}

\item{description}{A description of the endpoint. If not specified, the existing value is retained.}

\item{clientToken}{A unique, case-sensitive identifier to ensure idempotency of the request.}
}
\description{
Operation to update a harness endpoint.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_update_harness_endpoint/} for full documentation.
}
\keyword{internal}
