% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_update_memory}
\alias{bedrockagentcorecontrol_update_memory}
\title{Update an Amazon Bedrock AgentCore Memory resource memory}
\usage{
bedrockagentcorecontrol_update_memory(
  clientToken = NULL,
  memoryId,
  description = NULL,
  eventExpiryDuration = NULL,
  memoryExecutionRoleArn = NULL,
  memoryStrategies = NULL,
  addIndexedKeys = NULL,
  namespaceKeys = NULL,
  streamDeliveryResources = NULL
)
}
\arguments{
\item{clientToken}{A client token is used for keeping track of idempotent requests. It can contain a session id which can be around 250 chars, combined with a unique AWS identifier.}

\item{memoryId}{[required] The unique identifier of the memory to update.}

\item{description}{The updated description of the AgentCore Memory resource.}

\item{eventExpiryDuration}{The number of days after which memory events will expire, between 7 and 365 days.}

\item{memoryExecutionRoleArn}{The ARN of the IAM role that provides permissions for the AgentCore Memory resource.}

\item{memoryStrategies}{The memory strategies to add, modify, or delete.}

\item{addIndexedKeys}{Additional metadata keys to index. Previously indexed keys cannot be removed.}

\item{namespaceKeys}{The namespace variable key definitions with validation rules for this memory. This value fully replaces the existing set — any key you omit is removed. Any referenced \code{namespaceKey} omission will throw ValidationException.}

\item{streamDeliveryResources}{Configuration for streaming memory record data to external resources.}
}
\description{
Update an Amazon Bedrock AgentCore Memory resource memory.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_update_memory/} for full documentation.
}
\keyword{internal}
