% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_update_oauth_2_credential_provider}
\alias{bedrockagentcorecontrol_update_oauth_2_credential_provider}
\title{Updates an existing OAuth2 credential provider}
\usage{
bedrockagentcorecontrol_update_oauth_2_credential_provider(
  name,
  credentialProviderVendor,
  oauth2ProviderConfigInput
)
}
\arguments{
\item{name}{[required] The name of the OAuth2 credential provider to update.}

\item{credentialProviderVendor}{[required] The vendor of the OAuth2 credential provider.}

\item{oauth2ProviderConfigInput}{[required] The configuration input for the OAuth2 provider.}
}
\description{
Updates an existing OAuth2 credential provider.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_update_oauth_2_credential_provider/} for full documentation.
}
\keyword{internal}
