% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_update_payment_connector}
\alias{bedrockagentcorecontrol_update_payment_connector}
\title{Updates an existing payment connector}
\usage{
bedrockagentcorecontrol_update_payment_connector(
  paymentManagerId,
  paymentConnectorId,
  description = NULL,
  type = NULL,
  credentialProviderConfigurations = NULL,
  clientToken = NULL
)
}
\arguments{
\item{paymentManagerId}{[required] The unique identifier of the parent payment manager.}

\item{paymentConnectorId}{[required] The unique identifier of the payment connector to update.}

\item{description}{The updated description of the payment connector.}

\item{type}{The updated type of the payment connector.}

\item{credentialProviderConfigurations}{The updated credential provider configurations for the payment
connector.}

\item{clientToken}{A unique, case-sensitive identifier to ensure that the API request
completes no more than one time. If you don't specify this field, a
value is randomly generated for you. If this token matches a previous
request, the service ignores the request, but doesn't return an error.
For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring idempotency}.}
}
\description{
Updates an existing payment connector. This operation uses PATCH semantics, so you only need to specify the fields you want to change.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_update_payment_connector/} for full documentation.
}
\keyword{internal}
