% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_update_payment_credential_provider}
\alias{bedrockagentcorecontrol_update_payment_credential_provider}
\title{Updates an existing payment credential provider with new authentication
credentials}
\usage{
bedrockagentcorecontrol_update_payment_credential_provider(
  name,
  credentialProviderVendor,
  providerConfigurationInput
)
}
\arguments{
\item{name}{[required] The name of the payment credential provider to update.}

\item{credentialProviderVendor}{[required]}

\item{providerConfigurationInput}{[required] Configuration specific to the vendor, including API credentials}
}
\description{
Updates an existing payment credential provider with new authentication credentials.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_update_payment_credential_provider/} for full documentation.
}
\keyword{internal}
