% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_update_policy_engine}
\alias{bedrockagentcorecontrol_update_policy_engine}
\title{Updates an existing policy engine within the AgentCore Policy system}
\usage{
bedrockagentcorecontrol_update_policy_engine(
  policyEngineId,
  description = NULL
)
}
\arguments{
\item{policyEngineId}{[required] The unique identifier of the policy engine to be updated.}

\item{description}{The new description for the policy engine.}
}
\description{
Updates an existing policy engine within the AgentCore Policy system. This operation allows modification of the policy engine description while maintaining its identity. This is an asynchronous operation. Use the \code{\link[=bedrockagentcorecontrol_get_policy_engine]{get_policy_engine}} operation to poll the \code{status} field to track completion.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_update_policy_engine/} for full documentation.
}
\keyword{internal}
