% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_update_registry_record_status}
\alias{bedrockagentcorecontrol_update_registry_record_status}
\title{Updates the status of a registry record}
\usage{
bedrockagentcorecontrol_update_registry_record_status(
  registryId,
  recordId,
  status,
  statusReason
)
}
\arguments{
\item{registryId}{[required] The identifier of the registry containing the record. You can specify
either the Amazon Resource Name (ARN) or the ID of the registry.}

\item{recordId}{[required] The identifier of the registry record to update the status for. You can
specify either the Amazon Resource Name (ARN) or the ID of the record.}

\item{status}{[required] The target status for the registry record.}

\item{statusReason}{[required] The reason for the status change, such as why the record was approved or
rejected.}
}
\description{
Updates the status of a registry record. Use this operation to approve, reject, or deprecate a registry record.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_update_registry_record_status/} for full documentation.
}
\keyword{internal}
