% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_update_workload_identity}
\alias{bedrockagentcorecontrol_update_workload_identity}
\title{Updates an existing workload identity}
\usage{
bedrockagentcorecontrol_update_workload_identity(
  name,
  allowedResourceOauth2ReturnUrls = NULL
)
}
\arguments{
\item{name}{[required] The name of the workload identity to update.}

\item{allowedResourceOauth2ReturnUrls}{The new list of allowed OAuth2 return URLs for resources associated with
this workload identity. This list replaces the existing list.}
}
\description{
Updates an existing workload identity.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_update_workload_identity/} for full documentation.
}
\keyword{internal}
