% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentruntime_operations.R
\name{bedrockagentruntime_delete_agent_memory}
\alias{bedrockagentruntime_delete_agent_memory}
\title{Deletes memory from the specified memory identifier}
\usage{
bedrockagentruntime_delete_agent_memory(
  agentAliasId,
  agentId,
  memoryId = NULL,
  sessionId = NULL
)
}
\arguments{
\item{agentAliasId}{[required] The unique identifier of an alias of an agent.}

\item{agentId}{[required] The unique identifier of the agent to which the alias belongs.}

\item{memoryId}{The unique identifier of the memory.}

\item{sessionId}{The unique session identifier of the memory.}
}
\description{
Deletes memory from the specified memory identifier.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentruntime_delete_agent_memory/} for full documentation.
}
\keyword{internal}
