% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentruntime_operations.R
\name{bedrockagentruntime_get_agent_memory}
\alias{bedrockagentruntime_get_agent_memory}
\title{Gets the sessions stored in the memory of the agent}
\usage{
bedrockagentruntime_get_agent_memory(
  agentAliasId,
  agentId,
  maxItems = NULL,
  memoryId,
  memoryType,
  nextToken = NULL
)
}
\arguments{
\item{agentAliasId}{[required] The unique identifier of an alias of an agent.}

\item{agentId}{[required] The unique identifier of the agent to which the alias belongs.}

\item{maxItems}{The maximum number of items to return in the response. If the total
number of results is greater than this value, use the token returned in
the response in the \code{nextToken} field when making another request to
return the next batch of results.}

\item{memoryId}{[required] The unique identifier of the memory.}

\item{memoryType}{[required] The type of memory.}

\item{nextToken}{If the total number of results is greater than the maxItems value
provided in the request, enter the token returned in the \code{nextToken}
field in the response in this field to return the next batch of results.}
}
\description{
Gets the sessions stored in the memory of the agent.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentruntime_get_agent_memory/} for full documentation.
}
\keyword{internal}
