% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentruntime_operations.R
\name{bedrockagentruntime_get_execution_flow_snapshot}
\alias{bedrockagentruntime_get_execution_flow_snapshot}
\title{Retrieves the flow definition snapshot used for a flow execution}
\usage{
bedrockagentruntime_get_execution_flow_snapshot(
  executionIdentifier,
  flowAliasIdentifier,
  flowIdentifier
)
}
\arguments{
\item{executionIdentifier}{[required] The unique identifier of the flow execution.}

\item{flowAliasIdentifier}{[required] The unique identifier of the flow alias used for the flow execution.}

\item{flowIdentifier}{[required] The unique identifier of the flow.}
}
\description{
Retrieves the flow definition snapshot used for a flow execution. The snapshot represents the flow metadata and definition as it existed at the time the execution was started. Note that even if the flow is edited after an execution starts, the snapshot connected to the execution remains unchanged.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentruntime_get_execution_flow_snapshot/} for full documentation.
}
\keyword{internal}
