% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentruntime_operations.R
\name{bedrockagentruntime_get_flow_execution}
\alias{bedrockagentruntime_get_flow_execution}
\title{Retrieves details about a specific flow execution, including its status,
start and end times, and any errors that occurred during execution}
\usage{
bedrockagentruntime_get_flow_execution(
  executionIdentifier,
  flowAliasIdentifier,
  flowIdentifier
)
}
\arguments{
\item{executionIdentifier}{[required] The unique identifier of the flow execution to retrieve.}

\item{flowAliasIdentifier}{[required] The unique identifier of the flow alias used for the execution.}

\item{flowIdentifier}{[required] The unique identifier of the flow.}
}
\description{
Retrieves details about a specific flow execution, including its status, start and end times, and any errors that occurred during execution.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentruntime_get_flow_execution/} for full documentation.
}
\keyword{internal}
