% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentruntime_operations.R
\name{bedrockagentruntime_get_ingested_document_acl}
\alias{bedrockagentruntime_get_ingested_document_acl}
\title{Retrieves the ingested access control list (ACL) for a specific document
in a knowledge base}
\usage{
bedrockagentruntime_get_ingested_document_acl(
  dataSourceId,
  documentId,
  knowledgeBaseId
)
}
\arguments{
\item{dataSourceId}{[required] The unique identifier of the data source that contains the document.}

\item{documentId}{[required] The unique identifier of the document to retrieve the ingested access control list (ACL) for.}

\item{knowledgeBaseId}{[required] The unique identifier of the knowledge base that contains the document.}
}
\description{
Retrieves the ingested access control list (ACL) for a specific document in a knowledge base. Use this operation to inspect the allow and deny lists that were ingested for a document to troubleshoot access control issues. To use this operation, you must have the \code{bedrock:GetIngestedDocumentAcl} permission.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentruntime_get_ingested_document_acl/} for full documentation.
}
\keyword{internal}
