% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentruntime_operations.R
\name{bedrockagentruntime_get_session}
\alias{bedrockagentruntime_get_session}
\title{Retrieves details about a specific session}
\usage{
bedrockagentruntime_get_session(sessionIdentifier)
}
\arguments{
\item{sessionIdentifier}{[required] A unique identifier for the session to retrieve. You can specify either
the session's \code{sessionId} or its Amazon Resource Name (ARN).}
}
\description{
Retrieves details about a specific session. For more information about sessions, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/sessions.html}{Store and retrieve conversation history and context with Amazon Bedrock sessions}.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentruntime_get_session/} for full documentation.
}
\keyword{internal}
