% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentruntime_operations.R
\name{bedrockagentruntime_invoke_agent}
\alias{bedrockagentruntime_invoke_agent}
\title{Amazon Bedrock Agents (now Amazon Bedrock Agents Classic) is no longer
open to new customers}
\usage{
bedrockagentruntime_invoke_agent(
  agentAliasId,
  agentId,
  bedrockModelConfigurations = NULL,
  enableTrace = NULL,
  endSession = NULL,
  inputText = NULL,
  memoryId = NULL,
  promptCreationConfigurations = NULL,
  sessionId,
  sessionState = NULL,
  sourceArn = NULL,
  streamingConfigurations = NULL
)
}
\arguments{
\item{agentAliasId}{[required] The alias of the agent to use.}

\item{agentId}{[required] The unique identifier of the agent to use.}

\item{bedrockModelConfigurations}{Model performance settings for the request.}

\item{enableTrace}{Specifies whether to turn on the trace or not to track the agent's reasoning process. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/agents-test.html#trace-events}{Trace enablement}.}

\item{endSession}{Specifies whether to end the session with the agent or not.}

\item{inputText}{The prompt text to send the agent.

If you include \code{returnControlInvocationResults} in the \code{sessionState} field, the \code{inputText} field will be ignored.}

\item{memoryId}{The unique identifier of the agent memory.}

\item{promptCreationConfigurations}{Specifies parameters that control how the service populates the agent prompt for an \code{\link[=bedrockagentruntime_invoke_agent]{invoke_agent}} request. You can control which aspects of previous invocations in the same agent session the service uses to populate the agent prompt. This gives you more granular control over the contextual history that is used to process the current request.}

\item{sessionId}{[required] The unique identifier of the session. Use the same value across requests to continue the same conversation.}

\item{sessionState}{Contains parameters that specify various attributes of the session. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/agents-session-state.html}{Control session context}.

If you include \code{returnControlInvocationResults} in the \code{sessionState} field, the \code{inputText} field will be ignored.}

\item{sourceArn}{The ARN of the resource making the request.}

\item{streamingConfigurations}{Specifies the configurations for streaming.

To use agent streaming, you need permissions to perform the \code{bedrock:InvokeModelWithResponseStream} action.}
}
\description{
Amazon Bedrock Agents (now Amazon Bedrock Agents Classic) is no longer open to new customers. For capabilities similar to Bedrock Agents Classic, explore Amazon Bedrock AgentCore. Existing customers can continue to use the service as normal. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/agents-classic-maintenance-mode.html}{Amazon Bedrock Agents Classic availability change}.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentruntime_invoke_agent/} for full documentation.
}
\keyword{internal}
