% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentruntime_operations.R
\name{bedrockagentruntime_list_tags_for_resource}
\alias{bedrockagentruntime_list_tags_for_resource}
\title{List all the tags for the resource you specify}
\usage{
bedrockagentruntime_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource for which to list tags.}
}
\description{
List all the tags for the resource you specify.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentruntime_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
