% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentruntime_operations.R
\name{bedrockagentruntime_retrieve}
\alias{bedrockagentruntime_retrieve}
\title{Queries a knowledge base and retrieves information from it}
\usage{
bedrockagentruntime_retrieve(
  guardrailConfiguration = NULL,
  knowledgeBaseId,
  nextToken = NULL,
  retrievalConfiguration = NULL,
  retrievalQuery,
  userContext = NULL
)
}
\arguments{
\item{guardrailConfiguration}{Guardrail settings.}

\item{knowledgeBaseId}{[required] The unique identifier of the knowledge base to query.}

\item{nextToken}{If there are more results than can fit in the response, the response returns a \code{nextToken}. Use this token in the \code{nextToken} field of another request to retrieve the next batch of results.}

\item{retrievalConfiguration}{Contains configurations for the knowledge base query and retrieval process. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/kb-test-config.html}{Query configurations}.}

\item{retrievalQuery}{[required] Contains the query to send the knowledge base.}

\item{userContext}{Contains information about the user making the request. This is used for access control filtering to ensure that retrieval results only include documents the user is authorized to access.}
}
\description{
Queries a knowledge base and retrieves information from it.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentruntime_retrieve/} for full documentation.
}
\keyword{internal}
