% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentruntime_operations.R
\name{bedrockagentruntime_start_flow_execution}
\alias{bedrockagentruntime_start_flow_execution}
\title{Starts an execution of an Amazon Bedrock flow}
\usage{
bedrockagentruntime_start_flow_execution(
  flowAliasIdentifier,
  flowExecutionName = NULL,
  flowIdentifier,
  inputs,
  modelPerformanceConfiguration = NULL
)
}
\arguments{
\item{flowAliasIdentifier}{[required] The unique identifier of the flow alias to use for the flow execution.}

\item{flowExecutionName}{The unique name for the flow execution. If you don't provide one, a
system-generated name is used.}

\item{flowIdentifier}{[required] The unique identifier of the flow to execute.}

\item{inputs}{[required] The input data required for the flow execution. This must match the
input schema defined in the flow.}

\item{modelPerformanceConfiguration}{The performance settings for the foundation model used in the flow
execution.}
}
\description{
Starts an execution of an Amazon Bedrock flow. Unlike flows that run until completion or time out after five minutes, flow executions let you run flows asynchronously for longer durations. Flow executions also yield control so that your application can perform other tasks.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentruntime_start_flow_execution/} for full documentation.
}
\keyword{internal}
