% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentruntime_operations.R
\name{bedrockagentruntime_stop_flow_execution}
\alias{bedrockagentruntime_stop_flow_execution}
\title{Stops an Amazon Bedrock flow's execution}
\usage{
bedrockagentruntime_stop_flow_execution(
  executionIdentifier,
  flowAliasIdentifier,
  flowIdentifier
)
}
\arguments{
\item{executionIdentifier}{[required] The unique identifier of the flow execution to stop.}

\item{flowAliasIdentifier}{[required] The unique identifier of the flow alias used for the execution.}

\item{flowIdentifier}{[required] The unique identifier of the flow.}
}
\description{
Stops an Amazon Bedrock flow's execution. This operation prevents further processing of the flow and changes the execution status to \code{Aborted}.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentruntime_stop_flow_execution/} for full documentation.
}
\keyword{internal}
