% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_copy_blueprint_stage}
\alias{bedrockdataautomation_copy_blueprint_stage}
\title{Copies a Blueprint from one stage to another}
\usage{
bedrockdataautomation_copy_blueprint_stage(
  blueprintArn,
  sourceStage,
  targetStage,
  clientToken = NULL
)
}
\arguments{
\item{blueprintArn}{[required] Blueprint to be copied}

\item{sourceStage}{[required] Source stage to copy from}

\item{targetStage}{[required] Target stage to copy to}

\item{clientToken}{Client token for idempotency}
}
\description{
Copies a Blueprint from one stage to another

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_copy_blueprint_stage/} for full documentation.
}
\keyword{internal}
