% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_create_blueprint}
\alias{bedrockdataautomation_create_blueprint}
\title{Creates an Amazon Bedrock Data Automation Blueprint}
\usage{
bedrockdataautomation_create_blueprint(
  blueprintName,
  type,
  blueprintStage = NULL,
  schema,
  clientToken = NULL,
  encryptionConfiguration = NULL,
  tags = NULL
)
}
\arguments{
\item{blueprintName}{[required] Name of the Blueprint}

\item{type}{[required] Type}

\item{blueprintStage}{Stage of the Blueprint}

\item{schema}{[required] Schema of the blueprint}

\item{clientToken}{Client specified token used for idempotency checks}

\item{encryptionConfiguration}{KMS Encryption Configuration}

\item{tags}{List of tags}
}
\description{
Creates an Amazon Bedrock Data Automation Blueprint

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_create_blueprint/} for full documentation.
}
\keyword{internal}
