% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_create_blueprint_version}
\alias{bedrockdataautomation_create_blueprint_version}
\title{Creates a new version of an existing Amazon Bedrock Data Automation
Blueprint}
\usage{
bedrockdataautomation_create_blueprint_version(
  blueprintArn,
  clientToken = NULL
)
}
\arguments{
\item{blueprintArn}{[required] ARN generated at the server side when a Blueprint is created}

\item{clientToken}{}
}
\description{
Creates a new version of an existing Amazon Bedrock Data Automation Blueprint

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_create_blueprint_version/} for full documentation.
}
\keyword{internal}
