% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_create_data_automation_library}
\alias{bedrockdataautomation_create_data_automation_library}
\title{Creates an Amazon Bedrock Data Automation Library}
\usage{
bedrockdataautomation_create_data_automation_library(
  libraryName,
  libraryDescription = NULL,
  clientToken = NULL,
  encryptionConfiguration = NULL,
  tags = NULL
)
}
\arguments{
\item{libraryName}{[required]}
}
\description{
Creates an Amazon Bedrock Data Automation Library

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_create_data_automation_library/} for full documentation.
}
\keyword{internal}
