% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_create_data_automation_project}
\alias{bedrockdataautomation_create_data_automation_project}
\title{Creates an Amazon Bedrock Data Automation Project}
\usage{
bedrockdataautomation_create_data_automation_project(
  projectName,
  projectDescription = NULL,
  projectStage = NULL,
  projectType = NULL,
  standardOutputConfiguration,
  customOutputConfiguration = NULL,
  overrideConfiguration = NULL,
  dataAutomationLibraryConfiguration = NULL,
  clientToken = NULL,
  encryptionConfiguration = NULL,
  tags = NULL
)
}
\arguments{
\item{projectName}{[required] Name of the DataAutomationProject}

\item{projectDescription}{Description of the DataAutomationProject}

\item{projectStage}{Stage of the Project}

\item{projectType}{Type of the DataAutomationProject}

\item{standardOutputConfiguration}{[required] Standard output configuration}

\item{customOutputConfiguration}{Custom output configuration}

\item{overrideConfiguration}{Override configuration}

\item{dataAutomationLibraryConfiguration}{DataAutomation Library configuration}

\item{clientToken}{Client specified token used for idempotency checks}

\item{encryptionConfiguration}{KMS Encryption Configuration}

\item{tags}{List of tags}
}
\description{
Creates an Amazon Bedrock Data Automation Project

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_create_data_automation_project/} for full documentation.
}
\keyword{internal}
