% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_delete_blueprint}
\alias{bedrockdataautomation_delete_blueprint}
\title{Deletes an existing Amazon Bedrock Keystone Blueprint}
\usage{
bedrockdataautomation_delete_blueprint(blueprintArn, blueprintVersion = NULL)
}
\arguments{
\item{blueprintArn}{[required] ARN generated at the server side when a Blueprint is created}

\item{blueprintVersion}{Optional field to delete a specific Blueprint version}
}
\description{
Deletes an existing Amazon Bedrock Keystone Blueprint

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_delete_blueprint/} for full documentation.
}
\keyword{internal}
