% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_get_blueprint}
\alias{bedrockdataautomation_get_blueprint}
\title{Gets an existing Amazon Bedrock Keystone Blueprint}
\usage{
bedrockdataautomation_get_blueprint(
  blueprintArn,
  blueprintVersion = NULL,
  blueprintStage = NULL
)
}
\arguments{
\item{blueprintArn}{[required] ARN generated at the server side when a Blueprint is created}

\item{blueprintVersion}{Optional field to get a specific Blueprint version}

\item{blueprintStage}{Optional field to get a specific Blueprint stage}
}
\description{
Gets an existing Amazon Bedrock Keystone Blueprint

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_get_blueprint/} for full documentation.
}
\keyword{internal}
