% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_get_blueprint_optimization_status}
\alias{bedrockdataautomation_get_blueprint_optimization_status}
\title{API used to get blueprint optimization status}
\usage{
bedrockdataautomation_get_blueprint_optimization_status(invocationArn)
}
\arguments{
\item{invocationArn}{[required] Invocation arn.}
}
\description{
API used to get blueprint optimization status.

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_get_blueprint_optimization_status/} for full documentation.
}
\keyword{internal}
