% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_get_data_automation_library_ingestion_job}
\alias{bedrockdataautomation_get_data_automation_library_ingestion_job}
\title{API used to get status of data automation library ingestion job}
\usage{
bedrockdataautomation_get_data_automation_library_ingestion_job(
  libraryArn,
  jobArn
)
}
\arguments{
\item{libraryArn}{[required] ARN generated at the server side when a DataAutomationLibrary is created}

\item{jobArn}{[required] ARN of the DataAutomationLibraryIngestionJob}
}
\description{
API used to get status of data automation library ingestion job

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_get_data_automation_library_ingestion_job/} for full documentation.
}
\keyword{internal}
