% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_get_data_automation_library_entity}
\alias{bedrockdataautomation_get_data_automation_library_entity}
\title{Gets an existing entity based on entity type from the library}
\usage{
bedrockdataautomation_get_data_automation_library_entity(
  libraryArn,
  entityType,
  entityId
)
}
\arguments{
\item{libraryArn}{[required] ARN generated at the server side when a DataAutomationLibrary is created}

\item{entityType}{[required] The entity type for which the entity is requested}

\item{entityId}{[required] Unique identifier for the entity}
}
\description{
Gets an existing entity based on entity type from the library

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_get_data_automation_library_entity/} for full documentation.
}
\keyword{internal}
