% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_get_data_automation_project}
\alias{bedrockdataautomation_get_data_automation_project}
\title{Gets an existing Amazon Bedrock Keystone DataAutomationProject}
\usage{
bedrockdataautomation_get_data_automation_project(
  projectArn,
  projectStage = NULL
)
}
\arguments{
\item{projectArn}{[required] ARN generated at the server side when a DataAutomationProject is created}

\item{projectStage}{Optional field to delete a specific DataAutomationProject stage}
}
\description{
Gets an existing Amazon Bedrock Keystone DataAutomationProject

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_get_data_automation_project/} for full documentation.
}
\keyword{internal}
