% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_invoke_data_automation_library_ingestion_job}
\alias{bedrockdataautomation_invoke_data_automation_library_ingestion_job}
\title{Async API: Invoke data automation library ingestion job}
\usage{
bedrockdataautomation_invoke_data_automation_library_ingestion_job(
  libraryArn,
  clientToken = NULL,
  inputConfiguration,
  entityType,
  operationType,
  outputConfiguration,
  notificationConfiguration = NULL,
  tags = NULL
)
}
\arguments{
\item{libraryArn}{[required] ARN generated at the server side when a DataAutomationLibrary is created}

\item{clientToken}{Idempotency token}

\item{inputConfiguration}{[required] Input configuration of DataAutomationLibraryIngestionJob request}

\item{entityType}{[required] The entity type for which DataAutomationLibraryIngestionJob is being run}

\item{operationType}{[required] The operation to be performed by DataAutomationLibraryIngestionJob}

\item{outputConfiguration}{[required] Output configuration of DataAutomationLibraryIngestionJob}

\item{notificationConfiguration}{Notification configuration.}

\item{tags}{List of tags}
}
\description{
Async API: Invoke data automation library ingestion job

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_invoke_data_automation_library_ingestion_job/} for full documentation.
}
\keyword{internal}
