% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_invoke_blueprint_optimization_async}
\alias{bedrockdataautomation_invoke_blueprint_optimization_async}
\title{Invoke an async job to perform Blueprint Optimization}
\usage{
bedrockdataautomation_invoke_blueprint_optimization_async(
  blueprint,
  samples,
  outputConfiguration,
  dataAutomationProfileArn,
  encryptionConfiguration = NULL,
  tags = NULL
)
}
\arguments{
\item{blueprint}{[required] Blueprint to be optimized}

\item{samples}{[required] List of Blueprint Optimization Samples}

\item{outputConfiguration}{[required] Output configuration where the results should be placed}

\item{dataAutomationProfileArn}{[required] Data automation profile ARN}

\item{encryptionConfiguration}{Encryption configuration.}

\item{tags}{List of tags.}
}
\description{
Invoke an async job to perform Blueprint Optimization

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_invoke_blueprint_optimization_async/} for full documentation.
}
\keyword{internal}
