% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_list_blueprints}
\alias{bedrockdataautomation_list_blueprints}
\title{Lists all existing Amazon Bedrock Data Automation Blueprints}
\usage{
bedrockdataautomation_list_blueprints(
  blueprintArn = NULL,
  resourceOwner = NULL,
  blueprintStageFilter = NULL,
  maxResults = NULL,
  nextToken = NULL,
  projectFilter = NULL
)
}
\arguments{
\item{blueprintArn}{ARN of a Blueprint}

\item{resourceOwner}{Resource Owner}

\item{blueprintStageFilter}{Blueprint Stage filter}

\item{maxResults}{Max Results}

\item{nextToken}{Pagination token}

\item{projectFilter}{Data Automation Project Filter}
}
\description{
Lists all existing Amazon Bedrock Data Automation Blueprints

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_list_blueprints/} for full documentation.
}
\keyword{internal}
