% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_list_data_automation_library_ingestion_jobs}
\alias{bedrockdataautomation_list_data_automation_library_ingestion_jobs}
\title{Lists all data automation library ingestion jobs}
\usage{
bedrockdataautomation_list_data_automation_library_ingestion_jobs(
  libraryArn,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{libraryArn}{[required] ARN generated at the server side when a DataAutomationLibrary is created}

\item{nextToken}{Pagination token for retrieving the next set of results}
}
\description{
Lists all data automation library ingestion jobs

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_list_data_automation_library_ingestion_jobs/} for full documentation.
}
\keyword{internal}
