% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_list_data_automation_projects}
\alias{bedrockdataautomation_list_data_automation_projects}
\title{Lists all existing Amazon Bedrock Keystone DataAutomationProjects}
\usage{
bedrockdataautomation_list_data_automation_projects(
  maxResults = NULL,
  nextToken = NULL,
  projectStageFilter = NULL,
  blueprintFilter = NULL,
  resourceOwner = NULL
)
}
\arguments{
\item{maxResults}{}

\item{nextToken}{}

\item{projectStageFilter}{}

\item{blueprintFilter}{}

\item{resourceOwner}{}
}
\description{
Lists all existing Amazon Bedrock Keystone DataAutomationProjects

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_list_data_automation_projects/} for full documentation.
}
\keyword{internal}
