% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_list_tags_for_resource}
\alias{bedrockdataautomation_list_tags_for_resource}
\title{List tags for an Amazon Bedrock Data Automation resource}
\usage{
bedrockdataautomation_list_tags_for_resource(resourceARN)
}
\arguments{
\item{resourceARN}{[required]}
}
\description{
List tags for an Amazon Bedrock Data Automation resource

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
