% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_tag_resource}
\alias{bedrockdataautomation_tag_resource}
\title{Tag an Amazon Bedrock Data Automation resource}
\usage{
bedrockdataautomation_tag_resource(resourceARN, tags)
}
\arguments{
\item{resourceARN}{[required]}

\item{tags}{[required]}
}
\description{
Tag an Amazon Bedrock Data Automation resource

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_tag_resource/} for full documentation.
}
\keyword{internal}
