% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_untag_resource}
\alias{bedrockdataautomation_untag_resource}
\title{Untag an Amazon Bedrock Data Automation resource}
\usage{
bedrockdataautomation_untag_resource(resourceARN, tagKeys)
}
\arguments{
\item{resourceARN}{[required]}

\item{tagKeys}{[required]}
}
\description{
Untag an Amazon Bedrock Data Automation resource

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_untag_resource/} for full documentation.
}
\keyword{internal}
