% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_update_blueprint}
\alias{bedrockdataautomation_update_blueprint}
\title{Updates an existing Amazon Bedrock Data Automation Blueprint}
\usage{
bedrockdataautomation_update_blueprint(
  blueprintArn,
  schema,
  blueprintStage = NULL
)
}
\arguments{
\item{blueprintArn}{[required] ARN generated at the server side when a Blueprint is created}

\item{schema}{[required]}

\item{blueprintStage}{}
}
\description{
Updates an existing Amazon Bedrock Data Automation Blueprint

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_update_blueprint/} for full documentation.
}
\keyword{internal}
