% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_update_data_automation_library}
\alias{bedrockdataautomation_update_data_automation_library}
\title{Updates an existing Amazon Bedrock Data Automation Library}
\usage{
bedrockdataautomation_update_data_automation_library(
  libraryArn,
  libraryDescription = NULL,
  clientToken = NULL
)
}
\arguments{
\item{libraryArn}{[required] ARN generated at the server side when a DataAutomationLibrary is created}
}
\description{
Updates an existing Amazon Bedrock Data Automation Library

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_update_data_automation_library/} for full documentation.
}
\keyword{internal}
