% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_update_data_automation_project}
\alias{bedrockdataautomation_update_data_automation_project}
\title{Updates an existing Amazon Bedrock DataAutomationProject}
\usage{
bedrockdataautomation_update_data_automation_project(
  projectArn,
  projectStage = NULL,
  projectDescription = NULL,
  standardOutputConfiguration,
  customOutputConfiguration = NULL,
  overrideConfiguration = NULL
)
}
\arguments{
\item{projectArn}{[required] ARN generated at the server side when a DataAutomationProject is created}

\item{projectStage}{}

\item{projectDescription}{}

\item{standardOutputConfiguration}{[required]}

\item{customOutputConfiguration}{}

\item{overrideConfiguration}{}
}
\description{
Updates an existing Amazon Bedrock DataAutomationProject

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_update_data_automation_project/} for full documentation.
}
\keyword{internal}
