% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomationruntime_operations.R
\name{bedrockdataautomationruntime_get_data_automation_status}
\alias{bedrockdataautomationruntime_get_data_automation_status}
\title{API used to get data automation status}
\usage{
bedrockdataautomationruntime_get_data_automation_status(invocationArn)
}
\arguments{
\item{invocationArn}{[required] Invocation arn.}
}
\description{
API used to get data automation status.

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomationruntime_get_data_automation_status/} for full documentation.
}
\keyword{internal}
