% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomationruntime_operations.R
\name{bedrockdataautomationruntime_invoke_data_automation}
\alias{bedrockdataautomationruntime_invoke_data_automation}
\title{Sync API: Invoke data automation}
\usage{
bedrockdataautomationruntime_invoke_data_automation(
  inputConfiguration,
  dataAutomationConfiguration = NULL,
  blueprints = NULL,
  dataAutomationProfileArn,
  encryptionConfiguration = NULL
)
}
\arguments{
\item{inputConfiguration}{[required] Input configuration.}

\item{dataAutomationConfiguration}{Data automation configuration.}

\item{blueprints}{Blueprint list.}

\item{dataAutomationProfileArn}{[required] Data automation profile ARN}

\item{encryptionConfiguration}{Encryption configuration.}
}
\description{
Sync API: Invoke data automation.

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomationruntime_invoke_data_automation/} for full documentation.
}
\keyword{internal}
