% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomationruntime_operations.R
\name{bedrockdataautomationruntime_invoke_data_automation_async}
\alias{bedrockdataautomationruntime_invoke_data_automation_async}
\title{Async API: Invoke data automation}
\usage{
bedrockdataautomationruntime_invoke_data_automation_async(
  clientToken = NULL,
  inputConfiguration,
  outputConfiguration,
  dataAutomationConfiguration = NULL,
  encryptionConfiguration = NULL,
  notificationConfiguration = NULL,
  blueprints = NULL,
  dataAutomationProfileArn,
  tags = NULL
)
}
\arguments{
\item{clientToken}{Idempotency token.}

\item{inputConfiguration}{[required] Input configuration.}

\item{outputConfiguration}{[required] Output configuration.}

\item{dataAutomationConfiguration}{Data automation configuration.}

\item{encryptionConfiguration}{Encryption configuration.}

\item{notificationConfiguration}{Notification configuration.}

\item{blueprints}{Blueprint list.}

\item{dataAutomationProfileArn}{[required] Data automation profile ARN}

\item{tags}{List of tags.}
}
\description{
Async API: Invoke data automation.

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomationruntime_invoke_data_automation_async/} for full documentation.
}
\keyword{internal}
