% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockruntime_operations.R
\name{bedrockruntime_apply_guardrail}
\alias{bedrockruntime_apply_guardrail}
\title{The action to apply a guardrail}
\usage{
bedrockruntime_apply_guardrail(
  guardrailIdentifier,
  guardrailVersion,
  source,
  content
)
}
\arguments{
\item{guardrailIdentifier}{[required] The guardrail identifier used in the request to apply the guardrail.}

\item{guardrailVersion}{[required] The guardrail version used in the request to apply the guardrail.}

\item{source}{[required] The source of data used in the request to apply the guardrail.}

\item{content}{[required] The content details used in the request to apply the guardrail.}
}
\description{
The action to apply a guardrail.

See \url{https://www.paws-r-sdk.com/docs/bedrockruntime_apply_guardrail/} for full documentation.
}
\keyword{internal}
