% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockruntime_operations.R
\name{bedrockruntime_invoke_model_with_response_stream}
\alias{bedrockruntime_invoke_model_with_response_stream}
\title{Invoke the specified Bedrock model to run inference using the input
provided}
\usage{
bedrockruntime_invoke_model_with_response_stream(
  accept = NULL,
  body,
  contentType = NULL,
  modelId
)
}
\arguments{
\item{accept}{The desired MIME type of the inference body in the response. The default
value is \code{application/json}.}

\item{body}{[required] Inference input in the format specified by the content-type. To see the
format and content of this field for different models, refer to
\href{https://docs.aws.amazon.com/bedrock/latest/userguide/model-parameters.html}{Inference parameters}.}

\item{contentType}{The MIME type of the input data in the request. The default value is
\code{application/json}.}

\item{modelId}{[required] Id of the model to invoke using the streaming request.}
}
\description{
Invoke the specified Bedrock model to run inference using the input provided. Return the response in a stream.

See \url{https://www.paws-r-sdk.com/docs/bedrockruntime_invoke_model_with_response_stream/} for full documentation.
}
\keyword{internal}
