% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockruntime_operations.R
\name{bedrockruntime_list_async_invokes}
\alias{bedrockruntime_list_async_invokes}
\title{Lists asynchronous invocations}
\usage{
bedrockruntime_list_async_invokes(
  submitTimeAfter = NULL,
  submitTimeBefore = NULL,
  statusEquals = NULL,
  maxResults = NULL,
  nextToken = NULL,
  sortBy = NULL,
  sortOrder = NULL
)
}
\arguments{
\item{submitTimeAfter}{Include invocations submitted after this time.}

\item{submitTimeBefore}{Include invocations submitted before this time.}

\item{statusEquals}{Filter invocations by status.}

\item{maxResults}{The maximum number of invocations to return in one page of results.}

\item{nextToken}{Specify the pagination token from a previous request to retrieve the
next page of results.}

\item{sortBy}{How to sort the response.}

\item{sortOrder}{The sorting order for the response.}
}
\description{
Lists asynchronous invocations.

See \url{https://www.paws-r-sdk.com/docs/bedrockruntime_list_async_invokes/} for full documentation.
}
\keyword{internal}
