% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockruntime_operations.R
\name{bedrockruntime_start_async_invoke}
\alias{bedrockruntime_start_async_invoke}
\title{Starts an asynchronous invocation}
\usage{
bedrockruntime_start_async_invoke(
  clientRequestToken = NULL,
  modelId,
  modelInput,
  outputDataConfig,
  tags = NULL
)
}
\arguments{
\item{clientRequestToken}{Specify idempotency token to ensure that requests are not duplicated.}

\item{modelId}{[required] The model to invoke.}

\item{modelInput}{[required] Input to send to the model.}

\item{outputDataConfig}{[required] Where to store the output.}

\item{tags}{Tags to apply to the invocation.}
}
\description{
Starts an asynchronous invocation.

See \url{https://www.paws-r-sdk.com/docs/bedrockruntime_start_async_invoke/} for full documentation.
}
\keyword{internal}
