% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_batch_detect_targeted_sentiment}
\alias{comprehend_batch_detect_targeted_sentiment}
\title{Inspects a batch of documents and returns a sentiment analysis for each
entity identified in the documents}
\usage{
comprehend_batch_detect_targeted_sentiment(TextList, LanguageCode)
}
\arguments{
\item{TextList}{[required] A list containing the UTF-8 encoded text of the input documents. The
list can contain a maximum of 25 documents. The maximum size of each
document is 5 KB.}

\item{LanguageCode}{[required] The language of the input documents. Currently, English is the only
supported language.}
}
\description{
Inspects a batch of documents and returns a sentiment analysis for each entity identified in the documents.

See \url{https://www.paws-r-sdk.com/docs/comprehend_batch_detect_targeted_sentiment/} for full documentation.
}
\keyword{internal}
