% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_contains_pii_entities}
\alias{comprehend_contains_pii_entities}
\title{Analyzes input text for the presence of personally identifiable
information (PII) and returns the labels of identified PII entity types
such as name, address, bank account number, or phone number}
\usage{
comprehend_contains_pii_entities(Text, LanguageCode)
}
\arguments{
\item{Text}{[required] Creates a new document classification request to analyze a single
document in real-time, returning personally identifiable information
(PII) entity labels.}

\item{LanguageCode}{[required] The language of the input documents.}
}
\value{
A list with the following syntax:\preformatted{list(
  Labels = list(
    list(
      Name = "BANK_ACCOUNT_NUMBER"|"BANK_ROUTING"|"CREDIT_DEBIT_NUMBER"|"CREDIT_DEBIT_CVV"|"CREDIT_DEBIT_EXPIRY"|"PIN"|"EMAIL"|"ADDRESS"|"NAME"|"PHONE"|"SSN"|"DATE_TIME"|"PASSPORT_NUMBER"|"DRIVER_ID"|"URL"|"AGE"|"USERNAME"|"PASSWORD"|"AWS_ACCESS_KEY"|"AWS_SECRET_KEY"|"IP_ADDRESS"|"MAC_ADDRESS"|"ALL",
      Score = 123.0
    )
  )
)
}
}
\description{
Analyzes input text for the presence of personally identifiable
information (PII) and returns the labels of identified PII entity types
such as name, address, bank account number, or phone number.
}
\section{Request syntax}{
\preformatted{svc$contains_pii_entities(
  Text = "string",
  LanguageCode = "en"|"es"|"fr"|"de"|"it"|"pt"|"ar"|"hi"|"ja"|"ko"|"zh"|"zh-TW"
)
}
}

\keyword{internal}
