% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_delete_document_classifier}
\alias{comprehend_delete_document_classifier}
\title{Deletes a previously created document classifier}
\usage{
comprehend_delete_document_classifier(DocumentClassifierArn)
}
\arguments{
\item{DocumentClassifierArn}{[required] The Amazon Resource Name (ARN) that identifies the document classifier.}
}
\description{
Deletes a previously created document classifier

See \url{https://paws-r.github.io/docs/comprehend/delete_document_classifier.html} for full documentation.
}
\keyword{internal}
