% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_delete_flywheel}
\alias{comprehend_delete_flywheel}
\title{Deletes a flywheel}
\usage{
comprehend_delete_flywheel(FlywheelArn)
}
\arguments{
\item{FlywheelArn}{[required] The Amazon Resource Number (ARN) of the flywheel to delete.}
}
\description{
Deletes a flywheel. When you delete the flywheel, Amazon Comprehend does not delete the data lake or the model associated with the flywheel.

See \url{https://paws-r.github.io/docs/comprehend/delete_flywheel.html} for full documentation.
}
\keyword{internal}
