% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_delete_resource_policy}
\alias{comprehend_delete_resource_policy}
\title{Deletes a resource-based policy that is attached to a custom model}
\usage{
comprehend_delete_resource_policy(ResourceArn, PolicyRevisionId = NULL)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the custom model version that has the
policy to delete.}

\item{PolicyRevisionId}{The revision ID of the policy to delete.}
}
\description{
Deletes a resource-based policy that is attached to a custom model.

See \url{https://paws-r.github.io/docs/comprehend/delete_resource_policy.html} for full documentation.
}
\keyword{internal}
