% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_describe_dataset}
\alias{comprehend_describe_dataset}
\title{Returns information about the dataset that you specify}
\usage{
comprehend_describe_dataset(DatasetArn)
}
\arguments{
\item{DatasetArn}{[required] The ARN of the dataset.}
}
\description{
Returns information about the dataset that you specify. For more information about datasets, see \href{https://docs.aws.amazon.com/comprehend/latest/dg/flywheels-about.html}{Flywheel overview} in the \emph{Amazon Comprehend Developer Guide}.

See \url{https://paws-r.github.io/docs/comprehend/describe_dataset.html} for full documentation.
}
\keyword{internal}
