% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_describe_entity_recognizer}
\alias{comprehend_describe_entity_recognizer}
\title{Provides details about an entity recognizer including status, S3 buckets
containing training data, recognizer metadata, metrics, and so on}
\usage{
comprehend_describe_entity_recognizer(EntityRecognizerArn)
}
\arguments{
\item{EntityRecognizerArn}{[required] The Amazon Resource Name (ARN) that identifies the entity recognizer.}
}
\description{
Provides details about an entity recognizer including status, S3 buckets containing training data, recognizer metadata, metrics, and so on.

See \url{https://www.paws-r-sdk.com/docs/comprehend_describe_entity_recognizer/} for full documentation.
}
\keyword{internal}
