% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_describe_events_detection_job}
\alias{comprehend_describe_events_detection_job}
\title{Gets the status and details of an events detection job}
\usage{
comprehend_describe_events_detection_job(JobId)
}
\arguments{
\item{JobId}{[required] The identifier of the events detection job.}
}
\description{
Gets the status and details of an events detection job.
}
\section{Request syntax}{
\preformatted{svc$describe_events_detection_job(
  JobId = "string"
)
}
}

\keyword{internal}
