% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_describe_flywheel}
\alias{comprehend_describe_flywheel}
\title{Provides configuration information about the flywheel}
\usage{
comprehend_describe_flywheel(FlywheelArn)
}
\arguments{
\item{FlywheelArn}{[required] The Amazon Resource Number (ARN) of the flywheel.}
}
\description{
Provides configuration information about the flywheel. For more information about flywheels, see \href{https://docs.aws.amazon.com/comprehend/latest/dg/flywheels-about.html}{Flywheel overview} in the \emph{Amazon Comprehend Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/comprehend_describe_flywheel/} for full documentation.
}
\keyword{internal}
