% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_describe_flywheel_iteration}
\alias{comprehend_describe_flywheel_iteration}
\title{Retrieve the configuration properties of a flywheel iteration}
\usage{
comprehend_describe_flywheel_iteration(FlywheelArn, FlywheelIterationId)
}
\arguments{
\item{FlywheelArn}{[required]}

\item{FlywheelIterationId}{[required]}
}
\description{
Retrieve the configuration properties of a flywheel iteration. For more information about flywheels, see \href{https://docs.aws.amazon.com/comprehend/latest/dg/flywheels-about.html}{Flywheel overview} in the \emph{Amazon Comprehend Developer Guide}.

See \url{https://paws-r.github.io/docs/comprehend/describe_flywheel_iteration.html} for full documentation.
}
\keyword{internal}
