% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_describe_resource_policy}
\alias{comprehend_describe_resource_policy}
\title{Gets the details of a resource-based policy that is attached to a custom
model, including the JSON body of the policy}
\usage{
comprehend_describe_resource_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the custom model version that has the
resource policy.}
}
\description{
Gets the details of a resource-based policy that is attached to a custom model, including the JSON body of the policy.

See \url{https://paws-r.github.io/docs/comprehend/describe_resource_policy.html} for full documentation.
}
\keyword{internal}
