% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_describe_sentiment_detection_job}
\alias{comprehend_describe_sentiment_detection_job}
\title{Gets the properties associated with a sentiment detection job}
\usage{
comprehend_describe_sentiment_detection_job(JobId)
}
\arguments{
\item{JobId}{[required] The identifier that Amazon Comprehend generated for the job. The
operation returns this identifier in its response.}
}
\description{
Gets the properties associated with a sentiment detection job. Use this
operation to get the status of a detection job.
}
\section{Request syntax}{
\preformatted{svc$describe_sentiment_detection_job(
  JobId = "string"
)
}
}

\keyword{internal}
