% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_describe_topics_detection_job}
\alias{comprehend_describe_topics_detection_job}
\title{Gets the properties associated with a topic detection job}
\usage{
comprehend_describe_topics_detection_job(JobId)
}
\arguments{
\item{JobId}{[required] The identifier assigned by the user to the detection job.}
}
\description{
Gets the properties associated with a topic detection job. Use this operation to get the status of a detection job.

See \url{https://paws-r.github.io/docs/comprehend/describe_topics_detection_job.html} for full documentation.
}
\keyword{internal}
