% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_detect_dominant_language}
\alias{comprehend_detect_dominant_language}
\title{Determines the dominant language of the input text}
\usage{
comprehend_detect_dominant_language(Text)
}
\arguments{
\item{Text}{[required] A UTF-8 text string. Each string should contain at least 20 characters
and must contain fewer that 5,000 bytes of UTF-8 encoded characters.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Languages = list(
    list(
      LanguageCode = "string",
      Score = 123.0
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Determines the dominant language of the input text. For a list of
languages that Amazon Comprehend can detect, see \href{https://docs.aws.amazon.com/comprehend/latest/dg/how-languages.html}{Amazon Comprehend Supported Languages}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$detect_dominant_language(
  Text = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
