% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_detect_key_phrases}
\alias{comprehend_detect_key_phrases}
\title{Detects the key noun phrases found in the text}
\usage{
comprehend_detect_key_phrases(Text, LanguageCode)
}
\arguments{
\item{Text}{[required] A UTF-8 text string. Each string must contain fewer that 5,000 bytes of
UTF-8 encoded characters.}

\item{LanguageCode}{[required] The language of the input documents. You can specify any of the primary
languages supported by Amazon Comprehend. All documents must be in the
same language.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  KeyPhrases = list(
    list(
      Score = 123.0,
      Text = "string",
      BeginOffset = 123,
      EndOffset = 123
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Detects the key noun phrases found in the text.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$detect_key_phrases(
  Text = "string",
  LanguageCode = "en"|"es"|"fr"|"de"|"it"|"pt"|"ar"|"hi"|"ja"|"ko"|"zh"|"zh-TW"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
