% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_detect_pii_entities}
\alias{comprehend_detect_pii_entities}
\title{Inspects the input text for entities that contain personally
identifiable information (PII) and returns information about them}
\usage{
comprehend_detect_pii_entities(Text, LanguageCode)
}
\arguments{
\item{Text}{[required] A UTF-8 text string. Each string must contain fewer that 5,000 bytes of
UTF-8 encoded characters.}

\item{LanguageCode}{[required] The language of the input documents.}
}
\description{
Inspects the input text for entities that contain personally
identifiable information (PII) and returns information about them.
}
\section{Request syntax}{
\preformatted{svc$detect_pii_entities(
  Text = "string",
  LanguageCode = "en"|"es"|"fr"|"de"|"it"|"pt"|"ar"|"hi"|"ja"|"ko"|"zh"|"zh-TW"
)
}
}

\keyword{internal}
