% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_list_datasets}
\alias{comprehend_list_datasets}
\title{List the datasets that you have configured in this Region}
\usage{
comprehend_list_datasets(
  FlywheelArn = NULL,
  Filter = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{FlywheelArn}{The Amazon Resource Number (ARN) of the flywheel.}

\item{Filter}{Filters the datasets to be returned in the response.}

\item{NextToken}{Identifies the next page of results to return.}

\item{MaxResults}{Maximum number of results to return in a response. The default is 100.}
}
\description{
List the datasets that you have configured in this Region. For more information about datasets, see \href{https://docs.aws.amazon.com/comprehend/latest/dg/flywheels-about.html}{Flywheel overview} in the \emph{Amazon Comprehend Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/comprehend_list_datasets/} for full documentation.
}
\keyword{internal}
