% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_list_entity_recognizer_summaries}
\alias{comprehend_list_entity_recognizer_summaries}
\title{Gets a list of summaries for the entity recognizers that you have
created}
\usage{
comprehend_list_entity_recognizer_summaries(
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{NextToken}{Identifies the next page of results to return.}

\item{MaxResults}{The maximum number of results to return on each page. The default is
100.}
}
\description{
Gets a list of summaries for the entity recognizers that you have created.

See \url{https://www.paws-r-sdk.com/docs/comprehend_list_entity_recognizer_summaries/} for full documentation.
}
\keyword{internal}
