% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_list_flywheel_iteration_history}
\alias{comprehend_list_flywheel_iteration_history}
\title{Information about the history of a flywheel iteration}
\usage{
comprehend_list_flywheel_iteration_history(
  FlywheelArn,
  Filter = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{FlywheelArn}{[required] The ARN of the flywheel.}

\item{Filter}{Filter the flywheel iteration history based on creation time.}

\item{NextToken}{Next token}

\item{MaxResults}{Maximum number of iteration history results to return}
}
\description{
Information about the history of a flywheel iteration. For more information about flywheels, see \href{https://docs.aws.amazon.com/comprehend/latest/dg/flywheels-about.html}{Flywheel overview} in the \emph{Amazon Comprehend Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/comprehend_list_flywheel_iteration_history/} for full documentation.
}
\keyword{internal}
