% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_list_flywheels}
\alias{comprehend_list_flywheels}
\title{Gets a list of the flywheels that you have created}
\usage{
comprehend_list_flywheels(Filter = NULL, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{Filter}{Filters the flywheels that are returned. You can filter flywheels on
their status, or the date and time that they were submitted. You can
only set one filter at a time.}

\item{NextToken}{Identifies the next page of results to return.}

\item{MaxResults}{Maximum number of results to return in a response. The default is 100.}
}
\description{
Gets a list of the flywheels that you have created.

See \url{https://www.paws-r-sdk.com/docs/comprehend_list_flywheels/} for full documentation.
}
\keyword{internal}
