% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_list_tags_for_resource}
\alias{comprehend_list_tags_for_resource}
\title{Lists all tags associated with a given Amazon Comprehend resource}
\usage{
comprehend_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the given Amazon Comprehend resource
you are querying.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ResourceArn = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Lists all tags associated with a given Amazon Comprehend resource.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_tags_for_resource(
  ResourceArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
