% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_start_document_classification_job}
\alias{comprehend_start_document_classification_job}
\title{Starts an asynchronous document classification job}
\usage{
comprehend_start_document_classification_job(
  JobName = NULL,
  DocumentClassifierArn = NULL,
  InputDataConfig,
  OutputDataConfig,
  DataAccessRoleArn,
  ClientRequestToken = NULL,
  VolumeKmsKeyId = NULL,
  VpcConfig = NULL,
  Tags = NULL,
  FlywheelArn = NULL
)
}
\arguments{
\item{JobName}{The identifier of the job.}

\item{DocumentClassifierArn}{The Amazon Resource Name (ARN) of the document classifier to use to
process the job.}

\item{InputDataConfig}{[required] Specifies the format and location of the input data for the job.}

\item{OutputDataConfig}{[required] Specifies where to send the output files.}

\item{DataAccessRoleArn}{[required] The Amazon Resource Name (ARN) of the IAM role that grants Amazon
Comprehend read access to your input data.}

\item{ClientRequestToken}{A unique identifier for the request. If you do not set the client
request token, Amazon Comprehend generates one.}

\item{VolumeKmsKeyId}{ID for the Amazon Web Services Key Management Service (KMS) key that
Amazon Comprehend uses to encrypt data on the storage volume attached to
the ML compute instance(s) that process the analysis job. The
VolumeKmsKeyId can be either of the following formats:
\itemize{
\item KMS Key ID: \code{"1234abcd-12ab-34cd-56ef-1234567890ab"}
\item Amazon Resource Name (ARN) of a KMS Key:
\code{"arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"}
}}

\item{VpcConfig}{Configuration parameters for an optional private Virtual Private Cloud
(VPC) containing the resources you are using for your document
classification job. For more information, see \href{https://docs.aws.amazon.com/vpc/latest/userguide/what-is-amazon-vpc.html}{Amazon VPC}.}

\item{Tags}{Tags to associate with the document classification job. A tag is a
key-value pair that adds metadata to a resource used by Amazon
Comprehend. For example, a tag with "Sales" as the key might be added to
a resource to indicate its use by the sales department.}

\item{FlywheelArn}{The Amazon Resource Number (ARN) of the flywheel associated with the
model to use.}
}
\description{
Starts an asynchronous document classification job. Use the \code{\link[=comprehend_describe_document_classification_job]{describe_document_classification_job}} operation to track the progress of the job.

See \url{https://paws-r.github.io/docs/comprehend/start_document_classification_job.html} for full documentation.
}
\keyword{internal}
