% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_start_flywheel_iteration}
\alias{comprehend_start_flywheel_iteration}
\title{Start the flywheel iteration}
\usage{
comprehend_start_flywheel_iteration(FlywheelArn, ClientRequestToken = NULL)
}
\arguments{
\item{FlywheelArn}{[required] The ARN of the flywheel.}

\item{ClientRequestToken}{A unique identifier for the request. If you don't set the client request
token, Amazon Comprehend generates one.}
}
\description{
Start the flywheel iteration.This operation uses any new datasets to train a new model version. For more information about flywheels, see \href{https://docs.aws.amazon.com/comprehend/latest/dg/flywheels-about.html}{Flywheel overview} in the \emph{Amazon Comprehend Developer Guide}.

See \url{https://paws-r.github.io/docs/comprehend/start_flywheel_iteration.html} for full documentation.
}
\keyword{internal}
