% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_stop_entities_detection_job}
\alias{comprehend_stop_entities_detection_job}
\title{Stops an entities detection job in progress}
\usage{
comprehend_stop_entities_detection_job(JobId)
}
\arguments{
\item{JobId}{[required] The identifier of the entities detection job to stop.}
}
\description{
Stops an entities detection job in progress.

See \url{https://paws-r.github.io/docs/comprehend/stop_entities_detection_job.html} for full documentation.
}
\keyword{internal}
