% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_stop_events_detection_job}
\alias{comprehend_stop_events_detection_job}
\title{Stops an events detection job in progress}
\usage{
comprehend_stop_events_detection_job(JobId)
}
\arguments{
\item{JobId}{[required] The identifier of the events detection job to stop.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  JobId = "string",
  JobStatus = "SUBMITTED"|"IN_PROGRESS"|"COMPLETED"|"FAILED"|"STOP_REQUESTED"|"STOPPED"
)
}\if{html}{\out{</div>}}
}
\description{
Stops an events detection job in progress.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$stop_events_detection_job(
  JobId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
