% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_stop_pii_entities_detection_job}
\alias{comprehend_stop_pii_entities_detection_job}
\title{Stops a PII entities detection job in progress}
\usage{
comprehend_stop_pii_entities_detection_job(JobId)
}
\arguments{
\item{JobId}{[required] The identifier of the PII entities detection job to stop.}
}
\value{
A list with the following syntax:\preformatted{list(
  JobId = "string",
  JobStatus = "SUBMITTED"|"IN_PROGRESS"|"COMPLETED"|"FAILED"|"STOP_REQUESTED"|"STOPPED"
)
}
}
\description{
Stops a PII entities detection job in progress.
}
\section{Request syntax}{
\preformatted{svc$stop_pii_entities_detection_job(
  JobId = "string"
)
}
}

\keyword{internal}
