% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_untag_resource}
\alias{comprehend_untag_resource}
\title{Removes a specific tag associated with an Amazon Comprehend resource}
\usage{
comprehend_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the given Amazon Comprehend resource
from which you want to remove the tags.}

\item{TagKeys}{[required] The initial part of a key-value pair that forms a tag being removed from
a given resource. For example, a tag with "Sales" as the key might be
added to a resource to indicate its use by the sales department. Keys
must be unique and cannot be duplicated for a particular resource.}
}
\description{
Removes a specific tag associated with an Amazon Comprehend resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
