% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_update_flywheel}
\alias{comprehend_update_flywheel}
\title{Update the configuration information for an existing flywheel}
\usage{
comprehend_update_flywheel(
  FlywheelArn,
  ActiveModelArn = NULL,
  DataAccessRoleArn = NULL,
  DataSecurityConfig = NULL
)
}
\arguments{
\item{FlywheelArn}{[required] The Amazon Resource Number (ARN) of the flywheel to update.}

\item{ActiveModelArn}{The Amazon Resource Number (ARN) of the active model version.}

\item{DataAccessRoleArn}{The Amazon Resource Name (ARN) of the IAM role that grants Amazon
Comprehend permission to access the flywheel data.}

\item{DataSecurityConfig}{Flywheel data security configuration.}
}
\description{
Update the configuration information for an existing flywheel.

See \url{https://www.paws-r-sdk.com/docs/comprehend_update_flywheel/} for full documentation.
}
\keyword{internal}
