% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehendmedical_operations.R
\name{comprehendmedical_describe_entities_detection_v2_job}
\alias{comprehendmedical_describe_entities_detection_v2_job}
\title{Gets the properties associated with a medical entities detection job}
\usage{
comprehendmedical_describe_entities_detection_v2_job(JobId)
}
\arguments{
\item{JobId}{[required] The identifier that Amazon Comprehend Medical generated for the job. The
\code{StartEntitiesDetectionV2Job} operation returns this identifier in its
response.}
}
\description{
Gets the properties associated with a medical entities detection job.
Use this operation to get the status of a detection job.
}
\section{Request syntax}{
\preformatted{svc$describe_entities_detection_v2_job(
  JobId = "string"
)
}
}

\keyword{internal}
