% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehendmedical_operations.R
\name{comprehendmedical_detect_entities}
\alias{comprehendmedical_detect_entities}
\title{The DetectEntities operation is deprecated}
\usage{
comprehendmedical_detect_entities(Text)
}
\arguments{
\item{Text}{[required] A UTF-8 text string containing the clinical content being examined for
entities.}
}
\description{
The \code{\link[=comprehendmedical_detect_entities]{detect_entities}} operation is deprecated. You should use the \code{\link[=comprehendmedical_detect_entities_v2]{detect_entities_v2}} operation instead.

See \url{https://www.paws-r-sdk.com/docs/comprehendmedical_detect_entities/} for full documentation.
}
\keyword{internal}
