% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehendmedical_operations.R
\name{comprehendmedical_infer_icd10cm}
\alias{comprehendmedical_infer_icd10cm}
\title{InferICD10CM detects medical conditions as entities listed in a patient
record and links those entities to normalized concept identifiers in the
ICD-10-CM knowledge base from the Centers for Disease Control}
\usage{
comprehendmedical_infer_icd10cm(Text)
}
\arguments{
\item{Text}{[required] The input text used for analysis. The input for InferICD10CM is a string
from 1 to 10000 characters.}
}
\description{
InferICD10CM detects medical conditions as entities listed in a patient record and links those entities to normalized concept identifiers in the ICD-10-CM knowledge base from the Centers for Disease Control. Amazon Comprehend Medical only detects medical entities in English language texts.

See \url{https://paws-r.github.io/docs/comprehendmedical/infer_icd10cm.html} for full documentation.
}
\keyword{internal}
