% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehendmedical_operations.R
\name{comprehendmedical_infer_snomedct}
\alias{comprehendmedical_infer_snomedct}
\title{InferSNOMEDCT detects possible medical concepts as entities and links
them to codes from the Systematized Nomenclature of Medicine, Clinical
Terms (SNOMED-CT) ontology}
\usage{
comprehendmedical_infer_snomedct(Text)
}
\arguments{
\item{Text}{[required] The input text to be analyzed using InferSNOMEDCT.}
}
\description{
InferSNOMEDCT detects possible medical concepts as entities and links them to codes from the Systematized Nomenclature of Medicine, Clinical Terms (SNOMED-CT) ontology

See \url{https://www.paws-r-sdk.com/docs/comprehendmedical_infer_snomedct/} for full documentation.
}
\keyword{internal}
