% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehendmedical_operations.R
\name{comprehendmedical_list_icd10cm_inference_jobs}
\alias{comprehendmedical_list_icd10cm_inference_jobs}
\title{Gets a list of InferICD10CM jobs that you have submitted}
\usage{
comprehendmedical_list_icd10cm_inference_jobs(
  Filter = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Filter}{Filters the jobs that are returned. You can filter jobs based on their
names, status, or the date and time that they were submitted. You can
only set one filter at a time.}

\item{NextToken}{Identifies the next page of results to return.}

\item{MaxResults}{The maximum number of results to return in each page. The default is
100.}
}
\description{
Gets a list of InferICD10CM jobs that you have submitted.

See \url{https://www.paws-r-sdk.com/docs/comprehendmedical_list_icd10cm_inference_jobs/} for full documentation.
}
\keyword{internal}
