% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehendmedical_operations.R
\name{comprehendmedical_list_snomedct_inference_jobs}
\alias{comprehendmedical_list_snomedct_inference_jobs}
\title{Gets a list of InferSNOMEDCT jobs a user has submitted}
\usage{
comprehendmedical_list_snomedct_inference_jobs(
  Filter = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Filter}{}

\item{NextToken}{Identifies the next page of InferSNOMEDCT results to return.}

\item{MaxResults}{The maximum number of results to return in each page. The default is
100.}
}
\description{
Gets a list of InferSNOMEDCT jobs a user has submitted.

See \url{https://www.paws-r-sdk.com/docs/comprehendmedical_list_snomedct_inference_jobs/} for full documentation.
}
\keyword{internal}
