% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehendmedical_operations.R
\name{comprehendmedical_stop_entities_detection_v2_job}
\alias{comprehendmedical_stop_entities_detection_v2_job}
\title{Stops a medical entities detection job in progress}
\usage{
comprehendmedical_stop_entities_detection_v2_job(JobId)
}
\arguments{
\item{JobId}{[required] The identifier of the medical entities job to stop.}
}
\description{
Stops a medical entities detection job in progress.

See \url{https://www.paws-r-sdk.com/docs/comprehendmedical_stop_entities_detection_v2_job/} for full documentation.
}
\keyword{internal}
