% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehendmedical_operations.R
\name{comprehendmedical_stop_icd10cm_inference_job}
\alias{comprehendmedical_stop_icd10cm_inference_job}
\title{Stops an InferICD10CM inference job in progress}
\usage{
comprehendmedical_stop_icd10cm_inference_job(JobId)
}
\arguments{
\item{JobId}{[required] The identifier of the job.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  JobId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Stops an InferICD10CM inference job in progress.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$stop_icd10cm_inference_job(
  JobId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
