% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehendmedical_operations.R
\name{comprehendmedical_stop_phi_detection_job}
\alias{comprehendmedical_stop_phi_detection_job}
\title{Stops a protected health information (PHI) detection job in progress}
\usage{
comprehendmedical_stop_phi_detection_job(JobId)
}
\arguments{
\item{JobId}{[required] The identifier of the PHI detection job to stop.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  JobId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Stops a protected health information (PHI) detection job in progress.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$stop_phi_detection_job(
  JobId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
