% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehendmedical_operations.R
\name{comprehendmedical_stop_rx_norm_inference_job}
\alias{comprehendmedical_stop_rx_norm_inference_job}
\title{Stops an InferRxNorm inference job in progress}
\usage{
comprehendmedical_stop_rx_norm_inference_job(JobId)
}
\arguments{
\item{JobId}{[required] The identifier of the job.}
}
\value{
A list with the following syntax:\preformatted{list(
  JobId = "string"
)
}
}
\description{
Stops an InferRxNorm inference job in progress.
}
\section{Request syntax}{
\preformatted{svc$stop_rx_norm_inference_job(
  JobId = "string"
)
}
}

\keyword{internal}
