% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticinference_operations.R
\name{elasticinference_describe_accelerator_types}
\alias{elasticinference_describe_accelerator_types}
\title{Amazon Elastic Inference is no longer available}
\usage{
elasticinference_describe_accelerator_types()
}
\description{
Amazon Elastic Inference is no longer available.

See \url{https://www.paws-r-sdk.com/docs/elasticinference_describe_accelerator_types/} for full documentation.
}
\keyword{internal}
