% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticinference_operations.R
\name{elasticinference_list_tags_for_resource}
\alias{elasticinference_list_tags_for_resource}
\title{Returns all tags of an Elastic Inference Accelerator}
\usage{
elasticinference_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the Elastic Inference Accelerator to list the tags for.}
}
\description{
Returns all tags of an Elastic Inference Accelerator.

See \url{https://paws-r.github.io/docs/elasticinference/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
