% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticinference_operations.R
\name{elasticinference_tag_resource}
\alias{elasticinference_tag_resource}
\title{Amazon Elastic Inference is no longer available}
\usage{
elasticinference_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The ARN of the Elastic Inference Accelerator to tag.}

\item{tags}{[required] The tags to add to the Elastic Inference Accelerator.}
}
\description{
Amazon Elastic Inference is no longer available.

See \url{https://www.paws-r-sdk.com/docs/elasticinference_tag_resource/} for full documentation.
}
\keyword{internal}
