% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_create_monitor}
\alias{forecastservice_create_monitor}
\title{Creates a predictor monitor resource for an existing auto predictor}
\usage{
forecastservice_create_monitor(MonitorName, ResourceArn, Tags = NULL)
}
\arguments{
\item{MonitorName}{[required] The name of the monitor resource.}

\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the predictor to monitor.}

\item{Tags}{A list of
\href{https://docs.aws.amazon.com/forecast/latest/dg/tagging-forecast-resources.html}{tags}
to apply to the monitor resource.}
}
\description{
Creates a predictor monitor resource for an existing auto predictor. Predictor monitoring allows you to see how your predictor's performance changes over time. For more information, see \href{https://docs.aws.amazon.com/forecast/latest/dg/predictor-monitoring.html}{Predictor Monitoring}.

See \url{https://paws-r.github.io/docs/forecastservice/create_monitor.html} for full documentation.
}
\keyword{internal}
