% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_delete_dataset}
\alias{forecastservice_delete_dataset}
\title{Deletes an Amazon Forecast dataset that was created using the
CreateDataset operation}
\usage{
forecastservice_delete_dataset(DatasetArn)
}
\arguments{
\item{DatasetArn}{[required] The Amazon Resource Name (ARN) of the dataset to delete.}
}
\description{
Deletes an Amazon Forecast dataset that was created using the \code{\link[=forecastservice_create_dataset]{create_dataset}} operation. You can only delete datasets that have a status of \code{ACTIVE} or \code{CREATE_FAILED}. To get the status use the \code{\link[=forecastservice_describe_dataset]{describe_dataset}} operation.

See \url{https://www.paws-r-sdk.com/docs/forecastservice_delete_dataset/} for full documentation.
}
\keyword{internal}
