% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_delete_dataset_group}
\alias{forecastservice_delete_dataset_group}
\title{Deletes a dataset group created using the CreateDatasetGroup operation}
\usage{
forecastservice_delete_dataset_group(DatasetGroupArn)
}
\arguments{
\item{DatasetGroupArn}{[required] The Amazon Resource Name (ARN) of the dataset group to delete.}
}
\description{
Deletes a dataset group created using the \code{\link[=forecastservice_create_dataset_group]{create_dataset_group}} operation. You can only delete dataset groups that have a status of \code{ACTIVE}, \code{CREATE_FAILED}, or \code{UPDATE_FAILED}. To get the status, use the \code{\link[=forecastservice_describe_dataset_group]{describe_dataset_group}} operation.

See \url{https://www.paws-r-sdk.com/docs/forecastservice_delete_dataset_group/} for full documentation.
}
\keyword{internal}
