% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_delete_explainability_export}
\alias{forecastservice_delete_explainability_export}
\title{Deletes an Explainability export}
\usage{
forecastservice_delete_explainability_export(ExplainabilityExportArn)
}
\arguments{
\item{ExplainabilityExportArn}{[required] The Amazon Resource Name (ARN) of the Explainability export to delete.}
}
\description{
Deletes an Explainability export.

See \url{https://www.paws-r-sdk.com/docs/forecastservice_delete_explainability_export/} for full documentation.
}
\keyword{internal}
