% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_delete_monitor}
\alias{forecastservice_delete_monitor}
\title{Deletes a monitor resource}
\usage{
forecastservice_delete_monitor(MonitorArn)
}
\arguments{
\item{MonitorArn}{[required] The Amazon Resource Name (ARN) of the monitor resource to delete.}
}
\description{
Deletes a monitor resource. You can only delete a monitor resource with a status of \code{ACTIVE}, \code{ACTIVE_STOPPED}, \code{CREATE_FAILED}, or \code{CREATE_STOPPED}.

See \url{https://www.paws-r-sdk.com/docs/forecastservice_delete_monitor/} for full documentation.
}
\keyword{internal}
