% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_delete_predictor_backtest_export_job}
\alias{forecastservice_delete_predictor_backtest_export_job}
\title{Deletes a predictor backtest export job}
\usage{
forecastservice_delete_predictor_backtest_export_job(
  PredictorBacktestExportJobArn
)
}
\arguments{
\item{PredictorBacktestExportJobArn}{[required] The Amazon Resource Name (ARN) of the predictor backtest export job to
delete.}
}
\description{
Deletes a predictor backtest export job.

See \url{https://paws-r.github.io/docs/forecastservice/delete_predictor_backtest_export_job.html} for full documentation.
}
\keyword{internal}
