% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_delete_what_if_analysis}
\alias{forecastservice_delete_what_if_analysis}
\title{Deletes a what-if analysis created using the CreateWhatIfAnalysis
operation}
\usage{
forecastservice_delete_what_if_analysis(WhatIfAnalysisArn)
}
\arguments{
\item{WhatIfAnalysisArn}{[required] The Amazon Resource Name (ARN) of the what-if analysis that you want to
delete.}
}
\description{
Deletes a what-if analysis created using the \code{\link[=forecastservice_create_what_if_analysis]{create_what_if_analysis}} operation. You can delete only what-if analyses that have a status of \code{ACTIVE} or \code{CREATE_FAILED}. To get the status, use the \code{\link[=forecastservice_describe_what_if_analysis]{describe_what_if_analysis}} operation.

See \url{https://paws-r.github.io/docs/forecastservice/delete_what_if_analysis.html} for full documentation.
}
\keyword{internal}
