% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_describe_dataset}
\alias{forecastservice_describe_dataset}
\title{Describes an Amazon Forecast dataset created using the CreateDataset
operation}
\usage{
forecastservice_describe_dataset(DatasetArn)
}
\arguments{
\item{DatasetArn}{[required] The Amazon Resource Name (ARN) of the dataset.}
}
\description{
Describes an Amazon Forecast dataset created using the \code{\link[=forecastservice_create_dataset]{create_dataset}} operation.

See \url{https://www.paws-r-sdk.com/docs/forecastservice_describe_dataset/} for full documentation.
}
\keyword{internal}
