% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_describe_dataset_import_job}
\alias{forecastservice_describe_dataset_import_job}
\title{Describes a dataset import job created using the CreateDatasetImportJob
operation}
\usage{
forecastservice_describe_dataset_import_job(DatasetImportJobArn)
}
\arguments{
\item{DatasetImportJobArn}{[required] The Amazon Resource Name (ARN) of the dataset import job.}
}
\description{
Describes a dataset import job created using the \code{\link[=forecastservice_create_dataset_import_job]{create_dataset_import_job}} operation.

See \url{https://paws-r.github.io/docs/forecastservice/describe_dataset_import_job.html} for full documentation.
}
\keyword{internal}
