% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_describe_explainability}
\alias{forecastservice_describe_explainability}
\title{Describes an Explainability resource created using the
CreateExplainability operation}
\usage{
forecastservice_describe_explainability(ExplainabilityArn)
}
\arguments{
\item{ExplainabilityArn}{[required] The Amazon Resource Name (ARN) of the Explaianability to describe.}
}
\description{
Describes an Explainability resource created using the \code{\link[=forecastservice_create_explainability]{create_explainability}} operation.

See \url{https://paws-r.github.io/docs/forecastservice/describe_explainability.html} for full documentation.
}
\keyword{internal}
