% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_describe_forecast}
\alias{forecastservice_describe_forecast}
\title{Describes a forecast created using the CreateForecast operation}
\usage{
forecastservice_describe_forecast(ForecastArn)
}
\arguments{
\item{ForecastArn}{[required] The Amazon Resource Name (ARN) of the forecast.}
}
\description{
Describes a forecast created using the \code{\link[=forecastservice_create_forecast]{create_forecast}} operation.

See \url{https://paws-r.github.io/docs/forecastservice/describe_forecast.html} for full documentation.
}
\keyword{internal}
