% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_describe_forecast_export_job}
\alias{forecastservice_describe_forecast_export_job}
\title{Describes a forecast export job created using the
CreateForecastExportJob operation}
\usage{
forecastservice_describe_forecast_export_job(ForecastExportJobArn)
}
\arguments{
\item{ForecastExportJobArn}{[required] The Amazon Resource Name (ARN) of the forecast export job.}
}
\description{
Describes a forecast export job created using the \code{\link[=forecastservice_create_forecast_export_job]{create_forecast_export_job}} operation.

See \url{https://paws-r.github.io/docs/forecastservice/describe_forecast_export_job.html} for full documentation.
}
\keyword{internal}
