% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_describe_monitor}
\alias{forecastservice_describe_monitor}
\title{Describes a monitor resource}
\usage{
forecastservice_describe_monitor(MonitorArn)
}
\arguments{
\item{MonitorArn}{[required] The Amazon Resource Name (ARN) of the monitor resource to describe.}
}
\description{
Describes a monitor resource. In addition to listing the properties provided in the \code{\link[=forecastservice_create_monitor]{create_monitor}} request, this operation lists the following properties:

See \url{https://www.paws-r-sdk.com/docs/forecastservice_describe_monitor/} for full documentation.
}
\keyword{internal}
