% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_describe_predictor}
\alias{forecastservice_describe_predictor}
\title{This operation is only valid for legacy predictors created with
CreatePredictor}
\usage{
forecastservice_describe_predictor(PredictorArn)
}
\arguments{
\item{PredictorArn}{[required] The Amazon Resource Name (ARN) of the predictor that you want
information about.}
}
\description{
This operation is only valid for legacy predictors created with CreatePredictor. If you are not using a legacy predictor, use \code{\link[=forecastservice_describe_auto_predictor]{describe_auto_predictor}}.

See \url{https://www.paws-r-sdk.com/docs/forecastservice_describe_predictor/} for full documentation.
}
\keyword{internal}
