% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_describe_predictor_backtest_export_job}
\alias{forecastservice_describe_predictor_backtest_export_job}
\title{Describes a predictor backtest export job created using the
CreatePredictorBacktestExportJob operation}
\usage{
forecastservice_describe_predictor_backtest_export_job(
  PredictorBacktestExportJobArn
)
}
\arguments{
\item{PredictorBacktestExportJobArn}{[required] The Amazon Resource Name (ARN) of the predictor backtest export job.}
}
\description{
Describes a predictor backtest export job created using the \code{\link[=forecastservice_create_predictor_backtest_export_job]{create_predictor_backtest_export_job}} operation.

See \url{https://www.paws-r-sdk.com/docs/forecastservice_describe_predictor_backtest_export_job/} for full documentation.
}
\keyword{internal}
