% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_describe_what_if_analysis}
\alias{forecastservice_describe_what_if_analysis}
\title{Describes the what-if analysis created using the CreateWhatIfAnalysis
operation}
\usage{
forecastservice_describe_what_if_analysis(WhatIfAnalysisArn)
}
\arguments{
\item{WhatIfAnalysisArn}{[required] The Amazon Resource Name (ARN) of the what-if analysis that you are
interested in.}
}
\description{
Describes the what-if analysis created using the \code{\link[=forecastservice_create_what_if_analysis]{create_what_if_analysis}} operation.

See \url{https://www.paws-r-sdk.com/docs/forecastservice_describe_what_if_analysis/} for full documentation.
}
\keyword{internal}
