% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_describe_what_if_forecast}
\alias{forecastservice_describe_what_if_forecast}
\title{Describes the what-if forecast created using the CreateWhatIfForecast
operation}
\usage{
forecastservice_describe_what_if_forecast(WhatIfForecastArn)
}
\arguments{
\item{WhatIfForecastArn}{[required] The Amazon Resource Name (ARN) of the what-if forecast that you are
interested in.}
}
\description{
Describes the what-if forecast created using the \code{\link[=forecastservice_create_what_if_forecast]{create_what_if_forecast}} operation.

See \url{https://www.paws-r-sdk.com/docs/forecastservice_describe_what_if_forecast/} for full documentation.
}
\keyword{internal}
