% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_describe_what_if_forecast_export}
\alias{forecastservice_describe_what_if_forecast_export}
\title{Describes the what-if forecast export created using the
CreateWhatIfForecastExport operation}
\usage{
forecastservice_describe_what_if_forecast_export(WhatIfForecastExportArn)
}
\arguments{
\item{WhatIfForecastExportArn}{[required] The Amazon Resource Name (ARN) of the what-if forecast export that you
are interested in.}
}
\description{
Describes the what-if forecast export created using the \code{\link[=forecastservice_create_what_if_forecast_export]{create_what_if_forecast_export}} operation.

See \url{https://paws-r.github.io/docs/forecastservice/describe_what_if_forecast_export.html} for full documentation.
}
\keyword{internal}
