% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_get_accuracy_metrics}
\alias{forecastservice_get_accuracy_metrics}
\title{Provides metrics on the accuracy of the models that were trained by the
CreatePredictor operation}
\usage{
forecastservice_get_accuracy_metrics(PredictorArn)
}
\arguments{
\item{PredictorArn}{[required] The Amazon Resource Name (ARN) of the predictor to get metrics for.}
}
\description{
Provides metrics on the accuracy of the models that were trained by the \code{\link[=forecastservice_create_predictor]{create_predictor}} operation. Use metrics to see how well the model performed and to decide whether to use the predictor to generate a forecast. For more information, see \href{https://docs.aws.amazon.com/forecast/latest/dg/metrics.html}{Predictor Metrics}.

See \url{https://paws-r.github.io/docs/forecastservice/get_accuracy_metrics.html} for full documentation.
}
\keyword{internal}
