% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_list_dataset_groups}
\alias{forecastservice_list_dataset_groups}
\title{Returns a list of dataset groups created using the CreateDatasetGroup
operation}
\usage{
forecastservice_list_dataset_groups(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{If the result of the previous request was truncated, the response
includes a \code{NextToken}. To retrieve the next set of results, use the
token in the next request. Tokens expire after 24 hours.}

\item{MaxResults}{The number of items to return in the response.}
}
\description{
Returns a list of dataset groups created using the \code{\link[=forecastservice_create_dataset_group]{create_dataset_group}} operation. For each dataset group, this operation returns a summary of its properties, including its Amazon Resource Name (ARN). You can retrieve the complete set of properties by using the dataset group ARN with the \code{\link[=forecastservice_describe_dataset_group]{describe_dataset_group}} operation.

See \url{https://www.paws-r-sdk.com/docs/forecastservice_list_dataset_groups/} for full documentation.
}
\keyword{internal}
