% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_list_predictor_backtest_export_jobs}
\alias{forecastservice_list_predictor_backtest_export_jobs}
\title{Returns a list of predictor backtest export jobs created using the
CreatePredictorBacktestExportJob operation}
\usage{
forecastservice_list_predictor_backtest_export_jobs(
  NextToken = NULL,
  MaxResults = NULL,
  Filters = NULL
)
}
\arguments{
\item{NextToken}{If the result of the previous request was truncated, the response
includes a NextToken. To retrieve the next set of results, use the token
in the next request. Tokens expire after 24 hours.}

\item{MaxResults}{The number of items to return in the response.}

\item{Filters}{An array of filters. For each filter, provide a condition and a match
statement. The condition is either \code{IS} or \code{IS_NOT}, which specifies
whether to include or exclude the predictor backtest export jobs that
match the statement from the list. The match statement consists of a key
and a value.

\strong{Filter properties}
\itemize{
\item \code{Condition} - The condition to apply. Valid values are \code{IS} and
\code{IS_NOT}. To include the predictor backtest export jobs that match
the statement, specify \code{IS}. To exclude matching predictor backtest
export jobs, specify \code{IS_NOT}.
\item \code{Key} - The name of the parameter to filter on. Valid values are
\code{PredictorArn} and \code{Status}.
\item \code{Value} - The value to match.
}}
}
\description{
Returns a list of predictor backtest export jobs created using the \code{\link[=forecastservice_create_predictor_backtest_export_job]{create_predictor_backtest_export_job}} operation. This operation returns a summary for each backtest export job. You can filter the list using an array of Filter objects.

See \url{https://www.paws-r-sdk.com/docs/forecastservice_list_predictor_backtest_export_jobs/} for full documentation.
}
\keyword{internal}
