% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_stop_resource}
\alias{forecastservice_stop_resource}
\title{Stops a resource}
\usage{
forecastservice_stop_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) that identifies the resource to stop. The
supported ARNs are \code{DatasetImportJobArn}, \code{PredictorArn},
\code{PredictorBacktestExportJobArn}, \code{ForecastArn}, \code{ForecastExportJobArn},
\code{ExplainabilityArn}, and \code{ExplainabilityExportArn}.}
}
\description{
Stops a resource.

See \url{https://paws-r.github.io/docs/forecastservice/stop_resource.html} for full documentation.
}
\keyword{internal}
