% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_untag_resource}
\alias{forecastservice_untag_resource}
\title{Deletes the specified tags from a resource}
\usage{
forecastservice_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) that identifies the resource for which to
list the tags.}

\item{TagKeys}{[required] The keys of the tags to be removed.}
}
\description{
Deletes the specified tags from a resource.

See \url{https://paws-r.github.io/docs/forecastservice/untag_resource.html} for full documentation.
}
\keyword{internal}
