% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_update_dataset_group}
\alias{forecastservice_update_dataset_group}
\title{Replaces the datasets in a dataset group with the specified datasets}
\usage{
forecastservice_update_dataset_group(DatasetGroupArn, DatasetArns)
}
\arguments{
\item{DatasetGroupArn}{[required] The ARN of the dataset group.}

\item{DatasetArns}{[required] An array of the Amazon Resource Names (ARNs) of the datasets to add to
the dataset group.}
}
\description{
Replaces the datasets in a dataset group with the specified datasets.

See \url{https://paws-r.github.io/docs/forecastservice/update_dataset_group.html} for full documentation.
}
\keyword{internal}
