% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_batch_create_variable}
\alias{frauddetector_batch_create_variable}
\title{Creates a batch of variables}
\usage{
frauddetector_batch_create_variable(variableEntries, tags = NULL)
}
\arguments{
\item{variableEntries}{[required] The list of variables for the batch create variable request.}

\item{tags}{A collection of key and value pairs.}
}
\description{
Creates a batch of variables.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_batch_create_variable/} for full documentation.
}
\keyword{internal}
