% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_batch_get_variable}
\alias{frauddetector_batch_get_variable}
\title{Gets a batch of variables}
\usage{
frauddetector_batch_get_variable(names)
}
\arguments{
\item{names}{[required] The list of variable names to get.}
}
\description{
Gets a batch of variables.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_batch_get_variable/} for full documentation.
}
\keyword{internal}
