% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_cancel_batch_prediction_job}
\alias{frauddetector_cancel_batch_prediction_job}
\title{Cancels the specified batch prediction job}
\usage{
frauddetector_cancel_batch_prediction_job(jobId)
}
\arguments{
\item{jobId}{[required] The ID of the batch prediction job to cancel.}
}
\description{
Cancels the specified batch prediction job.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_cancel_batch_prediction_job/} for full documentation.
}
\keyword{internal}
