% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_create_list}
\alias{frauddetector_create_list}
\title{Creates a list}
\usage{
frauddetector_create_list(
  name,
  elements = NULL,
  variableType = NULL,
  description = NULL,
  tags = NULL
)
}
\arguments{
\item{name}{[required] The name of the list.}

\item{elements}{The names of the elements, if providing. You can also create an empty
list and add elements later using the
\code{\link[=frauddetector_update_list]{update_list}} API.}

\item{variableType}{The variable type of the list. You can only assign the variable type
with String data type. For more information, see \href{https://docs.aws.amazon.com/frauddetector/latest/ug/create-a-variable.html#variable-types}{Variable types}.}

\item{description}{The description of the list.}

\item{tags}{A collection of the key and value pairs.}
}
\description{
Creates a list.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_create_list/} for full documentation.
}
\keyword{internal}
