% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_batch_import_job}
\alias{frauddetector_delete_batch_import_job}
\title{Deletes the specified batch import job ID record}
\usage{
frauddetector_delete_batch_import_job(jobId)
}
\arguments{
\item{jobId}{[required] The ID of the batch import job to delete.}
}
\description{
Deletes the specified batch import job ID record. This action does not delete the data that was batch imported.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_delete_batch_import_job/} for full documentation.
}
\keyword{internal}
