% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_batch_prediction_job}
\alias{frauddetector_delete_batch_prediction_job}
\title{Deletes a batch prediction job}
\usage{
frauddetector_delete_batch_prediction_job(jobId)
}
\arguments{
\item{jobId}{[required] The ID of the batch prediction job to delete.}
}
\description{
Deletes a batch prediction job.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_delete_batch_prediction_job/} for full documentation.
}
\keyword{internal}
