% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_detector}
\alias{frauddetector_delete_detector}
\title{Deletes the detector}
\usage{
frauddetector_delete_detector(detectorId)
}
\arguments{
\item{detectorId}{[required] The ID of the detector to delete.}
}
\description{
Deletes the detector. Before deleting a detector, you must first delete all detector versions and rule versions associated with the detector.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_delete_detector/} for full documentation.
}
\keyword{internal}
