% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_detector_version}
\alias{frauddetector_delete_detector_version}
\title{Deletes the detector version}
\usage{
frauddetector_delete_detector_version(detectorId, detectorVersionId)
}
\arguments{
\item{detectorId}{[required] The ID of the parent detector for the detector version to delete.}

\item{detectorVersionId}{[required] The ID of the detector version to delete.}
}
\description{
Deletes the detector version. You cannot delete detector versions that are in \code{ACTIVE} status.

See \url{https://paws-r.github.io/docs/frauddetector/delete_detector_version.html} for full documentation.
}
\keyword{internal}
