% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_entity_type}
\alias{frauddetector_delete_entity_type}
\title{Deletes an entity type}
\usage{
frauddetector_delete_entity_type(name)
}
\arguments{
\item{name}{[required] The name of the entity type to delete.}
}
\description{
Deletes an entity type.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_delete_entity_type/} for full documentation.
}
\keyword{internal}
