% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_event_type}
\alias{frauddetector_delete_event_type}
\title{Deletes an event type}
\usage{
frauddetector_delete_event_type(name)
}
\arguments{
\item{name}{[required] The name of the event type to delete.}
}
\description{
Deletes an event type.

See \url{https://paws-r.github.io/docs/frauddetector/delete_event_type.html} for full documentation.
}
\keyword{internal}
