% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_events_by_event_type}
\alias{frauddetector_delete_events_by_event_type}
\title{Deletes all events of a particular event type}
\usage{
frauddetector_delete_events_by_event_type(eventTypeName)
}
\arguments{
\item{eventTypeName}{[required] The name of the event type.}
}
\description{
Deletes all events of a particular event type.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_delete_events_by_event_type/} for full documentation.
}
\keyword{internal}
