% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_label}
\alias{frauddetector_delete_label}
\title{Deletes a label}
\usage{
frauddetector_delete_label(name)
}
\arguments{
\item{name}{[required] The name of the label to delete.}
}
\description{
Deletes a label.

See \url{https://paws-r.github.io/docs/frauddetector/delete_label.html} for full documentation.
}
\keyword{internal}
