% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_model}
\alias{frauddetector_delete_model}
\title{Deletes a model}
\usage{
frauddetector_delete_model(modelId, modelType)
}
\arguments{
\item{modelId}{[required] The model ID of the model to delete.}

\item{modelType}{[required] The model type of the model to delete.}
}
\description{
Deletes a model.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_delete_model/} for full documentation.
}
\keyword{internal}
