% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_model_version}
\alias{frauddetector_delete_model_version}
\title{Deletes a model version}
\usage{
frauddetector_delete_model_version(modelId, modelType, modelVersionNumber)
}
\arguments{
\item{modelId}{[required] The model ID of the model version to delete.}

\item{modelType}{[required] The model type of the model version to delete.}

\item{modelVersionNumber}{[required] The model version number of the model version to delete.}
}
\description{
Deletes a model version.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_delete_model_version/} for full documentation.
}
\keyword{internal}
