% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_rule}
\alias{frauddetector_delete_rule}
\title{Deletes the rule}
\usage{
frauddetector_delete_rule(rule)
}
\arguments{
\item{rule}{[required]}
}
\description{
Deletes the rule. You cannot delete a rule if it is used by an \code{ACTIVE} or \code{INACTIVE} detector version.

See \url{https://paws-r.github.io/docs/frauddetector/delete_rule.html} for full documentation.
}
\keyword{internal}
