% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_variable}
\alias{frauddetector_delete_variable}
\title{Deletes a variable}
\usage{
frauddetector_delete_variable(name)
}
\arguments{
\item{name}{[required] The name of the variable to delete.}
}
\description{
Deletes a variable.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_delete_variable/} for full documentation.
}
\keyword{internal}
