% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_describe_detector}
\alias{frauddetector_describe_detector}
\title{Gets all versions for a specified detector}
\usage{
frauddetector_describe_detector(
  detectorId,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{detectorId}{[required] The detector ID.}

\item{nextToken}{The next token from the previous response.}

\item{maxResults}{The maximum number of results to return for the request.}
}
\description{
Gets all versions for a specified detector.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_describe_detector/} for full documentation.
}
\keyword{internal}
