% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_describe_model_versions}
\alias{frauddetector_describe_model_versions}
\title{Gets all of the model versions for the specified model type or for the
specified model type and model ID}
\usage{
frauddetector_describe_model_versions(
  modelId = NULL,
  modelVersionNumber = NULL,
  modelType = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{modelId}{The model ID.}

\item{modelVersionNumber}{The model version number.}

\item{modelType}{The model type.}

\item{nextToken}{The next token from the previous results.}

\item{maxResults}{The maximum number of results to return.}
}
\description{
Gets all of the model versions for the specified model type or for the specified model type and model ID. You can also get details for a single, specified model version.

See \url{https://paws-r.github.io/docs/frauddetector/describe_model_versions.html} for full documentation.
}
\keyword{internal}
