% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_get_batch_import_jobs}
\alias{frauddetector_get_batch_import_jobs}
\title{Gets all batch import jobs or a specific job of the specified ID}
\usage{
frauddetector_get_batch_import_jobs(
  jobId = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{jobId}{The ID of the batch import job to get.}

\item{maxResults}{The maximum number of objects to return for request.}

\item{nextToken}{The next token from the previous request.}
}
\description{
Gets all batch import jobs or a specific job of the specified ID. This is a paginated API. If you provide a null \code{maxResults}, this action retrieves a maximum of 50 records per page. If you provide a \code{maxResults}, the value must be between 1 and 50. To get the next page results, provide the pagination token from the \code{GetBatchImportJobsResponse} as part of your request. A null pagination token fetches the records from the beginning.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_get_batch_import_jobs/} for full documentation.
}
\keyword{internal}
