% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_get_batch_prediction_jobs}
\alias{frauddetector_get_batch_prediction_jobs}
\title{Gets all batch prediction jobs or a specific job if you specify a job ID}
\usage{
frauddetector_get_batch_prediction_jobs(
  jobId = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{jobId}{The batch prediction job for which to get the details.}

\item{maxResults}{The maximum number of objects to return for the request.}

\item{nextToken}{The next token from the previous request.}
}
\description{
Gets all batch prediction jobs or a specific job if you specify a job ID. This is a paginated API. If you provide a null maxResults, this action retrieves a maximum of 50 records per page. If you provide a maxResults, the value must be between 1 and 50. To get the next page results, provide the pagination token from the GetBatchPredictionJobsResponse as part of your request. A null pagination token fetches the records from the beginning.

See \url{https://paws-r.github.io/docs/frauddetector/get_batch_prediction_jobs.html} for full documentation.
}
\keyword{internal}
