% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_get_delete_events_by_event_type_status}
\alias{frauddetector_get_delete_events_by_event_type_status}
\title{Retrieves the status of a DeleteEventsByEventType action}
\usage{
frauddetector_get_delete_events_by_event_type_status(eventTypeName)
}
\arguments{
\item{eventTypeName}{[required] Name of event type for which to get the deletion status.}
}
\description{
Retrieves the status of a \code{\link[=frauddetector_delete_events_by_event_type]{delete_events_by_event_type}} action.

See \url{https://paws-r.github.io/docs/frauddetector/get_delete_events_by_event_type_status.html} for full documentation.
}
\keyword{internal}
