% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_get_detector_version}
\alias{frauddetector_get_detector_version}
\title{Gets a particular detector version}
\usage{
frauddetector_get_detector_version(detectorId, detectorVersionId)
}
\arguments{
\item{detectorId}{[required] The detector ID.}

\item{detectorVersionId}{[required] The detector version ID.}
}
\description{
Gets a particular detector version.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_get_detector_version/} for full documentation.
}
\keyword{internal}
