% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_get_event}
\alias{frauddetector_get_event}
\title{Retrieves details of events stored with Amazon Fraud Detector}
\usage{
frauddetector_get_event(eventId, eventTypeName)
}
\arguments{
\item{eventId}{[required] The ID of the event to retrieve.}

\item{eventTypeName}{[required] The event type of the event to retrieve.}
}
\description{
Retrieves details of events stored with Amazon Fraud Detector. This action does not retrieve prediction results.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_get_event/} for full documentation.
}
\keyword{internal}
