% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_get_event_types}
\alias{frauddetector_get_event_types}
\title{Gets all event types or a specific event type if name is provided}
\usage{
frauddetector_get_event_types(name = NULL, nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{name}{The name.}

\item{nextToken}{The next token for the subsequent request.}

\item{maxResults}{The maximum number of objects to return for the request.}
}
\description{
Gets all event types or a specific event type if name is provided. This is a paginated API. If you provide a null \code{maxResults}, this action retrieves a maximum of 10 records per page. If you provide a \code{maxResults}, the value must be between 5 and 10. To get the next page results, provide the pagination token from the \code{GetEventTypesResponse} as part of your request. A null pagination token fetches the records from the beginning.

See \url{https://paws-r.github.io/docs/frauddetector/get_event_types.html} for full documentation.
}
\keyword{internal}
