% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_get_kms_encryption_key}
\alias{frauddetector_get_kms_encryption_key}
\title{Gets the encryption key if a KMS key has been specified to be used to
encrypt content in Amazon Fraud Detector}
\usage{
frauddetector_get_kms_encryption_key()
}
\description{
Gets the encryption key if a KMS key has been specified to be used to encrypt content in Amazon Fraud Detector.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_get_kms_encryption_key/} for full documentation.
}
\keyword{internal}
