% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_get_list_elements}
\alias{frauddetector_get_list_elements}
\title{Gets all the elements in the specified list}
\usage{
frauddetector_get_list_elements(name, nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{name}{[required] The name of the list.}

\item{nextToken}{The next token for the subsequent request.}

\item{maxResults}{The maximum number of objects to return for the request.}
}
\description{
Gets all the elements in the specified list.

See \url{https://paws-r.github.io/docs/frauddetector/get_list_elements.html} for full documentation.
}
\keyword{internal}
