% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_get_lists_metadata}
\alias{frauddetector_get_lists_metadata}
\title{Gets the metadata of either all the lists under the account or the
specified list}
\usage{
frauddetector_get_lists_metadata(
  name = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{name}{The name of the list.}

\item{nextToken}{The next token for the subsequent request.}

\item{maxResults}{The maximum number of objects to return for the request.}
}
\description{
Gets the metadata of either all the lists under the account or the specified list.

See \url{https://paws-r.github.io/docs/frauddetector/get_lists_metadata.html} for full documentation.
}
\keyword{internal}
