% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_get_model_version}
\alias{frauddetector_get_model_version}
\title{Gets the details of the specified model version}
\usage{
frauddetector_get_model_version(modelId, modelType, modelVersionNumber)
}
\arguments{
\item{modelId}{[required] The model ID.}

\item{modelType}{[required] The model type.}

\item{modelVersionNumber}{[required] The model version number.}
}
\description{
Gets the details of the specified model version.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_get_model_version/} for full documentation.
}
\keyword{internal}
