% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_get_models}
\alias{frauddetector_get_models}
\title{Gets one or more models}
\usage{
frauddetector_get_models(
  modelId = NULL,
  modelType = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{modelId}{The model ID.}

\item{modelType}{The model type.}

\item{nextToken}{The next token for the subsequent request.}

\item{maxResults}{The maximum number of objects to return for the request.}
}
\description{
Gets one or more models. Gets all models for the Amazon Web Services account if no model type and no model id provided. Gets all models for the Amazon Web Services account and model type, if the model type is specified but model id is not provided. Gets a specific model if (model type, model id) tuple is specified.

See \url{https://paws-r.github.io/docs/frauddetector/get_models.html} for full documentation.
}
\keyword{internal}
