% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_get_variables}
\alias{frauddetector_get_variables}
\title{Gets all of the variables or the specific variable}
\usage{
frauddetector_get_variables(name = NULL, nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{name}{The name of the variable.}

\item{nextToken}{The next page token of the get variable request.}

\item{maxResults}{The max size per page determined for the get variable request.}
}
\description{
Gets all of the variables or the specific variable. This is a paginated API. Providing null \code{maxSizePerPage} results in retrieving maximum of 100 records per page. If you provide \code{maxSizePerPage} the value must be between 50 and 100. To get the next page result, a provide a pagination token from \code{GetVariablesResult} as part of your request. Null pagination token fetches the records from the beginning.

See \url{https://paws-r.github.io/docs/frauddetector/get_variables.html} for full documentation.
}
\keyword{internal}
