% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_list_tags_for_resource}
\alias{frauddetector_list_tags_for_resource}
\title{Lists all tags associated with the resource}
\usage{
frauddetector_list_tags_for_resource(
  resourceARN,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{resourceARN}{[required] The ARN that specifies the resource whose tags you want to list.}

\item{nextToken}{The next token from the previous results.}

\item{maxResults}{The maximum number of objects to return for the request.}
}
\description{
Lists all tags associated with the resource. This is a paginated API. To get the next page results, provide the pagination token from the response as part of your request. A null pagination token fetches the records from the beginning.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
