% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_put_detector}
\alias{frauddetector_put_detector}
\title{Creates or updates a detector}
\usage{
frauddetector_put_detector(
  detectorId,
  description = NULL,
  eventTypeName,
  tags = NULL
)
}
\arguments{
\item{detectorId}{[required] The detector ID.}

\item{description}{The description of the detector.}

\item{eventTypeName}{[required] The name of the event type.}

\item{tags}{A collection of key and value pairs.}
}
\description{
Creates or updates a detector.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_put_detector/} for full documentation.
}
\keyword{internal}
