% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_put_entity_type}
\alias{frauddetector_put_entity_type}
\title{Creates or updates an entity type}
\usage{
frauddetector_put_entity_type(name, description = NULL, tags = NULL)
}
\arguments{
\item{name}{[required] The name of the entity type.}

\item{description}{The description.}

\item{tags}{A collection of key and value pairs.}
}
\description{
Creates or updates an entity type. An entity represents who is performing the event. As part of a fraud prediction, you pass the entity ID to indicate the specific entity who performed the event. An entity type classifies the entity. Example classifications include customer, merchant, or account.

See \url{https://paws-r.github.io/docs/frauddetector/put_entity_type.html} for full documentation.
}
\keyword{internal}
