% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_put_external_model}
\alias{frauddetector_put_external_model}
\title{Creates or updates an Amazon SageMaker model endpoint}
\usage{
frauddetector_put_external_model(
  modelEndpoint,
  modelSource,
  invokeModelEndpointRoleArn,
  inputConfiguration,
  outputConfiguration,
  modelEndpointStatus,
  tags = NULL
)
}
\arguments{
\item{modelEndpoint}{[required] The model endpoints name.}

\item{modelSource}{[required] The source of the model.}

\item{invokeModelEndpointRoleArn}{[required] The IAM role used to invoke the model endpoint.}

\item{inputConfiguration}{[required] The model endpoint input configuration.}

\item{outputConfiguration}{[required] The model endpoint output configuration.}

\item{modelEndpointStatus}{[required] The model endpoint’s status in Amazon Fraud Detector.}

\item{tags}{A collection of key and value pairs.}
}
\description{
Creates or updates an Amazon SageMaker model endpoint. You can also use this action to update the configuration of the model endpoint, including the IAM role and/or the mapped variables.

See \url{https://paws-r.github.io/docs/frauddetector/put_external_model.html} for full documentation.
}
\keyword{internal}
