% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_put_kms_encryption_key}
\alias{frauddetector_put_kms_encryption_key}
\title{Specifies the KMS key to be used to encrypt content in Amazon Fraud
Detector}
\usage{
frauddetector_put_kms_encryption_key(kmsEncryptionKeyArn)
}
\arguments{
\item{kmsEncryptionKeyArn}{[required] The KMS encryption key ARN.

The KMS key must be single-Region key. Amazon Fraud Detector does not
support multi-Region KMS key.}
}
\description{
Specifies the KMS key to be used to encrypt content in Amazon Fraud Detector.

See \url{https://paws-r.github.io/docs/frauddetector/put_kms_encryption_key.html} for full documentation.
}
\keyword{internal}
