% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_put_outcome}
\alias{frauddetector_put_outcome}
\title{Creates or updates an outcome}
\usage{
frauddetector_put_outcome(name, description = NULL, tags = NULL)
}
\arguments{
\item{name}{[required] The name of the outcome.}

\item{description}{The outcome description.}

\item{tags}{A collection of key and value pairs.}
}
\description{
Creates or updates an outcome.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_put_outcome/} for full documentation.
}
\keyword{internal}
