% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_tag_resource}
\alias{frauddetector_tag_resource}
\title{Assigns tags to a resource}
\usage{
frauddetector_tag_resource(resourceARN, tags)
}
\arguments{
\item{resourceARN}{[required] The resource ARN.}

\item{tags}{[required] The tags to assign to the resource.}
}
\description{
Assigns tags to a resource.

See \url{https://paws-r.github.io/docs/frauddetector/tag_resource.html} for full documentation.
}
\keyword{internal}
