% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_untag_resource}
\alias{frauddetector_untag_resource}
\title{Removes tags from a resource}
\usage{
frauddetector_untag_resource(resourceARN, tagKeys)
}
\arguments{
\item{resourceARN}{[required] The ARN of the resource from which to remove the tag.}

\item{tagKeys}{[required] The resource ARN.}
}
\description{
Removes tags from a resource.

See \url{https://paws-r.github.io/docs/frauddetector/untag_resource.html} for full documentation.
}
\keyword{internal}
