% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_update_event_label}
\alias{frauddetector_update_event_label}
\title{Updates the specified event with a new label}
\usage{
frauddetector_update_event_label(
  eventId,
  eventTypeName,
  assignedLabel,
  labelTimestamp
)
}
\arguments{
\item{eventId}{[required] The ID of the event associated with the label to update.}

\item{eventTypeName}{[required] The event type of the event associated with the label to update.}

\item{assignedLabel}{[required] The new label to assign to the event.}

\item{labelTimestamp}{[required] The timestamp associated with the label. The timestamp must be specified
using ISO 8601 standard in UTC.}
}
\description{
Updates the specified event with a new label.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_update_event_label/} for full documentation.
}
\keyword{internal}
