% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_update_list}
\alias{frauddetector_update_list}
\title{Updates a list}
\usage{
frauddetector_update_list(
  name,
  elements = NULL,
  description = NULL,
  updateMode = NULL,
  variableType = NULL
)
}
\arguments{
\item{name}{[required] The name of the list to update.}

\item{elements}{One or more list elements to add or replace. If you are providing the
elements, make sure to specify the \code{updateMode} to use.

If you are deleting all elements from the list, use \code{REPLACE} for the
\code{updateMode} and provide an empty list (0 elements).}

\item{description}{The new description.}

\item{updateMode}{The update mode (type).
\itemize{
\item Use \code{APPEND} if you are adding elements to the list.
\item Use \code{REPLACE} if you replacing existing elements in the list.
\item Use \code{REMOVE} if you are removing elements from the list.
}}

\item{variableType}{The variable type you want to assign to the list.

You cannot update a variable type of a list that already has a variable
type assigned to it. You can assign a variable type to a list only if
the list does not already have a variable type.}
}
\description{
Updates a list.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_update_list/} for full documentation.
}
\keyword{internal}
