% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_create_slot_type_version}
\alias{lexmodelbuildingservice_create_slot_type_version}
\title{Creates a new version of a slot type based on the $LATEST version of the
specified slot type}
\usage{
lexmodelbuildingservice_create_slot_type_version(name, checksum)
}
\arguments{
\item{name}{[required] The name of the slot type that you want to create a new version for. The
name is case sensitive.}

\item{checksum}{Checksum for the \code{$LATEST} version of the slot type that you want to
publish. If you specify a checksum and the \code{$LATEST} version of the slot
type has a different checksum, Amazon Lex returns a
\code{PreconditionFailedException} exception and doesn't publish the new
version. If you don't specify a checksum, Amazon Lex publishes the
\code{$LATEST} version.}
}
\description{
Creates a new version of a slot type based on the \code{$LATEST} version of
the specified slot type. If the \code{$LATEST} version of this resource has
not changed since the last version that you created, Amazon Lex doesn't
create a new version. It returns the last version that you created.
}
\details{
You can update only the \code{$LATEST} version of a slot type. You can't
update the numbered versions that you create with the
\code{CreateSlotTypeVersion} operation.

When you create a version of a slot type, Amazon Lex sets the version to
\enumerate{
\item Subsequent versions increment by 1. For more information, see
versioning-intro.
}

This operation requires permissions for the \code{lex:CreateSlotTypeVersion}
action.
}
\section{Request syntax}{
\preformatted{svc$create_slot_type_version(
  name = "string",
  checksum = "string"
)
}
}

\keyword{internal}
