% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_delete_bot}
\alias{lexmodelbuildingservice_delete_bot}
\title{Deletes all versions of the bot, including the $LATEST version}
\usage{
lexmodelbuildingservice_delete_bot(name)
}
\arguments{
\item{name}{[required] The name of the bot. The name is case sensitive.}
}
\description{
Deletes all versions of the bot, including the \verb{$LATEST} version. To delete a specific version of the bot, use the \code{\link[=lexmodelbuildingservice_delete_bot_version]{delete_bot_version}} operation. The \code{\link[=lexmodelbuildingservice_delete_bot]{delete_bot}} operation doesn't immediately remove the bot schema. Instead, it is marked for deletion and removed later.

See \url{https://www.paws-r-sdk.com/docs/lexmodelbuildingservice_delete_bot/} for full documentation.
}
\keyword{internal}
