% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_delete_bot_alias}
\alias{lexmodelbuildingservice_delete_bot_alias}
\title{Deletes an alias for the specified bot}
\usage{
lexmodelbuildingservice_delete_bot_alias(name, botName)
}
\arguments{
\item{name}{[required] The name of the alias to delete. The name is case sensitive.}

\item{botName}{[required] The name of the bot that the alias points to.}
}
\description{
Deletes an alias for the specified bot.
}
\details{
You can\'t delete an alias that is used in the association between a bot
and a messaging channel. If an alias is used in a channel association,
the \code{DeleteBot} operation returns a \code{ResourceInUseException} exception
that includes a reference to the channel association that refers to the
bot. You can remove the reference to the alias by deleting the channel
association. If you get the same exception again, delete the referring
association until the \code{DeleteBotAlias} operation is successful.
}
\section{Request syntax}{
\preformatted{svc$delete_bot_alias(
  name = "string",
  botName = "string"
)
}
}

\keyword{internal}
