% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_delete_bot_channel_association}
\alias{lexmodelbuildingservice_delete_bot_channel_association}
\title{Deletes the association between an Amazon Lex bot and a messaging
platform}
\usage{
lexmodelbuildingservice_delete_bot_channel_association(name, botName, botAlias)
}
\arguments{
\item{name}{[required] The name of the association. The name is case sensitive.}

\item{botName}{[required] The name of the Amazon Lex bot.}

\item{botAlias}{[required] An alias that points to the specific version of the Amazon Lex bot to
which this association is being made.}
}
\description{
Deletes the association between an Amazon Lex bot and a messaging platform.

See \url{https://www.paws-r-sdk.com/docs/lexmodelbuildingservice_delete_bot_channel_association/} for full documentation.
}
\keyword{internal}
