% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_delete_bot_version}
\alias{lexmodelbuildingservice_delete_bot_version}
\title{Deletes a specific version of a bot}
\usage{
lexmodelbuildingservice_delete_bot_version(name, version)
}
\arguments{
\item{name}{[required] The name of the bot.}

\item{version}{[required] The version of the bot to delete. You cannot delete the \verb{$LATEST}
version of the bot. To delete the \verb{$LATEST} version, use the
\code{\link[=lexmodelbuildingservice_delete_bot]{delete_bot}} operation.}
}
\value{
An empty list.
}
\description{
Deletes a specific version of a bot. To delete all versions of a bot,
use the \code{\link[=lexmodelbuildingservice_delete_bot]{delete_bot}} operation.

This operation requires permissions for the \code{lex:DeleteBotVersion}
action.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_bot_version(
  name = "string",
  version = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
