% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_delete_intent}
\alias{lexmodelbuildingservice_delete_intent}
\title{Deletes all versions of the intent, including the $LATEST version}
\usage{
lexmodelbuildingservice_delete_intent(name)
}
\arguments{
\item{name}{[required] The name of the intent. The name is case sensitive.}
}
\description{
Deletes all versions of the intent, including the \verb{$LATEST} version. To delete a specific version of the intent, use the \code{\link[=lexmodelbuildingservice_delete_intent_version]{delete_intent_version}} operation.

See \url{https://paws-r.github.io/docs/lexmodelbuildingservice/delete_intent.html} for full documentation.
}
\keyword{internal}
