% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_delete_slot_type}
\alias{lexmodelbuildingservice_delete_slot_type}
\title{Deletes all versions of the slot type, including the $LATEST version}
\usage{
lexmodelbuildingservice_delete_slot_type(name)
}
\arguments{
\item{name}{[required] The name of the slot type. The name is case sensitive.}
}
\description{
Deletes all versions of the slot type, including the \verb{$LATEST} version. To delete a specific version of the slot type, use the \code{\link[=lexmodelbuildingservice_delete_slot_type_version]{delete_slot_type_version}} operation.

See \url{https://www.paws-r-sdk.com/docs/lexmodelbuildingservice_delete_slot_type/} for full documentation.
}
\keyword{internal}
