% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_delete_slot_type_version}
\alias{lexmodelbuildingservice_delete_slot_type_version}
\title{Deletes a specific version of a slot type}
\usage{
lexmodelbuildingservice_delete_slot_type_version(name, version)
}
\arguments{
\item{name}{[required] The name of the slot type.}

\item{version}{[required] The version of the slot type to delete. You cannot delete the \verb{$LATEST}
version of the slot type. To delete the \verb{$LATEST} version, use the
\code{\link[=lexmodelbuildingservice_delete_slot_type]{delete_slot_type}}
operation.}
}
\description{
Deletes a specific version of a slot type. To delete all versions of a slot type, use the \code{\link[=lexmodelbuildingservice_delete_slot_type]{delete_slot_type}} operation.

See \url{https://paws-r.github.io/docs/lexmodelbuildingservice/delete_slot_type_version.html} for full documentation.
}
\keyword{internal}
