% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_get_bot}
\alias{lexmodelbuildingservice_get_bot}
\title{Returns metadata information for a specific bot}
\usage{
lexmodelbuildingservice_get_bot(name, versionOrAlias)
}
\arguments{
\item{name}{[required] The name of the bot. The name is case sensitive.}

\item{versionOrAlias}{[required] The version or alias of the bot.}
}
\description{
Returns metadata information for a specific bot. You must provide the bot name and the bot version or alias.

See \url{https://www.paws-r-sdk.com/docs/lexmodelbuildingservice_get_bot/} for full documentation.
}
\keyword{internal}
