% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_get_bot_alias}
\alias{lexmodelbuildingservice_get_bot_alias}
\title{Returns information about an Amazon Lex bot alias}
\usage{
lexmodelbuildingservice_get_bot_alias(name, botName)
}
\arguments{
\item{name}{[required] The name of the bot alias. The name is case sensitive.}

\item{botName}{[required] The name of the bot.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  name = "string",
  description = "string",
  botVersion = "string",
  botName = "string",
  lastUpdatedDate = as.POSIXct(
    "2015-01-01"
  ),
  createdDate = as.POSIXct(
    "2015-01-01"
  ),
  checksum = "string",
  conversationLogs = list(
    logSettings = list(
      list(
        logType = "AUDIO"|"TEXT",
        destination = "CLOUDWATCH_LOGS"|"S3",
        kmsKeyArn = "string",
        resourceArn = "string",
        resourcePrefix = "string"
      )
    ),
    iamRoleArn = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns information about an Amazon Lex bot alias. For more information
about aliases, see versioning-aliases.

This operation requires permissions for the \code{lex:GetBotAlias} action.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_bot_alias(
  name = "string",
  botName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
