% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_get_bots}
\alias{lexmodelbuildingservice_get_bots}
\title{Returns bot information as follows:}
\usage{
lexmodelbuildingservice_get_bots(
  nextToken = NULL,
  maxResults = NULL,
  nameContains = NULL
)
}
\arguments{
\item{nextToken}{A pagination token that fetches the next page of bots. If the response
to this call is truncated, Amazon Lex returns a pagination token in the
response. To fetch the next page of bots, specify the pagination token
in the next request.}

\item{maxResults}{The maximum number of bots to return in the response that the request
will return. The default is 10.}

\item{nameContains}{Substring to match in bot names. A bot will be returned if any part of
its name matches the substring. For example, "xyz" matches both "xyzabc"
and "abcxyz."}
}
\description{
Returns bot information as follows:

See \url{https://www.paws-r-sdk.com/docs/lexmodelbuildingservice_get_bots/} for full documentation.
}
\keyword{internal}
