% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_get_export}
\alias{lexmodelbuildingservice_get_export}
\title{Exports the contents of a Amazon Lex resource in a specified format}
\usage{
lexmodelbuildingservice_get_export(name, version, resourceType, exportType)
}
\arguments{
\item{name}{[required] The name of the bot to export.}

\item{version}{[required] The version of the bot to export.}

\item{resourceType}{[required] The type of resource to export.}

\item{exportType}{[required] The format of the exported data.}
}
\description{
Exports the contents of a Amazon Lex resource in a specified format.

See \url{https://www.paws-r-sdk.com/docs/lexmodelbuildingservice_get_export/} for full documentation.
}
\keyword{internal}
