% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_get_intents}
\alias{lexmodelbuildingservice_get_intents}
\title{Returns intent information as follows:}
\usage{
lexmodelbuildingservice_get_intents(
  nextToken = NULL,
  maxResults = NULL,
  nameContains = NULL
)
}
\arguments{
\item{nextToken}{A pagination token that fetches the next page of intents. If the
response to this API call is truncated, Amazon Lex returns a pagination
token in the response. To fetch the next page of intents, specify the
pagination token in the next request.}

\item{maxResults}{The maximum number of intents to return in the response. The default is
10.}

\item{nameContains}{Substring to match in intent names. An intent will be returned if any
part of its name matches the substring. For example, "xyz" matches both
"xyzabc" and "abcxyz."}
}
\description{
Returns intent information as follows:

See \url{https://paws-r.github.io/docs/lexmodelbuildingservice/get_intents.html} for full documentation.
}
\keyword{internal}
