% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_get_migration}
\alias{lexmodelbuildingservice_get_migration}
\title{Provides details about an ongoing or complete migration from an Amazon
Lex V1 bot to an Amazon Lex V2 bot}
\usage{
lexmodelbuildingservice_get_migration(migrationId)
}
\arguments{
\item{migrationId}{[required] The unique identifier of the migration to view. The \code{migrationID} is
returned by the operation.}
}
\description{
Provides details about an ongoing or complete migration from an Amazon Lex V1 bot to an Amazon Lex V2 bot. Use this operation to view the migration alerts and warnings related to the migration.

See \url{https://www.paws-r-sdk.com/docs/lexmodelbuildingservice_get_migration/} for full documentation.
}
\keyword{internal}
