% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_get_migrations}
\alias{lexmodelbuildingservice_get_migrations}
\title{Gets a list of migrations between Amazon Lex V1 and Amazon Lex V2}
\usage{
lexmodelbuildingservice_get_migrations(
  sortByAttribute = NULL,
  sortByOrder = NULL,
  v1BotNameContains = NULL,
  migrationStatusEquals = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{sortByAttribute}{The field to sort the list of migrations by. You can sort by the Amazon
Lex V1 bot name or the date and time that the migration was started.}

\item{sortByOrder}{The order so sort the list.}

\item{v1BotNameContains}{Filters the list to contain only bots whose name contains the specified
string. The string is matched anywhere in bot name.}

\item{migrationStatusEquals}{Filters the list to contain only migrations in the specified state.}

\item{maxResults}{The maximum number of migrations to return in the response. The default
is 10.}

\item{nextToken}{A pagination token that fetches the next page of migrations. If the
response to this operation is truncated, Amazon Lex returns a pagination
token in the response. To fetch the next page of migrations, specify the
pagination token in the request.}
}
\description{
Gets a list of migrations between Amazon Lex V1 and Amazon Lex V2.

See \url{https://www.paws-r-sdk.com/docs/lexmodelbuildingservice_get_migrations/} for full documentation.
}
\keyword{internal}
