% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_get_slot_type_versions}
\alias{lexmodelbuildingservice_get_slot_type_versions}
\title{Gets information about all versions of a slot type}
\usage{
lexmodelbuildingservice_get_slot_type_versions(
  name,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{name}{[required] The name of the slot type for which versions should be returned.}

\item{nextToken}{A pagination token for fetching the next page of slot type versions. If
the response to this call is truncated, Amazon Lex returns a pagination
token in the response. To fetch the next page of versions, specify the
pagination token in the next request.}

\item{maxResults}{The maximum number of slot type versions to return in the response. The
default is 10.}
}
\description{
Gets information about all versions of a slot type.

See \url{https://www.paws-r-sdk.com/docs/lexmodelbuildingservice_get_slot_type_versions/} for full documentation.
}
\keyword{internal}
