% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_put_bot_alias}
\alias{lexmodelbuildingservice_put_bot_alias}
\title{Creates an alias for the specified version of the bot or replaces an
alias for the specified bot}
\usage{
lexmodelbuildingservice_put_bot_alias(
  name,
  description = NULL,
  botVersion,
  botName,
  checksum = NULL,
  conversationLogs = NULL,
  tags = NULL
)
}
\arguments{
\item{name}{[required] The name of the alias. The name is \emph{not} case sensitive.}

\item{description}{A description of the alias.}

\item{botVersion}{[required] The version of the bot.}

\item{botName}{[required] The name of the bot.}

\item{checksum}{Identifies a specific revision of the \verb{$LATEST} version.

When you create a new bot alias, leave the \code{checksum} field blank. If
you specify a checksum you get a \code{BadRequestException} exception.

When you want to update a bot alias, set the \code{checksum} field to the
checksum of the most recent revision of the \verb{$LATEST} version. If you
don't specify the \code{ checksum} field, or if the checksum does not match
the \verb{$LATEST} version, you get a \code{PreconditionFailedException}
exception.}

\item{conversationLogs}{Settings for conversation logs for the alias.}

\item{tags}{A list of tags to add to the bot alias. You can only add tags when you
create an alias, you can't use the
\code{\link[=lexmodelbuildingservice_put_bot_alias]{put_bot_alias}} operation to
update the tags on a bot alias. To update tags, use the
\code{\link[=lexmodelbuildingservice_tag_resource]{tag_resource}} operation.}
}
\description{
Creates an alias for the specified version of the bot or replaces an alias for the specified bot. To change the version of the bot that the alias points to, replace the alias. For more information about aliases, see versioning-aliases.

See \url{https://paws-r.github.io/docs/lexmodelbuildingservice/put_bot_alias.html} for full documentation.
}
\keyword{internal}
