% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_start_import}
\alias{lexmodelbuildingservice_start_import}
\title{Starts a job to import a resource to Amazon Lex}
\usage{
lexmodelbuildingservice_start_import(payload, resourceType,
  mergeStrategy, tags)
}
\arguments{
\item{payload}{[required] A zip archive in binary format. The archive should contain one file, a
JSON file containing the resource to import. The resource should match
the type specified in the \code{resourceType} field.}

\item{resourceType}{[required] Specifies the type of resource to export. Each resource also exports any
resources that it depends on.
\itemize{
\item A bot exports dependent intents.
\item An intent exports dependent slot types.
}}

\item{mergeStrategy}{[required] Specifies the action that the
\code{\link[=lexmodelbuildingservice_start_import]{start_import}} operation should
take when there is an existing resource with the same name.
\itemize{
\item FAIL_ON_CONFLICT - The import operation is stopped on the first
conflict between a resource in the import file and an existing
resource. The name of the resource causing the conflict is in the
\code{failureReason} field of the response to the
\code{\link[=lexmodelbuildingservice_get_import]{get_import}} operation.

OVERWRITE_LATEST - The import operation proceeds even if there is a
conflict with an existing resource. The $LASTEST version of the
existing resource is overwritten with the data from the import file.
}}

\item{tags}{A list of tags to add to the imported bot. You can only add tags when
you import a bot, you can't add tags to an intent or slot type.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  name = "string",
  resourceType = "BOT"|"INTENT"|"SLOT_TYPE",
  mergeStrategy = "OVERWRITE_LATEST"|"FAIL_ON_CONFLICT",
  importId = "string",
  importStatus = "IN_PROGRESS"|"COMPLETE"|"FAILED",
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  ),
  createdDate = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Starts a job to import a resource to Amazon Lex.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$start_import(
  payload = raw,
  resourceType = "BOT"|"INTENT"|"SLOT_TYPE",
  mergeStrategy = "OVERWRITE_LATEST"|"FAIL_ON_CONFLICT",
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
