% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_tag_resource}
\alias{lexmodelbuildingservice_tag_resource}
\title{Adds the specified tags to the specified resource}
\usage{
lexmodelbuildingservice_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the bot, bot alias, or bot channel to
tag.}

\item{tags}{[required] A list of tag keys to add to the resource. If a tag key already exists,
the existing value is replaced with the new value.}
}
\value{
An empty list.
}
\description{
Adds the specified tags to the specified resource. If a tag key already
exists, the existing value is replaced with the new value.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$tag_resource(
  resourceArn = "string",
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
