% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_batch_delete_custom_vocabulary_item}
\alias{lexmodelsv2_batch_delete_custom_vocabulary_item}
\title{Delete a batch of custom vocabulary items for a given bot locale's
custom vocabulary}
\usage{
lexmodelsv2_batch_delete_custom_vocabulary_item(
  botId,
  botVersion,
  localeId,
  customVocabularyItemList
)
}
\arguments{
\item{botId}{[required] The identifier of the bot associated with this custom vocabulary.}

\item{botVersion}{[required] The identifier of the version of the bot associated with this custom
vocabulary.}

\item{localeId}{[required] The identifier of the language and locale where this custom vocabulary
is used. The string must match one of the supported locales. For more
information, see \href{https://docs.aws.amazon.com/lexv2/latest/dg/how-languages.html}{Supported Languages}
.}

\item{customVocabularyItemList}{[required] A list of custom vocabulary items requested to be deleted. Each entry
must contain the unique custom vocabulary entry identifier.}
}
\description{
Delete a batch of custom vocabulary items for a given bot locale's custom vocabulary.

See \url{https://paws-r.github.io/docs/lexmodelsv2/batch_delete_custom_vocabulary_item.html} for full documentation.
}
\keyword{internal}
