% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_batch_update_custom_vocabulary_item}
\alias{lexmodelsv2_batch_update_custom_vocabulary_item}
\title{Update a batch of custom vocabulary items for a given bot locale's
custom vocabulary}
\usage{
lexmodelsv2_batch_update_custom_vocabulary_item(
  botId,
  botVersion,
  localeId,
  customVocabularyItemList
)
}
\arguments{
\item{botId}{[required] The identifier of the bot associated with this custom vocabulary}

\item{botVersion}{[required] The identifier of the version of the bot associated with this custom
vocabulary.}

\item{localeId}{[required] The identifier of the language and locale where this custom vocabulary
is used. The string must match one of the supported locales. For more
information, see \href{https://docs.aws.amazon.com/lexv2/latest/dg/how-languages.html}{Supported Languages}
.}

\item{customVocabularyItemList}{[required] A list of custom vocabulary items with updated fields. Each entry must
contain a phrase and can optionally contain a displayAs and/or a weight.}
}
\description{
Update a batch of custom vocabulary items for a given bot locale's custom vocabulary.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_batch_update_custom_vocabulary_item/} for full documentation.
}
\keyword{internal}
