% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_create_bot_replica}
\alias{lexmodelsv2_create_bot_replica}
\title{Action to create a replication of the source bot in the secondary region}
\usage{
lexmodelsv2_create_bot_replica(botId, replicaRegion)
}
\arguments{
\item{botId}{[required] The request for the unique bot ID of the source bot to be replicated in
the secondary region.}

\item{replicaRegion}{[required] The request for the secondary region that will be used in the
replication of the source bot.}
}
\description{
Action to create a replication of the source bot in the secondary region.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_create_bot_replica/} for full documentation.
}
\keyword{internal}
