% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_create_test_set_discrepancy_report}
\alias{lexmodelsv2_create_test_set_discrepancy_report}
\title{Create a report that describes the differences between the bot and the
test set}
\usage{
lexmodelsv2_create_test_set_discrepancy_report(testSetId, target)
}
\arguments{
\item{testSetId}{[required] The test set Id for the test set discrepancy report.}

\item{target}{[required] The target bot for the test set discrepancy report.}
}
\description{
Create a report that describes the differences between the bot and the test set.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_create_test_set_discrepancy_report/} for full documentation.
}
\keyword{internal}
