% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_create_upload_url}
\alias{lexmodelsv2_create_upload_url}
\title{Gets a pre-signed S3 write URL that you use to upload the zip archive
when importing a bot or a bot locale}
\usage{
lexmodelsv2_create_upload_url()
}
\description{
Gets a pre-signed S3 write URL that you use to upload the zip archive when importing a bot or a bot locale.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_create_upload_url/} for full documentation.
}
\keyword{internal}
