% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_delete_bot}
\alias{lexmodelsv2_delete_bot}
\title{Deletes all versions of a bot, including the Draft version}
\usage{
lexmodelsv2_delete_bot(botId, skipResourceInUseCheck = NULL)
}
\arguments{
\item{botId}{[required] The identifier of the bot to delete.}

\item{skipResourceInUseCheck}{When \code{true}, Amazon Lex doesn't check to see if another resource, such
as an alias, is using the bot before it is deleted.}
}
\description{
Deletes all versions of a bot, including the \code{Draft} version. To delete a specific version, use the \code{\link[=lexmodelsv2_delete_bot_version]{delete_bot_version}} operation.

See \url{https://paws-r.github.io/docs/lexmodelsv2/delete_bot.html} for full documentation.
}
\keyword{internal}
