% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_delete_bot_alias}
\alias{lexmodelsv2_delete_bot_alias}
\title{Deletes the specified bot alias}
\usage{
lexmodelsv2_delete_bot_alias(botAliasId, botId, skipResourceInUseCheck = NULL)
}
\arguments{
\item{botAliasId}{[required] The unique identifier of the bot alias to delete.}

\item{botId}{[required] The unique identifier of the bot associated with the alias to delete.}

\item{skipResourceInUseCheck}{By default, Amazon Lex checks if any other resource, such as a bot
network, is using the bot alias before it is deleted and throws a
\code{ResourceInUseException} exception if the alias is being used by another
resource. Set this parameter to \code{true} to skip this check and remove the
alias even if it is being used by another resource.}
}
\description{
Deletes the specified bot alias.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_delete_bot_alias/} for full documentation.
}
\keyword{internal}
