% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_delete_bot_analyzer_recommendation}
\alias{lexmodelsv2_delete_bot_analyzer_recommendation}
\title{Permanently deletes the recommendations and analysis results for a
specific bot analysis request}
\usage{
lexmodelsv2_delete_bot_analyzer_recommendation(botId, botAnalyzerRequestId)
}
\arguments{
\item{botId}{[required] The unique identifier of the bot.}

\item{botAnalyzerRequestId}{[required] The unique identifier of the analysis request whose recommendations
should be deleted.}
}
\description{
Permanently deletes the recommendations and analysis results for a specific bot analysis request. This operation is provided for GDPR compliance and cannot be undone.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_delete_bot_analyzer_recommendation/} for full documentation.
}
\keyword{internal}
