% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_delete_bot_locale}
\alias{lexmodelsv2_delete_bot_locale}
\title{Removes a locale from a bot}
\usage{
lexmodelsv2_delete_bot_locale(botId, botVersion, localeId)
}
\arguments{
\item{botId}{[required] The unique identifier of the bot that contains the locale.}

\item{botVersion}{[required] The version of the bot that contains the locale.}

\item{localeId}{[required] The identifier of the language and locale that will be deleted. The
string must match one of the supported locales. For more information,
see \href{https://docs.aws.amazon.com/lexv2/latest/dg/how-languages.html}{Supported languages}.}
}
\description{
Removes a locale from a bot.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_delete_bot_locale/} for full documentation.
}
\keyword{internal}
