% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_delete_bot_replica}
\alias{lexmodelsv2_delete_bot_replica}
\title{The action to delete the replicated bot in the secondary region}
\usage{
lexmodelsv2_delete_bot_replica(botId, replicaRegion)
}
\arguments{
\item{botId}{[required] The unique ID of the replicated bot to be deleted from the secondary
region}

\item{replicaRegion}{[required] The secondary region of the replicated bot that will be deleted.}
}
\description{
The action to delete the replicated bot in the secondary region.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_delete_bot_replica/} for full documentation.
}
\keyword{internal}
