% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_delete_bot_version}
\alias{lexmodelsv2_delete_bot_version}
\title{Deletes a specific version of a bot}
\usage{
lexmodelsv2_delete_bot_version(
  botId,
  botVersion,
  skipResourceInUseCheck = NULL
)
}
\arguments{
\item{botId}{[required] The identifier of the bot that contains the version.}

\item{botVersion}{[required] The version of the bot to delete.}

\item{skipResourceInUseCheck}{By default, the \code{\link[=lexmodelsv2_delete_bot_version]{delete_bot_version}}
operations throws a \code{ResourceInUseException} exception if you try to
delete a bot version that has an alias pointing at it. Set the
\code{skipResourceInUseCheck} parameter to \code{true} to skip this check and
remove the version even if an alias points to it.}
}
\description{
Deletes a specific version of a bot. To delete all version of a bot, use the \code{\link[=lexmodelsv2_delete_bot]{delete_bot}} operation.

See \url{https://paws-r.github.io/docs/lexmodelsv2/delete_bot_version.html} for full documentation.
}
\keyword{internal}
