% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_delete_custom_vocabulary}
\alias{lexmodelsv2_delete_custom_vocabulary}
\title{Removes a custom vocabulary from the specified locale in the specified
bot}
\usage{
lexmodelsv2_delete_custom_vocabulary(botId, botVersion, localeId)
}
\arguments{
\item{botId}{[required] The unique identifier of the bot to remove the custom vocabulary from.}

\item{botVersion}{[required] The version of the bot to remove the custom vocabulary from.}

\item{localeId}{[required] The locale identifier for the locale that contains the custom vocabulary
to remove.}
}
\description{
Removes a custom vocabulary from the specified locale in the specified bot.

See \url{https://paws-r.github.io/docs/lexmodelsv2/delete_custom_vocabulary.html} for full documentation.
}
\keyword{internal}
