% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_delete_export}
\alias{lexmodelsv2_delete_export}
\title{Removes a previous export and the associated files stored in an S3
bucket}
\usage{
lexmodelsv2_delete_export(exportId)
}
\arguments{
\item{exportId}{[required] The unique identifier of the export to delete.}
}
\description{
Removes a previous export and the associated files stored in an S3 bucket.

See \url{https://paws-r.github.io/docs/lexmodelsv2/delete_export.html} for full documentation.
}
\keyword{internal}
