% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_delete_import}
\alias{lexmodelsv2_delete_import}
\title{Removes a previous import and the associated file stored in an S3 bucket}
\usage{
lexmodelsv2_delete_import(importId)
}
\arguments{
\item{importId}{[required] The unique identifier of the import to delete.}
}
\description{
Removes a previous import and the associated file stored in an S3 bucket.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_delete_import/} for full documentation.
}
\keyword{internal}
