% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_delete_intent}
\alias{lexmodelsv2_delete_intent}
\title{Removes the specified intent}
\usage{
lexmodelsv2_delete_intent(intentId, botId, botVersion, localeId)
}
\arguments{
\item{intentId}{[required] The unique identifier of the intent to delete.}

\item{botId}{[required] The identifier of the bot associated with the intent.}

\item{botVersion}{[required] The version of the bot associated with the intent.}

\item{localeId}{[required] The identifier of the language and locale where the bot will be deleted.
The string must match one of the supported locales. For more
information, see \href{https://docs.aws.amazon.com/lexv2/latest/dg/how-languages.html}{Supported languages}.}
}
\description{
Removes the specified intent.

See \url{https://paws-r.github.io/docs/lexmodelsv2/delete_intent.html} for full documentation.
}
\keyword{internal}
