% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_delete_slot}
\alias{lexmodelsv2_delete_slot}
\title{Deletes the specified slot from an intent}
\usage{
lexmodelsv2_delete_slot(slotId, botId, botVersion, localeId, intentId)
}
\arguments{
\item{slotId}{[required] The identifier of the slot to delete.}

\item{botId}{[required] The identifier of the bot associated with the slot to delete.}

\item{botVersion}{[required] The version of the bot associated with the slot to delete.}

\item{localeId}{[required] The identifier of the language and locale that the slot will be deleted
from. The string must match one of the supported locales. For more
information, see \href{https://docs.aws.amazon.com/lexv2/latest/dg/how-languages.html}{Supported languages}.}

\item{intentId}{[required] The identifier of the intent associated with the slot.}
}
\description{
Deletes the specified slot from an intent.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_delete_slot/} for full documentation.
}
\keyword{internal}
