% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_delete_test_set}
\alias{lexmodelsv2_delete_test_set}
\title{The action to delete the selected test set}
\usage{
lexmodelsv2_delete_test_set(testSetId)
}
\arguments{
\item{testSetId}{[required] The test set Id of the test set to be deleted.}
}
\description{
The action to delete the selected test set.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_delete_test_set/} for full documentation.
}
\keyword{internal}
