% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_delete_utterances}
\alias{lexmodelsv2_delete_utterances}
\title{Deletes stored utterances}
\usage{
lexmodelsv2_delete_utterances(botId, localeId = NULL, sessionId = NULL)
}
\arguments{
\item{botId}{[required] The unique identifier of the bot that contains the utterances.}

\item{localeId}{The identifier of the language and locale where the utterances were
collected. The string must match one of the supported locales. For more
information, see \href{https://docs.aws.amazon.com/lexv2/latest/dg/how-languages.html}{Supported languages}.}

\item{sessionId}{The unique identifier of the session with the user. The ID is returned
in the response from the
\href{https://docs.aws.amazon.com/lexv2/latest/APIReference/API_runtime_RecognizeText.html}{RecognizeText}
and
\href{https://docs.aws.amazon.com/lexv2/latest/APIReference/API_runtime_RecognizeUtterance.html}{RecognizeUtterance}
operations.}
}
\description{
Deletes stored utterances.

See \url{https://paws-r.github.io/docs/lexmodelsv2/delete_utterances.html} for full documentation.
}
\keyword{internal}
