% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_describe_bot}
\alias{lexmodelsv2_describe_bot}
\title{Provides metadata information about a bot}
\usage{
lexmodelsv2_describe_bot(botId)
}
\arguments{
\item{botId}{[required] The unique identifier of the bot to describe.}
}
\description{
Provides metadata information about a bot.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_describe_bot/} for full documentation.
}
\keyword{internal}
