% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_describe_bot_alias}
\alias{lexmodelsv2_describe_bot_alias}
\title{Get information about a specific bot alias}
\usage{
lexmodelsv2_describe_bot_alias(botAliasId, botId)
}
\arguments{
\item{botAliasId}{[required] The identifier of the bot alias to describe.}

\item{botId}{[required] The identifier of the bot associated with the bot alias to describe.}
}
\description{
Get information about a specific bot alias.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_describe_bot_alias/} for full documentation.
}
\keyword{internal}
