% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_describe_bot_analyzer_recommendation}
\alias{lexmodelsv2_describe_bot_analyzer_recommendation}
\title{Retrieves the analysis results and recommendations for bot optimization}
\usage{
lexmodelsv2_describe_bot_analyzer_recommendation(
  botId,
  botAnalyzerRequestId,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{botId}{[required] The unique identifier of the bot.}

\item{botAnalyzerRequestId}{[required] The unique identifier of the analysis request.}

\item{nextToken}{If the response from a previous request was truncated, the \code{nextToken}
value is used to retrieve the next page of recommendations.}

\item{maxResults}{The maximum number of recommendations to return in the response. The
default is 5.}
}
\description{
Retrieves the analysis results and recommendations for bot optimization. The analysis must be in \code{Available} status before recommendations can be retrieved.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_describe_bot_analyzer_recommendation/} for full documentation.
}
\keyword{internal}
