% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_describe_bot_locale}
\alias{lexmodelsv2_describe_bot_locale}
\title{Describes the settings that a bot has for a specific locale}
\usage{
lexmodelsv2_describe_bot_locale(botId, botVersion, localeId)
}
\arguments{
\item{botId}{[required] The identifier of the bot associated with the locale.}

\item{botVersion}{[required] The version of the bot associated with the locale.}

\item{localeId}{[required] The unique identifier of the locale to describe. The string must match
one of the supported locales. For more information, see \href{https://docs.aws.amazon.com/lexv2/latest/dg/how-languages.html}{Supported languages}.}
}
\description{
Describes the settings that a bot has for a specific locale.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_describe_bot_locale/} for full documentation.
}
\keyword{internal}
