% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_describe_bot_replica}
\alias{lexmodelsv2_describe_bot_replica}
\title{Monitors the bot replication status through the UI console}
\usage{
lexmodelsv2_describe_bot_replica(botId, replicaRegion)
}
\arguments{
\item{botId}{[required] The request for the unique bot ID of the replicated bot being monitored.}

\item{replicaRegion}{[required] The request for the region of the replicated bot being monitored.}
}
\description{
Monitors the bot replication status through the UI console.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_describe_bot_replica/} for full documentation.
}
\keyword{internal}
