% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_describe_bot_version}
\alias{lexmodelsv2_describe_bot_version}
\title{Provides metadata about a version of a bot}
\usage{
lexmodelsv2_describe_bot_version(botId, botVersion)
}
\arguments{
\item{botId}{[required] The identifier of the bot containing the version to return metadata for.}

\item{botVersion}{[required] The version of the bot to return metadata for.}
}
\description{
Provides metadata about a version of a bot.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_describe_bot_version/} for full documentation.
}
\keyword{internal}
