% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_describe_custom_vocabulary_metadata}
\alias{lexmodelsv2_describe_custom_vocabulary_metadata}
\title{Provides metadata information about a custom vocabulary}
\usage{
lexmodelsv2_describe_custom_vocabulary_metadata(botId, botVersion, localeId)
}
\arguments{
\item{botId}{[required] The unique identifier of the bot that contains the custom vocabulary.}

\item{botVersion}{[required] The bot version of the bot to return metadata for.}

\item{localeId}{[required] The locale to return the custom vocabulary information for. The locale
must be \code{en_GB}.}
}
\description{
Provides metadata information about a custom vocabulary.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_describe_custom_vocabulary_metadata/} for full documentation.
}
\keyword{internal}
