% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_describe_export}
\alias{lexmodelsv2_describe_export}
\title{Gets information about a specific export}
\usage{
lexmodelsv2_describe_export(exportId)
}
\arguments{
\item{exportId}{[required] The unique identifier of the export to describe.}
}
\description{
Gets information about a specific export.

See \url{https://paws-r.github.io/docs/lexmodelsv2/describe_export.html} for full documentation.
}
\keyword{internal}
