% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_describe_import}
\alias{lexmodelsv2_describe_import}
\title{Gets information about a specific import}
\usage{
lexmodelsv2_describe_import(importId)
}
\arguments{
\item{importId}{[required] The unique identifier of the import to describe.}
}
\description{
Gets information about a specific import.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_describe_import/} for full documentation.
}
\keyword{internal}
