% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_describe_intent}
\alias{lexmodelsv2_describe_intent}
\title{Returns metadata about an intent}
\usage{
lexmodelsv2_describe_intent(intentId, botId, botVersion, localeId)
}
\arguments{
\item{intentId}{[required] The identifier of the intent to describe.}

\item{botId}{[required] The identifier of the bot associated with the intent.}

\item{botVersion}{[required] The version of the bot associated with the intent.}

\item{localeId}{[required] The identifier of the language and locale of the intent to describe. The
string must match one of the supported locales. For more information,
see \href{https://docs.aws.amazon.com/lexv2/latest/dg/how-languages.html}{Supported languages}.}
}
\description{
Returns metadata about an intent.

See \url{https://paws-r.github.io/docs/lexmodelsv2/describe_intent.html} for full documentation.
}
\keyword{internal}
