% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_describe_resource_policy}
\alias{lexmodelsv2_describe_resource_policy}
\title{Gets the resource policy and policy revision for a bot or bot alias}
\usage{
lexmodelsv2_describe_resource_policy(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the bot or bot alias that the resource
policy is attached to.}
}
\description{
Gets the resource policy and policy revision for a bot or bot alias.

See \url{https://paws-r.github.io/docs/lexmodelsv2/describe_resource_policy.html} for full documentation.
}
\keyword{internal}
