% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_describe_slot}
\alias{lexmodelsv2_describe_slot}
\title{Gets metadata information about a slot}
\usage{
lexmodelsv2_describe_slot(slotId, botId, botVersion, localeId, intentId)
}
\arguments{
\item{slotId}{[required] The unique identifier for the slot.}

\item{botId}{[required] The identifier of the bot associated with the slot.}

\item{botVersion}{[required] The version of the bot associated with the slot.}

\item{localeId}{[required] The identifier of the language and locale of the slot to describe. The
string must match one of the supported locales. For more information,
see \href{https://docs.aws.amazon.com/lexv2/latest/dg/how-languages.html}{Supported languages}.}

\item{intentId}{[required] The identifier of the intent that contains the slot.}
}
\description{
Gets metadata information about a slot.

See \url{https://paws-r.github.io/docs/lexmodelsv2/describe_slot.html} for full documentation.
}
\keyword{internal}
