% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_describe_slot_type}
\alias{lexmodelsv2_describe_slot_type}
\title{Gets metadata information about a slot type}
\usage{
lexmodelsv2_describe_slot_type(slotTypeId, botId, botVersion, localeId)
}
\arguments{
\item{slotTypeId}{[required] The identifier of the slot type.}

\item{botId}{[required] The identifier of the bot associated with the slot type.}

\item{botVersion}{[required] The version of the bot associated with the slot type.}

\item{localeId}{[required] The identifier of the language and locale of the slot type to describe.
The string must match one of the supported locales. For more
information, see \href{https://docs.aws.amazon.com/lexv2/latest/dg/how-languages.html}{Supported languages}.}
}
\description{
Gets metadata information about a slot type.

See \url{https://paws-r.github.io/docs/lexmodelsv2/describe_slot_type.html} for full documentation.
}
\keyword{internal}
