% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_describe_test_execution}
\alias{lexmodelsv2_describe_test_execution}
\title{Gets metadata information about the test execution}
\usage{
lexmodelsv2_describe_test_execution(testExecutionId)
}
\arguments{
\item{testExecutionId}{[required] The execution Id of the test set execution.}
}
\description{
Gets metadata information about the test execution.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_describe_test_execution/} for full documentation.
}
\keyword{internal}
