% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_describe_test_set}
\alias{lexmodelsv2_describe_test_set}
\title{Gets metadata information about the test set}
\usage{
lexmodelsv2_describe_test_set(testSetId)
}
\arguments{
\item{testSetId}{[required] The test set Id for the test set request.}
}
\description{
Gets metadata information about the test set.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_describe_test_set/} for full documentation.
}
\keyword{internal}
