% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_describe_test_set_discrepancy_report}
\alias{lexmodelsv2_describe_test_set_discrepancy_report}
\title{Gets metadata information about the test set discrepancy report}
\usage{
lexmodelsv2_describe_test_set_discrepancy_report(testSetDiscrepancyReportId)
}
\arguments{
\item{testSetDiscrepancyReportId}{[required] The unique identifier of the test set discrepancy report.}
}
\description{
Gets metadata information about the test set discrepancy report.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_describe_test_set_discrepancy_report/} for full documentation.
}
\keyword{internal}
