% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_describe_test_set_generation}
\alias{lexmodelsv2_describe_test_set_generation}
\title{Gets metadata information about the test set generation}
\usage{
lexmodelsv2_describe_test_set_generation(testSetGenerationId)
}
\arguments{
\item{testSetGenerationId}{[required] The unique identifier of the test set generation.}
}
\description{
Gets metadata information about the test set generation.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_describe_test_set_generation/} for full documentation.
}
\keyword{internal}
