% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_generate_bot_element}
\alias{lexmodelsv2_generate_bot_element}
\title{Generates sample utterances for an intent}
\usage{
lexmodelsv2_generate_bot_element(intentId, botId, botVersion, localeId)
}
\arguments{
\item{intentId}{[required] The intent unique Id for the bot request to generate utterances.}

\item{botId}{[required] The bot unique Id for the bot request to generate utterances.}

\item{botVersion}{[required] The bot version for the bot request to generate utterances.}

\item{localeId}{[required] The unique locale Id for the bot request to generate utterances.}
}
\description{
Generates sample utterances for an intent.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_generate_bot_element/} for full documentation.
}
\keyword{internal}
