% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_get_test_execution_artifacts_url}
\alias{lexmodelsv2_get_test_execution_artifacts_url}
\title{The pre-signed Amazon S3 URL to download the test execution result
artifacts}
\usage{
lexmodelsv2_get_test_execution_artifacts_url(testExecutionId)
}
\arguments{
\item{testExecutionId}{[required] The unique identifier of the completed test execution.}
}
\description{
The pre-signed Amazon S3 URL to download the test execution result artifacts.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_get_test_execution_artifacts_url/} for full documentation.
}
\keyword{internal}
