% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_list_bot_aliases}
\alias{lexmodelsv2_list_bot_aliases}
\title{Gets a list of aliases for the specified bot}
\usage{
lexmodelsv2_list_bot_aliases(botId, maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{botId}{[required] The identifier of the bot to list aliases for.}

\item{maxResults}{The maximum number of aliases to return in each page of results. If
there are fewer results than the max page size, only the actual number
of results are returned.}

\item{nextToken}{If the response from the
\code{\link[=lexmodelsv2_list_bot_aliases]{list_bot_aliases}} operation contains
more results than specified in the \code{maxResults} parameter, a token is
returned in the response. Use that token in the \code{nextToken} parameter to
return the next page of results.}
}
\description{
Gets a list of aliases for the specified bot.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_list_bot_aliases/} for full documentation.
}
\keyword{internal}
