% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_list_bot_analyzer_history}
\alias{lexmodelsv2_list_bot_analyzer_history}
\title{Retrieves a list of historical bot analysis executions for a specific
bot}
\usage{
lexmodelsv2_list_bot_analyzer_history(
  botId,
  localeId = NULL,
  botVersion = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{botId}{[required] The unique identifier of the bot.}

\item{localeId}{The locale identifier to filter the history. If not specified, returns
history for all locales.}

\item{botVersion}{The bot version to filter the history. If not specified, defaults to
\code{DRAFT}.}

\item{nextToken}{If the response from a previous request was truncated, the \code{nextToken}
value is used to retrieve the next page of history entries.}

\item{maxResults}{The maximum number of history entries to return in the response. The
default is 10.}
}
\description{
Retrieves a list of historical bot analysis executions for a specific bot. You can filter the results by locale and bot version.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_list_bot_analyzer_history/} for full documentation.
}
\keyword{internal}
