% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_list_bot_replicas}
\alias{lexmodelsv2_list_bot_replicas}
\title{The action to list the replicated bots}
\usage{
lexmodelsv2_list_bot_replicas(botId)
}
\arguments{
\item{botId}{[required] The request for the unique bot IDs in the list of replicated bots.}
}
\description{
The action to list the replicated bots.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_list_bot_replicas/} for full documentation.
}
\keyword{internal}
