% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_list_bot_resource_generations}
\alias{lexmodelsv2_list_bot_resource_generations}
\title{Lists the generation requests made for a bot locale}
\usage{
lexmodelsv2_list_bot_resource_generations(
  botId,
  botVersion,
  localeId,
  sortBy = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{botId}{[required] The unique identifier of the bot whose generation requests you want to
view.}

\item{botVersion}{[required] The version of the bot whose generation requests you want to view.}

\item{localeId}{[required] The locale of the bot whose generation requests you want to view.}

\item{sortBy}{An object containing information about the attribute and the method by
which to sort the results}

\item{maxResults}{The maximum number of results to return in the response.}

\item{nextToken}{If the total number of results is greater than the number specified in
the \code{maxResults}, the response returns a token in the \code{nextToken} field.
Use this token when making a request to return the next batch of
results.}
}
\description{
Lists the generation requests made for a bot locale.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_list_bot_resource_generations/} for full documentation.
}
\keyword{internal}
