% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_list_bot_version_replicas}
\alias{lexmodelsv2_list_bot_version_replicas}
\title{Contains information about all the versions replication statuses
applicable for Global Resiliency}
\usage{
lexmodelsv2_list_bot_version_replicas(
  botId,
  replicaRegion,
  maxResults = NULL,
  nextToken = NULL,
  sortBy = NULL
)
}
\arguments{
\item{botId}{[required] The request for the unique ID in the list of replicated bots.}

\item{replicaRegion}{[required] The request for the region used in the list of replicated bots.}

\item{maxResults}{The maximum results given in the list of replicated bots.}

\item{nextToken}{The next token given in the list of replicated bots.}

\item{sortBy}{The requested sort category for the list of replicated bots.}
}
\description{
Contains information about all the versions replication statuses applicable for Global Resiliency.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_list_bot_version_replicas/} for full documentation.
}
\keyword{internal}
