% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_list_bots}
\alias{lexmodelsv2_list_bots}
\title{Gets a list of available bots}
\usage{
lexmodelsv2_list_bots(
  sortBy = NULL,
  filters = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{sortBy}{Specifies sorting parameters for the list of bots. You can specify that
the list be sorted by bot name in ascending or descending order.}

\item{filters}{Provides the specification of a filter used to limit the bots in the
response to only those that match the filter specification. You can only
specify one filter and one string to filter on.}

\item{maxResults}{The maximum number of bots to return in each page of results. If there
are fewer results than the maximum page size, only the actual number of
results are returned.}

\item{nextToken}{If the response from the \code{\link[=lexmodelsv2_list_bots]{list_bots}} operation
contains more results than specified in the \code{maxResults} parameter, a
token is returned in the response.

Use the returned token in the \code{nextToken} parameter of a
\code{\link[=lexmodelsv2_list_bots]{list_bots}} request to return the next page of
results. For a complete set of results, call the
\code{\link[=lexmodelsv2_list_bots]{list_bots}} operation until the \code{nextToken}
returned in the response is null.}
}
\description{
Gets a list of available bots.

See \url{https://paws-r.github.io/docs/lexmodelsv2/list_bots.html} for full documentation.
}
\keyword{internal}
