% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_list_intent_metrics}
\alias{lexmodelsv2_list_intent_metrics}
\title{Retrieves summary metrics for the intents in your bot}
\usage{
lexmodelsv2_list_intent_metrics(
  botId,
  startDateTime,
  endDateTime,
  metrics,
  binBy = NULL,
  groupBy = NULL,
  filters = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{botId}{[required] The identifier for the bot for which you want to retrieve intent
metrics.}

\item{startDateTime}{[required] The timestamp that marks the beginning of the range of time for which
you want to see intent metrics.}

\item{endDateTime}{[required] The date and time that marks the end of the range of time for which you
want to see intent metrics.}

\item{metrics}{[required] A list of objects, each of which contains a metric you want to list, the
statistic for the metric you want to return, and the order by which to
organize the results.}

\item{binBy}{A list of objects, each of which contains specifications for organizing
the results by time.}

\item{groupBy}{A list of objects, each of which specifies how to group the results. You
can group by the following criteria:
\itemize{
\item \code{IntentName} – The name of the intent.
\item \code{IntentEndState} – The final state of the intent. The possible end
states are detailed in Key definitions in the user guide.
}}

\item{filters}{A list of objects, each of which describes a condition by which you want
to filter the results.}

\item{maxResults}{The maximum number of results to return in each page of results. If
there are fewer results than the maximum page size, only the actual
number of results are returned.}

\item{nextToken}{If the response from the ListIntentMetrics operation contains more
results than specified in the maxResults parameter, a token is returned
in the response.

Use the returned token in the nextToken parameter of a ListIntentMetrics
request to return the next page of results. For a complete set of
results, call the ListIntentMetrics operation until the nextToken
returned in the response is null.}
}
\description{
Retrieves summary metrics for the intents in your bot. The following fields are required:

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_list_intent_metrics/} for full documentation.
}
\keyword{internal}
