% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_list_test_executions}
\alias{lexmodelsv2_list_test_executions}
\title{The list of test set executions}
\usage{
lexmodelsv2_list_test_executions(
  sortBy = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{sortBy}{The sort order of the test set executions.}

\item{maxResults}{The maximum number of test executions to return in each page. If there
are fewer results than the max page size, only the actual number of
results are returned.}

\item{nextToken}{If the response from the ListTestExecutions operation contains more
results than specified in the maxResults parameter, a token is returned
in the response. Use that token in the nextToken parameter to return the
next page of results.}
}
\description{
The list of test set executions.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_list_test_executions/} for full documentation.
}
\keyword{internal}
