% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_list_test_set_records}
\alias{lexmodelsv2_list_test_set_records}
\title{The list of test set records}
\usage{
lexmodelsv2_list_test_set_records(
  testSetId,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{testSetId}{[required] The identifier of the test set to list its test set records.}

\item{maxResults}{The maximum number of test set records to return in each page. If there
are fewer records than the max page size, only the actual number of
records are returned.}

\item{nextToken}{If the response from the ListTestSetRecords operation contains more
results than specified in the maxResults parameter, a token is returned
in the response. Use that token in the nextToken parameter to return the
next page of results.}
}
\description{
The list of test set records.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_list_test_set_records/} for full documentation.
}
\keyword{internal}
