% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_start_test_set_generation}
\alias{lexmodelsv2_start_test_set_generation}
\title{The action to start the generation of test set}
\usage{
lexmodelsv2_start_test_set_generation(
  testSetName,
  description = NULL,
  storageLocation,
  generationDataSource,
  roleArn,
  testSetTags = NULL
)
}
\arguments{
\item{testSetName}{[required] The test set name for the test set generation request.}

\item{description}{The test set description for the test set generation request.}

\item{storageLocation}{[required] The Amazon S3 storage location for the test set generation.}

\item{generationDataSource}{[required] The data source for the test set generation.}

\item{roleArn}{[required] The roleARN used for any operation in the test set to access resources
in the Amazon Web Services account.}

\item{testSetTags}{A list of tags to add to the test set. You can only add tags when you
import/generate a new test set. You can't use the
\code{\link[=lexmodelsv2_update_test_set]{update_test_set}} operation to update
tags. To update tags, use the \code{\link[=lexmodelsv2_tag_resource]{tag_resource}}
operation.}
}
\description{
The action to start the generation of test set.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_start_test_set_generation/} for full documentation.
}
\keyword{internal}
