% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_stop_bot_analyzer}
\alias{lexmodelsv2_stop_bot_analyzer}
\title{Cancels an ongoing bot analysis execution}
\usage{
lexmodelsv2_stop_bot_analyzer(botId, botAnalyzerRequestId)
}
\arguments{
\item{botId}{[required] The unique identifier of the bot.}

\item{botAnalyzerRequestId}{[required] The unique identifier of the analysis request to stop.}
}
\description{
Cancels an ongoing bot analysis execution. Once stopped, the analysis cannot be resumed and no recommendations will be generated.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_stop_bot_analyzer/} for full documentation.
}
\keyword{internal}
