% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_stop_bot_recommendation}
\alias{lexmodelsv2_stop_bot_recommendation}
\title{Stop an already running Bot Recommendation request}
\usage{
lexmodelsv2_stop_bot_recommendation(
  botId,
  botVersion,
  localeId,
  botRecommendationId
)
}
\arguments{
\item{botId}{[required] The unique identifier of the bot containing the bot recommendation to be
stopped.}

\item{botVersion}{[required] The version of the bot containing the bot recommendation.}

\item{localeId}{[required] The identifier of the language and locale of the bot recommendation to
stop. The string must match one of the supported locales. For more
information, see \href{https://docs.aws.amazon.com/lexv2/latest/dg/how-languages.html}{Supported languages}}

\item{botRecommendationId}{[required] The unique identifier of the bot recommendation to be stopped.}
}
\description{
Stop an already running Bot Recommendation request.

See \url{https://paws-r.github.io/docs/lexmodelsv2/stop_bot_recommendation.html} for full documentation.
}
\keyword{internal}
