% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_tag_resource}
\alias{lexmodelsv2_tag_resource}
\title{Adds the specified tags to the specified resource}
\usage{
lexmodelsv2_tag_resource(resourceARN, tags)
}
\arguments{
\item{resourceARN}{[required] The Amazon Resource Name (ARN) of the bot, bot alias, or bot channel to
tag.}

\item{tags}{[required] A list of tag keys to add to the resource. If a tag key already exists,
the existing value is replaced with the new value.}
}
\description{
Adds the specified tags to the specified resource. If a tag key already exists, the existing value is replaced with the new value.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_tag_resource/} for full documentation.
}
\keyword{internal}
