% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_untag_resource}
\alias{lexmodelsv2_untag_resource}
\title{Removes tags from a bot, bot alias, or bot channel}
\usage{
lexmodelsv2_untag_resource(resourceARN, tagKeys)
}
\arguments{
\item{resourceARN}{[required] The Amazon Resource Name (ARN) of the resource to remove the tags from.}

\item{tagKeys}{[required] A list of tag keys to remove from the resource. If a tag key does not
exist on the resource, it is ignored.}
}
\description{
Removes tags from a bot, bot alias, or bot channel.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_untag_resource/} for full documentation.
}
\keyword{internal}
