% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_update_bot_alias}
\alias{lexmodelsv2_update_bot_alias}
\title{Updates the configuration of an existing bot alias}
\usage{
lexmodelsv2_update_bot_alias(
  botAliasId,
  botAliasName,
  description = NULL,
  botVersion = NULL,
  botAliasLocaleSettings = NULL,
  conversationLogSettings = NULL,
  sentimentAnalysisSettings = NULL,
  botId
)
}
\arguments{
\item{botAliasId}{[required] The unique identifier of the bot alias.}

\item{botAliasName}{[required] The new name to assign to the bot alias.}

\item{description}{The new description to assign to the bot alias.}

\item{botVersion}{The new bot version to assign to the bot alias.}

\item{botAliasLocaleSettings}{The new Lambda functions to use in each locale for the bot alias.}

\item{conversationLogSettings}{The new settings for storing conversation logs in Amazon CloudWatch Logs
and Amazon S3 buckets.}

\item{sentimentAnalysisSettings}{}

\item{botId}{[required] The identifier of the bot with the updated alias.}
}
\description{
Updates the configuration of an existing bot alias.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_update_bot_alias/} for full documentation.
}
\keyword{internal}
