% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_update_export}
\alias{lexmodelsv2_update_export}
\title{Updates the password used to protect an export zip archive}
\usage{
lexmodelsv2_update_export(exportId, filePassword = NULL)
}
\arguments{
\item{exportId}{[required] The unique identifier Amazon Lex assigned to the export.}

\item{filePassword}{The new password to use to encrypt the export zip archive.}
}
\description{
Updates the password used to protect an export zip archive.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_update_export/} for full documentation.
}
\keyword{internal}
