% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_update_test_set}
\alias{lexmodelsv2_update_test_set}
\title{The action to update the test set}
\usage{
lexmodelsv2_update_test_set(testSetId, testSetName, description = NULL)
}
\arguments{
\item{testSetId}{[required] The test set Id for which update test operation to be performed.}

\item{testSetName}{[required] The new test set name.}

\item{description}{The new test set description.}
}
\description{
The action to update the test set.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_update_test_set/} for full documentation.
}
\keyword{internal}
