% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexruntimeservice_operations.R
\name{lexruntimeservice_delete_session}
\alias{lexruntimeservice_delete_session}
\title{Removes session information for a specified bot, alias, and user ID}
\usage{
lexruntimeservice_delete_session(botName, botAlias, userId)
}
\arguments{
\item{botName}{[required] The name of the bot that contains the session data.}

\item{botAlias}{[required] The alias in use for the bot that contains the session data.}

\item{userId}{[required] The identifier of the user associated with the session data.}
}
\description{
Removes session information for a specified bot, alias, and user ID.

See \url{https://paws-r.github.io/docs/lexruntimeservice/delete_session.html} for full documentation.
}
\keyword{internal}
