% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexruntimev2_operations.R
\name{lexruntimev2_delete_session}
\alias{lexruntimev2_delete_session}
\title{Removes session information for a specified bot, alias, and user ID}
\usage{
lexruntimev2_delete_session(botId, botAliasId, localeId, sessionId)
}
\arguments{
\item{botId}{[required] The identifier of the bot that contains the session data.}

\item{botAliasId}{[required] The alias identifier in use for the bot that contains the session data.}

\item{localeId}{[required] The locale where the session is in use.}

\item{sessionId}{[required] The identifier of the session to delete.}
}
\description{
Removes session information for a specified bot, alias, and user ID.

See \url{https://www.paws-r-sdk.com/docs/lexruntimev2_delete_session/} for full documentation.
}
\keyword{internal}
