% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexruntimev2_operations.R
\name{lexruntimev2_get_session}
\alias{lexruntimev2_get_session}
\title{Returns session information for a specified bot, alias, and user}
\usage{
lexruntimev2_get_session(botId, botAliasId, localeId, sessionId)
}
\arguments{
\item{botId}{[required] The identifier of the bot that contains the session data.}

\item{botAliasId}{[required] The alias identifier in use for the bot that contains the session data.}

\item{localeId}{[required] The locale where the session is in use.}

\item{sessionId}{[required] The identifier of the session to return.}
}
\description{
Returns session information for a specified bot, alias, and user.

See \url{https://paws-r.github.io/docs/lexruntimev2/get_session.html} for full documentation.
}
\keyword{internal}
