% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_create_label}
\alias{lookoutequipment_create_label}
\title{Creates a label for an event}
\usage{
lookoutequipment_create_label(
  LabelGroupName,
  StartTime,
  EndTime,
  Rating,
  FaultCode = NULL,
  Notes = NULL,
  Equipment = NULL,
  ClientToken
)
}
\arguments{
\item{LabelGroupName}{[required] The name of a group of labels.

Data in this field will be retained for service usage. Follow best
practices for the security of your data.}

\item{StartTime}{[required] The start time of the labeled event.}

\item{EndTime}{[required] The end time of the labeled event.}

\item{Rating}{[required] Indicates whether a labeled event represents an anomaly.}

\item{FaultCode}{Provides additional information about the label. The fault code must be
defined in the FaultCodes attribute of the label group.

Data in this field will be retained for service usage. Follow best
practices for the security of your data.}

\item{Notes}{Metadata providing additional information about the label.

Data in this field will be retained for service usage. Follow best
practices for the security of your data.}

\item{Equipment}{Indicates that a label pertains to a particular piece of equipment.

Data in this field will be retained for service usage. Follow best
practices for the security of your data.}

\item{ClientToken}{[required] A unique identifier for the request to create a label. If you do not set
the client request token, Lookout for Equipment generates one.}
}
\description{
Creates a label for an event.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_create_label/} for full documentation.
}
\keyword{internal}
