% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_create_label_group}
\alias{lookoutequipment_create_label_group}
\title{Creates a group of labels}
\usage{
lookoutequipment_create_label_group(
  LabelGroupName,
  FaultCodes = NULL,
  ClientToken,
  Tags = NULL
)
}
\arguments{
\item{LabelGroupName}{[required] Names a group of labels.

Data in this field will be retained for service usage. Follow best
practices for the security of your data.}

\item{FaultCodes}{The acceptable fault codes (indicating the type of anomaly associated
with the label) that can be used with this label group.

Data in this field will be retained for service usage. Follow best
practices for the security of your data.}

\item{ClientToken}{[required] A unique identifier for the request to create a label group. If you do
not set the client request token, Lookout for Equipment generates one.}

\item{Tags}{Tags that provide metadata about the label group you are creating.

Data in this field will be retained for service usage. Follow best
practices for the security of your data.}
}
\description{
Creates a group of labels.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_create_label_group/} for full documentation.
}
\keyword{internal}
