% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_delete_dataset}
\alias{lookoutequipment_delete_dataset}
\title{Deletes a dataset and associated artifacts}
\usage{
lookoutequipment_delete_dataset(DatasetName)
}
\arguments{
\item{DatasetName}{[required] The name of the dataset to be deleted.}
}
\description{
Deletes a dataset and associated artifacts. The operation will check to see if any inference scheduler or data ingestion job is currently using the dataset, and if there isn't, the dataset, its metadata, and any associated data stored in S3 will be deleted. This does not affect any models that used this dataset for training and evaluation, but does prevent it from being used in the future.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_delete_dataset/} for full documentation.
}
\keyword{internal}
