% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_delete_inference_scheduler}
\alias{lookoutequipment_delete_inference_scheduler}
\title{Deletes an inference scheduler that has been set up}
\usage{
lookoutequipment_delete_inference_scheduler(InferenceSchedulerName)
}
\arguments{
\item{InferenceSchedulerName}{[required] The name of the inference scheduler to be deleted.}
}
\description{
Deletes an inference scheduler that has been set up. Prior inference results will not be deleted.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_delete_inference_scheduler/} for full documentation.
}
\keyword{internal}
