% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_delete_label}
\alias{lookoutequipment_delete_label}
\title{Deletes a label}
\usage{
lookoutequipment_delete_label(LabelGroupName, LabelId)
}
\arguments{
\item{LabelGroupName}{[required] The name of the label group that contains the label that you want to
delete. Data in this field will be retained for service usage. Follow
best practices for the security of your data.}

\item{LabelId}{[required] The ID of the label that you want to delete.}
}
\description{
Deletes a label.

See \url{https://paws-r.github.io/docs/lookoutequipment/delete_label.html} for full documentation.
}
\keyword{internal}
