% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_delete_label_group}
\alias{lookoutequipment_delete_label_group}
\title{Deletes a group of labels}
\usage{
lookoutequipment_delete_label_group(LabelGroupName)
}
\arguments{
\item{LabelGroupName}{[required] The name of the label group that you want to delete. Data in this field
will be retained for service usage. Follow best practices for the
security of your data.}
}
\description{
Deletes a group of labels.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_delete_label_group/} for full documentation.
}
\keyword{internal}
