% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_delete_model}
\alias{lookoutequipment_delete_model}
\title{Deletes a machine learning model currently available for Amazon Lookout
for Equipment}
\usage{
lookoutequipment_delete_model(ModelName)
}
\arguments{
\item{ModelName}{[required] The name of the machine learning model to be deleted.}
}
\description{
Deletes a machine learning model currently available for Amazon Lookout for Equipment. This will prevent it from being used with an inference scheduler, even one that is already set up.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_delete_model/} for full documentation.
}
\keyword{internal}
