% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_delete_resource_policy}
\alias{lookoutequipment_delete_resource_policy}
\title{Deletes the resource policy attached to the resource}
\usage{
lookoutequipment_delete_resource_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource for which the resource
policy should be deleted.}
}
\description{
Deletes the resource policy attached to the resource.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_delete_resource_policy/} for full documentation.
}
\keyword{internal}
