% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_delete_retraining_scheduler}
\alias{lookoutequipment_delete_retraining_scheduler}
\title{Deletes a retraining scheduler from a model}
\usage{
lookoutequipment_delete_retraining_scheduler(ModelName)
}
\arguments{
\item{ModelName}{[required] The name of the model whose retraining scheduler you want to delete.}
}
\description{
Deletes a retraining scheduler from a model. The retraining scheduler must be in the \code{STOPPED} status.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_delete_retraining_scheduler/} for full documentation.
}
\keyword{internal}
