% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_describe_data_ingestion_job}
\alias{lookoutequipment_describe_data_ingestion_job}
\title{Provides information on a specific data ingestion job such as creation
time, dataset ARN, and status}
\usage{
lookoutequipment_describe_data_ingestion_job(JobId)
}
\arguments{
\item{JobId}{[required] The job ID of the data ingestion job.}
}
\description{
Provides information on a specific data ingestion job such as creation time, dataset ARN, and status.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_describe_data_ingestion_job/} for full documentation.
}
\keyword{internal}
