% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_describe_dataset}
\alias{lookoutequipment_describe_dataset}
\title{Provides a JSON description of the data in each time series dataset,
including names, column names, and data types}
\usage{
lookoutequipment_describe_dataset(DatasetName)
}
\arguments{
\item{DatasetName}{[required] The name of the dataset to be described.}
}
\description{
Provides a JSON description of the data in each time series dataset, including names, column names, and data types.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_describe_dataset/} for full documentation.
}
\keyword{internal}
