% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_describe_label}
\alias{lookoutequipment_describe_label}
\title{Returns the name of the label}
\usage{
lookoutequipment_describe_label(LabelGroupName, LabelId)
}
\arguments{
\item{LabelGroupName}{[required] Returns the name of the group containing the label.}

\item{LabelId}{[required] Returns the ID of the label.}
}
\description{
Returns the name of the label.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_describe_label/} for full documentation.
}
\keyword{internal}
