% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_describe_label_group}
\alias{lookoutequipment_describe_label_group}
\title{Returns information about the label group}
\usage{
lookoutequipment_describe_label_group(LabelGroupName)
}
\arguments{
\item{LabelGroupName}{[required] Returns the name of the label group.}
}
\description{
Returns information about the label group.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_describe_label_group/} for full documentation.
}
\keyword{internal}
