% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_describe_model_version}
\alias{lookoutequipment_describe_model_version}
\title{Retrieves information about a specific machine learning model version}
\usage{
lookoutequipment_describe_model_version(ModelName, ModelVersion)
}
\arguments{
\item{ModelName}{[required] The name of the machine learning model that this version belongs to.}

\item{ModelVersion}{[required] The version of the machine learning model.}
}
\description{
Retrieves information about a specific machine learning model version.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_describe_model_version/} for full documentation.
}
\keyword{internal}
