% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_describe_resource_policy}
\alias{lookoutequipment_describe_resource_policy}
\title{Provides the details of a resource policy attached to a resource}
\usage{
lookoutequipment_describe_resource_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource that is associated with
the resource policy.}
}
\description{
Provides the details of a resource policy attached to a resource.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_describe_resource_policy/} for full documentation.
}
\keyword{internal}
