% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_import_dataset}
\alias{lookoutequipment_import_dataset}
\title{Imports a dataset}
\usage{
lookoutequipment_import_dataset(
  SourceDatasetArn,
  DatasetName = NULL,
  ClientToken,
  ServerSideKmsKeyId = NULL,
  Tags = NULL
)
}
\arguments{
\item{SourceDatasetArn}{[required] The Amazon Resource Name (ARN) of the dataset to import.}

\item{DatasetName}{The name of the machine learning dataset to be created. If the dataset
already exists, Amazon Lookout for Equipment overwrites the existing
dataset. If you don't specify this field, it is filled with the name of
the source dataset.}

\item{ClientToken}{[required] A unique identifier for the request. If you do not set the client
request token, Amazon Lookout for Equipment generates one.}

\item{ServerSideKmsKeyId}{Provides the identifier of the KMS key key used to encrypt model data by
Amazon Lookout for Equipment.}

\item{Tags}{Any tags associated with the dataset to be created.}
}
\description{
Imports a dataset.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_import_dataset/} for full documentation.
}
\keyword{internal}
