% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_list_data_ingestion_jobs}
\alias{lookoutequipment_list_data_ingestion_jobs}
\title{Provides a list of all data ingestion jobs, including dataset name and
ARN, S3 location of the input data, status, and so on}
\usage{
lookoutequipment_list_data_ingestion_jobs(
  DatasetName = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  Status = NULL
)
}
\arguments{
\item{DatasetName}{The name of the dataset being used for the data ingestion job.}

\item{NextToken}{An opaque pagination token indicating where to continue the listing of
data ingestion jobs.}

\item{MaxResults}{Specifies the maximum number of data ingestion jobs to list.}

\item{Status}{Indicates the status of the data ingestion job.}
}
\description{
Provides a list of all data ingestion jobs, including dataset name and ARN, S3 location of the input data, status, and so on.

See \url{https://paws-r.github.io/docs/lookoutequipment/list_data_ingestion_jobs.html} for full documentation.
}
\keyword{internal}
