% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_list_datasets}
\alias{lookoutequipment_list_datasets}
\title{Lists all datasets currently available in your account, filtering on the
dataset name}
\usage{
lookoutequipment_list_datasets(
  NextToken = NULL,
  MaxResults = NULL,
  DatasetNameBeginsWith = NULL
)
}
\arguments{
\item{NextToken}{An opaque pagination token indicating where to continue the listing of
datasets.}

\item{MaxResults}{Specifies the maximum number of datasets to list.}

\item{DatasetNameBeginsWith}{The beginning of the name of the datasets to be listed.}
}
\description{
Lists all datasets currently available in your account, filtering on the dataset name.

See \url{https://paws-r.github.io/docs/lookoutequipment/list_datasets.html} for full documentation.
}
\keyword{internal}
