% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_list_inference_executions}
\alias{lookoutequipment_list_inference_executions}
\title{Lists all inference executions that have been performed by the specified
inference scheduler}
\usage{
lookoutequipment_list_inference_executions(
  NextToken = NULL,
  MaxResults = NULL,
  InferenceSchedulerName,
  DataStartTimeAfter = NULL,
  DataEndTimeBefore = NULL,
  Status = NULL
)
}
\arguments{
\item{NextToken}{An opaque pagination token indicating where to continue the listing of
inference executions.}

\item{MaxResults}{Specifies the maximum number of inference executions to list.}

\item{InferenceSchedulerName}{[required] The name of the inference scheduler for the inference execution listed.}

\item{DataStartTimeAfter}{The time reference in the inferenced dataset after which Amazon Lookout
for Equipment started the inference execution.}

\item{DataEndTimeBefore}{The time reference in the inferenced dataset before which Amazon Lookout
for Equipment stopped the inference execution.}

\item{Status}{The status of the inference execution.}
}
\description{
Lists all inference executions that have been performed by the specified inference scheduler.

See \url{https://paws-r.github.io/docs/lookoutequipment/list_inference_executions.html} for full documentation.
}
\keyword{internal}
