% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_list_inference_schedulers}
\alias{lookoutequipment_list_inference_schedulers}
\title{Retrieves a list of all inference schedulers currently available for
your account}
\usage{
lookoutequipment_list_inference_schedulers(
  NextToken = NULL,
  MaxResults = NULL,
  InferenceSchedulerNameBeginsWith = NULL,
  ModelName = NULL,
  Status = NULL
)
}
\arguments{
\item{NextToken}{An opaque pagination token indicating where to continue the listing of
inference schedulers.}

\item{MaxResults}{Specifies the maximum number of inference schedulers to list.}

\item{InferenceSchedulerNameBeginsWith}{The beginning of the name of the inference schedulers to be listed.}

\item{ModelName}{The name of the ML model used by the inference scheduler to be listed.}

\item{Status}{Specifies the current status of the inference schedulers.}
}
\description{
Retrieves a list of all inference schedulers currently available for your account.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_list_inference_schedulers/} for full documentation.
}
\keyword{internal}
