% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_list_label_groups}
\alias{lookoutequipment_list_label_groups}
\title{Returns a list of the label groups}
\usage{
lookoutequipment_list_label_groups(
  LabelGroupNameBeginsWith = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{LabelGroupNameBeginsWith}{The beginning of the name of the label groups to be listed.}

\item{NextToken}{An opaque pagination token indicating where to continue the listing of
label groups.}

\item{MaxResults}{Specifies the maximum number of label groups to list.}
}
\description{
Returns a list of the label groups.

See \url{https://paws-r.github.io/docs/lookoutequipment/list_label_groups.html} for full documentation.
}
\keyword{internal}
