% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_list_labels}
\alias{lookoutequipment_list_labels}
\title{Provides a list of labels}
\usage{
lookoutequipment_list_labels(
  LabelGroupName,
  IntervalStartTime = NULL,
  IntervalEndTime = NULL,
  FaultCode = NULL,
  Equipment = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{LabelGroupName}{[required] Retruns the name of the label group.}

\item{IntervalStartTime}{Returns all the labels with a end time equal to or later than the start
time given.}

\item{IntervalEndTime}{Returns all labels with a start time earlier than the end time given.}

\item{FaultCode}{Returns labels with a particular fault code.}

\item{Equipment}{Lists the labels that pertain to a particular piece of equipment.}

\item{NextToken}{An opaque pagination token indicating where to continue the listing of
label groups.}

\item{MaxResults}{Specifies the maximum number of labels to list.}
}
\description{
Provides a list of labels.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_list_labels/} for full documentation.
}
\keyword{internal}
