% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_list_models}
\alias{lookoutequipment_list_models}
\title{Generates a list of all models in the account, including model name and
ARN, dataset, and status}
\usage{
lookoutequipment_list_models(
  NextToken = NULL,
  MaxResults = NULL,
  Status = NULL,
  ModelNameBeginsWith = NULL,
  DatasetNameBeginsWith = NULL
)
}
\arguments{
\item{NextToken}{An opaque pagination token indicating where to continue the listing of
machine learning models.}

\item{MaxResults}{Specifies the maximum number of machine learning models to list.}

\item{Status}{The status of the machine learning model.}

\item{ModelNameBeginsWith}{The beginning of the name of the machine learning models being listed.}

\item{DatasetNameBeginsWith}{The beginning of the name of the dataset of the machine learning models
to be listed.}
}
\description{
Generates a list of all models in the account, including model name and ARN, dataset, and status.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_list_models/} for full documentation.
}
\keyword{internal}
