% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_list_retraining_schedulers}
\alias{lookoutequipment_list_retraining_schedulers}
\title{Lists all retraining schedulers in your account, filtering by model name
prefix and status}
\usage{
lookoutequipment_list_retraining_schedulers(
  ModelNameBeginsWith = NULL,
  Status = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ModelNameBeginsWith}{Specify this field to only list retraining schedulers whose machine
learning models begin with the value you specify.}

\item{Status}{Specify this field to only list retraining schedulers whose status
matches the value you specify.}

\item{NextToken}{If the number of results exceeds the maximum, a pagination token is
returned. Use the token in the request to show the next page of
retraining schedulers.}

\item{MaxResults}{Specifies the maximum number of retraining schedulers to list.}
}
\description{
Lists all retraining schedulers in your account, filtering by model name prefix and status.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_list_retraining_schedulers/} for full documentation.
}
\keyword{internal}
