% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_list_tags_for_resource}
\alias{lookoutequipment_list_tags_for_resource}
\title{Lists all the tags for a specified resource, including key and value}
\usage{
lookoutequipment_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource (such as the dataset or
model) that is the focus of the
\code{\link[=lookoutequipment_list_tags_for_resource]{list_tags_for_resource}}
operation.}
}
\description{
Lists all the tags for a specified resource, including key and value.

See \url{https://paws-r.github.io/docs/lookoutequipment/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
