% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_put_resource_policy}
\alias{lookoutequipment_put_resource_policy}
\title{Creates a resource control policy for a given resource}
\usage{
lookoutequipment_put_resource_policy(
  ResourceArn,
  ResourcePolicy,
  PolicyRevisionId = NULL,
  ClientToken
)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource for which the policy is
being created.}

\item{ResourcePolicy}{[required] The JSON-formatted resource policy to create.}

\item{PolicyRevisionId}{A unique identifier for a revision of the resource policy.}

\item{ClientToken}{[required] A unique identifier for the request. If you do not set the client
request token, Amazon Lookout for Equipment generates one.}
}
\description{
Creates a resource control policy for a given resource.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_put_resource_policy/} for full documentation.
}
\keyword{internal}
