% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_tag_resource}
\alias{lookoutequipment_tag_resource}
\title{Associates a given tag to a resource in your account}
\usage{
lookoutequipment_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the specific resource to which the tag
should be associated.}

\item{Tags}{[required] The tag or tags to be associated with a specific resource. Both the tag
key and value are specified.}
}
\description{
Associates a given tag to a resource in your account. A tag is a key-value pair which can be added to an Amazon Lookout for Equipment resource as metadata. Tags can be used for organizing your resources as well as helping you to search and filter by tag. Multiple tags can be added to a resource, either when you create it, or later. Up to 50 tags can be associated with each resource.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_tag_resource/} for full documentation.
}
\keyword{internal}
