% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_untag_resource}
\alias{lookoutequipment_untag_resource}
\title{Removes a specific tag from a given resource}
\usage{
lookoutequipment_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource to which the tag is
currently associated.}

\item{TagKeys}{[required] Specifies the key of the tag to be removed from a specified resource.}
}
\description{
Removes a specific tag from a given resource. The tag is specified by its key.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_untag_resource/} for full documentation.
}
\keyword{internal}
