% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_update_active_model_version}
\alias{lookoutequipment_update_active_model_version}
\title{Sets the active model version for a given machine learning model}
\usage{
lookoutequipment_update_active_model_version(ModelName, ModelVersion)
}
\arguments{
\item{ModelName}{[required] The name of the machine learning model for which the active model
version is being set.}

\item{ModelVersion}{[required] The version of the machine learning model for which the active model
version is being set.}
}
\description{
Sets the active model version for a given machine learning model.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_update_active_model_version/} for full documentation.
}
\keyword{internal}
