% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_update_label_group}
\alias{lookoutequipment_update_label_group}
\title{Updates the label group}
\usage{
lookoutequipment_update_label_group(LabelGroupName, FaultCodes = NULL)
}
\arguments{
\item{LabelGroupName}{[required] The name of the label group to be updated.}

\item{FaultCodes}{Updates the code indicating the type of anomaly associated with the
label.

Data in this field will be retained for service usage. Follow best
practices for the security of your data.}
}
\description{
Updates the label group.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_update_label_group/} for full documentation.
}
\keyword{internal}
