% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_update_model}
\alias{lookoutequipment_update_model}
\title{Updates a model in the account}
\usage{
lookoutequipment_update_model(
  ModelName,
  LabelsInputConfiguration = NULL,
  RoleArn = NULL,
  ModelDiagnosticsOutputConfiguration = NULL
)
}
\arguments{
\item{ModelName}{[required] The name of the model to update.}

\item{LabelsInputConfiguration}{}

\item{RoleArn}{The ARN of the model to update.}

\item{ModelDiagnosticsOutputConfiguration}{The Amazon S3 location where you want Amazon Lookout for Equipment to
save the pointwise model diagnostics for the model. You must also
specify the \code{RoleArn} request parameter.}
}
\description{
Updates a model in the account.

See \url{https://www.paws-r-sdk.com/docs/lookoutequipment_update_model/} for full documentation.
}
\keyword{internal}
