% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_create_anomaly_detector}
\alias{lookoutmetrics_create_anomaly_detector}
\title{Creates an anomaly detector}
\usage{
lookoutmetrics_create_anomaly_detector(
  AnomalyDetectorName,
  AnomalyDetectorDescription = NULL,
  AnomalyDetectorConfig,
  KmsKeyArn = NULL,
  Tags = NULL
)
}
\arguments{
\item{AnomalyDetectorName}{[required] The name of the detector.}

\item{AnomalyDetectorDescription}{A description of the detector.}

\item{AnomalyDetectorConfig}{[required] Contains information about the configuration of the anomaly detector.}

\item{KmsKeyArn}{The ARN of the KMS key to use to encrypt your data.}

\item{Tags}{A list of
\href{https://docs.aws.amazon.com/lookoutmetrics/latest/dev/detectors-tags.html}{tags}
to apply to the anomaly detector.}
}
\description{
Creates an anomaly detector.

See \url{https://www.paws-r-sdk.com/docs/lookoutmetrics_create_anomaly_detector/} for full documentation.
}
\keyword{internal}
