% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_deactivate_anomaly_detector}
\alias{lookoutmetrics_deactivate_anomaly_detector}
\title{Deactivates an anomaly detector}
\usage{
lookoutmetrics_deactivate_anomaly_detector(AnomalyDetectorArn)
}
\arguments{
\item{AnomalyDetectorArn}{[required] The Amazon Resource Name (ARN) of the anomaly detector.}
}
\description{
Deactivates an anomaly detector.

See \url{https://paws-r.github.io/docs/lookoutmetrics/deactivate_anomaly_detector.html} for full documentation.
}
\keyword{internal}
