% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_delete_alert}
\alias{lookoutmetrics_delete_alert}
\title{Deletes an alert}
\usage{
lookoutmetrics_delete_alert(AlertArn)
}
\arguments{
\item{AlertArn}{[required] The ARN of the alert to delete.}
}
\description{
Deletes an alert.

See \url{https://www.paws-r-sdk.com/docs/lookoutmetrics_delete_alert/} for full documentation.
}
\keyword{internal}
