% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_delete_anomaly_detector}
\alias{lookoutmetrics_delete_anomaly_detector}
\title{Deletes a detector}
\usage{
lookoutmetrics_delete_anomaly_detector(AnomalyDetectorArn)
}
\arguments{
\item{AnomalyDetectorArn}{[required] The ARN of the detector to delete.}
}
\description{
Deletes a detector. Deleting an anomaly detector will delete all of its corresponding resources including any configured datasets and alerts.

See \url{https://paws-r.github.io/docs/lookoutmetrics/delete_anomaly_detector.html} for full documentation.
}
\keyword{internal}
