% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_describe_anomaly_detection_executions}
\alias{lookoutmetrics_describe_anomaly_detection_executions}
\title{Returns information about the status of the specified anomaly detection
jobs}
\usage{
lookoutmetrics_describe_anomaly_detection_executions(
  AnomalyDetectorArn,
  Timestamp = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AnomalyDetectorArn}{[required] The Amazon Resource Name (ARN) of the anomaly detector.}

\item{Timestamp}{The timestamp of the anomaly detection job.}

\item{MaxResults}{The number of items to return in the response.}

\item{NextToken}{Specify the pagination token that's returned by a previous request to
retrieve the next page of results.}
}
\description{
Returns information about the status of the specified anomaly detection jobs.

See \url{https://paws-r.github.io/docs/lookoutmetrics/describe_anomaly_detection_executions.html} for full documentation.
}
\keyword{internal}
