% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_describe_anomaly_detector}
\alias{lookoutmetrics_describe_anomaly_detector}
\title{Describes a detector}
\usage{
lookoutmetrics_describe_anomaly_detector(AnomalyDetectorArn)
}
\arguments{
\item{AnomalyDetectorArn}{[required] The ARN of the detector to describe.}
}
\description{
Describes a detector.

See \url{https://paws-r.github.io/docs/lookoutmetrics/describe_anomaly_detector.html} for full documentation.
}
\keyword{internal}
