% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_detect_metric_set_config}
\alias{lookoutmetrics_detect_metric_set_config}
\title{Detects an Amazon S3 dataset's file format, interval, and offset}
\usage{
lookoutmetrics_detect_metric_set_config(
  AnomalyDetectorArn,
  AutoDetectionMetricSource
)
}
\arguments{
\item{AnomalyDetectorArn}{[required] An anomaly detector ARN.}

\item{AutoDetectionMetricSource}{[required] A data source.}
}
\description{
Detects an Amazon S3 dataset's file format, interval, and offset.

See \url{https://paws-r.github.io/docs/lookoutmetrics/detect_metric_set_config.html} for full documentation.
}
\keyword{internal}
