% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_get_anomaly_group}
\alias{lookoutmetrics_get_anomaly_group}
\title{Returns details about a group of anomalous metrics}
\usage{
lookoutmetrics_get_anomaly_group(AnomalyGroupId, AnomalyDetectorArn)
}
\arguments{
\item{AnomalyGroupId}{[required] The ID of the anomaly group.}

\item{AnomalyDetectorArn}{[required] The Amazon Resource Name (ARN) of the anomaly detector.}
}
\description{
Returns details about a group of anomalous metrics.

See \url{https://paws-r.github.io/docs/lookoutmetrics/get_anomaly_group.html} for full documentation.
}
\keyword{internal}
