% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_get_feedback}
\alias{lookoutmetrics_get_feedback}
\title{Get feedback for an anomaly group}
\usage{
lookoutmetrics_get_feedback(
  AnomalyDetectorArn,
  AnomalyGroupTimeSeriesFeedback,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AnomalyDetectorArn}{[required] The Amazon Resource Name (ARN) of the anomaly detector.}

\item{AnomalyGroupTimeSeriesFeedback}{[required] The anomalous metric and group ID.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{Specify the pagination token that's returned by a previous request to
retrieve the next page of results.}
}
\description{
Get feedback for an anomaly group.

See \url{https://paws-r.github.io/docs/lookoutmetrics/get_feedback.html} for full documentation.
}
\keyword{internal}
