% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_list_anomaly_detectors}
\alias{lookoutmetrics_list_anomaly_detectors}
\title{Lists the detectors in the current AWS Region}
\usage{
lookoutmetrics_list_anomaly_detectors(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{If the result of the previous request was truncated, the response
includes a \code{NextToken}. To retrieve the next set of results, use the
token in the next request. Tokens expire after 24 hours.}
}
\description{
Lists the detectors in the current AWS Region.

See \url{https://www.paws-r-sdk.com/docs/lookoutmetrics_list_anomaly_detectors/} for full documentation.
}
\keyword{internal}
