% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_list_anomaly_group_summaries}
\alias{lookoutmetrics_list_anomaly_group_summaries}
\title{Returns a list of anomaly groups}
\usage{
lookoutmetrics_list_anomaly_group_summaries(
  AnomalyDetectorArn,
  SensitivityThreshold,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AnomalyDetectorArn}{[required] The Amazon Resource Name (ARN) of the anomaly detector.}

\item{SensitivityThreshold}{[required] The minimum severity score for inclusion in the output.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{Specify the pagination token that's returned by a previous request to
retrieve the next page of results.}
}
\description{
Returns a list of anomaly groups.

See \url{https://paws-r.github.io/docs/lookoutmetrics/list_anomaly_group_summaries.html} for full documentation.
}
\keyword{internal}
