% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_list_anomaly_group_time_series}
\alias{lookoutmetrics_list_anomaly_group_time_series}
\title{Gets a list of anomalous metrics for a measure in an anomaly group}
\usage{
lookoutmetrics_list_anomaly_group_time_series(
  AnomalyDetectorArn,
  AnomalyGroupId,
  MetricName,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AnomalyDetectorArn}{[required] The Amazon Resource Name (ARN) of the anomaly detector.}

\item{AnomalyGroupId}{[required] The ID of the anomaly group.}

\item{MetricName}{[required] The name of the measure field.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{Specify the pagination token that's returned by a previous request to
retrieve the next page of results.}
}
\description{
Gets a list of anomalous metrics for a measure in an anomaly group.

See \url{https://paws-r.github.io/docs/lookoutmetrics/list_anomaly_group_time_series.html} for full documentation.
}
\keyword{internal}
