% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_list_metric_sets}
\alias{lookoutmetrics_list_metric_sets}
\title{Lists the datasets in the current AWS Region}
\usage{
lookoutmetrics_list_metric_sets(
  AnomalyDetectorArn = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AnomalyDetectorArn}{The ARN of the anomaly detector containing the metrics sets to list.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{If the result of the previous request was truncated, the response
includes a \code{NextToken}. To retrieve the next set of results, use the
token in the next request. Tokens expire after 24 hours.}
}
\description{
Lists the datasets in the current AWS Region.

See \url{https://paws-r.github.io/docs/lookoutmetrics/list_metric_sets.html} for full documentation.
}
\keyword{internal}
