% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_list_tags_for_resource}
\alias{lookoutmetrics_list_tags_for_resource}
\title{Gets a list of tags for a detector, dataset, or alert}
\usage{
lookoutmetrics_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The resource's Amazon Resource Name (ARN).}
}
\description{
Gets a list of \href{https://docs.aws.amazon.com/lookoutmetrics/latest/dev/detectors-tags.html}{tags} for a detector, dataset, or alert.

See \url{https://www.paws-r-sdk.com/docs/lookoutmetrics_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
