% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_put_feedback}
\alias{lookoutmetrics_put_feedback}
\title{Add feedback for an anomalous metric}
\usage{
lookoutmetrics_put_feedback(AnomalyDetectorArn, AnomalyGroupTimeSeriesFeedback)
}
\arguments{
\item{AnomalyDetectorArn}{[required] The Amazon Resource Name (ARN) of the anomaly detector.}

\item{AnomalyGroupTimeSeriesFeedback}{[required] Feedback for an anomalous metric.}
}
\description{
Add feedback for an anomalous metric.

See \url{https://www.paws-r-sdk.com/docs/lookoutmetrics_put_feedback/} for full documentation.
}
\keyword{internal}
