% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_tag_resource}
\alias{lookoutmetrics_tag_resource}
\title{Adds tags to a detector, dataset, or alert}
\usage{
lookoutmetrics_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The resource's Amazon Resource Name (ARN).}

\item{Tags}{[required] Tags to apply to the resource. Tag keys and values can contain letters,
numbers, spaces, and the following symbols: \verb{_.:/=+@-}}
}
\description{
Adds \href{https://docs.aws.amazon.com/lookoutmetrics/latest/dev/detectors-tags.html}{tags} to a detector, dataset, or alert.

See \url{https://www.paws-r-sdk.com/docs/lookoutmetrics_tag_resource/} for full documentation.
}
\keyword{internal}
