% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_untag_resource}
\alias{lookoutmetrics_untag_resource}
\title{Removes tags from a detector, dataset, or alert}
\usage{
lookoutmetrics_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The resource's Amazon Resource Name (ARN).}

\item{TagKeys}{[required] Keys to remove from the resource's tags.}
}
\description{
Removes \href{https://docs.aws.amazon.com/lookoutmetrics/latest/dev/detectors-tags.html}{tags} from a detector, dataset, or alert.

See \url{https://www.paws-r-sdk.com/docs/lookoutmetrics_untag_resource/} for full documentation.
}
\keyword{internal}
