% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_update_alert}
\alias{lookoutmetrics_update_alert}
\title{Make changes to an existing alert}
\usage{
lookoutmetrics_update_alert(
  AlertArn,
  AlertDescription = NULL,
  AlertSensitivityThreshold = NULL,
  Action = NULL,
  AlertFilters = NULL
)
}
\arguments{
\item{AlertArn}{[required] The ARN of the alert to update.}

\item{AlertDescription}{A description of the alert.}

\item{AlertSensitivityThreshold}{An integer from 0 to 100 specifying the alert sensitivity threshold.}

\item{Action}{Action that will be triggered when there is an alert.}

\item{AlertFilters}{The configuration of the alert filters, containing MetricList and
DimensionFilterList.}
}
\description{
Make changes to an existing alert.

See \url{https://www.paws-r-sdk.com/docs/lookoutmetrics_update_alert/} for full documentation.
}
\keyword{internal}
