% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_update_anomaly_detector}
\alias{lookoutmetrics_update_anomaly_detector}
\title{Updates a detector}
\usage{
lookoutmetrics_update_anomaly_detector(
  AnomalyDetectorArn,
  KmsKeyArn = NULL,
  AnomalyDetectorDescription = NULL,
  AnomalyDetectorConfig = NULL
)
}
\arguments{
\item{AnomalyDetectorArn}{[required] The ARN of the detector to update.}

\item{KmsKeyArn}{The Amazon Resource Name (ARN) of an AWS KMS encryption key.}

\item{AnomalyDetectorDescription}{The updated detector description.}

\item{AnomalyDetectorConfig}{Contains information about the configuration to which the detector will
be updated.}
}
\description{
Updates a detector. After activation, you can only change a detector's ingestion delay and description.

See \url{https://paws-r.github.io/docs/lookoutmetrics/update_anomaly_detector.html} for full documentation.
}
\keyword{internal}
