% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning_operations.R
\name{machinelearning_create_realtime_endpoint}
\alias{machinelearning_create_realtime_endpoint}
\title{Creates a real-time endpoint for the MLModel}
\usage{
machinelearning_create_realtime_endpoint(MLModelId)
}
\arguments{
\item{MLModelId}{[required] The ID assigned to the \code{MLModel} during creation.}
}
\description{
Creates a real-time endpoint for the \code{MLModel}. The endpoint contains
the URI of the \code{MLModel}; that is, the location to send real-time
prediction requests for the specified \code{MLModel}.
}
\section{Request syntax}{
\preformatted{svc$create_realtime_endpoint(
  MLModelId = "string"
)
}
}

\keyword{internal}
