% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning_operations.R
\name{machinelearning_delete_batch_prediction}
\alias{machinelearning_delete_batch_prediction}
\title{Assigns the DELETED status to a BatchPrediction, rendering it unusable}
\usage{
machinelearning_delete_batch_prediction(BatchPredictionId)
}
\arguments{
\item{BatchPredictionId}{[required] A user-supplied ID that uniquely identifies the \code{BatchPrediction}.}
}
\description{
Assigns the DELETED status to a \code{BatchPrediction}, rendering it unusable.

See \url{https://www.paws-r-sdk.com/docs/machinelearning_delete_batch_prediction/} for full documentation.
}
\keyword{internal}
