% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning_operations.R
\name{machinelearning_delete_data_source}
\alias{machinelearning_delete_data_source}
\title{Assigns the DELETED status to a DataSource, rendering it unusable}
\usage{
machinelearning_delete_data_source(DataSourceId)
}
\arguments{
\item{DataSourceId}{[required] A user-supplied ID that uniquely identifies the \code{DataSource}.}
}
\description{
Assigns the DELETED status to a \code{DataSource}, rendering it unusable.

See \url{https://www.paws-r-sdk.com/docs/machinelearning_delete_data_source/} for full documentation.
}
\keyword{internal}
