% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning_operations.R
\name{machinelearning_delete_evaluation}
\alias{machinelearning_delete_evaluation}
\title{Assigns the DELETED status to an Evaluation, rendering it unusable}
\usage{
machinelearning_delete_evaluation(EvaluationId)
}
\arguments{
\item{EvaluationId}{[required] A user-supplied ID that uniquely identifies the \code{Evaluation} to delete.}
}
\description{
Assigns the \code{DELETED} status to an \code{Evaluation}, rendering it unusable.

See \url{https://paws-r.github.io/docs/machinelearning/delete_evaluation.html} for full documentation.
}
\keyword{internal}
