% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning_operations.R
\name{machinelearning_delete_ml_model}
\alias{machinelearning_delete_ml_model}
\title{Assigns the DELETED status to an MLModel, rendering it unusable}
\usage{
machinelearning_delete_ml_model(MLModelId)
}
\arguments{
\item{MLModelId}{[required] A user-supplied ID that uniquely identifies the \code{MLModel}.}
}
\description{
Assigns the \code{DELETED} status to an \code{MLModel}, rendering it unusable.

See \url{https://www.paws-r-sdk.com/docs/machinelearning_delete_ml_model/} for full documentation.
}
\keyword{internal}
