% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning_operations.R
\name{machinelearning_describe_evaluations}
\alias{machinelearning_describe_evaluations}
\title{Returns a list of DescribeEvaluations that match the search criteria in
the request}
\usage{
machinelearning_describe_evaluations(FilterVariable, EQ, GT, LT, GE, LE,
  NE, Prefix, SortOrder, NextToken, Limit)
}
\arguments{
\item{FilterVariable}{Use one of the following variable to filter a list of \code{Evaluation}
objects:
\itemize{
\item \code{CreatedAt} - Sets the search criteria to the \code{Evaluation} creation
date.
\item \code{Status} - Sets the search criteria to the \code{Evaluation} status.
\item \code{Name} - Sets the search criteria to the contents of \code{Evaluation}
\code{Name}.
\item \code{IAMUser} - Sets the search criteria to the user account that
invoked an \code{Evaluation}.
\item \code{MLModelId} - Sets the search criteria to the \code{MLModel} that was
evaluated.
\item \code{DataSourceId} - Sets the search criteria to the \code{DataSource} used
in \code{Evaluation}.
\item \code{DataUri} - Sets the search criteria to the data file(s) used in
\code{Evaluation}. The URL can identify either a file or an Amazon Simple
Storage Solution (Amazon S3) bucket or directory.
}}

\item{EQ}{The equal to operator. The \code{Evaluation} results will have
\code{FilterVariable} values that exactly match the value specified with
\code{EQ}.}

\item{GT}{The greater than operator. The \code{Evaluation} results will have
\code{FilterVariable} values that are greater than the value specified with
\code{GT}.}

\item{LT}{The less than operator. The \code{Evaluation} results will have
\code{FilterVariable} values that are less than the value specified with
\code{LT}.}

\item{GE}{The greater than or equal to operator. The \code{Evaluation} results will
have \code{FilterVariable} values that are greater than or equal to the value
specified with \code{GE}.}

\item{LE}{The less than or equal to operator. The \code{Evaluation} results will have
\code{FilterVariable} values that are less than or equal to the value
specified with \code{LE}.}

\item{NE}{The not equal to operator. The \code{Evaluation} results will have
\code{FilterVariable} values not equal to the value specified with \code{NE}.}

\item{Prefix}{A string that is found at the beginning of a variable, such as \code{Name} or
\code{Id}.

For example, an \code{Evaluation} could have the \code{Name}
\code{2014-09-09-HolidayGiftMailer}. To search for this \code{Evaluation}, select
\code{Name} for the \code{FilterVariable} and any of the following strings for the
\code{Prefix}:
\itemize{
\item 2014-09
\item 2014-09-09
\item 2014-09-09-Holiday
}}

\item{SortOrder}{A two-value parameter that determines the sequence of the resulting list
of \code{Evaluation}.
\itemize{
\item \code{asc} - Arranges the list in ascending order (A-Z, 0-9).
\item \code{dsc} - Arranges the list in descending order (Z-A, 9-0).
}

Results are sorted by \code{FilterVariable}.}

\item{NextToken}{The ID of the page in the paginated results.}

\item{Limit}{The maximum number of \code{Evaluation} to include in the result.}
}
\value{
A list with the following syntax:\preformatted{list(
  Results = list(
    list(
      EvaluationId = "string",
      MLModelId = "string",
      EvaluationDataSourceId = "string",
      InputDataLocationS3 = "string",
      CreatedByIamUser = "string",
      CreatedAt = as.POSIXct(
        "2015-01-01"
      ),
      LastUpdatedAt = as.POSIXct(
        "2015-01-01"
      ),
      Name = "string",
      Status = "PENDING"|"INPROGRESS"|"FAILED"|"COMPLETED"|"DELETED",
      PerformanceMetrics = list(
        Properties = list(
          "string"
        )
      ),
      Message = "string",
      ComputeTime = 123,
      FinishedAt = as.POSIXct(
        "2015-01-01"
      ),
      StartedAt = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Returns a list of
\code{\link[=machinelearning_describe_evaluations]{describe_evaluations}} that
match the search criteria in the request.
}
\section{Request syntax}{
\preformatted{svc$describe_evaluations(
  FilterVariable = "CreatedAt"|"LastUpdatedAt"|"Status"|"Name"|"IAMUser"|"MLModelId"|"DataSourceId"|"DataURI",
  EQ = "string",
  GT = "string",
  LT = "string",
  GE = "string",
  LE = "string",
  NE = "string",
  Prefix = "string",
  SortOrder = "asc"|"dsc",
  NextToken = "string",
  Limit = 123
)
}
}

\keyword{internal}
