% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning_operations.R
\name{machinelearning_describe_tags}
\alias{machinelearning_describe_tags}
\title{Describes one or more of the tags for your Amazon ML object}
\usage{
machinelearning_describe_tags(ResourceId, ResourceType)
}
\arguments{
\item{ResourceId}{[required] The ID of the ML object. For example, \code{exampleModelId}.}

\item{ResourceType}{[required] The type of the ML object.}
}
\value{
A list with the following syntax:\preformatted{list(
  ResourceId = "string",
  ResourceType = "BatchPrediction"|"DataSource"|"Evaluation"|"MLModel",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}
\description{
Describes one or more of the tags for your Amazon ML object.
}
\section{Request syntax}{
\preformatted{svc$describe_tags(
  ResourceId = "string",
  ResourceType = "BatchPrediction"|"DataSource"|"Evaluation"|"MLModel"
)
}
}

\keyword{internal}
