% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning_operations.R
\name{machinelearning_get_data_source}
\alias{machinelearning_get_data_source}
\title{Returns a DataSource that includes metadata and data file information,
as well as the current status of the DataSource}
\usage{
machinelearning_get_data_source(DataSourceId, Verbose = NULL)
}
\arguments{
\item{DataSourceId}{[required] The ID assigned to the \code{DataSource} at creation.}

\item{Verbose}{Specifies whether the
\code{\link[=machinelearning_get_data_source]{get_data_source}} operation should
return \code{DataSourceSchema}.

If true, \code{DataSourceSchema} is returned.

If false, \code{DataSourceSchema} is not returned.}
}
\description{
Returns a \code{DataSource} that includes metadata and data file information, as well as the current status of the \code{DataSource}.

See \url{https://www.paws-r-sdk.com/docs/machinelearning_get_data_source/} for full documentation.
}
\keyword{internal}
