% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning_operations.R
\name{machinelearning_update_batch_prediction}
\alias{machinelearning_update_batch_prediction}
\title{Updates the BatchPredictionName of a BatchPrediction}
\usage{
machinelearning_update_batch_prediction(BatchPredictionId,
  BatchPredictionName)
}
\arguments{
\item{BatchPredictionId}{[required] The ID assigned to the \code{BatchPrediction} during creation.}

\item{BatchPredictionName}{[required] A new user-supplied name or description of the \code{BatchPrediction}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  BatchPredictionId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Updates the \code{BatchPredictionName} of a \code{BatchPrediction}.

You can use the
\code{\link[=machinelearning_get_batch_prediction]{get_batch_prediction}} operation
to view the contents of the updated data element.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_batch_prediction(
  BatchPredictionId = "string",
  BatchPredictionName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
