% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning_operations.R
\name{machinelearning_update_data_source}
\alias{machinelearning_update_data_source}
\title{Updates the DataSourceName of a DataSource}
\usage{
machinelearning_update_data_source(DataSourceId, DataSourceName)
}
\arguments{
\item{DataSourceId}{[required] The ID assigned to the \code{DataSource} during creation.}

\item{DataSourceName}{[required] A new user-supplied name or description of the \code{DataSource} that will
replace the current description.}
}
\description{
Updates the \code{DataSourceName} of a \code{DataSource}.

You can use the \code{\link[=machinelearning_get_data_source]{get_data_source}}
operation to view the contents of the updated data element.
}
\section{Request syntax}{
\preformatted{svc$update_data_source(
  DataSourceId = "string",
  DataSourceName = "string"
)
}
}

\keyword{internal}
