% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning_operations.R
\name{machinelearning_update_evaluation}
\alias{machinelearning_update_evaluation}
\title{Updates the EvaluationName of an Evaluation}
\usage{
machinelearning_update_evaluation(EvaluationId, EvaluationName)
}
\arguments{
\item{EvaluationId}{[required] The ID assigned to the \code{Evaluation} during creation.}

\item{EvaluationName}{[required] A new user-supplied name or description of the \code{Evaluation} that will
replace the current content.}
}
\description{
Updates the \code{EvaluationName} of an \code{Evaluation}.

See \url{https://www.paws-r-sdk.com/docs/machinelearning_update_evaluation/} for full documentation.
}
\keyword{internal}
