% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_create_application_instance}
\alias{panorama_create_application_instance}
\title{Creates an application instance and deploys it to a device}
\usage{
panorama_create_application_instance(
  ApplicationInstanceIdToReplace = NULL,
  DefaultRuntimeContextDevice,
  Description = NULL,
  ManifestOverridesPayload = NULL,
  ManifestPayload,
  Name = NULL,
  RuntimeRoleArn = NULL,
  Tags = NULL
)
}
\arguments{
\item{ApplicationInstanceIdToReplace}{The ID of an application instance to replace with the new instance.}

\item{DefaultRuntimeContextDevice}{[required] A device's ID.}

\item{Description}{A description for the application instance.}

\item{ManifestOverridesPayload}{Setting overrides for the application manifest.}

\item{ManifestPayload}{[required] The application's manifest document.}

\item{Name}{A name for the application instance.}

\item{RuntimeRoleArn}{The ARN of a runtime role for the application instance.}

\item{Tags}{Tags for the application instance.}
}
\description{
Creates an application instance and deploys it to a device.

See \url{https://www.paws-r-sdk.com/docs/panorama_create_application_instance/} for full documentation.
}
\keyword{internal}
