% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_create_job_for_devices}
\alias{panorama_create_job_for_devices}
\title{Creates a job to run on a device}
\usage{
panorama_create_job_for_devices(DeviceIds, DeviceJobConfig = NULL, JobType)
}
\arguments{
\item{DeviceIds}{[required] ID of target device.}

\item{DeviceJobConfig}{Configuration settings for a software update job.}

\item{JobType}{[required] The type of job to run.}
}
\description{
Creates a job to run on a device. A job can update a device's software or reboot it.

See \url{https://www.paws-r-sdk.com/docs/panorama_create_job_for_devices/} for full documentation.
}
\keyword{internal}
