% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_create_node_from_template_job}
\alias{panorama_create_node_from_template_job}
\title{Creates a camera stream node}
\usage{
panorama_create_node_from_template_job(
  JobTags = NULL,
  NodeDescription = NULL,
  NodeName,
  OutputPackageName,
  OutputPackageVersion,
  TemplateParameters,
  TemplateType
)
}
\arguments{
\item{JobTags}{Tags for the job.}

\item{NodeDescription}{A description for the node.}

\item{NodeName}{[required] A name for the node.}

\item{OutputPackageName}{[required] An output package name for the node.}

\item{OutputPackageVersion}{[required] An output package version for the node.}

\item{TemplateParameters}{[required] Template parameters for the node.}

\item{TemplateType}{[required] The type of node.}
}
\description{
Creates a camera stream node.

See \url{https://www.paws-r-sdk.com/docs/panorama_create_node_from_template_job/} for full documentation.
}
\keyword{internal}
