% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_create_package}
\alias{panorama_create_package}
\title{Creates a package and storage location in an Amazon S3 access point}
\usage{
panorama_create_package(PackageName, Tags = NULL)
}
\arguments{
\item{PackageName}{[required] A name for the package.}

\item{Tags}{Tags for the package.}
}
\description{
Creates a package and storage location in an Amazon S3 access point.

See \url{https://www.paws-r-sdk.com/docs/panorama_create_package/} for full documentation.
}
\keyword{internal}
