% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_create_package_import_job}
\alias{panorama_create_package_import_job}
\title{Imports a node package}
\usage{
panorama_create_package_import_job(
  ClientToken,
  InputConfig,
  JobTags = NULL,
  JobType,
  OutputConfig
)
}
\arguments{
\item{ClientToken}{[required] A client token for the package import job.}

\item{InputConfig}{[required] An input config for the package import job.}

\item{JobTags}{Tags for the package import job.}

\item{JobType}{[required] A job type for the package import job.}

\item{OutputConfig}{[required] An output config for the package import job.}
}
\description{
Imports a node package.

See \url{https://www.paws-r-sdk.com/docs/panorama_create_package_import_job/} for full documentation.
}
\keyword{internal}
