% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_delete_device}
\alias{panorama_delete_device}
\title{Deletes a device}
\usage{
panorama_delete_device(DeviceId)
}
\arguments{
\item{DeviceId}{[required] The device's ID.}
}
\description{
Deletes a device.

See \url{https://paws-r.github.io/docs/panorama/delete_device.html} for full documentation.
}
\keyword{internal}
