% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_deregister_package_version}
\alias{panorama_deregister_package_version}
\title{Deregisters a package version}
\usage{
panorama_deregister_package_version(
  OwnerAccount = NULL,
  PackageId,
  PackageVersion,
  PatchVersion,
  UpdatedLatestPatchVersion = NULL
)
}
\arguments{
\item{OwnerAccount}{An owner account.}

\item{PackageId}{[required] A package ID.}

\item{PackageVersion}{[required] A package version.}

\item{PatchVersion}{[required] A patch version.}

\item{UpdatedLatestPatchVersion}{If the version was marked latest, the new version to maker as latest.}
}
\description{
Deregisters a package version.

See \url{https://paws-r.github.io/docs/panorama/deregister_package_version.html} for full documentation.
}
\keyword{internal}
