% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_describe_application_instance}
\alias{panorama_describe_application_instance}
\title{Returns information about an application instance on a device}
\usage{
panorama_describe_application_instance(ApplicationInstanceId)
}
\arguments{
\item{ApplicationInstanceId}{[required] The application instance's ID.}
}
\description{
Returns information about an application instance on a device.

See \url{https://www.paws-r-sdk.com/docs/panorama_describe_application_instance/} for full documentation.
}
\keyword{internal}
