% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_describe_application_instance_details}
\alias{panorama_describe_application_instance_details}
\title{Returns information about an application instance's configuration
manifest}
\usage{
panorama_describe_application_instance_details(ApplicationInstanceId)
}
\arguments{
\item{ApplicationInstanceId}{[required] The application instance's ID.}
}
\description{
Returns information about an application instance's configuration manifest.

See \url{https://www.paws-r-sdk.com/docs/panorama_describe_application_instance_details/} for full documentation.
}
\keyword{internal}
