% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_describe_device}
\alias{panorama_describe_device}
\title{Returns information about a device}
\usage{
panorama_describe_device(DeviceId)
}
\arguments{
\item{DeviceId}{[required] The device's ID.}
}
\description{
Returns information about a device.

See \url{https://www.paws-r-sdk.com/docs/panorama_describe_device/} for full documentation.
}
\keyword{internal}
