% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_describe_node}
\alias{panorama_describe_node}
\title{Returns information about a node}
\usage{
panorama_describe_node(NodeId, OwnerAccount = NULL)
}
\arguments{
\item{NodeId}{[required] The node's ID.}

\item{OwnerAccount}{The account ID of the node's owner.}
}
\description{
Returns information about a node.

See \url{https://paws-r.github.io/docs/panorama/describe_node.html} for full documentation.
}
\keyword{internal}
