% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_describe_node_from_template_job}
\alias{panorama_describe_node_from_template_job}
\title{Returns information about a job to create a camera stream node}
\usage{
panorama_describe_node_from_template_job(JobId)
}
\arguments{
\item{JobId}{[required] The job's ID.}
}
\description{
Returns information about a job to create a camera stream node.

See \url{https://www.paws-r-sdk.com/docs/panorama_describe_node_from_template_job/} for full documentation.
}
\keyword{internal}
