% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_describe_package}
\alias{panorama_describe_package}
\title{Returns information about a package}
\usage{
panorama_describe_package(PackageId)
}
\arguments{
\item{PackageId}{[required] The package's ID.}
}
\description{
Returns information about a package.

See \url{https://paws-r.github.io/docs/panorama/describe_package.html} for full documentation.
}
\keyword{internal}
