% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_describe_package_import_job}
\alias{panorama_describe_package_import_job}
\title{Returns information about a package import job}
\usage{
panorama_describe_package_import_job(JobId)
}
\arguments{
\item{JobId}{[required] The job's ID.}
}
\description{
Returns information about a package import job.

See \url{https://www.paws-r-sdk.com/docs/panorama_describe_package_import_job/} for full documentation.
}
\keyword{internal}
