% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_list_application_instances}
\alias{panorama_list_application_instances}
\title{Returns a list of application instances}
\usage{
panorama_list_application_instances(
  DeviceId = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  StatusFilter = NULL
)
}
\arguments{
\item{DeviceId}{The application instances' device ID.}

\item{MaxResults}{The maximum number of application instances to return in one page of
results.}

\item{NextToken}{Specify the pagination token from a previous request to retrieve the
next page of results.}

\item{StatusFilter}{Only include instances with a specific status.}
}
\description{
Returns a list of application instances.

See \url{https://www.paws-r-sdk.com/docs/panorama_list_application_instances/} for full documentation.
}
\keyword{internal}
