% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_list_devices_jobs}
\alias{panorama_list_devices_jobs}
\title{Returns a list of jobs}
\usage{
panorama_list_devices_jobs(
  DeviceId = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DeviceId}{Filter results by the job's target device ID.}

\item{MaxResults}{The maximum number of device jobs to return in one page of results.}

\item{NextToken}{Specify the pagination token from a previous request to retrieve the
next page of results.}
}
\description{
Returns a list of jobs.

See \url{https://paws-r.github.io/docs/panorama/list_devices_jobs.html} for full documentation.
}
\keyword{internal}
