% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_list_package_import_jobs}
\alias{panorama_list_package_import_jobs}
\title{Returns a list of package import jobs}
\usage{
panorama_list_package_import_jobs(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of package import jobs to return in one page of
results.}

\item{NextToken}{Specify the pagination token from a previous request to retrieve the
next page of results.}
}
\description{
Returns a list of package import jobs.

See \url{https://paws-r.github.io/docs/panorama/list_package_import_jobs.html} for full documentation.
}
\keyword{internal}
