% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_list_tags_for_resource}
\alias{panorama_list_tags_for_resource}
\title{Returns a list of tags for a resource}
\usage{
panorama_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The resource's ARN.}
}
\description{
Returns a list of tags for a resource.

See \url{https://paws-r.github.io/docs/panorama/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
