% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_register_package_version}
\alias{panorama_register_package_version}
\title{Registers a package version}
\usage{
panorama_register_package_version(
  MarkLatest = NULL,
  OwnerAccount = NULL,
  PackageId,
  PackageVersion,
  PatchVersion
)
}
\arguments{
\item{MarkLatest}{Whether to mark the new version as the latest version.}

\item{OwnerAccount}{An owner account.}

\item{PackageId}{[required] A package ID.}

\item{PackageVersion}{[required] A package version.}

\item{PatchVersion}{[required] A patch version.}
}
\description{
Registers a package version.

See \url{https://paws-r.github.io/docs/panorama/register_package_version.html} for full documentation.
}
\keyword{internal}
