% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_signal_application_instance_node_instances}
\alias{panorama_signal_application_instance_node_instances}
\title{Signal camera nodes to stop or resume}
\usage{
panorama_signal_application_instance_node_instances(
  ApplicationInstanceId,
  NodeSignals
)
}
\arguments{
\item{ApplicationInstanceId}{[required] An application instance ID.}

\item{NodeSignals}{[required] A list of signals.}
}
\description{
Signal camera nodes to stop or resume.

See \url{https://www.paws-r-sdk.com/docs/panorama_signal_application_instance_node_instances/} for full documentation.
}
\keyword{internal}
