% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_tag_resource}
\alias{panorama_tag_resource}
\title{Tags a resource}
\usage{
panorama_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The resource's ARN.}

\item{Tags}{[required] Tags for the resource.}
}
\description{
Tags a resource.

See \url{https://www.paws-r-sdk.com/docs/panorama_tag_resource/} for full documentation.
}
\keyword{internal}
