% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_update_device_metadata}
\alias{panorama_update_device_metadata}
\title{Updates a device's metadata}
\usage{
panorama_update_device_metadata(Description = NULL, DeviceId)
}
\arguments{
\item{Description}{A description for the device.}

\item{DeviceId}{[required] The device's ID.}
}
\description{
Updates a device's metadata.

See \url{https://paws-r.github.io/docs/panorama/update_device_metadata.html} for full documentation.
}
\keyword{internal}
