% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_create_event_tracker}
\alias{personalize_create_event_tracker}
\title{Creates an event tracker that you use when adding event data to a
specified dataset group using the PutEvents API}
\usage{
personalize_create_event_tracker(name, datasetGroupArn, tags = NULL)
}
\arguments{
\item{name}{[required] The name for the event tracker.}

\item{datasetGroupArn}{[required] The Amazon Resource Name (ARN) of the dataset group that receives the
event data.}

\item{tags}{A list of
\href{https://docs.aws.amazon.com/personalize/latest/dg/tagging-resources.html}{tags}
to apply to the event tracker.}
}
\description{
Creates an event tracker that you use when adding event data to a specified dataset group using the \href{https://docs.aws.amazon.com/personalize/latest/dg/API_UBS_PutEvents.html}{PutEvents} API.

See \url{https://www.paws-r-sdk.com/docs/personalize_create_event_tracker/} for full documentation.
}
\keyword{internal}
