% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_create_filter}
\alias{personalize_create_filter}
\title{Creates a recommendation filter}
\usage{
personalize_create_filter(name, datasetGroupArn, filterExpression)
}
\arguments{
\item{name}{[required] The name of the filter to create.}

\item{datasetGroupArn}{[required] The ARN of the dataset group that the filter will belong to.}

\item{filterExpression}{[required] The filter expression defines which items are included or excluded from
recommendations. Filter expression must follow specific format rules.
For information about filter expression structure and syntax, see
filter-expressions.}
}
\value{
A list with the following syntax:\preformatted{list(
  filterArn = "string"
)
}
}
\description{
Creates a recommendation filter. For more information, see filter.
}
\section{Request syntax}{
\preformatted{svc$create_filter(
  name = "string",
  datasetGroupArn = "string",
  filterExpression = "string"
)
}
}

\keyword{internal}
