% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_create_schema}
\alias{personalize_create_schema}
\title{Creates an Amazon Personalize schema from the specified schema string}
\usage{
personalize_create_schema(name, schema, domain = NULL)
}
\arguments{
\item{name}{[required] The name for the schema.}

\item{schema}{[required] A schema in Avro JSON format.}

\item{domain}{The domain for the schema. If you are creating a schema for a dataset in
a Domain dataset group, specify the domain you chose when you created
the Domain dataset group.}
}
\description{
Creates an Amazon Personalize schema from the specified schema string. The schema you create must be in Avro JSON format.

See \url{https://www.paws-r-sdk.com/docs/personalize_create_schema/} for full documentation.
}
\keyword{internal}
