% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_delete_campaign}
\alias{personalize_delete_campaign}
\title{Removes a campaign by deleting the solution deployment}
\usage{
personalize_delete_campaign(campaignArn)
}
\arguments{
\item{campaignArn}{[required] The Amazon Resource Name (ARN) of the campaign to delete.}
}
\description{
Removes a campaign by deleting the solution deployment. The solution that the campaign is based on is not deleted and can be redeployed when needed. A deleted campaign can no longer be specified in a \href{https://docs.aws.amazon.com/personalize/latest/dg/API_RS_GetRecommendations.html}{GetRecommendations} request. For information on creating campaigns, see \code{\link[=personalize_create_campaign]{create_campaign}}.

See \url{https://paws-r.github.io/docs/personalize/delete_campaign.html} for full documentation.
}
\keyword{internal}
