% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_delete_dataset}
\alias{personalize_delete_dataset}
\title{Deletes a dataset}
\usage{
personalize_delete_dataset(datasetArn)
}
\arguments{
\item{datasetArn}{[required] The Amazon Resource Name (ARN) of the dataset to delete.}
}
\description{
Deletes a dataset. You can't delete a dataset if an associated \code{DatasetImportJob} or \code{SolutionVersion} is in the CREATE PENDING or IN PROGRESS state. For more information about deleting datasets, see \href{https://docs.aws.amazon.com/personalize/latest/dg/delete-dataset.html}{Deleting a dataset}.

See \url{https://www.paws-r-sdk.com/docs/personalize_delete_dataset/} for full documentation.
}
\keyword{internal}
