% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_delete_dataset_group}
\alias{personalize_delete_dataset_group}
\title{Deletes a dataset group}
\usage{
personalize_delete_dataset_group(datasetGroupArn)
}
\arguments{
\item{datasetGroupArn}{[required] The ARN of the dataset group to delete.}
}
\description{
Deletes a dataset group. Before you delete a dataset group, you must delete the following:

See \url{https://www.paws-r-sdk.com/docs/personalize_delete_dataset_group/} for full documentation.
}
\keyword{internal}
