% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_delete_metric_attribution}
\alias{personalize_delete_metric_attribution}
\title{Deletes a metric attribution}
\usage{
personalize_delete_metric_attribution(metricAttributionArn)
}
\arguments{
\item{metricAttributionArn}{[required] The metric attribution's Amazon Resource Name (ARN).}
}
\description{
Deletes a metric attribution.

See \url{https://www.paws-r-sdk.com/docs/personalize_delete_metric_attribution/} for full documentation.
}
\keyword{internal}
