% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_delete_recommender}
\alias{personalize_delete_recommender}
\title{Deactivates and removes a recommender}
\usage{
personalize_delete_recommender(recommenderArn)
}
\arguments{
\item{recommenderArn}{[required] The Amazon Resource Name (ARN) of the recommender to delete.}
}
\description{
Deactivates and removes a recommender. A deleted recommender can no longer be specified in a \href{https://docs.aws.amazon.com/personalize/latest/dg/API_RS_GetRecommendations.html}{GetRecommendations} request.

See \url{https://www.paws-r-sdk.com/docs/personalize_delete_recommender/} for full documentation.
}
\keyword{internal}
