% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_delete_schema}
\alias{personalize_delete_schema}
\title{Deletes a schema}
\usage{
personalize_delete_schema(schemaArn)
}
\arguments{
\item{schemaArn}{[required] The Amazon Resource Name (ARN) of the schema to delete.}
}
\description{
Deletes a schema. Before deleting a schema, you must delete all datasets referencing the schema. For more information on schemas, see \code{\link[=personalize_create_schema]{create_schema}}.

See \url{https://www.paws-r-sdk.com/docs/personalize_delete_schema/} for full documentation.
}
\keyword{internal}
