% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_delete_solution}
\alias{personalize_delete_solution}
\title{Deletes all versions of a solution and the Solution object itself}
\usage{
personalize_delete_solution(solutionArn)
}
\arguments{
\item{solutionArn}{[required] The ARN of the solution to delete.}
}
\value{
An empty list.
}
\description{
Deletes all versions of a solution and the \code{Solution} object itself.
Before deleting a solution, you must delete all campaigns based on the
solution. To determine what campaigns are using the solution, call
\code{\link[=personalize_list_campaigns]{list_campaigns}} and supply the Amazon
Resource Name (ARN) of the solution. You can't delete a solution if an
associated \code{SolutionVersion} is in the CREATE PENDING or IN PROGRESS
state. For more information on solutions, see
\code{\link[=personalize_create_solution]{create_solution}}.
}
\section{Request syntax}{
\preformatted{svc$delete_solution(
  solutionArn = "string"
)
}
}

\keyword{internal}
