% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_describe_algorithm}
\alias{personalize_describe_algorithm}
\title{Describes the given algorithm}
\usage{
personalize_describe_algorithm(algorithmArn)
}
\arguments{
\item{algorithmArn}{[required] The Amazon Resource Name (ARN) of the algorithm to describe.}
}
\description{
Describes the given algorithm.
}
\section{Request syntax}{
\preformatted{svc$describe_algorithm(
  algorithmArn = "string"
)
}
}

\keyword{internal}
