% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_describe_batch_segment_job}
\alias{personalize_describe_batch_segment_job}
\title{Gets the properties of a batch segment job including name, Amazon
Resource Name (ARN), status, input and output configurations, and the
ARN of the solution version used to generate segments}
\usage{
personalize_describe_batch_segment_job(batchSegmentJobArn)
}
\arguments{
\item{batchSegmentJobArn}{[required] The ARN of the batch segment job to describe.}
}
\description{
Gets the properties of a batch segment job including name, Amazon Resource Name (ARN), status, input and output configurations, and the ARN of the solution version used to generate segments.

See \url{https://paws-r.github.io/docs/personalize/describe_batch_segment_job.html} for full documentation.
}
\keyword{internal}
