% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_describe_data_deletion_job}
\alias{personalize_describe_data_deletion_job}
\title{Describes the data deletion job created by CreateDataDeletionJob,
including the job status}
\usage{
personalize_describe_data_deletion_job(dataDeletionJobArn)
}
\arguments{
\item{dataDeletionJobArn}{[required] The Amazon Resource Name (ARN) of the data deletion job.}
}
\description{
Describes the data deletion job created by \code{\link[=personalize_create_data_deletion_job]{create_data_deletion_job}}, including the job status.

See \url{https://www.paws-r-sdk.com/docs/personalize_describe_data_deletion_job/} for full documentation.
}
\keyword{internal}
