% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_describe_dataset}
\alias{personalize_describe_dataset}
\title{Describes the given dataset}
\usage{
personalize_describe_dataset(datasetArn)
}
\arguments{
\item{datasetArn}{[required] The Amazon Resource Name (ARN) of the dataset to describe.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  dataset = list(
    name = "string",
    datasetArn = "string",
    datasetGroupArn = "string",
    datasetType = "string",
    schemaArn = "string",
    status = "string",
    creationDateTime = as.POSIXct(
      "2015-01-01"
    ),
    lastUpdatedDateTime = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes the given dataset. For more information on datasets, see
\code{\link[=personalize_create_dataset]{create_dataset}}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_dataset(
  datasetArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
