% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_describe_dataset_import_job}
\alias{personalize_describe_dataset_import_job}
\title{Describes the dataset import job created by CreateDatasetImportJob,
including the import job status}
\usage{
personalize_describe_dataset_import_job(datasetImportJobArn)
}
\arguments{
\item{datasetImportJobArn}{[required] The Amazon Resource Name (ARN) of the dataset import job to describe.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  datasetImportJob = list(
    jobName = "string",
    datasetImportJobArn = "string",
    datasetArn = "string",
    dataSource = list(
      dataLocation = "string"
    ),
    roleArn = "string",
    status = "string",
    creationDateTime = as.POSIXct(
      "2015-01-01"
    ),
    lastUpdatedDateTime = as.POSIXct(
      "2015-01-01"
    ),
    failureReason = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes the dataset import job created by
\code{\link[=personalize_create_dataset_import_job]{create_dataset_import_job}},
including the import job status.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_dataset_import_job(
  datasetImportJobArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
