% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_describe_feature_transformation}
\alias{personalize_describe_feature_transformation}
\title{Describes the given feature transformation}
\usage{
personalize_describe_feature_transformation(featureTransformationArn)
}
\arguments{
\item{featureTransformationArn}{[required] The Amazon Resource Name (ARN) of the feature transformation to
describe.}
}
\value{
A list with the following syntax:\preformatted{list(
  featureTransformation = list(
    name = "string",
    featureTransformationArn = "string",
    defaultParameters = list(
      "string"
    ),
    creationDateTime = as.POSIXct(
      "2015-01-01"
    ),
    lastUpdatedDateTime = as.POSIXct(
      "2015-01-01"
    ),
    status = "string"
  )
)
}
}
\description{
Describes the given feature transformation.
}
\section{Request syntax}{
\preformatted{svc$describe_feature_transformation(
  featureTransformationArn = "string"
)
}
}

\keyword{internal}
