% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_describe_filter}
\alias{personalize_describe_filter}
\title{Describes a filter's properties}
\usage{
personalize_describe_filter(filterArn)
}
\arguments{
\item{filterArn}{[required] The ARN of the filter to describe.}
}
\value{
A list with the following syntax:\preformatted{list(
  filter = list(
    name = "string",
    filterArn = "string",
    creationDateTime = as.POSIXct(
      "2015-01-01"
    ),
    lastUpdatedDateTime = as.POSIXct(
      "2015-01-01"
    ),
    datasetGroupArn = "string",
    failureReason = "string",
    filterExpression = "string",
    status = "string"
  )
)
}
}
\description{
Describes a filter's properties.
}
\section{Request syntax}{
\preformatted{svc$describe_filter(
  filterArn = "string"
)
}
}

\keyword{internal}
