% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_describe_metric_attribution}
\alias{personalize_describe_metric_attribution}
\title{Describes a metric attribution}
\usage{
personalize_describe_metric_attribution(metricAttributionArn)
}
\arguments{
\item{metricAttributionArn}{[required] The metric attribution's Amazon Resource Name (ARN).}
}
\description{
Describes a metric attribution.

See \url{https://paws-r.github.io/docs/personalize/describe_metric_attribution.html} for full documentation.
}
\keyword{internal}
