% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_describe_recommender}
\alias{personalize_describe_recommender}
\title{Describes the given recommender, including its status}
\usage{
personalize_describe_recommender(recommenderArn)
}
\arguments{
\item{recommenderArn}{[required] The Amazon Resource Name (ARN) of the recommender to describe.}
}
\description{
Describes the given recommender, including its status.

See \url{https://paws-r.github.io/docs/personalize/describe_recommender.html} for full documentation.
}
\keyword{internal}
