% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_get_solution_metrics}
\alias{personalize_get_solution_metrics}
\title{Gets the metrics for the specified solution version}
\usage{
personalize_get_solution_metrics(solutionVersionArn)
}
\arguments{
\item{solutionVersionArn}{[required] The Amazon Resource Name (ARN) of the solution version for which to get
metrics.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  solutionVersionArn = "string",
  metrics = list(
    123.0
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets the metrics for the specified solution version.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_solution_metrics(
  solutionVersionArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
