% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_list_batch_segment_jobs}
\alias{personalize_list_batch_segment_jobs}
\title{Gets a list of the batch segment jobs that have been performed off of a
solution version that you specify}
\usage{
personalize_list_batch_segment_jobs(
  solutionVersionArn = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{solutionVersionArn}{The Amazon Resource Name (ARN) of the solution version that the batch
segment jobs used to generate batch segments.}

\item{nextToken}{The token to request the next page of results.}

\item{maxResults}{The maximum number of batch segment job results to return in each page.
The default value is 100.}
}
\description{
Gets a list of the batch segment jobs that have been performed off of a solution version that you specify.

See \url{https://paws-r.github.io/docs/personalize/list_batch_segment_jobs.html} for full documentation.
}
\keyword{internal}
