% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_list_datasets}
\alias{personalize_list_datasets}
\title{Returns the list of datasets contained in the given dataset group}
\usage{
personalize_list_datasets(datasetGroupArn, nextToken, maxResults)
}
\arguments{
\item{datasetGroupArn}{The Amazon Resource Name (ARN) of the dataset group that contains the
datasets to list.}

\item{nextToken}{A token returned from the previous call to
\code{\link[=personalize_list_dataset_import_jobs]{list_dataset_import_jobs}} for
getting the next set of dataset import jobs (if they exist).}

\item{maxResults}{The maximum number of datasets to return.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  datasets = list(
    list(
      name = "string",
      datasetArn = "string",
      datasetType = "string",
      status = "string",
      creationDateTime = as.POSIXct(
        "2015-01-01"
      ),
      lastUpdatedDateTime = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns the list of datasets contained in the given dataset group. The
response provides the properties for each dataset, including the Amazon
Resource Name (ARN). For more information on datasets, see
\code{\link[=personalize_create_dataset]{create_dataset}}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_datasets(
  datasetGroupArn = "string",
  nextToken = "string",
  maxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
