% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_list_metric_attribution_metrics}
\alias{personalize_list_metric_attribution_metrics}
\title{Lists the metrics for the metric attribution}
\usage{
personalize_list_metric_attribution_metrics(
  metricAttributionArn = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{metricAttributionArn}{The Amazon Resource Name (ARN) of the metric attribution to retrieve
attributes for.}

\item{nextToken}{Specify the pagination token from a previous request to retrieve the
next page of results.}

\item{maxResults}{The maximum number of metrics to return in one page of results.}
}
\description{
Lists the metrics for the metric attribution.

See \url{https://paws-r.github.io/docs/personalize/list_metric_attribution_metrics.html} for full documentation.
}
\keyword{internal}
