% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_list_metric_attributions}
\alias{personalize_list_metric_attributions}
\title{Lists metric attributions}
\usage{
personalize_list_metric_attributions(
  datasetGroupArn = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{datasetGroupArn}{The metric attributions' dataset group Amazon Resource Name (ARN).}

\item{nextToken}{Specify the pagination token from a previous request to retrieve the
next page of results.}

\item{maxResults}{The maximum number of metric attributions to return in one page of
results.}
}
\description{
Lists metric attributions.

See \url{https://paws-r.github.io/docs/personalize/list_metric_attributions.html} for full documentation.
}
\keyword{internal}
