% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_list_solution_versions}
\alias{personalize_list_solution_versions}
\title{Returns a list of solution versions for the given solution}
\usage{
personalize_list_solution_versions(
  solutionArn = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{solutionArn}{The Amazon Resource Name (ARN) of the solution.}

\item{nextToken}{A token returned from the previous call to
\code{\link[=personalize_list_solution_versions]{list_solution_versions}} for
getting the next set of solution versions (if they exist).}

\item{maxResults}{The maximum number of solution versions to return.}
}
\description{
Returns a list of solution versions for the given solution. When a solution is not specified, all the solution versions associated with the account are listed. The response provides the properties for each solution version, including the Amazon Resource Name (ARN).

See \url{https://paws-r.github.io/docs/personalize/list_solution_versions.html} for full documentation.
}
\keyword{internal}
