% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_list_tags_for_resource}
\alias{personalize_list_tags_for_resource}
\title{Get a list of tags attached to a resource}
\usage{
personalize_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The resource's Amazon Resource Name.}
}
\description{
Get a list of tags attached to a resource.

See \url{https://paws-r.github.io/docs/personalize/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
