% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_tag_resource}
\alias{personalize_tag_resource}
\title{Add a list of tags to a resource}
\usage{
personalize_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The resource's Amazon Resource Name (ARN).}

\item{tags}{[required] Tags to apply to the resource. For more information see Tagging
Personalize resources.}
}
\description{
Add a list of tags to a resource.

See \url{https://paws-r.github.io/docs/personalize/tag_resource.html} for full documentation.
}
\keyword{internal}
