% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_untag_resource}
\alias{personalize_untag_resource}
\title{Remove tags that are attached to a resource}
\usage{
personalize_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The resource's Amazon Resource Name (ARN).}

\item{tagKeys}{[required] Keys to remove from the resource's tags.}
}
\description{
Remove tags that are attached to a resource.

See \url{https://paws-r.github.io/docs/personalize/untag_resource.html} for full documentation.
}
\keyword{internal}
