% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_update_dataset}
\alias{personalize_update_dataset}
\title{Update a dataset to replace its schema with a new or existing one}
\usage{
personalize_update_dataset(datasetArn, schemaArn)
}
\arguments{
\item{datasetArn}{[required] The Amazon Resource Name (ARN) of the dataset that you want to update.}

\item{schemaArn}{[required] The Amazon Resource Name (ARN) of the new schema you want use.}
}
\description{
Update a dataset to replace its schema with a new or existing one. For more information, see \href{https://docs.aws.amazon.com/personalize/latest/dg/updating-dataset-schema.html}{Replacing a dataset's schema}.

See \url{https://www.paws-r-sdk.com/docs/personalize_update_dataset/} for full documentation.
}
\keyword{internal}
