% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_update_metric_attribution}
\alias{personalize_update_metric_attribution}
\title{Updates a metric attribution}
\usage{
personalize_update_metric_attribution(
  addMetrics = NULL,
  removeMetrics = NULL,
  metricsOutputConfig = NULL,
  metricAttributionArn = NULL
)
}
\arguments{
\item{addMetrics}{Add new metric attributes to the metric attribution.}

\item{removeMetrics}{Remove metric attributes from the metric attribution.}

\item{metricsOutputConfig}{An output config for the metric attribution.}

\item{metricAttributionArn}{The Amazon Resource Name (ARN) for the metric attribution to update.}
}
\description{
Updates a metric attribution.

See \url{https://www.paws-r-sdk.com/docs/personalize_update_metric_attribution/} for full documentation.
}
\keyword{internal}
