% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_update_recommender}
\alias{personalize_update_recommender}
\title{Updates the recommender to modify the recommender configuration}
\usage{
personalize_update_recommender(recommenderArn, recommenderConfig)
}
\arguments{
\item{recommenderArn}{[required] The Amazon Resource Name (ARN) of the recommender to modify.}

\item{recommenderConfig}{[required] The configuration details of the recommender.}
}
\description{
Updates the recommender to modify the recommender configuration.

See \url{https://paws-r.github.io/docs/personalize/update_recommender.html} for full documentation.
}
\keyword{internal}
