% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalizeevents_operations.R
\name{personalizeevents_put_action_interactions}
\alias{personalizeevents_put_action_interactions}
\title{Records action interaction event data}
\usage{
personalizeevents_put_action_interactions(trackingId, actionInteractions)
}
\arguments{
\item{trackingId}{[required] The ID of your action interaction event tracker. When you create an
Action interactions dataset, Amazon Personalize creates an action
interaction event tracker for you. For more information, see \href{https://docs.aws.amazon.com/personalize/latest/dg/action-interaction-tracker-id.html}{Action interaction event tracker ID}.}

\item{actionInteractions}{[required] A list of action interaction events from the session.}
}
\description{
Records action interaction event data. An \emph{action interaction} event is an interaction between a user and an \emph{action}. For example, a user taking an action, such a enrolling in a membership program or downloading your app.

See \url{https://www.paws-r-sdk.com/docs/personalizeevents_put_action_interactions/} for full documentation.
}
\keyword{internal}
