% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalizeevents_operations.R
\name{personalizeevents_put_actions}
\alias{personalizeevents_put_actions}
\title{Adds one or more actions to an Actions dataset}
\usage{
personalizeevents_put_actions(datasetArn, actions)
}
\arguments{
\item{datasetArn}{[required] The Amazon Resource Name (ARN) of the Actions dataset you are adding the
action or actions to.}

\item{actions}{[required] A list of action data.}
}
\description{
Adds one or more actions to an Actions dataset. For more information see \href{https://docs.aws.amazon.com/personalize/latest/dg/importing-actions.html}{Importing actions individually}.

See \url{https://www.paws-r-sdk.com/docs/personalizeevents_put_actions/} for full documentation.
}
\keyword{internal}
