% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalizeevents_operations.R
\name{personalizeevents_put_items}
\alias{personalizeevents_put_items}
\title{Adds one or more items to an Items dataset}
\usage{
personalizeevents_put_items(datasetArn, items)
}
\arguments{
\item{datasetArn}{[required] The Amazon Resource Name (ARN) of the Items dataset you are adding the
item or items to.}

\item{items}{[required] A list of item data.}
}
\description{
Adds one or more items to an Items dataset. For more information see \href{https://docs.aws.amazon.com/personalize/latest/dg/importing-items.html}{Importing Items Incrementally}.

See \url{https://www.paws-r-sdk.com/docs/personalizeevents_put_items/} for full documentation.
}
\keyword{internal}
