% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalizeevents_operations.R
\name{personalizeevents_put_users}
\alias{personalizeevents_put_users}
\title{Adds one or more users to a Users dataset}
\usage{
personalizeevents_put_users(datasetArn, users)
}
\arguments{
\item{datasetArn}{[required] The Amazon Resource Name (ARN) of the Users dataset you are adding the
user or users to.}

\item{users}{[required] A list of user data.}
}
\value{
An empty list.
}
\description{
Adds one or more users to a Users dataset. For more information see
\href{https://docs.aws.amazon.com/personalize/latest/dg/importing-users.html}{Importing Users Incrementally}.
}
\section{Request syntax}{
\preformatted{svc$put_users(
  datasetArn = "string",
  users = list(
    list(
      userId = "string",
      properties = "string"
    )
  )
)
}
}

\keyword{internal}
