% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polly_operations.R
\name{polly_delete_lexicon}
\alias{polly_delete_lexicon}
\title{Deletes the specified pronunciation lexicon stored in an Amazon Web
Services Region}
\usage{
polly_delete_lexicon(Name)
}
\arguments{
\item{Name}{[required] The name of the lexicon to delete. Must be an existing lexicon in the
region.}
}
\description{
Deletes the specified pronunciation lexicon stored in an Amazon Web Services Region. A lexicon which has been deleted is not available for speech synthesis, nor is it possible to retrieve it using either the \code{\link[=polly_get_lexicon]{get_lexicon}} or \code{ListLexicon} APIs.

See \url{https://www.paws-r-sdk.com/docs/polly_delete_lexicon/} for full documentation.
}
\keyword{internal}
