% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polly_operations.R
\name{polly_get_speech_synthesis_task}
\alias{polly_get_speech_synthesis_task}
\title{Retrieves a specific SpeechSynthesisTask object based on its TaskID}
\usage{
polly_get_speech_synthesis_task(TaskId)
}
\arguments{
\item{TaskId}{[required] The Amazon Polly generated identifier for a speech synthesis task.}
}
\description{
Retrieves a specific SpeechSynthesisTask object based on its TaskID. This object contains information about the given speech synthesis task, including the status of the task, and a link to the S3 bucket containing the output of the task.

See \url{https://www.paws-r-sdk.com/docs/polly_get_speech_synthesis_task/} for full documentation.
}
\keyword{internal}
