% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polly_operations.R
\name{polly_list_speech_synthesis_tasks}
\alias{polly_list_speech_synthesis_tasks}
\title{Returns a list of SpeechSynthesisTask objects ordered by their creation
date}
\usage{
polly_list_speech_synthesis_tasks(
  MaxResults = NULL,
  NextToken = NULL,
  Status = NULL
)
}
\arguments{
\item{MaxResults}{Maximum number of speech synthesis tasks returned in a List operation.}

\item{NextToken}{The pagination token to use in the next request to continue the listing
of speech synthesis tasks.}

\item{Status}{Status of the speech synthesis tasks returned in a List operation}
}
\description{
Returns a list of SpeechSynthesisTask objects ordered by their creation date. This operation can filter the tasks by their status, for example, allowing users to list only tasks that are completed.

See \url{https://www.paws-r-sdk.com/docs/polly_list_speech_synthesis_tasks/} for full documentation.
}
\keyword{internal}
