% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polly_operations.R
\name{polly_start_speech_synthesis_stream}
\alias{polly_start_speech_synthesis_stream}
\title{Synthesizes UTF-8 input, plain text, or SSML over a bidirectional
streaming connection}
\usage{
polly_start_speech_synthesis_stream(
  Engine,
  LanguageCode = NULL,
  LexiconNames = NULL,
  OutputFormat,
  SampleRate = NULL,
  VoiceId,
  ActionStream = NULL
)
}
\arguments{
\item{Engine}{[required] Specifies the engine for Amazon Polly to use when processing input text
for speech synthesis. Currently, only the \code{generative} engine is
supported. If you specify a voice that the selected engine doesn't
support, Amazon Polly returns an error.}

\item{LanguageCode}{An optional parameter that sets the language code for the speech
synthesis request. Specify this parameter only when using a bilingual
voice. If a bilingual voice is used and no language code is specified,
Amazon Polly uses the default language of the bilingual voice.}

\item{LexiconNames}{The names of one or more pronunciation lexicons for the service to apply
during synthesis. Amazon Polly applies lexicons only when the lexicon
language matches the voice language.}

\item{OutputFormat}{[required] The audio format for the synthesized speech. Currently, Amazon Polly
does not support JSON speech marks.}

\item{SampleRate}{The audio frequency, specified in Hz.}

\item{VoiceId}{[required] The voice to use in synthesis. To get a list of available voice IDs, use
the \code{\link[=polly_describe_voices]{describe_voices}} operation.}

\item{ActionStream}{The input event stream that contains text events and stream control
events.}
}
\description{
Synthesizes UTF-8 input, plain text, or SSML over a bidirectional streaming connection. Specify synthesis parameters in HTTP/2 headers, send text incrementally as events on the input stream, and receive synthesized audio as it becomes available.

See \url{https://www.paws-r-sdk.com/docs/polly_start_speech_synthesis_stream/} for full documentation.
}
\keyword{internal}
