% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_service.R
\name{rekognition}
\alias{rekognition}
\title{Amazon Rekognition}
\usage{
rekognition(
  config = list(),
  credentials = list(),
  endpoint = NULL,
  region = NULL
)
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.
\itemize{
\item{\strong{credentials}: \itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}
\item{\strong{endpoint}: The complete URL to use for the constructed client.}
\item{\strong{region}: The AWS Region used in instantiating the client.}
\item{\strong{close_connection}: Immediately close all HTTP connections.}
\item{\strong{timeout}: The time in seconds till a timeout exception is thrown when attempting to make a connection. The default is 60 seconds.}
\item{\strong{s3_force_path_style}: Set this to \code{true} to force the request to use path-style addressing, i.e. \verb{http://s3.amazonaws.com/BUCKET/KEY}.}
\item{\strong{sts_regional_endpoint}: Set sts regional endpoint resolver to regional or legacy \url{https://docs.aws.amazon.com/sdkref/latest/guide/feature-sts-regionalized-endpoints.html}}
\item{\strong{use_dual_stack}: Set this to \code{true} to use the dualstack (IPv4 and IPv6) endpoint for a service, where available, falling back to the regular endpoint when it isn't. Defaults to the \code{AWS_USE_DUALSTACK_ENDPOINT} environment variable when unset.}
}}

\item{credentials}{Optional credentials shorthand for the config parameter
\itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}

\item{endpoint}{Optional shorthand for complete URL to use for the constructed client.}

\item{region}{Optional shorthand for AWS Region used in instantiating the client.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
This is the API Reference for \href{https://docs.aws.amazon.com/rekognition/latest/dg/images.html}{Amazon Rekognition Image}, \href{https://docs.aws.amazon.com/rekognition/latest/dg/bulk-analysis.html}{Amazon Rekognition Bulk Image Analysis}, \href{https://docs.aws.amazon.com/rekognition/latest/customlabels-dg/what-is.html}{Amazon Rekognition Custom Labels}, \href{https://docs.aws.amazon.com/rekognition/latest/dg/video.html}{Amazon Rekognition Stored Video}, \href{https://docs.aws.amazon.com/rekognition/latest/dg/face-liveness.html}{Amazon Rekognition Face Liveness}, \href{https://docs.aws.amazon.com/rekognition/latest/dg/streaming-video.html}{Amazon Rekognition Streaming Video}. It provides descriptions of actions, data types, common parameters, and common errors.

\strong{Amazon Rekognition Image}
\itemize{
\item \code{\link[=rekognition_associate_faces]{associate_faces}}
\item \code{\link[=rekognition_compare_faces]{compare_faces}}
\item \code{\link[=rekognition_create_collection]{create_collection}}
\item \code{\link[=rekognition_create_user]{create_user}}
\item \code{\link[=rekognition_delete_collection]{delete_collection}}
\item \code{\link[=rekognition_delete_faces]{delete_faces}}
\item \code{\link[=rekognition_delete_user]{delete_user}}
\item \code{\link[=rekognition_describe_collection]{describe_collection}}
\item \code{\link[=rekognition_detect_faces]{detect_faces}}
\item \code{\link[=rekognition_detect_labels]{detect_labels}}
\item \code{\link[=rekognition_detect_moderation_labels]{detect_moderation_labels}}
\item \code{\link[=rekognition_detect_protective_equipment]{detect_protective_equipment}}
\item \code{\link[=rekognition_detect_text]{detect_text}}
\item \code{\link[=rekognition_disassociate_faces]{disassociate_faces}}
\item \code{\link[=rekognition_get_celebrity_info]{get_celebrity_info}}
\item \code{\link[=rekognition_index_faces]{index_faces}}
\item \code{\link[=rekognition_list_collections]{list_collections}}
\item \code{\link[=rekognition_list_faces]{list_faces}}
\item \code{\link[=rekognition_list_users]{list_users}}
\item \code{\link[=rekognition_recognize_celebrities]{recognize_celebrities}}
\item \code{\link[=rekognition_search_faces]{search_faces}}
\item \code{\link[=rekognition_search_faces_by_image]{search_faces_by_image}}
\item \code{\link[=rekognition_search_users]{search_users}}
\item \code{\link[=rekognition_search_users_by_image]{search_users_by_image}}
}

\strong{Amazon Rekognition Bulk Image Analysis}
\itemize{
\item \code{\link[=rekognition_get_media_analysis_job]{get_media_analysis_job}}
\item \href{https://docs.aws.amazon.com/rekognition/latest/APIReference/}{ListMediaAnalysisJob}
\item \code{\link[=rekognition_start_media_analysis_job]{start_media_analysis_job}}
}

\strong{Amazon Rekognition Custom Labels}
\itemize{
\item \code{\link[=rekognition_copy_project_version]{copy_project_version}}
\item \code{\link[=rekognition_create_dataset]{create_dataset}}
\item \code{\link[=rekognition_create_project]{create_project}}
\item \code{\link[=rekognition_create_project_version]{create_project_version}}
\item \code{\link[=rekognition_delete_dataset]{delete_dataset}}
\item \code{\link[=rekognition_delete_project]{delete_project}}
\item \code{\link[=rekognition_delete_project_policy]{delete_project_policy}}
\item \code{\link[=rekognition_delete_project_version]{delete_project_version}}
\item \code{\link[=rekognition_describe_dataset]{describe_dataset}}
\item \code{\link[=rekognition_describe_projects]{describe_projects}}
\item \code{\link[=rekognition_describe_project_versions]{describe_project_versions}}
\item \code{\link[=rekognition_detect_custom_labels]{detect_custom_labels}}
\item \code{\link[=rekognition_distribute_dataset_entries]{distribute_dataset_entries}}
\item \code{\link[=rekognition_list_dataset_entries]{list_dataset_entries}}
\item \code{\link[=rekognition_list_dataset_labels]{list_dataset_labels}}
\item \code{\link[=rekognition_list_project_policies]{list_project_policies}}
\item \code{\link[=rekognition_put_project_policy]{put_project_policy}}
\item \code{\link[=rekognition_start_project_version]{start_project_version}}
\item \code{\link[=rekognition_stop_project_version]{stop_project_version}}
\item \code{\link[=rekognition_update_dataset_entries]{update_dataset_entries}}
}

\strong{Amazon Rekognition Stored Video}
\itemize{
\item \code{\link[=rekognition_get_celebrity_recognition]{get_celebrity_recognition}}
\item \code{\link[=rekognition_get_content_moderation]{get_content_moderation}}
\item \code{\link[=rekognition_get_face_detection]{get_face_detection}}
\item \code{\link[=rekognition_get_face_search]{get_face_search}}
\item \code{\link[=rekognition_get_label_detection]{get_label_detection}}
\item \code{\link[=rekognition_get_person_tracking]{get_person_tracking}}
\item \code{\link[=rekognition_get_segment_detection]{get_segment_detection}}
\item \code{\link[=rekognition_get_text_detection]{get_text_detection}}
\item \code{\link[=rekognition_start_celebrity_recognition]{start_celebrity_recognition}}
\item \code{\link[=rekognition_start_content_moderation]{start_content_moderation}}
\item \code{\link[=rekognition_start_face_detection]{start_face_detection}}
\item \code{\link[=rekognition_start_face_search]{start_face_search}}
\item \code{\link[=rekognition_start_label_detection]{start_label_detection}}
\item \code{\link[=rekognition_start_person_tracking]{start_person_tracking}}
\item \code{\link[=rekognition_start_segment_detection]{start_segment_detection}}
\item \code{\link[=rekognition_start_text_detection]{start_text_detection}}
}

\strong{Amazon Rekognition Face Liveness}
\itemize{
\item \code{\link[=rekognition_create_face_liveness_session]{create_face_liveness_session}}
\item \code{\link[=rekognition_get_face_liveness_session_results]{get_face_liveness_session_results}}
\item \href{https://docs.aws.amazon.com/rekognition/latest/APIReference/API_rekognitionstreaming_StartFaceLivenessSession.html}{StartFaceLivenessSession}
}

\strong{Amazon Rekognition Streaming Video}
\itemize{
\item \code{\link[=rekognition_create_stream_processor]{create_stream_processor}}
\item \code{\link[=rekognition_delete_stream_processor]{delete_stream_processor}}
\item \code{\link[=rekognition_describe_stream_processor]{describe_stream_processor}}
\item \code{\link[=rekognition_list_stream_processors]{list_stream_processors}}
\item \code{\link[=rekognition_start_stream_processor]{start_stream_processor}}
\item \code{\link[=rekognition_stop_stream_processor]{stop_stream_processor}}
\item \code{\link[=rekognition_update_stream_processor]{update_stream_processor}}
}
}
\section{Service syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc <- rekognition(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string",
      anonymous = "logical"
    ),
    endpoint = "string",
    region = "string",
    close_connection = "logical",
    timeout = "numeric",
    s3_force_path_style = "logical",
    sts_regional_endpoint = "string",
    use_dual_stack = "logical"
  ),
  credentials = list(
    creds = list(
      access_key_id = "string",
      secret_access_key = "string",
      session_token = "string"
    ),
    profile = "string",
    anonymous = "logical"
  ),
  endpoint = "string",
  region = "string"
)
}\if{html}{\out{</div>}}
}

\section{Operations}{

\tabular{ll}{
\link[=rekognition_associate_faces]{associate_faces} \tab Associates one or more faces with an existing UserID\cr
\link[=rekognition_compare_faces]{compare_faces} \tab Compares a face in the source input image with each of the 100 largest faces detected in the target input image\cr
\link[=rekognition_copy_project_version]{copy_project_version} \tab This operation applies only to Amazon Rekognition Custom Labels\cr
\link[=rekognition_create_collection]{create_collection} \tab Creates a collection in an AWS Region\cr
\link[=rekognition_create_dataset]{create_dataset} \tab This operation applies only to Amazon Rekognition Custom Labels\cr
\link[=rekognition_create_face_liveness_session]{create_face_liveness_session} \tab This API operation initiates a Face Liveness session\cr
\link[=rekognition_create_project]{create_project} \tab Creates a new Amazon Rekognition project\cr
\link[=rekognition_create_project_version]{create_project_version} \tab Creates a new version of Amazon Rekognition project (like a Custom Labels model or a custom adapter) and begins training\cr
\link[=rekognition_create_stream_processor]{create_stream_processor} \tab Service availability notice: Streaming Video and Bulk Image Analysis is no longer available to new customers\cr
\link[=rekognition_create_user]{create_user} \tab Creates a new User within a collection specified by CollectionId\cr
\link[=rekognition_delete_collection]{delete_collection} \tab Deletes the specified collection\cr
\link[=rekognition_delete_dataset]{delete_dataset} \tab This operation applies only to Amazon Rekognition Custom Labels\cr
\link[=rekognition_delete_faces]{delete_faces} \tab Deletes faces from a collection\cr
\link[=rekognition_delete_project]{delete_project} \tab Deletes a Amazon Rekognition project\cr
\link[=rekognition_delete_project_policy]{delete_project_policy} \tab This operation applies only to Amazon Rekognition Custom Labels\cr
\link[=rekognition_delete_project_version]{delete_project_version} \tab Deletes a Rekognition project model or project version, like a Amazon Rekognition Custom Labels model or a custom adapter\cr
\link[=rekognition_delete_stream_processor]{delete_stream_processor} \tab Service availability notice: Streaming Video and Bulk Image Analysis is no longer available to new customers\cr
\link[=rekognition_delete_user]{delete_user} \tab Deletes the specified UserID within the collection\cr
\link[=rekognition_describe_collection]{describe_collection} \tab Describes the specified collection\cr
\link[=rekognition_describe_dataset]{describe_dataset} \tab This operation applies only to Amazon Rekognition Custom Labels\cr
\link[=rekognition_describe_projects]{describe_projects} \tab Gets information about your Rekognition projects\cr
\link[=rekognition_describe_project_versions]{describe_project_versions} \tab Lists and describes the versions of an Amazon Rekognition project\cr
\link[=rekognition_describe_stream_processor]{describe_stream_processor} \tab Service availability notice: Streaming Video and Bulk Image Analysis is no longer available to new customers\cr
\link[=rekognition_detect_custom_labels]{detect_custom_labels} \tab This operation applies only to Amazon Rekognition Custom Labels\cr
\link[=rekognition_detect_faces]{detect_faces} \tab Detects faces within an image that is provided as input\cr
\link[=rekognition_detect_labels]{detect_labels} \tab Detects instances of real-world entities within an image (JPEG or PNG) provided as input\cr
\link[=rekognition_detect_moderation_labels]{detect_moderation_labels} \tab Detects unsafe content in a specified JPEG or PNG format image\cr
\link[=rekognition_detect_protective_equipment]{detect_protective_equipment} \tab Detects Personal Protective Equipment (PPE) worn by people detected in an image\cr
\link[=rekognition_detect_text]{detect_text} \tab Detects text in the input image and converts it into machine-readable text\cr
\link[=rekognition_disassociate_faces]{disassociate_faces} \tab Removes the association between a Face supplied in an array of FaceIds and the User\cr
\link[=rekognition_distribute_dataset_entries]{distribute_dataset_entries} \tab This operation applies only to Amazon Rekognition Custom Labels\cr
\link[=rekognition_get_celebrity_info]{get_celebrity_info} \tab Gets the name and additional information about a celebrity based on their Amazon Rekognition ID\cr
\link[=rekognition_get_celebrity_recognition]{get_celebrity_recognition} \tab Gets the celebrity recognition results for a Amazon Rekognition Video analysis started by StartCelebrityRecognition\cr
\link[=rekognition_get_content_moderation]{get_content_moderation} \tab Gets the inappropriate, unwanted, or offensive content analysis results for a Amazon Rekognition Video analysis started by StartContentModeration\cr
\link[=rekognition_get_face_detection]{get_face_detection} \tab Gets face detection results for a Amazon Rekognition Video analysis started by StartFaceDetection\cr
\link[=rekognition_get_face_liveness_session_results]{get_face_liveness_session_results} \tab Retrieves the results of a specific Face Liveness session\cr
\link[=rekognition_get_face_search]{get_face_search} \tab Gets the face search results for Amazon Rekognition Video face search started by StartFaceSearch\cr
\link[=rekognition_get_label_detection]{get_label_detection} \tab Gets the label detection results of a Amazon Rekognition Video analysis started by StartLabelDetection\cr
\link[=rekognition_get_media_analysis_job]{get_media_analysis_job} \tab Service availability notice: Streaming Video and Bulk Image Analysis is no longer available to new customers\cr
\link[=rekognition_get_person_tracking]{get_person_tracking} \tab End of support notice: On October 31, 2025, AWS will discontinue support for Amazon Rekognition People Pathing\cr
\link[=rekognition_get_segment_detection]{get_segment_detection} \tab Gets the segment detection results of a Amazon Rekognition Video analysis started by StartSegmentDetection\cr
\link[=rekognition_get_text_detection]{get_text_detection} \tab Gets the text detection results of a Amazon Rekognition Video analysis started by StartTextDetection\cr
\link[=rekognition_index_faces]{index_faces} \tab Detects faces in the input image and adds them to the specified collection\cr
\link[=rekognition_list_collections]{list_collections} \tab Returns list of collection IDs in your account\cr
\link[=rekognition_list_dataset_entries]{list_dataset_entries} \tab This operation applies only to Amazon Rekognition Custom Labels\cr
\link[=rekognition_list_dataset_labels]{list_dataset_labels} \tab This operation applies only to Amazon Rekognition Custom Labels\cr
\link[=rekognition_list_faces]{list_faces} \tab Returns metadata for faces in the specified collection\cr
\link[=rekognition_list_media_analysis_jobs]{list_media_analysis_jobs} \tab Service availability notice: Streaming Video and Bulk Image Analysis is no longer available to new customers\cr
\link[=rekognition_list_project_policies]{list_project_policies} \tab This operation applies only to Amazon Rekognition Custom Labels\cr
\link[=rekognition_list_stream_processors]{list_stream_processors} \tab Service availability notice: Streaming Video and Bulk Image Analysis is no longer available to new customers\cr
\link[=rekognition_list_tags_for_resource]{list_tags_for_resource} \tab Returns a list of tags in an Amazon Rekognition collection, stream processor, or Custom Labels model\cr
\link[=rekognition_list_users]{list_users} \tab Returns metadata of the User such as UserID in the specified collection\cr
\link[=rekognition_put_project_policy]{put_project_policy} \tab This operation applies only to Amazon Rekognition Custom Labels\cr
\link[=rekognition_recognize_celebrities]{recognize_celebrities} \tab Returns an array of celebrities recognized in the input image\cr
\link[=rekognition_search_faces]{search_faces} \tab For a given input face ID, searches for matching faces in the collection the face belongs to\cr
\link[=rekognition_search_faces_by_image]{search_faces_by_image} \tab For a given input image, first detects the largest face in the image, and then searches the specified collection for matching faces\cr
\link[=rekognition_search_users]{search_users} \tab Searches for UserIDs within a collection based on a FaceId or UserId\cr
\link[=rekognition_search_users_by_image]{search_users_by_image} \tab Searches for UserIDs using a supplied image\cr
\link[=rekognition_start_celebrity_recognition]{start_celebrity_recognition} \tab Starts asynchronous recognition of celebrities in a stored video\cr
\link[=rekognition_start_content_moderation]{start_content_moderation} \tab Starts asynchronous detection of inappropriate, unwanted, or offensive content in a stored video\cr
\link[=rekognition_start_face_detection]{start_face_detection} \tab Starts asynchronous detection of faces in a stored video\cr
\link[=rekognition_start_face_search]{start_face_search} \tab Starts the asynchronous search for faces in a collection that match the faces of persons detected in a stored video\cr
\link[=rekognition_start_label_detection]{start_label_detection} \tab Starts asynchronous detection of labels in a stored video\cr
\link[=rekognition_start_media_analysis_job]{start_media_analysis_job} \tab Service availability notice: Streaming Video and Bulk Image Analysis is no longer available to new customers\cr
\link[=rekognition_start_person_tracking]{start_person_tracking} \tab End of support notice: On October 31, 2025, AWS will discontinue support for Amazon Rekognition People Pathing\cr
\link[=rekognition_start_project_version]{start_project_version} \tab This operation applies only to Amazon Rekognition Custom Labels\cr
\link[=rekognition_start_segment_detection]{start_segment_detection} \tab Starts asynchronous detection of segment detection in a stored video\cr
\link[=rekognition_start_stream_processor]{start_stream_processor} \tab Service availability notice: Streaming Video and Bulk Image Analysis is no longer available to new customers\cr
\link[=rekognition_start_text_detection]{start_text_detection} \tab Starts asynchronous detection of text in a stored video\cr
\link[=rekognition_stop_project_version]{stop_project_version} \tab This operation applies only to Amazon Rekognition Custom Labels\cr
\link[=rekognition_stop_stream_processor]{stop_stream_processor} \tab Service availability notice: Streaming Video and Bulk Image Analysis is no longer available to new customers\cr
\link[=rekognition_tag_resource]{tag_resource} \tab Adds one or more key-value tags to an Amazon Rekognition collection, stream processor, or Custom Labels model\cr
\link[=rekognition_untag_resource]{untag_resource} \tab Removes one or more tags from an Amazon Rekognition collection, stream processor, or Custom Labels model\cr
\link[=rekognition_update_dataset_entries]{update_dataset_entries} \tab This operation applies only to Amazon Rekognition Custom Labels\cr
\link[=rekognition_update_stream_processor]{update_stream_processor} \tab Service availability notice: Streaming Video and Bulk Image Analysis is no longer available to new customers
}
}

\examples{
\dontrun{
svc <- rekognition()
# This operation compares the largest face detected in the source image
# with each face detected in the target image.
svc$compare_faces(
  SimilarityThreshold = 90L,
  SourceImage = list(
    S3Object = list(
      Bucket = "mybucket",
      Name = "mysourceimage"
    )
  ),
  TargetImage = list(
    S3Object = list(
      Bucket = "mybucket",
      Name = "mytargetimage"
    )
  )
)
}

}
