% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_create_collection}
\alias{rekognition_create_collection}
\title{Creates a collection in an AWS Region}
\usage{
rekognition_create_collection(CollectionId, Tags = NULL)
}
\arguments{
\item{CollectionId}{[required] ID for the collection that you are creating.}

\item{Tags}{A set of tags (key-value pairs) that you want to attach to the
collection.}
}
\description{
Creates a collection in an AWS Region. You can add faces to the collection using the \code{\link[=rekognition_index_faces]{index_faces}} operation.

See \url{https://www.paws-r-sdk.com/docs/rekognition_create_collection/} for full documentation.
}
\keyword{internal}
