% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_create_project}
\alias{rekognition_create_project}
\title{Creates a new Amazon Rekognition Custom Labels project}
\usage{
rekognition_create_project(ProjectName)
}
\arguments{
\item{ProjectName}{[required] The name of the project to create.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ProjectArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Creates a new Amazon Rekognition Custom Labels project. A project is a
logical grouping of resources (images, Labels, models) and operations
(training, evaluation and detection).

This operation requires permissions to perform the
\code{rekognition:CreateProject} action.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_project(
  ProjectName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
