% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_delete_collection}
\alias{rekognition_delete_collection}
\title{Deletes the specified collection}
\usage{
rekognition_delete_collection(CollectionId)
}
\arguments{
\item{CollectionId}{[required] ID of the collection to delete.}
}
\description{
Deletes the specified collection. Note that this operation removes all faces in the collection. For an example, see \href{https://docs.aws.amazon.com/rekognition/latest/dg/delete-collection-procedure.html}{Deleting a collection}.

See \url{https://www.paws-r-sdk.com/docs/rekognition_delete_collection/} for full documentation.
}
\keyword{internal}
