% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_delete_dataset}
\alias{rekognition_delete_dataset}
\title{Deletes an existing Amazon Rekognition Custom Labels dataset}
\usage{
rekognition_delete_dataset(DatasetArn)
}
\arguments{
\item{DatasetArn}{[required] The ARN of the Amazon Rekognition Custom Labels dataset that you want to
delete.}
}
\description{
Deletes an existing Amazon Rekognition Custom Labels dataset. Deleting a dataset might take while. Use \code{\link[=rekognition_describe_dataset]{describe_dataset}} to check the current status. The dataset is still deleting if the value of \code{Status} is \code{DELETE_IN_PROGRESS}. If you try to access the dataset after it is deleted, you get a \code{ResourceNotFoundException} exception.

See \url{https://paws-r.github.io/docs/rekognition/delete_dataset.html} for full documentation.
}
\keyword{internal}
