% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_describe_dataset}
\alias{rekognition_describe_dataset}
\title{This operation applies only to Amazon Rekognition Custom Labels}
\usage{
rekognition_describe_dataset(DatasetArn)
}
\arguments{
\item{DatasetArn}{[required] The Amazon Resource Name (ARN) of the dataset that you want to describe.}
}
\description{
This operation applies only to Amazon Rekognition Custom Labels.

See \url{https://www.paws-r-sdk.com/docs/rekognition_describe_dataset/} for full documentation.
}
\keyword{internal}
