% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_describe_projects}
\alias{rekognition_describe_projects}
\title{Lists and gets information about your Amazon Rekognition Custom Labels
projects}
\usage{
rekognition_describe_projects(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{If the previous response was incomplete (because there is more results
to retrieve), Amazon Rekognition Custom Labels returns a pagination
token in the response. You can use this pagination token to retrieve the
next set of results.}

\item{MaxResults}{The maximum number of results to return per paginated call. The largest
value you can specify is 100. If you specify a value greater than 100, a
ValidationException error occurs. The default value is 100.}
}
\description{
Lists and gets information about your Amazon Rekognition Custom Labels
projects.

This operation requires permissions to perform the
\code{rekognition:DescribeProjects} action.
}
\section{Request syntax}{
\preformatted{svc$describe_projects(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
