% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_describe_stream_processor}
\alias{rekognition_describe_stream_processor}
\title{Service availability notice: Streaming Video and Bulk Image Analysis is
no longer available to new customers}
\usage{
rekognition_describe_stream_processor(Name)
}
\arguments{
\item{Name}{[required] Name of the stream processor for which you want information.}
}
\description{
Service availability notice: Streaming Video and Bulk Image Analysis is no longer available to new customers. For more information, see \href{https://docs.aws.amazon.com/rekognition/latest/dg/rekognition-availability-changes.html}{Rekognition feature availability changes}.

See \url{https://www.paws-r-sdk.com/docs/rekognition_describe_stream_processor/} for full documentation.
}
\keyword{internal}
