% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_detect_protective_equipment}
\alias{rekognition_detect_protective_equipment}
\title{Detects Personal Protective Equipment (PPE) worn by people detected in
an image}
\usage{
rekognition_detect_protective_equipment(Image, SummarizationAttributes = NULL)
}
\arguments{
\item{Image}{[required] The image in which you want to detect PPE on detected persons. The image
can be passed as image bytes or you can reference an image stored in an
Amazon S3 bucket.}

\item{SummarizationAttributes}{An array of PPE types that you want to summarize.}
}
\description{
Detects Personal Protective Equipment (PPE) worn by people detected in an image. Amazon Rekognition can detect the following types of PPE.

See \url{https://www.paws-r-sdk.com/docs/rekognition_detect_protective_equipment/} for full documentation.
}
\keyword{internal}
