% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_distribute_dataset_entries}
\alias{rekognition_distribute_dataset_entries}
\title{Distributes the entries (images) in a training dataset across the
training dataset and the test dataset for a project}
\usage{
rekognition_distribute_dataset_entries(Datasets)
}
\arguments{
\item{Datasets}{[required] The ARNS for the training dataset and test dataset that you want to use.
The datasets must belong to the same project. The test dataset must be
empty.}
}
\description{
Distributes the entries (images) in a training dataset across the training dataset and the test dataset for a project. \code{\link[=rekognition_distribute_dataset_entries]{distribute_dataset_entries}} moves 20\% of the training dataset images to the test dataset. An entry is a JSON Line that describes an image.

See \url{https://www.paws-r-sdk.com/docs/rekognition_distribute_dataset_entries/} for full documentation.
}
\keyword{internal}
