% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_get_media_analysis_job}
\alias{rekognition_get_media_analysis_job}
\title{Service availability notice: Streaming Video and Bulk Image Analysis is
no longer available to new customers}
\usage{
rekognition_get_media_analysis_job(JobId)
}
\arguments{
\item{JobId}{[required] Unique identifier for the media analysis job for which you want to retrieve results.}
}
\description{
Service availability notice: Streaming Video and Bulk Image Analysis is no longer available to new customers. For more information, see \href{https://docs.aws.amazon.com/rekognition/latest/dg/rekognition-availability-changes.html}{Rekognition feature availability changes}.

See \url{https://www.paws-r-sdk.com/docs/rekognition_get_media_analysis_job/} for full documentation.
}
\keyword{internal}
