% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_get_segment_detection}
\alias{rekognition_get_segment_detection}
\title{Gets the segment detection results of a Amazon Rekognition Video
analysis started by StartSegmentDetection}
\usage{
rekognition_get_segment_detection(JobId, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{JobId}{[required] Job identifier for the text detection operation for which you want
results returned. You get the job identifer from an initial call to
\code{\link[=rekognition_start_segment_detection]{start_segment_detection}}.}

\item{MaxResults}{Maximum number of results to return per paginated call. The largest
value you can specify is 1000.}

\item{NextToken}{If the response is truncated, Amazon Rekognition Video returns this
token that you can use in the subsequent request to retrieve the next
set of text.}
}
\description{
Gets the segment detection results of a Amazon Rekognition Video analysis started by \code{\link[=rekognition_start_segment_detection]{start_segment_detection}}.

See \url{https://paws-r.github.io/docs/rekognition/get_segment_detection.html} for full documentation.
}
\keyword{internal}
