% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_list_dataset_labels}
\alias{rekognition_list_dataset_labels}
\title{Lists the labels in a dataset}
\usage{
rekognition_list_dataset_labels(
  DatasetArn,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{DatasetArn}{[required] The Amazon Resource Name (ARN) of the dataset that you want to use.}

\item{NextToken}{If the previous response was incomplete (because there is more results
to retrieve), Amazon Rekognition Custom Labels returns a pagination
token in the response. You can use this pagination token to retrieve the
next set of results.}

\item{MaxResults}{The maximum number of results to return per paginated call. The largest
value you can specify is 100. If you specify a value greater than 100, a
ValidationException error occurs. The default value is 100.}
}
\description{
Lists the labels in a dataset. Amazon Rekognition Custom Labels uses labels to describe images. For more information, see \href{https://docs.aws.amazon.com/rekognition/latest/customlabels-dg/md-labeling-images.html}{Labeling images}.

See \url{https://paws-r.github.io/docs/rekognition/list_dataset_labels.html} for full documentation.
}
\keyword{internal}
