% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_list_project_policies}
\alias{rekognition_list_project_policies}
\title{Gets a list of the project policies attached to a project}
\usage{
rekognition_list_project_policies(
  ProjectArn,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ProjectArn}{[required] The ARN of the project for which you want to list the project policies.}

\item{NextToken}{If the previous response was incomplete (because there is more results
to retrieve), Amazon Rekognition Custom Labels returns a pagination
token in the response. You can use this pagination token to retrieve the
next set of results.}

\item{MaxResults}{The maximum number of results to return per paginated call. The largest
value you can specify is 5. If you specify a value greater than 5, a
ValidationException error occurs. The default value is 5.}
}
\description{
Gets a list of the project policies attached to a project.

See \url{https://www.paws-r-sdk.com/docs/rekognition_list_project_policies/} for full documentation.
}
\keyword{internal}
