% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_list_stream_processors}
\alias{rekognition_list_stream_processors}
\title{Service availability notice: Streaming Video and Bulk Image Analysis is
no longer available to new customers}
\usage{
rekognition_list_stream_processors(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{If the previous response was incomplete (because there are more stream processors to retrieve), Amazon Rekognition Video returns a pagination token in the response. You can use this pagination token to retrieve the next set of stream processors.}

\item{MaxResults}{Maximum number of stream processors you want Amazon Rekognition Video to return in the response. The default is 1000.}
}
\description{
Service availability notice: Streaming Video and Bulk Image Analysis is no longer available to new customers. For more information, see \href{https://docs.aws.amazon.com/rekognition/latest/dg/rekognition-availability-changes.html}{Rekognition feature availability changes}.

See \url{https://www.paws-r-sdk.com/docs/rekognition_list_stream_processors/} for full documentation.
}
\keyword{internal}
