% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_list_tags_for_resource}
\alias{rekognition_list_tags_for_resource}
\title{Returns a list of tags in an Amazon Rekognition collection, stream
processor, or Custom Labels model}
\usage{
rekognition_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] Amazon Resource Name (ARN) of the model, collection, or stream processor
that contains the tags that you want a list of.}
}
\value{
A list with the following syntax:\preformatted{list(
  Tags = list(
    "string"
  )
)
}
}
\description{
Returns a list of tags in an Amazon Rekognition collection, stream
processor, or Custom Labels model.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceArn = "string"
)
}
}

\keyword{internal}
