% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_list_users}
\alias{rekognition_list_users}
\title{Returns metadata of the User such as UserID in the specified collection}
\usage{
rekognition_list_users(CollectionId, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{CollectionId}{[required] The ID of an existing collection.}

\item{MaxResults}{Maximum number of UsersID to return.}

\item{NextToken}{Pagingation token to receive the next set of UsersID.}
}
\description{
Returns metadata of the User such as \code{UserID} in the specified collection. Anonymous User (to reserve faces without any identity) is not returned as part of this request. The results are sorted by system generated primary key ID. If the response is truncated, \code{NextToken} is returned in the response that can be used in the subsequent request to retrieve the next set of identities.

See \url{https://www.paws-r-sdk.com/docs/rekognition_list_users/} for full documentation.
}
\keyword{internal}
