% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_recognize_celebrities}
\alias{rekognition_recognize_celebrities}
\title{Returns an array of celebrities recognized in the input image}
\usage{
rekognition_recognize_celebrities(Image)
}
\arguments{
\item{Image}{[required] The input image as base64-encoded bytes or an S3 object. If you use the
AWS CLI to call Amazon Rekognition operations, passing base64-encoded
image bytes is not supported.

If you are using an AWS SDK to call Amazon Rekognition, you might not
need to base64-encode image bytes passed using the \code{Bytes} field. For
more information, see Images in the Amazon Rekognition developer guide.}
}
\description{
Returns an array of celebrities recognized in the input image. For more information, see Recognizing celebrities in the Amazon Rekognition Developer Guide.

See \url{https://www.paws-r-sdk.com/docs/rekognition_recognize_celebrities/} for full documentation.
}
\keyword{internal}
