% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_search_users_by_image}
\alias{rekognition_search_users_by_image}
\title{Searches for UserIDs using a supplied image}
\usage{
rekognition_search_users_by_image(
  CollectionId,
  Image,
  UserMatchThreshold = NULL,
  MaxUsers = NULL,
  QualityFilter = NULL
)
}
\arguments{
\item{CollectionId}{[required] The ID of an existing collection containing the UserID.}

\item{Image}{[required] Provides the input image either as bytes or an S3 object.

You pass image bytes to an Amazon Rekognition API operation by using the \code{Bytes} property. For example, you would use the \code{Bytes} property to pass an image loaded from a local file system. Image bytes passed by using the \code{Bytes} property must be base64-encoded. Your code may not need to encode image bytes if you are using an AWS SDK to call Amazon Rekognition API operations.

For more information, see Analyzing an Image Loaded from a Local File System in the Amazon Rekognition Developer Guide.

You pass images stored in an S3 bucket to an Amazon Rekognition API operation by using the \code{S3Object} property. Images stored in an S3 bucket do not need to be base64-encoded.

The region for the S3 bucket containing the S3 object must match the region you use for Amazon Rekognition operations.

If you use the AWS CLI to call Amazon Rekognition operations, passing image bytes using the Bytes property is not supported. You must first upload the image to an Amazon S3 bucket and then call the operation using the S3Object property.

For Amazon Rekognition to process an S3 object, the user must have permission to access the S3 object. For more information, see How Amazon Rekognition works with IAM in the Amazon Rekognition Developer Guide.}

\item{UserMatchThreshold}{Specifies the minimum confidence in the UserID match to return. Default value is 80.}

\item{MaxUsers}{Maximum number of UserIDs to return.}

\item{QualityFilter}{A filter that specifies a quality bar for how much filtering is done to identify faces. Filtered faces aren't searched for in the collection. The default value is NONE.}
}
\description{
Searches for UserIDs using a supplied image. It first detects the largest face in the image, and then searches a specified collection for matching UserIDs.

See \url{https://www.paws-r-sdk.com/docs/rekognition_search_users_by_image/} for full documentation.
}
\keyword{internal}
