% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_start_media_analysis_job}
\alias{rekognition_start_media_analysis_job}
\title{Service availability notice: Streaming Video and Bulk Image Analysis is
no longer available to new customers}
\usage{
rekognition_start_media_analysis_job(
  ClientRequestToken = NULL,
  JobName = NULL,
  OperationsConfig,
  Input,
  OutputConfig,
  KmsKeyId = NULL
)
}
\arguments{
\item{ClientRequestToken}{Idempotency token used to prevent the accidental creation of duplicate versions. If you use the same token with multiple \code{StartMediaAnalysisJobRequest} requests, the same response is returned. Use \code{ClientRequestToken} to prevent the same request from being processed more than once.}

\item{JobName}{The name of the job. Does not have to be unique.}

\item{OperationsConfig}{[required] Configuration options for the media analysis job to be created.}

\item{Input}{[required] Input data to be analyzed by the job.}

\item{OutputConfig}{[required] The Amazon S3 bucket location to store the results.}

\item{KmsKeyId}{The identifier of customer managed AWS KMS key (name or ARN). The key is used to encrypt images copied into the service. The key is also used to encrypt results and manifest files written to the output Amazon S3 bucket.}
}
\description{
Service availability notice: Streaming Video and Bulk Image Analysis is no longer available to new customers. For more information, see \href{https://docs.aws.amazon.com/rekognition/latest/dg/rekognition-availability-changes.html}{Rekognition feature availability changes}.

See \url{https://www.paws-r-sdk.com/docs/rekognition_start_media_analysis_job/} for full documentation.
}
\keyword{internal}
