% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_start_stream_processor}
\alias{rekognition_start_stream_processor}
\title{Service availability notice: Streaming Video and Bulk Image Analysis is
no longer available to new customers}
\usage{
rekognition_start_stream_processor(
  Name,
  StartSelector = NULL,
  StopSelector = NULL
)
}
\arguments{
\item{Name}{[required] The name of the stream processor to start processing.}

\item{StartSelector}{Specifies the starting point in the Kinesis stream to start processing. You can use the producer timestamp or the fragment number. If you use the producer timestamp, you must put the time in milliseconds. For more information about fragment numbers, see \href{https://docs.aws.amazon.com/kinesisvideostreams/latest/APIReference/API_reader_Fragment.html}{Fragment}.

This is a required parameter for label detection stream processors and should not be used to start a face search stream processor.}

\item{StopSelector}{Specifies when to stop processing the stream. You can specify a maximum amount of time to process the video.

This is a required parameter for label detection stream processors and should not be used to start a face search stream processor.}
}
\description{
Service availability notice: Streaming Video and Bulk Image Analysis is no longer available to new customers. For more information, see \href{https://docs.aws.amazon.com/rekognition/latest/dg/rekognition-availability-changes.html}{Rekognition feature availability changes}.

See \url{https://www.paws-r-sdk.com/docs/rekognition_start_stream_processor/} for full documentation.
}
\keyword{internal}
