% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_stop_project_version}
\alias{rekognition_stop_project_version}
\title{This operation applies only to Amazon Rekognition Custom Labels}
\usage{
rekognition_stop_project_version(ProjectVersionArn)
}
\arguments{
\item{ProjectVersionArn}{[required] The Amazon Resource Name (ARN) of the model version that you want to
stop.

This operation requires permissions to perform the
\code{rekognition:StopProjectVersion} action.}
}
\description{
This operation applies only to Amazon Rekognition Custom Labels.

See \url{https://www.paws-r-sdk.com/docs/rekognition_stop_project_version/} for full documentation.
}
\keyword{internal}
