% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_untag_resource}
\alias{rekognition_untag_resource}
\title{Removes one or more tags from an Amazon Rekognition collection, stream
processor, or Custom Labels model}
\usage{
rekognition_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] Amazon Resource Name (ARN) of the model, collection, or stream processor
that you want to remove the tags from.}

\item{TagKeys}{[required] A list of the tags that you want to remove.}
}
\description{
Removes one or more tags from an Amazon Rekognition collection, stream processor, or Custom Labels model.

See \url{https://www.paws-r-sdk.com/docs/rekognition_untag_resource/} for full documentation.
}
\keyword{internal}
