% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_update_dataset_entries}
\alias{rekognition_update_dataset_entries}
\title{This operation applies only to Amazon Rekognition Custom Labels}
\usage{
rekognition_update_dataset_entries(DatasetArn, Changes)
}
\arguments{
\item{DatasetArn}{[required] The Amazon Resource Name (ARN) of the dataset that you want to update.}

\item{Changes}{[required] The changes that you want to make to the dataset.}
}
\description{
This operation applies only to Amazon Rekognition Custom Labels.

See \url{https://www.paws-r-sdk.com/docs/rekognition_update_dataset_entries/} for full documentation.
}
\keyword{internal}
