% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_update_stream_processor}
\alias{rekognition_update_stream_processor}
\title{Service availability notice: Streaming Video and Bulk Image Analysis is
no longer available to new customers}
\usage{
rekognition_update_stream_processor(
  Name,
  SettingsForUpdate = NULL,
  RegionsOfInterestForUpdate = NULL,
  DataSharingPreferenceForUpdate = NULL,
  ParametersToDelete = NULL
)
}
\arguments{
\item{Name}{[required] Name of the stream processor that you want to update.}

\item{SettingsForUpdate}{The stream processor settings that you want to update. Label detection settings can be updated to detect different labels with a different minimum confidence.}

\item{RegionsOfInterestForUpdate}{Specifies locations in the frames where Amazon Rekognition checks for objects or people. This is an optional parameter for label detection stream processors.}

\item{DataSharingPreferenceForUpdate}{Shows whether you are sharing data with Rekognition to improve model performance. You can choose this option at the account level or on a per-stream basis. Note that if you opt out at the account level this setting is ignored on individual streams.}

\item{ParametersToDelete}{A list of parameters you want to delete from the stream processor.}
}
\description{
Service availability notice: Streaming Video and Bulk Image Analysis is no longer available to new customers. For more information, see \href{https://docs.aws.amazon.com/rekognition/latest/dg/rekognition-availability-changes.html}{Rekognition feature availability changes}.

See \url{https://www.paws-r-sdk.com/docs/rekognition_update_stream_processor/} for full documentation.
}
\keyword{internal}
