% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_attach_cluster_node_network_interface}
\alias{sagemaker_attach_cluster_node_network_interface}
\title{Attaches an elastic network interface (ENI) to a node in a HyperPod
cluster}
\usage{
sagemaker_attach_cluster_node_network_interface(
  ClusterName,
  NodeId,
  NetworkInterfaceId
)
}
\arguments{
\item{ClusterName}{[required] The name or Amazon Resource Name (ARN) of the SageMaker HyperPod cluster that contains the target node.}

\item{NodeId}{[required] The unique identifier of the cluster node to which you want to attach the network interface. The node must belong to your specified HyperPod cluster and cannot be part of a Restricted Instance Group (RIG).}

\item{NetworkInterfaceId}{[required] The unique identifier of the elastic network interface (ENI) to attach.}
}
\description{
Attaches an elastic network interface (ENI) to a node in a HyperPod cluster.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_attach_cluster_node_network_interface/} for full documentation.
}
\keyword{internal}
