% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_batch_delete_cluster_nodes}
\alias{sagemaker_batch_delete_cluster_nodes}
\title{Deletes specific nodes within a SageMaker HyperPod cluster}
\usage{
sagemaker_batch_delete_cluster_nodes(ClusterName, NodeIds)
}
\arguments{
\item{ClusterName}{[required] The name of the SageMaker HyperPod cluster from which to delete the
specified nodes.}

\item{NodeIds}{[required] A list of node IDs to be deleted from the specified cluster.

For SageMaker HyperPod clusters using the Slurm workload manager, you
cannot remove instances that are configured as Slurm controller nodes.}
}
\description{
Deletes specific nodes within a SageMaker HyperPod cluster. \code{\link[=sagemaker_batch_delete_cluster_nodes]{batch_delete_cluster_nodes}} accepts a cluster name and a list of node IDs.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_batch_delete_cluster_nodes/} for full documentation.
}
\keyword{internal}
