% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_batch_reboot_cluster_nodes}
\alias{sagemaker_batch_reboot_cluster_nodes}
\title{Reboots specific nodes within a SageMaker HyperPod cluster using a soft
recovery mechanism}
\usage{
sagemaker_batch_reboot_cluster_nodes(
  ClusterName,
  NodeIds = NULL,
  NodeLogicalIds = NULL
)
}
\arguments{
\item{ClusterName}{[required] The name or Amazon Resource Name (ARN) of the SageMaker HyperPod cluster
containing the nodes to reboot.}

\item{NodeIds}{A list of EC2 instance IDs to reboot using soft recovery. You can
specify between 1 and 25 instance IDs.
\itemize{
\item Either \code{NodeIds} or \code{NodeLogicalIds} must be provided (or both), but
at least one is required.
\item Each instance ID must follow the pattern \verb{i-} followed by 17
hexadecimal characters (for example, \verb{i-0123456789abcdef0}).
}}

\item{NodeLogicalIds}{A list of logical node IDs to reboot using soft recovery. You can
specify between 1 and 25 logical node IDs.

The \code{NodeLogicalId} is a unique identifier that persists throughout the
node's lifecycle and can be used to track nodes that are still being
provisioned and don't yet have an EC2 instance ID assigned.
\itemize{
\item This parameter is only supported for clusters using \code{Continuous} as
the \code{NodeProvisioningMode}. For clusters using the default
provisioning mode, use \code{NodeIds} instead.
\item Either \code{NodeIds} or \code{NodeLogicalIds} must be provided (or both), but
at least one is required.
}}
}
\description{
Reboots specific nodes within a SageMaker HyperPod cluster using a soft recovery mechanism. \code{\link[=sagemaker_batch_reboot_cluster_nodes]{batch_reboot_cluster_nodes}} performs a graceful reboot of the specified nodes by calling the Amazon Elastic Compute Cloud \code{RebootInstances} API, which attempts to cleanly shut down the operating system before restarting the instance.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_batch_reboot_cluster_nodes/} for full documentation.
}
\keyword{internal}
