% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_create_cluster}
\alias{sagemaker_create_cluster}
\title{Creates a SageMaker HyperPod cluster}
\usage{
sagemaker_create_cluster(
  ClusterName,
  InstanceGroups,
  VpcConfig = NULL,
  Tags = NULL,
  Orchestrator = NULL,
  NodeRecovery = NULL
)
}
\arguments{
\item{ClusterName}{[required] The name for the new SageMaker HyperPod cluster.}

\item{InstanceGroups}{[required] The instance groups to be created in the SageMaker HyperPod cluster.}

\item{VpcConfig}{Specifies the Amazon Virtual Private Cloud (VPC) that is associated with
the Amazon SageMaker HyperPod cluster. You can control access to and
from your resources by configuring your VPC. For more information, see
\href{https://docs.aws.amazon.com/sagemaker/latest/dg/infrastructure-give-access.html}{Give SageMaker access to resources in your Amazon VPC}.

If you configure your VPC with IPv6 support and specify subnets with
IPv6 addressing enabled in your VPC configuration, the cluster
automatically uses IPv6 addressing for network communication.

For information about adding IPv6 support for your VPC, see \href{https://docs.aws.amazon.com/vpc/latest/userguide/vpc-migrate-ipv6.html}{IPv6 support for your VPC}.

For information about creating a new VPC for use with IPv6, see \href{https://docs.aws.amazon.com/vpc/latest/userguide/create-vpc.html}{Create a VPC}.}

\item{Tags}{Custom tags for managing the SageMaker HyperPod cluster as an Amazon Web
Services resource. You can add tags to your cluster in the same way you
add them in other Amazon Web Services services that support tagging. To
learn more about tagging Amazon Web Services resources in general, see
\href{https://docs.aws.amazon.com/tag-editor/latest/userguide/tagging.html}{Tagging Amazon Web Services Resources User Guide}.}

\item{Orchestrator}{The type of orchestrator to use for the SageMaker HyperPod cluster.
Currently, the only supported value is \code{"eks"}, which is to use an
Amazon Elastic Kubernetes Service (EKS) cluster as the orchestrator.}

\item{NodeRecovery}{The node recovery mode for the SageMaker HyperPod cluster. When set to
\code{Automatic}, SageMaker HyperPod will automatically reboot or replace
faulty nodes when issues are detected. When set to \code{None}, cluster
administrators will need to manually manage any faulty cluster
instances.}
}
\description{
Creates a SageMaker HyperPod cluster. SageMaker HyperPod is a capability of SageMaker for creating and managing persistent clusters for developing large machine learning models, such as large language models (LLMs) and diffusion models. To learn more, see \href{https://docs.aws.amazon.com/sagemaker/latest/dg/sagemaker-hyperpod.html}{Amazon SageMaker HyperPod} in the \emph{Amazon SageMaker Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_create_cluster/} for full documentation.
}
\keyword{internal}
