% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_create_cluster_scheduler_config}
\alias{sagemaker_create_cluster_scheduler_config}
\title{Create cluster policy configuration}
\usage{
sagemaker_create_cluster_scheduler_config(
  Name,
  ClusterArn,
  SchedulerConfig,
  Description = NULL,
  Tags = NULL
)
}
\arguments{
\item{Name}{[required] The name for the cluster policy. The name must be unique within the SageMaker AI HyperPod cluster specified by \code{ClusterArn}. You can use the same name in other clusters within a Region or across Regions.}

\item{ClusterArn}{[required] ARN of the cluster.}

\item{SchedulerConfig}{[required] Configuration about the monitoring schedule.}

\item{Description}{Description of the cluster policy.}

\item{Tags}{Tags of the cluster policy.}
}
\description{
Create cluster policy configuration. This policy is used for task prioritization and fair-share allocation of idle compute. This helps prioritize critical workloads and distributes idle compute across entities.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_create_cluster_scheduler_config/} for full documentation.
}
\keyword{internal}
