% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_create_data_quality_job_definition}
\alias{sagemaker_create_data_quality_job_definition}
\title{Creates a definition for a job that monitors data quality and drift}
\usage{
sagemaker_create_data_quality_job_definition(
  JobDefinitionName,
  DataQualityBaselineConfig = NULL,
  DataQualityAppSpecification,
  DataQualityJobInput,
  DataQualityJobOutputConfig,
  JobResources,
  NetworkConfig = NULL,
  RoleArn,
  StoppingCondition = NULL,
  Tags = NULL
)
}
\arguments{
\item{JobDefinitionName}{[required] The name for the monitoring job definition.}

\item{DataQualityBaselineConfig}{Configures the constraints and baselines for the monitoring job.}

\item{DataQualityAppSpecification}{[required] Specifies the container that runs the monitoring job.}

\item{DataQualityJobInput}{[required] A list of inputs for the monitoring job. Currently endpoints are supported as monitoring inputs.}

\item{DataQualityJobOutputConfig}{[required] The output configuration for monitoring jobs.}

\item{JobResources}{[required] Identifies the resources to deploy for a monitoring job.}

\item{NetworkConfig}{Specifies networking configuration for the monitoring job.}

\item{RoleArn}{[required] The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker AI can assume to perform tasks on your behalf.}

\item{StoppingCondition}{Specifies a limit to how long a job can run. When the job reaches the time limit, SageMaker ends the job. Use this API to cap costs.

To stop a training job, SageMaker sends the algorithm the \code{SIGTERM} signal, which delays job termination for 120 seconds. Algorithms can use this 120-second window to save the model artifacts, so the results of training are not lost.

The training algorithms provided by SageMaker automatically save the intermediate results of a model training job when possible. This attempt to save artifacts is only a best effort case as model might not be in a state from which it can be saved. For example, if training has just started, the model might not be ready to save. When saved, this intermediate data is a valid model artifact. You can use it to create a model with \code{\link[=sagemaker_create_model]{create_model}}.

The Neural Topic Model (NTM) currently does not support saving intermediate model artifacts. When training NTMs, make sure that the maximum runtime is sufficient for the training job to complete.}

\item{Tags}{(Optional) An array of key-value pairs. For more information, see \href{https://docs.aws.amazon.com/awsaccountbilling/latest/aboutv2/cost-alloc-tags.html#allocation-whatURL}{Using Cost Allocation Tags} in the \emph{Amazon Web Services Billing and Cost Management User Guide}.}
}
\description{
Creates a definition for a job that monitors data quality and drift. For information about model monitor, see \href{https://docs.aws.amazon.com/sagemaker/latest/dg/model-monitor.html}{Amazon SageMaker AI Model Monitor}.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_create_data_quality_job_definition/} for full documentation.
}
\keyword{internal}
