% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_create_edge_deployment_stage}
\alias{sagemaker_create_edge_deployment_stage}
\title{Creates a new stage in an existing edge deployment plan}
\usage{
sagemaker_create_edge_deployment_stage(EdgeDeploymentPlanName, Stages)
}
\arguments{
\item{EdgeDeploymentPlanName}{[required] The name of the edge deployment plan.}

\item{Stages}{[required] List of stages to be added to the edge deployment plan.}
}
\description{
Creates a new stage in an existing edge deployment plan.

See \url{https://paws-r.github.io/docs/sagemaker/create_edge_deployment_stage.html} for full documentation.
}
\keyword{internal}
