% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_create_hub}
\alias{sagemaker_create_hub}
\title{Create a hub}
\usage{
sagemaker_create_hub(
  HubName,
  HubDescription,
  HubDisplayName = NULL,
  HubSearchKeywords = NULL,
  S3StorageConfig = NULL,
  Tags = NULL
)
}
\arguments{
\item{HubName}{[required] The name of the hub to create.}

\item{HubDescription}{[required] A description of the hub.}

\item{HubDisplayName}{The display name of the hub.}

\item{HubSearchKeywords}{The searchable keywords for the hub.}

\item{S3StorageConfig}{The Amazon S3 storage configuration for the hub.}

\item{Tags}{Any tags to associate with the hub.}
}
\description{
Create a hub.

See \url{https://paws-r.github.io/docs/sagemaker/create_hub.html} for full documentation.
}
\keyword{internal}
