% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_create_job}
\alias{sagemaker_create_job}
\title{Creates a model customization job in Amazon SageMaker}
\usage{
sagemaker_create_job(
  JobName,
  RoleArn,
  JobCategory,
  JobConfigSchemaVersion,
  JobConfigDocument,
  Tags = NULL
)
}
\arguments{
\item{JobName}{[required] The name of the job. The name must be unique within your account and Amazon Web Services Region.}

\item{RoleArn}{[required] The Amazon Resource Name (ARN) of the IAM role that Amazon SageMaker assumes to perform the job. The role must have the necessary permissions to access the resources required by the job configuration.}

\item{JobCategory}{[required] The category of the job. The category determines the type of workload that the job runs.}

\item{JobConfigSchemaVersion}{[required] The version of the configuration schema to use for the job configuration document. Use \code{\link[=sagemaker_list_job_schema_versions]{list_job_schema_versions}} to get available schema versions for a job category.}

\item{JobConfigDocument}{[required] The JSON configuration document for the job. The document must conform to the schema specified by \code{JobConfigSchemaVersion}. Use \code{\link[=sagemaker_describe_job_schema_version]{describe_job_schema_version}} to retrieve the schema for validation.}

\item{Tags}{An array of key-value pairs to apply to the job as tags. For more information, see \href{https://docs.aws.amazon.com/tag-editor/latest/userguide/tagging.html}{Tagging Amazon Web Services Resources}.}
}
\description{
Creates a model customization job in Amazon SageMaker. A job runs a workload based on the job category and configuration you provide. You specify the job category, a schema-versioned configuration document, and an IAM role that grants Amazon SageMaker permission to access resources on your behalf.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_create_job/} for full documentation.
}
\keyword{internal}
