% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_create_model_bias_job_definition}
\alias{sagemaker_create_model_bias_job_definition}
\title{Creates the definition for a model bias job}
\usage{
sagemaker_create_model_bias_job_definition(
  JobDefinitionName,
  ModelBiasBaselineConfig = NULL,
  ModelBiasAppSpecification,
  ModelBiasJobInput,
  ModelBiasJobOutputConfig,
  JobResources,
  NetworkConfig = NULL,
  RoleArn,
  StoppingCondition = NULL,
  Tags = NULL
)
}
\arguments{
\item{JobDefinitionName}{[required] The name of the bias job definition. The name must be unique within an
Amazon Web Services Region in the Amazon Web Services account.}

\item{ModelBiasBaselineConfig}{The baseline configuration for a model bias job.}

\item{ModelBiasAppSpecification}{[required] Configures the model bias job to run a specified Docker container image.}

\item{ModelBiasJobInput}{[required] Inputs for the model bias job.}

\item{ModelBiasJobOutputConfig}{[required]}

\item{JobResources}{[required]}

\item{NetworkConfig}{Networking options for a model bias job.}

\item{RoleArn}{[required] The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can
assume to perform tasks on your behalf.}

\item{StoppingCondition}{}

\item{Tags}{(Optional) An array of key-value pairs. For more information, see Using
Cost Allocation Tags in the \emph{Amazon Web Services Billing and Cost
Management User Guide}.}
}
\description{
Creates the definition for a model bias job.

See \url{https://paws-r.github.io/docs/sagemaker/create_model_bias_job_definition.html} for full documentation.
}
\keyword{internal}
