% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_create_model_card_export_job}
\alias{sagemaker_create_model_card_export_job}
\title{Creates an Amazon SageMaker Model Card export job}
\usage{
sagemaker_create_model_card_export_job(
  ModelCardName,
  ModelCardVersion = NULL,
  ModelCardExportJobName,
  OutputConfig
)
}
\arguments{
\item{ModelCardName}{[required] The name of the model card to export.}

\item{ModelCardVersion}{The version of the model card to export. If a version is not provided,
then the latest version of the model card is exported.}

\item{ModelCardExportJobName}{[required] The name of the model card export job.}

\item{OutputConfig}{[required] The model card output configuration that specifies the Amazon S3 path
for exporting.}
}
\description{
Creates an Amazon SageMaker Model Card export job.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_create_model_card_export_job/} for full documentation.
}
\keyword{internal}
