% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_create_model_package_group}
\alias{sagemaker_create_model_package_group}
\title{Creates a model group}
\usage{
sagemaker_create_model_package_group(
  ModelPackageGroupName,
  ModelPackageGroupDescription = NULL,
  Tags = NULL
)
}
\arguments{
\item{ModelPackageGroupName}{[required] The name of the model group.}

\item{ModelPackageGroupDescription}{A description for the model group.}

\item{Tags}{A list of key value pairs associated with the model group. For more
information, see \href{https://docs.aws.amazon.com/tag-editor/latest/userguide/tagging.html}{Tagging Amazon Web Services resources}
in the \emph{Amazon Web Services General Reference Guide}.}
}
\description{
Creates a model group. A model group contains a group of model versions.

See \url{https://paws-r.github.io/docs/sagemaker/create_model_package_group.html} for full documentation.
}
\keyword{internal}
