% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_create_monitoring_schedule}
\alias{sagemaker_create_monitoring_schedule}
\title{Creates a schedule that regularly starts Amazon SageMaker Processing
Jobs to monitor the data captured for an Amazon SageMaker Endoint}
\usage{
sagemaker_create_monitoring_schedule(
  MonitoringScheduleName,
  MonitoringScheduleConfig,
  Tags = NULL
)
}
\arguments{
\item{MonitoringScheduleName}{[required] The name of the monitoring schedule. The name must be unique within an
Amazon Web Services Region within an Amazon Web Services account.}

\item{MonitoringScheduleConfig}{[required] The configuration object that specifies the monitoring schedule and
defines the monitoring job.}

\item{Tags}{(Optional) An array of key-value pairs. For more information, see Using
Cost Allocation Tags in the \emph{Amazon Web Services Billing and Cost
Management User Guide}.}
}
\description{
Creates a schedule that regularly starts Amazon SageMaker Processing Jobs to monitor the data captured for an Amazon SageMaker Endoint.

See \url{https://paws-r.github.io/docs/sagemaker/create_monitoring_schedule.html} for full documentation.
}
\keyword{internal}
