% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_create_notebook_instance_lifecycle_config}
\alias{sagemaker_create_notebook_instance_lifecycle_config}
\title{Creates a lifecycle configuration that you can associate with a notebook
instance}
\usage{
sagemaker_create_notebook_instance_lifecycle_config(
  NotebookInstanceLifecycleConfigName,
  OnCreate = NULL,
  OnStart = NULL
)
}
\arguments{
\item{NotebookInstanceLifecycleConfigName}{[required] The name of the lifecycle configuration.}

\item{OnCreate}{A shell script that runs only once, when you create a notebook instance.
The shell script must be a base64-encoded string.}

\item{OnStart}{A shell script that runs every time you start a notebook instance,
including when you create the notebook instance. The shell script must
be a base64-encoded string.}
}
\description{
Creates a lifecycle configuration that you can associate with a notebook instance. A \emph{lifecycle configuration} is a collection of shell scripts that run when you create or start a notebook instance.

See \url{https://paws-r.github.io/docs/sagemaker/create_notebook_instance_lifecycle_config.html} for full documentation.
}
\keyword{internal}
