% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_create_training_plan}
\alias{sagemaker_create_training_plan}
\title{Creates a new training plan in SageMaker to reserve compute capacity}
\usage{
sagemaker_create_training_plan(
  TrainingPlanName,
  TrainingPlanOfferingId,
  SpareInstanceCountPerUltraServer = NULL,
  Tags = NULL
)
}
\arguments{
\item{TrainingPlanName}{[required] The name of the training plan to create.}

\item{TrainingPlanOfferingId}{[required] The unique identifier of the training plan offering to use for creating
this plan.}

\item{SpareInstanceCountPerUltraServer}{Number of spare instances to reserve per UltraServer for enhanced
resiliency. Default is 1.}

\item{Tags}{An array of key-value pairs to apply to this training plan.}
}
\description{
Creates a new training plan in SageMaker to reserve compute capacity.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_create_training_plan/} for full documentation.
}
\keyword{internal}
