% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_action}
\alias{sagemaker_delete_action}
\title{Deletes an action}
\usage{
sagemaker_delete_action(ActionName)
}
\arguments{
\item{ActionName}{[required] The name of the action to delete.}
}
\description{
Deletes an action.

See \url{https://paws-r.github.io/docs/sagemaker/delete_action.html} for full documentation.
}
\keyword{internal}
