% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_ai_benchmark_job}
\alias{sagemaker_delete_ai_benchmark_job}
\title{Deletes the specified AI benchmark job}
\usage{
sagemaker_delete_ai_benchmark_job(AIBenchmarkJobName)
}
\arguments{
\item{AIBenchmarkJobName}{[required] The name of the AI benchmark job to delete.}
}
\description{
Deletes the specified AI benchmark job.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_delete_ai_benchmark_job/} for full documentation.
}
\keyword{internal}
