% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_ai_recommendation_job}
\alias{sagemaker_delete_ai_recommendation_job}
\title{Deletes the specified AI recommendation job}
\usage{
sagemaker_delete_ai_recommendation_job(AIRecommendationJobName)
}
\arguments{
\item{AIRecommendationJobName}{[required] The name of the AI recommendation job to delete.}
}
\description{
Deletes the specified AI recommendation job.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_delete_ai_recommendation_job/} for full documentation.
}
\keyword{internal}
