% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_ai_workload_config}
\alias{sagemaker_delete_ai_workload_config}
\title{Deletes the specified AI workload configuration}
\usage{
sagemaker_delete_ai_workload_config(AIWorkloadConfigName)
}
\arguments{
\item{AIWorkloadConfigName}{[required] The name of the AI workload configuration to delete.}
}
\description{
Deletes the specified AI workload configuration. You cannot delete a configuration that is referenced by an active benchmark job.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_delete_ai_workload_config/} for full documentation.
}
\keyword{internal}
