% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_algorithm}
\alias{sagemaker_delete_algorithm}
\title{Removes the specified algorithm from your account}
\usage{
sagemaker_delete_algorithm(AlgorithmName)
}
\arguments{
\item{AlgorithmName}{[required] The name of the algorithm to delete.}
}
\description{
Removes the specified algorithm from your account.
}
\section{Request syntax}{
\preformatted{svc$delete_algorithm(
  AlgorithmName = "string"
)
}
}

\keyword{internal}
