% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_app_image_config}
\alias{sagemaker_delete_app_image_config}
\title{Deletes an AppImageConfig}
\usage{
sagemaker_delete_app_image_config(AppImageConfigName)
}
\arguments{
\item{AppImageConfigName}{[required] The name of the AppImageConfig to delete.}
}
\description{
Deletes an AppImageConfig.

See \url{https://paws-r.github.io/docs/sagemaker/delete_app_image_config.html} for full documentation.
}
\keyword{internal}
