% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_association}
\alias{sagemaker_delete_association}
\title{Deletes an association}
\usage{
sagemaker_delete_association(SourceArn, DestinationArn)
}
\arguments{
\item{SourceArn}{[required] The ARN of the source.}

\item{DestinationArn}{[required] The Amazon Resource Name (ARN) of the destination.}
}
\value{
A list with the following syntax:\preformatted{list(
  SourceArn = "string",
  DestinationArn = "string"
)
}
}
\description{
Deletes an association.
}
\section{Request syntax}{
\preformatted{svc$delete_association(
  SourceArn = "string",
  DestinationArn = "string"
)
}
}

\keyword{internal}
