% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_cluster}
\alias{sagemaker_delete_cluster}
\title{Delete a SageMaker HyperPod cluster}
\usage{
sagemaker_delete_cluster(ClusterName)
}
\arguments{
\item{ClusterName}{[required] The string name or the Amazon Resource Name (ARN) of the SageMaker
HyperPod cluster to delete.}
}
\description{
Delete a SageMaker HyperPod cluster.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_delete_cluster/} for full documentation.
}
\keyword{internal}
