% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_cluster_scheduler_config}
\alias{sagemaker_delete_cluster_scheduler_config}
\title{Deletes the cluster policy of the cluster}
\usage{
sagemaker_delete_cluster_scheduler_config(ClusterSchedulerConfigId)
}
\arguments{
\item{ClusterSchedulerConfigId}{[required] ID of the cluster policy.}
}
\description{
Deletes the cluster policy of the cluster.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_delete_cluster_scheduler_config/} for full documentation.
}
\keyword{internal}
