% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_code_repository}
\alias{sagemaker_delete_code_repository}
\title{Deletes the specified Git repository from your account}
\usage{
sagemaker_delete_code_repository(CodeRepositoryName)
}
\arguments{
\item{CodeRepositoryName}{[required] The name of the Git repository to delete.}
}
\description{
Deletes the specified Git repository from your account.

See \url{https://paws-r.github.io/docs/sagemaker/delete_code_repository.html} for full documentation.
}
\keyword{internal}
