% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_compilation_job}
\alias{sagemaker_delete_compilation_job}
\title{Deletes the specified compilation job}
\usage{
sagemaker_delete_compilation_job(CompilationJobName)
}
\arguments{
\item{CompilationJobName}{[required] The name of the compilation job to delete.}
}
\description{
Deletes the specified compilation job. This action deletes only the compilation job resource in Amazon SageMaker AI. It doesn't delete other resources that are related to that job, such as the model artifacts that the job creates, the compilation logs in CloudWatch, the compiled model, or the IAM role.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_delete_compilation_job/} for full documentation.
}
\keyword{internal}
