% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_compute_quota}
\alias{sagemaker_delete_compute_quota}
\title{Deletes the compute allocation from the cluster}
\usage{
sagemaker_delete_compute_quota(ComputeQuotaId)
}
\arguments{
\item{ComputeQuotaId}{[required] ID of the compute allocation definition.}
}
\description{
Deletes the compute allocation from the cluster.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_delete_compute_quota/} for full documentation.
}
\keyword{internal}
