% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_context}
\alias{sagemaker_delete_context}
\title{Deletes an context}
\usage{
sagemaker_delete_context(ContextName)
}
\arguments{
\item{ContextName}{[required] The name of the context to delete.}
}
\description{
Deletes an context.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_delete_context/} for full documentation.
}
\keyword{internal}
