% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_data_quality_job_definition}
\alias{sagemaker_delete_data_quality_job_definition}
\title{Deletes a data quality monitoring job definition}
\usage{
sagemaker_delete_data_quality_job_definition(JobDefinitionName)
}
\arguments{
\item{JobDefinitionName}{[required] The name of the data quality monitoring job definition to delete.}
}
\value{
An empty list.
}
\description{
Deletes a data quality monitoring job definition.
}
\section{Request syntax}{
\preformatted{svc$delete_data_quality_job_definition(
  JobDefinitionName = "string"
)
}
}

\keyword{internal}
