% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_device_fleet}
\alias{sagemaker_delete_device_fleet}
\title{Deletes a fleet}
\usage{
sagemaker_delete_device_fleet(DeviceFleetName)
}
\arguments{
\item{DeviceFleetName}{[required] The name of the fleet to delete.}
}
\description{
Deletes a fleet.
}
\section{Request syntax}{
\preformatted{svc$delete_device_fleet(
  DeviceFleetName = "string"
)
}
}

\keyword{internal}
