% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_edge_deployment_plan}
\alias{sagemaker_delete_edge_deployment_plan}
\title{Deletes an edge deployment plan if (and only if) all the stages in the
plan are inactive or there are no stages in the plan}
\usage{
sagemaker_delete_edge_deployment_plan(EdgeDeploymentPlanName)
}
\arguments{
\item{EdgeDeploymentPlanName}{[required] The name of the edge deployment plan to delete.}
}
\description{
Deletes an edge deployment plan if (and only if) all the stages in the plan are inactive or there are no stages in the plan.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_delete_edge_deployment_plan/} for full documentation.
}
\keyword{internal}
