% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_edge_deployment_stage}
\alias{sagemaker_delete_edge_deployment_stage}
\title{Delete a stage in an edge deployment plan if (and only if) the stage is
inactive}
\usage{
sagemaker_delete_edge_deployment_stage(EdgeDeploymentPlanName, StageName)
}
\arguments{
\item{EdgeDeploymentPlanName}{[required] The name of the edge deployment plan from which the stage will be
deleted.}

\item{StageName}{[required] The name of the stage.}
}
\description{
Delete a stage in an edge deployment plan if (and only if) the stage is inactive.

See \url{https://paws-r.github.io/docs/sagemaker/delete_edge_deployment_stage.html} for full documentation.
}
\keyword{internal}
