% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_endpoint}
\alias{sagemaker_delete_endpoint}
\title{Deletes an endpoint}
\usage{
sagemaker_delete_endpoint(EndpointName)
}
\arguments{
\item{EndpointName}{[required] The name of the endpoint that you want to delete.}
}
\description{
Deletes an endpoint. SageMaker frees up all of the resources that were deployed when the endpoint was created.

See \url{https://paws-r.github.io/docs/sagemaker/delete_endpoint.html} for full documentation.
}
\keyword{internal}
