% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_flow_definition}
\alias{sagemaker_delete_flow_definition}
\title{Deletes the specified flow definition}
\usage{
sagemaker_delete_flow_definition(FlowDefinitionName)
}
\arguments{
\item{FlowDefinitionName}{[required] The name of the flow definition you are deleting.}
}
\description{
Deletes the specified flow definition.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_delete_flow_definition/} for full documentation.
}
\keyword{internal}
