% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_hub}
\alias{sagemaker_delete_hub}
\title{Delete a hub}
\usage{
sagemaker_delete_hub(HubName)
}
\arguments{
\item{HubName}{[required] The name of the hub to delete.}
}
\description{
Delete a hub.

See \url{https://paws-r.github.io/docs/sagemaker/delete_hub.html} for full documentation.
}
\keyword{internal}
