% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_hub_content}
\alias{sagemaker_delete_hub_content}
\title{Delete the contents of a hub}
\usage{
sagemaker_delete_hub_content(
  HubName,
  HubContentType,
  HubContentName,
  HubContentVersion
)
}
\arguments{
\item{HubName}{[required] The name of the hub that you want to delete content in.}

\item{HubContentType}{[required] The type of content that you want to delete from a hub.}

\item{HubContentName}{[required] The name of the content that you want to delete from a hub.}

\item{HubContentVersion}{[required] The version of the content that you want to delete from a hub.}
}
\description{
Delete the contents of a hub.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_delete_hub_content/} for full documentation.
}
\keyword{internal}
