% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_hub_content_reference}
\alias{sagemaker_delete_hub_content_reference}
\title{Delete a hub content reference in order to remove a model from a private
hub}
\usage{
sagemaker_delete_hub_content_reference(HubName, HubContentType, HubContentName)
}
\arguments{
\item{HubName}{[required] The name of the hub to delete the hub content reference from.}

\item{HubContentType}{[required] The type of hub content reference to delete. The only supported type of
hub content reference to delete is \code{ModelReference}.}

\item{HubContentName}{[required] The name of the hub content to delete.}
}
\description{
Delete a hub content reference in order to remove a model from a private hub.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_delete_hub_content_reference/} for full documentation.
}
\keyword{internal}
