% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_image_version}
\alias{sagemaker_delete_image_version}
\title{Deletes a version of a SageMaker image}
\usage{
sagemaker_delete_image_version(ImageName, Version)
}
\arguments{
\item{ImageName}{[required] The name of the image.}

\item{Version}{[required] The version to delete.}
}
\value{
An empty list.
}
\description{
Deletes a version of a SageMaker image. The container image the version
represents isn't deleted.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_image_version(
  ImageName = "string",
  Version = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
