% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_inference_experiment}
\alias{sagemaker_delete_inference_experiment}
\title{Deletes an inference experiment}
\usage{
sagemaker_delete_inference_experiment(Name)
}
\arguments{
\item{Name}{[required] The name of the inference experiment you want to delete.}
}
\description{
Deletes an inference experiment.

See \url{https://paws-r.github.io/docs/sagemaker/delete_inference_experiment.html} for full documentation.
}
\keyword{internal}
