% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_job}
\alias{sagemaker_delete_job}
\title{Deletes a job}
\usage{
sagemaker_delete_job(JobName, JobCategory)
}
\arguments{
\item{JobName}{[required] The name of the job to delete.}

\item{JobCategory}{[required] The category of the job to delete.}
}
\description{
Deletes a job. This operation is idempotent. If the job is currently running, you must stop it before deleting it by calling \code{\link[=sagemaker_stop_job]{stop_job}}.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_delete_job/} for full documentation.
}
\keyword{internal}
