% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_mlflow_tracking_server}
\alias{sagemaker_delete_mlflow_tracking_server}
\title{Deletes an MLflow Tracking Server}
\usage{
sagemaker_delete_mlflow_tracking_server(TrackingServerName)
}
\arguments{
\item{TrackingServerName}{[required] The name of the the tracking server to delete.}
}
\description{
Deletes an MLflow Tracking Server. For more information, see \href{https://docs.aws.amazon.com/sagemaker/latest/dg/}{Clean up MLflow resources}.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_delete_mlflow_tracking_server/} for full documentation.
}
\keyword{internal}
