% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_model_package}
\alias{sagemaker_delete_model_package}
\title{Deletes a model package}
\usage{
sagemaker_delete_model_package(ModelPackageName)
}
\arguments{
\item{ModelPackageName}{[required] The name or Amazon Resource Name (ARN) of the model package to delete.

When you specify a name, the name must have 1 to 63 characters. Valid
characters are a-z, A-Z, 0-9, and - (hyphen).}
}
\description{
Deletes a model package.

See \url{https://paws-r.github.io/docs/sagemaker/delete_model_package.html} for full documentation.
}
\keyword{internal}
