% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_model_package_group}
\alias{sagemaker_delete_model_package_group}
\title{Deletes the specified model group}
\usage{
sagemaker_delete_model_package_group(ModelPackageGroupName)
}
\arguments{
\item{ModelPackageGroupName}{[required] The name of the model group to delete.}
}
\description{
Deletes the specified model group.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_delete_model_package_group/} for full documentation.
}
\keyword{internal}
