% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_model_quality_job_definition}
\alias{sagemaker_delete_model_quality_job_definition}
\title{Deletes the secified model quality monitoring job definition}
\usage{
sagemaker_delete_model_quality_job_definition(JobDefinitionName)
}
\arguments{
\item{JobDefinitionName}{[required] The name of the model quality monitoring job definition to delete.}
}
\description{
Deletes the secified model quality monitoring job definition.

See \url{https://paws-r.github.io/docs/sagemaker/delete_model_quality_job_definition.html} for full documentation.
}
\keyword{internal}
