% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_partner_app}
\alias{sagemaker_delete_partner_app}
\title{Deletes a SageMaker Partner AI App}
\usage{
sagemaker_delete_partner_app(Arn, ClientToken = NULL)
}
\arguments{
\item{Arn}{[required] The ARN of the SageMaker Partner AI App to delete.}

\item{ClientToken}{A unique token that guarantees that the call to this API is idempotent.}
}
\description{
Deletes a SageMaker Partner AI App.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_delete_partner_app/} for full documentation.
}
\keyword{internal}
