% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_processing_job}
\alias{sagemaker_delete_processing_job}
\title{Deletes a processing job}
\usage{
sagemaker_delete_processing_job(ProcessingJobName)
}
\arguments{
\item{ProcessingJobName}{[required] The name of the processing job to delete.}
}
\description{
Deletes a processing job. After Amazon SageMaker deletes a processing job, all of the metadata for the processing job is lost. You can delete only processing jobs that are in a terminal state (\code{Stopped}, \code{Failed}, or \code{Completed}). You cannot delete a job that is in the \code{InProgress} or \code{Stopping} state. After deleting the job, you can reuse its name to create another processing job.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_delete_processing_job/} for full documentation.
}
\keyword{internal}
