% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_project}
\alias{sagemaker_delete_project}
\title{Delete the specified project}
\usage{
sagemaker_delete_project(ProjectName)
}
\arguments{
\item{ProjectName}{[required] The name of the project to delete.}
}
\description{
Delete the specified project.

See \url{https://paws-r.github.io/docs/sagemaker/delete_project.html} for full documentation.
}
\keyword{internal}
