% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_space}
\alias{sagemaker_delete_space}
\title{Used to delete a space}
\usage{
sagemaker_delete_space(DomainId, SpaceName)
}
\arguments{
\item{DomainId}{[required] The ID of the associated Domain.}

\item{SpaceName}{[required] The name of the space.}
}
\description{
Used to delete a space.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_delete_space/} for full documentation.
}
\keyword{internal}
