% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_training_job}
\alias{sagemaker_delete_training_job}
\title{Deletes a training job}
\usage{
sagemaker_delete_training_job(TrainingJobName)
}
\arguments{
\item{TrainingJobName}{[required] The name of the training job to delete.}
}
\description{
Deletes a training job. After SageMaker deletes a training job, all of the metadata for the training job is lost. You can delete only training jobs that are in a terminal state (\code{Stopped}, \code{Failed}, or \code{Completed}) and don't retain an \code{Available} \href{https://docs.aws.amazon.com/sagemaker/latest/dg/train-warm-pools.html}{managed warm pool}. You cannot delete a job that is in the \code{InProgress} or \code{Stopping} state. After deleting the job, you can reuse its name to create another training job.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_delete_training_job/} for full documentation.
}
\keyword{internal}
