% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_deregister_devices}
\alias{sagemaker_deregister_devices}
\title{Deregisters the specified devices}
\usage{
sagemaker_deregister_devices(DeviceFleetName, DeviceNames)
}
\arguments{
\item{DeviceFleetName}{[required] The name of the fleet the devices belong to.}

\item{DeviceNames}{[required] The unique IDs of the devices.}
}
\description{
Deregisters the specified devices. After you deregister a device, you will need to re-register the devices.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_deregister_devices/} for full documentation.
}
\keyword{internal}
