% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_ai_benchmark_job}
\alias{sagemaker_describe_ai_benchmark_job}
\title{Returns details of an AI benchmark job, including its status,
configuration, target endpoint, and timing information}
\usage{
sagemaker_describe_ai_benchmark_job(AIBenchmarkJobName)
}
\arguments{
\item{AIBenchmarkJobName}{[required] The name of the AI benchmark job to describe.}
}
\description{
Returns details of an AI benchmark job, including its status, configuration, target endpoint, and timing information.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_ai_benchmark_job/} for full documentation.
}
\keyword{internal}
