% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_ai_recommendation_job}
\alias{sagemaker_describe_ai_recommendation_job}
\title{Returns details of an AI recommendation job, including its status, model
source, performance targets, optimization recommendations, and
deployment configurations}
\usage{
sagemaker_describe_ai_recommendation_job(AIRecommendationJobName)
}
\arguments{
\item{AIRecommendationJobName}{[required] The name of the AI recommendation job to describe.}
}
\description{
Returns details of an AI recommendation job, including its status, model source, performance targets, optimization recommendations, and deployment configurations.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_ai_recommendation_job/} for full documentation.
}
\keyword{internal}
