% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_ai_workload_config}
\alias{sagemaker_describe_ai_workload_config}
\title{Returns details of an AI workload configuration, including the dataset
configuration, benchmark tool settings, tags, and creation time}
\usage{
sagemaker_describe_ai_workload_config(AIWorkloadConfigName)
}
\arguments{
\item{AIWorkloadConfigName}{[required] The name of the AI workload configuration to describe.}
}
\description{
Returns details of an AI workload configuration, including the dataset configuration, benchmark tool settings, tags, and creation time.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_ai_workload_config/} for full documentation.
}
\keyword{internal}
