% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_app_image_config}
\alias{sagemaker_describe_app_image_config}
\title{Describes an AppImageConfig}
\usage{
sagemaker_describe_app_image_config(AppImageConfigName)
}
\arguments{
\item{AppImageConfigName}{[required] The name of the AppImageConfig to describe.}
}
\description{
Describes an AppImageConfig.

See \url{https://paws-r.github.io/docs/sagemaker/describe_app_image_config.html} for full documentation.
}
\keyword{internal}
