% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_auto_ml_job_v2}
\alias{sagemaker_describe_auto_ml_job_v2}
\title{Returns information about an Amazon SageMaker AutoML V2 job}
\usage{
sagemaker_describe_auto_ml_job_v2(AutoMLJobName)
}
\arguments{
\item{AutoMLJobName}{[required] Requests information about an AutoML V2 job using its unique name.}
}
\description{
Returns information about an Amazon SageMaker AutoML V2 job.

See \url{https://paws-r.github.io/docs/sagemaker/describe_auto_ml_job_v2.html} for full documentation.
}
\keyword{internal}
