% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_cluster}
\alias{sagemaker_describe_cluster}
\title{Retrieves information of a SageMaker HyperPod cluster}
\usage{
sagemaker_describe_cluster(ClusterName)
}
\arguments{
\item{ClusterName}{[required] The string name or the Amazon Resource Name (ARN) of the SageMaker
HyperPod cluster.}
}
\description{
Retrieves information of a SageMaker HyperPod cluster.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_cluster/} for full documentation.
}
\keyword{internal}
