% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_cluster_event}
\alias{sagemaker_describe_cluster_event}
\title{Retrieves detailed information about a specific event for a given
HyperPod cluster}
\usage{
sagemaker_describe_cluster_event(EventId, ClusterName)
}
\arguments{
\item{EventId}{[required] The unique identifier (UUID) of the event to describe. This ID can be
obtained from the \code{\link[=sagemaker_list_cluster_events]{list_cluster_events}}
operation.}

\item{ClusterName}{[required] The name or Amazon Resource Name (ARN) of the HyperPod cluster
associated with the event.}
}
\description{
Retrieves detailed information about a specific event for a given HyperPod cluster. This functionality is only supported when the \code{NodeProvisioningMode} is set to \code{Continuous}.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_cluster_event/} for full documentation.
}
\keyword{internal}
