% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_cluster_node}
\alias{sagemaker_describe_cluster_node}
\title{Retrieves information of a node (also called a instance interchangeably)
of a SageMaker HyperPod cluster}
\usage{
sagemaker_describe_cluster_node(ClusterName, NodeId)
}
\arguments{
\item{ClusterName}{[required] The string name or the Amazon Resource Name (ARN) of the SageMaker
HyperPod cluster in which the node is.}

\item{NodeId}{[required] The ID of the SageMaker HyperPod cluster node.}
}
\description{
Retrieves information of a node (also called a \emph{instance} interchangeably) of a SageMaker HyperPod cluster.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_cluster_node/} for full documentation.
}
\keyword{internal}
