% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_compilation_job}
\alias{sagemaker_describe_compilation_job}
\title{Returns information about a model compilation job}
\usage{
sagemaker_describe_compilation_job(CompilationJobName)
}
\arguments{
\item{CompilationJobName}{[required] The name of the model compilation job that you want information about.}
}
\description{
Returns information about a model compilation job.

To create a model compilation job, use
\code{\link[=sagemaker_create_compilation_job]{create_compilation_job}}. To get
information about multiple model compilation jobs, use
\code{\link[=sagemaker_list_compilation_jobs]{list_compilation_jobs}}.
}
\section{Request syntax}{
\preformatted{svc$describe_compilation_job(
  CompilationJobName = "string"
)
}
}

\keyword{internal}
