% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_compute_quota}
\alias{sagemaker_describe_compute_quota}
\title{Description of the compute allocation definition}
\usage{
sagemaker_describe_compute_quota(ComputeQuotaId, ComputeQuotaVersion = NULL)
}
\arguments{
\item{ComputeQuotaId}{[required] ID of the compute allocation definition.}

\item{ComputeQuotaVersion}{Version of the compute allocation definition.}
}
\description{
Description of the compute allocation definition.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_compute_quota/} for full documentation.
}
\keyword{internal}
