% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_data_quality_job_definition}
\alias{sagemaker_describe_data_quality_job_definition}
\title{Gets the details of a data quality monitoring job definition}
\usage{
sagemaker_describe_data_quality_job_definition(JobDefinitionName)
}
\arguments{
\item{JobDefinitionName}{[required] The name of the data quality monitoring job definition to describe.}
}
\description{
Gets the details of a data quality monitoring job definition.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_data_quality_job_definition/} for full documentation.
}
\keyword{internal}
