% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_device}
\alias{sagemaker_describe_device}
\title{Describes the device}
\usage{
sagemaker_describe_device(NextToken = NULL, DeviceName, DeviceFleetName)
}
\arguments{
\item{NextToken}{Next token of device description.}

\item{DeviceName}{[required] The unique ID of the device.}

\item{DeviceFleetName}{[required] The name of the fleet the devices belong to.}
}
\description{
Describes the device.

See \url{https://paws-r.github.io/docs/sagemaker/describe_device.html} for full documentation.
}
\keyword{internal}
