% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_device_fleet}
\alias{sagemaker_describe_device_fleet}
\title{A description of the fleet the device belongs to}
\usage{
sagemaker_describe_device_fleet(DeviceFleetName)
}
\arguments{
\item{DeviceFleetName}{[required] The name of the fleet.}
}
\value{
A list with the following syntax:\preformatted{list(
  DeviceFleetName = "string",
  DeviceFleetArn = "string",
  OutputConfig = list(
    S3OutputLocation = "string",
    KmsKeyId = "string"
  ),
  Description = "string",
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  LastModifiedTime = as.POSIXct(
    "2015-01-01"
  ),
  RoleArn = "string",
  IotRoleAlias = "string"
)
}
}
\description{
A description of the fleet the device belongs to.
}
\section{Request syntax}{
\preformatted{svc$describe_device_fleet(
  DeviceFleetName = "string"
)
}
}

\keyword{internal}
