% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_edge_packaging_job}
\alias{sagemaker_describe_edge_packaging_job}
\title{A description of edge packaging jobs}
\usage{
sagemaker_describe_edge_packaging_job(EdgePackagingJobName)
}
\arguments{
\item{EdgePackagingJobName}{[required] The name of the edge packaging job.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  EdgePackagingJobArn = "string",
  EdgePackagingJobName = "string",
  CompilationJobName = "string",
  ModelName = "string",
  ModelVersion = "string",
  RoleArn = "string",
  OutputConfig = list(
    S3OutputLocation = "string",
    KmsKeyId = "string"
  ),
  ResourceKey = "string",
  EdgePackagingJobStatus = "STARTING"|"INPROGRESS"|"COMPLETED"|"FAILED"|"STOPPING"|"STOPPED",
  EdgePackagingJobStatusMessage = "string",
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  LastModifiedTime = as.POSIXct(
    "2015-01-01"
  ),
  ModelArtifact = "string",
  ModelSignature = "string"
)
}\if{html}{\out{</div>}}
}
\description{
A description of edge packaging jobs.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_edge_packaging_job(
  EdgePackagingJobName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
