% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_endpoint}
\alias{sagemaker_describe_endpoint}
\title{Returns the description of an endpoint}
\usage{
sagemaker_describe_endpoint(EndpointName)
}
\arguments{
\item{EndpointName}{[required] The name of the endpoint.}
}
\description{
Returns the description of an endpoint.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_endpoint/} for full documentation.
}
\keyword{internal}
