% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_experiment}
\alias{sagemaker_describe_experiment}
\title{Provides a list of an experiment's properties}
\usage{
sagemaker_describe_experiment(ExperimentName)
}
\arguments{
\item{ExperimentName}{[required] The name of the experiment to describe.}
}
\description{
Provides a list of an experiment's properties.

See \url{https://paws-r.github.io/docs/sagemaker/describe_experiment.html} for full documentation.
}
\keyword{internal}
