% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_feature_metadata}
\alias{sagemaker_describe_feature_metadata}
\title{Shows the metadata for a feature within a feature group}
\usage{
sagemaker_describe_feature_metadata(FeatureGroupName, FeatureName)
}
\arguments{
\item{FeatureGroupName}{[required] The name of the feature group containing the feature.}

\item{FeatureName}{[required] The name of the feature.}
}
\description{
Shows the metadata for a feature within a feature group.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_feature_metadata/} for full documentation.
}
\keyword{internal}
