% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_hub}
\alias{sagemaker_describe_hub}
\title{Describe a hub}
\usage{
sagemaker_describe_hub(HubName)
}
\arguments{
\item{HubName}{[required] The name of the hub to describe.}
}
\description{
Describe a hub.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_hub/} for full documentation.
}
\keyword{internal}
