% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_hub_content}
\alias{sagemaker_describe_hub_content}
\title{Describe the content of a hub}
\usage{
sagemaker_describe_hub_content(
  HubName,
  HubContentType,
  HubContentName,
  HubContentVersion = NULL
)
}
\arguments{
\item{HubName}{[required] The name of the hub that contains the content to describe.}

\item{HubContentType}{[required] The type of content in the hub.}

\item{HubContentName}{[required] The name of the content to describe.}

\item{HubContentVersion}{The version of the content to describe.}
}
\description{
Describe the content of a hub.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_hub_content/} for full documentation.
}
\keyword{internal}
