% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_human_task_ui}
\alias{sagemaker_describe_human_task_ui}
\title{Returns information about the requested human task user interface
(worker task template)}
\usage{
sagemaker_describe_human_task_ui(HumanTaskUiName)
}
\arguments{
\item{HumanTaskUiName}{[required] The name of the human task user interface (worker task template) you
want information about.}
}
\description{
Returns information about the requested human task user interface (worker task template).

See \url{https://paws-r.github.io/docs/sagemaker/describe_human_task_ui.html} for full documentation.
}
\keyword{internal}
