% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_image}
\alias{sagemaker_describe_image}
\title{Describes a SageMaker image}
\usage{
sagemaker_describe_image(ImageName)
}
\arguments{
\item{ImageName}{[required] The name of the image to describe.}
}
\description{
Describes a SageMaker image.

See \url{https://paws-r-sdk.com/docs/describe_image/} for full documentation.
}
\keyword{internal}
