% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_inference_component}
\alias{sagemaker_describe_inference_component}
\title{Returns information about an inference component}
\usage{
sagemaker_describe_inference_component(InferenceComponentName)
}
\arguments{
\item{InferenceComponentName}{[required] The name of the inference component.}
}
\description{
Returns information about an inference component.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_inference_component/} for full documentation.
}
\keyword{internal}
