% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_inference_experiment}
\alias{sagemaker_describe_inference_experiment}
\title{Returns details about an inference experiment}
\usage{
sagemaker_describe_inference_experiment(Name)
}
\arguments{
\item{Name}{[required] The name of the inference experiment to describe.}
}
\description{
Returns details about an inference experiment.

See \url{https://paws-r.github.io/docs/sagemaker/describe_inference_experiment.html} for full documentation.
}
\keyword{internal}
