% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_inference_recommendations_job}
\alias{sagemaker_describe_inference_recommendations_job}
\title{Provides the results of the Inference Recommender job}
\usage{
sagemaker_describe_inference_recommendations_job(JobName)
}
\arguments{
\item{JobName}{[required] The name of the job. The name must be unique within an Amazon Web
Services Region in the Amazon Web Services account.}
}
\description{
Provides the results of the Inference Recommender job. One or more recommendation jobs are returned.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_inference_recommendations_job/} for full documentation.
}
\keyword{internal}
