% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_labeling_job}
\alias{sagemaker_describe_labeling_job}
\title{Gets information about a labeling job}
\usage{
sagemaker_describe_labeling_job(LabelingJobName)
}
\arguments{
\item{LabelingJobName}{[required] The name of the labeling job to return information for.}
}
\description{
Gets information about a labeling job.

See \url{https://paws-r.github.io/docs/sagemaker/describe_labeling_job.html} for full documentation.
}
\keyword{internal}
