% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_mlflow_app}
\alias{sagemaker_describe_mlflow_app}
\title{Returns information about an MLflow App}
\usage{
sagemaker_describe_mlflow_app(Arn)
}
\arguments{
\item{Arn}{[required] The ARN of the MLflow App for which to get information.}
}
\description{
Returns information about an MLflow App.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_mlflow_app/} for full documentation.
}
\keyword{internal}
