% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_mlflow_tracking_server}
\alias{sagemaker_describe_mlflow_tracking_server}
\title{Returns information about an MLflow Tracking Server}
\usage{
sagemaker_describe_mlflow_tracking_server(TrackingServerName)
}
\arguments{
\item{TrackingServerName}{[required] The name of the MLflow Tracking Server to describe.}
}
\description{
Returns information about an MLflow Tracking Server.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_mlflow_tracking_server/} for full documentation.
}
\keyword{internal}
