% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_model_card_export_job}
\alias{sagemaker_describe_model_card_export_job}
\title{Describes an Amazon SageMaker Model Card export job}
\usage{
sagemaker_describe_model_card_export_job(ModelCardExportJobArn)
}
\arguments{
\item{ModelCardExportJobArn}{[required] The Amazon Resource Name (ARN) of the model card export job to describe.}
}
\description{
Describes an Amazon SageMaker Model Card export job.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_model_card_export_job/} for full documentation.
}
\keyword{internal}
