% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_monitoring_schedule}
\alias{sagemaker_describe_monitoring_schedule}
\title{Describes the schedule for a monitoring job}
\usage{
sagemaker_describe_monitoring_schedule(MonitoringScheduleName)
}
\arguments{
\item{MonitoringScheduleName}{[required] Name of a previously created monitoring schedule.}
}
\description{
Describes the schedule for a monitoring job.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_monitoring_schedule/} for full documentation.
}
\keyword{internal}
