% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_notebook_instance}
\alias{sagemaker_describe_notebook_instance}
\title{Returns information about a notebook instance}
\usage{
sagemaker_describe_notebook_instance(NotebookInstanceName)
}
\arguments{
\item{NotebookInstanceName}{[required] The name of the notebook instance that you want information about.}
}
\description{
Returns information about a notebook instance.
}
\section{Request syntax}{
\preformatted{svc$describe_notebook_instance(
  NotebookInstanceName = "string"
)
}
}

\keyword{internal}
