% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_optimization_job}
\alias{sagemaker_describe_optimization_job}
\title{Provides the properties of the specified optimization job}
\usage{
sagemaker_describe_optimization_job(OptimizationJobName)
}
\arguments{
\item{OptimizationJobName}{[required] The name that you assigned to the optimization job.}
}
\description{
Provides the properties of the specified optimization job.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_optimization_job/} for full documentation.
}
\keyword{internal}
