% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_partner_app}
\alias{sagemaker_describe_partner_app}
\title{Gets information about a SageMaker Partner AI App}
\usage{
sagemaker_describe_partner_app(Arn, IncludeAvailableUpgrade = NULL)
}
\arguments{
\item{Arn}{[required] The ARN of the SageMaker Partner AI App to describe.}

\item{IncludeAvailableUpgrade}{When set to \code{TRUE}, the response includes available upgrade information
for the SageMaker Partner AI App. Default is \code{FALSE}.}
}
\description{
Gets information about a SageMaker Partner AI App.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_partner_app/} for full documentation.
}
\keyword{internal}
