% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_pipeline}
\alias{sagemaker_describe_pipeline}
\title{Describes the details of a pipeline}
\usage{
sagemaker_describe_pipeline(PipelineName)
}
\arguments{
\item{PipelineName}{[required] The name of the pipeline to describe.}
}
\value{
A list with the following syntax:\preformatted{list(
  PipelineArn = "string",
  PipelineName = "string",
  PipelineDisplayName = "string",
  PipelineDefinition = "string",
  PipelineDescription = "string",
  RoleArn = "string",
  PipelineStatus = "Active",
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  LastModifiedTime = as.POSIXct(
    "2015-01-01"
  ),
  LastRunTime = as.POSIXct(
    "2015-01-01"
  ),
  CreatedBy = list(
    UserProfileArn = "string",
    UserProfileName = "string",
    DomainId = "string"
  ),
  LastModifiedBy = list(
    UserProfileArn = "string",
    UserProfileName = "string",
    DomainId = "string"
  )
)
}
}
\description{
Describes the details of a pipeline.
}
\section{Request syntax}{
\preformatted{svc$describe_pipeline(
  PipelineName = "string"
)
}
}

\keyword{internal}
