% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_pipeline_definition_for_execution}
\alias{sagemaker_describe_pipeline_definition_for_execution}
\title{Describes the details of an execution's pipeline definition}
\usage{
sagemaker_describe_pipeline_definition_for_execution(
  PipelineExecutionArn)
}
\arguments{
\item{PipelineExecutionArn}{[required] The Amazon Resource Name (ARN) of the pipeline execution.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  PipelineDefinition = "string",
  CreationTime = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes the details of an execution's pipeline definition.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_pipeline_definition_for_execution(
  PipelineExecutionArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
