% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_processing_job}
\alias{sagemaker_describe_processing_job}
\title{Returns a description of a processing job}
\usage{
sagemaker_describe_processing_job(ProcessingJobName)
}
\arguments{
\item{ProcessingJobName}{[required] The name of the processing job. The name must be unique within an Amazon
Web Services Region in the Amazon Web Services account.}
}
\description{
Returns a description of a processing job.

See \url{https://paws-r.github.io/docs/sagemaker/describe_processing_job.html} for full documentation.
}
\keyword{internal}
