% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_reserved_capacity}
\alias{sagemaker_describe_reserved_capacity}
\title{Retrieves details about a reserved capacity}
\usage{
sagemaker_describe_reserved_capacity(ReservedCapacityArn)
}
\arguments{
\item{ReservedCapacityArn}{[required] ARN of the reserved capacity to describe.}
}
\description{
Retrieves details about a reserved capacity.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_reserved_capacity/} for full documentation.
}
\keyword{internal}
