% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_studio_lifecycle_config}
\alias{sagemaker_describe_studio_lifecycle_config}
\title{Describes the Amazon SageMaker AI Studio Lifecycle Configuration}
\usage{
sagemaker_describe_studio_lifecycle_config(StudioLifecycleConfigName)
}
\arguments{
\item{StudioLifecycleConfigName}{[required] The name of the Amazon SageMaker AI Studio Lifecycle Configuration to
describe.}
}
\description{
Describes the Amazon SageMaker AI Studio Lifecycle Configuration.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_studio_lifecycle_config/} for full documentation.
}
\keyword{internal}
