% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_subscribed_workteam}
\alias{sagemaker_describe_subscribed_workteam}
\title{Gets information about a work team provided by a vendor}
\usage{
sagemaker_describe_subscribed_workteam(WorkteamArn)
}
\arguments{
\item{WorkteamArn}{[required] The Amazon Resource Name (ARN) of the subscribed work team to describe.}
}
\description{
Gets information about a work team provided by a vendor. It returns details about the subscription with a vendor in the Amazon Web Services Marketplace.

See \url{https://paws-r.github.io/docs/sagemaker/describe_subscribed_workteam.html} for full documentation.
}
\keyword{internal}
