% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_training_job}
\alias{sagemaker_describe_training_job}
\title{Returns information about a training job}
\usage{
sagemaker_describe_training_job(TrainingJobName)
}
\arguments{
\item{TrainingJobName}{[required] The name of the training job.}
}
\description{
Returns information about a training job.

See \url{https://paws-r.github.io/docs/sagemaker/describe_training_job.html} for full documentation.
}
\keyword{internal}
