% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_training_plan}
\alias{sagemaker_describe_training_plan}
\title{Retrieves detailed information about a specific training plan}
\usage{
sagemaker_describe_training_plan(TrainingPlanName)
}
\arguments{
\item{TrainingPlanName}{[required] The name of the training plan to describe.}
}
\description{
Retrieves detailed information about a specific training plan.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_training_plan/} for full documentation.
}
\keyword{internal}
