% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_training_plan_extension_history}
\alias{sagemaker_describe_training_plan_extension_history}
\title{Retrieves the extension history for a specified training plan}
\usage{
sagemaker_describe_training_plan_extension_history(
  TrainingPlanArn,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{TrainingPlanArn}{[required] The Amazon Resource Name (ARN); of the training plan to retrieve
extension history for.}

\item{NextToken}{A token to continue pagination if more results are available.}

\item{MaxResults}{The maximum number of extensions to return in the response.}
}
\description{
Retrieves the extension history for a specified training plan. The response includes details about each extension, such as the offering ID, start and end dates, status, payment status, and cost information.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_training_plan_extension_history/} for full documentation.
}
\keyword{internal}
