% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_trial_component}
\alias{sagemaker_describe_trial_component}
\title{Provides a list of a trials component's properties}
\usage{
sagemaker_describe_trial_component(TrialComponentName)
}
\arguments{
\item{TrialComponentName}{[required] The name of the trial component to describe.}
}
\description{
Provides a list of a trials component's properties.
}
\section{Request syntax}{
\preformatted{svc$describe_trial_component(
  TrialComponentName = "string"
)
}
}

\keyword{internal}
