% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_user_profile}
\alias{sagemaker_describe_user_profile}
\title{Describes a user profile}
\usage{
sagemaker_describe_user_profile(DomainId, UserProfileName)
}
\arguments{
\item{DomainId}{[required] The domain ID.}

\item{UserProfileName}{[required] The user profile name. This value is not case sensitive.}
}
\description{
Describes a user profile. For more information, see \code{\link[=sagemaker_create_user_profile]{create_user_profile}}.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_user_profile/} for full documentation.
}
\keyword{internal}
