% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_disable_sagemaker_servicecatalog_portfolio}
\alias{sagemaker_disable_sagemaker_servicecatalog_portfolio}
\title{Disables using Service Catalog in SageMaker}
\usage{
sagemaker_disable_sagemaker_servicecatalog_portfolio()
}
\value{
An empty list.
}
\description{
Disables using Service Catalog in SageMaker. Service Catalog is used to
create SageMaker projects.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$disable_sagemaker_servicecatalog_portfolio()
}\if{html}{\out{</div>}}
}

\keyword{internal}
