% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_enable_sagemaker_servicecatalog_portfolio}
\alias{sagemaker_enable_sagemaker_servicecatalog_portfolio}
\title{Enables using Service Catalog in SageMaker}
\usage{
sagemaker_enable_sagemaker_servicecatalog_portfolio()
}
\description{
Enables using Service Catalog in SageMaker. Service Catalog is used to create SageMaker projects.

See \url{https://paws-r.github.io/docs/sagemaker/enable_sagemaker_servicecatalog_portfolio.html} for full documentation.
}
\keyword{internal}
