% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_get_lineage_group_policy}
\alias{sagemaker_get_lineage_group_policy}
\title{The resource policy for the lineage group}
\usage{
sagemaker_get_lineage_group_policy(LineageGroupName)
}
\arguments{
\item{LineageGroupName}{[required] The name or Amazon Resource Name (ARN) of the lineage group.}
}
\description{
The resource policy for the lineage group.

See \url{https://paws-r.github.io/docs/sagemaker/get_lineage_group_policy.html} for full documentation.
}
\keyword{internal}
