% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_get_model_package_group_policy}
\alias{sagemaker_get_model_package_group_policy}
\title{Gets a resource policy that manages access for a model group}
\usage{
sagemaker_get_model_package_group_policy(ModelPackageGroupName)
}
\arguments{
\item{ModelPackageGroupName}{[required] The name of the model group for which to get the resource policy.}
}
\description{
Gets a resource policy that manages access for a model group. For information about resource policies, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/access_policies_identity-vs-resource.html}{Identity-based policies and resource-based policies} in the \emph{Amazon Web Services Identity and Access Management User Guide.}.

See \url{https://paws-r.github.io/docs/sagemaker/get_model_package_group_policy.html} for full documentation.
}
\keyword{internal}
