% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_get_search_suggestions}
\alias{sagemaker_get_search_suggestions}
\title{An auto-complete API for the search functionality in the Amazon
SageMaker console}
\usage{
sagemaker_get_search_suggestions(Resource, SuggestionQuery)
}
\arguments{
\item{Resource}{[required] The name of the Amazon SageMaker resource to search for.}

\item{SuggestionQuery}{Limits the property names that are included in the response.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  PropertyNameSuggestions = list(
    list(
      PropertyName = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
An auto-complete API for the search functionality in the Amazon
SageMaker console. It returns suggestions of possible matches for the
property name to use in \code{\link[=sagemaker_search]{search}} queries. Provides
suggestions for \code{HyperParameters}, \code{Tags}, and \code{Metrics}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_search_suggestions(
  Resource = "TrainingJob"|"Experiment"|"ExperimentTrial"|"ExperimentTrialComponent"|"Endpoint"|"ModelPackage"|"ModelPackageGroup"|"Pipeline"|"PipelineExecution"|"FeatureGroup",
  SuggestionQuery = list(
    PropertyNameQuery = list(
      PropertyNameHint = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
