% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_ai_benchmark_jobs}
\alias{sagemaker_list_ai_benchmark_jobs}
\title{Returns a list of AI benchmark jobs in your account}
\usage{
sagemaker_list_ai_benchmark_jobs(
  MaxResults = NULL,
  NextToken = NULL,
  NameContains = NULL,
  StatusEquals = NULL,
  CreationTimeAfter = NULL,
  CreationTimeBefore = NULL,
  SortBy = NULL,
  SortOrder = NULL
)
}
\arguments{
\item{MaxResults}{The maximum number of benchmark jobs to return in the response.}

\item{NextToken}{If the previous call to
\code{\link[=sagemaker_list_ai_benchmark_jobs]{list_ai_benchmark_jobs}} didn't
return the full set of jobs, the call returns a token for getting the
next set.}

\item{NameContains}{A string in the job name. This filter returns only jobs whose name
contains the specified string.}

\item{StatusEquals}{A filter that returns only benchmark jobs with the specified status.}

\item{CreationTimeAfter}{A filter that returns only jobs created after the specified time.}

\item{CreationTimeBefore}{A filter that returns only jobs created before the specified time.}

\item{SortBy}{The field to sort results by. The default is \code{CreationTime}.}

\item{SortOrder}{The sort order for results. The default is \code{Descending}.}
}
\description{
Returns a list of AI benchmark jobs in your account. You can filter the results by name, status, and creation time, and sort the results. The response is paginated.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_list_ai_benchmark_jobs/} for full documentation.
}
\keyword{internal}
