% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_ai_workload_configs}
\alias{sagemaker_list_ai_workload_configs}
\title{Returns a list of AI workload configurations in your account}
\usage{
sagemaker_list_ai_workload_configs(
  MaxResults = NULL,
  NextToken = NULL,
  NameContains = NULL,
  CreationTimeAfter = NULL,
  CreationTimeBefore = NULL,
  SortBy = NULL,
  SortOrder = NULL
)
}
\arguments{
\item{MaxResults}{The maximum number of AI workload configurations to return in the
response.}

\item{NextToken}{If the previous call to
\code{\link[=sagemaker_list_ai_workload_configs]{list_ai_workload_configs}} didn't
return the full set of configurations, the call returns a token for
getting the next set of configurations.}

\item{NameContains}{A string in the configuration name. This filter returns only
configurations whose name contains the specified string.}

\item{CreationTimeAfter}{A filter that returns only configurations created after the specified
time.}

\item{CreationTimeBefore}{A filter that returns only configurations created before the specified
time.}

\item{SortBy}{The field to sort results by. The default is \code{CreationTime}.}

\item{SortOrder}{The sort order for results. The default is \code{Descending}.}
}
\description{
Returns a list of AI workload configurations in your account. You can filter the results by name and creation time, and sort the results. The response is paginated.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_list_ai_workload_configs/} for full documentation.
}
\keyword{internal}
