% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_algorithms}
\alias{sagemaker_list_algorithms}
\title{Lists the machine learning algorithms that have been created}
\usage{
sagemaker_list_algorithms(
  CreationTimeAfter = NULL,
  CreationTimeBefore = NULL,
  MaxResults = NULL,
  NameContains = NULL,
  NextToken = NULL,
  SortBy = NULL,
  SortOrder = NULL
)
}
\arguments{
\item{CreationTimeAfter}{A filter that returns only algorithms created after the specified time
(timestamp).}

\item{CreationTimeBefore}{A filter that returns only algorithms created before the specified time
(timestamp).}

\item{MaxResults}{The maximum number of algorithms to return in the response.}

\item{NameContains}{A string in the algorithm name. This filter returns only algorithms
whose name contains the specified string.}

\item{NextToken}{If the response to a previous
\code{\link[=sagemaker_list_algorithms]{list_algorithms}} request was truncated,
the response includes a \code{NextToken}. To retrieve the next set of
algorithms, use the token in the next request.}

\item{SortBy}{The parameter by which to sort the results. The default is
\code{CreationTime}.}

\item{SortOrder}{The sort order for the results. The default is \code{Ascending}.}
}
\description{
Lists the machine learning algorithms that have been created.

See \url{https://paws-r.github.io/docs/sagemaker/list_algorithms.html} for full documentation.
}
\keyword{internal}
