% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_aliases}
\alias{sagemaker_list_aliases}
\title{Lists the aliases of a specified image or image version}
\usage{
sagemaker_list_aliases(
  ImageName,
  Alias = NULL,
  Version = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ImageName}{[required] The name of the image.}

\item{Alias}{The alias of the image version.}

\item{Version}{The version of the image. If image version is not specified, the aliases
of all versions of the image are listed.}

\item{MaxResults}{The maximum number of aliases to return.}

\item{NextToken}{If the previous call to \code{\link[=sagemaker_list_aliases]{list_aliases}} didn't
return the full set of aliases, the call returns a token for retrieving
the next set of aliases.}
}
\description{
Lists the aliases of a specified image or image version.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_list_aliases/} for full documentation.
}
\keyword{internal}
