% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_associations}
\alias{sagemaker_list_associations}
\title{Lists the associations in your account and their properties}
\usage{
sagemaker_list_associations(SourceArn, DestinationArn, SourceType,
  DestinationType, AssociationType, CreatedAfter, CreatedBefore, SortBy,
  SortOrder, NextToken, MaxResults)
}
\arguments{
\item{SourceArn}{A filter that returns only associations with the specified source ARN.}

\item{DestinationArn}{A filter that returns only associations with the specified destination
Amazon Resource Name (ARN).}

\item{SourceType}{A filter that returns only associations with the specified source type.}

\item{DestinationType}{A filter that returns only associations with the specified destination
type.}

\item{AssociationType}{A filter that returns only associations of the specified type.}

\item{CreatedAfter}{A filter that returns only associations created on or after the
specified time.}

\item{CreatedBefore}{A filter that returns only associations created on or before the
specified time.}

\item{SortBy}{The property used to sort results. The default value is \code{CreationTime}.}

\item{SortOrder}{The sort order. The default value is \code{Descending}.}

\item{NextToken}{If the previous call to
\code{\link[=sagemaker_list_associations]{list_associations}} didn't return the
full set of associations, the call returns a token for getting the next
set of associations.}

\item{MaxResults}{The maximum number of associations to return in the response. The
default value is 10.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AssociationSummaries = list(
    list(
      SourceArn = "string",
      DestinationArn = "string",
      SourceType = "string",
      DestinationType = "string",
      AssociationType = "ContributedTo"|"AssociatedWith"|"DerivedFrom"|"Produced",
      SourceName = "string",
      DestinationName = "string",
      CreationTime = as.POSIXct(
        "2015-01-01"
      ),
      CreatedBy = list(
        UserProfileArn = "string",
        UserProfileName = "string",
        DomainId = "string"
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the associations in your account and their properties.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_associations(
  SourceArn = "string",
  DestinationArn = "string",
  SourceType = "string",
  DestinationType = "string",
  AssociationType = "ContributedTo"|"AssociatedWith"|"DerivedFrom"|"Produced",
  CreatedAfter = as.POSIXct(
    "2015-01-01"
  ),
  CreatedBefore = as.POSIXct(
    "2015-01-01"
  ),
  SortBy = "SourceArn"|"DestinationArn"|"SourceType"|"DestinationType"|"CreationTime",
  SortOrder = "Ascending"|"Descending",
  NextToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
