% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_candidates_for_auto_ml_job}
\alias{sagemaker_list_candidates_for_auto_ml_job}
\title{List the candidates created for the job}
\usage{
sagemaker_list_candidates_for_auto_ml_job(
  AutoMLJobName,
  StatusEquals = NULL,
  CandidateNameEquals = NULL,
  SortOrder = NULL,
  SortBy = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AutoMLJobName}{[required] List the candidates created for the job by providing the job's name.}

\item{StatusEquals}{List the candidates for the job and filter by status.}

\item{CandidateNameEquals}{List the candidates for the job and filter by candidate name.}

\item{SortOrder}{The sort order for the results. The default is \code{Ascending}.}

\item{SortBy}{The parameter by which to sort the results. The default is \code{Descending}.}

\item{MaxResults}{List the job's candidates up to a specified limit.}

\item{NextToken}{If the previous response was truncated, you receive this token. Use it
in your next request to receive the next set of results.}
}
\description{
List the candidates created for the job.

See \url{https://paws-r.github.io/docs/sagemaker/list_candidates_for_auto_ml_job.html} for full documentation.
}
\keyword{internal}
