% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_cluster_scheduler_configs}
\alias{sagemaker_list_cluster_scheduler_configs}
\title{List the cluster policy configurations}
\usage{
sagemaker_list_cluster_scheduler_configs(
  CreatedAfter = NULL,
  CreatedBefore = NULL,
  NameContains = NULL,
  ClusterArn = NULL,
  Status = NULL,
  SortBy = NULL,
  SortOrder = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{CreatedAfter}{Filter for after this creation time. The input for this parameter is a
Unix timestamp. To convert a date and time into a Unix timestamp, see
\href{https://www.epochconverter.com/}{EpochConverter}.}

\item{CreatedBefore}{Filter for before this creation time. The input for this parameter is a
Unix timestamp. To convert a date and time into a Unix timestamp, see
\href{https://www.epochconverter.com/}{EpochConverter}.}

\item{NameContains}{Filter for name containing this string.}

\item{ClusterArn}{Filter for ARN of the cluster.}

\item{Status}{Filter for status.}

\item{SortBy}{Filter for sorting the list by a given value. For example, sort by name,
creation time, or status.}

\item{SortOrder}{The order of the list. By default, listed in \code{Descending} order
according to by \code{SortBy}. To change the list order, you can specify
\code{SortOrder} to be \code{Ascending}.}

\item{NextToken}{If the previous response was truncated, you will receive this token. Use
it in your next request to receive the next set of results.}

\item{MaxResults}{The maximum number of cluster policies to list.}
}
\description{
List the cluster policy configurations.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_list_cluster_scheduler_configs/} for full documentation.
}
\keyword{internal}
