% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_compilation_jobs}
\alias{sagemaker_list_compilation_jobs}
\title{Lists model compilation jobs that satisfy various filters}
\usage{
sagemaker_list_compilation_jobs(NextToken, MaxResults,
  CreationTimeAfter, CreationTimeBefore, LastModifiedTimeAfter,
  LastModifiedTimeBefore, NameContains, StatusEquals, SortBy, SortOrder)
}
\arguments{
\item{NextToken}{If the result of the previous
\code{\link[=sagemaker_list_compilation_jobs]{list_compilation_jobs}} request was
truncated, the response includes a \code{NextToken}. To retrieve the next set
of model compilation jobs, use the token in the next request.}

\item{MaxResults}{The maximum number of model compilation jobs to return in the response.}

\item{CreationTimeAfter}{A filter that returns the model compilation jobs that were created after
a specified time.}

\item{CreationTimeBefore}{A filter that returns the model compilation jobs that were created
before a specified time.}

\item{LastModifiedTimeAfter}{A filter that returns the model compilation jobs that were modified
after a specified time.}

\item{LastModifiedTimeBefore}{A filter that returns the model compilation jobs that were modified
before a specified time.}

\item{NameContains}{A filter that returns the model compilation jobs whose name contains a
specified string.}

\item{StatusEquals}{A filter that retrieves model compilation jobs with a specific
DescribeCompilationJobResponse$CompilationJobStatus status.}

\item{SortBy}{The field by which to sort results. The default is \code{CreationTime}.}

\item{SortOrder}{The sort order for results. The default is \code{Ascending}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  CompilationJobSummaries = list(
    list(
      CompilationJobName = "string",
      CompilationJobArn = "string",
      CreationTime = as.POSIXct(
        "2015-01-01"
      ),
      CompilationStartTime = as.POSIXct(
        "2015-01-01"
      ),
      CompilationEndTime = as.POSIXct(
        "2015-01-01"
      ),
      CompilationTargetDevice = "lambda"|"ml_m4"|"ml_m5"|"ml_c4"|"ml_c5"|"ml_p2"|"ml_p3"|"ml_g4dn"|"ml_inf1"|"jetson_tx1"|"jetson_tx2"|"jetson_nano"|"jetson_xavier"|"rasp3b"|"imx8qm"|"deeplens"|"rk3399"|"rk3288"|"aisage"|"sbe_c"|"qcs605"|"qcs603"|"sitara_am57x"|"amba_cv22"|"x86_win32"|"x86_win64"|"coreml"|"jacinto_tda4vm",
      CompilationTargetPlatformOs = "ANDROID"|"LINUX",
      CompilationTargetPlatformArch = "X86_64"|"X86"|"ARM64"|"ARM_EABI"|"ARM_EABIHF",
      CompilationTargetPlatformAccelerator = "INTEL_GRAPHICS"|"MALI"|"NVIDIA",
      LastModifiedTime = as.POSIXct(
        "2015-01-01"
      ),
      CompilationJobStatus = "INPROGRESS"|"COMPLETED"|"FAILED"|"STARTING"|"STOPPING"|"STOPPED"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists model compilation jobs that satisfy various filters.

To create a model compilation job, use
\code{\link[=sagemaker_create_compilation_job]{create_compilation_job}}. To get
information about a particular model compilation job you have created,
use \code{\link[=sagemaker_describe_compilation_job]{describe_compilation_job}}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_compilation_jobs(
  NextToken = "string",
  MaxResults = 123,
  CreationTimeAfter = as.POSIXct(
    "2015-01-01"
  ),
  CreationTimeBefore = as.POSIXct(
    "2015-01-01"
  ),
  LastModifiedTimeAfter = as.POSIXct(
    "2015-01-01"
  ),
  LastModifiedTimeBefore = as.POSIXct(
    "2015-01-01"
  ),
  NameContains = "string",
  StatusEquals = "INPROGRESS"|"COMPLETED"|"FAILED"|"STARTING"|"STOPPING"|"STOPPED",
  SortBy = "Name"|"CreationTime"|"Status",
  SortOrder = "Ascending"|"Descending"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
