% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_devices}
\alias{sagemaker_list_devices}
\title{A list of devices}
\usage{
sagemaker_list_devices(
  NextToken = NULL,
  MaxResults = NULL,
  LatestHeartbeatAfter = NULL,
  ModelName = NULL,
  DeviceFleetName = NULL
)
}
\arguments{
\item{NextToken}{The response from the last list when returning a list large enough to
need tokening.}

\item{MaxResults}{Maximum number of results to select.}

\item{LatestHeartbeatAfter}{Select fleets where the job was updated after X}

\item{ModelName}{A filter that searches devices that contains this name in any of their
models.}

\item{DeviceFleetName}{Filter for fleets containing this name in their device fleet name.}
}
\description{
A list of devices.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_list_devices/} for full documentation.
}
\keyword{internal}
