% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_edge_packaging_jobs}
\alias{sagemaker_list_edge_packaging_jobs}
\title{Returns a list of edge packaging jobs}
\usage{
sagemaker_list_edge_packaging_jobs(
  NextToken = NULL,
  MaxResults = NULL,
  CreationTimeAfter = NULL,
  CreationTimeBefore = NULL,
  LastModifiedTimeAfter = NULL,
  LastModifiedTimeBefore = NULL,
  NameContains = NULL,
  ModelNameContains = NULL,
  StatusEquals = NULL,
  SortBy = NULL,
  SortOrder = NULL
)
}
\arguments{
\item{NextToken}{The response from the last list when returning a list large enough to
need tokening.}

\item{MaxResults}{Maximum number of results to select.}

\item{CreationTimeAfter}{Select jobs where the job was created after specified time.}

\item{CreationTimeBefore}{Select jobs where the job was created before specified time.}

\item{LastModifiedTimeAfter}{Select jobs where the job was updated after specified time.}

\item{LastModifiedTimeBefore}{Select jobs where the job was updated before specified time.}

\item{NameContains}{Filter for jobs containing this name in their packaging job name.}

\item{ModelNameContains}{Filter for jobs where the model name contains this string.}

\item{StatusEquals}{The job status to filter for.}

\item{SortBy}{Use to specify what column to sort by.}

\item{SortOrder}{What direction to sort by.}
}
\description{
Returns a list of edge packaging jobs.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_list_edge_packaging_jobs/} for full documentation.
}
\keyword{internal}
