% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_inference_recommendations_job_steps}
\alias{sagemaker_list_inference_recommendations_job_steps}
\title{Returns a list of the subtasks for an Inference Recommender job}
\usage{
sagemaker_list_inference_recommendations_job_steps(
  JobName,
  Status = NULL,
  StepType = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{JobName}{[required] The name for the Inference Recommender job.}

\item{Status}{A filter to return benchmarks of a specified status. If this field is
left empty, then all benchmarks are returned.}

\item{StepType}{A filter to return details about the specified type of subtask.

\code{BENCHMARK}: Evaluate the performance of your model on different
instance types.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{A token that you can specify to return more results from the list.
Specify this field if you have a token that was returned from a previous
request.}
}
\description{
Returns a list of the subtasks for an Inference Recommender job.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_list_inference_recommendations_job_steps/} for full documentation.
}
\keyword{internal}
