% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_job_schema_versions}
\alias{sagemaker_list_job_schema_versions}
\title{Lists available configuration schema versions for a specified job
category}
\usage{
sagemaker_list_job_schema_versions(
  JobCategory,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{JobCategory}{[required] The category of job schemas to list.}

\item{NextToken}{If the previous response was truncated, this token retrieves the next set of results.}

\item{MaxResults}{The maximum number of schema versions to return in the response. The default value is 5.}
}
\description{
Lists available configuration schema versions for a specified job category. Use the schema versions with \code{\link[=sagemaker_describe_job_schema_version]{describe_job_schema_version}} to retrieve the full schema document.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_list_job_schema_versions/} for full documentation.
}
\keyword{internal}
