% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_jobs}
\alias{sagemaker_list_jobs}
\title{Lists jobs in a specified category}
\usage{
sagemaker_list_jobs(
  JobCategory,
  NextToken = NULL,
  MaxResults = NULL,
  CreationTimeAfter = NULL,
  CreationTimeBefore = NULL,
  LastModifiedTimeAfter = NULL,
  LastModifiedTimeBefore = NULL,
  NameContains = NULL,
  SortBy = NULL,
  SortOrder = NULL,
  StatusEquals = NULL
)
}
\arguments{
\item{JobCategory}{[required] The category of jobs to list.}

\item{NextToken}{If the previous response was truncated, this token retrieves the next set of results.}

\item{MaxResults}{The maximum number of jobs to return in the response. The default value is 50.}

\item{CreationTimeAfter}{A filter that returns only jobs created after the specified time.}

\item{CreationTimeBefore}{A filter that returns only jobs created before the specified time.}

\item{LastModifiedTimeAfter}{A filter that returns only jobs modified after the specified time.}

\item{LastModifiedTimeBefore}{A filter that returns only jobs modified before the specified time.}

\item{NameContains}{A string in the job name to filter results. Only jobs whose name contains the specified string are returned.}

\item{SortBy}{The field to sort results by.}

\item{SortOrder}{The sort order for results. Valid values are \code{Ascending} and \code{Descending}.}

\item{StatusEquals}{A filter that returns only jobs with the specified status.}
}
\description{
Lists jobs in a specified category. You can filter results by creation time, last modified time, name, and status. Results are sorted by the field you specify in \code{SortBy}. Use pagination to retrieve large result sets efficiently.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_list_jobs/} for full documentation.
}
\keyword{internal}
