% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_mlflow_apps}
\alias{sagemaker_list_mlflow_apps}
\title{Lists all MLflow Apps}
\usage{
sagemaker_list_mlflow_apps(
  CreatedAfter = NULL,
  CreatedBefore = NULL,
  Status = NULL,
  MlflowVersion = NULL,
  DefaultForDomainId = NULL,
  AccountDefaultStatus = NULL,
  SortBy = NULL,
  SortOrder = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{CreatedAfter}{Use the \code{CreatedAfter} filter to only list MLflow Apps created after a
specific date and time. Listed MLflow Apps are shown with a date and
time such as \code{"2024-03-16T01:46:56+00:00"}. The \code{CreatedAfter} parameter
takes in a Unix timestamp.}

\item{CreatedBefore}{Use the \code{CreatedBefore} filter to only list MLflow Apps created before a
specific date and time. Listed MLflow Apps are shown with a date and
time such as \code{"2024-03-16T01:46:56+00:00"}. The \code{CreatedAfter} parameter
takes in a Unix timestamp.}

\item{Status}{Filter for Mlflow apps with a specific creation status.}

\item{MlflowVersion}{Filter for Mlflow Apps with the specified version.}

\item{DefaultForDomainId}{Filter for MLflow Apps with the specified default SageMaker Domain ID.}

\item{AccountDefaultStatus}{Filter for MLflow Apps with the specified \code{AccountDefaultStatus}.}

\item{SortBy}{Filter for MLflow Apps sorting by name, creation time, or creation
status.}

\item{SortOrder}{Change the order of the listed MLflow Apps. By default, MLflow Apps are
listed in \code{Descending} order by creation time. To change the list order,
specify \code{SortOrder} to be \code{Ascending}.}

\item{NextToken}{If the previous response was truncated, use this token in your next
request to receive the next set of results.}

\item{MaxResults}{The maximum number of MLflow Apps to list.}
}
\description{
Lists all MLflow Apps

See \url{https://www.paws-r-sdk.com/docs/sagemaker_list_mlflow_apps/} for full documentation.
}
\keyword{internal}
