% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_model_metadata}
\alias{sagemaker_list_model_metadata}
\title{Lists the domain, framework, task, and model name of standard machine
learning models found in common model zoos}
\usage{
sagemaker_list_model_metadata(
  SearchExpression = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{SearchExpression}{One or more filters that searches for the specified resource or
resources in a search. All resource objects that satisfy the
expression's condition are included in the search results. Specify the
Framework, FrameworkVersion, Domain or Task to filter supported. Filter
names and values are case-sensitive.}

\item{NextToken}{If the response to a previous \code{ListModelMetadataResponse} request was
truncated, the response includes a NextToken. To retrieve the next set
of model metadata, use the token in the next request.}

\item{MaxResults}{The maximum number of models to return in the response.}
}
\description{
Lists the domain, framework, task, and model name of standard machine learning models found in common model zoos.

See \url{https://paws-r.github.io/docs/sagemaker/list_model_metadata.html} for full documentation.
}
\keyword{internal}
