% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_monitoring_alerts}
\alias{sagemaker_list_monitoring_alerts}
\title{Gets the alerts for a single monitoring schedule}
\usage{
sagemaker_list_monitoring_alerts(
  MonitoringScheduleName,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{MonitoringScheduleName}{[required] The name of a monitoring schedule.}

\item{NextToken}{If the result of the previous
\code{\link[=sagemaker_list_monitoring_alerts]{list_monitoring_alerts}} request was
truncated, the response includes a \code{NextToken}. To retrieve the next set
of alerts in the history, use the token in the next request.}

\item{MaxResults}{The maximum number of results to display. The default is 100.}
}
\description{
Gets the alerts for a single monitoring schedule.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_list_monitoring_alerts/} for full documentation.
}
\keyword{internal}
