% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_notebook_instance_lifecycle_configs}
\alias{sagemaker_list_notebook_instance_lifecycle_configs}
\title{Lists notebook instance lifestyle configurations created with the
CreateNotebookInstanceLifecycleConfig API}
\usage{
sagemaker_list_notebook_instance_lifecycle_configs(NextToken,
  MaxResults, SortBy, SortOrder, NameContains, CreationTimeBefore,
  CreationTimeAfter, LastModifiedTimeBefore, LastModifiedTimeAfter)
}
\arguments{
\item{NextToken}{If the result of a \code{ListNotebookInstanceLifecycleConfigs} request was
truncated, the response includes a \code{NextToken}. To get the next set of
lifecycle configurations, use the token in the next request.}

\item{MaxResults}{The maximum number of lifecycle configurations to return in the
response.}

\item{SortBy}{Sorts the list of results. The default is \code{CreationTime}.}

\item{SortOrder}{The sort order for results.}

\item{NameContains}{A string in the lifecycle configuration name. This filter returns only
lifecycle configurations whose name contains the specified string.}

\item{CreationTimeBefore}{A filter that returns only lifecycle configurations that were created
before the specified time (timestamp).}

\item{CreationTimeAfter}{A filter that returns only lifecycle configurations that were created
after the specified time (timestamp).}

\item{LastModifiedTimeBefore}{A filter that returns only lifecycle configurations that were modified
before the specified time (timestamp).}

\item{LastModifiedTimeAfter}{A filter that returns only lifecycle configurations that were modified
after the specified time (timestamp).}
}
\description{
Lists notebook instance lifestyle configurations created with the
CreateNotebookInstanceLifecycleConfig API.
}
\section{Request syntax}{
\preformatted{svc$list_notebook_instance_lifecycle_configs(
  NextToken = "string",
  MaxResults = 123,
  SortBy = "Name"|"CreationTime"|"LastModifiedTime",
  SortOrder = "Ascending"|"Descending",
  NameContains = "string",
  CreationTimeBefore = as.POSIXct(
    "2015-01-01"
  ),
  CreationTimeAfter = as.POSIXct(
    "2015-01-01"
  ),
  LastModifiedTimeBefore = as.POSIXct(
    "2015-01-01"
  ),
  LastModifiedTimeAfter = as.POSIXct(
    "2015-01-01"
  )
)
}
}

\keyword{internal}
