% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_notebook_instances}
\alias{sagemaker_list_notebook_instances}
\title{Returns a list of the Amazon SageMaker notebook instances in the
requester's account in an AWS Region}
\usage{
sagemaker_list_notebook_instances(NextToken, MaxResults, SortBy,
  SortOrder, NameContains, CreationTimeBefore, CreationTimeAfter,
  LastModifiedTimeBefore, LastModifiedTimeAfter, StatusEquals,
  NotebookInstanceLifecycleConfigNameContains,
  DefaultCodeRepositoryContains, AdditionalCodeRepositoryEquals)
}
\arguments{
\item{NextToken}{If the previous call to the
\code{\link[=sagemaker_list_notebook_instances]{list_notebook_instances}} is
truncated, the response includes a \code{NextToken}. You can use this token
in your subsequent
\code{\link[=sagemaker_list_notebook_instances]{list_notebook_instances}} request
to fetch the next set of notebook instances.

You might specify a filter or a sort order in your request. When
response is truncated, you must use the same values for the filer and
sort order in the next request.}

\item{MaxResults}{The maximum number of notebook instances to return.}

\item{SortBy}{The field to sort results by. The default is \code{Name}.}

\item{SortOrder}{The sort order for results.}

\item{NameContains}{A string in the notebook instances' name. This filter returns only
notebook instances whose name contains the specified string.}

\item{CreationTimeBefore}{A filter that returns only notebook instances that were created before
the specified time (timestamp).}

\item{CreationTimeAfter}{A filter that returns only notebook instances that were created after
the specified time (timestamp).}

\item{LastModifiedTimeBefore}{A filter that returns only notebook instances that were modified before
the specified time (timestamp).}

\item{LastModifiedTimeAfter}{A filter that returns only notebook instances that were modified after
the specified time (timestamp).}

\item{StatusEquals}{A filter that returns only notebook instances with the specified status.}

\item{NotebookInstanceLifecycleConfigNameContains}{A string in the name of a notebook instances lifecycle configuration
associated with this notebook instance. This filter returns only
notebook instances associated with a lifecycle configuration with a name
that contains the specified string.}

\item{DefaultCodeRepositoryContains}{A string in the name or URL of a Git repository associated with this
notebook instance. This filter returns only notebook instances
associated with a git repository with a name that contains the specified
string.}

\item{AdditionalCodeRepositoryEquals}{A filter that returns only notebook instances with associated with the
specified git repository.}
}
\value{
A list with the following syntax:\preformatted{list(
  NextToken = "string",
  NotebookInstances = list(
    list(
      NotebookInstanceName = "string",
      NotebookInstanceArn = "string",
      NotebookInstanceStatus = "Pending"|"InService"|"Stopping"|"Stopped"|"Failed"|"Deleting"|"Updating",
      Url = "string",
      InstanceType = "ml.t2.medium"|"ml.t2.large"|"ml.t2.xlarge"|"ml.t2.2xlarge"|"ml.t3.medium"|"ml.t3.large"|"ml.t3.xlarge"|"ml.t3.2xlarge"|"ml.m4.xlarge"|"ml.m4.2xlarge"|"ml.m4.4xlarge"|"ml.m4.10xlarge"|"ml.m4.16xlarge"|"ml.m5.xlarge"|"ml.m5.2xlarge"|"ml.m5.4xlarge"|"ml.m5.12xlarge"|"ml.m5.24xlarge"|"ml.c4.xlarge"|"ml.c4.2xlarge"|"ml.c4.4xlarge"|"ml.c4.8xlarge"|"ml.c5.xlarge"|"ml.c5.2xlarge"|"ml.c5.4xlarge"|"ml.c5.9xlarge"|"ml.c5.18xlarge"|"ml.c5d.xlarge"|"ml.c5d.2xlarge"|"ml.c5d.4xlarge"|"ml.c5d.9xlarge"|"ml.c5d.18xlarge"|"ml.p2.xlarge"|"ml.p2.8xlarge"|"ml.p2.16xlarge"|"ml.p3.2xlarge"|"ml.p3.8xlarge"|"ml.p3.16xlarge",
      CreationTime = as.POSIXct(
        "2015-01-01"
      ),
      LastModifiedTime = as.POSIXct(
        "2015-01-01"
      ),
      NotebookInstanceLifecycleConfigName = "string",
      DefaultCodeRepository = "string",
      AdditionalCodeRepositories = list(
        "string"
      )
    )
  )
)
}
}
\description{
Returns a list of the Amazon SageMaker notebook instances in the
requester's account in an AWS Region.
}
\section{Request syntax}{
\preformatted{svc$list_notebook_instances(
  NextToken = "string",
  MaxResults = 123,
  SortBy = "Name"|"CreationTime"|"Status",
  SortOrder = "Ascending"|"Descending",
  NameContains = "string",
  CreationTimeBefore = as.POSIXct(
    "2015-01-01"
  ),
  CreationTimeAfter = as.POSIXct(
    "2015-01-01"
  ),
  LastModifiedTimeBefore = as.POSIXct(
    "2015-01-01"
  ),
  LastModifiedTimeAfter = as.POSIXct(
    "2015-01-01"
  ),
  StatusEquals = "Pending"|"InService"|"Stopping"|"Stopped"|"Failed"|"Deleting"|"Updating",
  NotebookInstanceLifecycleConfigNameContains = "string",
  DefaultCodeRepositoryContains = "string",
  AdditionalCodeRepositoryEquals = "string"
)
}
}

\keyword{internal}
