% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_pipeline_versions}
\alias{sagemaker_list_pipeline_versions}
\title{Gets a list of all versions of the pipeline}
\usage{
sagemaker_list_pipeline_versions(
  PipelineName,
  CreatedAfter = NULL,
  CreatedBefore = NULL,
  SortOrder = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{PipelineName}{[required] The Amazon Resource Name (ARN) of the pipeline.}

\item{CreatedAfter}{A filter that returns the pipeline versions that were created after a
specified time.}

\item{CreatedBefore}{A filter that returns the pipeline versions that were created before a
specified time.}

\item{SortOrder}{The sort order for the results.}

\item{NextToken}{If the result of the previous
\code{\link[=sagemaker_list_pipeline_versions]{list_pipeline_versions}} request was
truncated, the response includes a \code{NextToken}. To retrieve the next set
of pipeline versions, use this token in your next request.}

\item{MaxResults}{The maximum number of pipeline versions to return in the response.}
}
\description{
Gets a list of all versions of the pipeline.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_list_pipeline_versions/} for full documentation.
}
\keyword{internal}
