% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_pipelines}
\alias{sagemaker_list_pipelines}
\title{Gets a list of pipelines}
\usage{
sagemaker_list_pipelines(PipelineNamePrefix, CreatedAfter,
  CreatedBefore, SortBy, SortOrder, NextToken, MaxResults)
}
\arguments{
\item{PipelineNamePrefix}{The prefix of the pipeline name.}

\item{CreatedAfter}{A filter that returns the pipelines that were created after a specified
time.}

\item{CreatedBefore}{A filter that returns the pipelines that were created before a specified
time.}

\item{SortBy}{The field by which to sort results. The default is \code{CreatedTime}.}

\item{SortOrder}{The sort order for results.}

\item{NextToken}{If the result of the previous
\code{\link[=sagemaker_list_pipelines]{list_pipelines}} request was truncated, the
response includes a \code{NextToken}. To retrieve the next set of pipelines,
use the token in the next request.}

\item{MaxResults}{The maximum number of pipelines to return in the response.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  PipelineSummaries = list(
    list(
      PipelineArn = "string",
      PipelineName = "string",
      PipelineDisplayName = "string",
      PipelineDescription = "string",
      RoleArn = "string",
      CreationTime = as.POSIXct(
        "2015-01-01"
      ),
      LastModifiedTime = as.POSIXct(
        "2015-01-01"
      ),
      LastExecutionTime = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets a list of pipelines.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_pipelines(
  PipelineNamePrefix = "string",
  CreatedAfter = as.POSIXct(
    "2015-01-01"
  ),
  CreatedBefore = as.POSIXct(
    "2015-01-01"
  ),
  SortBy = "Name"|"CreationTime",
  SortOrder = "Ascending"|"Descending",
  NextToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
