% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_spaces}
\alias{sagemaker_list_spaces}
\title{Lists spaces}
\usage{
sagemaker_list_spaces(
  NextToken = NULL,
  MaxResults = NULL,
  SortOrder = NULL,
  SortBy = NULL,
  DomainIdEquals = NULL,
  SpaceNameContains = NULL
)
}
\arguments{
\item{NextToken}{If the previous response was truncated, you will receive this token. Use
it in your next request to receive the next set of results.}

\item{MaxResults}{This parameter defines the maximum number of results that can be return
in a single response. The \code{MaxResults} parameter is an upper bound, not
a target. If there are more results available than the value specified,
a \code{NextToken} is provided in the response. The \code{NextToken} indicates
that the user should get the next set of results by providing this token
as a part of a subsequent call. The default value for \code{MaxResults} is
10.}

\item{SortOrder}{The sort order for the results. The default is \code{Ascending}.}

\item{SortBy}{The parameter by which to sort the results. The default is
\code{CreationTime}.}

\item{DomainIdEquals}{A parameter to search for the domain ID.}

\item{SpaceNameContains}{A parameter by which to filter the results.}
}
\description{
Lists spaces.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_list_spaces/} for full documentation.
}
\keyword{internal}
