% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_stage_devices}
\alias{sagemaker_list_stage_devices}
\title{Lists devices allocated to the stage, containing detailed device
information and deployment status}
\usage{
sagemaker_list_stage_devices(
  NextToken = NULL,
  MaxResults = NULL,
  EdgeDeploymentPlanName,
  ExcludeDevicesDeployedInOtherStage = NULL,
  StageName
)
}
\arguments{
\item{NextToken}{The response from the last list when returning a list large enough to
neeed tokening.}

\item{MaxResults}{The maximum number of requests to select.}

\item{EdgeDeploymentPlanName}{[required] The name of the edge deployment plan.}

\item{ExcludeDevicesDeployedInOtherStage}{Toggle for excluding devices deployed in other stages.}

\item{StageName}{[required] The name of the stage in the deployment.}
}
\description{
Lists devices allocated to the stage, containing detailed device information and deployment status.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_list_stage_devices/} for full documentation.
}
\keyword{internal}
