% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_training_jobs}
\alias{sagemaker_list_training_jobs}
\title{Lists training jobs}
\usage{
sagemaker_list_training_jobs(
  NextToken = NULL,
  MaxResults = NULL,
  CreationTimeAfter = NULL,
  CreationTimeBefore = NULL,
  LastModifiedTimeAfter = NULL,
  LastModifiedTimeBefore = NULL,
  NameContains = NULL,
  StatusEquals = NULL,
  SortBy = NULL,
  SortOrder = NULL,
  WarmPoolStatusEquals = NULL,
  TrainingPlanArnEquals = NULL
)
}
\arguments{
\item{NextToken}{If the result of the previous
\code{\link[=sagemaker_list_training_jobs]{list_training_jobs}} request was
truncated, the response includes a \code{NextToken}. To retrieve the next set
of training jobs, use the token in the next request.}

\item{MaxResults}{The maximum number of training jobs to return in the response.}

\item{CreationTimeAfter}{A filter that returns only training jobs created after the specified
time (timestamp).}

\item{CreationTimeBefore}{A filter that returns only training jobs created before the specified
time (timestamp).}

\item{LastModifiedTimeAfter}{A filter that returns only training jobs modified after the specified
time (timestamp).}

\item{LastModifiedTimeBefore}{A filter that returns only training jobs modified before the specified
time (timestamp).}

\item{NameContains}{A string in the training job name. This filter returns only training
jobs whose name contains the specified string.}

\item{StatusEquals}{A filter that retrieves only training jobs with a specific status.}

\item{SortBy}{The field to sort results by. The default is \code{CreationTime}.}

\item{SortOrder}{The sort order for results. The default is \code{Ascending}.}

\item{WarmPoolStatusEquals}{A filter that retrieves only training jobs with a specific warm pool
status.}

\item{TrainingPlanArnEquals}{The Amazon Resource Name (ARN); of the training plan to filter training
jobs by. For more information about reserving GPU capacity for your
SageMaker training jobs using Amazon SageMaker Training Plan, see
\code{\link[=sagemaker_create_training_plan]{create_training_plan}}.}
}
\description{
Lists training jobs.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_list_training_jobs/} for full documentation.
}
\keyword{internal}
