% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_ultra_servers_by_reserved_capacity}
\alias{sagemaker_list_ultra_servers_by_reserved_capacity}
\title{Lists all UltraServers that are part of a specified reserved capacity}
\usage{
sagemaker_list_ultra_servers_by_reserved_capacity(
  ReservedCapacityArn,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ReservedCapacityArn}{[required] The ARN of the reserved capacity to list UltraServers for.}

\item{MaxResults}{The maximum number of UltraServers to return in the response. The
default value is 10.}

\item{NextToken}{If the previous response was truncated, you receive this token. Use it
in your next request to receive the next set of results.}
}
\description{
Lists all UltraServers that are part of a specified reserved capacity.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_list_ultra_servers_by_reserved_capacity/} for full documentation.
}
\keyword{internal}
