% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_put_model_package_group_policy}
\alias{sagemaker_put_model_package_group_policy}
\title{Adds a resouce policy to control access to a model group}
\usage{
sagemaker_put_model_package_group_policy(ModelPackageGroupName,
  ResourcePolicy)
}
\arguments{
\item{ModelPackageGroupName}{[required] The name of the model group to add a resource policy to.}

\item{ResourcePolicy}{[required] The resource policy for the model group.}
}
\description{
Adds a resouce policy to control access to a model group. For
information about resoure policies, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/access_policies_identity-vs-resource.html}{Identity-based policies and resource-based policies}
in the \emph{AWS Identity and Access Management User Guide.}.
}
\section{Request syntax}{
\preformatted{svc$put_model_package_group_policy(
  ModelPackageGroupName = "string",
  ResourcePolicy = "string"
)
}
}

\keyword{internal}
