% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_start_edge_deployment_stage}
\alias{sagemaker_start_edge_deployment_stage}
\title{Starts a stage in an edge deployment plan}
\usage{
sagemaker_start_edge_deployment_stage(EdgeDeploymentPlanName, StageName)
}
\arguments{
\item{EdgeDeploymentPlanName}{[required] The name of the edge deployment plan to start.}

\item{StageName}{[required] The name of the stage to start.}
}
\description{
Starts a stage in an edge deployment plan.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_start_edge_deployment_stage/} for full documentation.
}
\keyword{internal}
