% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_start_pipeline_execution}
\alias{sagemaker_start_pipeline_execution}
\title{Starts a pipeline execution}
\usage{
sagemaker_start_pipeline_execution(
  PipelineName,
  PipelineExecutionDisplayName = NULL,
  PipelineParameters = NULL,
  PipelineExecutionDescription = NULL,
  ClientRequestToken,
  ParallelismConfiguration = NULL
)
}
\arguments{
\item{PipelineName}{[required] The name of the pipeline.}

\item{PipelineExecutionDisplayName}{The display name of the pipeline execution.}

\item{PipelineParameters}{Contains a list of pipeline parameters. This list can be empty.}

\item{PipelineExecutionDescription}{The description of the pipeline execution.}

\item{ClientRequestToken}{[required] A unique, case-sensitive identifier that you provide to ensure the
idempotency of the operation. An idempotent operation completes no more
than once.}

\item{ParallelismConfiguration}{This configuration, if specified, overrides the parallelism
configuration of the parent pipeline for this specific run.}
}
\description{
Starts a pipeline execution.

See \url{https://paws-r.github.io/docs/sagemaker/start_pipeline_execution.html} for full documentation.
}
\keyword{internal}
