% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_start_session}
\alias{sagemaker_start_session}
\title{Initiates a remote connection session between a local integrated
development environments (IDEs) and a remote SageMaker space}
\usage{
sagemaker_start_session(ResourceIdentifier)
}
\arguments{
\item{ResourceIdentifier}{[required] The Amazon Resource Name (ARN) of the resource to which the remote
connection will be established. For example, this identifies the
specific ARN space application you want to connect to from your local
IDE.}
}
\description{
Initiates a remote connection session between a local integrated development environments (IDEs) and a remote SageMaker space.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_start_session/} for full documentation.
}
\keyword{internal}
