% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_stop_ai_recommendation_job}
\alias{sagemaker_stop_ai_recommendation_job}
\title{Stops a running AI recommendation job}
\usage{
sagemaker_stop_ai_recommendation_job(AIRecommendationJobName)
}
\arguments{
\item{AIRecommendationJobName}{[required] The name of the AI recommendation job to stop.}
}
\description{
Stops a running AI recommendation job.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_stop_ai_recommendation_job/} for full documentation.
}
\keyword{internal}
