% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_stop_edge_deployment_stage}
\alias{sagemaker_stop_edge_deployment_stage}
\title{Stops a stage in an edge deployment plan}
\usage{
sagemaker_stop_edge_deployment_stage(EdgeDeploymentPlanName, StageName)
}
\arguments{
\item{EdgeDeploymentPlanName}{[required] The name of the edge deployment plan to stop.}

\item{StageName}{[required] The name of the stage to stop.}
}
\description{
Stops a stage in an edge deployment plan.

See \url{https://paws-r.github.io/docs/sagemaker/stop_edge_deployment_stage.html} for full documentation.
}
\keyword{internal}
