% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_stop_hyper_parameter_tuning_job}
\alias{sagemaker_stop_hyper_parameter_tuning_job}
\title{Stops a running hyperparameter tuning job and all running training jobs
that the tuning job launched}
\usage{
sagemaker_stop_hyper_parameter_tuning_job(HyperParameterTuningJobName)
}
\arguments{
\item{HyperParameterTuningJobName}{[required] The name of the tuning job to stop.}
}
\description{
Stops a running hyperparameter tuning job and all running training jobs that the tuning job launched.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_stop_hyper_parameter_tuning_job/} for full documentation.
}
\keyword{internal}
