% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_stop_inference_recommendations_job}
\alias{sagemaker_stop_inference_recommendations_job}
\title{Stops an Inference Recommender job}
\usage{
sagemaker_stop_inference_recommendations_job(JobName)
}
\arguments{
\item{JobName}{[required] The name of the job you want to stop.}
}
\description{
Stops an Inference Recommender job.

See \url{https://paws-r.github.io/docs/sagemaker/stop_inference_recommendations_job.html} for full documentation.
}
\keyword{internal}
