% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_stop_job}
\alias{sagemaker_stop_job}
\title{Stops a running job}
\usage{
sagemaker_stop_job(JobName, JobCategory)
}
\arguments{
\item{JobName}{[required] The name of the job to stop.}

\item{JobCategory}{[required] The category of the job to stop.}
}
\description{
Stops a running job. When you call \code{\link[=sagemaker_stop_job]{stop_job}}, Amazon SageMaker sets the job status to \code{Stopping}. After the job stops, the status changes to \code{Stopped}. Partial results may be available in the output location if the job was in progress. To delete a stopped job, call \code{\link[=sagemaker_delete_job]{delete_job}}.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_stop_job/} for full documentation.
}
\keyword{internal}
