% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_stop_mlflow_tracking_server}
\alias{sagemaker_stop_mlflow_tracking_server}
\title{Programmatically stop an MLflow Tracking Server}
\usage{
sagemaker_stop_mlflow_tracking_server(TrackingServerName)
}
\arguments{
\item{TrackingServerName}{[required] The name of the tracking server to stop.}
}
\description{
Programmatically stop an MLflow Tracking Server.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_stop_mlflow_tracking_server/} for full documentation.
}
\keyword{internal}
