% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_stop_monitoring_schedule}
\alias{sagemaker_stop_monitoring_schedule}
\title{Stops a previously started monitoring schedule}
\usage{
sagemaker_stop_monitoring_schedule(MonitoringScheduleName)
}
\arguments{
\item{MonitoringScheduleName}{[required] The name of the schedule to stop.}
}
\description{
Stops a previously started monitoring schedule.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_stop_monitoring_schedule/} for full documentation.
}
\keyword{internal}
