% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_stop_optimization_job}
\alias{sagemaker_stop_optimization_job}
\title{Ends a running inference optimization job}
\usage{
sagemaker_stop_optimization_job(OptimizationJobName)
}
\arguments{
\item{OptimizationJobName}{[required] The name that you assigned to the optimization job.}
}
\description{
Ends a running inference optimization job.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_stop_optimization_job/} for full documentation.
}
\keyword{internal}
