% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_stop_transform_job}
\alias{sagemaker_stop_transform_job}
\title{Stops a batch transform job}
\usage{
sagemaker_stop_transform_job(TransformJobName)
}
\arguments{
\item{TransformJobName}{[required] The name of the batch transform job to stop.}
}
\description{
Stops a batch transform job.

See \url{https://paws-r.github.io/docs/sagemaker/stop_transform_job.html} for full documentation.
}
\keyword{internal}
