% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_app_image_config}
\alias{sagemaker_update_app_image_config}
\title{Updates the properties of an AppImageConfig}
\usage{
sagemaker_update_app_image_config(AppImageConfigName,
  KernelGatewayImageConfig)
}
\arguments{
\item{AppImageConfigName}{[required] The name of the AppImageConfig to update.}

\item{KernelGatewayImageConfig}{The new KernelGateway app to run on the image.}
}
\value{
A list with the following syntax:\preformatted{list(
  AppImageConfigArn = "string"
)
}
}
\description{
Updates the properties of an AppImageConfig.
}
\section{Request syntax}{
\preformatted{svc$update_app_image_config(
  AppImageConfigName = "string",
  KernelGatewayImageConfig = list(
    KernelSpecs = list(
      list(
        Name = "string",
        DisplayName = "string"
      )
    ),
    FileSystemConfig = list(
      MountPath = "string",
      DefaultUid = 123,
      DefaultGid = 123
    )
  )
)
}
}

\keyword{internal}
