% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_cluster}
\alias{sagemaker_update_cluster}
\title{Update a SageMaker HyperPod cluster}
\usage{
sagemaker_update_cluster(ClusterName, InstanceGroups)
}
\arguments{
\item{ClusterName}{[required] Specify the name of the SageMaker HyperPod cluster you want to update.}

\item{InstanceGroups}{[required] Specify the instance groups to update.}
}
\description{
Update a SageMaker HyperPod cluster.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_update_cluster/} for full documentation.
}
\keyword{internal}
