% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_cluster_scheduler_config}
\alias{sagemaker_update_cluster_scheduler_config}
\title{Update the cluster policy configuration}
\usage{
sagemaker_update_cluster_scheduler_config(
  ClusterSchedulerConfigId,
  TargetVersion,
  SchedulerConfig = NULL,
  Description = NULL
)
}
\arguments{
\item{ClusterSchedulerConfigId}{[required] ID of the cluster policy.}

\item{TargetVersion}{[required] Target version.}

\item{SchedulerConfig}{Cluster policy configuration.}

\item{Description}{Description of the cluster policy.}
}
\description{
Update the cluster policy configuration.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_update_cluster_scheduler_config/} for full documentation.
}
\keyword{internal}
