% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_cluster_software}
\alias{sagemaker_update_cluster_software}
\title{Updates the platform software of a SageMaker HyperPod cluster for
security patching}
\usage{
sagemaker_update_cluster_software(ClusterName)
}
\arguments{
\item{ClusterName}{[required] Specify the name or the Amazon Resource Name (ARN) of the SageMaker
HyperPod cluster you want to update for security patching.}
}
\description{
Updates the platform software of a SageMaker HyperPod cluster for security patching. To learn how to use this API, see \href{https://docs.aws.amazon.com/sagemaker/latest/dg/sagemaker-hyperpod-operate.html#sagemaker-hyperpod-operate-cli-command-update-cluster-software}{Update the SageMaker HyperPod platform software of a cluster}.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_update_cluster_software/} for full documentation.
}
\keyword{internal}
