% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_compute_quota}
\alias{sagemaker_update_compute_quota}
\title{Update the compute allocation definition}
\usage{
sagemaker_update_compute_quota(
  ComputeQuotaId,
  TargetVersion,
  ComputeQuotaConfig = NULL,
  ComputeQuotaTarget = NULL,
  ActivationState = NULL,
  Description = NULL
)
}
\arguments{
\item{ComputeQuotaId}{[required] ID of the compute allocation definition.}

\item{TargetVersion}{[required] Target version.}

\item{ComputeQuotaConfig}{Configuration of the compute allocation definition. This includes the
resource sharing option, and the setting to preempt low priority tasks.}

\item{ComputeQuotaTarget}{The target entity to allocate compute resources to.}

\item{ActivationState}{The state of the compute allocation being described. Use to enable or
disable compute allocation.

Default is \code{Enabled}.}

\item{Description}{Description of the compute allocation definition.}
}
\description{
Update the compute allocation definition.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_update_compute_quota/} for full documentation.
}
\keyword{internal}
