% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_experiment}
\alias{sagemaker_update_experiment}
\title{Adds, updates, or removes the description of an experiment}
\usage{
sagemaker_update_experiment(
  ExperimentName,
  DisplayName = NULL,
  Description = NULL
)
}
\arguments{
\item{ExperimentName}{[required] The name of the experiment to update.}

\item{DisplayName}{The name of the experiment as displayed. The name doesn't need to be
unique. If \code{DisplayName} isn't specified, \code{ExperimentName} is displayed.}

\item{Description}{The description of the experiment.}
}
\description{
Adds, updates, or removes the description of an experiment. Updates the display name of an experiment.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_update_experiment/} for full documentation.
}
\keyword{internal}
