% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_feature_metadata}
\alias{sagemaker_update_feature_metadata}
\title{Updates the description and parameters of the feature group}
\usage{
sagemaker_update_feature_metadata(
  FeatureGroupName,
  FeatureName,
  Description = NULL,
  ParameterAdditions = NULL,
  ParameterRemovals = NULL
)
}
\arguments{
\item{FeatureGroupName}{[required] The name of the feature group containing the feature that you're
updating.}

\item{FeatureName}{[required] The name of the feature that you're updating.}

\item{Description}{A description that you can write to better describe the feature.}

\item{ParameterAdditions}{A list of key-value pairs that you can add to better describe the
feature.}

\item{ParameterRemovals}{A list of parameter keys that you can specify to remove parameters that
describe your feature.}
}
\description{
Updates the description and parameters of the feature group.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_update_feature_metadata/} for full documentation.
}
\keyword{internal}
