% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_hub}
\alias{sagemaker_update_hub}
\title{Update a hub}
\usage{
sagemaker_update_hub(
  HubName,
  HubDescription = NULL,
  HubDisplayName = NULL,
  HubSearchKeywords = NULL
)
}
\arguments{
\item{HubName}{[required] The name of the hub to update.}

\item{HubDescription}{A description of the updated hub.}

\item{HubDisplayName}{The display name of the hub.}

\item{HubSearchKeywords}{The searchable keywords for the hub.}
}
\description{
Update a hub.

See \url{https://paws-r.github.io/docs/sagemaker/update_hub.html} for full documentation.
}
\keyword{internal}
