% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_hub_content}
\alias{sagemaker_update_hub_content}
\title{Updates SageMaker hub content (either a Model or Notebook resource)}
\usage{
sagemaker_update_hub_content(
  HubName,
  HubContentName,
  HubContentType,
  HubContentVersion,
  HubContentDisplayName = NULL,
  HubContentDescription = NULL,
  HubContentMarkdown = NULL,
  HubContentSearchKeywords = NULL,
  SupportStatus = NULL
)
}
\arguments{
\item{HubName}{[required] The name of the SageMaker hub that contains the hub content you want to
update. You can optionally use the hub ARN instead.}

\item{HubContentName}{[required] The name of the hub content resource that you want to update.}

\item{HubContentType}{[required] The content type of the resource that you want to update. Only specify a
\code{Model} or \code{Notebook} resource for this API. To update a
\code{ModelReference}, use the
\code{\link[=sagemaker_update_hub_content_reference]{update_hub_content_reference}}
API instead.}

\item{HubContentVersion}{[required] The hub content version that you want to update. For example, if you
have two versions of a resource in your hub, you can update the second
version.}

\item{HubContentDisplayName}{The display name of the hub content.}

\item{HubContentDescription}{The description of the hub content.}

\item{HubContentMarkdown}{A string that provides a description of the hub content. This string can
include links, tables, and standard markdown formatting.}

\item{HubContentSearchKeywords}{The searchable keywords of the hub content.}

\item{SupportStatus}{Indicates the current status of the hub content resource.}
}
\description{
Updates SageMaker hub content (either a \code{Model} or \code{Notebook} resource).

See \url{https://www.paws-r-sdk.com/docs/sagemaker_update_hub_content/} for full documentation.
}
\keyword{internal}
