% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_inference_component}
\alias{sagemaker_update_inference_component}
\title{Updates an inference component}
\usage{
sagemaker_update_inference_component(
  InferenceComponentName,
  Specification = NULL,
  Specifications = NULL,
  RuntimeConfig = NULL,
  DeploymentConfig = NULL
)
}
\arguments{
\item{InferenceComponentName}{[required] The name of the inference component.}

\item{Specification}{Details about the resources to deploy with this inference component,
including the model, container, and compute resources.}

\item{Specifications}{A list of specification objects for the inference component, one per
instance type. Use this parameter when you want to specify different
model or resource configurations for the inference component on each
instance type. You can use either this parameter or the singular
\code{Specification} parameter, but not both.}

\item{RuntimeConfig}{Runtime settings for a model that is deployed with an inference
component.}

\item{DeploymentConfig}{The deployment configuration for the inference component. The
configuration contains the desired deployment strategy and rollback
settings.}
}
\description{
Updates an inference component.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_update_inference_component/} for full documentation.
}
\keyword{internal}
