% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_inference_component_runtime_config}
\alias{sagemaker_update_inference_component_runtime_config}
\title{Runtime settings for a model that is deployed with an inference
component}
\usage{
sagemaker_update_inference_component_runtime_config(
  InferenceComponentName,
  DesiredRuntimeConfig
)
}
\arguments{
\item{InferenceComponentName}{[required] The name of the inference component to update.}

\item{DesiredRuntimeConfig}{[required] Runtime settings for a model that is deployed with an inference
component.}
}
\description{
Runtime settings for a model that is deployed with an inference component.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_update_inference_component_runtime_config/} for full documentation.
}
\keyword{internal}
