% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_mlflow_app}
\alias{sagemaker_update_mlflow_app}
\title{Updates an MLflow App}
\usage{
sagemaker_update_mlflow_app(
  Arn,
  Name = NULL,
  ArtifactStoreUri = NULL,
  ModelRegistrationMode = NULL,
  WeeklyMaintenanceWindowStart = NULL,
  DefaultDomainIdList = NULL,
  AccountDefaultStatus = NULL
)
}
\arguments{
\item{Arn}{[required] The ARN of the MLflow App to update.}

\item{Name}{The name of the MLflow App to update.}

\item{ArtifactStoreUri}{The new S3 URI for the general purpose bucket to use as the artifact
store for the MLflow App.}

\item{ModelRegistrationMode}{Whether to enable or disable automatic registration of new MLflow models
to the SageMaker Model Registry. To enable automatic model registration,
set this value to \code{AutoModelRegistrationEnabled}. To disable automatic
model registration, set this value to \code{AutoModelRegistrationDisabled}.
If not specified, \code{AutomaticModelRegistration} defaults to
\code{AutoModelRegistrationEnabled}}

\item{WeeklyMaintenanceWindowStart}{The new weekly maintenance window start day and time to update. The
maintenance window day and time should be in Coordinated Universal Time
(UTC) 24-hour standard time. For example: TUE:03:30.}

\item{DefaultDomainIdList}{List of SageMaker Domain IDs for which this MLflow App is the default.}

\item{AccountDefaultStatus}{Indicates whether this this MLflow App is the default for the account.}
}
\description{
Updates an MLflow App.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_update_mlflow_app/} for full documentation.
}
\keyword{internal}
