% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_model_package}
\alias{sagemaker_update_model_package}
\title{Updates a versioned model}
\usage{
sagemaker_update_model_package(ModelPackageArn, ModelApprovalStatus,
  ApprovalDescription)
}
\arguments{
\item{ModelPackageArn}{[required] The Amazon Resource Name (ARN) of the model.}

\item{ModelApprovalStatus}{[required] The approval status of the model.}

\item{ApprovalDescription}{A description for the approval status of the model.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ModelPackageArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Updates a versioned model.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_model_package(
  ModelPackageArn = "string",
  ModelApprovalStatus = "Approved"|"Rejected"|"PendingManualApproval",
  ApprovalDescription = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
