% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_monitoring_alert}
\alias{sagemaker_update_monitoring_alert}
\title{Update the parameters of a model monitor alert}
\usage{
sagemaker_update_monitoring_alert(
  MonitoringScheduleName,
  MonitoringAlertName,
  DatapointsToAlert,
  EvaluationPeriod
)
}
\arguments{
\item{MonitoringScheduleName}{[required] The name of a monitoring schedule.}

\item{MonitoringAlertName}{[required] The name of a monitoring alert.}

\item{DatapointsToAlert}{[required] Within \code{EvaluationPeriod}, how many execution failures will raise an
alert.}

\item{EvaluationPeriod}{[required] The number of most recent monitoring executions to consider when
evaluating alert status.}
}
\description{
Update the parameters of a model monitor alert.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_update_monitoring_alert/} for full documentation.
}
\keyword{internal}
