% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_monitoring_schedule}
\alias{sagemaker_update_monitoring_schedule}
\title{Updates a previously created schedule}
\usage{
sagemaker_update_monitoring_schedule(
  MonitoringScheduleName,
  MonitoringScheduleConfig
)
}
\arguments{
\item{MonitoringScheduleName}{[required] The name of the monitoring schedule. The name must be unique within an
Amazon Web Services Region within an Amazon Web Services account.}

\item{MonitoringScheduleConfig}{[required] The configuration object that specifies the monitoring schedule and
defines the monitoring job.}
}
\description{
Updates a previously created schedule.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_update_monitoring_schedule/} for full documentation.
}
\keyword{internal}
