% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_notebook_instance_lifecycle_config}
\alias{sagemaker_update_notebook_instance_lifecycle_config}
\title{Updates a notebook instance lifecycle configuration created with the
CreateNotebookInstanceLifecycleConfig API}
\usage{
sagemaker_update_notebook_instance_lifecycle_config(
  NotebookInstanceLifecycleConfigName,
  OnCreate = NULL,
  OnStart = NULL
)
}
\arguments{
\item{NotebookInstanceLifecycleConfigName}{[required] The name of the lifecycle configuration.}

\item{OnCreate}{The shell script that runs only once, when you create a notebook
instance. The shell script must be a base64-encoded string.}

\item{OnStart}{The shell script that runs every time you start a notebook instance,
including when you create the notebook instance. The shell script must
be a base64-encoded string.}
}
\description{
Updates a notebook instance lifecycle configuration created with the \code{\link[=sagemaker_create_notebook_instance_lifecycle_config]{create_notebook_instance_lifecycle_config}} API.

See \url{https://paws-r.github.io/docs/sagemaker/update_notebook_instance_lifecycle_config.html} for full documentation.
}
\keyword{internal}
