% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_partner_app}
\alias{sagemaker_update_partner_app}
\title{Updates all of the SageMaker Partner AI Apps in an account}
\usage{
sagemaker_update_partner_app(
  Arn,
  MaintenanceConfig = NULL,
  Tier = NULL,
  ApplicationConfig = NULL,
  IdcConfig = NULL,
  AuthType = NULL,
  EnableIamSessionBasedIdentity = NULL,
  EnableAutoMinorVersionUpgrade = NULL,
  AppVersion = NULL,
  ClientToken = NULL,
  Tags = NULL
)
}
\arguments{
\item{Arn}{[required] The ARN of the SageMaker Partner AI App to update.}

\item{MaintenanceConfig}{Maintenance configuration settings for the SageMaker Partner AI App.}

\item{Tier}{Indicates the instance type and size of the cluster attached to the SageMaker Partner AI App.}

\item{ApplicationConfig}{Configuration settings for the SageMaker Partner AI App.}

\item{IdcConfig}{Specifies the Amazon Web Services IAM Identity Center configuration for the SageMaker Partner AI App. Specify this parameter when \code{AuthType} is \code{IDC}. Apps that use \code{IAM} authorization don't use this parameter.}

\item{AuthType}{The authorization type that users use to access the SageMaker Partner AI App. Use this parameter to migrate an existing SageMaker Partner AI App from \code{IAM} authorization to \code{IDC} authorization. Valid values:
\itemize{
\item \code{IAM}: Users access the SageMaker Partner AI App with their Amazon Web Services IAM identity.
\item \code{IDC}: Users access the SageMaker Partner AI App with their Amazon Web Services IAM Identity Center identity. Specify the Identity Center instance to use in \code{IdcConfig}.
}}

\item{EnableIamSessionBasedIdentity}{When set to \code{TRUE}, the SageMaker Partner AI App sets the Amazon Web Services IAM session name or the authenticated IAM user as the identity of the SageMaker Partner AI App user.}

\item{EnableAutoMinorVersionUpgrade}{When set to \code{TRUE}, the SageMaker Partner AI App is automatically upgraded to the latest minor version during the next scheduled maintenance window, if one is available.}

\item{AppVersion}{The semantic version to upgrade the SageMaker Partner AI App to. Must be the same semantic version returned in the \code{AvailableUpgrade} field from \code{\link[=sagemaker_describe_partner_app]{describe_partner_app}}. Version skipping and downgrades are not supported.}

\item{ClientToken}{A unique token that guarantees that the call to this API is idempotent.}

\item{Tags}{Each tag consists of a key and an optional value. Tag keys must be unique per resource.}
}
\description{
Updates all of the SageMaker Partner AI Apps in an account.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_update_partner_app/} for full documentation.
}
\keyword{internal}
