% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_pipeline}
\alias{sagemaker_update_pipeline}
\title{Updates a pipeline}
\usage{
sagemaker_update_pipeline(PipelineName, PipelineDisplayName,
  PipelineDefinition, PipelineDescription, RoleArn)
}
\arguments{
\item{PipelineName}{[required] The name of the pipeline to update.}

\item{PipelineDisplayName}{The display name of the pipeline.}

\item{PipelineDefinition}{The JSON pipeline definition.}

\item{PipelineDescription}{The description of the pipeline.}

\item{RoleArn}{The Amazon Resource Name (ARN) that the pipeline uses to execute.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  PipelineArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Updates a pipeline.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_pipeline(
  PipelineName = "string",
  PipelineDisplayName = "string",
  PipelineDefinition = "string",
  PipelineDescription = "string",
  RoleArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
