% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_pipeline_version}
\alias{sagemaker_update_pipeline_version}
\title{Updates a pipeline version}
\usage{
sagemaker_update_pipeline_version(
  PipelineArn,
  PipelineVersionId,
  PipelineVersionDisplayName = NULL,
  PipelineVersionDescription = NULL
)
}
\arguments{
\item{PipelineArn}{[required] The Amazon Resource Name (ARN) of the pipeline.}

\item{PipelineVersionId}{[required] The pipeline version ID to update.}

\item{PipelineVersionDisplayName}{The display name of the pipeline version.}

\item{PipelineVersionDescription}{The description of the pipeline version.}
}
\description{
Updates a pipeline version.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_update_pipeline_version/} for full documentation.
}
\keyword{internal}
