% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_training_job}
\alias{sagemaker_update_training_job}
\title{Update a model training job to request a new Debugger profiling
configuration}
\usage{
sagemaker_update_training_job(
  TrainingJobName,
  ProfilerConfig = NULL,
  ProfilerRuleConfigurations = NULL
)
}
\arguments{
\item{TrainingJobName}{[required] The name of a training job to update the Debugger profiling
configuration.}

\item{ProfilerConfig}{Configuration information for Debugger system monitoring, framework
profiling, and storage paths.}

\item{ProfilerRuleConfigurations}{Configuration information for Debugger rules for profiling system and
framework metrics.}
}
\description{
Update a model training job to request a new Debugger profiling configuration.

See \url{https://paws-r.github.io/docs/sagemaker/update_training_job.html} for full documentation.
}
\keyword{internal}
