% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_trial}
\alias{sagemaker_update_trial}
\title{Updates the display name of a trial}
\usage{
sagemaker_update_trial(TrialName, DisplayName)
}
\arguments{
\item{TrialName}{[required] The name of the trial to update.}

\item{DisplayName}{The name of the trial as displayed. The name doesn't need to be unique.
If \code{DisplayName} isn't specified, \code{TrialName} is displayed.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  TrialArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Updates the display name of a trial.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_trial(
  TrialName = "string",
  DisplayName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
