% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakeredgemanager_operations.R
\name{sagemakeredgemanager_get_device_registration}
\alias{sagemakeredgemanager_get_device_registration}
\title{Use to check if a device is registered with SageMaker Edge Manager}
\usage{
sagemakeredgemanager_get_device_registration(DeviceName, DeviceFleetName)
}
\arguments{
\item{DeviceName}{[required] The unique name of the device you want to get the registration status
from.}

\item{DeviceFleetName}{[required] The name of the fleet that the device belongs to.}
}
\description{
Use to check if a device is registered with SageMaker Edge Manager.

See \url{https://paws-r.github.io/docs/sagemakeredgemanager/get_device_registration.html} for full documentation.
}
\keyword{internal}
