% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakeredgemanager_operations.R
\name{sagemakeredgemanager_send_heartbeat}
\alias{sagemakeredgemanager_send_heartbeat}
\title{Use to get the current status of devices registered on SageMaker Edge
Manager}
\usage{
sagemakeredgemanager_send_heartbeat(
  AgentMetrics = NULL,
  Models = NULL,
  AgentVersion,
  DeviceName,
  DeviceFleetName,
  DeploymentResult = NULL
)
}
\arguments{
\item{AgentMetrics}{For internal use. Returns a list of SageMaker Edge Manager agent
operating metrics.}

\item{Models}{Returns a list of models deployed on the the device.}

\item{AgentVersion}{[required] Returns the version of the agent.}

\item{DeviceName}{[required] The unique name of the device.}

\item{DeviceFleetName}{[required] The name of the fleet that the device belongs to.}

\item{DeploymentResult}{Returns the result of a deployment on the device.}
}
\description{
Use to get the current status of devices registered on SageMaker Edge Manager.

See \url{https://paws-r.github.io/docs/sagemakeredgemanager/send_heartbeat.html} for full documentation.
}
\keyword{internal}
