% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakerfeaturestoreruntime_operations.R
\name{sagemakerfeaturestoreruntime_batch_write_record}
\alias{sagemakerfeaturestoreruntime_batch_write_record}
\title{Writes a batch of Records to one or more FeatureGroups}
\usage{
sagemakerfeaturestoreruntime_batch_write_record(Entries, TtlDuration = NULL)
}
\arguments{
\item{Entries}{[required] A list of records to write. Each entry specifies the \code{FeatureGroup}, the record data, and optionally target stores and a TTL duration.}

\item{TtlDuration}{Time to live duration applied to all entries in the batch that do not specify their own \code{TtlDuration}; \code{ExpiresAt} = \code{EventTime} + \code{TtlDuration}. For information on HardDelete, see the \code{\link[=sagemakerfeaturestoreruntime_delete_record]{delete_record}} API in the Amazon SageMaker API Reference guide.}
}
\description{
Writes a batch of \code{Records} to one or more \code{FeatureGroup}s. Use this API for bulk ingestion of records into the \code{OnlineStore} and \code{OfflineStore}.

See \url{https://www.paws-r-sdk.com/docs/sagemakerfeaturestoreruntime_batch_write_record/} for full documentation.
}
\keyword{internal}
