% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakerfeaturestoreruntime_operations.R
\name{sagemakerfeaturestoreruntime_list_records}
\alias{sagemakerfeaturestoreruntime_list_records}
\title{Lists the RecordIdentifier values of all records stored in a
FeatureGroup's OnlineStore}
\usage{
sagemakerfeaturestoreruntime_list_records(
  FeatureGroupName,
  MaxResults = NULL,
  NextToken = NULL,
  IncludeSoftDeletedRecords = NULL
)
}
\arguments{
\item{FeatureGroupName}{[required] The name or Amazon Resource Name (ARN) of the feature group to list records from.}

\item{MaxResults}{The maximum number of record identifiers to return in a single page of results. For the \code{InMemory} tier, this value is a hint and not a strict requirement. The response may contain more or fewer results than the specified \code{MaxResults}.}

\item{NextToken}{A token to resume pagination of \code{\link[=sagemakerfeaturestoreruntime_list_records]{list_records}} results.}

\item{IncludeSoftDeletedRecords}{If set to \code{true}, the result includes records that have been soft deleted.}
}
\description{
Lists the \code{RecordIdentifier} values of all records stored in a \code{FeatureGroup}'s \code{OnlineStore}. This enables you to discover which records exist without retrieving the full record data.

See \url{https://www.paws-r-sdk.com/docs/sagemakerfeaturestoreruntime_list_records/} for full documentation.
}
\keyword{internal}
