% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakerfeaturestoreruntime_operations.R
\name{sagemakerfeaturestoreruntime_update_record}
\alias{sagemakerfeaturestoreruntime_update_record}
\title{Updates one or more feature values for an existing record in the
specified feature group}
\usage{
sagemakerfeaturestoreruntime_update_record(
  FeatureGroupName,
  RecordIdentifierValueAsString,
  Features,
  TargetStores = NULL,
  TtlDuration = NULL
)
}
\arguments{
\item{FeatureGroupName}{[required] The identifier for the feature group that contains the record to update. You can specify one of the following:
\itemize{
\item The feature group name.
\item The feature group Amazon Resource Name (ARN).
}}

\item{RecordIdentifierValueAsString}{[required] The value that uniquely identifies the record in the feature group. This must match the value defined by the feature group's record identifier feature.}

\item{Features}{[required] The feature values to write to the record.}

\item{TargetStores}{The target stores for the record update. By default, Amazon SageMaker Feature Store updates the record in all stores associated with the \code{FeatureGroup}.}

\item{TtlDuration}{The time-to-live (TTL) duration for the record. Amazon SageMaker Feature Store deletes the record when \code{EventTime} + \code{TtlDuration} elapses. If you omit this parameter, the record's existing TTL setting remains unchanged. For information about \code{HardDelete}, see the \code{\link[=sagemakerfeaturestoreruntime_delete_record]{delete_record}} operation in the Amazon SageMaker API Reference.}
}
\description{
Updates one or more feature values for an existing record in the specified feature group. Features that you do not include in the request remain unchanged. You can update up to 100 features per call.

See \url{https://www.paws-r-sdk.com/docs/sagemakerfeaturestoreruntime_update_record/} for full documentation.
}
\keyword{internal}
