% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakergeospatialcapabilities_operations.R
\name{sagemakergeospatialcapabilities_delete_earth_observation_job}
\alias{sagemakergeospatialcapabilities_delete_earth_observation_job}
\title{Use this operation to delete an Earth Observation job}
\usage{
sagemakergeospatialcapabilities_delete_earth_observation_job(Arn)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the Earth Observation job being
deleted.}
}
\description{
Use this operation to delete an Earth Observation job.

See \url{https://www.paws-r-sdk.com/docs/sagemakergeospatialcapabilities_delete_earth_observation_job/} for full documentation.
}
\keyword{internal}
