% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakergeospatialcapabilities_operations.R
\name{sagemakergeospatialcapabilities_delete_vector_enrichment_job}
\alias{sagemakergeospatialcapabilities_delete_vector_enrichment_job}
\title{Use this operation to delete a Vector Enrichment job}
\usage{
sagemakergeospatialcapabilities_delete_vector_enrichment_job(Arn)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the Vector Enrichment job being
deleted.}
}
\description{
Use this operation to delete a Vector Enrichment job.

See \url{https://www.paws-r-sdk.com/docs/sagemakergeospatialcapabilities_delete_vector_enrichment_job/} for full documentation.
}
\keyword{internal}
