% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakergeospatialcapabilities_operations.R
\name{sagemakergeospatialcapabilities_export_earth_observation_job}
\alias{sagemakergeospatialcapabilities_export_earth_observation_job}
\title{Use this operation to export results of an Earth Observation job and
optionally source images used as input to the EOJ to an Amazon S3
location}
\usage{
sagemakergeospatialcapabilities_export_earth_observation_job(
  Arn,
  ClientToken = NULL,
  ExecutionRoleArn,
  ExportSourceImages = NULL,
  OutputConfig
)
}
\arguments{
\item{Arn}{[required] The input Amazon Resource Name (ARN) of the Earth Observation job being
exported.}

\item{ClientToken}{A unique token that guarantees that the call to this API is idempotent.}

\item{ExecutionRoleArn}{[required] The Amazon Resource Name (ARN) of the IAM role that you specified for
the job.}

\item{ExportSourceImages}{The source images provided to the Earth Observation job being exported.}

\item{OutputConfig}{[required] An object containing information about the output file.}
}
\description{
Use this operation to export results of an Earth Observation job and optionally source images used as input to the EOJ to an Amazon S3 location.

See \url{https://www.paws-r-sdk.com/docs/sagemakergeospatialcapabilities_export_earth_observation_job/} for full documentation.
}
\keyword{internal}
