% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakergeospatialcapabilities_operations.R
\name{sagemakergeospatialcapabilities_export_vector_enrichment_job}
\alias{sagemakergeospatialcapabilities_export_vector_enrichment_job}
\title{Use this operation to copy results of a Vector Enrichment job to an
Amazon S3 location}
\usage{
sagemakergeospatialcapabilities_export_vector_enrichment_job(
  Arn,
  ClientToken = NULL,
  ExecutionRoleArn,
  OutputConfig
)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the Vector Enrichment job.}

\item{ClientToken}{A unique token that guarantees that the call to this API is idempotent.}

\item{ExecutionRoleArn}{[required] The Amazon Resource Name (ARN) of the IAM rolewith permission to upload
to the location in OutputConfig.}

\item{OutputConfig}{[required] Output location information for exporting Vector Enrichment Job results.}
}
\description{
Use this operation to copy results of a Vector Enrichment job to an Amazon S3 location.

See \url{https://www.paws-r-sdk.com/docs/sagemakergeospatialcapabilities_export_vector_enrichment_job/} for full documentation.
}
\keyword{internal}
