% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakergeospatialcapabilities_operations.R
\name{sagemakergeospatialcapabilities_get_earth_observation_job}
\alias{sagemakergeospatialcapabilities_get_earth_observation_job}
\title{Get the details for a previously initiated Earth Observation job}
\usage{
sagemakergeospatialcapabilities_get_earth_observation_job(Arn)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the Earth Observation job.}
}
\description{
Get the details for a previously initiated Earth Observation job.

See \url{https://www.paws-r-sdk.com/docs/sagemakergeospatialcapabilities_get_earth_observation_job/} for full documentation.
}
\keyword{internal}
