% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakergeospatialcapabilities_operations.R
\name{sagemakergeospatialcapabilities_get_raster_data_collection}
\alias{sagemakergeospatialcapabilities_get_raster_data_collection}
\title{Use this operation to get details of a specific raster data collection}
\usage{
sagemakergeospatialcapabilities_get_raster_data_collection(Arn)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the raster data collection.}
}
\description{
Use this operation to get details of a specific raster data collection.

See \url{https://www.paws-r-sdk.com/docs/sagemakergeospatialcapabilities_get_raster_data_collection/} for full documentation.
}
\keyword{internal}
