% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakergeospatialcapabilities_operations.R
\name{sagemakergeospatialcapabilities_get_vector_enrichment_job}
\alias{sagemakergeospatialcapabilities_get_vector_enrichment_job}
\title{Retrieves details of a Vector Enrichment Job for a given job Amazon
Resource Name (ARN)}
\usage{
sagemakergeospatialcapabilities_get_vector_enrichment_job(Arn)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the Vector Enrichment job.}
}
\description{
Retrieves details of a Vector Enrichment Job for a given job Amazon Resource Name (ARN).

See \url{https://www.paws-r-sdk.com/docs/sagemakergeospatialcapabilities_get_vector_enrichment_job/} for full documentation.
}
\keyword{internal}
