% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakergeospatialcapabilities_operations.R
\name{sagemakergeospatialcapabilities_list_raster_data_collections}
\alias{sagemakergeospatialcapabilities_list_raster_data_collections}
\title{Use this operation to get raster data collections}
\usage{
sagemakergeospatialcapabilities_list_raster_data_collections(
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MaxResults}{The total number of items to return.}

\item{NextToken}{If the previous response was truncated, you receive this token. Use it
in your next request to receive the next set of results.}
}
\description{
Use this operation to get raster data collections.

See \url{https://www.paws-r-sdk.com/docs/sagemakergeospatialcapabilities_list_raster_data_collections/} for full documentation.
}
\keyword{internal}
