% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakergeospatialcapabilities_operations.R
\name{sagemakergeospatialcapabilities_list_tags_for_resource}
\alias{sagemakergeospatialcapabilities_list_tags_for_resource}
\title{Lists the tags attached to the resource}
\usage{
sagemakergeospatialcapabilities_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource you want to tag.}
}
\description{
Lists the tags attached to the resource.

See \url{https://www.paws-r-sdk.com/docs/sagemakergeospatialcapabilities_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
