% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakergeospatialcapabilities_operations.R
\name{sagemakergeospatialcapabilities_list_vector_enrichment_jobs}
\alias{sagemakergeospatialcapabilities_list_vector_enrichment_jobs}
\title{Retrieves a list of vector enrichment jobs}
\usage{
sagemakergeospatialcapabilities_list_vector_enrichment_jobs(
  MaxResults = NULL,
  NextToken = NULL,
  SortBy = NULL,
  SortOrder = NULL,
  StatusEquals = NULL
)
}
\arguments{
\item{MaxResults}{The maximum number of items to return.}

\item{NextToken}{If the previous response was truncated, you receive this token. Use it
in your next request to receive the next set of results.}

\item{SortBy}{The parameter by which to sort the results.}

\item{SortOrder}{An optional value that specifies whether you want the results sorted in
\code{Ascending} or \code{Descending} order.}

\item{StatusEquals}{A filter that retrieves only jobs with a specific status.}
}
\description{
Retrieves a list of vector enrichment jobs.

See \url{https://www.paws-r-sdk.com/docs/sagemakergeospatialcapabilities_list_vector_enrichment_jobs/} for full documentation.
}
\keyword{internal}
