% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakergeospatialcapabilities_operations.R
\name{sagemakergeospatialcapabilities_start_vector_enrichment_job}
\alias{sagemakergeospatialcapabilities_start_vector_enrichment_job}
\title{Creates a Vector Enrichment job for the supplied job type}
\usage{
sagemakergeospatialcapabilities_start_vector_enrichment_job(
  ClientToken = NULL,
  ExecutionRoleArn,
  InputConfig,
  JobConfig,
  KmsKeyId = NULL,
  Name,
  Tags = NULL
)
}
\arguments{
\item{ClientToken}{A unique token that guarantees that the call to this API is idempotent.}

\item{ExecutionRoleArn}{[required] The Amazon Resource Name (ARN) of the IAM role that you specified for
the job.}

\item{InputConfig}{[required] Input configuration information for the Vector Enrichment job.}

\item{JobConfig}{[required] An object containing information about the job configuration.}

\item{KmsKeyId}{The Key Management Service key ID for server-side encryption.}

\item{Name}{[required] The name of the Vector Enrichment job.}

\item{Tags}{Each tag consists of a key and a value.}
}
\description{
Creates a Vector Enrichment job for the supplied job type. Currently, there are two supported job types: reverse geocoding and map matching.

See \url{https://www.paws-r-sdk.com/docs/sagemakergeospatialcapabilities_start_vector_enrichment_job/} for full documentation.
}
\keyword{internal}
