% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakergeospatialcapabilities_operations.R
\name{sagemakergeospatialcapabilities_stop_earth_observation_job}
\alias{sagemakergeospatialcapabilities_stop_earth_observation_job}
\title{Use this operation to stop an existing earth observation job}
\usage{
sagemakergeospatialcapabilities_stop_earth_observation_job(Arn)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the Earth Observation job being
stopped.}
}
\description{
Use this operation to stop an existing earth observation job.

See \url{https://www.paws-r-sdk.com/docs/sagemakergeospatialcapabilities_stop_earth_observation_job/} for full documentation.
}
\keyword{internal}
