% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakergeospatialcapabilities_operations.R
\name{sagemakergeospatialcapabilities_stop_vector_enrichment_job}
\alias{sagemakergeospatialcapabilities_stop_vector_enrichment_job}
\title{Stops the Vector Enrichment job for a given job ARN}
\usage{
sagemakergeospatialcapabilities_stop_vector_enrichment_job(Arn)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the Vector Enrichment job.}
}
\description{
Stops the Vector Enrichment job for a given job ARN.

See \url{https://www.paws-r-sdk.com/docs/sagemakergeospatialcapabilities_stop_vector_enrichment_job/} for full documentation.
}
\keyword{internal}
