% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakergeospatialcapabilities_operations.R
\name{sagemakergeospatialcapabilities_tag_resource}
\alias{sagemakergeospatialcapabilities_tag_resource}
\title{The resource you want to tag}
\usage{
sagemakergeospatialcapabilities_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource you want to tag.}

\item{Tags}{[required] Each tag consists of a key and a value.}
}
\description{
The resource you want to tag.

See \url{https://www.paws-r-sdk.com/docs/sagemakergeospatialcapabilities_tag_resource/} for full documentation.
}
\keyword{internal}
