% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakergeospatialcapabilities_operations.R
\name{sagemakergeospatialcapabilities_untag_resource}
\alias{sagemakergeospatialcapabilities_untag_resource}
\title{The resource you want to untag}
\usage{
sagemakergeospatialcapabilities_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource you want to untag.}

\item{TagKeys}{[required] Keys of the tags you want to remove.}
}
\description{
The resource you want to untag.

See \url{https://www.paws-r-sdk.com/docs/sagemakergeospatialcapabilities_untag_resource/} for full documentation.
}
\keyword{internal}
