% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakermetrics_operations.R
\name{sagemakermetrics_batch_get_metrics}
\alias{sagemakermetrics_batch_get_metrics}
\title{Used to retrieve training metrics from SageMaker}
\usage{
sagemakermetrics_batch_get_metrics(MetricQueries)
}
\arguments{
\item{MetricQueries}{[required] Queries made to retrieve training metrics from SageMaker.}
}
\description{
Used to retrieve training metrics from SageMaker.

See \url{https://www.paws-r-sdk.com/docs/sagemakermetrics_batch_get_metrics/} for full documentation.
}
\keyword{internal}
