% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakermetrics_operations.R
\name{sagemakermetrics_batch_put_metrics}
\alias{sagemakermetrics_batch_put_metrics}
\title{Used to ingest training metrics into SageMaker}
\usage{
sagemakermetrics_batch_put_metrics(TrialComponentName, MetricData)
}
\arguments{
\item{TrialComponentName}{[required] The name of the Trial Component to associate with the metrics.}

\item{MetricData}{[required] A list of raw metric values to put.}
}
\description{
Used to ingest training metrics into SageMaker. These metrics can be visualized in SageMaker Studio and retrieved with the \code{GetMetrics} API.

See \url{https://www.paws-r-sdk.com/docs/sagemakermetrics_batch_put_metrics/} for full documentation.
}
\keyword{internal}
