% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textract_operations.R
\name{textract_delete_adapter}
\alias{textract_delete_adapter}
\title{Deletes an Amazon Textract adapter}
\usage{
textract_delete_adapter(AdapterId)
}
\arguments{
\item{AdapterId}{[required] A string containing a unique ID for the adapter to be deleted.}
}
\description{
Deletes an Amazon Textract adapter. Takes an AdapterId and deletes the adapter specified by the ID.

See \url{https://www.paws-r-sdk.com/docs/textract_delete_adapter/} for full documentation.
}
\keyword{internal}
