% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textract_operations.R
\name{textract_delete_adapter_version}
\alias{textract_delete_adapter_version}
\title{Deletes an Amazon Textract adapter version}
\usage{
textract_delete_adapter_version(AdapterId, AdapterVersion)
}
\arguments{
\item{AdapterId}{[required] A string containing a unique ID for the adapter version that will be
deleted.}

\item{AdapterVersion}{[required] Specifies the adapter version to be deleted.}
}
\description{
Deletes an Amazon Textract adapter version. Requires that you specify both an AdapterId and a AdapterVersion. Deletes the adapter version specified by the AdapterId and the AdapterVersion.

See \url{https://www.paws-r-sdk.com/docs/textract_delete_adapter_version/} for full documentation.
}
\keyword{internal}
