% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textract_operations.R
\name{textract_get_adapter}
\alias{textract_get_adapter}
\title{Gets configuration information for an adapter specified by an AdapterId,
returning information on AdapterName, Description, CreationTime,
AutoUpdate status, and FeatureTypes}
\usage{
textract_get_adapter(AdapterId)
}
\arguments{
\item{AdapterId}{[required] A string containing a unique ID for the adapter.}
}
\description{
Gets configuration information for an adapter specified by an AdapterId, returning information on AdapterName, Description, CreationTime, AutoUpdate status, and FeatureTypes.

See \url{https://www.paws-r-sdk.com/docs/textract_get_adapter/} for full documentation.
}
\keyword{internal}
