% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textract_operations.R
\name{textract_get_adapter_version}
\alias{textract_get_adapter_version}
\title{Gets configuration information for the specified adapter version,
including: AdapterId, AdapterVersion, FeatureTypes, Status,
StatusMessage, DatasetConfig, KMSKeyId, OutputConfig, Tags and
EvaluationMetrics}
\usage{
textract_get_adapter_version(AdapterId, AdapterVersion)
}
\arguments{
\item{AdapterId}{[required] A string specifying a unique ID for the adapter version you want to
retrieve information for.}

\item{AdapterVersion}{[required] A string specifying the adapter version you want to retrieve information
for.}
}
\description{
Gets configuration information for the specified adapter version, including: AdapterId, AdapterVersion, FeatureTypes, Status, StatusMessage, DatasetConfig, KMSKeyId, OutputConfig, Tags and EvaluationMetrics.

See \url{https://www.paws-r-sdk.com/docs/textract_get_adapter_version/} for full documentation.
}
\keyword{internal}
