% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textract_operations.R
\name{textract_get_lending_analysis}
\alias{textract_get_lending_analysis}
\title{Gets the results for an Amazon Textract asynchronous operation that
analyzes text in a lending document}
\usage{
textract_get_lending_analysis(JobId, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{JobId}{[required] A unique identifier for the lending or text-detection job. The \code{JobId}
is returned from
\code{\link[=textract_start_lending_analysis]{start_lending_analysis}}. A \code{JobId}
value is only valid for 7 days.}

\item{MaxResults}{The maximum number of results to return per paginated call. The largest
value that you can specify is 30. If you specify a value greater than
30, a maximum of 30 results is returned. The default value is 30.}

\item{NextToken}{If the previous response was incomplete, Amazon Textract returns a
pagination token in the response. You can use this pagination token to
retrieve the next set of lending results.}
}
\description{
Gets the results for an Amazon Textract asynchronous operation that analyzes text in a lending document.

See \url{https://www.paws-r-sdk.com/docs/textract_get_lending_analysis/} for full documentation.
}
\keyword{internal}
