% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textract_operations.R
\name{textract_list_adapter_versions}
\alias{textract_list_adapter_versions}
\title{List all version of an adapter that meet the specified filtration
criteria}
\usage{
textract_list_adapter_versions(
  AdapterId = NULL,
  AfterCreationTime = NULL,
  BeforeCreationTime = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AdapterId}{A string containing a unique ID for the adapter to match for when
listing adapter versions.}

\item{AfterCreationTime}{Specifies the lower bound for the ListAdapterVersions operation. Ensures
ListAdapterVersions returns only adapter versions created after the
specified creation time.}

\item{BeforeCreationTime}{Specifies the upper bound for the ListAdapterVersions operation. Ensures
ListAdapterVersions returns only adapter versions created after the
specified creation time.}

\item{MaxResults}{The maximum number of results to return when listing adapter versions.}

\item{NextToken}{Identifies the next page of results to return when listing adapter
versions.}
}
\description{
List all version of an adapter that meet the specified filtration criteria.

See \url{https://www.paws-r-sdk.com/docs/textract_list_adapter_versions/} for full documentation.
}
\keyword{internal}
