% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textract_operations.R
\name{textract_list_adapters}
\alias{textract_list_adapters}
\title{Lists all adapters that match the specified filtration criteria}
\usage{
textract_list_adapters(
  AfterCreationTime = NULL,
  BeforeCreationTime = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AfterCreationTime}{Specifies the lower bound for the ListAdapters operation. Ensures
ListAdapters returns only adapters created after the specified creation
time.}

\item{BeforeCreationTime}{Specifies the upper bound for the ListAdapters operation. Ensures
ListAdapters returns only adapters created before the specified creation
time.}

\item{MaxResults}{The maximum number of results to return when listing adapters.}

\item{NextToken}{Identifies the next page of results to return when listing adapters.}
}
\description{
Lists all adapters that match the specified filtration criteria.

See \url{https://www.paws-r-sdk.com/docs/textract_list_adapters/} for full documentation.
}
\keyword{internal}
