% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textract_operations.R
\name{textract_list_tags_for_resource}
\alias{textract_list_tags_for_resource}
\title{Lists all tags for an Amazon Textract resource}
\usage{
textract_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) that specifies the resource to list tags
for.}
}
\description{
Lists all tags for an Amazon Textract resource.

See \url{https://www.paws-r-sdk.com/docs/textract_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
