% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textract_operations.R
\name{textract_tag_resource}
\alias{textract_tag_resource}
\title{Adds one or more tags to the specified resource}
\usage{
textract_tag_resource(ResourceARN, Tags)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) that specifies the resource to be tagged.}

\item{Tags}{[required] A set of tags (key-value pairs) that you want to assign to the resource.}
}
\description{
Adds one or more tags to the specified resource.

See \url{https://www.paws-r-sdk.com/docs/textract_tag_resource/} for full documentation.
}
\keyword{internal}
