% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textract_operations.R
\name{textract_untag_resource}
\alias{textract_untag_resource}
\title{Removes any tags with the specified keys from the specified resource}
\usage{
textract_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) that specifies the resource to be
untagged.}

\item{TagKeys}{[required] Specifies the tags to be removed from the resource specified by the
ResourceARN.}
}
\description{
Removes any tags with the specified keys from the specified resource.

See \url{https://www.paws-r-sdk.com/docs/textract_untag_resource/} for full documentation.
}
\keyword{internal}
