% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textract_operations.R
\name{textract_update_adapter}
\alias{textract_update_adapter}
\title{Update the configuration for an adapter}
\usage{
textract_update_adapter(
  AdapterId,
  Description = NULL,
  AdapterName = NULL,
  AutoUpdate = NULL
)
}
\arguments{
\item{AdapterId}{[required] A string containing a unique ID for the adapter that will be updated.}

\item{Description}{The new description to be applied to the adapter.}

\item{AdapterName}{The new name to be applied to the adapter.}

\item{AutoUpdate}{The new auto-update status to be applied to the adapter.}
}
\description{
Update the configuration for an adapter. FeatureTypes configurations cannot be updated. At least one new parameter must be specified as an argument.

See \url{https://www.paws-r-sdk.com/docs/textract_update_adapter/} for full documentation.
}
\keyword{internal}
