% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_create_vocabulary}
\alias{transcribeservice_create_vocabulary}
\title{Creates a new custom vocabulary}
\usage{
transcribeservice_create_vocabulary(
  VocabularyName,
  LanguageCode,
  Phrases = NULL,
  VocabularyFileUri = NULL,
  Tags = NULL,
  DataAccessRoleArn = NULL,
  EncryptionConfiguration = NULL
)
}
\arguments{
\item{VocabularyName}{[required] A unique name, chosen by you, for your new custom vocabulary.

This name is case sensitive, cannot contain spaces, and must be unique within an Amazon Web Services account. If you try to create a new custom vocabulary with the same name as an existing custom vocabulary, you get a \code{ConflictException} error.}

\item{LanguageCode}{[required] The language code that represents the language of the entries in your custom vocabulary. Each custom vocabulary must contain terms in only one language.

A custom vocabulary can only be used to transcribe files in the same language as the custom vocabulary. For example, if you create a custom vocabulary using US English (\code{en-US}), you can only apply this custom vocabulary to files that contain English audio.

For a list of supported languages and their associated language codes, refer to the \href{https://docs.aws.amazon.com/transcribe/latest/dg/supported-languages.html}{Supported languages} table.}

\item{Phrases}{Use this parameter if you want to create your custom vocabulary by including all desired terms, as comma-separated values, within your request. The other option for creating your custom vocabulary is to save your entries in a text file and upload them to an Amazon S3 bucket, then specify the location of your file using the \code{VocabularyFileUri} parameter.

Note that if you include \code{Phrases} in your request, you cannot use \code{VocabularyFileUri}; you must choose one or the other.

Each language has a character set that contains all allowed characters for that specific language. If you use unsupported characters, your custom vocabulary filter request fails. Refer to \href{https://docs.aws.amazon.com/transcribe/latest/dg/charsets.html}{Character Sets for Custom Vocabularies} to get the character set for your language.}

\item{VocabularyFileUri}{The Amazon S3 location of the text file that contains your custom vocabulary. The URI must be located in the same Amazon Web Services Region as the resource you're calling.

Here's an example URI path: \verb{s3://DOC-EXAMPLE-BUCKET/my-vocab-file.txt}

Note that if you include \code{VocabularyFileUri} in your request, you cannot use the \code{Phrases} flag; you must choose one or the other.}

\item{Tags}{Adds one or more custom tags, each in the form of a key:value pair, to a new custom vocabulary at the time you create this new custom vocabulary.

To learn more about using tags with Amazon Transcribe, refer to \href{https://docs.aws.amazon.com/transcribe/latest/dg/tagging.html}{Tagging resources}.}

\item{DataAccessRoleArn}{The Amazon Resource Name (ARN) of an IAM role that has permissions to access the Amazon S3 bucket that contains your input files (in this case, your custom vocabulary). If you include \code{EncryptionConfiguration} in your request, this role must also have permissions to access the specified KMS key. If the role that you specify doesn’t have the appropriate permissions, your request fails.

IAM role ARNs have the format \code{arn:partition:iam::account:role/role-name-with-path}. For example: \verb{arn:aws:iam::111122223333:role/Admin}.

For more information, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference_identifiers.html#identifiers-arns}{IAM ARNs}.}

\item{EncryptionConfiguration}{Specifies the encryption configuration for your custom vocabulary. Your vocabulary artifacts are encrypted with the specified KMS key or with an AWS-owned key if a key is not supplied.}
}
\description{
Creates a new custom vocabulary.

See \url{https://www.paws-r-sdk.com/docs/transcribeservice_create_vocabulary/} for full documentation.
}
\keyword{internal}
