% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_delete_call_analytics_category}
\alias{transcribeservice_delete_call_analytics_category}
\title{Deletes a Call Analytics category}
\usage{
transcribeservice_delete_call_analytics_category(CategoryName)
}
\arguments{
\item{CategoryName}{[required] The name of the Call Analytics category you want to delete. Category
names are case sensitive.}
}
\description{
Deletes a Call Analytics category. To use this operation, specify the name of the category you want to delete using \code{CategoryName}. Category names are case sensitive.

See \url{https://www.paws-r-sdk.com/docs/transcribeservice_delete_call_analytics_category/} for full documentation.
}
\keyword{internal}
