% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_delete_call_analytics_job}
\alias{transcribeservice_delete_call_analytics_job}
\title{Deletes a Call Analytics job}
\usage{
transcribeservice_delete_call_analytics_job(CallAnalyticsJobName)
}
\arguments{
\item{CallAnalyticsJobName}{[required] The name of the Call Analytics job you want to delete. Job names are
case sensitive.}
}
\description{
Deletes a Call Analytics job. To use this operation, specify the name of the job you want to delete using \code{CallAnalyticsJobName}. Job names are case sensitive.

See \url{https://paws-r.github.io/docs/transcribeservice/delete_call_analytics_job.html} for full documentation.
}
\keyword{internal}
