% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_delete_language_model}
\alias{transcribeservice_delete_language_model}
\title{Deletes a custom language model}
\usage{
transcribeservice_delete_language_model(ModelName)
}
\arguments{
\item{ModelName}{[required] The name of the custom language model you want to delete. Model names
are case sensitive.}
}
\description{
Deletes a custom language model. To use this operation, specify the name of the language model you want to delete using \code{ModelName}. custom language model names are case sensitive.

See \url{https://www.paws-r-sdk.com/docs/transcribeservice_delete_language_model/} for full documentation.
}
\keyword{internal}
