% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_delete_medical_scribe_job}
\alias{transcribeservice_delete_medical_scribe_job}
\title{Deletes a Medical Scribe job}
\usage{
transcribeservice_delete_medical_scribe_job(MedicalScribeJobName)
}
\arguments{
\item{MedicalScribeJobName}{[required] The name of the Medical Scribe job you want to delete. Job names are
case sensitive.}
}
\description{
Deletes a Medical Scribe job. To use this operation, specify the name of the job you want to delete using \code{MedicalScribeJobName}. Job names are case sensitive.

See \url{https://www.paws-r-sdk.com/docs/transcribeservice_delete_medical_scribe_job/} for full documentation.
}
\keyword{internal}
