% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_delete_medical_transcription_job}
\alias{transcribeservice_delete_medical_transcription_job}
\title{Deletes a medical transcription job}
\usage{
transcribeservice_delete_medical_transcription_job(MedicalTranscriptionJobName)
}
\arguments{
\item{MedicalTranscriptionJobName}{[required] The name of the medical transcription job you want to delete. Job names
are case sensitive.}
}
\description{
Deletes a medical transcription job. To use this operation, specify the name of the job you want to delete using \code{MedicalTranscriptionJobName}. Job names are case sensitive.

See \url{https://www.paws-r-sdk.com/docs/transcribeservice_delete_medical_transcription_job/} for full documentation.
}
\keyword{internal}
