% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_get_call_analytics_category}
\alias{transcribeservice_get_call_analytics_category}
\title{Provides information about the specified Call Analytics category}
\usage{
transcribeservice_get_call_analytics_category(CategoryName)
}
\arguments{
\item{CategoryName}{[required] The name of the Call Analytics category you want information about.
Category names are case sensitive.}
}
\description{
Provides information about the specified Call Analytics category.

See \url{https://www.paws-r-sdk.com/docs/transcribeservice_get_call_analytics_category/} for full documentation.
}
\keyword{internal}
