% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_get_call_analytics_job}
\alias{transcribeservice_get_call_analytics_job}
\title{Provides information about the specified Call Analytics job}
\usage{
transcribeservice_get_call_analytics_job(CallAnalyticsJobName)
}
\arguments{
\item{CallAnalyticsJobName}{[required] The name of the Call Analytics job you want information about. Job names
are case sensitive.}
}
\description{
Provides information about the specified Call Analytics job.

See \url{https://www.paws-r-sdk.com/docs/transcribeservice_get_call_analytics_job/} for full documentation.
}
\keyword{internal}
