% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_get_medical_scribe_job}
\alias{transcribeservice_get_medical_scribe_job}
\title{Provides information about the specified Medical Scribe job}
\usage{
transcribeservice_get_medical_scribe_job(MedicalScribeJobName)
}
\arguments{
\item{MedicalScribeJobName}{[required] The name of the Medical Scribe job you want information about. Job names
are case sensitive.}
}
\description{
Provides information about the specified Medical Scribe job.

See \url{https://www.paws-r-sdk.com/docs/transcribeservice_get_medical_scribe_job/} for full documentation.
}
\keyword{internal}
