% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_get_medical_vocabulary}
\alias{transcribeservice_get_medical_vocabulary}
\title{Provides information about the specified custom medical vocabulary}
\usage{
transcribeservice_get_medical_vocabulary(VocabularyName)
}
\arguments{
\item{VocabularyName}{[required] The name of the custom medical vocabulary you want information about.
Vocabulary names are case sensitive.}
}
\description{
Provides information about the specified custom medical vocabulary.

See \url{https://paws-r.github.io/docs/transcribeservice/get_medical_vocabulary.html} for full documentation.
}
\keyword{internal}
