% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_get_transcription_job}
\alias{transcribeservice_get_transcription_job}
\title{Provides information about the specified transcription job}
\usage{
transcribeservice_get_transcription_job(TranscriptionJobName)
}
\arguments{
\item{TranscriptionJobName}{[required] The name of the transcription job you want information about. Job names
are case sensitive.}
}
\description{
Provides information about the specified transcription job.

See \url{https://www.paws-r-sdk.com/docs/transcribeservice_get_transcription_job/} for full documentation.
}
\keyword{internal}
