% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_get_vocabulary}
\alias{transcribeservice_get_vocabulary}
\title{Provides information about the specified custom vocabulary}
\usage{
transcribeservice_get_vocabulary(VocabularyName)
}
\arguments{
\item{VocabularyName}{[required] The name of the custom vocabulary you want information about. Custom
vocabulary names are case sensitive.}
}
\description{
Provides information about the specified custom vocabulary.

See \url{https://www.paws-r-sdk.com/docs/transcribeservice_get_vocabulary/} for full documentation.
}
\keyword{internal}
