% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_get_vocabulary_filter}
\alias{transcribeservice_get_vocabulary_filter}
\title{Returns information about a vocabulary filter}
\usage{
transcribeservice_get_vocabulary_filter(VocabularyFilterName)
}
\arguments{
\item{VocabularyFilterName}{[required] The name of the vocabulary filter for which to return information.}
}
\description{
Returns information about a vocabulary filter.
}
\section{Request syntax}{
\preformatted{svc$get_vocabulary_filter(
  VocabularyFilterName = "string"
)
}
}

\keyword{internal}
