% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_update_language_model}
\alias{transcribeservice_update_language_model}
\title{Updates the encryption configuration for an existing custom language
model}
\usage{
transcribeservice_update_language_model(
  ModelName,
  DataAccessRoleArn = NULL,
  EncryptionConfiguration = NULL
)
}
\arguments{
\item{ModelName}{[required] The name of the custom language model you want to update. Model names are case sensitive.}

\item{DataAccessRoleArn}{The Amazon Resource Name (ARN) of an IAM role. If you include \code{EncryptionConfiguration} in your request, this role must have permissions to access the specified KMS key. If the role that you specify doesn't have the appropriate permissions, your request fails.

IAM role ARNs have the format \code{arn:partition:iam::account:role/role-name-with-path}. For example: \verb{arn:aws:iam::111122223333:role/Admin}.

For more information, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference_identifiers.html#identifiers-arns}{IAM ARNs}.}

\item{EncryptionConfiguration}{Specifies the new encryption configuration for your custom language model. The model artifacts are re-encrypted in place using the specified KMS key or with an AWS-owned key if a key is not supplied.}
}
\description{
Updates the encryption configuration for an existing custom language model. You can use this operation to change the KMS key used to encrypt your model artifacts. The model artifacts are re-encrypted in place. No model training is required.

See \url{https://www.paws-r-sdk.com/docs/transcribeservice_update_language_model/} for full documentation.
}
\keyword{internal}
