% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_operations.R
\name{translate_list_languages}
\alias{translate_list_languages}
\title{Provides a list of languages (RFC-5646 codes and names) that Amazon
Translate supports}
\usage{
translate_list_languages(
  DisplayLanguageCode = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{DisplayLanguageCode}{The language code for the language to use to display the language names
in the response. The language code is \code{en} by default.}

\item{NextToken}{Include the NextToken value to fetch the next group of supported
languages.}

\item{MaxResults}{The maximum number of results to return in each response.}
}
\description{
Provides a list of languages (RFC-5646 codes and names) that Amazon Translate supports.

See \url{https://www.paws-r-sdk.com/docs/translate_list_languages/} for full documentation.
}
\keyword{internal}
