% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_operations.R
\name{translate_list_tags_for_resource}
\alias{translate_list_tags_for_resource}
\title{Lists all tags associated with a given Amazon Translate resource}
\usage{
translate_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the given Amazon Translate resource
you are querying.}
}
\description{
Lists all tags associated with a given Amazon Translate resource. For more information, see \href{https://docs.aws.amazon.com/translate/latest/dg/tagging.html}{Tagging your resources}.

See \url{https://www.paws-r-sdk.com/docs/translate_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
