% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_operations.R
\name{translate_list_text_translation_jobs}
\alias{translate_list_text_translation_jobs}
\title{Gets a list of the batch translation jobs that you have submitted}
\usage{
translate_list_text_translation_jobs(
  Filter = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Filter}{The parameters that specify which batch translation jobs to retrieve.
Filters include job name, job status, and submission time. You can only
set one filter at a time.}

\item{NextToken}{The token to request the next page of results.}

\item{MaxResults}{The maximum number of results to return in each page. The default value
is 100.}
}
\description{
Gets a list of the batch translation jobs that you have submitted.

See \url{https://paws-r.github.io/docs/translate/list_text_translation_jobs.html} for full documentation.
}
\keyword{internal}
