% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_operations.R
\name{translate_update_parallel_data}
\alias{translate_update_parallel_data}
\title{Updates a previously created parallel data resource by importing a new
input file from Amazon S3}
\usage{
translate_update_parallel_data(Name, Description, ParallelDataConfig,
  ClientToken)
}
\arguments{
\item{Name}{[required] The name of the parallel data resource being updated.}

\item{Description}{A custom description for the parallel data resource in Amazon Translate.}

\item{ParallelDataConfig}{[required] Specifies the format and S3 location of the parallel data input file.}

\item{ClientToken}{[required] A unique identifier for the request. This token is automatically
generated when you use Amazon Translate through an AWS SDK.}
}
\description{
Updates a previously created parallel data resource by importing a new
input file from Amazon S3.
}
\section{Request syntax}{
\preformatted{svc$update_parallel_data(
  Name = "string",
  Description = "string",
  ParallelDataConfig = list(
    S3Uri = "string",
    Format = "TSV"|"CSV"|"TMX"
  ),
  ClientToken = "string"
)
}
}

\keyword{internal}
