% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiceid_operations.R
\name{voiceid_associate_fraudster}
\alias{voiceid_associate_fraudster}
\title{Associates the fraudsters with the watchlist specified in the same
domain}
\usage{
voiceid_associate_fraudster(DomainId, FraudsterId, WatchlistId)
}
\arguments{
\item{DomainId}{[required] The identifier of the domain that contains the fraudster.}

\item{FraudsterId}{[required] The identifier of the fraudster to be associated with the watchlist.}

\item{WatchlistId}{[required] The identifier of the watchlist you want to associate with the
fraudster.}
}
\description{
Associates the fraudsters with the watchlist specified in the same domain.

See \url{https://paws-r.github.io/docs/voiceid/associate_fraudster.html} for full documentation.
}
\keyword{internal}
