% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiceid_operations.R
\name{voiceid_create_watchlist}
\alias{voiceid_create_watchlist}
\title{Creates a watchlist that fraudsters can be a part of}
\usage{
voiceid_create_watchlist(
  ClientToken = NULL,
  Description = NULL,
  DomainId,
  Name
)
}
\arguments{
\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If not provided, the Amazon Web Services SDK
populates this field. For more information about idempotency, see
\href{https://aws.amazon.com/builders-library/making-retries-safe-with-idempotent-APIs/}{Making retries safe with idempotent APIs}.}

\item{Description}{A brief description of this watchlist.}

\item{DomainId}{[required] The identifier of the domain that contains the watchlist.}

\item{Name}{[required] The name of the watchlist.}
}
\description{
Creates a watchlist that fraudsters can be a part of.

See \url{https://www.paws-r-sdk.com/docs/voiceid_create_watchlist/} for full documentation.
}
\keyword{internal}
