% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiceid_operations.R
\name{voiceid_delete_domain}
\alias{voiceid_delete_domain}
\title{Deletes the specified domain from Voice ID}
\usage{
voiceid_delete_domain(DomainId)
}
\arguments{
\item{DomainId}{[required] The identifier of the domain you want to delete.}
}
\description{
Deletes the specified domain from Voice ID.

See \url{https://paws-r.github.io/docs/voiceid/delete_domain.html} for full documentation.
}
\keyword{internal}
