% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiceid_operations.R
\name{voiceid_delete_fraudster}
\alias{voiceid_delete_fraudster}
\title{Deletes the specified fraudster from Voice ID}
\usage{
voiceid_delete_fraudster(DomainId, FraudsterId)
}
\arguments{
\item{DomainId}{[required] The identifier of the domain containing the fraudster.}

\item{FraudsterId}{[required] The identifier of the fraudster you want to delete.}
}
\description{
Deletes the specified fraudster from Voice ID.

See \url{https://paws-r.github.io/docs/voiceid/delete_fraudster.html} for full documentation.
}
\keyword{internal}
