% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiceid_operations.R
\name{voiceid_delete_speaker}
\alias{voiceid_delete_speaker}
\title{Deletes the specified speaker from Voice ID}
\usage{
voiceid_delete_speaker(DomainId, SpeakerId)
}
\arguments{
\item{DomainId}{[required] The identifier of the domain that contains the speaker.}

\item{SpeakerId}{[required] The identifier of the speaker you want to delete.}
}
\description{
Deletes the specified speaker from Voice ID.

See \url{https://www.paws-r-sdk.com/docs/voiceid_delete_speaker/} for full documentation.
}
\keyword{internal}
