% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiceid_operations.R
\name{voiceid_delete_watchlist}
\alias{voiceid_delete_watchlist}
\title{Deletes the specified watchlist from Voice ID}
\usage{
voiceid_delete_watchlist(DomainId, WatchlistId)
}
\arguments{
\item{DomainId}{[required] The identifier of the domain that contains the watchlist.}

\item{WatchlistId}{[required] The identifier of the watchlist to be deleted.}
}
\description{
Deletes the specified watchlist from Voice ID. This API throws an exception when there are fraudsters in the watchlist that you are trying to delete. You must delete the fraudsters, and then delete the watchlist. Every domain has a default watchlist which cannot be deleted.

See \url{https://www.paws-r-sdk.com/docs/voiceid_delete_watchlist/} for full documentation.
}
\keyword{internal}
