% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiceid_operations.R
\name{voiceid_describe_domain}
\alias{voiceid_describe_domain}
\title{Describes the specified domain}
\usage{
voiceid_describe_domain(DomainId)
}
\arguments{
\item{DomainId}{[required] The identifier of the domain that you are describing.}
}
\description{
Describes the specified domain.

See \url{https://paws-r.github.io/docs/voiceid/describe_domain.html} for full documentation.
}
\keyword{internal}
