% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiceid_operations.R
\name{voiceid_describe_fraudster}
\alias{voiceid_describe_fraudster}
\title{Describes the specified fraudster}
\usage{
voiceid_describe_fraudster(DomainId, FraudsterId)
}
\arguments{
\item{DomainId}{[required] The identifier of the domain that contains the fraudster.}

\item{FraudsterId}{[required] The identifier of the fraudster you are describing.}
}
\description{
Describes the specified fraudster.

See \url{https://www.paws-r-sdk.com/docs/voiceid_describe_fraudster/} for full documentation.
}
\keyword{internal}
