% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiceid_operations.R
\name{voiceid_describe_fraudster_registration_job}
\alias{voiceid_describe_fraudster_registration_job}
\title{Describes the specified fraudster registration job}
\usage{
voiceid_describe_fraudster_registration_job(DomainId, JobId)
}
\arguments{
\item{DomainId}{[required] The identifier of the domain that contains the fraudster registration
job.}

\item{JobId}{[required] The identifier of the fraudster registration job you are describing.}
}
\description{
Describes the specified fraudster registration job.

See \url{https://www.paws-r-sdk.com/docs/voiceid_describe_fraudster_registration_job/} for full documentation.
}
\keyword{internal}
