% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiceid_operations.R
\name{voiceid_describe_speaker}
\alias{voiceid_describe_speaker}
\title{Describes the specified speaker}
\usage{
voiceid_describe_speaker(DomainId, SpeakerId)
}
\arguments{
\item{DomainId}{[required] The identifier of the domain that contains the speaker.}

\item{SpeakerId}{[required] The identifier of the speaker you are describing.}
}
\description{
Describes the specified speaker.

See \url{https://paws-r.github.io/docs/voiceid/describe_speaker.html} for full documentation.
}
\keyword{internal}
