% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiceid_operations.R
\name{voiceid_describe_speaker_enrollment_job}
\alias{voiceid_describe_speaker_enrollment_job}
\title{Describes the specified speaker enrollment job}
\usage{
voiceid_describe_speaker_enrollment_job(DomainId, JobId)
}
\arguments{
\item{DomainId}{[required] The identifier of the domain that contains the speaker enrollment job.}

\item{JobId}{[required] The identifier of the speaker enrollment job you are describing.}
}
\description{
Describes the specified speaker enrollment job.

See \url{https://paws-r.github.io/docs/voiceid/describe_speaker_enrollment_job.html} for full documentation.
}
\keyword{internal}
