% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiceid_operations.R
\name{voiceid_describe_watchlist}
\alias{voiceid_describe_watchlist}
\title{Describes the specified watchlist}
\usage{
voiceid_describe_watchlist(DomainId, WatchlistId)
}
\arguments{
\item{DomainId}{[required] The identifier of the domain that contains the watchlist.}

\item{WatchlistId}{[required] The identifier of the watchlist that you are describing.}
}
\description{
Describes the specified watchlist.

See \url{https://paws-r.github.io/docs/voiceid/describe_watchlist.html} for full documentation.
}
\keyword{internal}
