% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiceid_operations.R
\name{voiceid_disassociate_fraudster}
\alias{voiceid_disassociate_fraudster}
\title{Disassociates the fraudsters from the watchlist specified}
\usage{
voiceid_disassociate_fraudster(DomainId, FraudsterId, WatchlistId)
}
\arguments{
\item{DomainId}{[required] The identifier of the domain that contains the fraudster.}

\item{FraudsterId}{[required] The identifier of the fraudster to be disassociated from the watchlist.}

\item{WatchlistId}{[required] The identifier of the watchlist that you want to disassociate from the
fraudster.}
}
\description{
Disassociates the fraudsters from the watchlist specified. Voice ID always expects a fraudster to be a part of at least one watchlist. If you try to disassociate a fraudster from its only watchlist, a \code{ValidationException} is thrown.

See \url{https://www.paws-r-sdk.com/docs/voiceid_disassociate_fraudster/} for full documentation.
}
\keyword{internal}
