% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiceid_operations.R
\name{voiceid_evaluate_session}
\alias{voiceid_evaluate_session}
\title{Evaluates a specified session based on audio data accumulated during a
streaming Amazon Connect Voice ID call}
\usage{
voiceid_evaluate_session(DomainId, SessionNameOrId)
}
\arguments{
\item{DomainId}{[required] The identifier of the domain where the session started.}

\item{SessionNameOrId}{[required] The session identifier, or name of the session, that you want to
evaluate. In Voice ID integration, this is the Contact-Id.}
}
\description{
Evaluates a specified session based on audio data accumulated during a streaming Amazon Connect Voice ID call.

See \url{https://paws-r.github.io/docs/voiceid/evaluate_session.html} for full documentation.
}
\keyword{internal}
