% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiceid_operations.R
\name{voiceid_list_domains}
\alias{voiceid_list_domains}
\title{Lists all the domains in the Amazon Web Services account}
\usage{
voiceid_list_domains(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of results that are returned per call. You can use
\code{NextToken} to obtain further pages of results. The default is 100; the
maximum allowed page size is also 100.}

\item{NextToken}{If \code{NextToken} is returned, there are more results available. The value
of \code{NextToken} is a unique pagination token for each page. Make the call
again using the returned token to retrieve the next page. Keep all other
arguments unchanged. Each pagination token expires after 24 hours.}
}
\description{
Lists all the domains in the Amazon Web Services account.

See \url{https://paws-r.github.io/docs/voiceid/list_domains.html} for full documentation.
}
\keyword{internal}
