% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiceid_operations.R
\name{voiceid_list_tags_for_resource}
\alias{voiceid_list_tags_for_resource}
\title{Lists all tags associated with a specified Voice ID resource}
\usage{
voiceid_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the Voice ID resource for which you
want to list the tags.}
}
\description{
Lists all tags associated with a specified Voice ID resource.

See \url{https://paws-r.github.io/docs/voiceid/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
