% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiceid_operations.R
\name{voiceid_opt_out_speaker}
\alias{voiceid_opt_out_speaker}
\title{Opts out a speaker from Voice ID}
\usage{
voiceid_opt_out_speaker(DomainId, SpeakerId)
}
\arguments{
\item{DomainId}{[required] The identifier of the domain that contains the speaker.}

\item{SpeakerId}{[required] The identifier of the speaker you want opted-out.}
}
\description{
Opts out a speaker from Voice ID. A speaker can be opted out regardless of whether or not they already exist in Voice ID. If they don't yet exist, a new speaker is created in an opted out state. If they already exist, their existing status is overridden and they are opted out. Enrollment and evaluation authentication requests are rejected for opted out speakers, and opted out speakers have no voice embeddings stored in Voice ID.

See \url{https://www.paws-r-sdk.com/docs/voiceid_opt_out_speaker/} for full documentation.
}
\keyword{internal}
