% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiceid_operations.R
\name{voiceid_tag_resource}
\alias{voiceid_tag_resource}
\title{Tags a Voice ID resource with the provided list of tags}
\usage{
voiceid_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the Voice ID resource you want to tag.}

\item{Tags}{[required] The list of tags to assign to the specified resource.}
}
\description{
Tags a Voice ID resource with the provided list of tags.

See \url{https://paws-r.github.io/docs/voiceid/tag_resource.html} for full documentation.
}
\keyword{internal}
