% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiceid_operations.R
\name{voiceid_untag_resource}
\alias{voiceid_untag_resource}
\title{Removes specified tags from a specified Amazon Connect Voice ID resource}
\usage{
voiceid_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the Voice ID resource you want to
remove tags from.}

\item{TagKeys}{[required] The list of tag keys you want to remove from the specified resource.}
}
\description{
Removes specified tags from a specified Amazon Connect Voice ID resource.

See \url{https://www.paws-r-sdk.com/docs/voiceid_untag_resource/} for full documentation.
}
\keyword{internal}
