% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiceid_operations.R
\name{voiceid_update_watchlist}
\alias{voiceid_update_watchlist}
\title{Updates the specified watchlist}
\usage{
voiceid_update_watchlist(
  Description = NULL,
  DomainId,
  Name = NULL,
  WatchlistId
)
}
\arguments{
\item{Description}{A brief description about this watchlist.}

\item{DomainId}{[required] The identifier of the domain that contains the watchlist.}

\item{Name}{The name of the watchlist.}

\item{WatchlistId}{[required] The identifier of the watchlist to be updated.}
}
\description{
Updates the specified watchlist. Every domain has a default watchlist which cannot be updated.

See \url{https://www.paws-r-sdk.com/docs/voiceid_update_watchlist/} for full documentation.
}
\keyword{internal}
