% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_create_advanced_prompt_optimization_job}
\alias{bedrock_create_advanced_prompt_optimization_job}
\title{Creates an asynchronous batch job for advanced prompt optimization}
\usage{
bedrock_create_advanced_prompt_optimization_job(
  jobName,
  jobDescription = NULL,
  clientToken = NULL,
  inputConfig,
  outputConfig,
  encryptionKeyArn = NULL,
  tags = NULL,
  modelConfigurations
)
}
\arguments{
\item{jobName}{[required] Name of the advanced prompt optimization job.}

\item{jobDescription}{Description of the advanced prompt optimization job.}

\item{clientToken}{Idempotency token for the request.}

\item{inputConfig}{[required] Input data configuration for the advanced prompt optimization job.}

\item{outputConfig}{[required] Output data configuration for the advanced prompt optimization job.}

\item{encryptionKeyArn}{KMS key ARN for encrypting output data.}

\item{tags}{Tags to associate with the job.}

\item{modelConfigurations}{[required] Model configurations for advanced prompt optimization.}
}
\description{
Creates an asynchronous batch job for advanced prompt optimization.

See \url{https://www.paws-r-sdk.com/docs/bedrock_create_advanced_prompt_optimization_job/} for full documentation.
}
\keyword{internal}
