% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_delete_custom_model}
\alias{bedrock_delete_custom_model}
\title{Deletes a custom model that you created earlier}
\usage{
bedrock_delete_custom_model(modelIdentifier)
}
\arguments{
\item{modelIdentifier}{[required] Name of the model to delete.}
}
\description{
Deletes a custom model that you created earlier. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/custom-models.html}{Custom models} in the Amazon Bedrock User Guide.

See \url{https://www.paws-r-sdk.com/docs/bedrock_delete_custom_model/} for full documentation.
}
\keyword{internal}
