% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_delete_provisioned_model_throughput}
\alias{bedrock_delete_provisioned_model_throughput}
\title{Deletes a Provisioned Throughput}
\usage{
bedrock_delete_provisioned_model_throughput(provisionedModelId)
}
\arguments{
\item{provisionedModelId}{[required] The Amazon Resource Name (ARN) or name of the Provisioned Throughput.}
}
\description{
Deletes a Provisioned Throughput. You can't delete a Provisioned Throughput before the commitment term is over. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/prov-throughput.html}{Provisioned Throughput} in the Amazon Bedrock User Guide.

See \url{https://www.paws-r-sdk.com/docs/bedrock_delete_provisioned_model_throughput/} for full documentation.
}
\keyword{internal}
