% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_get_model_copy_job}
\alias{bedrock_get_model_copy_job}
\title{Retrieves information about a model copy job}
\usage{
bedrock_get_model_copy_job(jobArn)
}
\arguments{
\item{jobArn}{[required] The Amazon Resource Name (ARN) of the model copy job.}
}
\description{
Retrieves information about a model copy job. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/copy-model.html}{Copy models to be used in other regions} in the \href{https://docs.aws.amazon.com/bedrock/latest/userguide/what-is-bedrock.html?trkcampaign=awsomedayonlinehk}{Amazon Bedrock User Guide}.

See \url{https://www.paws-r-sdk.com/docs/bedrock_get_model_copy_job/} for full documentation.
}
\keyword{internal}
