% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_get_model_customization_job}
\alias{bedrock_get_model_customization_job}
\title{Retrieves the properties associated with a model-customization job,
including the status of the job}
\usage{
bedrock_get_model_customization_job(jobIdentifier)
}
\arguments{
\item{jobIdentifier}{[required] Identifier for the customization job.}
}
\description{
Retrieves the properties associated with a model-customization job, including the status of the job. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/custom-models.html}{Custom models} in the \href{https://docs.aws.amazon.com/bedrock/latest/userguide/what-is-bedrock.html?sc_campaign=genaiwave&sc_channel=el&sc_content=working-with-your-live-data-using-langchain&sc_country=mult&sc_geo=mult&sc_outcome=acq}{Amazon Bedrock User Guide}.

See \url{https://www.paws-r-sdk.com/docs/bedrock_get_model_customization_job/} for full documentation.
}
\keyword{internal}
