% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_list_advanced_prompt_optimization_jobs}
\alias{bedrock_list_advanced_prompt_optimization_jobs}
\title{Lists all advanced prompt optimization jobs for the account}
\usage{
bedrock_list_advanced_prompt_optimization_jobs(
  maxResults = NULL,
  nextToken = NULL,
  sortBy = NULL,
  sortOrder = NULL
)
}
\arguments{
\item{maxResults}{Maximum number of results to return.}

\item{nextToken}{Pagination token for the next page of results.}

\item{sortBy}{Field to sort by in the returned list of jobs.}

\item{sortOrder}{Sort order for the results.}
}
\description{
Lists all advanced prompt optimization jobs for the account.

See \url{https://www.paws-r-sdk.com/docs/bedrock_list_advanced_prompt_optimization_jobs/} for full documentation.
}
\keyword{internal}
