% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_list_foundation_models}
\alias{bedrock_list_foundation_models}
\title{Lists Amazon Bedrock foundation models that you can use}
\usage{
bedrock_list_foundation_models(
  byProvider = NULL,
  byCustomizationType = NULL,
  byOutputModality = NULL,
  byInferenceType = NULL
)
}
\arguments{
\item{byProvider}{Return models belonging to the model provider that you specify.}

\item{byCustomizationType}{Return models that support the customization type that you specify. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/custom-models.html}{Custom models} in the \href{https://docs.aws.amazon.com/bedrock/latest/userguide/what-is-bedrock.html?trkcampaign=awsomedayonlinehk}{Amazon Bedrock User Guide}.}

\item{byOutputModality}{Return models that support the output modality that you specify.}

\item{byInferenceType}{Return models that support the inference type that you specify. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/prov-throughput.html}{Provisioned Throughput} in the \href{https://docs.aws.amazon.com/bedrock/latest/userguide/what-is-bedrock.html?trkcampaign=awsomedayonlinehk}{Amazon Bedrock User Guide}.}
}
\description{
Lists Amazon Bedrock foundation models that you can use. You can filter the results with the request parameters. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/}{Foundation models} in the \href{https://docs.aws.amazon.com/bedrock/latest/userguide/what-is-bedrock.html?trkcampaign=awsomedayonlinehk}{Amazon Bedrock User Guide}.

See \url{https://www.paws-r-sdk.com/docs/bedrock_list_foundation_models/} for full documentation.
}
\keyword{internal}
