% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_list_foundation_models}
\alias{bedrock_list_foundation_models}
\title{List of Bedrock foundation models that you can use}
\usage{
bedrock_list_foundation_models(
  byProvider = NULL,
  byCustomizationType = NULL,
  byOutputModality = NULL,
  byInferenceType = NULL
)
}
\arguments{
\item{byProvider}{A Bedrock model provider.}

\item{byCustomizationType}{List by customization type.}

\item{byOutputModality}{List by output modality type.}

\item{byInferenceType}{List by inference type.}
}
\description{
List of Bedrock foundation models that you can use. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/foundation-models.html}{Foundation models} in the Bedrock User Guide.

See \url{https://www.paws-r-sdk.com/docs/bedrock_list_foundation_models/} for full documentation.
}
\keyword{internal}
