% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_list_inference_profiles}
\alias{bedrock_list_inference_profiles}
\title{Returns a list of inference profiles that you can use}
\usage{
bedrock_list_inference_profiles(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of results to return in the response. If the total
number of results is greater than this value, use the token returned in
the response in the \code{nextToken} field when making another request to
return the next batch of results.}

\item{nextToken}{If the total number of results is greater than the \code{maxResults} value
provided in the request, enter the token returned in the \code{nextToken}
field in the response in this field to return the next batch of results.}
}
\description{
Returns a list of inference profiles that you can use.

See \url{https://www.paws-r-sdk.com/docs/bedrock_list_inference_profiles/} for full documentation.
}
\keyword{internal}
