% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_stop_advanced_prompt_optimization_job}
\alias{bedrock_stop_advanced_prompt_optimization_job}
\title{Stops an in-progress advanced prompt optimization job}
\usage{
bedrock_stop_advanced_prompt_optimization_job(jobIdentifier)
}
\arguments{
\item{jobIdentifier}{[required] ARN or ID of the advanced prompt optimization job to stop.}
}
\description{
Stops an in-progress advanced prompt optimization job.

See \url{https://www.paws-r-sdk.com/docs/bedrock_stop_advanced_prompt_optimization_job/} for full documentation.
}
\keyword{internal}
