% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_stop_model_customization_job}
\alias{bedrock_stop_model_customization_job}
\title{Stops an active model customization job}
\usage{
bedrock_stop_model_customization_job(jobIdentifier)
}
\arguments{
\item{jobIdentifier}{[required] Job identifier of the job to stop.}
}
\description{
Stops an active model customization job. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/custom-models.html}{Custom models} in the \href{https://docs.aws.amazon.com/bedrock/latest/userguide/what-is-bedrock.html?sc_campaign=genaiwave&sc_channel=el&sc_content=working-with-your-live-data-using-langchain&sc_country=mult&sc_geo=mult&sc_outcome=acq}{Amazon Bedrock User Guide}.

See \url{https://www.paws-r-sdk.com/docs/bedrock_stop_model_customization_job/} for full documentation.
}
\keyword{internal}
