% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_tag_resource}
\alias{bedrock_tag_resource}
\title{Associate tags with a resource}
\usage{
bedrock_tag_resource(resourceARN, tags)
}
\arguments{
\item{resourceARN}{[required] The Amazon Resource Name (ARN) of the resource to tag.}

\item{tags}{[required] Tags to associate with the resource.}
}
\description{
Associate tags with a resource. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/what-is-bedrock.html?sc_campaign=genaiwave&sc_channel=el&sc_content=working-with-your-live-data-using-langchain&sc_country=mult&sc_geo=mult&sc_outcome=acq}{Tagging resources} in the \href{https://docs.aws.amazon.com/bedrock/latest/userguide/what-is-bedrock.html?sc_campaign=genaiwave&sc_channel=el&sc_content=working-with-your-live-data-using-langchain&sc_country=mult&sc_geo=mult&sc_outcome=acq}{Amazon Bedrock User Guide}.

See \url{https://www.paws-r-sdk.com/docs/bedrock_tag_resource/} for full documentation.
}
\keyword{internal}
