% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_untag_resource}
\alias{bedrock_untag_resource}
\title{Remove one or more tags from a resource}
\usage{
bedrock_untag_resource(resourceARN, tagKeys)
}
\arguments{
\item{resourceARN}{[required] The Amazon Resource Name (ARN) of the resource to untag.}

\item{tagKeys}{[required] Tag keys of the tags to remove from the resource.}
}
\description{
Remove one or more tags from a resource. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/what-is-bedrock.html?sc_campaign=genaiwave&sc_channel=el&sc_content=working-with-your-live-data-using-langchain&sc_country=mult&sc_geo=mult&sc_outcome=acq}{Tagging resources} in the \href{https://docs.aws.amazon.com/bedrock/latest/userguide/what-is-bedrock.html?sc_campaign=genaiwave&sc_channel=el&sc_content=working-with-your-live-data-using-langchain&sc_country=mult&sc_geo=mult&sc_outcome=acq}{Amazon Bedrock User Guide}.

See \url{https://www.paws-r-sdk.com/docs/bedrock_untag_resource/} for full documentation.
}
\keyword{internal}
