% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrock_operations.R
\name{bedrock_update_provisioned_model_throughput}
\alias{bedrock_update_provisioned_model_throughput}
\title{Update a provisioned throughput}
\usage{
bedrock_update_provisioned_model_throughput(
  provisionedModelId,
  desiredProvisionedModelName = NULL,
  desiredModelId = NULL
)
}
\arguments{
\item{provisionedModelId}{[required] The ARN or name of the provisioned throughput to update.}

\item{desiredProvisionedModelName}{The new name for this provisioned throughput.}

\item{desiredModelId}{The ARN of the new model to associate with this provisioned throughput.}
}
\description{
Update a provisioned throughput. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/what-is-bedrock.html}{Provisioned throughput} in the Bedrock User Guide.

See \url{https://www.paws-r-sdk.com/docs/bedrock_update_provisioned_model_throughput/} for full documentation.
}
\keyword{internal}
