% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_associate_agent_collaborator}
\alias{bedrockagent_associate_agent_collaborator}
\title{Makes an agent a collaborator for another agent}
\usage{
bedrockagent_associate_agent_collaborator(
  agentId,
  agentVersion,
  agentDescriptor,
  collaboratorName,
  collaborationInstruction,
  relayConversationHistory = NULL,
  clientToken = NULL
)
}
\arguments{
\item{agentId}{[required] The agent's ID.}

\item{agentVersion}{[required] An agent version.}

\item{agentDescriptor}{[required] The alias of the collaborator agent.}

\item{collaboratorName}{[required] A name for the collaborator.}

\item{collaborationInstruction}{[required] Instruction for the collaborator.}

\item{relayConversationHistory}{A relay conversation history for the collaborator.}

\item{clientToken}{A client token.}
}
\description{
Makes an agent a collaborator for another agent.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_associate_agent_collaborator/} for full documentation.
}
\keyword{internal}
