% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_get_flow}
\alias{bedrockagent_get_flow}
\title{Retrieves information about a flow}
\usage{
bedrockagent_get_flow(flowIdentifier)
}
\arguments{
\item{flowIdentifier}{[required] The unique identifier of the flow.}
}
\description{
Retrieves information about a flow. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/}{Manage a flow in Amazon Bedrock} in the Amazon Bedrock User Guide.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_get_flow/} for full documentation.
}
\keyword{internal}
