% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagent_operations.R
\name{bedrockagent_tag_resource}
\alias{bedrockagent_tag_resource}
\title{Associate tags with a resource}
\usage{
bedrockagent_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource to tag.}

\item{tags}{[required] An object containing key-value pairs that define the tags to attach to the resource.}
}
\description{
Associate tags with a resource. For more information, see \href{https://docs.aws.amazon.com/bedrock/latest/userguide/what-is-bedrock.html?sc_campaign=genaiwave&sc_channel=el&sc_content=working-with-your-live-data-using-langchain&sc_country=mult&sc_geo=mult&sc_outcome=acq}{Tagging resources} in the Amazon Bedrock User Guide.

See \url{https://www.paws-r-sdk.com/docs/bedrockagent_tag_resource/} for full documentation.
}
\keyword{internal}
