% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_delete_memory_record}
\alias{bedrockagentcore_delete_memory_record}
\title{Deletes a memory record from an AgentCore Memory resource}
\usage{
bedrockagentcore_delete_memory_record(memoryId, memoryRecordId)
}
\arguments{
\item{memoryId}{[required] The identifier of the AgentCore Memory resource from which to delete the
memory record.}

\item{memoryRecordId}{[required] The identifier of the memory record to delete.}
}
\description{
Deletes a memory record from an AgentCore Memory resource. When you delete a memory record, it is permanently removed.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_delete_memory_record/} for full documentation.
}
\keyword{internal}
