% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_get_memory_record}
\alias{bedrockagentcore_get_memory_record}
\title{Retrieves a specific memory record from an AgentCore Memory resource}
\usage{
bedrockagentcore_get_memory_record(memoryId, memoryRecordId)
}
\arguments{
\item{memoryId}{[required] The identifier of the AgentCore Memory resource containing the memory
record.}

\item{memoryRecordId}{[required] The identifier of the memory record to retrieve.}
}
\description{
Retrieves a specific memory record from an AgentCore Memory resource.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_get_memory_record/} for full documentation.
}
\keyword{internal}
