% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_process_payment}
\alias{bedrockagentcore_process_payment}
\title{Process a payment transaction}
\usage{
bedrockagentcore_process_payment(
  userId = NULL,
  agentName = NULL,
  paymentManagerArn,
  paymentSessionId,
  paymentInstrumentId,
  paymentType,
  paymentInput,
  clientToken = NULL
)
}
\arguments{
\item{userId}{The user ID associated with this payment.}

\item{agentName}{The agent name associated with this request, used for observability.}

\item{paymentManagerArn}{[required] The ARN of the payment manager handling this payment.}

\item{paymentSessionId}{[required] The ID of the payment session for this transaction.}

\item{paymentInstrumentId}{[required] The ID of the payment instrument to use for this transaction.}

\item{paymentType}{[required] The type of payment being processed.}

\item{paymentInput}{[required] The payment input details specific to the payment type.}

\item{clientToken}{Idempotency token to ensure request uniqueness.}
}
\description{
Process a payment transaction

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_process_payment/} for full documentation.
}
\keyword{internal}
