% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcore_operations.R
\name{bedrockagentcore_start_batch_evaluation}
\alias{bedrockagentcore_start_batch_evaluation}
\title{Starts a batch evaluation job that evaluates agent performance across
multiple sessions}
\usage{
bedrockagentcore_start_batch_evaluation(
  batchEvaluationName,
  evaluators = NULL,
  dataSourceConfig,
  clientToken = NULL,
  evaluationMetadata = NULL,
  description = NULL
)
}
\arguments{
\item{batchEvaluationName}{[required] The name of the batch evaluation. Must be unique within your account.}

\item{evaluators}{The list of evaluators to apply during the batch evaluation. Can include
both built-in evaluators and custom evaluators. Maximum of 10
evaluators.}

\item{dataSourceConfig}{[required] The data source configuration that specifies where to pull agent session
traces from for evaluation.}

\item{clientToken}{A unique, case-sensitive identifier to ensure that the API request
completes no more than one time. If this token matches a previous
request, the service ignores the request, but does not return an error.}

\item{evaluationMetadata}{Optional metadata for the evaluation, including session-specific ground
truth data and test scenario identifiers.}

\item{description}{The description of the batch evaluation.}
}
\description{
Starts a batch evaluation job that evaluates agent performance across multiple sessions. Batch evaluations pull agent traces from CloudWatch Logs or an existing online evaluation configuration and run specified evaluators and insights against them.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcore_start_batch_evaluation/} for full documentation.
}
\keyword{internal}
