% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_create_agent_runtime}
\alias{bedrockagentcorecontrol_create_agent_runtime}
\title{Creates an Amazon Bedrock AgentCore Runtime}
\usage{
bedrockagentcorecontrol_create_agent_runtime(
  agentRuntimeName,
  agentRuntimeArtifact,
  roleArn,
  networkConfiguration,
  clientToken = NULL,
  description = NULL,
  authorizerConfiguration = NULL,
  requestHeaderConfiguration = NULL,
  protocolConfiguration = NULL,
  lifecycleConfiguration = NULL,
  environmentVariables = NULL,
  filesystemConfigurations = NULL,
  tags = NULL
)
}
\arguments{
\item{agentRuntimeName}{[required] The name of the AgentCore Runtime.}

\item{agentRuntimeArtifact}{[required] The artifact of the AgentCore Runtime.}

\item{roleArn}{[required] The IAM role ARN that provides permissions for the AgentCore Runtime.}

\item{networkConfiguration}{[required] The network configuration for the AgentCore Runtime.}

\item{clientToken}{A unique, case-sensitive identifier to ensure idempotency of the
request.}

\item{description}{The description of the AgentCore Runtime.}

\item{authorizerConfiguration}{The authorizer configuration for the AgentCore Runtime.}

\item{requestHeaderConfiguration}{Configuration for HTTP request headers that will be passed through to
the runtime.}

\item{lifecycleConfiguration}{The life cycle configuration for the AgentCore Runtime.}

\item{environmentVariables}{Environment variables to set in the AgentCore Runtime environment.}

\item{filesystemConfigurations}{The filesystem configurations to mount into the AgentCore Runtime. Use
filesystem configurations to provide persistent storage to your
AgentCore Runtime sessions.}

\item{tags}{A map of tag keys and values to assign to the agent runtime. Tags enable
you to categorize your resources in different ways, for example, by
purpose, owner, or environment.}
}
\description{
Creates an Amazon Bedrock AgentCore Runtime.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_create_agent_runtime/} for full documentation.
}
\keyword{internal}
