% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_create_api_key_credential_provider}
\alias{bedrockagentcorecontrol_create_api_key_credential_provider}
\title{Creates a new API key credential provider}
\usage{
bedrockagentcorecontrol_create_api_key_credential_provider(
  name,
  apiKey = NULL,
  apiKeySecretConfig = NULL,
  apiKeySecretSource = NULL,
  tags = NULL
)
}
\arguments{
\item{name}{[required] The name of the API key credential provider. The name must be unique within your account.}

\item{apiKey}{The API key to use for authentication. This value is encrypted and stored securely.}

\item{apiKeySecretConfig}{A reference to the Amazon Web Services Secrets Manager secret that stores the API key. This includes the secret ID and the JSON key used to extract the API key value from the secret. Required when \code{apiKeySecretSource} is set to \code{EXTERNAL}.}

\item{apiKeySecretSource}{The source type of the API key secret. Use \code{MANAGED} if the secret is managed by the service, or \code{EXTERNAL} if you manage the secret yourself in Amazon Web Services Secrets Manager.}

\item{tags}{A map of tag keys and values to assign to the API key credential provider. Tags enable you to categorize your resources in different ways, for example, by purpose, owner, or environment.}
}
\description{
Creates a new API key credential provider.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_create_api_key_credential_provider/} for full documentation.
}
\keyword{internal}
