% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_create_api_key_credential_provider}
\alias{bedrockagentcorecontrol_create_api_key_credential_provider}
\title{Creates a new API key credential provider}
\usage{
bedrockagentcorecontrol_create_api_key_credential_provider(
  name,
  apiKey,
  tags = NULL
)
}
\arguments{
\item{name}{[required] The name of the API key credential provider. The name must be unique
within your account.}

\item{apiKey}{[required] The API key to use for authentication. This value is encrypted and
stored securely.}

\item{tags}{A map of tag keys and values to assign to the API key credential
provider. Tags enable you to categorize your resources in different
ways, for example, by purpose, owner, or environment.}
}
\description{
Creates a new API key credential provider.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_create_api_key_credential_provider/} for full documentation.
}
\keyword{internal}
