% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_create_dataset_version}
\alias{bedrockagentcorecontrol_create_dataset_version}
\title{Publishes the current DRAFT as a new numbered version}
\usage{
bedrockagentcorecontrol_create_dataset_version(datasetId, clientToken = NULL)
}
\arguments{
\item{datasetId}{[required] The unique identifier of the dataset to publish a version for.}

\item{clientToken}{A unique, case-sensitive identifier to ensure that the API request completes no more than one time. If you don't specify this field, a value is randomly generated for you. If this token matches a previous request, the service ignores the request, but doesn't return an error. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring idempotency}.}
}
\description{
Publishes the current DRAFT as a new numbered version.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_create_dataset_version/} for full documentation.
}
\keyword{internal}
