% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_delete_dataset}
\alias{bedrockagentcorecontrol_delete_dataset}
\title{Deletes a dataset version or an entire dataset (all versions + name
claim)}
\usage{
bedrockagentcorecontrol_delete_dataset(datasetId, datasetVersion = NULL)
}
\arguments{
\item{datasetId}{[required] The unique identifier of the dataset to delete.}

\item{datasetVersion}{Optional version to delete. Use "DRAFT" or omit to delete the draft. Returns ResourceNotFoundException if the specified version does not exist.}
}
\description{
Deletes a dataset version or an entire dataset (all versions + name claim). Asynchronous 202.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_delete_dataset/} for full documentation.
}
\keyword{internal}
