% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_delete_dataset_examples}
\alias{bedrockagentcorecontrol_delete_dataset_examples}
\title{Deletes specific examples by ID from DRAFT}
\usage{
bedrockagentcorecontrol_delete_dataset_examples(
  datasetId,
  clientToken = NULL,
  exampleIds
)
}
\arguments{
\item{datasetId}{[required] The unique identifier of the dataset.}

\item{clientToken}{A unique, case-sensitive identifier to ensure that the API request completes no more than one time. If you don't specify this field, a value is randomly generated for you. If this token matches a previous request, the service ignores the request, but doesn't return an error. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring idempotency}.}

\item{exampleIds}{[required] The IDs of the examples to delete.}
}
\description{
Deletes specific examples by ID from DRAFT.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_delete_dataset_examples/} for full documentation.
}
\keyword{internal}
