% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_get_harness}
\alias{bedrockagentcorecontrol_get_harness}
\title{Operation to get a single Harness}
\usage{
bedrockagentcorecontrol_get_harness(harnessId)
}
\arguments{
\item{harnessId}{[required] The ID of the harness to retrieve.}
}
\description{
Operation to get a single Harness.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_get_harness/} for full documentation.
}
\keyword{internal}
