% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_update_agent_runtime}
\alias{bedrockagentcorecontrol_update_agent_runtime}
\title{Updates an existing Amazon Secure Agent}
\usage{
bedrockagentcorecontrol_update_agent_runtime(
  agentRuntimeId,
  agentRuntimeArtifact,
  roleArn,
  networkConfiguration = NULL,
  description = NULL,
  authorizerConfiguration = NULL,
  requestHeaderConfiguration = NULL,
  protocolConfiguration = NULL,
  lifecycleConfiguration = NULL,
  metadataConfiguration = NULL,
  environmentVariables = NULL,
  filesystemConfigurations = NULL,
  capacityProviderConfiguration = NULL,
  clientToken = NULL
)
}
\arguments{
\item{agentRuntimeId}{[required] The unique identifier of the AgentCore Runtime to update.}

\item{agentRuntimeArtifact}{[required] The updated artifact of the AgentCore Runtime.}

\item{roleArn}{[required] The updated IAM role ARN that provides permissions for the AgentCore Runtime.}

\item{networkConfiguration}{The updated network configuration for the AgentCore Runtime.}

\item{description}{The updated description of the AgentCore Runtime.}

\item{authorizerConfiguration}{The updated authorizer configuration for the AgentCore Runtime.}

\item{requestHeaderConfiguration}{The updated configuration for HTTP request headers that will be passed through to the runtime.}

\item{protocolConfiguration}{The protocol configuration for an agent runtime. This structure defines how the agent runtime communicates with clients.}

\item{lifecycleConfiguration}{The updated life cycle configuration for the AgentCore Runtime.}

\item{metadataConfiguration}{The updated configuration for microVM Metadata Service (MMDS) settings for the AgentCore Runtime.}

\item{environmentVariables}{Updated environment variables to set in the AgentCore Runtime environment.}

\item{filesystemConfigurations}{The updated filesystem configurations to mount into the AgentCore Runtime.}

\item{capacityProviderConfiguration}{The updated capacity provider configuration for the AgentCore Runtime.}

\item{clientToken}{A unique, case-sensitive identifier to ensure idempotency of the request.}
}
\description{
Updates an existing Amazon Secure Agent.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_update_agent_runtime/} for full documentation.
}
\keyword{internal}
