% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_update_api_key_credential_provider}
\alias{bedrockagentcorecontrol_update_api_key_credential_provider}
\title{Updates an existing API key credential provider}
\usage{
bedrockagentcorecontrol_update_api_key_credential_provider(name, apiKey)
}
\arguments{
\item{name}{[required] The name of the API key credential provider to update.}

\item{apiKey}{[required] The new API key to use for authentication. This value replaces the
existing API key and is encrypted and stored securely.}
}
\description{
Updates an existing API key credential provider.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_update_api_key_credential_provider/} for full documentation.
}
\keyword{internal}
