% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_update_configuration_bundle}
\alias{bedrockagentcorecontrol_update_configuration_bundle}
\title{Updates a configuration bundle by creating a new version with the
specified changes}
\usage{
bedrockagentcorecontrol_update_configuration_bundle(
  clientToken = NULL,
  bundleId,
  bundleName = NULL,
  description = NULL,
  components = NULL,
  parentVersionIds,
  branchName = NULL,
  commitMessage = NULL,
  createdBy = NULL,
  kmsKeyArn = NULL
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier to ensure that the API request completes no more than one time. If you don't specify this field, a value is randomly generated for you. If this token matches a previous request, the service ignores the request, but doesn't return an error. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring idempotency}.}

\item{bundleId}{[required] The unique identifier of the configuration bundle to update.}

\item{bundleName}{The updated name for the configuration bundle.}

\item{description}{The updated description for the configuration bundle.}

\item{components}{The updated component configurations. Creates a new version of the bundle.}

\item{parentVersionIds}{[required] A list of parent version identifiers for lineage tracking. Regular commits have a single parent. Merge commits have two parents: the target branch parent and the source branch parent. If the branch already exists, the first parent must be the latest version on that branch.}

\item{branchName}{The branch name for this version. If not specified, inherits the parent's branch or defaults to \code{mainline}.}

\item{commitMessage}{A commit message describing the changes in this version.}

\item{createdBy}{The source that created this version, including the source name and optional ARN.}

\item{kmsKeyArn}{Optional KMS key ARN for encrypting component configurations. If provided, components will be encrypted with this key. If the bundle already has a KMS key, this rotates to the new key.}
}
\description{
Updates a configuration bundle by creating a new version with the specified changes. Each update creates a new version in the version history.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_update_configuration_bundle/} for full documentation.
}
\keyword{internal}
