% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_update_dataset}
\alias{bedrockagentcorecontrol_update_dataset}
\title{Updates a dataset's metadata}
\usage{
bedrockagentcorecontrol_update_dataset(
  datasetId,
  clientToken = NULL,
  description = NULL
)
}
\arguments{
\item{datasetId}{[required] The unique identifier of the dataset to update.}

\item{clientToken}{A unique, case-sensitive identifier to ensure that the API request completes no more than one time. If you don't specify this field, a value is randomly generated for you. If this token matches a previous request, the service ignores the request, but doesn't return an error. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring idempotency}.}

\item{description}{The updated description for the dataset.}
}
\description{
Updates a dataset's metadata. Synchronous operation. Only provided fields are updated; omitted fields remain unchanged.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_update_dataset/} for full documentation.
}
\keyword{internal}
