% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockagentcorecontrol_operations.R
\name{bedrockagentcorecontrol_update_gateway_target}
\alias{bedrockagentcorecontrol_update_gateway_target}
\title{Updates an existing gateway target}
\usage{
bedrockagentcorecontrol_update_gateway_target(
  gatewayIdentifier,
  targetId,
  name = NULL,
  description = NULL,
  targetConfiguration,
  credentialProviderConfigurations = NULL,
  metadataConfiguration = NULL,
  privateEndpoint = NULL
)
}
\arguments{
\item{gatewayIdentifier}{[required] The unique identifier of the gateway associated with the target.}

\item{targetId}{[required] The unique identifier of the gateway target to update.}

\item{name}{The updated name for the gateway target.}

\item{description}{The updated description for the gateway target.}

\item{targetConfiguration}{[required] The configuration for a gateway target. This structure defines how the gateway connects to and interacts with the target endpoint.}

\item{credentialProviderConfigurations}{The updated credential provider configurations for the gateway target.}

\item{metadataConfiguration}{Configuration for HTTP header and query parameter propagation to the gateway target.}

\item{privateEndpoint}{The private endpoint configuration for the gateway target. Use this to connect the gateway to private resources in your VPC.}
}
\description{
Updates an existing gateway target.

See \url{https://www.paws-r-sdk.com/docs/bedrockagentcorecontrol_update_gateway_target/} for full documentation.
}
\keyword{internal}
