% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_create_blueprint}
\alias{bedrockdataautomation_create_blueprint}
\title{Creates an Amazon Bedrock Data Automation Blueprint}
\usage{
bedrockdataautomation_create_blueprint(
  blueprintName,
  type,
  blueprintStage = NULL,
  schema,
  clientToken = NULL,
  encryptionConfiguration = NULL
)
}
\arguments{
\item{blueprintName}{[required]}

\item{type}{[required]}

\item{blueprintStage}{}

\item{schema}{[required]}

\item{clientToken}{}

\item{encryptionConfiguration}{}
}
\description{
Creates an Amazon Bedrock Data Automation Blueprint

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_create_blueprint/} for full documentation.
}
\keyword{internal}
