% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_create_data_automation_project}
\alias{bedrockdataautomation_create_data_automation_project}
\title{Creates an Amazon Bedrock Keystone DataAutomationProject}
\usage{
bedrockdataautomation_create_data_automation_project(
  projectName,
  projectDescription = NULL,
  projectStage = NULL,
  standardOutputConfiguration,
  customOutputConfiguration = NULL,
  overrideConfiguration = NULL,
  clientToken = NULL,
  encryptionConfiguration = NULL
)
}
\arguments{
\item{projectName}{[required]}

\item{projectDescription}{}

\item{projectStage}{}

\item{standardOutputConfiguration}{[required]}

\item{customOutputConfiguration}{}

\item{overrideConfiguration}{}

\item{clientToken}{}

\item{encryptionConfiguration}{}
}
\description{
Creates an Amazon Bedrock Keystone DataAutomationProject

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_create_data_automation_project/} for full documentation.
}
\keyword{internal}
