% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomation_operations.R
\name{bedrockdataautomation_list_blueprints}
\alias{bedrockdataautomation_list_blueprints}
\title{Lists all existing Amazon Bedrock Keystone Blueprints}
\usage{
bedrockdataautomation_list_blueprints(
  blueprintArn = NULL,
  resourceOwner = NULL,
  blueprintStageFilter = NULL,
  maxResults = NULL,
  nextToken = NULL,
  projectFilter = NULL
)
}
\arguments{
\item{blueprintArn}{}

\item{resourceOwner}{}

\item{blueprintStageFilter}{}

\item{maxResults}{}

\item{nextToken}{}

\item{projectFilter}{}
}
\description{
Lists all existing Amazon Bedrock Keystone Blueprints

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomation_list_blueprints/} for full documentation.
}
\keyword{internal}
