% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockdataautomationruntime_operations.R
\name{bedrockdataautomationruntime_tag_resource}
\alias{bedrockdataautomationruntime_tag_resource}
\title{Tag an Amazon Bedrock Data Automation resource}
\usage{
bedrockdataautomationruntime_tag_resource(resourceARN, tags)
}
\arguments{
\item{resourceARN}{[required]}

\item{tags}{[required]}
}
\description{
Tag an Amazon Bedrock Data Automation resource

See \url{https://www.paws-r-sdk.com/docs/bedrockdataautomationruntime_tag_resource/} for full documentation.
}
\keyword{internal}
