% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_batch_detect_entities}
\alias{comprehend_batch_detect_entities}
\title{Inspects the text of a batch of documents for named entities and returns
information about them}
\usage{
comprehend_batch_detect_entities(TextList, LanguageCode)
}
\arguments{
\item{TextList}{[required] A list containing the UTF-8 encoded text of the input documents. The
list can contain a maximum of 25 documents. The maximum size of each
document is 5 KB.}

\item{LanguageCode}{[required] The language of the input documents. You can specify any of the primary
languages supported by Amazon Comprehend. All documents must be in the
same language.}
}
\description{
Inspects the text of a batch of documents for named entities and returns information about them. For more information about named entities, see \href{https://docs.aws.amazon.com/comprehend/latest/dg/how-entities.html}{Entities} in the Comprehend Developer Guide.

See \url{https://paws-r.github.io/docs/comprehend/batch_detect_entities.html} for full documentation.
}
\keyword{internal}
