% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_classify_document}
\alias{comprehend_classify_document}
\title{Creates a new document classification request to analyze a single
document in real-time, using a previously created and trained custom
model and an endpoint}
\usage{
comprehend_classify_document(Text, EndpointArn)
}
\arguments{
\item{Text}{[required] The document text to be analyzed.}

\item{EndpointArn}{[required] The Amazon Resource Number (ARN) of the endpoint. For information about
endpoints, see \href{https://docs.aws.amazon.com/comprehend/latest/dg/manage-endpoints.html}{Managing endpoints}.}
}
\description{
Creates a new document classification request to analyze a single document in real-time, using a previously created and trained custom model and an endpoint.

See \url{https://paws-r.github.io/docs/comprehend/classify_document.html} for full documentation.
}
\keyword{internal}
