% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_contains_pii_entities}
\alias{comprehend_contains_pii_entities}
\title{Analyzes input text for the presence of personally identifiable
information (PII) and returns the labels of identified PII entity types
such as name, address, bank account number, or phone number}
\usage{
comprehend_contains_pii_entities(Text, LanguageCode)
}
\arguments{
\item{Text}{[required] Creates a new document classification request to analyze a single
document in real-time, returning personally identifiable information
(PII) entity labels.}

\item{LanguageCode}{[required] The language of the input documents. Currently, English is the only
valid language.}
}
\description{
Analyzes input text for the presence of personally identifiable information (PII) and returns the labels of identified PII entity types such as name, address, bank account number, or phone number.

See \url{https://paws-r.github.io/docs/comprehend/contains_pii_entities.html} for full documentation.
}
\keyword{internal}
