% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_delete_entity_recognizer}
\alias{comprehend_delete_entity_recognizer}
\title{Deletes an entity recognizer}
\usage{
comprehend_delete_entity_recognizer(EntityRecognizerArn)
}
\arguments{
\item{EntityRecognizerArn}{[required] The Amazon Resource Name (ARN) that identifies the entity recognizer.}
}
\description{
Deletes an entity recognizer.

See \url{https://paws-r.github.io/docs/comprehend/delete_entity_recognizer.html} for full documentation.
}
\keyword{internal}
