% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_describe_key_phrases_detection_job}
\alias{comprehend_describe_key_phrases_detection_job}
\title{Gets the properties associated with a key phrases detection job}
\usage{
comprehend_describe_key_phrases_detection_job(JobId)
}
\arguments{
\item{JobId}{[required] The identifier that Amazon Comprehend generated for the job. The
operation returns this identifier in its response.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  KeyPhrasesDetectionJobProperties = list(
    JobId = "string",
    JobName = "string",
    JobStatus = "SUBMITTED"|"IN_PROGRESS"|"COMPLETED"|"FAILED"|"STOP_REQUESTED"|"STOPPED",
    Message = "string",
    SubmitTime = as.POSIXct(
      "2015-01-01"
    ),
    EndTime = as.POSIXct(
      "2015-01-01"
    ),
    InputDataConfig = list(
      S3Uri = "string",
      InputFormat = "ONE_DOC_PER_FILE"|"ONE_DOC_PER_LINE"
    ),
    OutputDataConfig = list(
      S3Uri = "string",
      KmsKeyId = "string"
    ),
    LanguageCode = "en"|"es"|"fr"|"de"|"it"|"pt"|"ar"|"hi"|"ja"|"ko"|"zh"|"zh-TW",
    DataAccessRoleArn = "string",
    VolumeKmsKeyId = "string",
    VpcConfig = list(
      SecurityGroupIds = list(
        "string"
      ),
      Subnets = list(
        "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets the properties associated with a key phrases detection job. Use
this operation to get the status of a detection job.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_key_phrases_detection_job(
  JobId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
