% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_describe_pii_entities_detection_job}
\alias{comprehend_describe_pii_entities_detection_job}
\title{Gets the properties associated with a PII entities detection job}
\usage{
comprehend_describe_pii_entities_detection_job(JobId)
}
\arguments{
\item{JobId}{[required] The identifier that Amazon Comprehend generated for the job. The
operation returns this identifier in its response.}
}
\description{
Gets the properties associated with a PII entities detection job. For
example, you can use this operation to get the job status.
}
\section{Request syntax}{
\preformatted{svc$describe_pii_entities_detection_job(
  JobId = "string"
)
}
}

\keyword{internal}
