% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_detect_dominant_language}
\alias{comprehend_detect_dominant_language}
\title{Determines the dominant language of the input text}
\usage{
comprehend_detect_dominant_language(Text)
}
\arguments{
\item{Text}{[required] A UTF-8 text string. The string must contain at least 20 characters. The
maximum string size is 100 KB.}
}
\description{
Determines the dominant language of the input text. For a list of languages that Amazon Comprehend can detect, see \href{https://docs.aws.amazon.com/comprehend/latest/dg/how-languages.html}{Amazon Comprehend Supported Languages}.

See \url{https://paws-r.github.io/docs/comprehend/detect_dominant_language.html} for full documentation.
}
\keyword{internal}
