% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_detect_key_phrases}
\alias{comprehend_detect_key_phrases}
\title{Detects the key noun phrases found in the text}
\usage{
comprehend_detect_key_phrases(Text, LanguageCode)
}
\arguments{
\item{Text}{[required] A UTF-8 text string. The string must contain less than 100 KB of UTF-8
encoded characters.}

\item{LanguageCode}{[required] The language of the input documents. You can specify any of the primary
languages supported by Amazon Comprehend. All documents must be in the
same language.}
}
\description{
Detects the key noun phrases found in the text.

See \url{https://paws-r.github.io/docs/comprehend/detect_key_phrases.html} for full documentation.
}
\keyword{internal}
