% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_detect_sentiment}
\alias{comprehend_detect_sentiment}
\title{Inspects text and returns an inference of the prevailing sentiment
(POSITIVE, NEUTRAL, MIXED, or NEGATIVE)}
\usage{
comprehend_detect_sentiment(Text, LanguageCode)
}
\arguments{
\item{Text}{[required] A UTF-8 text string. The maximum string size is 5 KB.}

\item{LanguageCode}{[required] The language of the input documents. You can specify any of the primary
languages supported by Amazon Comprehend. All documents must be in the
same language.}
}
\description{
Inspects text and returns an inference of the prevailing sentiment (\code{POSITIVE}, \code{NEUTRAL}, \code{MIXED}, or \code{NEGATIVE}).

See \url{https://paws-r.github.io/docs/comprehend/detect_sentiment.html} for full documentation.
}
\keyword{internal}
