% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_detect_sentiment}
\alias{comprehend_detect_sentiment}
\title{Inspects text and returns an inference of the prevailing sentiment
(POSITIVE, NEUTRAL, MIXED, or NEGATIVE)}
\usage{
comprehend_detect_sentiment(Text, LanguageCode)
}
\arguments{
\item{Text}{[required] A UTF-8 text string. Each string must contain fewer that 5,000 bytes of
UTF-8 encoded characters.}

\item{LanguageCode}{[required] The language of the input documents. You can specify any of the primary
languages supported by Amazon Comprehend. All documents must be in the
same language.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Sentiment = "POSITIVE"|"NEGATIVE"|"NEUTRAL"|"MIXED",
  SentimentScore = list(
    Positive = 123.0,
    Negative = 123.0,
    Neutral = 123.0,
    Mixed = 123.0
  )
)
}\if{html}{\out{</div>}}
}
\description{
Inspects text and returns an inference of the prevailing sentiment
(\code{POSITIVE}, \code{NEUTRAL}, \code{MIXED}, or \code{NEGATIVE}).
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$detect_sentiment(
  Text = "string",
  LanguageCode = "en"|"es"|"fr"|"de"|"it"|"pt"|"ar"|"hi"|"ja"|"ko"|"zh"|"zh-TW"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
