% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_detect_syntax}
\alias{comprehend_detect_syntax}
\title{Inspects text for syntax and the part of speech of words in the document}
\usage{
comprehend_detect_syntax(Text, LanguageCode)
}
\arguments{
\item{Text}{[required] A UTF-8 string. The maximum string size is 5 KB.}

\item{LanguageCode}{[required] The language code of the input documents. You can specify any of the
following languages supported by Amazon Comprehend: German ("de"),
English ("en"), Spanish ("es"), French ("fr"), Italian ("it"), or
Portuguese ("pt").}
}
\description{
Inspects text for syntax and the part of speech of words in the document. For more information, see \href{https://docs.aws.amazon.com/comprehend/latest/dg/how-syntax.html}{Syntax} in the Comprehend Developer Guide.

See \url{https://paws-r.github.io/docs/comprehend/detect_syntax.html} for full documentation.
}
\keyword{internal}
