% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_list_sentiment_detection_jobs}
\alias{comprehend_list_sentiment_detection_jobs}
\title{Gets a list of sentiment detection jobs that you have submitted}
\usage{
comprehend_list_sentiment_detection_jobs(
  Filter = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Filter}{Filters the jobs that are returned. You can filter jobs on their name,
status, or the date and time that they were submitted. You can only set
one filter at a time.}

\item{NextToken}{Identifies the next page of results to return.}

\item{MaxResults}{The maximum number of results to return in each page. The default is
100.}
}
\description{
Gets a list of sentiment detection jobs that you have submitted.

See \url{https://www.paws-r-sdk.com/docs/comprehend_list_sentiment_detection_jobs/} for full documentation.
}
\keyword{internal}
