% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_stop_pii_entities_detection_job}
\alias{comprehend_stop_pii_entities_detection_job}
\title{Stops a PII entities detection job in progress}
\usage{
comprehend_stop_pii_entities_detection_job(JobId)
}
\arguments{
\item{JobId}{[required] The identifier of the PII entities detection job to stop.}
}
\description{
Stops a PII entities detection job in progress.

See \url{https://www.paws-r-sdk.com/docs/comprehend_stop_pii_entities_detection_job/} for full documentation.
}
\keyword{internal}
