% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_stop_training_document_classifier}
\alias{comprehend_stop_training_document_classifier}
\title{Stops a document classifier training job while in progress}
\usage{
comprehend_stop_training_document_classifier(DocumentClassifierArn)
}
\arguments{
\item{DocumentClassifierArn}{[required] The Amazon Resource Name (ARN) that identifies the document classifier
currently being trained.}
}
\description{
Stops a document classifier training job while in progress.

See \url{https://paws-r.github.io/docs/comprehend/stop_training_document_classifier.html} for full documentation.
}
\keyword{internal}
