% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_stop_training_entity_recognizer}
\alias{comprehend_stop_training_entity_recognizer}
\title{Stops an entity recognizer training job while in progress}
\usage{
comprehend_stop_training_entity_recognizer(EntityRecognizerArn)
}
\arguments{
\item{EntityRecognizerArn}{[required] The Amazon Resource Name (ARN) that identifies the entity recognizer
currently being trained.}
}
\description{
Stops an entity recognizer training job while in progress.

See \url{https://www.paws-r-sdk.com/docs/comprehend_stop_training_entity_recognizer/} for full documentation.
}
\keyword{internal}
