% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_update_endpoint}
\alias{comprehend_update_endpoint}
\title{Updates information about the specified endpoint}
\usage{
comprehend_update_endpoint(EndpointArn, DesiredInferenceUnits)
}
\arguments{
\item{EndpointArn}{[required] The Amazon Resource Number (ARN) of the endpoint being updated.}

\item{DesiredInferenceUnits}{[required] The desired number of inference units to be used by the model using this
endpoint. Each inference unit represents of a throughput of 100
characters per second.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Updates information about the specified endpoint.
}
\section{Request syntax}{
\preformatted{svc$update_endpoint(
  EndpointArn = "string",
  DesiredInferenceUnits = 123
)
}
}

\keyword{internal}
