% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehendmedical_operations.R
\name{comprehendmedical_describe_entities_detection_v2_job}
\alias{comprehendmedical_describe_entities_detection_v2_job}
\title{Gets the properties associated with a medical entities detection job}
\usage{
comprehendmedical_describe_entities_detection_v2_job(JobId)
}
\arguments{
\item{JobId}{[required] The identifier that Comprehend Medical; generated for the job. The
\code{\link[=comprehendmedical_start_entities_detection_v2_job]{start_entities_detection_v2_job}}
operation returns this identifier in its response.}
}
\description{
Gets the properties associated with a medical entities detection job. Use this operation to get the status of a detection job.

See \url{https://www.paws-r-sdk.com/docs/comprehendmedical_describe_entities_detection_v2_job/} for full documentation.
}
\keyword{internal}
