% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehendmedical_operations.R
\name{comprehendmedical_describe_phi_detection_job}
\alias{comprehendmedical_describe_phi_detection_job}
\title{Gets the properties associated with a protected health information (PHI)
detection job}
\usage{
comprehendmedical_describe_phi_detection_job(JobId)
}
\arguments{
\item{JobId}{[required] The identifier that Amazon Comprehend Medical generated for the job. The
\code{\link[=comprehendmedical_start_phi_detection_job]{start_phi_detection_job}}
operation returns this identifier in its response.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ComprehendMedicalAsyncJobProperties = list(
    JobId = "string",
    JobName = "string",
    JobStatus = "SUBMITTED"|"IN_PROGRESS"|"COMPLETED"|"PARTIAL_SUCCESS"|"FAILED"|"STOP_REQUESTED"|"STOPPED",
    Message = "string",
    SubmitTime = as.POSIXct(
      "2015-01-01"
    ),
    EndTime = as.POSIXct(
      "2015-01-01"
    ),
    ExpirationTime = as.POSIXct(
      "2015-01-01"
    ),
    InputDataConfig = list(
      S3Bucket = "string",
      S3Key = "string"
    ),
    OutputDataConfig = list(
      S3Bucket = "string",
      S3Key = "string"
    ),
    LanguageCode = "en",
    DataAccessRoleArn = "string",
    ManifestFilePath = "string",
    KMSKey = "string",
    ModelVersion = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets the properties associated with a protected health information (PHI)
detection job. Use this operation to get the status of a detection job.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_phi_detection_job(
  JobId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
