% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehendmedical_operations.R
\name{comprehendmedical_detect_entities_v2}
\alias{comprehendmedical_detect_entities_v2}
\title{Inspects the clinical text for a variety of medical entities and returns
specific information about them such as entity category, location, and
confidence score on that information}
\usage{
comprehendmedical_detect_entities_v2(Text)
}
\arguments{
\item{Text}{[required] A UTF-8 string containing the clinical content being examined for
entities. Each string must contain fewer than 20,000 bytes of
characters.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Entities = list(
    list(
      Id = 123,
      BeginOffset = 123,
      EndOffset = 123,
      Score = 123.0,
      Text = "string",
      Category = "MEDICATION"|"MEDICAL_CONDITION"|"PROTECTED_HEALTH_INFORMATION"|"TEST_TREATMENT_PROCEDURE"|"ANATOMY"|"TIME_EXPRESSION",
      Type = "NAME"|"DOSAGE"|"ROUTE_OR_MODE"|"FORM"|"FREQUENCY"|"DURATION"|"GENERIC_NAME"|"BRAND_NAME"|"STRENGTH"|"RATE"|"ACUITY"|"TEST_NAME"|"TEST_VALUE"|"TEST_UNITS"|"PROCEDURE_NAME"|"TREATMENT_NAME"|"DATE"|"AGE"|"CONTACT_POINT"|"EMAIL"|"IDENTIFIER"|"URL"|"ADDRESS"|"PROFESSION"|"SYSTEM_ORGAN_SITE"|"DIRECTION"|"QUALITY"|"QUANTITY"|"TIME_EXPRESSION"|"TIME_TO_MEDICATION_NAME"|"TIME_TO_DX_NAME"|"TIME_TO_TEST_NAME"|"TIME_TO_PROCEDURE_NAME"|"TIME_TO_TREATMENT_NAME",
      Traits = list(
        list(
          Name = "SIGN"|"SYMPTOM"|"DIAGNOSIS"|"NEGATION",
          Score = 123.0
        )
      ),
      Attributes = list(
        list(
          Type = "NAME"|"DOSAGE"|"ROUTE_OR_MODE"|"FORM"|"FREQUENCY"|"DURATION"|"GENERIC_NAME"|"BRAND_NAME"|"STRENGTH"|"RATE"|"ACUITY"|"TEST_NAME"|"TEST_VALUE"|"TEST_UNITS"|"PROCEDURE_NAME"|"TREATMENT_NAME"|"DATE"|"AGE"|"CONTACT_POINT"|"EMAIL"|"IDENTIFIER"|"URL"|"ADDRESS"|"PROFESSION"|"SYSTEM_ORGAN_SITE"|"DIRECTION"|"QUALITY"|"QUANTITY"|"TIME_EXPRESSION"|"TIME_TO_MEDICATION_NAME"|"TIME_TO_DX_NAME"|"TIME_TO_TEST_NAME"|"TIME_TO_PROCEDURE_NAME"|"TIME_TO_TREATMENT_NAME",
          Score = 123.0,
          RelationshipScore = 123.0,
          RelationshipType = "EVERY"|"WITH_DOSAGE"|"ADMINISTERED_VIA"|"FOR"|"NEGATIVE"|"OVERLAP"|"DOSAGE"|"ROUTE_OR_MODE"|"FORM"|"FREQUENCY"|"DURATION"|"STRENGTH"|"RATE"|"ACUITY"|"TEST_VALUE"|"TEST_UNITS"|"DIRECTION"|"SYSTEM_ORGAN_SITE",
          Id = 123,
          BeginOffset = 123,
          EndOffset = 123,
          Text = "string",
          Category = "MEDICATION"|"MEDICAL_CONDITION"|"PROTECTED_HEALTH_INFORMATION"|"TEST_TREATMENT_PROCEDURE"|"ANATOMY"|"TIME_EXPRESSION",
          Traits = list(
            list(
              Name = "SIGN"|"SYMPTOM"|"DIAGNOSIS"|"NEGATION",
              Score = 123.0
            )
          )
        )
      )
    )
  ),
  UnmappedAttributes = list(
    list(
      Type = "MEDICATION"|"MEDICAL_CONDITION"|"PROTECTED_HEALTH_INFORMATION"|"TEST_TREATMENT_PROCEDURE"|"ANATOMY"|"TIME_EXPRESSION",
      Attribute = list(
        Type = "NAME"|"DOSAGE"|"ROUTE_OR_MODE"|"FORM"|"FREQUENCY"|"DURATION"|"GENERIC_NAME"|"BRAND_NAME"|"STRENGTH"|"RATE"|"ACUITY"|"TEST_NAME"|"TEST_VALUE"|"TEST_UNITS"|"PROCEDURE_NAME"|"TREATMENT_NAME"|"DATE"|"AGE"|"CONTACT_POINT"|"EMAIL"|"IDENTIFIER"|"URL"|"ADDRESS"|"PROFESSION"|"SYSTEM_ORGAN_SITE"|"DIRECTION"|"QUALITY"|"QUANTITY"|"TIME_EXPRESSION"|"TIME_TO_MEDICATION_NAME"|"TIME_TO_DX_NAME"|"TIME_TO_TEST_NAME"|"TIME_TO_PROCEDURE_NAME"|"TIME_TO_TREATMENT_NAME",
        Score = 123.0,
        RelationshipScore = 123.0,
        RelationshipType = "EVERY"|"WITH_DOSAGE"|"ADMINISTERED_VIA"|"FOR"|"NEGATIVE"|"OVERLAP"|"DOSAGE"|"ROUTE_OR_MODE"|"FORM"|"FREQUENCY"|"DURATION"|"STRENGTH"|"RATE"|"ACUITY"|"TEST_VALUE"|"TEST_UNITS"|"DIRECTION"|"SYSTEM_ORGAN_SITE",
        Id = 123,
        BeginOffset = 123,
        EndOffset = 123,
        Text = "string",
        Category = "MEDICATION"|"MEDICAL_CONDITION"|"PROTECTED_HEALTH_INFORMATION"|"TEST_TREATMENT_PROCEDURE"|"ANATOMY"|"TIME_EXPRESSION",
        Traits = list(
          list(
            Name = "SIGN"|"SYMPTOM"|"DIAGNOSIS"|"NEGATION",
            Score = 123.0
          )
        )
      )
    )
  ),
  PaginationToken = "string",
  ModelVersion = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Inspects the clinical text for a variety of medical entities and returns
specific information about them such as entity category, location, and
confidence score on that information. Amazon Comprehend Medical only
detects medical entities in English language texts.

The \code{\link[=comprehendmedical_detect_entities_v2]{detect_entities_v2}}
operation replaces the
\code{\link[=comprehendmedical_detect_entities]{detect_entities}} operation. This
new action uses a different model for determining the entities in your
medical text and changes the way that some entities are returned in the
output. You should use the
\code{\link[=comprehendmedical_detect_entities_v2]{detect_entities_v2}} operation
in all new applications.

The \code{\link[=comprehendmedical_detect_entities_v2]{detect_entities_v2}}
operation returns the \code{Acuity} and \code{Direction} entities as attributes
instead of types.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$detect_entities_v2(
  Text = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
