% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehendmedical_operations.R
\name{comprehendmedical_detect_phi}
\alias{comprehendmedical_detect_phi}
\title{Inspects the clinical text for protected health information (PHI)
entities and returns the entity category, location, and confidence score
for each entity}
\usage{
comprehendmedical_detect_phi(Text)
}
\arguments{
\item{Text}{[required] A UTF-8 text string containing the clinical content being examined for
PHI entities.}
}
\description{
Inspects the clinical text for protected health information (PHI) entities and returns the entity category, location, and confidence score for each entity. Amazon Comprehend Medical only detects entities in English language texts.

See \url{https://www.paws-r-sdk.com/docs/comprehendmedical_detect_phi/} for full documentation.
}
\keyword{internal}
