% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehendmedical_operations.R
\name{comprehendmedical_infer_rx_norm}
\alias{comprehendmedical_infer_rx_norm}
\title{InferRxNorm detects medications as entities listed in a patient record
and links to the normalized concept identifiers in the RxNorm database
from the National Library of Medicine}
\usage{
comprehendmedical_infer_rx_norm(Text)
}
\arguments{
\item{Text}{[required] The input text used for analysis.}
}
\description{
InferRxNorm detects medications as entities listed in a patient record and links to the normalized concept identifiers in the RxNorm database from the National Library of Medicine. Amazon Comprehend Medical only detects medical entities in English language texts.

See \url{https://www.paws-r-sdk.com/docs/comprehendmedical_infer_rx_norm/} for full documentation.
}
\keyword{internal}
