% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehendmedical_operations.R
\name{comprehendmedical_infer_snomedct}
\alias{comprehendmedical_infer_snomedct}
\title{InferSNOMEDCT detects possible medical concepts as entities and links
them to codes from the Systematized Nomenclature of Medicine, Clinical
Terms (SNOMED-CT) ontology}
\usage{
comprehendmedical_infer_snomedct(Text)
}
\arguments{
\item{Text}{[required] The input text to be analyzed using InferSNOMEDCT. The text should be a
string with 1 to 10000 characters.}
}
\description{
InferSNOMEDCT detects possible medical concepts as entities and links them to codes from the Systematized Nomenclature of Medicine, Clinical Terms (SNOMED-CT) ontology

See \url{https://paws-r.github.io/docs/comprehendmedical/infer_snomedct.html} for full documentation.
}
\keyword{internal}
