% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehendmedical_operations.R
\name{comprehendmedical_stop_entities_detection_v2_job}
\alias{comprehendmedical_stop_entities_detection_v2_job}
\title{Stops a medical entities detection job in progress}
\usage{
comprehendmedical_stop_entities_detection_v2_job(JobId)
}
\arguments{
\item{JobId}{[required] The identifier of the medical entities job to stop.}
}
\value{
A list with the following syntax:\preformatted{list(
  JobId = "string"
)
}
}
\description{
Stops a medical entities detection job in progress.
}
\section{Request syntax}{
\preformatted{svc$stop_entities_detection_v2_job(
  JobId = "string"
)
}
}

\keyword{internal}
