% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehendmedical_operations.R
\name{comprehendmedical_stop_phi_detection_job}
\alias{comprehendmedical_stop_phi_detection_job}
\title{Stops a protected health information (PHI) detection job in progress}
\usage{
comprehendmedical_stop_phi_detection_job(JobId)
}
\arguments{
\item{JobId}{[required] The identifier of the PHI detection job to stop.}
}
\description{
Stops a protected health information (PHI) detection job in progress.

See \url{https://www.paws-r-sdk.com/docs/comprehendmedical_stop_phi_detection_job/} for full documentation.
}
\keyword{internal}
