% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehendmedical_operations.R
\name{comprehendmedical_stop_rx_norm_inference_job}
\alias{comprehendmedical_stop_rx_norm_inference_job}
\title{Stops an InferRxNorm inference job in progress}
\usage{
comprehendmedical_stop_rx_norm_inference_job(JobId)
}
\arguments{
\item{JobId}{[required] The identifier of the job.}
}
\description{
Stops an InferRxNorm inference job in progress.

See \url{https://paws-r.github.io/docs/comprehendmedical/stop_rx_norm_inference_job.html} for full documentation.
}
\keyword{internal}
