% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticinference_operations.R
\name{elasticinference_list_tags_for_resource}
\alias{elasticinference_list_tags_for_resource}
\title{Amazon Elastic Inference is no longer available}
\usage{
elasticinference_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the Elastic Inference Accelerator to list the tags for.}
}
\description{
Amazon Elastic Inference is no longer available.

See \url{https://www.paws-r-sdk.com/docs/elasticinference_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
