% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticinference_operations.R
\name{elasticinference_tag_resource}
\alias{elasticinference_tag_resource}
\title{Adds the specified tags to an Elastic Inference Accelerator}
\usage{
elasticinference_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The ARN of the Elastic Inference Accelerator to tag.}

\item{tags}{[required] The tags to add to the Elastic Inference Accelerator.}
}
\description{
Adds the specified tags to an Elastic Inference Accelerator.

See \url{https://paws-r.github.io/docs/elasticinference/tag_resource.html} for full documentation.
}
\keyword{internal}
