% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticinference_operations.R
\name{elasticinference_untag_resource}
\alias{elasticinference_untag_resource}
\title{Amazon Elastic Inference is no longer available}
\usage{
elasticinference_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The ARN of the Elastic Inference Accelerator to untag.}

\item{tagKeys}{[required] The list of tags to remove from the Elastic Inference Accelerator.}
}
\description{
Amazon Elastic Inference is no longer available.

See \url{https://www.paws-r-sdk.com/docs/elasticinference_untag_resource/} for full documentation.
}
\keyword{internal}
