% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_delete_forecast}
\alias{forecastservice_delete_forecast}
\title{Deletes a forecast created using the CreateForecast operation}
\usage{
forecastservice_delete_forecast(ForecastArn)
}
\arguments{
\item{ForecastArn}{[required] The Amazon Resource Name (ARN) of the forecast to delete.}
}
\description{
Deletes a forecast created using the \code{\link[=forecastservice_create_forecast]{create_forecast}} operation. You can delete only forecasts that have a status of \code{ACTIVE} or \code{CREATE_FAILED}. To get the status, use the \code{\link[=forecastservice_describe_forecast]{describe_forecast}} operation.

See \url{https://www.paws-r-sdk.com/docs/forecastservice_delete_forecast/} for full documentation.
}
\keyword{internal}
