% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_delete_what_if_forecast}
\alias{forecastservice_delete_what_if_forecast}
\title{Deletes a what-if forecast created using the CreateWhatIfForecast
operation}
\usage{
forecastservice_delete_what_if_forecast(WhatIfForecastArn)
}
\arguments{
\item{WhatIfForecastArn}{[required] The Amazon Resource Name (ARN) of the what-if forecast that you want to
delete.}
}
\description{
Deletes a what-if forecast created using the \code{\link[=forecastservice_create_what_if_forecast]{create_what_if_forecast}} operation. You can delete only what-if forecasts that have a status of \code{ACTIVE} or \code{CREATE_FAILED}. To get the status, use the \code{\link[=forecastservice_describe_what_if_forecast]{describe_what_if_forecast}} operation.

See \url{https://www.paws-r-sdk.com/docs/forecastservice_delete_what_if_forecast/} for full documentation.
}
\keyword{internal}
