% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_describe_auto_predictor}
\alias{forecastservice_describe_auto_predictor}
\title{Describes a predictor created using the CreateAutoPredictor operation}
\usage{
forecastservice_describe_auto_predictor(PredictorArn)
}
\arguments{
\item{PredictorArn}{[required] The Amazon Resource Name (ARN) of the predictor.}
}
\description{
Describes a predictor created using the CreateAutoPredictor operation.

See \url{https://www.paws-r-sdk.com/docs/forecastservice_describe_auto_predictor/} for full documentation.
}
\keyword{internal}
