% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_describe_dataset_group}
\alias{forecastservice_describe_dataset_group}
\title{Describes a dataset group created using the CreateDatasetGroup operation}
\usage{
forecastservice_describe_dataset_group(DatasetGroupArn)
}
\arguments{
\item{DatasetGroupArn}{[required] The Amazon Resource Name (ARN) of the dataset group.}
}
\description{
Describes a dataset group created using the \code{\link[=forecastservice_create_dataset_group]{create_dataset_group}} operation.

See \url{https://paws-r.github.io/docs/forecastservice/describe_dataset_group.html} for full documentation.
}
\keyword{internal}
