% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastservice_operations.R
\name{forecastservice_describe_explainability_export}
\alias{forecastservice_describe_explainability_export}
\title{Describes an Explainability export created using the
CreateExplainabilityExport operation}
\usage{
forecastservice_describe_explainability_export(ExplainabilityExportArn)
}
\arguments{
\item{ExplainabilityExportArn}{[required] The Amazon Resource Name (ARN) of the Explainability export.}
}
\description{
Describes an Explainability export created using the \code{\link[=forecastservice_create_explainability_export]{create_explainability_export}} operation.

See \url{https://paws-r.github.io/docs/forecastservice/describe_explainability_export.html} for full documentation.
}
\keyword{internal}
