% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_external_model}
\alias{frauddetector_delete_external_model}
\title{Removes a SageMaker model from Amazon Fraud Detector}
\usage{
frauddetector_delete_external_model(modelEndpoint)
}
\arguments{
\item{modelEndpoint}{[required] The endpoint of the Amazon Sagemaker model to delete.}
}
\description{
Removes a SageMaker model from Amazon Fraud Detector.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_delete_external_model/} for full documentation.
}
\keyword{internal}
