% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_list}
\alias{frauddetector_delete_list}
\title{Deletes the list, provided it is not used in a rule}
\usage{
frauddetector_delete_list(name)
}
\arguments{
\item{name}{[required] The name of the list to delete.}
}
\description{
Deletes the list, provided it is not used in a rule.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_delete_list/} for full documentation.
}
\keyword{internal}
