% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_delete_outcome}
\alias{frauddetector_delete_outcome}
\title{Deletes an outcome}
\usage{
frauddetector_delete_outcome(name)
}
\arguments{
\item{name}{[required] The name of the outcome to delete.}
}
\description{
Deletes an outcome.

See \url{https://paws-r.github.io/docs/frauddetector/delete_outcome.html} for full documentation.
}
\keyword{internal}
