% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_update_detector_version_metadata}
\alias{frauddetector_update_detector_version_metadata}
\title{Updates the detector version's description}
\usage{
frauddetector_update_detector_version_metadata(
  detectorId,
  detectorVersionId,
  description
)
}
\arguments{
\item{detectorId}{[required] The detector ID.}

\item{detectorVersionId}{[required] The detector version ID.}

\item{description}{[required] The description.}
}
\description{
Updates the detector version's description. You can update the metadata for any detector version (\verb{DRAFT, ACTIVE,} or \code{INACTIVE}).

See \url{https://paws-r.github.io/docs/frauddetector/update_detector_version_metadata.html} for full documentation.
}
\keyword{internal}
