% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frauddetector_operations.R
\name{frauddetector_update_model}
\alias{frauddetector_update_model}
\title{Updates model description}
\usage{
frauddetector_update_model(modelId, modelType, description = NULL)
}
\arguments{
\item{modelId}{[required] The model ID.}

\item{modelType}{[required] The model type.}

\item{description}{The new model description.}
}
\description{
Updates model description.

See \url{https://www.paws-r-sdk.com/docs/frauddetector_update_model/} for full documentation.
}
\keyword{internal}
