% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_delete_bot_alias}
\alias{lexmodelbuildingservice_delete_bot_alias}
\title{Deletes an alias for the specified bot}
\usage{
lexmodelbuildingservice_delete_bot_alias(name, botName)
}
\arguments{
\item{name}{[required] The name of the alias to delete. The name is case sensitive.}

\item{botName}{[required] The name of the bot that the alias points to.}
}
\description{
Deletes an alias for the specified bot.

See \url{https://www.paws-r-sdk.com/docs/lexmodelbuildingservice_delete_bot_alias/} for full documentation.
}
\keyword{internal}
