% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_delete_bot_channel_association}
\alias{lexmodelbuildingservice_delete_bot_channel_association}
\title{Deletes the association between an Amazon Lex bot and a messaging
platform}
\usage{
lexmodelbuildingservice_delete_bot_channel_association(name, botName,
  botAlias)
}
\arguments{
\item{name}{[required] The name of the association. The name is case sensitive.}

\item{botName}{[required] The name of the Amazon Lex bot.}

\item{botAlias}{[required] An alias that points to the specific version of the Amazon Lex bot to
which this association is being made.}
}
\value{
An empty list.
}
\description{
Deletes the association between an Amazon Lex bot and a messaging
platform.

This operation requires permission for the
\code{lex:DeleteBotChannelAssociation} action.
}
\section{Request syntax}{
\preformatted{svc$delete_bot_channel_association(
  name = "string",
  botName = "string",
  botAlias = "string"
)
}
}

\keyword{internal}
