% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_get_builtin_intents}
\alias{lexmodelbuildingservice_get_builtin_intents}
\title{Gets a list of built-in intents that meet the specified criteria}
\usage{
lexmodelbuildingservice_get_builtin_intents(locale, signatureContains,
  nextToken, maxResults)
}
\arguments{
\item{locale}{A list of locales that the intent supports.}

\item{signatureContains}{Substring to match in built-in intent signatures. An intent will be
returned if any part of its signature matches the substring. For
example, "xyz" matches both "xyzabc" and "abcxyz." To find the signature
for an intent, see \href{https://developer.amazon.com/en-US/docs/alexa/custom-skills/standard-built-in-intents.html}{Standard Built-in Intents}
in the \emph{Alexa Skills Kit}.}

\item{nextToken}{A pagination token that fetches the next page of intents. If this API
call is truncated, Amazon Lex returns a pagination token in the
response. To fetch the next page of intents, use the pagination token in
the next request.}

\item{maxResults}{The maximum number of intents to return in the response. The default is
10.}
}
\value{
A list with the following syntax:\preformatted{list(
  intents = list(
    list(
      signature = "string",
      supportedLocales = list(
        "de-DE"|"en-AU"|"en-GB"|"en-US"|"es-419"|"es-ES"|"es-US"|"fr-FR"|"fr-CA"|"it-IT"
      )
    )
  ),
  nextToken = "string"
)
}
}
\description{
Gets a list of built-in intents that meet the specified criteria.

This operation requires permission for the \code{lex:GetBuiltinIntents}
action.
}
\section{Request syntax}{
\preformatted{svc$get_builtin_intents(
  locale = "de-DE"|"en-AU"|"en-GB"|"en-US"|"es-419"|"es-ES"|"es-US"|"fr-FR"|"fr-CA"|"it-IT",
  signatureContains = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
