% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_get_export}
\alias{lexmodelbuildingservice_get_export}
\title{Exports the contents of a Amazon Lex resource in a specified format}
\usage{
lexmodelbuildingservice_get_export(name, version, resourceType,
  exportType)
}
\arguments{
\item{name}{[required] The name of the bot to export.}

\item{version}{[required] The version of the bot to export.}

\item{resourceType}{[required] The type of resource to export.}

\item{exportType}{[required] The format of the exported data.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  name = "string",
  version = "string",
  resourceType = "BOT"|"INTENT"|"SLOT_TYPE",
  exportType = "ALEXA_SKILLS_KIT"|"LEX",
  exportStatus = "IN_PROGRESS"|"READY"|"FAILED",
  failureReason = "string",
  url = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Exports the contents of a Amazon Lex resource in a specified format.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_export(
  name = "string",
  version = "string",
  resourceType = "BOT"|"INTENT"|"SLOT_TYPE",
  exportType = "ALEXA_SKILLS_KIT"|"LEX"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
