% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_get_import}
\alias{lexmodelbuildingservice_get_import}
\title{Gets information about an import job started with the StartImport
operation}
\usage{
lexmodelbuildingservice_get_import(importId)
}
\arguments{
\item{importId}{[required] The identifier of the import job information to return.}
}
\description{
Gets information about an import job started with the \code{\link[=lexmodelbuildingservice_start_import]{start_import}} operation.

See \url{https://paws-r.github.io/docs/lexmodelbuildingservice/get_import.html} for full documentation.
}
\keyword{internal}
