% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_get_intent}
\alias{lexmodelbuildingservice_get_intent}
\title{Returns information about an intent}
\usage{
lexmodelbuildingservice_get_intent(name, version)
}
\arguments{
\item{name}{[required] The name of the intent. The name is case sensitive.}

\item{version}{[required] The version of the intent.}
}
\description{
Returns information about an intent. In addition to the intent name, you
must specify the intent version.
}
\details{
This operation requires permissions to perform the \code{lex:GetIntent}
action.
}
\section{Request syntax}{
\preformatted{svc$get_intent(
  name = "string",
  version = "string"
)
}
}

\examples{
\dontrun{
# This example shows how to get information about an intent.
svc$get_intent(
  version = "$LATEST",
  name = "DocOrderPizza"
)
}

}
\keyword{internal}
