% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_get_intent_versions}
\alias{lexmodelbuildingservice_get_intent_versions}
\title{Gets information about all of the versions of an intent}
\usage{
lexmodelbuildingservice_get_intent_versions(
  name,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{name}{[required] The name of the intent for which versions should be returned.}

\item{nextToken}{A pagination token for fetching the next page of intent versions. If the
response to this call is truncated, Amazon Lex returns a pagination
token in the response. To fetch the next page of versions, specify the
pagination token in the next request.}

\item{maxResults}{The maximum number of intent versions to return in the response. The
default is 10.}
}
\description{
Gets information about all of the versions of an intent.

See \url{https://paws-r.github.io/docs/lexmodelbuildingservice/get_intent_versions.html} for full documentation.
}
\keyword{internal}
