% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_get_slot_type}
\alias{lexmodelbuildingservice_get_slot_type}
\title{Returns information about a specific version of a slot type}
\usage{
lexmodelbuildingservice_get_slot_type(name, version)
}
\arguments{
\item{name}{[required] The name of the slot type. The name is case sensitive.}

\item{version}{[required] The version of the slot type.}
}
\description{
Returns information about a specific version of a slot type. In addition to specifying the slot type name, you must specify the slot type version.

See \url{https://paws-r.github.io/docs/lexmodelbuildingservice/get_slot_type.html} for full documentation.
}
\keyword{internal}
