% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_list_tags_for_resource}
\alias{lexmodelbuildingservice_list_tags_for_resource}
\title{Gets a list of tags associated with the specified resource}
\usage{
lexmodelbuildingservice_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource to get a list of tags
for.}
}
\description{
Gets a list of tags associated with the specified resource. Only bots, bot aliases, and bot channels can have tags associated with them.

See \url{https://www.paws-r-sdk.com/docs/lexmodelbuildingservice_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
