% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_delete_bot_alias}
\alias{lexmodelsv2_delete_bot_alias}
\title{Deletes the specified bot alias}
\usage{
lexmodelsv2_delete_bot_alias(botAliasId, botId, skipResourceInUseCheck = NULL)
}
\arguments{
\item{botAliasId}{[required] The unique identifier of the bot alias to delete.}

\item{botId}{[required] The unique identifier of the bot associated with the alias to delete.}

\item{skipResourceInUseCheck}{When this parameter is true, Amazon Lex doesn't check to see if any
other resource is using the alias before it is deleted.}
}
\description{
Deletes the specified bot alias.

See \url{https://paws-r.github.io/docs/lexmodelsv2/delete_bot_alias.html} for full documentation.
}
\keyword{internal}
