% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_describe_bot_alias}
\alias{lexmodelsv2_describe_bot_alias}
\title{Get information about a specific bot alias}
\usage{
lexmodelsv2_describe_bot_alias(botAliasId, botId)
}
\arguments{
\item{botAliasId}{[required] The identifier of the bot alias to describe.}

\item{botId}{[required] The identifier of the bot associated with the bot alias to describe.}
}
\description{
Get information about a specific bot alias.

See \url{https://paws-r.github.io/docs/lexmodelsv2/describe_bot_alias.html} for full documentation.
}
\keyword{internal}
