% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelsv2_operations.R
\name{lexmodelsv2_list_tags_for_resource}
\alias{lexmodelsv2_list_tags_for_resource}
\title{Gets a list of tags associated with a resource}
\usage{
lexmodelsv2_list_tags_for_resource(resourceARN)
}
\arguments{
\item{resourceARN}{[required] The Amazon Resource Name (ARN) of the resource to get a list of tags
for.}
}
\description{
Gets a list of tags associated with a resource. Only bots, bot aliases, and bot channels can have tags associated with them.

See \url{https://www.paws-r-sdk.com/docs/lexmodelsv2_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
