% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexruntimeservice_operations.R
\name{lexruntimeservice_delete_session}
\alias{lexruntimeservice_delete_session}
\title{Removes session information for a specified bot, alias, and user ID}
\usage{
lexruntimeservice_delete_session(botName, botAlias, userId)
}
\arguments{
\item{botName}{[required] The name of the bot that contains the session data.}

\item{botAlias}{[required] The alias in use for the bot that contains the session data.}

\item{userId}{[required] The identifier of the user associated with the session data.}
}
\value{
A list with the following syntax:\preformatted{list(
  botName = "string",
  botAlias = "string",
  userId = "string",
  sessionId = "string"
)
}
}
\description{
Removes session information for a specified bot, alias, and user ID.
}
\section{Request syntax}{
\preformatted{svc$delete_session(
  botName = "string",
  botAlias = "string",
  userId = "string"
)
}
}

\keyword{internal}
