% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_delete_model}
\alias{lookoutequipment_delete_model}
\title{Deletes an ML model currently available for Amazon Lookout for Equipment}
\usage{
lookoutequipment_delete_model(ModelName)
}
\arguments{
\item{ModelName}{[required] The name of the ML model to be deleted.}
}
\description{
Deletes an ML model currently available for Amazon Lookout for Equipment. This will prevent it from being used with an inference scheduler, even one that is already set up.

See \url{https://paws-r.github.io/docs/lookoutequipment/delete_model.html} for full documentation.
}
\keyword{internal}
