% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutequipment_operations.R
\name{lookoutequipment_describe_model}
\alias{lookoutequipment_describe_model}
\title{Provides a JSON containing the overall information about a specific ML
model, including model name and ARN, dataset, training and evaluation
information, status, and so on}
\usage{
lookoutequipment_describe_model(ModelName)
}
\arguments{
\item{ModelName}{[required] The name of the ML model to be described.}
}
\description{
Provides a JSON containing the overall information about a specific ML model, including model name and ARN, dataset, training and evaluation information, status, and so on.

See \url{https://paws-r.github.io/docs/lookoutequipment/describe_model.html} for full documentation.
}
\keyword{internal}
