% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_activate_anomaly_detector}
\alias{lookoutmetrics_activate_anomaly_detector}
\title{Activates an anomaly detector}
\usage{
lookoutmetrics_activate_anomaly_detector(AnomalyDetectorArn)
}
\arguments{
\item{AnomalyDetectorArn}{[required] The ARN of the anomaly detector.}
}
\description{
Activates an anomaly detector.

See \url{https://paws-r.github.io/docs/lookoutmetrics/activate_anomaly_detector.html} for full documentation.
}
\keyword{internal}
