% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_back_test_anomaly_detector}
\alias{lookoutmetrics_back_test_anomaly_detector}
\title{Runs a backtest for anomaly detection for the specified resource}
\usage{
lookoutmetrics_back_test_anomaly_detector(AnomalyDetectorArn)
}
\arguments{
\item{AnomalyDetectorArn}{[required] The Amazon Resource Name (ARN) of the anomaly detector.}
}
\description{
Runs a backtest for anomaly detection for the specified resource.

See \url{https://paws-r.github.io/docs/lookoutmetrics/back_test_anomaly_detector.html} for full documentation.
}
\keyword{internal}
