% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_delete_alert}
\alias{lookoutmetrics_delete_alert}
\title{Deletes an alert}
\usage{
lookoutmetrics_delete_alert(AlertArn)
}
\arguments{
\item{AlertArn}{[required] The ARN of the alert to delete.}
}
\description{
Deletes an alert.

See \url{https://paws-r.github.io/docs/lookoutmetrics/delete_alert.html} for full documentation.
}
\keyword{internal}
