% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_describe_alert}
\alias{lookoutmetrics_describe_alert}
\title{Describes an alert}
\usage{
lookoutmetrics_describe_alert(AlertArn)
}
\arguments{
\item{AlertArn}{[required] The ARN of the alert to describe.}
}
\description{
Describes an alert.

See \url{https://paws-r.github.io/docs/lookoutmetrics/describe_alert.html} for full documentation.
}
\keyword{internal}
