% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_describe_anomaly_detector}
\alias{lookoutmetrics_describe_anomaly_detector}
\title{Describes a detector}
\usage{
lookoutmetrics_describe_anomaly_detector(AnomalyDetectorArn)
}
\arguments{
\item{AnomalyDetectorArn}{[required] The ARN of the detector to describe.}
}
\description{
Describes a detector.

See \url{https://www.paws-r-sdk.com/docs/lookoutmetrics_describe_anomaly_detector/} for full documentation.
}
\keyword{internal}
