% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_describe_metric_set}
\alias{lookoutmetrics_describe_metric_set}
\title{Describes a dataset}
\usage{
lookoutmetrics_describe_metric_set(MetricSetArn)
}
\arguments{
\item{MetricSetArn}{[required] The ARN of the dataset.}
}
\description{
Describes a dataset.

See \url{https://paws-r.github.io/docs/lookoutmetrics/describe_metric_set.html} for full documentation.
}
\keyword{internal}
