% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_get_data_quality_metrics}
\alias{lookoutmetrics_get_data_quality_metrics}
\title{Returns details about the requested data quality metrics}
\usage{
lookoutmetrics_get_data_quality_metrics(
  AnomalyDetectorArn,
  MetricSetArn = NULL
)
}
\arguments{
\item{AnomalyDetectorArn}{[required] The Amazon Resource Name (ARN) of the anomaly detector that you want to
investigate.}

\item{MetricSetArn}{The Amazon Resource Name (ARN) of a specific data quality metric set.}
}
\description{
Returns details about the requested data quality metrics.

See \url{https://www.paws-r-sdk.com/docs/lookoutmetrics_get_data_quality_metrics/} for full documentation.
}
\keyword{internal}
