% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_get_sample_data}
\alias{lookoutmetrics_get_sample_data}
\title{Returns a selection of sample records from an Amazon S3 datasource}
\usage{
lookoutmetrics_get_sample_data(S3SourceConfig = NULL)
}
\arguments{
\item{S3SourceConfig}{A datasource bucket in Amazon S3.}
}
\description{
Returns a selection of sample records from an Amazon S3 datasource.

See \url{https://paws-r.github.io/docs/lookoutmetrics/get_sample_data.html} for full documentation.
}
\keyword{internal}
