% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookoutmetrics_operations.R
\name{lookoutmetrics_update_metric_set}
\alias{lookoutmetrics_update_metric_set}
\title{Updates a dataset}
\usage{
lookoutmetrics_update_metric_set(
  MetricSetArn,
  MetricSetDescription = NULL,
  MetricList = NULL,
  Offset = NULL,
  TimestampColumn = NULL,
  DimensionList = NULL,
  MetricSetFrequency = NULL,
  MetricSource = NULL,
  DimensionFilterList = NULL
)
}
\arguments{
\item{MetricSetArn}{[required] The ARN of the dataset to update.}

\item{MetricSetDescription}{The dataset's description.}

\item{MetricList}{The metric list.}

\item{Offset}{After an interval ends, the amount of seconds that the detector waits
before importing data. Offset is only supported for S3, Redshift, Athena
and datasources.}

\item{TimestampColumn}{The timestamp column.}

\item{DimensionList}{The dimension list.}

\item{MetricSetFrequency}{The dataset's interval.}

\item{MetricSource}{}

\item{DimensionFilterList}{Describes a list of filters for choosing specific dimensions and
specific values. Each filter consists of the dimension and one of its
values that you want to include. When multiple dimensions or values are
specified, the dimensions are joined with an AND operation and the
values are joined with an OR operation.}
}
\description{
Updates a dataset.

See \url{https://www.paws-r-sdk.com/docs/lookoutmetrics_update_metric_set/} for full documentation.
}
\keyword{internal}
