% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning_operations.R
\name{machinelearning_delete_batch_prediction}
\alias{machinelearning_delete_batch_prediction}
\title{Assigns the DELETED status to a BatchPrediction, rendering it unusable}
\usage{
machinelearning_delete_batch_prediction(BatchPredictionId)
}
\arguments{
\item{BatchPredictionId}{[required] A user-supplied ID that uniquely identifies the \code{BatchPrediction}.}
}
\value{
A list with the following syntax:\preformatted{list(
  BatchPredictionId = "string"
)
}
}
\description{
Assigns the DELETED status to a \code{BatchPrediction}, rendering it
unusable.

After using the
\code{\link[=machinelearning_delete_batch_prediction]{delete_batch_prediction}}
operation, you can use the
\code{\link[=machinelearning_get_batch_prediction]{get_batch_prediction}} operation
to verify that the status of the \code{BatchPrediction} changed to DELETED.

\strong{Caution:} The result of the
\code{\link[=machinelearning_delete_batch_prediction]{delete_batch_prediction}}
operation is irreversible.
}
\section{Request syntax}{
\preformatted{svc$delete_batch_prediction(
  BatchPredictionId = "string"
)
}
}

\keyword{internal}
