% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning_operations.R
\name{machinelearning_delete_evaluation}
\alias{machinelearning_delete_evaluation}
\title{Assigns the DELETED status to an Evaluation, rendering it unusable}
\usage{
machinelearning_delete_evaluation(EvaluationId)
}
\arguments{
\item{EvaluationId}{[required] A user-supplied ID that uniquely identifies the \code{Evaluation} to delete.}
}
\description{
Assigns the \code{DELETED} status to an \code{Evaluation}, rendering it unusable.
}
\details{
After invoking the \code{DeleteEvaluation} operation, you can use the
\code{GetEvaluation} operation to verify that the status of the \code{Evaluation}
changed to \code{DELETED}.

Caution

The results of the \code{DeleteEvaluation} operation are irreversible.
}
\section{Request syntax}{
\preformatted{svc$delete_evaluation(
  EvaluationId = "string"
)
}
}

\keyword{internal}
