% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning_operations.R
\name{machinelearning_delete_ml_model}
\alias{machinelearning_delete_ml_model}
\title{Assigns the DELETED status to an MLModel, rendering it unusable}
\usage{
machinelearning_delete_ml_model(MLModelId)
}
\arguments{
\item{MLModelId}{[required] A user-supplied ID that uniquely identifies the \code{MLModel}.}
}
\value{
A list with the following syntax:\preformatted{list(
  MLModelId = "string"
)
}
}
\description{
Assigns the \code{DELETED} status to an \code{MLModel}, rendering it unusable.

After using the \code{\link[=machinelearning_delete_ml_model]{delete_ml_model}}
operation, you can use the
\code{\link[=machinelearning_get_ml_model]{get_ml_model}} operation to verify that
the status of the \code{MLModel} changed to DELETED.

\strong{Caution:} The result of the
\code{\link[=machinelearning_delete_ml_model]{delete_ml_model}} operation is
irreversible.
}
\section{Request syntax}{
\preformatted{svc$delete_ml_model(
  MLModelId = "string"
)
}
}

\keyword{internal}
