% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning_operations.R
\name{machinelearning_delete_realtime_endpoint}
\alias{machinelearning_delete_realtime_endpoint}
\title{Deletes a real time endpoint of an MLModel}
\usage{
machinelearning_delete_realtime_endpoint(MLModelId)
}
\arguments{
\item{MLModelId}{[required] The ID assigned to the \code{MLModel} during creation.}
}
\description{
Deletes a real time endpoint of an \code{MLModel}.

See \url{https://www.paws-r-sdk.com/docs/machinelearning_delete_realtime_endpoint/} for full documentation.
}
\keyword{internal}
