% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning_operations.R
\name{machinelearning_get_evaluation}
\alias{machinelearning_get_evaluation}
\title{Returns an Evaluation that includes metadata as well as the current
status of the Evaluation}
\usage{
machinelearning_get_evaluation(EvaluationId)
}
\arguments{
\item{EvaluationId}{[required] The ID of the \code{Evaluation} to retrieve. The evaluation of each \code{MLModel}
is recorded and cataloged. The ID provides the means to access the
information.}
}
\description{
Returns an \code{Evaluation} that includes metadata as well as the current status of the \code{Evaluation}.

See \url{https://paws-r.github.io/docs/machinelearning/get_evaluation.html} for full documentation.
}
\keyword{internal}
