% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning_operations.R
\name{machinelearning_update_batch_prediction}
\alias{machinelearning_update_batch_prediction}
\title{Updates the BatchPredictionName of a BatchPrediction}
\usage{
machinelearning_update_batch_prediction(BatchPredictionId, BatchPredictionName)
}
\arguments{
\item{BatchPredictionId}{[required] The ID assigned to the \code{BatchPrediction} during creation.}

\item{BatchPredictionName}{[required] A new user-supplied name or description of the \code{BatchPrediction}.}
}
\description{
Updates the \code{BatchPredictionName} of a \code{BatchPrediction}.

See \url{https://paws-r.github.io/docs/machinelearning/update_batch_prediction.html} for full documentation.
}
\keyword{internal}
