% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_create_job_for_devices}
\alias{panorama_create_job_for_devices}
\title{Creates a job to run on one or more devices}
\usage{
panorama_create_job_for_devices(DeviceIds, DeviceJobConfig, JobType)
}
\arguments{
\item{DeviceIds}{[required] IDs of target devices.}

\item{DeviceJobConfig}{[required] Configuration settings for the job.}

\item{JobType}{[required] The type of job to run.}
}
\description{
Creates a job to run on one or more devices.

See \url{https://paws-r.github.io/docs/panorama/create_job_for_devices.html} for full documentation.
}
\keyword{internal}
