% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_list_devices}
\alias{panorama_list_devices}
\title{Returns a list of devices}
\usage{
panorama_list_devices(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of devices to return in one page of results.}

\item{NextToken}{Specify the pagination token from a previous request to retrieve the
next page of results.}
}
\description{
Returns a list of devices.

See \url{https://paws-r.github.io/docs/panorama/list_devices.html} for full documentation.
}
\keyword{internal}
