% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panorama_operations.R
\name{panorama_list_packages}
\alias{panorama_list_packages}
\title{Returns a list of packages}
\usage{
panorama_list_packages(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of packages to return in one page of results.}

\item{NextToken}{Specify the pagination token from a previous request to retrieve the
next page of results.}
}
\description{
Returns a list of packages.

See \url{https://www.paws-r-sdk.com/docs/panorama_list_packages/} for full documentation.
}
\keyword{internal}
