% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_create_dataset_export_job}
\alias{personalize_create_dataset_export_job}
\title{Creates a job that exports data from your dataset to an Amazon S3 bucket}
\usage{
personalize_create_dataset_export_job(jobName, datasetArn,
  ingestionMode, roleArn, jobOutput)
}
\arguments{
\item{jobName}{[required] The name for the dataset export job.}

\item{datasetArn}{[required] The Amazon Resource Name (ARN) of the dataset that contains the data to
export.}

\item{ingestionMode}{The data to export, based on how you imported the data. You can choose
to export only \code{BULK} data that you imported using a dataset import job,
only \code{PUT} data that you imported incrementally (using the console,
PutEvents, PutUsers and PutItems operations), or \code{ALL} for both types.
The default value is \code{PUT}.}

\item{roleArn}{[required] The Amazon Resource Name (ARN) of the AWS Identity and Access Management
service role that has permissions to add data to your output Amazon S3
bucket.}

\item{jobOutput}{[required] The path to the Amazon S3 bucket where the job's output is stored.}
}
\value{
A list with the following syntax:\preformatted{list(
  datasetExportJobArn = "string"
)
}
}
\description{
Creates a job that exports data from your dataset to an Amazon S3
bucket. To allow Amazon Personalize to export the training data, you
must specify an service-linked AWS Identity and Access Management (IAM)
role that gives Amazon Personalize \code{PutObject} permissions for your
Amazon S3 bucket. For information, see \href{https://docs.aws.amazon.com/personalize/latest/dg/export-data.html}{Exporting a dataset}
in the Amazon Personalize developer guide.

\strong{Status}

A dataset export job can be in one of the following states:
\itemize{
\item CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE
FAILED
}

To get the status of the export job, call
\code{\link[=personalize_describe_dataset_export_job]{describe_dataset_export_job}},
and specify the Amazon Resource Name (ARN) of the dataset export job.
The dataset export is complete when the status shows as ACTIVE. If the
status shows as CREATE FAILED, the response includes a \code{failureReason}
key, which describes why the job failed.
}
\section{Request syntax}{
\preformatted{svc$create_dataset_export_job(
  jobName = "string",
  datasetArn = "string",
  ingestionMode = "BULK"|"PUT"|"ALL",
  roleArn = "string",
  jobOutput = list(
    s3DataDestination = list(
      path = "string",
      kmsKeyArn = "string"
    )
  )
)
}
}

\keyword{internal}
