% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_create_event_tracker}
\alias{personalize_create_event_tracker}
\title{Creates an event tracker that you use when adding event data to a
specified dataset group using the PutEvents API}
\usage{
personalize_create_event_tracker(name, datasetGroupArn)
}
\arguments{
\item{name}{[required] The name for the event tracker.}

\item{datasetGroupArn}{[required] The Amazon Resource Name (ARN) of the dataset group that receives the
event data.}
}
\value{
A list with the following syntax:\preformatted{list(
  eventTrackerArn = "string",
  trackingId = "string"
)
}
}
\description{
Creates an event tracker that you use when adding event data to a
specified dataset group using the
\href{https://docs.aws.amazon.com/personalize/latest/dg/API_UBS_PutEvents.html}{PutEvents}
API.

Only one event tracker can be associated with a dataset group. You will
get an error if you call
\code{\link[=personalize_create_event_tracker]{create_event_tracker}} using the
same dataset group as an existing event tracker.

When you create an event tracker, the response includes a tracking ID,
which you pass as a parameter when you use the
\href{https://docs.aws.amazon.com/personalize/latest/dg/API_UBS_PutEvents.html}{PutEvents}
operation. Amazon Personalize then appends the event data to the
Interactions dataset of the dataset group you specify in your event
tracker.

The event tracker can be in one of the following states:
\itemize{
\item CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE
FAILED
\item DELETE PENDING > DELETE IN_PROGRESS
}

To get the status of the event tracker, call
\code{\link[=personalize_describe_event_tracker]{describe_event_tracker}}.

The event tracker must be in the ACTIVE state before using the tracking
ID.

\strong{Related APIs}
\itemize{
\item \code{\link[=personalize_list_event_trackers]{list_event_trackers}}
\item \code{\link[=personalize_describe_event_tracker]{describe_event_tracker}}
\item \code{\link[=personalize_delete_event_tracker]{delete_event_tracker}}
}
}
\section{Request syntax}{
\preformatted{svc$create_event_tracker(
  name = "string",
  datasetGroupArn = "string"
)
}
}

\keyword{internal}
