% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_create_filter}
\alias{personalize_create_filter}
\title{Creates a recommendation filter}
\usage{
personalize_create_filter(name, datasetGroupArn, filterExpression, tags = NULL)
}
\arguments{
\item{name}{[required] The name of the filter to create.}

\item{datasetGroupArn}{[required] The ARN of the dataset group that the filter will belong to.}

\item{filterExpression}{[required] The filter expression defines which items are included or excluded from
recommendations. Filter expression must follow specific format rules.
For information about filter expression structure and syntax, see
\href{https://docs.aws.amazon.com/personalize/latest/dg/filter-expressions.html}{Filter expressions}.}

\item{tags}{A list of
\href{https://docs.aws.amazon.com/personalize/latest/dg/tagging-resources.html}{tags}
to apply to the filter.}
}
\description{
Creates a recommendation filter. For more information, see \href{https://docs.aws.amazon.com/personalize/latest/dg/filter.html}{Filtering recommendations and user segments}.

See \url{https://www.paws-r-sdk.com/docs/personalize_create_filter/} for full documentation.
}
\keyword{internal}
