% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_delete_dataset}
\alias{personalize_delete_dataset}
\title{Deletes a dataset}
\usage{
personalize_delete_dataset(datasetArn)
}
\arguments{
\item{datasetArn}{[required] The Amazon Resource Name (ARN) of the dataset to delete.}
}
\description{
Deletes a dataset. You can't delete a dataset if an associated \code{DatasetImportJob} or \code{SolutionVersion} is in the CREATE PENDING or IN PROGRESS state. For more information on datasets, see \code{\link[=personalize_create_dataset]{create_dataset}}.

See \url{https://paws-r.github.io/docs/personalize/delete_dataset.html} for full documentation.
}
\keyword{internal}
