% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_delete_dataset_group}
\alias{personalize_delete_dataset_group}
\title{Deletes a dataset group}
\usage{
personalize_delete_dataset_group(datasetGroupArn)
}
\arguments{
\item{datasetGroupArn}{[required] The ARN of the dataset group to delete.}
}
\value{
An empty list.
}
\description{
Deletes a dataset group. Before you delete a dataset group, you must
delete the following:
\itemize{
\item All associated event trackers.
\item All associated solutions.
\item All datasets in the dataset group.
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_dataset_group(
  datasetGroupArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
