% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_delete_event_tracker}
\alias{personalize_delete_event_tracker}
\title{Deletes the event tracker}
\usage{
personalize_delete_event_tracker(eventTrackerArn)
}
\arguments{
\item{eventTrackerArn}{[required] The Amazon Resource Name (ARN) of the event tracker to delete.}
}
\description{
Deletes the event tracker. Does not delete the dataset from the dataset group. For more information on event trackers, see \code{\link[=personalize_create_event_tracker]{create_event_tracker}}.

See \url{https://www.paws-r-sdk.com/docs/personalize_delete_event_tracker/} for full documentation.
}
\keyword{internal}
