% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_describe_batch_inference_job}
\alias{personalize_describe_batch_inference_job}
\title{Gets the properties of a batch inference job including name, Amazon
Resource Name (ARN), status, input and output configurations, and the
ARN of the solution version used to generate the recommendations}
\usage{
personalize_describe_batch_inference_job(batchInferenceJobArn)
}
\arguments{
\item{batchInferenceJobArn}{[required] The ARN of the batch inference job to describe.}
}
\description{
Gets the properties of a batch inference job including name, Amazon Resource Name (ARN), status, input and output configurations, and the ARN of the solution version used to generate the recommendations.

See \url{https://www.paws-r-sdk.com/docs/personalize_describe_batch_inference_job/} for full documentation.
}
\keyword{internal}
