% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_describe_campaign}
\alias{personalize_describe_campaign}
\title{Describes the given campaign, including its status}
\usage{
personalize_describe_campaign(campaignArn)
}
\arguments{
\item{campaignArn}{[required] The Amazon Resource Name (ARN) of the campaign.}
}
\description{
Describes the given campaign, including its status.

See \url{https://www.paws-r-sdk.com/docs/personalize_describe_campaign/} for full documentation.
}
\keyword{internal}
