% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_describe_dataset_export_job}
\alias{personalize_describe_dataset_export_job}
\title{Describes the dataset export job created by CreateDatasetExportJob,
including the export job status}
\usage{
personalize_describe_dataset_export_job(datasetExportJobArn)
}
\arguments{
\item{datasetExportJobArn}{[required] The Amazon Resource Name (ARN) of the dataset export job to describe.}
}
\description{
Describes the dataset export job created by \code{\link[=personalize_create_dataset_export_job]{create_dataset_export_job}}, including the export job status.

See \url{https://www.paws-r-sdk.com/docs/personalize_describe_dataset_export_job/} for full documentation.
}
\keyword{internal}
