% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_describe_dataset_export_job}
\alias{personalize_describe_dataset_export_job}
\title{Describes the dataset export job created by CreateDatasetExportJob,
including the export job status}
\usage{
personalize_describe_dataset_export_job(datasetExportJobArn)
}
\arguments{
\item{datasetExportJobArn}{[required] The Amazon Resource Name (ARN) of the dataset export job to describe.}
}
\value{
A list with the following syntax:\preformatted{list(
  datasetExportJob = list(
    jobName = "string",
    datasetExportJobArn = "string",
    datasetArn = "string",
    ingestionMode = "BULK"|"PUT"|"ALL",
    roleArn = "string",
    status = "string",
    jobOutput = list(
      s3DataDestination = list(
        path = "string",
        kmsKeyArn = "string"
      )
    ),
    creationDateTime = as.POSIXct(
      "2015-01-01"
    ),
    lastUpdatedDateTime = as.POSIXct(
      "2015-01-01"
    ),
    failureReason = "string"
  )
)
}
}
\description{
Describes the dataset export job created by
\code{\link[=personalize_create_dataset_export_job]{create_dataset_export_job}},
including the export job status.
}
\section{Request syntax}{
\preformatted{svc$describe_dataset_export_job(
  datasetExportJobArn = "string"
)
}
}

\keyword{internal}
