% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_describe_dataset_import_job}
\alias{personalize_describe_dataset_import_job}
\title{Describes the dataset import job created by CreateDatasetImportJob,
including the import job status}
\usage{
personalize_describe_dataset_import_job(datasetImportJobArn)
}
\arguments{
\item{datasetImportJobArn}{[required] The Amazon Resource Name (ARN) of the dataset import job to describe.}
}
\description{
Describes the dataset import job created by \code{\link[=personalize_create_dataset_import_job]{create_dataset_import_job}}, including the import job status.

See \url{https://www.paws-r-sdk.com/docs/personalize_describe_dataset_import_job/} for full documentation.
}
\keyword{internal}
