% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_describe_event_tracker}
\alias{personalize_describe_event_tracker}
\title{Describes an event tracker}
\usage{
personalize_describe_event_tracker(eventTrackerArn)
}
\arguments{
\item{eventTrackerArn}{[required] The Amazon Resource Name (ARN) of the event tracker to describe.}
}
\description{
Describes an event tracker. The response includes the \code{trackingId} and
\code{status} of the event tracker. For more information on event trackers,
see CreateEventTracker.
}
\section{Request syntax}{
\preformatted{svc$describe_event_tracker(
  eventTrackerArn = "string"
)
}
}

\keyword{internal}
