% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_describe_feature_transformation}
\alias{personalize_describe_feature_transformation}
\title{Describes the given feature transformation}
\usage{
personalize_describe_feature_transformation(featureTransformationArn)
}
\arguments{
\item{featureTransformationArn}{[required] The Amazon Resource Name (ARN) of the feature transformation to
describe.}
}
\description{
Describes the given feature transformation.

See \url{https://www.paws-r-sdk.com/docs/personalize_describe_feature_transformation/} for full documentation.
}
\keyword{internal}
