% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_describe_filter}
\alias{personalize_describe_filter}
\title{Describes a filter's properties}
\usage{
personalize_describe_filter(filterArn)
}
\arguments{
\item{filterArn}{[required] The ARN of the filter to describe.}
}
\description{
Describes a filter's properties.
}
\section{Request syntax}{
\preformatted{svc$describe_filter(
  filterArn = "string"
)
}
}

\keyword{internal}
