% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_describe_recipe}
\alias{personalize_describe_recipe}
\title{Describes a recipe}
\usage{
personalize_describe_recipe(recipeArn)
}
\arguments{
\item{recipeArn}{[required] The Amazon Resource Name (ARN) of the recipe to describe.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  recipe = list(
    name = "string",
    recipeArn = "string",
    algorithmArn = "string",
    featureTransformationArn = "string",
    status = "string",
    description = "string",
    creationDateTime = as.POSIXct(
      "2015-01-01"
    ),
    recipeType = "string",
    lastUpdatedDateTime = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes a recipe.

A recipe contains three items:
\itemize{
\item An algorithm that trains a model.
\item Hyperparameters that govern the training.
\item Feature transformation information for modifying the input data
before training.
}

Amazon Personalize provides a set of predefined recipes. You specify a
recipe when you create a solution with the
\code{\link[=personalize_create_solution]{create_solution}} API.
\code{\link[=personalize_create_solution]{create_solution}} trains a model by using
the algorithm in the specified recipe and a training dataset. The
solution, when deployed as a campaign, can provide recommendations using
the
\href{https://docs.aws.amazon.com/personalize/latest/dg/API_RS_GetRecommendations.html}{GetRecommendations}
API.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_recipe(
  recipeArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
