% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_describe_recipe}
\alias{personalize_describe_recipe}
\title{Describes a recipe}
\usage{
personalize_describe_recipe(recipeArn)
}
\arguments{
\item{recipeArn}{[required] The Amazon Resource Name (ARN) of the recipe to describe.}
}
\description{
Describes a recipe.

See \url{https://www.paws-r-sdk.com/docs/personalize_describe_recipe/} for full documentation.
}
\keyword{internal}
