% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_describe_schema}
\alias{personalize_describe_schema}
\title{Describes a schema}
\usage{
personalize_describe_schema(schemaArn)
}
\arguments{
\item{schemaArn}{[required] The Amazon Resource Name (ARN) of the schema to retrieve.}
}
\description{
Describes a schema. For more information on schemas, see \code{\link[=personalize_create_schema]{create_schema}}.

See \url{https://paws-r.github.io/docs/personalize/describe_schema.html} for full documentation.
}
\keyword{internal}
