% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_get_solution_metrics}
\alias{personalize_get_solution_metrics}
\title{Gets the metrics for the specified solution version}
\usage{
personalize_get_solution_metrics(solutionVersionArn)
}
\arguments{
\item{solutionVersionArn}{[required] The Amazon Resource Name (ARN) of the solution version for which to get
metrics.}
}
\description{
Gets the metrics for the specified solution version.

See \url{https://paws-r.github.io/docs/personalize/get_solution_metrics.html} for full documentation.
}
\keyword{internal}
