% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_list_batch_inference_jobs}
\alias{personalize_list_batch_inference_jobs}
\title{Gets a list of the batch inference jobs that have been performed off of
a solution version}
\usage{
personalize_list_batch_inference_jobs(solutionVersionArn, nextToken,
  maxResults)
}
\arguments{
\item{solutionVersionArn}{The Amazon Resource Name (ARN) of the solution version from which the
batch inference jobs were created.}

\item{nextToken}{The token to request the next page of results.}

\item{maxResults}{The maximum number of batch inference job results to return in each
page. The default value is 100.}
}
\value{
A list with the following syntax:\preformatted{list(
  batchInferenceJobs = list(
    list(
      batchInferenceJobArn = "string",
      jobName = "string",
      status = "string",
      creationDateTime = as.POSIXct(
        "2015-01-01"
      ),
      lastUpdatedDateTime = as.POSIXct(
        "2015-01-01"
      ),
      failureReason = "string",
      solutionVersionArn = "string"
    )
  ),
  nextToken = "string"
)
}
}
\description{
Gets a list of the batch inference jobs that have been performed off of
a solution version.
}
\section{Request syntax}{
\preformatted{svc$list_batch_inference_jobs(
  solutionVersionArn = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
