% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_list_dataset_groups}
\alias{personalize_list_dataset_groups}
\title{Returns a list of dataset groups}
\usage{
personalize_list_dataset_groups(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{A token returned from the previous call to
\code{\link[=personalize_list_dataset_groups]{list_dataset_groups}} for getting the
next set of dataset groups (if they exist).}

\item{maxResults}{The maximum number of dataset groups to return.}
}
\description{
Returns a list of dataset groups. The response provides the properties for each dataset group, including the Amazon Resource Name (ARN). For more information on dataset groups, see \code{\link[=personalize_create_dataset_group]{create_dataset_group}}.

See \url{https://paws-r.github.io/docs/personalize/list_dataset_groups.html} for full documentation.
}
\keyword{internal}
