% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_list_dataset_import_jobs}
\alias{personalize_list_dataset_import_jobs}
\title{Returns a list of dataset import jobs that use the given dataset}
\usage{
personalize_list_dataset_import_jobs(
  datasetArn = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{datasetArn}{The Amazon Resource Name (ARN) of the dataset to list the dataset import
jobs for.}

\item{nextToken}{A token returned from the previous call to
\code{\link[=personalize_list_dataset_import_jobs]{list_dataset_import_jobs}} for
getting the next set of dataset import jobs (if they exist).}

\item{maxResults}{The maximum number of dataset import jobs to return.}
}
\description{
Returns a list of dataset import jobs that use the given dataset. When a dataset is not specified, all the dataset import jobs associated with the account are listed. The response provides the properties for each dataset import job, including the Amazon Resource Name (ARN). For more information on dataset import jobs, see \code{\link[=personalize_create_dataset_import_job]{create_dataset_import_job}}. For more information on datasets, see \code{\link[=personalize_create_dataset]{create_dataset}}.

See \url{https://www.paws-r-sdk.com/docs/personalize_list_dataset_import_jobs/} for full documentation.
}
\keyword{internal}
