% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_list_event_trackers}
\alias{personalize_list_event_trackers}
\title{Returns the list of event trackers associated with the account}
\usage{
personalize_list_event_trackers(datasetGroupArn, nextToken, maxResults)
}
\arguments{
\item{datasetGroupArn}{The ARN of a dataset group used to filter the response.}

\item{nextToken}{A token returned from the previous call to \code{ListEventTrackers} for
getting the next set of event trackers (if they exist).}

\item{maxResults}{The maximum number of event trackers to return.}
}
\description{
Returns the list of event trackers associated with the account. The
response provides the properties for each event tracker, including the
Amazon Resource Name (ARN) and tracking ID. For more information on
event trackers, see CreateEventTracker.
}
\section{Request syntax}{
\preformatted{svc$list_event_trackers(
  datasetGroupArn = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
