% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_list_tags_for_resource}
\alias{personalize_list_tags_for_resource}
\title{Get a list of tags attached to a resource}
\usage{
personalize_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The resource's Amazon Resource Name (ARN).}
}
\description{
Get a list of \href{https://docs.aws.amazon.com/personalize/latest/dg/tagging-resources.html}{tags} attached to a resource.

See \url{https://www.paws-r-sdk.com/docs/personalize_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
