% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_stop_recommender}
\alias{personalize_stop_recommender}
\title{Stops a recommender that is ACTIVE}
\usage{
personalize_stop_recommender(recommenderArn)
}
\arguments{
\item{recommenderArn}{[required] The Amazon Resource Name (ARN) of the recommender to stop.}
}
\description{
Stops a recommender that is ACTIVE. Stopping a recommender halts billing and automatic retraining for the recommender.

See \url{https://www.paws-r-sdk.com/docs/personalize_stop_recommender/} for full documentation.
}
\keyword{internal}
