% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_stop_solution_version_creation}
\alias{personalize_stop_solution_version_creation}
\title{Stops creating a solution version that is in a state of CREATE_PENDING
or CREATE IN_PROGRESS}
\usage{
personalize_stop_solution_version_creation(solutionVersionArn)
}
\arguments{
\item{solutionVersionArn}{[required] The Amazon Resource Name (ARN) of the solution version you want to stop
creating.}
}
\description{
Stops creating a solution version that is in a state of CREATE_PENDING or CREATE IN_PROGRESS.

See \url{https://www.paws-r-sdk.com/docs/personalize_stop_solution_version_creation/} for full documentation.
}
\keyword{internal}
