% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_update_campaign}
\alias{personalize_update_campaign}
\title{Updates a campaign by either deploying a new solution or changing the
value of the campaign's minProvisionedTPS parameter}
\usage{
personalize_update_campaign(
  campaignArn,
  solutionVersionArn = NULL,
  minProvisionedTPS = NULL,
  campaignConfig = NULL
)
}
\arguments{
\item{campaignArn}{[required] The Amazon Resource Name (ARN) of the campaign.}

\item{solutionVersionArn}{The ARN of a new solution version to deploy.}

\item{minProvisionedTPS}{Specifies the requested minimum provisioned transactions
(recommendations) per second that Amazon Personalize will support.}

\item{campaignConfig}{The configuration details of a campaign.}
}
\description{
Updates a campaign by either deploying a new solution or changing the value of the campaign's \code{minProvisionedTPS} parameter.

See \url{https://paws-r.github.io/docs/personalize/update_campaign.html} for full documentation.
}
\keyword{internal}
