% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalize_operations.R
\name{personalize_update_recommender}
\alias{personalize_update_recommender}
\title{Updates the recommender to modify the recommender configuration}
\usage{
personalize_update_recommender(recommenderArn, recommenderConfig)
}
\arguments{
\item{recommenderArn}{[required] The Amazon Resource Name (ARN) of the recommender to modify.}

\item{recommenderConfig}{[required] The configuration details of the recommender.}
}
\description{
Updates the recommender to modify the recommender configuration. If you update the recommender to modify the columns used in training, Amazon Personalize automatically starts a full retraining of the models backing your recommender. While the update completes, you can still get recommendations from the recommender. The recommender uses the previous configuration until the update completes. To track the status of this update, use the \code{latestRecommenderUpdate} returned in the \code{\link[=personalize_describe_recommender]{describe_recommender}} operation.

See \url{https://www.paws-r-sdk.com/docs/personalize_update_recommender/} for full documentation.
}
\keyword{internal}
