% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalizeevents_operations.R
\name{personalizeevents_put_items}
\alias{personalizeevents_put_items}
\title{Adds one or more items to an Items dataset}
\usage{
personalizeevents_put_items(datasetArn, items)
}
\arguments{
\item{datasetArn}{[required] The Amazon Resource Number (ARN) of the Items dataset you are adding the
item or items to.}

\item{items}{[required] A list of item data.}
}
\value{
An empty list.
}
\description{
Adds one or more items to an Items dataset. For more information see
importing-items.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$put_items(
  datasetArn = "string",
  items = list(
    list(
      itemId = "string",
      properties = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
