% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polly_operations.R
\name{polly_get_lexicon}
\alias{polly_get_lexicon}
\title{Returns the content of the specified pronunciation lexicon stored in an
Amazon Web Services Region}
\usage{
polly_get_lexicon(Name)
}
\arguments{
\item{Name}{[required] Name of the lexicon.}
}
\description{
Returns the content of the specified pronunciation lexicon stored in an Amazon Web Services Region. For more information, see \href{https://docs.aws.amazon.com/polly/latest/dg/managing-lexicons.html}{Managing Lexicons}.

See \url{https://paws-r.github.io/docs/polly/get_lexicon.html} for full documentation.
}
\keyword{internal}
