% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_create_dataset}
\alias{rekognition_create_dataset}
\title{Creates a new Amazon Rekognition Custom Labels dataset}
\usage{
rekognition_create_dataset(DatasetSource = NULL, DatasetType, ProjectArn)
}
\arguments{
\item{DatasetSource}{The source files for the dataset. You can specify the ARN of an existing
dataset or specify the Amazon S3 bucket location of an Amazon Sagemaker
format manifest file. If you don't specify \code{datasetSource}, an empty
dataset is created. To add labeled images to the dataset, You can use
the console or call
\code{\link[=rekognition_update_dataset_entries]{update_dataset_entries}}.}

\item{DatasetType}{[required] The type of the dataset. Specify \code{TRAIN} to create a training dataset.
Specify \code{TEST} to create a test dataset.}

\item{ProjectArn}{[required] The ARN of the Amazon Rekognition Custom Labels project to which you
want to asssign the dataset.}
}
\description{
Creates a new Amazon Rekognition Custom Labels dataset. You can create a dataset by using an Amazon Sagemaker format manifest file or by copying an existing Amazon Rekognition Custom Labels dataset.

See \url{https://www.paws-r-sdk.com/docs/rekognition_create_dataset/} for full documentation.
}
\keyword{internal}
