% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_create_project}
\alias{rekognition_create_project}
\title{Creates a new Amazon Rekognition Custom Labels project}
\usage{
rekognition_create_project(ProjectName)
}
\arguments{
\item{ProjectName}{[required] The name of the project to create.}
}
\description{
Creates a new Amazon Rekognition Custom Labels project. A project is a group of resources (datasets, model versions) that you use to create and manage Amazon Rekognition Custom Labels models.

See \url{https://paws-r.github.io/docs/rekognition/create_project.html} for full documentation.
}
\keyword{internal}
