% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_delete_collection}
\alias{rekognition_delete_collection}
\title{Deletes the specified collection}
\usage{
rekognition_delete_collection(CollectionId)
}
\arguments{
\item{CollectionId}{[required] ID of the collection to delete.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  StatusCode = 123
)
}\if{html}{\out{</div>}}
}
\description{
Deletes the specified collection. Note that this operation removes all
faces in the collection. For an example, see
delete-collection-procedure.

This operation requires permissions to perform the
\code{rekognition:DeleteCollection} action.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_collection(
  CollectionId = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# This operation deletes a Rekognition collection.
svc$delete_collection(
  CollectionId = "myphotos"
)
}

}
\keyword{internal}
