% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_delete_dataset}
\alias{rekognition_delete_dataset}
\title{This operation applies only to Amazon Rekognition Custom Labels}
\usage{
rekognition_delete_dataset(DatasetArn)
}
\arguments{
\item{DatasetArn}{[required] The ARN of the Amazon Rekognition Custom Labels dataset that you want to
delete.}
}
\description{
This operation applies only to Amazon Rekognition Custom Labels.

See \url{https://www.paws-r-sdk.com/docs/rekognition_delete_dataset/} for full documentation.
}
\keyword{internal}
