% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_delete_faces}
\alias{rekognition_delete_faces}
\title{Deletes faces from a collection}
\usage{
rekognition_delete_faces(CollectionId, FaceIds)
}
\arguments{
\item{CollectionId}{[required] Collection from which to remove the specific faces.}

\item{FaceIds}{[required] An array of face IDs to delete.}
}
\description{
Deletes faces from a collection. You specify a collection ID and an array of face IDs to remove from the collection.

See \url{https://www.paws-r-sdk.com/docs/rekognition_delete_faces/} for full documentation.
}
\keyword{internal}
