% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_delete_project_version}
\alias{rekognition_delete_project_version}
\title{Deletes an Amazon Rekognition Custom Labels model}
\usage{
rekognition_delete_project_version(ProjectVersionArn)
}
\arguments{
\item{ProjectVersionArn}{[required] The Amazon Resource Name (ARN) of the model version that you want to
delete.}
}
\description{
Deletes an Amazon Rekognition Custom Labels model.

See \url{https://paws-r.github.io/docs/rekognition/delete_project_version.html} for full documentation.
}
\keyword{internal}
