% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_delete_project_version}
\alias{rekognition_delete_project_version}
\title{Deletes a Rekognition project model or project version, like a Amazon
Rekognition Custom Labels model or a custom adapter}
\usage{
rekognition_delete_project_version(ProjectVersionArn)
}
\arguments{
\item{ProjectVersionArn}{[required] The Amazon Resource Name (ARN) of the project version that you want to
delete.}
}
\description{
Deletes a Rekognition project model or project version, like a Amazon Rekognition Custom Labels model or a custom adapter.

See \url{https://www.paws-r-sdk.com/docs/rekognition_delete_project_version/} for full documentation.
}
\keyword{internal}
