% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_describe_collection}
\alias{rekognition_describe_collection}
\title{Describes the specified collection}
\usage{
rekognition_describe_collection(CollectionId)
}
\arguments{
\item{CollectionId}{[required] The ID of the collection to describe.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  FaceCount = 123,
  FaceModelVersion = "string",
  CollectionARN = "string",
  CreationTimestamp = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes the specified collection. You can use
\code{\link[=rekognition_describe_collection]{describe_collection}} to get
information, such as the number of faces indexed into a collection and
the version of the model used by the collection for face detection.

For more information, see Describing a Collection in the Amazon
Rekognition Developer Guide.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_collection(
  CollectionId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
