% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_describe_collection}
\alias{rekognition_describe_collection}
\title{Describes the specified collection}
\usage{
rekognition_describe_collection(CollectionId)
}
\arguments{
\item{CollectionId}{[required] The ID of the collection to describe.}
}
\description{
Describes the specified collection. You can use \code{DescribeCollection} to
get information, such as the number of faces indexed into a collection
and the version of the model used by the collection for face detection.

For more information, see Describing a Collection in the Amazon
Rekognition Developer Guide.
}
\section{Request syntax}{
\preformatted{svc$describe_collection(
  CollectionId = "string"
)
}
}

\keyword{internal}
