% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_describe_project_versions}
\alias{rekognition_describe_project_versions}
\title{Lists and describes the models in an Amazon Rekognition Custom Labels
project}
\usage{
rekognition_describe_project_versions(ProjectArn, VersionNames,
  NextToken, MaxResults)
}
\arguments{
\item{ProjectArn}{[required] The Amazon Resource Name (ARN) of the project that contains the models
you want to describe.}

\item{VersionNames}{A list of model version names that you want to describe. You can add up
to 10 model version names to the list. If you don\'t specify a value,
all model descriptions are returned.}

\item{NextToken}{If the previous response was incomplete (because there is more results
to retrieve), Amazon Rekognition Custom Labels returns a pagination
token in the response. You can use this pagination token to retrieve the
next set of results.}

\item{MaxResults}{The maximum number of results to return per paginated call. The largest
value you can specify is 100. If you specify a value greater than 100, a
ValidationException error occurs. The default value is 100.}
}
\description{
Lists and describes the models in an Amazon Rekognition Custom Labels
project. You can specify up to 10 model versions in
\code{ProjectVersionArns}. If you don\'t specify a value, descriptions for
all models are returned.
}
\details{
This operation requires permissions to perform the
\code{rekognition:DescribeProjectVersions} action.
}
\section{Request syntax}{
\preformatted{svc$describe_project_versions(
  ProjectArn = "string",
  VersionNames = list(
    "string"
  ),
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
