% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_describe_projects}
\alias{rekognition_describe_projects}
\title{Lists and gets information about your Amazon Rekognition Custom Labels
projects}
\usage{
rekognition_describe_projects(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{If the previous response was incomplete (because there is more results
to retrieve), Amazon Rekognition Custom Labels returns a pagination
token in the response. You can use this pagination token to retrieve the
next set of results.}

\item{MaxResults}{The maximum number of results to return per paginated call. The largest
value you can specify is 100. If you specify a value greater than 100, a
ValidationException error occurs. The default value is 100.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ProjectDescriptions = list(
    list(
      ProjectArn = "string",
      CreationTimestamp = as.POSIXct(
        "2015-01-01"
      ),
      Status = "CREATING"|"CREATED"|"DELETING"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists and gets information about your Amazon Rekognition Custom Labels
projects.

This operation requires permissions to perform the
\code{rekognition:DescribeProjects} action.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_projects(
  NextToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
