% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_describe_stream_processor}
\alias{rekognition_describe_stream_processor}
\title{Provides information about a stream processor created by
CreateStreamProcessor}
\usage{
rekognition_describe_stream_processor(Name)
}
\arguments{
\item{Name}{[required] Name of the stream processor for which you want information.}
}
\description{
Provides information about a stream processor created by \code{\link[=rekognition_create_stream_processor]{create_stream_processor}}. You can get information about the input and output streams, the input parameters for the face recognition being performed, and the current status of the stream processor.

See \url{https://paws-r.github.io/docs/rekognition/describe_stream_processor.html} for full documentation.
}
\keyword{internal}
