% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_detect_custom_labels}
\alias{rekognition_detect_custom_labels}
\title{This operation applies only to Amazon Rekognition Custom Labels}
\usage{
rekognition_detect_custom_labels(
  ProjectVersionArn,
  Image,
  MaxResults = NULL,
  MinConfidence = NULL
)
}
\arguments{
\item{ProjectVersionArn}{[required] The ARN of the model version that you want to use. Only models
associated with Custom Labels projects accepted by the operation. If a
provided ARN refers to a model version associated with a project for a
different feature type, then an InvalidParameterException is returned.}

\item{Image}{[required]}

\item{MaxResults}{Maximum number of results you want the service to return in the
response. The service returns the specified number of highest confidence
labels ranked from highest confidence to lowest.}

\item{MinConfidence}{Specifies the minimum confidence level for the labels to return.
\code{\link[=rekognition_detect_custom_labels]{detect_custom_labels}} doesn't
return any labels with a confidence value that's lower than this
specified value. If you specify a value of 0,
\code{\link[=rekognition_detect_custom_labels]{detect_custom_labels}} returns all
labels, regardless of the assumed threshold applied to each label. If
you don't specify a value for \code{MinConfidence},
\code{\link[=rekognition_detect_custom_labels]{detect_custom_labels}} returns
labels based on the assumed threshold of each label.}
}
\description{
This operation applies only to Amazon Rekognition Custom Labels.

See \url{https://www.paws-r-sdk.com/docs/rekognition_detect_custom_labels/} for full documentation.
}
\keyword{internal}
