% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_distribute_dataset_entries}
\alias{rekognition_distribute_dataset_entries}
\title{This operation applies only to Amazon Rekognition Custom Labels}
\usage{
rekognition_distribute_dataset_entries(Datasets)
}
\arguments{
\item{Datasets}{[required] The ARNS for the training dataset and test dataset that you want to use.
The datasets must belong to the same project. The test dataset must be
empty.}
}
\description{
This operation applies only to Amazon Rekognition Custom Labels.

See \url{https://www.paws-r-sdk.com/docs/rekognition_distribute_dataset_entries/} for full documentation.
}
\keyword{internal}
