% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_get_celebrity_info}
\alias{rekognition_get_celebrity_info}
\title{Gets the name and additional information about a celebrity based on his
or her Amazon Rekognition ID}
\usage{
rekognition_get_celebrity_info(Id)
}
\arguments{
\item{Id}{[required] The ID for the celebrity. You get the celebrity ID from a call to the
\code{\link[=rekognition_recognize_celebrities]{recognize_celebrities}} operation,
which recognizes celebrities in an image.}
}
\value{
A list with the following syntax:\preformatted{list(
  Urls = list(
    "string"
  ),
  Name = "string"
)
}
}
\description{
Gets the name and additional information about a celebrity based on his
or her Amazon Rekognition ID. The additional information is returned as
an array of URLs. If there is no additional information about the
celebrity, this list is empty.

For more information, see Recognizing Celebrities in an Image in the
Amazon Rekognition Developer Guide.

This operation requires permissions to perform the
\code{rekognition:GetCelebrityInfo} action.
}
\section{Request syntax}{
\preformatted{svc$get_celebrity_info(
  Id = "string"
)
}
}

\keyword{internal}
