% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_get_media_analysis_job}
\alias{rekognition_get_media_analysis_job}
\title{Retrieves the results for a given media analysis job}
\usage{
rekognition_get_media_analysis_job(JobId)
}
\arguments{
\item{JobId}{[required] Unique identifier for the media analysis job for which you want to
retrieve results.}
}
\description{
Retrieves the results for a given media analysis job. Takes a \code{JobId} returned by StartMediaAnalysisJob.

See \url{https://www.paws-r-sdk.com/docs/rekognition_get_media_analysis_job/} for full documentation.
}
\keyword{internal}
