% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_start_stream_processor}
\alias{rekognition_start_stream_processor}
\title{Starts processing a stream processor}
\usage{
rekognition_start_stream_processor(Name)
}
\arguments{
\item{Name}{[required] The name of the stream processor to start processing.}
}
\value{
An empty list.
}
\description{
Starts processing a stream processor. You create a stream processor by
calling
\code{\link[=rekognition_create_stream_processor]{create_stream_processor}}. To
tell \code{\link[=rekognition_start_stream_processor]{start_stream_processor}}
which stream processor to start, use the value of the \code{Name} field
specified in the call to
\code{\link[=rekognition_create_stream_processor]{create_stream_processor}}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$start_stream_processor(
  Name = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
