% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_stop_stream_processor}
\alias{rekognition_stop_stream_processor}
\title{Stops a running stream processor that was created by
CreateStreamProcessor}
\usage{
rekognition_stop_stream_processor(Name)
}
\arguments{
\item{Name}{[required] The name of a stream processor created by
\code{\link[=rekognition_create_stream_processor]{create_stream_processor}}.}
}
\description{
Stops a running stream processor that was created by \code{\link[=rekognition_create_stream_processor]{create_stream_processor}}.

See \url{https://paws-r.github.io/docs/rekognition/stop_stream_processor.html} for full documentation.
}
\keyword{internal}
