% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_tag_resource}
\alias{rekognition_tag_resource}
\title{Adds one or more key-value tags to an Amazon Rekognition collection,
stream processor, or Custom Labels model}
\usage{
rekognition_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] Amazon Resource Name (ARN) of the model, collection, or stream processor
that you want to assign the tags to.}

\item{Tags}{[required] The key-value tags to assign to the resource.}
}
\description{
Adds one or more key-value tags to an Amazon Rekognition collection, stream processor, or Custom Labels model. For more information, see \href{https://docs.aws.amazon.com/tag-editor/latest/userguide/tagging.html}{Tagging AWS Resources}.

See \url{https://paws-r.github.io/docs/rekognition/tag_resource.html} for full documentation.
}
\keyword{internal}
