% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_batch_describe_model_package}
\alias{sagemaker_batch_describe_model_package}
\title{This action batch describes a list of versioned model packages}
\usage{
sagemaker_batch_describe_model_package(ModelPackageArnList)
}
\arguments{
\item{ModelPackageArnList}{[required] The list of Amazon Resource Name (ARN) of the model package groups.}
}
\description{
This action batch describes a list of versioned model packages

See \url{https://www.paws-r-sdk.com/docs/sagemaker_batch_describe_model_package/} for full documentation.
}
\keyword{internal}
