% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_create_app_image_config}
\alias{sagemaker_create_app_image_config}
\title{Creates a configuration for running a SageMaker image as a KernelGateway
app}
\usage{
sagemaker_create_app_image_config(
  AppImageConfigName,
  Tags = NULL,
  KernelGatewayImageConfig = NULL
)
}
\arguments{
\item{AppImageConfigName}{[required] The name of the AppImageConfig. Must be unique to your account.}

\item{Tags}{A list of tags to apply to the AppImageConfig.}

\item{KernelGatewayImageConfig}{The KernelGatewayImageConfig. You can only specify one image kernel in
the AppImageConfig API. This kernel will be shown to users before the
image starts. Once the image runs, all kernels are visible in
JupyterLab.}
}
\description{
Creates a configuration for running a SageMaker image as a KernelGateway app. The configuration specifies the Amazon Elastic File System (EFS) storage volume on the image, and a list of the kernels in the image.

See \url{https://paws-r.github.io/docs/sagemaker/create_app_image_config.html} for full documentation.
}
\keyword{internal}
