% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_create_model_package_group}
\alias{sagemaker_create_model_package_group}
\title{Creates a model group}
\usage{
sagemaker_create_model_package_group(ModelPackageGroupName,
  ModelPackageGroupDescription, Tags)
}
\arguments{
\item{ModelPackageGroupName}{[required] The name of the model group.}

\item{ModelPackageGroupDescription}{A description for the model group.}

\item{Tags}{A list of key value pairs associated with the model group. For more
information, see \href{https://docs.aws.amazon.com/general/latest/gr/aws_tagging.html}{Tagging AWS resources}
in the \emph{AWS General Reference Guide}.}
}
\value{
A list with the following syntax:\preformatted{list(
  ModelPackageGroupArn = "string"
)
}
}
\description{
Creates a model group. A model group contains a group of model versions.
}
\section{Request syntax}{
\preformatted{svc$create_model_package_group(
  ModelPackageGroupName = "string",
  ModelPackageGroupDescription = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
