% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_create_notebook_instance_lifecycle_config}
\alias{sagemaker_create_notebook_instance_lifecycle_config}
\title{Creates a lifecycle configuration that you can associate with a notebook
instance}
\usage{
sagemaker_create_notebook_instance_lifecycle_config(
  NotebookInstanceLifecycleConfigName,
  OnCreate = NULL,
  OnStart = NULL,
  Tags = NULL
)
}
\arguments{
\item{NotebookInstanceLifecycleConfigName}{[required] The name of the lifecycle configuration.}

\item{OnCreate}{A shell script that runs only once, when you create a notebook instance.
The shell script must be a base64-encoded string.}

\item{OnStart}{A shell script that runs every time you start a notebook instance,
including when you create the notebook instance. The shell script must
be a base64-encoded string.}

\item{Tags}{An array of key-value pairs. You can use tags to categorize your Amazon
Web Services resources in different ways, for example, by purpose,
owner, or environment. For more information, see \href{https://docs.aws.amazon.com/tag-editor/latest/userguide/tagging.html}{Tagging Amazon Web Services Resources}.}
}
\description{
Creates a lifecycle configuration that you can associate with a notebook instance. A \emph{lifecycle configuration} is a collection of shell scripts that run when you create or start a notebook instance.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_create_notebook_instance_lifecycle_config/} for full documentation.
}
\keyword{internal}
