% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_create_workteam}
\alias{sagemaker_create_workteam}
\title{Creates a new work team for labeling your data}
\usage{
sagemaker_create_workteam(WorkteamName, WorkforceName,
  MemberDefinitions, Description, NotificationConfiguration, Tags)
}
\arguments{
\item{WorkteamName}{[required] The name of the work team. Use this name to identify the work team.}

\item{WorkforceName}{The name of the workforce.}

\item{MemberDefinitions}{[required] A list of \code{MemberDefinition} objects that contains objects that identify
the workers that make up the work team.

Workforces can be created using Amazon Cognito or your own OIDC Identity
Provider (IdP). For private workforces created using Amazon Cognito use
\code{CognitoMemberDefinition}. For workforces created using your own OIDC
identity provider (IdP) use \code{OidcMemberDefinition}. Do not provide input
for both of these parameters in a single request.

For workforces created using Amazon Cognito, private work teams
correspond to Amazon Cognito \emph{user groups} within the user pool used to
create a workforce. All of the \code{CognitoMemberDefinition} objects that
make up the member definition must have the same \code{ClientId} and
\code{UserPool} values. To add a Amazon Cognito user group to an existing
worker pool, see Adding groups to a User Pool. For more information
about user pools, see \href{https://docs.aws.amazon.com/cognito/latest/developerguide/cognito-user-identity-pools.html}{Amazon Cognito User Pools}.

For workforces created using your own OIDC IdP, specify the user groups
that you want to include in your private work team in
\code{OidcMemberDefinition} by listing those groups in \code{Groups}.}

\item{Description}{[required] A description of the work team.}

\item{NotificationConfiguration}{Configures notification of workers regarding available or expiring work
items.}

\item{Tags}{An array of key-value pairs.

For more information, see \href{https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-resource-tags.html}{Resource Tag}
and \href{https://docs.aws.amazon.com/awsaccountbilling/latest/aboutv2/cost-alloc-tags.html#allocation-what}{Using Cost Allocation Tags}
in the \emph{AWS Billing and Cost Management User Guide}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  WorkteamArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Creates a new work team for labeling your data. A work team is defined
by one or more Amazon Cognito user pools. You must first create the user
pools before you can create a work team.

You cannot create more than 25 work teams in an account and region.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_workteam(
  WorkteamName = "string",
  WorkforceName = "string",
  MemberDefinitions = list(
    list(
      CognitoMemberDefinition = list(
        UserPool = "string",
        UserGroup = "string",
        ClientId = "string"
      ),
      OidcMemberDefinition = list(
        Groups = list(
          "string"
        )
      )
    )
  ),
  Description = "string",
  NotificationConfiguration = list(
    NotificationTopicArn = "string"
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
