% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_action}
\alias{sagemaker_delete_action}
\title{Deletes an action}
\usage{
sagemaker_delete_action(ActionName)
}
\arguments{
\item{ActionName}{[required] The name of the action to delete.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ActionArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Deletes an action.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_action(
  ActionName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
