% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_app_image_config}
\alias{sagemaker_delete_app_image_config}
\title{Deletes an AppImageConfig}
\usage{
sagemaker_delete_app_image_config(AppImageConfigName)
}
\arguments{
\item{AppImageConfigName}{[required] The name of the AppImageConfig to delete.}
}
\value{
An empty list.
}
\description{
Deletes an AppImageConfig.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_app_image_config(
  AppImageConfigName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
