% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_artifact}
\alias{sagemaker_delete_artifact}
\title{Deletes an artifact}
\usage{
sagemaker_delete_artifact(ArtifactArn = NULL, Source = NULL)
}
\arguments{
\item{ArtifactArn}{The Amazon Resource Name (ARN) of the artifact to delete.}

\item{Source}{The URI of the source.}
}
\description{
Deletes an artifact. Either \code{ArtifactArn} or \code{Source} must be specified.

See \url{https://paws-r.github.io/docs/sagemaker/delete_artifact.html} for full documentation.
}
\keyword{internal}
