% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_code_repository}
\alias{sagemaker_delete_code_repository}
\title{Deletes the specified Git repository from your account}
\usage{
sagemaker_delete_code_repository(CodeRepositoryName)
}
\arguments{
\item{CodeRepositoryName}{[required] The name of the Git repository to delete.}
}
\value{
An empty list.
}
\description{
Deletes the specified Git repository from your account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_code_repository(
  CodeRepositoryName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
