% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_data_quality_job_definition}
\alias{sagemaker_delete_data_quality_job_definition}
\title{Deletes a data quality monitoring job definition}
\usage{
sagemaker_delete_data_quality_job_definition(JobDefinitionName)
}
\arguments{
\item{JobDefinitionName}{[required] The name of the data quality monitoring job definition to delete.}
}
\description{
Deletes a data quality monitoring job definition.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_delete_data_quality_job_definition/} for full documentation.
}
\keyword{internal}
