% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_endpoint_config}
\alias{sagemaker_delete_endpoint_config}
\title{Deletes an endpoint configuration}
\usage{
sagemaker_delete_endpoint_config(EndpointConfigName)
}
\arguments{
\item{EndpointConfigName}{[required] The name of the endpoint configuration that you want to delete.}
}
\description{
Deletes an endpoint configuration. The \code{\link[=sagemaker_delete_endpoint_config]{delete_endpoint_config}} API deletes only the specified configuration. It does not delete endpoints created using the configuration.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_delete_endpoint_config/} for full documentation.
}
\keyword{internal}
