% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_flow_definition}
\alias{sagemaker_delete_flow_definition}
\title{Deletes the specified flow definition}
\usage{
sagemaker_delete_flow_definition(FlowDefinitionName)
}
\arguments{
\item{FlowDefinitionName}{[required] The name of the flow definition you are deleting.}
}
\description{
Deletes the specified flow definition.
}
\section{Request syntax}{
\preformatted{svc$delete_flow_definition(
  FlowDefinitionName = "string"
)
}
}

\keyword{internal}
