% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_human_task_ui}
\alias{sagemaker_delete_human_task_ui}
\title{Use this operation to delete a human task user interface (worker task
template)}
\usage{
sagemaker_delete_human_task_ui(HumanTaskUiName)
}
\arguments{
\item{HumanTaskUiName}{[required] The name of the human task user interface (work task template) you want
to delete.}
}
\description{
Use this operation to delete a human task user interface (worker task
template).

To see a list of human task user interfaces (work task templates) in
your account, use . When you delete a worker task template, it no longer
appears when you call \code{ListHumanTaskUis}.
}
\section{Request syntax}{
\preformatted{svc$delete_human_task_ui(
  HumanTaskUiName = "string"
)
}
}

\keyword{internal}
