% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_image}
\alias{sagemaker_delete_image}
\title{Deletes a SageMaker image and all versions of the image}
\usage{
sagemaker_delete_image(ImageName)
}
\arguments{
\item{ImageName}{[required] The name of the image to delete.}
}
\description{
Deletes a SageMaker image and all versions of the image. The container
images aren't deleted.
}
\section{Request syntax}{
\preformatted{svc$delete_image(
  ImageName = "string"
)
}
}

\keyword{internal}
