% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_model}
\alias{sagemaker_delete_model}
\title{Deletes a model}
\usage{
sagemaker_delete_model(ModelName)
}
\arguments{
\item{ModelName}{[required] The name of the model to delete.}
}
\description{
Deletes a model. The \code{\link[=sagemaker_delete_model]{delete_model}} API deletes only the model entry that was created in SageMaker when you called the \code{\link[=sagemaker_create_model]{create_model}} API. It does not delete model artifacts, inference code, or the IAM role that you specified when creating the model.

See \url{https://paws-r.github.io/docs/sagemaker/delete_model.html} for full documentation.
}
\keyword{internal}
