% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_model_bias_job_definition}
\alias{sagemaker_delete_model_bias_job_definition}
\title{Deletes an Amazon SageMaker model bias job definition}
\usage{
sagemaker_delete_model_bias_job_definition(JobDefinitionName)
}
\arguments{
\item{JobDefinitionName}{[required] The name of the model bias job definition to delete.}
}
\value{
An empty list.
}
\description{
Deletes an Amazon SageMaker model bias job definition.
}
\section{Request syntax}{
\preformatted{svc$delete_model_bias_job_definition(
  JobDefinitionName = "string"
)
}
}

\keyword{internal}
