% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_model_card}
\alias{sagemaker_delete_model_card}
\title{Deletes an Amazon SageMaker Model Card}
\usage{
sagemaker_delete_model_card(ModelCardName)
}
\arguments{
\item{ModelCardName}{[required] The name of the model card to delete.}
}
\description{
Deletes an Amazon SageMaker Model Card.

See \url{https://paws-r.github.io/docs/sagemaker/delete_model_card.html} for full documentation.
}
\keyword{internal}
