% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_model_explainability_job_definition}
\alias{sagemaker_delete_model_explainability_job_definition}
\title{Deletes an Amazon SageMaker model explainability job definition}
\usage{
sagemaker_delete_model_explainability_job_definition(JobDefinitionName)
}
\arguments{
\item{JobDefinitionName}{[required] The name of the model explainability job definition to delete.}
}
\description{
Deletes an Amazon SageMaker model explainability job definition.

See \url{https://paws-r.github.io/docs/sagemaker/delete_model_explainability_job_definition.html} for full documentation.
}
\keyword{internal}
