% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_model_package}
\alias{sagemaker_delete_model_package}
\title{Deletes a model package}
\usage{
sagemaker_delete_model_package(ModelPackageName)
}
\arguments{
\item{ModelPackageName}{[required] The name of the model package. The name must have 1 to 63 characters.
Valid characters are a-z, A-Z, 0-9, and - (hyphen).}
}
\description{
Deletes a model package.

A model package is used to create Amazon SageMaker models or list on AWS
Marketplace. Buyers can subscribe to model packages listed on AWS
Marketplace to create models in Amazon SageMaker.
}
\section{Request syntax}{
\preformatted{svc$delete_model_package(
  ModelPackageName = "string"
)
}
}

\keyword{internal}
