% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_model_package_group_policy}
\alias{sagemaker_delete_model_package_group_policy}
\title{Deletes a model group resource policy}
\usage{
sagemaker_delete_model_package_group_policy(ModelPackageGroupName)
}
\arguments{
\item{ModelPackageGroupName}{[required] The name of the model group for which to delete the policy.}
}
\value{
An empty list.
}
\description{
Deletes a model group resource policy.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_model_package_group_policy(
  ModelPackageGroupName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
