% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_notebook_instance}
\alias{sagemaker_delete_notebook_instance}
\title{Deletes an SageMaker notebook instance}
\usage{
sagemaker_delete_notebook_instance(NotebookInstanceName)
}
\arguments{
\item{NotebookInstanceName}{[required] The name of the SageMaker notebook instance to delete.}
}
\description{
Deletes an SageMaker notebook instance. Before you can delete a notebook instance, you must call the \code{\link[=sagemaker_stop_notebook_instance]{stop_notebook_instance}} API.

See \url{https://paws-r.github.io/docs/sagemaker/delete_notebook_instance.html} for full documentation.
}
\keyword{internal}
