% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_pipeline}
\alias{sagemaker_delete_pipeline}
\title{Deletes a pipeline if there are no running instances of the pipeline}
\usage{
sagemaker_delete_pipeline(PipelineName, ClientRequestToken)
}
\arguments{
\item{PipelineName}{[required] The name of the pipeline to delete.}

\item{ClientRequestToken}{[required] A unique, case-sensitive identifier that you provide to ensure the
idempotency of the operation. An idempotent operation completes no more
than one time.}
}
\value{
A list with the following syntax:\preformatted{list(
  PipelineArn = "string"
)
}
}
\description{
Deletes a pipeline if there are no running instances of the pipeline. To
delete a pipeline, you must stop all running instances of the pipeline
using the \code{\link[=sagemaker_stop_pipeline_execution]{stop_pipeline_execution}}
API. When you delete a pipeline, all instances of the pipeline are
deleted.
}
\section{Request syntax}{
\preformatted{svc$delete_pipeline(
  PipelineName = "string",
  ClientRequestToken = "string"
)
}
}

\keyword{internal}
