% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_tags}
\alias{sagemaker_delete_tags}
\title{Deletes the specified tags from an SageMaker resource}
\usage{
sagemaker_delete_tags(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource whose tags you want to
delete.}

\item{TagKeys}{[required] An array or one or more tag keys to delete.}
}
\description{
Deletes the specified tags from an SageMaker resource.

See \url{https://paws-r.github.io/docs/sagemaker/delete_tags.html} for full documentation.
}
\keyword{internal}
