% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_trial}
\alias{sagemaker_delete_trial}
\title{Deletes the specified trial}
\usage{
sagemaker_delete_trial(TrialName)
}
\arguments{
\item{TrialName}{[required] The name of the trial to delete.}
}
\value{
A list with the following syntax:\preformatted{list(
  TrialArn = "string"
)
}
}
\description{
Deletes the specified trial. All trial components that make up the trial
must be deleted first. Use the
\code{\link[=sagemaker_describe_trial_component]{describe_trial_component}} API to
get the list of trial components.
}
\section{Request syntax}{
\preformatted{svc$delete_trial(
  TrialName = "string"
)
}
}

\keyword{internal}
