% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_trial_component}
\alias{sagemaker_delete_trial_component}
\title{Deletes the specified trial component}
\usage{
sagemaker_delete_trial_component(TrialComponentName)
}
\arguments{
\item{TrialComponentName}{[required] The name of the component to delete.}
}
\description{
Deletes the specified trial component. A trial component must be disassociated from all trials before the trial component can be deleted. To disassociate a trial component from a trial, call the \code{\link[=sagemaker_disassociate_trial_component]{disassociate_trial_component}} API.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_delete_trial_component/} for full documentation.
}
\keyword{internal}
