% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_trial_component}
\alias{sagemaker_delete_trial_component}
\title{Deletes the specified trial component}
\usage{
sagemaker_delete_trial_component(TrialComponentName)
}
\arguments{
\item{TrialComponentName}{[required] The name of the component to delete.}
}
\description{
Deletes the specified trial component. A trial component must be
disassociated from all trials before the trial component can be deleted.
To disassociate a trial component from a trial, call the
DisassociateTrialComponent API.
}
\section{Request syntax}{
\preformatted{svc$delete_trial_component(
  TrialComponentName = "string"
)
}
}

\keyword{internal}
