% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_user_profile}
\alias{sagemaker_delete_user_profile}
\title{Deletes a user profile}
\usage{
sagemaker_delete_user_profile(DomainId, UserProfileName)
}
\arguments{
\item{DomainId}{[required] The domain ID.}

\item{UserProfileName}{[required] The user profile name.}
}
\description{
Deletes a user profile. When a user profile is deleted, the user loses access to their EFS volume, including data, notebooks, and other artifacts.

See \url{https://paws-r.github.io/docs/sagemaker/delete_user_profile.html} for full documentation.
}
\keyword{internal}
