% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_workforce}
\alias{sagemaker_delete_workforce}
\title{Use this operation to delete a workforce}
\usage{
sagemaker_delete_workforce(WorkforceName)
}
\arguments{
\item{WorkforceName}{[required] The name of the workforce.}
}
\description{
Use this operation to delete a workforce.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_delete_workforce/} for full documentation.
}
\keyword{internal}
