% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_delete_workteam}
\alias{sagemaker_delete_workteam}
\title{Deletes an existing work team}
\usage{
sagemaker_delete_workteam(WorkteamName)
}
\arguments{
\item{WorkteamName}{[required] The name of the work team to delete.}
}
\description{
Deletes an existing work team. This operation can't be undone.
}
\section{Request syntax}{
\preformatted{svc$delete_workteam(
  WorkteamName = "string"
)
}
}

\keyword{internal}
