% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_action}
\alias{sagemaker_describe_action}
\title{Describes an action}
\usage{
sagemaker_describe_action(ActionName)
}
\arguments{
\item{ActionName}{[required] The name of the action to describe.}
}
\description{
Describes an action.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_action/} for full documentation.
}
\keyword{internal}
