% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_app}
\alias{sagemaker_describe_app}
\title{Describes the app}
\usage{
sagemaker_describe_app(
  DomainId,
  UserProfileName = NULL,
  SpaceName = NULL,
  AppType,
  AppName
)
}
\arguments{
\item{DomainId}{[required] The domain ID.}

\item{UserProfileName}{The user profile name. If this value is not set, then \code{SpaceName} must
be set.}

\item{SpaceName}{The name of the space.}

\item{AppType}{[required] The type of app.}

\item{AppName}{[required] The name of the app.}
}
\description{
Describes the app.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_app/} for full documentation.
}
\keyword{internal}
