% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_artifact}
\alias{sagemaker_describe_artifact}
\title{Describes an artifact}
\usage{
sagemaker_describe_artifact(ArtifactArn)
}
\arguments{
\item{ArtifactArn}{[required] The Amazon Resource Name (ARN) of the artifact to describe.}
}
\description{
Describes an artifact.

See \url{https://paws-r.github.io/docs/sagemaker/describe_artifact.html} for full documentation.
}
\keyword{internal}
