% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_auto_ml_job}
\alias{sagemaker_describe_auto_ml_job}
\title{Returns information about an Amazon SageMaker AutoML job}
\usage{
sagemaker_describe_auto_ml_job(AutoMLJobName)
}
\arguments{
\item{AutoMLJobName}{[required] Requests information about an AutoML job using its unique name.}
}
\description{
Returns information about an Amazon SageMaker AutoML job.

See \url{https://paws-r.github.io/docs/sagemaker/describe_auto_ml_job.html} for full documentation.
}
\keyword{internal}
