% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_code_repository}
\alias{sagemaker_describe_code_repository}
\title{Gets details about the specified Git repository}
\usage{
sagemaker_describe_code_repository(CodeRepositoryName)
}
\arguments{
\item{CodeRepositoryName}{[required] The name of the Git repository to describe.}
}
\description{
Gets details about the specified Git repository.
}
\section{Request syntax}{
\preformatted{svc$describe_code_repository(
  CodeRepositoryName = "string"
)
}
}

\keyword{internal}
