% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_compilation_job}
\alias{sagemaker_describe_compilation_job}
\title{Returns information about a model compilation job}
\usage{
sagemaker_describe_compilation_job(CompilationJobName)
}
\arguments{
\item{CompilationJobName}{[required] The name of the model compilation job that you want information about.}
}
\description{
Returns information about a model compilation job.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_compilation_job/} for full documentation.
}
\keyword{internal}
