% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_context}
\alias{sagemaker_describe_context}
\title{Describes a context}
\usage{
sagemaker_describe_context(ContextName)
}
\arguments{
\item{ContextName}{[required] The name of the context to describe.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ContextName = "string",
  ContextArn = "string",
  Source = list(
    SourceUri = "string",
    SourceType = "string",
    SourceId = "string"
  ),
  ContextType = "string",
  Description = "string",
  Properties = list(
    "string"
  ),
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  CreatedBy = list(
    UserProfileArn = "string",
    UserProfileName = "string",
    DomainId = "string"
  ),
  LastModifiedTime = as.POSIXct(
    "2015-01-01"
  ),
  LastModifiedBy = list(
    UserProfileArn = "string",
    UserProfileName = "string",
    DomainId = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes a context.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_context(
  ContextName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
