% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_device_fleet}
\alias{sagemaker_describe_device_fleet}
\title{A description of the fleet the device belongs to}
\usage{
sagemaker_describe_device_fleet(DeviceFleetName)
}
\arguments{
\item{DeviceFleetName}{[required] The name of the fleet.}
}
\description{
A description of the fleet the device belongs to.

See \url{https://paws-r.github.io/docs/sagemaker/describe_device_fleet.html} for full documentation.
}
\keyword{internal}
