% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_domain}
\alias{sagemaker_describe_domain}
\title{The description of the domain}
\usage{
sagemaker_describe_domain(DomainId)
}
\arguments{
\item{DomainId}{[required] The domain ID.}
}
\description{
The description of the domain.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_domain/} for full documentation.
}
\keyword{internal}
