% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_edge_packaging_job}
\alias{sagemaker_describe_edge_packaging_job}
\title{A description of edge packaging jobs}
\usage{
sagemaker_describe_edge_packaging_job(EdgePackagingJobName)
}
\arguments{
\item{EdgePackagingJobName}{[required] The name of the edge packaging job.}
}
\description{
A description of edge packaging jobs.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_edge_packaging_job/} for full documentation.
}
\keyword{internal}
