% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_endpoint_config}
\alias{sagemaker_describe_endpoint_config}
\title{Returns the description of an endpoint configuration created using the
CreateEndpointConfig API}
\usage{
sagemaker_describe_endpoint_config(EndpointConfigName)
}
\arguments{
\item{EndpointConfigName}{[required] The name of the endpoint configuration.}
}
\description{
Returns the description of an endpoint configuration created using the \code{\link[=sagemaker_create_endpoint_config]{create_endpoint_config}} API.

See \url{https://paws-r.github.io/docs/sagemaker/describe_endpoint_config.html} for full documentation.
}
\keyword{internal}
