% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_feature_group}
\alias{sagemaker_describe_feature_group}
\title{Use this operation to describe a FeatureGroup}
\usage{
sagemaker_describe_feature_group(FeatureGroupName, NextToken = NULL)
}
\arguments{
\item{FeatureGroupName}{[required] The name or Amazon Resource Name (ARN) of the \code{FeatureGroup} you want
described.}

\item{NextToken}{A token to resume pagination of the list of \code{Features}
(\code{FeatureDefinitions}). 2,500 \code{Features} are returned by default.}
}
\description{
Use this operation to describe a \code{FeatureGroup}. The response includes information on the creation time, \code{FeatureGroup} name, the unique identifier for each \code{FeatureGroup}, and more.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_feature_group/} for full documentation.
}
\keyword{internal}
