% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_flow_definition}
\alias{sagemaker_describe_flow_definition}
\title{Returns information about the specified flow definition}
\usage{
sagemaker_describe_flow_definition(FlowDefinitionName)
}
\arguments{
\item{FlowDefinitionName}{[required] The name of the flow definition.}
}
\description{
Returns information about the specified flow definition.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_flow_definition/} for full documentation.
}
\keyword{internal}
