% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_hyper_parameter_tuning_job}
\alias{sagemaker_describe_hyper_parameter_tuning_job}
\title{Returns a description of a hyperparameter tuning job, depending on the
fields selected}
\usage{
sagemaker_describe_hyper_parameter_tuning_job(HyperParameterTuningJobName)
}
\arguments{
\item{HyperParameterTuningJobName}{[required] The name of the tuning job.}
}
\description{
Returns a description of a hyperparameter tuning job, depending on the fields selected. These fields can include the name, Amazon Resource Name (ARN), job status of your tuning job and more.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_hyper_parameter_tuning_job/} for full documentation.
}
\keyword{internal}
