% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_image_version}
\alias{sagemaker_describe_image_version}
\title{Describes a version of a SageMaker image}
\usage{
sagemaker_describe_image_version(ImageName, Version = NULL, Alias = NULL)
}
\arguments{
\item{ImageName}{[required] The name of the image.}

\item{Version}{The version of the image. If not specified, the latest version is
described.}

\item{Alias}{The alias of the image version.}
}
\description{
Describes a version of a SageMaker image.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_image_version/} for full documentation.
}
\keyword{internal}
