% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_lineage_group}
\alias{sagemaker_describe_lineage_group}
\title{Provides a list of properties for the requested lineage group}
\usage{
sagemaker_describe_lineage_group(LineageGroupName)
}
\arguments{
\item{LineageGroupName}{[required] The name of the lineage group.}
}
\description{
Provides a list of properties for the requested lineage group. For more information, see \href{https://docs.aws.amazon.com/sagemaker/latest/dg/xaccount-lineage-tracking.html}{Cross-Account Lineage Tracking} in the \emph{Amazon SageMaker Developer Guide}.

See \url{https://paws-r.github.io/docs/sagemaker/describe_lineage_group.html} for full documentation.
}
\keyword{internal}
