% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_model_card}
\alias{sagemaker_describe_model_card}
\title{Describes the content, creation time, and security configuration of an
Amazon SageMaker Model Card}
\usage{
sagemaker_describe_model_card(ModelCardName, ModelCardVersion = NULL)
}
\arguments{
\item{ModelCardName}{[required] The name of the model card to describe.}

\item{ModelCardVersion}{The version of the model card to describe. If a version is not provided,
then the latest version of the model card is described.}
}
\description{
Describes the content, creation time, and security configuration of an Amazon SageMaker Model Card.

See \url{https://paws-r.github.io/docs/sagemaker/describe_model_card.html} for full documentation.
}
\keyword{internal}
