% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_model_explainability_job_definition}
\alias{sagemaker_describe_model_explainability_job_definition}
\title{Returns a description of a model explainability job definition}
\usage{
sagemaker_describe_model_explainability_job_definition(JobDefinitionName)
}
\arguments{
\item{JobDefinitionName}{[required] The name of the model explainability job definition. The name must be
unique within an Amazon Web Services Region in the Amazon Web Services
account.}
}
\description{
Returns a description of a model explainability job definition.

See \url{https://paws-r.github.io/docs/sagemaker/describe_model_explainability_job_definition.html} for full documentation.
}
\keyword{internal}
