% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_model_package_group}
\alias{sagemaker_describe_model_package_group}
\title{Gets a description for the specified model group}
\usage{
sagemaker_describe_model_package_group(ModelPackageGroupName)
}
\arguments{
\item{ModelPackageGroupName}{[required] The name of gthe model group to describe.}
}
\description{
Gets a description for the specified model group.

See \url{https://paws-r.github.io/docs/sagemaker/describe_model_package_group.html} for full documentation.
}
\keyword{internal}
