% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_notebook_instance}
\alias{sagemaker_describe_notebook_instance}
\title{Returns information about a notebook instance}
\usage{
sagemaker_describe_notebook_instance(NotebookInstanceName)
}
\arguments{
\item{NotebookInstanceName}{[required] The name of the notebook instance that you want information about.}
}
\description{
Returns information about a notebook instance.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_notebook_instance/} for full documentation.
}
\keyword{internal}
