% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_notebook_instance_lifecycle_config}
\alias{sagemaker_describe_notebook_instance_lifecycle_config}
\title{Returns a description of a notebook instance lifecycle configuration}
\usage{
sagemaker_describe_notebook_instance_lifecycle_config(
  NotebookInstanceLifecycleConfigName
)
}
\arguments{
\item{NotebookInstanceLifecycleConfigName}{[required] The name of the lifecycle configuration to describe.}
}
\description{
Returns a description of a notebook instance lifecycle configuration.

See \url{https://paws-r.github.io/docs/sagemaker/describe_notebook_instance_lifecycle_config.html} for full documentation.
}
\keyword{internal}
