% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_pipeline}
\alias{sagemaker_describe_pipeline}
\title{Describes the details of a pipeline}
\usage{
sagemaker_describe_pipeline(PipelineName)
}
\arguments{
\item{PipelineName}{[required] The name of the pipeline to describe.}
}
\description{
Describes the details of a pipeline.
}
\section{Request syntax}{
\preformatted{svc$describe_pipeline(
  PipelineName = "string"
)
}
}

\keyword{internal}
