% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_pipeline}
\alias{sagemaker_describe_pipeline}
\title{Describes the details of a pipeline}
\usage{
sagemaker_describe_pipeline(PipelineName, PipelineVersionId = NULL)
}
\arguments{
\item{PipelineName}{[required] The name or Amazon Resource Name (ARN) of the pipeline to describe.}

\item{PipelineVersionId}{The ID of the pipeline version to describe.}
}
\description{
Describes the details of a pipeline.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_pipeline/} for full documentation.
}
\keyword{internal}
