% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_pipeline_definition_for_execution}
\alias{sagemaker_describe_pipeline_definition_for_execution}
\title{Describes the details of an execution's pipeline definition}
\usage{
sagemaker_describe_pipeline_definition_for_execution(PipelineExecutionArn)
}
\arguments{
\item{PipelineExecutionArn}{[required] The Amazon Resource Name (ARN) of the pipeline execution.}
}
\description{
Describes the details of an execution's pipeline definition.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_pipeline_definition_for_execution/} for full documentation.
}
\keyword{internal}
