% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_pipeline_execution}
\alias{sagemaker_describe_pipeline_execution}
\title{Describes the details of a pipeline execution}
\usage{
sagemaker_describe_pipeline_execution(PipelineExecutionArn)
}
\arguments{
\item{PipelineExecutionArn}{[required] The Amazon Resource Name (ARN) of the pipeline execution.}
}
\description{
Describes the details of a pipeline execution.
}
\section{Request syntax}{
\preformatted{svc$describe_pipeline_execution(
  PipelineExecutionArn = "string"
)
}
}

\keyword{internal}
