% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_processing_job}
\alias{sagemaker_describe_processing_job}
\title{Returns a description of a processing job}
\usage{
sagemaker_describe_processing_job(ProcessingJobName)
}
\arguments{
\item{ProcessingJobName}{[required] The name of the processing job. The name must be unique within an AWS
Region in the AWS account.}
}
\value{
A list with the following syntax:\preformatted{list(
  ProcessingInputs = list(
    list(
      InputName = "string",
      AppManaged = TRUE|FALSE,
      S3Input = list(
        S3Uri = "string",
        LocalPath = "string",
        S3DataType = "ManifestFile"|"S3Prefix",
        S3InputMode = "Pipe"|"File",
        S3DataDistributionType = "FullyReplicated"|"ShardedByS3Key",
        S3CompressionType = "None"|"Gzip"
      ),
      DatasetDefinition = list(
        AthenaDatasetDefinition = list(
          Catalog = "string",
          Database = "string",
          QueryString = "string",
          WorkGroup = "string",
          OutputS3Uri = "string",
          KmsKeyId = "string",
          OutputFormat = "PARQUET"|"ORC"|"AVRO"|"JSON"|"TEXTFILE",
          OutputCompression = "GZIP"|"SNAPPY"|"ZLIB"
        ),
        RedshiftDatasetDefinition = list(
          ClusterId = "string",
          Database = "string",
          DbUser = "string",
          QueryString = "string",
          ClusterRoleArn = "string",
          OutputS3Uri = "string",
          KmsKeyId = "string",
          OutputFormat = "PARQUET"|"CSV",
          OutputCompression = "None"|"GZIP"|"BZIP2"|"ZSTD"|"SNAPPY"
        ),
        LocalPath = "string",
        DataDistributionType = "FullyReplicated"|"ShardedByS3Key",
        InputMode = "Pipe"|"File"
      )
    )
  ),
  ProcessingOutputConfig = list(
    Outputs = list(
      list(
        OutputName = "string",
        S3Output = list(
          S3Uri = "string",
          LocalPath = "string",
          S3UploadMode = "Continuous"|"EndOfJob"
        ),
        FeatureStoreOutput = list(
          FeatureGroupName = "string"
        ),
        AppManaged = TRUE|FALSE
      )
    ),
    KmsKeyId = "string"
  ),
  ProcessingJobName = "string",
  ProcessingResources = list(
    ClusterConfig = list(
      InstanceCount = 123,
      InstanceType = "ml.t3.medium"|"ml.t3.large"|"ml.t3.xlarge"|"ml.t3.2xlarge"|"ml.m4.xlarge"|"ml.m4.2xlarge"|"ml.m4.4xlarge"|"ml.m4.10xlarge"|"ml.m4.16xlarge"|"ml.c4.xlarge"|"ml.c4.2xlarge"|"ml.c4.4xlarge"|"ml.c4.8xlarge"|"ml.p2.xlarge"|"ml.p2.8xlarge"|"ml.p2.16xlarge"|"ml.p3.2xlarge"|"ml.p3.8xlarge"|"ml.p3.16xlarge"|"ml.c5.xlarge"|"ml.c5.2xlarge"|"ml.c5.4xlarge"|"ml.c5.9xlarge"|"ml.c5.18xlarge"|"ml.m5.large"|"ml.m5.xlarge"|"ml.m5.2xlarge"|"ml.m5.4xlarge"|"ml.m5.12xlarge"|"ml.m5.24xlarge"|"ml.r5.large"|"ml.r5.xlarge"|"ml.r5.2xlarge"|"ml.r5.4xlarge"|"ml.r5.8xlarge"|"ml.r5.12xlarge"|"ml.r5.16xlarge"|"ml.r5.24xlarge",
      VolumeSizeInGB = 123,
      VolumeKmsKeyId = "string"
    )
  ),
  StoppingCondition = list(
    MaxRuntimeInSeconds = 123
  ),
  AppSpecification = list(
    ImageUri = "string",
    ContainerEntrypoint = list(
      "string"
    ),
    ContainerArguments = list(
      "string"
    )
  ),
  Environment = list(
    "string"
  ),
  NetworkConfig = list(
    EnableInterContainerTrafficEncryption = TRUE|FALSE,
    EnableNetworkIsolation = TRUE|FALSE,
    VpcConfig = list(
      SecurityGroupIds = list(
        "string"
      ),
      Subnets = list(
        "string"
      )
    )
  ),
  RoleArn = "string",
  ExperimentConfig = list(
    ExperimentName = "string",
    TrialName = "string",
    TrialComponentDisplayName = "string"
  ),
  ProcessingJobArn = "string",
  ProcessingJobStatus = "InProgress"|"Completed"|"Failed"|"Stopping"|"Stopped",
  ExitMessage = "string",
  FailureReason = "string",
  ProcessingEndTime = as.POSIXct(
    "2015-01-01"
  ),
  ProcessingStartTime = as.POSIXct(
    "2015-01-01"
  ),
  LastModifiedTime = as.POSIXct(
    "2015-01-01"
  ),
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  MonitoringScheduleArn = "string",
  AutoMLJobArn = "string",
  TrainingJobArn = "string"
)
}
}
\description{
Returns a description of a processing job.
}
\section{Request syntax}{
\preformatted{svc$describe_processing_job(
  ProcessingJobName = "string"
)
}
}

\keyword{internal}
