% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_space}
\alias{sagemaker_describe_space}
\title{Describes the space}
\usage{
sagemaker_describe_space(DomainId, SpaceName)
}
\arguments{
\item{DomainId}{[required] The ID of the associated domain.}

\item{SpaceName}{[required] The name of the space.}
}
\description{
Describes the space.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_space/} for full documentation.
}
\keyword{internal}
