% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_studio_lifecycle_config}
\alias{sagemaker_describe_studio_lifecycle_config}
\title{Describes the Studio Lifecycle Configuration}
\usage{
sagemaker_describe_studio_lifecycle_config(StudioLifecycleConfigName)
}
\arguments{
\item{StudioLifecycleConfigName}{[required] The name of the Studio Lifecycle Configuration to describe.}
}
\description{
Describes the Studio Lifecycle Configuration.

See \url{https://paws-r.github.io/docs/sagemaker/describe_studio_lifecycle_config.html} for full documentation.
}
\keyword{internal}
