% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_subscribed_workteam}
\alias{sagemaker_describe_subscribed_workteam}
\title{Gets information about a work team provided by a vendor}
\usage{
sagemaker_describe_subscribed_workteam(WorkteamArn)
}
\arguments{
\item{WorkteamArn}{[required] The Amazon Resource Name (ARN) of the subscribed work team to describe.}
}
\value{
A list with the following syntax:\preformatted{list(
  SubscribedWorkteam = list(
    WorkteamArn = "string",
    MarketplaceTitle = "string",
    SellerName = "string",
    MarketplaceDescription = "string",
    ListingId = "string"
  )
)
}
}
\description{
Gets information about a work team provided by a vendor. It returns
details about the subscription with a vendor in the AWS Marketplace.
}
\section{Request syntax}{
\preformatted{svc$describe_subscribed_workteam(
  WorkteamArn = "string"
)
}
}

\keyword{internal}
