% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_transform_job}
\alias{sagemaker_describe_transform_job}
\title{Returns information about a transform job}
\usage{
sagemaker_describe_transform_job(TransformJobName)
}
\arguments{
\item{TransformJobName}{[required] The name of the transform job that you want to view details of.}
}
\description{
Returns information about a transform job.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_describe_transform_job/} for full documentation.
}
\keyword{internal}
