% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_trial_component}
\alias{sagemaker_describe_trial_component}
\title{Provides a list of a trials component's properties}
\usage{
sagemaker_describe_trial_component(TrialComponentName)
}
\arguments{
\item{TrialComponentName}{[required] The name of the trial component to describe.}
}
\description{
Provides a list of a trials component's properties.

See \url{https://paws-r.github.io/docs/sagemaker/describe_trial_component.html} for full documentation.
}
\keyword{internal}
