% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_disable_sagemaker_servicecatalog_portfolio}
\alias{sagemaker_disable_sagemaker_servicecatalog_portfolio}
\title{Disables using Service Catalog in SageMaker}
\usage{
sagemaker_disable_sagemaker_servicecatalog_portfolio()
}
\description{
Disables using Service Catalog in SageMaker. Service Catalog is used to create SageMaker projects.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_disable_sagemaker_servicecatalog_portfolio/} for full documentation.
}
\keyword{internal}
