% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_get_device_fleet_report}
\alias{sagemaker_get_device_fleet_report}
\title{Describes a fleet}
\usage{
sagemaker_get_device_fleet_report(DeviceFleetName)
}
\arguments{
\item{DeviceFleetName}{[required] The name of the fleet.}
}
\description{
Describes a fleet.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_get_device_fleet_report/} for full documentation.
}
\keyword{internal}
