% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_get_model_package_group_policy}
\alias{sagemaker_get_model_package_group_policy}
\title{Gets a resource policy that manages access for a model group}
\usage{
sagemaker_get_model_package_group_policy(ModelPackageGroupName)
}
\arguments{
\item{ModelPackageGroupName}{[required] The name of the model group for which to get the resource policy.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ResourcePolicy = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets a resource policy that manages access for a model group. For
information about resource policies, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/access_policies_identity-vs-resource.html}{Identity-based policies and resource-based policies}
in the \emph{AWS Identity and Access Management User Guide.}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_model_package_group_policy(
  ModelPackageGroupName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
