% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_get_sagemaker_servicecatalog_portfolio_status}
\alias{sagemaker_get_sagemaker_servicecatalog_portfolio_status}
\title{Gets the status of Service Catalog in SageMaker}
\usage{
sagemaker_get_sagemaker_servicecatalog_portfolio_status()
}
\description{
Gets the status of Service Catalog in SageMaker. Service Catalog is used to create SageMaker projects.

See \url{https://paws-r.github.io/docs/sagemaker/get_sagemaker_servicecatalog_portfolio_status.html} for full documentation.
}
\keyword{internal}
