% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_get_sagemaker_servicecatalog_portfolio_status}
\alias{sagemaker_get_sagemaker_servicecatalog_portfolio_status}
\title{Gets the status of Service Catalog in SageMaker}
\usage{
sagemaker_get_sagemaker_servicecatalog_portfolio_status()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Status = "Enabled"|"Disabled"
)
}\if{html}{\out{</div>}}
}
\description{
Gets the status of Service Catalog in SageMaker. Service Catalog is used
to create SageMaker projects.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_sagemaker_servicecatalog_portfolio_status()
}\if{html}{\out{</div>}}
}

\keyword{internal}
