% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_auto_ml_jobs}
\alias{sagemaker_list_auto_ml_jobs}
\title{Request a list of jobs}
\usage{
sagemaker_list_auto_ml_jobs(
  CreationTimeAfter = NULL,
  CreationTimeBefore = NULL,
  LastModifiedTimeAfter = NULL,
  LastModifiedTimeBefore = NULL,
  NameContains = NULL,
  StatusEquals = NULL,
  SortOrder = NULL,
  SortBy = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{CreationTimeAfter}{Request a list of jobs, using a filter for time.}

\item{CreationTimeBefore}{Request a list of jobs, using a filter for time.}

\item{LastModifiedTimeAfter}{Request a list of jobs, using a filter for time.}

\item{LastModifiedTimeBefore}{Request a list of jobs, using a filter for time.}

\item{NameContains}{Request a list of jobs, using a search filter for name.}

\item{StatusEquals}{Request a list of jobs, using a filter for status.}

\item{SortOrder}{The sort order for the results. The default is \code{Descending}.}

\item{SortBy}{The parameter by which to sort the results. The default is \code{Name}.}

\item{MaxResults}{Request a list of jobs up to a specified limit.}

\item{NextToken}{If the previous response was truncated, you receive this token. Use it
in your next request to receive the next set of results.}
}
\description{
Request a list of jobs.

See \url{https://paws-r.github.io/docs/sagemaker/list_auto_ml_jobs.html} for full documentation.
}
\keyword{internal}
