% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_contexts}
\alias{sagemaker_list_contexts}
\title{Lists the contexts in your account and their properties}
\usage{
sagemaker_list_contexts(SourceUri, ContextType, CreatedAfter,
  CreatedBefore, SortBy, SortOrder, NextToken, MaxResults)
}
\arguments{
\item{SourceUri}{A filter that returns only contexts with the specified source URI.}

\item{ContextType}{A filter that returns only contexts of the specified type.}

\item{CreatedAfter}{A filter that returns only contexts created on or after the specified
time.}

\item{CreatedBefore}{A filter that returns only contexts created on or before the specified
time.}

\item{SortBy}{The property used to sort results. The default value is \code{CreationTime}.}

\item{SortOrder}{The sort order. The default value is \code{Descending}.}

\item{NextToken}{If the previous call to \code{\link[=sagemaker_list_contexts]{list_contexts}}
didn't return the full set of contexts, the call returns a token for
getting the next set of contexts.}

\item{MaxResults}{The maximum number of contexts to return in the response. The default
value is 10.}
}
\value{
A list with the following syntax:\preformatted{list(
  ContextSummaries = list(
    list(
      ContextArn = "string",
      ContextName = "string",
      Source = list(
        SourceUri = "string",
        SourceType = "string",
        SourceId = "string"
      ),
      ContextType = "string",
      CreationTime = as.POSIXct(
        "2015-01-01"
      ),
      LastModifiedTime = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Lists the contexts in your account and their properties.
}
\section{Request syntax}{
\preformatted{svc$list_contexts(
  SourceUri = "string",
  ContextType = "string",
  CreatedAfter = as.POSIXct(
    "2015-01-01"
  ),
  CreatedBefore = as.POSIXct(
    "2015-01-01"
  ),
  SortBy = "Name"|"CreationTime",
  SortOrder = "Ascending"|"Descending",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
