% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_domains}
\alias{sagemaker_list_domains}
\title{Lists the domains}
\usage{
sagemaker_list_domains(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{If the previous response was truncated, you will receive this token. Use
it in your next request to receive the next set of results.}

\item{MaxResults}{The total number of items to return in the response. If the total number
of items available is more than the value specified, a \code{NextToken} is
provided in the response. To resume pagination, provide the \code{NextToken}
value in the as part of a subsequent call. The default value is 10.}
}
\description{
Lists the domains.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_list_domains/} for full documentation.
}
\keyword{internal}
