% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_domains}
\alias{sagemaker_list_domains}
\title{Lists the domains}
\usage{
sagemaker_list_domains(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{If the previous response was truncated, you will receive this token. Use
it in your next request to receive the next set of results.}

\item{MaxResults}{Returns a list up to a specified limit.}
}
\description{
Lists the domains.

See \url{https://paws-r.github.io/docs/sagemaker/list_domains.html} for full documentation.
}
\keyword{internal}
