% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_flow_definitions}
\alias{sagemaker_list_flow_definitions}
\title{Returns information about the flow definitions in your account}
\usage{
sagemaker_list_flow_definitions(CreationTimeAfter, CreationTimeBefore,
  SortOrder, NextToken, MaxResults)
}
\arguments{
\item{CreationTimeAfter}{A filter that returns only flow definitions with a creation time greater
than or equal to the specified timestamp.}

\item{CreationTimeBefore}{A filter that returns only flow definitions that were created before the
specified timestamp.}

\item{SortOrder}{An optional value that specifies whether you want the results sorted in
\code{Ascending} or \code{Descending} order.}

\item{NextToken}{A token to resume pagination.}

\item{MaxResults}{The total number of items to return. If the total number of available
items is more than the value specified in \code{MaxResults}, then a
\code{NextToken} will be provided in the output that you can use to resume
pagination.}
}
\value{
A list with the following syntax:\preformatted{list(
  FlowDefinitionSummaries = list(
    list(
      FlowDefinitionName = "string",
      FlowDefinitionArn = "string",
      FlowDefinitionStatus = "Initializing"|"Active"|"Failed"|"Deleting",
      CreationTime = as.POSIXct(
        "2015-01-01"
      ),
      FailureReason = "string"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Returns information about the flow definitions in your account.
}
\section{Request syntax}{
\preformatted{svc$list_flow_definitions(
  CreationTimeAfter = as.POSIXct(
    "2015-01-01"
  ),
  CreationTimeBefore = as.POSIXct(
    "2015-01-01"
  ),
  SortOrder = "Ascending"|"Descending",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
