% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_human_task_uis}
\alias{sagemaker_list_human_task_uis}
\title{Returns information about the human task user interfaces in your account}
\usage{
sagemaker_list_human_task_uis(
  CreationTimeAfter = NULL,
  CreationTimeBefore = NULL,
  SortOrder = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{CreationTimeAfter}{A filter that returns only human task user interfaces with a creation
time greater than or equal to the specified timestamp.}

\item{CreationTimeBefore}{A filter that returns only human task user interfaces that were created
before the specified timestamp.}

\item{SortOrder}{An optional value that specifies whether you want the results sorted in
\code{Ascending} or \code{Descending} order.}

\item{NextToken}{A token to resume pagination.}

\item{MaxResults}{The total number of items to return. If the total number of available
items is more than the value specified in \code{MaxResults}, then a
\code{NextToken} will be provided in the output that you can use to resume
pagination.}
}
\description{
Returns information about the human task user interfaces in your account.

See \url{https://paws-r.github.io/docs/sagemaker/list_human_task_uis.html} for full documentation.
}
\keyword{internal}
