% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_monitoring_executions}
\alias{sagemaker_list_monitoring_executions}
\title{Returns list of all monitoring job executions}
\usage{
sagemaker_list_monitoring_executions(MonitoringScheduleName,
  EndpointName, SortBy, SortOrder, NextToken, MaxResults,
  ScheduledTimeBefore, ScheduledTimeAfter, CreationTimeBefore,
  CreationTimeAfter, LastModifiedTimeBefore, LastModifiedTimeAfter,
  StatusEquals, MonitoringJobDefinitionName, MonitoringTypeEquals)
}
\arguments{
\item{MonitoringScheduleName}{Name of a specific schedule to fetch jobs for.}

\item{EndpointName}{Name of a specific endpoint to fetch jobs for.}

\item{SortBy}{Whether to sort results by \code{Status}, \code{CreationTime}, \code{ScheduledTime}
field. The default is \code{CreationTime}.}

\item{SortOrder}{Whether to sort the results in \code{Ascending} or \code{Descending} order. The
default is \code{Descending}.}

\item{NextToken}{The token returned if the response is truncated. To retrieve the next
set of job executions, use it in the next request.}

\item{MaxResults}{The maximum number of jobs to return in the response. The default value
is 10.}

\item{ScheduledTimeBefore}{Filter for jobs scheduled before a specified time.}

\item{ScheduledTimeAfter}{Filter for jobs scheduled after a specified time.}

\item{CreationTimeBefore}{A filter that returns only jobs created before a specified time.}

\item{CreationTimeAfter}{A filter that returns only jobs created after a specified time.}

\item{LastModifiedTimeBefore}{A filter that returns only jobs modified after a specified time.}

\item{LastModifiedTimeAfter}{A filter that returns only jobs modified before a specified time.}

\item{StatusEquals}{A filter that retrieves only jobs with a specific status.}

\item{MonitoringJobDefinitionName}{Gets a list of the monitoring job runs of the specified monitoring job
definitions.}

\item{MonitoringTypeEquals}{A filter that returns only the monitoring job runs of the specified
monitoring type.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  MonitoringExecutionSummaries = list(
    list(
      MonitoringScheduleName = "string",
      ScheduledTime = as.POSIXct(
        "2015-01-01"
      ),
      CreationTime = as.POSIXct(
        "2015-01-01"
      ),
      LastModifiedTime = as.POSIXct(
        "2015-01-01"
      ),
      MonitoringExecutionStatus = "Pending"|"Completed"|"CompletedWithViolations"|"InProgress"|"Failed"|"Stopping"|"Stopped",
      ProcessingJobArn = "string",
      EndpointName = "string",
      FailureReason = "string",
      MonitoringJobDefinitionName = "string",
      MonitoringType = "DataQuality"|"ModelQuality"|"ModelBias"|"ModelExplainability"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns list of all monitoring job executions.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_monitoring_executions(
  MonitoringScheduleName = "string",
  EndpointName = "string",
  SortBy = "CreationTime"|"ScheduledTime"|"Status",
  SortOrder = "Ascending"|"Descending",
  NextToken = "string",
  MaxResults = 123,
  ScheduledTimeBefore = as.POSIXct(
    "2015-01-01"
  ),
  ScheduledTimeAfter = as.POSIXct(
    "2015-01-01"
  ),
  CreationTimeBefore = as.POSIXct(
    "2015-01-01"
  ),
  CreationTimeAfter = as.POSIXct(
    "2015-01-01"
  ),
  LastModifiedTimeBefore = as.POSIXct(
    "2015-01-01"
  ),
  LastModifiedTimeAfter = as.POSIXct(
    "2015-01-01"
  ),
  StatusEquals = "Pending"|"Completed"|"CompletedWithViolations"|"InProgress"|"Failed"|"Stopping"|"Stopped",
  MonitoringJobDefinitionName = "string",
  MonitoringTypeEquals = "DataQuality"|"ModelQuality"|"ModelBias"|"ModelExplainability"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
