% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_pipeline_execution_steps}
\alias{sagemaker_list_pipeline_execution_steps}
\title{Gets a list of PipeLineExecutionStep objects}
\usage{
sagemaker_list_pipeline_execution_steps(
  PipelineExecutionArn = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  SortOrder = NULL
)
}
\arguments{
\item{PipelineExecutionArn}{The Amazon Resource Name (ARN) of the pipeline execution.}

\item{NextToken}{If the result of the previous
\code{\link[=sagemaker_list_pipeline_execution_steps]{list_pipeline_execution_steps}}
request was truncated, the response includes a \code{NextToken}. To retrieve
the next set of pipeline execution steps, use the token in the next
request.}

\item{MaxResults}{The maximum number of pipeline execution steps to return in the
response.}

\item{SortOrder}{The field by which to sort results. The default is \code{CreatedTime}.}
}
\description{
Gets a list of \code{PipeLineExecutionStep} objects.

See \url{https://paws-r.github.io/docs/sagemaker/list_pipeline_execution_steps.html} for full documentation.
}
\keyword{internal}
