% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_pipelines}
\alias{sagemaker_list_pipelines}
\title{Gets a list of pipelines}
\usage{
sagemaker_list_pipelines(
  PipelineNamePrefix = NULL,
  CreatedAfter = NULL,
  CreatedBefore = NULL,
  SortBy = NULL,
  SortOrder = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{PipelineNamePrefix}{The prefix of the pipeline name.}

\item{CreatedAfter}{A filter that returns the pipelines that were created after a specified
time.}

\item{CreatedBefore}{A filter that returns the pipelines that were created before a specified
time.}

\item{SortBy}{The field by which to sort results. The default is \code{CreatedTime}.}

\item{SortOrder}{The sort order for results.}

\item{NextToken}{If the result of the previous
\code{\link[=sagemaker_list_pipelines]{list_pipelines}} request was truncated, the
response includes a \code{NextToken}. To retrieve the next set of pipelines,
use the token in the next request.}

\item{MaxResults}{The maximum number of pipelines to return in the response.}
}
\description{
Gets a list of pipelines.

See \url{https://paws-r.github.io/docs/sagemaker/list_pipelines.html} for full documentation.
}
\keyword{internal}
