% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_spaces}
\alias{sagemaker_list_spaces}
\title{Lists spaces}
\usage{
sagemaker_list_spaces(
  NextToken = NULL,
  MaxResults = NULL,
  SortOrder = NULL,
  SortBy = NULL,
  DomainIdEquals = NULL,
  SpaceNameContains = NULL
)
}
\arguments{
\item{NextToken}{If the previous response was truncated, you will receive this token. Use
it in your next request to receive the next set of results.}

\item{MaxResults}{Returns a list up to a specified limit.}

\item{SortOrder}{The sort order for the results. The default is \code{Ascending}.}

\item{SortBy}{The parameter by which to sort the results. The default is
\code{CreationTime}.}

\item{DomainIdEquals}{A parameter to search for the Domain ID.}

\item{SpaceNameContains}{A parameter by which to filter the results.}
}
\description{
Lists spaces.

See \url{https://paws-r.github.io/docs/sagemaker/list_spaces.html} for full documentation.
}
\keyword{internal}
