% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_list_subscribed_workteams}
\alias{sagemaker_list_subscribed_workteams}
\title{Gets a list of the work teams that you are subscribed to in the Amazon
Web Services Marketplace}
\usage{
sagemaker_list_subscribed_workteams(
  NameContains = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{NameContains}{A string in the work team name. This filter returns only work teams
whose name contains the specified string.}

\item{NextToken}{If the result of the previous
\code{\link[=sagemaker_list_subscribed_workteams]{list_subscribed_workteams}}
request was truncated, the response includes a \code{NextToken}. To retrieve
the next set of labeling jobs, use the token in the next request.}

\item{MaxResults}{The maximum number of work teams to return in each page of the response.}
}
\description{
Gets a list of the work teams that you are subscribed to in the Amazon Web Services Marketplace. The list may be empty if no work team satisfies the filter specified in the \code{NameContains} parameter.

See \url{https://paws-r.github.io/docs/sagemaker/list_subscribed_workteams.html} for full documentation.
}
\keyword{internal}
