% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_render_ui_template}
\alias{sagemaker_render_ui_template}
\title{Renders the UI template so that you can preview the worker's experience}
\usage{
sagemaker_render_ui_template(
  UiTemplate = NULL,
  Task,
  RoleArn,
  HumanTaskUiArn = NULL
)
}
\arguments{
\item{UiTemplate}{A \code{Template} object containing the worker UI template to render.}

\item{Task}{[required] A \code{RenderableTask} object containing a representative task to render.}

\item{RoleArn}{[required] The Amazon Resource Name (ARN) that has access to the S3 objects that
are used by the template.}

\item{HumanTaskUiArn}{The \code{HumanTaskUiArn} of the worker UI that you want to render. Do not
provide a \code{HumanTaskUiArn} if you use the \code{UiTemplate} parameter.

See a list of available Human Ui Amazon Resource Names (ARNs) in
\href{https://docs.aws.amazon.com/sagemaker/latest/APIReference/API_UiConfig.html}{UiConfig}.}
}
\description{
Renders the UI template so that you can preview the worker's experience.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_render_ui_template/} for full documentation.
}
\keyword{internal}
