% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_search}
\alias{sagemaker_search}
\title{Finds SageMaker resources that match a search query}
\usage{
sagemaker_search(
  Resource,
  SearchExpression = NULL,
  SortBy = NULL,
  SortOrder = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  CrossAccountFilterOption = NULL,
  VisibilityConditions = NULL
)
}
\arguments{
\item{Resource}{[required] The name of the SageMaker resource to search for.}

\item{SearchExpression}{A Boolean conditional statement. Resources must satisfy this condition
to be included in search results. You must provide at least one
subexpression, filter, or nested filter. The maximum number of recursive
\code{SubExpressions}, \code{NestedFilters}, and \code{Filters} that can be included in
a \code{SearchExpression} object is 50.}

\item{SortBy}{The name of the resource property used to sort the \code{SearchResults}. The
default is \code{LastModifiedTime}.}

\item{SortOrder}{How \code{SearchResults} are ordered. Valid values are \code{Ascending} or
\code{Descending}. The default is \code{Descending}.}

\item{NextToken}{If more than \code{MaxResults} resources match the specified
\code{SearchExpression}, the response includes a \code{NextToken}. The \code{NextToken}
can be passed to the next \code{SearchRequest} to continue retrieving
results.}

\item{MaxResults}{The maximum number of results to return.}

\item{CrossAccountFilterOption}{A cross account filter option. When the value is \code{"CrossAccount"} the
search results will only include resources made discoverable to you from
other accounts. When the value is \code{"SameAccount"} or \code{null} the search
results will only include resources from your account. Default is
\code{null}. For more information on searching for resources made
discoverable to your account, see \href{https://docs.aws.amazon.com/sagemaker/latest/dg/feature-store-cross-account-discoverability-use.html}{Search discoverable resources}
in the SageMaker Developer Guide. The maximum number of
\code{ResourceCatalog}s viewable is 1000.}

\item{VisibilityConditions}{Limits the results of your search request to the resources that you can
access.}
}
\description{
Finds SageMaker resources that match a search query. Matching resources are returned as a list of \code{SearchRecord} objects in the response. You can sort the search results by any resource property in a ascending or descending order.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_search/} for full documentation.
}
\keyword{internal}
