% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_start_inference_experiment}
\alias{sagemaker_start_inference_experiment}
\title{Starts an inference experiment}
\usage{
sagemaker_start_inference_experiment(Name)
}
\arguments{
\item{Name}{[required] The name of the inference experiment to start.}
}
\description{
Starts an inference experiment.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_start_inference_experiment/} for full documentation.
}
\keyword{internal}
