% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_start_pipeline_execution}
\alias{sagemaker_start_pipeline_execution}
\title{Starts a pipeline execution}
\usage{
sagemaker_start_pipeline_execution(
  PipelineName,
  PipelineExecutionDisplayName = NULL,
  PipelineParameters = NULL,
  PipelineExecutionDescription = NULL,
  ClientRequestToken,
  ParallelismConfiguration = NULL,
  SelectiveExecutionConfig = NULL,
  PipelineVersionId = NULL,
  MlflowExperimentName = NULL
)
}
\arguments{
\item{PipelineName}{[required] The name or Amazon Resource Name (ARN) of the pipeline.}

\item{PipelineExecutionDisplayName}{The display name of the pipeline execution.}

\item{PipelineParameters}{Contains a list of pipeline parameters. This list can be empty.}

\item{PipelineExecutionDescription}{The description of the pipeline execution.}

\item{ClientRequestToken}{[required] A unique, case-sensitive identifier that you provide to ensure the
idempotency of the operation. An idempotent operation completes no more
than once.}

\item{ParallelismConfiguration}{This configuration, if specified, overrides the parallelism
configuration of the parent pipeline for this specific run.}

\item{SelectiveExecutionConfig}{The selective execution configuration applied to the pipeline run.}

\item{PipelineVersionId}{The ID of the pipeline version to start execution from.}

\item{MlflowExperimentName}{The MLflow experiment name of the start execution.}
}
\description{
Starts a pipeline execution.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_start_pipeline_execution/} for full documentation.
}
\keyword{internal}
