% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_stop_auto_ml_job}
\alias{sagemaker_stop_auto_ml_job}
\title{A method for forcing a running job to shut down}
\usage{
sagemaker_stop_auto_ml_job(AutoMLJobName)
}
\arguments{
\item{AutoMLJobName}{[required] The name of the object you are requesting.}
}
\description{
A method for forcing a running job to shut down.

See \url{https://paws-r.github.io/docs/sagemaker/stop_auto_ml_job.html} for full documentation.
}
\keyword{internal}
