% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_stop_auto_ml_job}
\alias{sagemaker_stop_auto_ml_job}
\title{A method for forcing the termination of a running job}
\usage{
sagemaker_stop_auto_ml_job(AutoMLJobName)
}
\arguments{
\item{AutoMLJobName}{[required] The name of the object you are requesting.}
}
\description{
A method for forcing the termination of a running job.
}
\section{Request syntax}{
\preformatted{svc$stop_auto_ml_job(
  AutoMLJobName = "string"
)
}
}

\keyword{internal}
