% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_stop_edge_packaging_job}
\alias{sagemaker_stop_edge_packaging_job}
\title{Request to stop an edge packaging job}
\usage{
sagemaker_stop_edge_packaging_job(EdgePackagingJobName)
}
\arguments{
\item{EdgePackagingJobName}{[required] The name of the edge packaging job.}
}
\value{
An empty list.
}
\description{
Request to stop an edge packaging job.
}
\section{Request syntax}{
\preformatted{svc$stop_edge_packaging_job(
  EdgePackagingJobName = "string"
)
}
}

\keyword{internal}
