% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_stop_monitoring_schedule}
\alias{sagemaker_stop_monitoring_schedule}
\title{Stops a previously started monitoring schedule}
\usage{
sagemaker_stop_monitoring_schedule(MonitoringScheduleName)
}
\arguments{
\item{MonitoringScheduleName}{[required] The name of the schedule to stop.}
}
\value{
An empty list.
}
\description{
Stops a previously started monitoring schedule.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$stop_monitoring_schedule(
  MonitoringScheduleName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
