% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_stop_pipeline_execution}
\alias{sagemaker_stop_pipeline_execution}
\title{Stops a pipeline execution}
\usage{
sagemaker_stop_pipeline_execution(PipelineExecutionArn, ClientRequestToken)
}
\arguments{
\item{PipelineExecutionArn}{[required] The Amazon Resource Name (ARN) of the pipeline execution.}

\item{ClientRequestToken}{[required] A unique, case-sensitive identifier that you provide to ensure the
idempotency of the operation. An idempotent operation completes no more
than once.}
}
\description{
Stops a pipeline execution.

See \url{https://paws-r.github.io/docs/sagemaker/stop_pipeline_execution.html} for full documentation.
}
\keyword{internal}
