% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_stop_transform_job}
\alias{sagemaker_stop_transform_job}
\title{Stops a transform job}
\usage{
sagemaker_stop_transform_job(TransformJobName)
}
\arguments{
\item{TransformJobName}{[required] The name of the transform job to stop.}
}
\value{
An empty list.
}
\description{
Stops a transform job.

When Amazon SageMaker receives a
\code{\link[=sagemaker_stop_transform_job]{stop_transform_job}} request, the status
of the job changes to \code{Stopping}. After Amazon SageMaker stops the job,
the status is set to \code{Stopped}. When you stop a transform job before it
is completed, Amazon SageMaker doesn't store the job's output in Amazon
S3.
}
\section{Request syntax}{
\preformatted{svc$stop_transform_job(
  TransformJobName = "string"
)
}
}

\keyword{internal}
