% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_action}
\alias{sagemaker_update_action}
\title{Updates an action}
\usage{
sagemaker_update_action(ActionName, Description, Status, Properties,
  PropertiesToRemove)
}
\arguments{
\item{ActionName}{[required] The name of the action to update.}

\item{Description}{The new description for the action.}

\item{Status}{The new status for the action.}

\item{Properties}{The new list of properties. Overwrites the current property list.}

\item{PropertiesToRemove}{A list of properties to remove.}
}
\description{
Updates an action.
}
\section{Request syntax}{
\preformatted{svc$update_action(
  ActionName = "string",
  Description = "string",
  Status = "Unknown"|"InProgress"|"Completed"|"Failed"|"Stopping"|"Stopped",
  Properties = list(
    "string"
  ),
  PropertiesToRemove = list(
    "string"
  )
)
}
}

\keyword{internal}
