% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_app_image_config}
\alias{sagemaker_update_app_image_config}
\title{Updates the properties of an AppImageConfig}
\usage{
sagemaker_update_app_image_config(
  AppImageConfigName,
  KernelGatewayImageConfig = NULL,
  JupyterLabAppImageConfig = NULL
)
}
\arguments{
\item{AppImageConfigName}{[required] The name of the AppImageConfig to update.}

\item{KernelGatewayImageConfig}{The new KernelGateway app to run on the image.}

\item{JupyterLabAppImageConfig}{The JupyterLab app running on the image.}
}
\description{
Updates the properties of an AppImageConfig.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_update_app_image_config/} for full documentation.
}
\keyword{internal}
