% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_cluster}
\alias{sagemaker_update_cluster}
\title{Updates a SageMaker HyperPod cluster}
\usage{
sagemaker_update_cluster(
  ClusterName,
  InstanceGroups,
  NodeRecovery = NULL,
  InstanceGroupsToDelete = NULL
)
}
\arguments{
\item{ClusterName}{[required] Specify the name of the SageMaker HyperPod cluster you want to update.}

\item{InstanceGroups}{[required] Specify the instance groups to update.}

\item{NodeRecovery}{The node recovery mode to be applied to the SageMaker HyperPod cluster.}

\item{InstanceGroupsToDelete}{Specify the names of the instance groups to delete. Use a single \verb{,} as
the separator between multiple names.}
}
\description{
Updates a SageMaker HyperPod cluster.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_update_cluster/} for full documentation.
}
\keyword{internal}
