% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_inference_component}
\alias{sagemaker_update_inference_component}
\title{Updates an inference component}
\usage{
sagemaker_update_inference_component(
  InferenceComponentName,
  Specification = NULL,
  RuntimeConfig = NULL
)
}
\arguments{
\item{InferenceComponentName}{[required] The name of the inference component.}

\item{Specification}{Details about the resources to deploy with this inference component,
including the model, container, and compute resources.}

\item{RuntimeConfig}{Runtime settings for a model that is deployed with an inference
component.}
}
\description{
Updates an inference component.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_update_inference_component/} for full documentation.
}
\keyword{internal}
