% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_pipeline}
\alias{sagemaker_update_pipeline}
\title{Updates a pipeline}
\usage{
sagemaker_update_pipeline(
  PipelineName,
  PipelineDisplayName = NULL,
  PipelineDefinition = NULL,
  PipelineDefinitionS3Location = NULL,
  PipelineDescription = NULL,
  RoleArn = NULL,
  ParallelismConfiguration = NULL
)
}
\arguments{
\item{PipelineName}{[required] The name of the pipeline to update.}

\item{PipelineDisplayName}{The display name of the pipeline.}

\item{PipelineDefinition}{The JSON pipeline definition.}

\item{PipelineDefinitionS3Location}{The location of the pipeline definition stored in Amazon S3. If
specified, SageMaker will retrieve the pipeline definition from this
location.}

\item{PipelineDescription}{The description of the pipeline.}

\item{RoleArn}{The Amazon Resource Name (ARN) that the pipeline uses to execute.}

\item{ParallelismConfiguration}{If specified, it applies to all executions of this pipeline by default.}
}
\description{
Updates a pipeline.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_update_pipeline/} for full documentation.
}
\keyword{internal}
