% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_pipeline_execution}
\alias{sagemaker_update_pipeline_execution}
\title{Updates a pipeline execution}
\usage{
sagemaker_update_pipeline_execution(PipelineExecutionArn,
  PipelineExecutionDescription, PipelineExecutionDisplayName)
}
\arguments{
\item{PipelineExecutionArn}{[required] The Amazon Resource Name (ARN) of the pipeline execution.}

\item{PipelineExecutionDescription}{The description of the pipeline execution.}

\item{PipelineExecutionDisplayName}{The display name of the pipeline execution.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  PipelineExecutionArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Updates a pipeline execution.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_pipeline_execution(
  PipelineExecutionArn = "string",
  PipelineExecutionDescription = "string",
  PipelineExecutionDisplayName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
