% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_pipeline_execution}
\alias{sagemaker_update_pipeline_execution}
\title{Updates a pipeline execution}
\usage{
sagemaker_update_pipeline_execution(
  PipelineExecutionArn,
  PipelineExecutionDescription = NULL,
  PipelineExecutionDisplayName = NULL,
  ParallelismConfiguration = NULL
)
}
\arguments{
\item{PipelineExecutionArn}{[required] The Amazon Resource Name (ARN) of the pipeline execution.}

\item{PipelineExecutionDescription}{The description of the pipeline execution.}

\item{PipelineExecutionDisplayName}{The display name of the pipeline execution.}

\item{ParallelismConfiguration}{This configuration, if specified, overrides the parallelism
configuration of the parent pipeline for this specific run.}
}
\description{
Updates a pipeline execution.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_update_pipeline_execution/} for full documentation.
}
\keyword{internal}
