% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_space}
\alias{sagemaker_update_space}
\title{Updates the settings of a space}
\usage{
sagemaker_update_space(DomainId, SpaceName, SpaceSettings = NULL)
}
\arguments{
\item{DomainId}{[required] The ID of the associated Domain.}

\item{SpaceName}{[required] The name of the space.}

\item{SpaceSettings}{A collection of space settings.}
}
\description{
Updates the settings of a space.

See \url{https://paws-r.github.io/docs/sagemaker/update_space.html} for full documentation.
}
\keyword{internal}
