% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_update_user_profile}
\alias{sagemaker_update_user_profile}
\title{Updates a user profile}
\usage{
sagemaker_update_user_profile(DomainId, UserProfileName, UserSettings = NULL)
}
\arguments{
\item{DomainId}{[required] The domain ID.}

\item{UserProfileName}{[required] The user profile name.}

\item{UserSettings}{A collection of settings.}
}
\description{
Updates a user profile.

See \url{https://www.paws-r-sdk.com/docs/sagemaker_update_user_profile/} for full documentation.
}
\keyword{internal}
