% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakeredgemanager_operations.R
\name{sagemakeredgemanager_get_deployments}
\alias{sagemakeredgemanager_get_deployments}
\title{Use to get the active deployments from a device}
\usage{
sagemakeredgemanager_get_deployments(DeviceName, DeviceFleetName)
}
\arguments{
\item{DeviceName}{[required] The unique name of the device you want to get the configuration of
active deployments from.}

\item{DeviceFleetName}{[required] The name of the fleet that the device belongs to.}
}
\description{
Use to get the active deployments from a device.

See \url{https://paws-r.github.io/docs/sagemakeredgemanager/get_deployments.html} for full documentation.
}
\keyword{internal}
