% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemakerfeaturestoreruntime_operations.R
\name{sagemakerfeaturestoreruntime_batch_get_record}
\alias{sagemakerfeaturestoreruntime_batch_get_record}
\title{Retrieves a batch of Records from a FeatureGroup}
\usage{
sagemakerfeaturestoreruntime_batch_get_record(Identifiers)
}
\arguments{
\item{Identifiers}{[required] A list of \code{FeatureGroup} names, with their corresponding
\code{RecordIdentifier} value, and Feature name that have been requested to
be retrieved in batch.}
}
\description{
Retrieves a batch of \code{Records} from a \code{FeatureGroup}.

See \url{https://www.paws-r-sdk.com/docs/sagemakerfeaturestoreruntime_batch_get_record/} for full documentation.
}
\keyword{internal}
