% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textract_operations.R
\name{textract_analyze_expense}
\alias{textract_analyze_expense}
\title{AnalyzeExpense synchronously analyzes an input document for financially
related relationships between text}
\usage{
textract_analyze_expense(Document)
}
\arguments{
\item{Document}{[required]}
}
\description{
\code{\link[=textract_analyze_expense]{analyze_expense}} synchronously analyzes an input document for financially related relationships between text.

See \url{https://paws-r.github.io/docs/textract/analyze_expense.html} for full documentation.
}
\keyword{internal}
