% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_delete_language_model}
\alias{transcribeservice_delete_language_model}
\title{Deletes a custom language model using its name}
\usage{
transcribeservice_delete_language_model(ModelName)
}
\arguments{
\item{ModelName}{[required] The name of the model you're choosing to delete.}
}
\value{
An empty list.
}
\description{
Deletes a custom language model using its name.
}
\section{Request syntax}{
\preformatted{svc$delete_language_model(
  ModelName = "string"
)
}
}

\keyword{internal}
