% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_delete_medical_vocabulary}
\alias{transcribeservice_delete_medical_vocabulary}
\title{Deletes a custom medical vocabulary}
\usage{
transcribeservice_delete_medical_vocabulary(VocabularyName)
}
\arguments{
\item{VocabularyName}{[required] The name of the custom medical vocabulary you want to delete. Vocabulary
names are case sensitive.}
}
\description{
Deletes a custom medical vocabulary. To use this operation, specify the name of the vocabulary you want to delete using \code{VocabularyName}. Vocabulary names are case sensitive.

See \url{https://paws-r.github.io/docs/transcribeservice/delete_medical_vocabulary.html} for full documentation.
}
\keyword{internal}
