% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_delete_vocabulary}
\alias{transcribeservice_delete_vocabulary}
\title{Deletes a custom vocabulary}
\usage{
transcribeservice_delete_vocabulary(VocabularyName)
}
\arguments{
\item{VocabularyName}{[required] The name of the custom vocabulary you want to delete. Custom vocabulary
names are case sensitive.}
}
\description{
Deletes a custom vocabulary. To use this operation, specify the name of the custom vocabulary you want to delete using \code{VocabularyName}. Custom vocabulary names are case sensitive.

See \url{https://www.paws-r-sdk.com/docs/transcribeservice_delete_vocabulary/} for full documentation.
}
\keyword{internal}
