% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_delete_vocabulary_filter}
\alias{transcribeservice_delete_vocabulary_filter}
\title{Deletes a vocabulary filter}
\usage{
transcribeservice_delete_vocabulary_filter(VocabularyFilterName)
}
\arguments{
\item{VocabularyFilterName}{[required] The name of the custom vocabulary filter you want to delete. Vocabulary
filter names are case sensitive.}
}
\description{
Deletes a vocabulary filter. To use this operation, specify the name of the vocabulary filter you want to delete using \code{VocabularyFilterName}. Vocabulary filter names are case sensitive.

See \url{https://paws-r.github.io/docs/transcribeservice/delete_vocabulary_filter.html} for full documentation.
}
\keyword{internal}
