% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_delete_vocabulary_filter}
\alias{transcribeservice_delete_vocabulary_filter}
\title{Deletes a custom vocabulary filter}
\usage{
transcribeservice_delete_vocabulary_filter(VocabularyFilterName)
}
\arguments{
\item{VocabularyFilterName}{[required] The name of the custom vocabulary filter you want to delete. Custom
vocabulary filter names are case sensitive.}
}
\description{
Deletes a custom vocabulary filter. To use this operation, specify the name of the custom vocabulary filter you want to delete using \code{VocabularyFilterName}. Custom vocabulary filter names are case sensitive.

See \url{https://www.paws-r-sdk.com/docs/transcribeservice_delete_vocabulary_filter/} for full documentation.
}
\keyword{internal}
