% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_describe_language_model}
\alias{transcribeservice_describe_language_model}
\title{Provides information about the specified custom language model}
\usage{
transcribeservice_describe_language_model(ModelName)
}
\arguments{
\item{ModelName}{[required] The name of the custom language model you want information about. Model
names are case sensitive.}
}
\description{
Provides information about the specified custom language model.

See \url{https://www.paws-r-sdk.com/docs/transcribeservice_describe_language_model/} for full documentation.
}
\keyword{internal}
