% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_get_medical_transcription_job}
\alias{transcribeservice_get_medical_transcription_job}
\title{Provides information about the specified medical transcription job}
\usage{
transcribeservice_get_medical_transcription_job(MedicalTranscriptionJobName)
}
\arguments{
\item{MedicalTranscriptionJobName}{[required] The name of the medical transcription job you want information about.
Job names are case sensitive.}
}
\description{
Provides information about the specified medical transcription job.

See \url{https://paws-r.github.io/docs/transcribeservice/get_medical_transcription_job.html} for full documentation.
}
\keyword{internal}
