% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_get_vocabulary_filter}
\alias{transcribeservice_get_vocabulary_filter}
\title{Provides information about the specified custom vocabulary filter}
\usage{
transcribeservice_get_vocabulary_filter(VocabularyFilterName)
}
\arguments{
\item{VocabularyFilterName}{[required] The name of the custom vocabulary filter you want information about.
Custom vocabulary filter names are case sensitive.}
}
\description{
Provides information about the specified custom vocabulary filter.

See \url{https://www.paws-r-sdk.com/docs/transcribeservice_get_vocabulary_filter/} for full documentation.
}
\keyword{internal}
