% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_list_medical_vocabularies}
\alias{transcribeservice_list_medical_vocabularies}
\title{Provides a list of custom medical vocabularies that match the specified
criteria}
\usage{
transcribeservice_list_medical_vocabularies(
  NextToken = NULL,
  MaxResults = NULL,
  StateEquals = NULL,
  NameContains = NULL
)
}
\arguments{
\item{NextToken}{If your
\code{\link[=transcribeservice_list_medical_vocabularies]{list_medical_vocabularies}}
request returns more results than can be displayed, \code{NextToken} is
displayed in the response with an associated string. To get the next
page of results, copy this string and repeat your request, including
\code{NextToken} with the value of the copied string. Repeat as needed to
view all your results.}

\item{MaxResults}{The maximum number of custom medical vocabularies to return in each page
of results. If there are fewer results than the value that you specify,
only the actual results are returned. If you don't specify a value, a
default of 5 is used.}

\item{StateEquals}{Returns only custom medical vocabularies with the specified state.
Custom vocabularies are ordered by creation date, with the newest
vocabulary first. If you don't include \code{StateEquals}, all custom medical
vocabularies are returned.}

\item{NameContains}{Returns only the custom medical vocabularies that contain the specified
string. The search is not case sensitive.}
}
\description{
Provides a list of custom medical vocabularies that match the specified criteria. If no criteria are specified, all custom medical vocabularies are returned.

See \url{https://www.paws-r-sdk.com/docs/transcribeservice_list_medical_vocabularies/} for full documentation.
}
\keyword{internal}
