% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeservice_operations.R
\name{transcribeservice_update_call_analytics_category}
\alias{transcribeservice_update_call_analytics_category}
\title{Updates the specified Call Analytics category with new rules}
\usage{
transcribeservice_update_call_analytics_category(CategoryName, Rules)
}
\arguments{
\item{CategoryName}{[required] The name of the Call Analytics category you want to update. Category
names are case sensitive.}

\item{Rules}{[required] The rules used for the updated Call Analytics category. The rules you
provide in this field replace the ones that are currently being used in
the specified category.}
}
\description{
Updates the specified Call Analytics category with new rules. Note that the \code{\link[=transcribeservice_update_call_analytics_category]{update_call_analytics_category}} operation overwrites all existing rules contained in the specified category. You cannot append additional rules onto an existing category.

See \url{https://paws-r.github.io/docs/transcribeservice/update_call_analytics_category.html} for full documentation.
}
\keyword{internal}
