% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_operations.R
\name{translate_describe_text_translation_job}
\alias{translate_describe_text_translation_job}
\title{Gets the properties associated with an asynchronous batch translation
job including name, ID, status, source and target languages,
input/output S3 buckets, and so on}
\usage{
translate_describe_text_translation_job(JobId)
}
\arguments{
\item{JobId}{[required] The identifier that Amazon Translate generated for the job. The
\code{\link[=translate_start_text_translation_job]{start_text_translation_job}}
operation returns this identifier in its response.}
}
\description{
Gets the properties associated with an asynchronous batch translation job including name, ID, status, source and target languages, input/output S3 buckets, and so on.

See \url{https://www.paws-r-sdk.com/docs/translate_describe_text_translation_job/} for full documentation.
}
\keyword{internal}
