% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_operations.R
\name{translate_get_parallel_data}
\alias{translate_get_parallel_data}
\title{Provides information about a parallel data resource}
\usage{
translate_get_parallel_data(Name)
}
\arguments{
\item{Name}{[required] The name of the parallel data resource that is being retrieved.}
}
\description{
Provides information about a parallel data resource.
}
\section{Request syntax}{
\preformatted{svc$get_parallel_data(
  Name = "string"
)
}
}

\keyword{internal}
