% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_operations.R
\name{translate_get_terminology}
\alias{translate_get_terminology}
\title{Retrieves a custom terminology}
\usage{
translate_get_terminology(Name, TerminologyDataFormat = NULL)
}
\arguments{
\item{Name}{[required] The name of the custom terminology being retrieved.}

\item{TerminologyDataFormat}{The data format of the custom terminology being retrieved.

If you don't specify this parameter, Amazon Translate returns a file
with the same format as the file that was imported to create the
terminology.

If you specify this parameter when you retrieve a multi-directional
terminology resource, you must specify the same format as the input file
that was imported to create it. Otherwise, Amazon Translate throws an
error.}
}
\description{
Retrieves a custom terminology.

See \url{https://www.paws-r-sdk.com/docs/translate_get_terminology/} for full documentation.
}
\keyword{internal}
