% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_operations.R
\name{translate_list_parallel_data}
\alias{translate_list_parallel_data}
\title{Provides a list of your parallel data resources in Amazon Translate}
\usage{
translate_list_parallel_data(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{A string that specifies the next page of results to return in a
paginated response.}

\item{MaxResults}{The maximum number of parallel data resources returned for each request.}
}
\description{
Provides a list of your parallel data resources in Amazon Translate.

See \url{https://www.paws-r-sdk.com/docs/translate_list_parallel_data/} for full documentation.
}
\keyword{internal}
