% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_operations.R
\name{translate_stop_text_translation_job}
\alias{translate_stop_text_translation_job}
\title{Stops an asynchronous batch translation job that is in progress}
\usage{
translate_stop_text_translation_job(JobId)
}
\arguments{
\item{JobId}{[required] The job ID of the job to be stopped.}
}
\description{
Stops an asynchronous batch translation job that is in progress.

See \url{https://www.paws-r-sdk.com/docs/translate_stop_text_translation_job/} for full documentation.
}
\keyword{internal}
